AWSSDK.Athena.jl

AWSSDK.Athena

Amazon Athena is an interactive query service that lets you use standard SQL to analyze data directly in Amazon S3. You can point Athena at your data in Amazon S3 and run ad-hoc queries and get results in seconds. Athena is serverless, so there is no infrastructure to set up or manage. You pay only for the queries you run. Athena scales automatically—executing queries in parallel—so results are fast, even with large datasets and complex queries. For more information, see What is Amazon Athena in the Amazon Athena User Guide.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

This document is generated from apis/athena-2017-05-18.normal.json. See JuliaCloud/AWSCore.jl.

AWSSDK.Athena.batch_get_named_query

AWSSDK.Athena.batch_get_query_execution

AWSSDK.Athena.create_named_query

AWSSDK.Athena.delete_named_query

AWSSDK.Athena.get_named_query

AWSSDK.Athena.get_query_execution

AWSSDK.Athena.get_query_results

AWSSDK.Athena.list_named_queries

AWSSDK.Athena.list_query_executions

AWSSDK.Athena.start_query_execution

AWSSDK.Athena.stop_query_execution

AWSSDK.Athena.batch_get_named_query — Function.

using AWSSDK.Athena.batch_get_named_query
batch_get_named_query([::AWSConfig], arguments::Dict)
batch_get_named_query([::AWSConfig]; NamedQueryIds=)

using AWSCore.Services.athena
athena([::AWSConfig], "BatchGetNamedQuery", arguments::Dict)
athena([::AWSConfig], "BatchGetNamedQuery", NamedQueryIds=)

BatchGetNamedQuery Operation

Returns the details of a single named query or a list of up to 50 queries, which you provide as an array of query ID strings. Use ListNamedQueries to get the list of named query IDs. If information could not be retrieved for a submitted query ID, information about the query ID submitted is listed under UnprocessedNamedQueryId. Named queries are different from executed queries. Use BatchGetQueryExecution to get details about each unique query execution, and ListQueryExecutions to get a list of query execution IDs.

Arguments

NamedQueryIds = [::String, ...] – Required

An array of query IDs.

Returns

BatchGetNamedQueryOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.batch_get_query_execution — Function.

using AWSSDK.Athena.batch_get_query_execution
batch_get_query_execution([::AWSConfig], arguments::Dict)
batch_get_query_execution([::AWSConfig]; QueryExecutionIds=)

using AWSCore.Services.athena
athena([::AWSConfig], "BatchGetQueryExecution", arguments::Dict)
athena([::AWSConfig], "BatchGetQueryExecution", QueryExecutionIds=)

BatchGetQueryExecution Operation

Returns the details of a single query execution or a list of up to 50 query executions, which you provide as an array of query execution ID strings. To get a list of query execution IDs, use ListQueryExecutions. Query executions are different from named (saved) queries. Use BatchGetNamedQuery to get details about named queries.

Arguments

QueryExecutionIds = [::String, ...] – Required

An array of query execution IDs.

Returns

BatchGetQueryExecutionOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.create_named_query — Function.

using AWSSDK.Athena.create_named_query
create_named_query([::AWSConfig], arguments::Dict)
create_named_query([::AWSConfig]; Name=, Database=, QueryString=, <keyword arguments>)

using AWSCore.Services.athena
athena([::AWSConfig], "CreateNamedQuery", arguments::Dict)
athena([::AWSConfig], "CreateNamedQuery", Name=, Database=, QueryString=, <keyword arguments>)

CreateNamedQuery Operation

Creates a named query.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

Arguments

Name = ::String – Required

The plain language name for the query.

Description = ::String

A brief explanation of the query.

Database = ::String – Required

The database to which the query belongs.

QueryString = ::String – Required

The text of the query itself. In other words, all query statements.

ClientRequestToken = ::String

A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another CreateNamedQuery request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString, an error is returned.

Important

This token is listed as not required because AWS SDKs (for example the AWS SDK for Java) auto-generate the token for users. If you are not using the AWS SDK or the AWS CLI, you must provide this token or the action will fail.

Returns

CreateNamedQueryOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.delete_named_query — Function.

using AWSSDK.Athena.delete_named_query
delete_named_query([::AWSConfig], arguments::Dict)
delete_named_query([::AWSConfig]; NamedQueryId=)

using AWSCore.Services.athena
athena([::AWSConfig], "DeleteNamedQuery", arguments::Dict)
athena([::AWSConfig], "DeleteNamedQuery", NamedQueryId=)

DeleteNamedQuery Operation

Deletes a named query.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

Arguments

NamedQueryId = ::String – Required

The unique ID of the query to delete.

Returns

DeleteNamedQueryOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.get_named_query — Function.

using AWSSDK.Athena.get_named_query
get_named_query([::AWSConfig], arguments::Dict)
get_named_query([::AWSConfig]; NamedQueryId=)

using AWSCore.Services.athena
athena([::AWSConfig], "GetNamedQuery", arguments::Dict)
athena([::AWSConfig], "GetNamedQuery", NamedQueryId=)

GetNamedQuery Operation

Returns information about a single query.

Arguments

NamedQueryId = ::String – Required

The unique ID of the query. Use ListNamedQueries to get query IDs.

Returns

GetNamedQueryOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.get_query_execution — Function.

using AWSSDK.Athena.get_query_execution
get_query_execution([::AWSConfig], arguments::Dict)
get_query_execution([::AWSConfig]; QueryExecutionId=)

using AWSCore.Services.athena
athena([::AWSConfig], "GetQueryExecution", arguments::Dict)
athena([::AWSConfig], "GetQueryExecution", QueryExecutionId=)

GetQueryExecution Operation

Returns information about a single execution of a query. Each time a query executes, information about the query execution is saved with a unique ID.

Arguments

QueryExecutionId = ::String – Required

The unique ID of the query execution.

Returns

GetQueryExecutionOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.get_query_results — Function.

using AWSSDK.Athena.get_query_results
get_query_results([::AWSConfig], arguments::Dict)
get_query_results([::AWSConfig]; QueryExecutionId=, <keyword arguments>)

using AWSCore.Services.athena
athena([::AWSConfig], "GetQueryResults", arguments::Dict)
athena([::AWSConfig], "GetQueryResults", QueryExecutionId=, <keyword arguments>)

GetQueryResults Operation

Returns the results of a single query execution specified by QueryExecutionId. This request does not execute the query but returns results. Use StartQueryExecution to run a query.

Arguments

QueryExecutionId = ::String – Required

The unique ID of the query execution.

NextToken = ::String

The token that specifies where to start pagination if a previous request was truncated.

MaxResults = ::Int

The maximum number of results (rows) to return in this request.

Returns

GetQueryResultsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.list_named_queries — Function.

using AWSSDK.Athena.list_named_queries
list_named_queries([::AWSConfig], arguments::Dict)
list_named_queries([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.athena
athena([::AWSConfig], "ListNamedQueries", arguments::Dict)
athena([::AWSConfig], "ListNamedQueries", <keyword arguments>)

ListNamedQueries Operation

Provides a list of all available query IDs.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

Arguments

NextToken = ::String

The token that specifies where to start pagination if a previous request was truncated.

MaxResults = ::Int

The maximum number of queries to return in this request.

Returns

ListNamedQueriesOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.list_query_executions — Function.

using AWSSDK.Athena.list_query_executions
list_query_executions([::AWSConfig], arguments::Dict)
list_query_executions([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.athena
athena([::AWSConfig], "ListQueryExecutions", arguments::Dict)
athena([::AWSConfig], "ListQueryExecutions", <keyword arguments>)

ListQueryExecutions Operation

Provides a list of all available query execution IDs.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

Arguments

NextToken = ::String

The token that specifies where to start pagination if a previous request was truncated.

MaxResults = ::Int

The maximum number of query executions to return in this request.

Returns

ListQueryExecutionsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AWSSDK.Athena.start_query_execution — Function.

using AWSSDK.Athena.start_query_execution
start_query_execution([::AWSConfig], arguments::Dict)
start_query_execution([::AWSConfig]; QueryString=, ResultConfiguration=, <keyword arguments>)

using AWSCore.Services.athena
athena([::AWSConfig], "StartQueryExecution", arguments::Dict)
athena([::AWSConfig], "StartQueryExecution", QueryString=, ResultConfiguration=, <keyword arguments>)

StartQueryExecution Operation

Runs (executes) the SQL query statements contained in the Query string.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

Arguments

QueryString = ::String – Required

The SQL query statements to be executed.

ClientRequestToken = ::String

A unique case-sensitive string used to ensure the request to create the query is idempotent (executes only once). If another StartQueryExecution request is received, the same response is returned and another query is not created. If a parameter has changed, for example, the QueryString, an error is returned.

Important

This token is listed as not required because AWS SDKs (for example the AWS SDK for Java) auto-generate the token for users. If you are not using the AWS SDK or the AWS CLI, you must provide this token or the action will fail.

QueryExecutionContext = ["Database" => ::String]

The database within which the query executes.

ResultConfiguration = [ ... ] – Required

Specifies information about where and how to save the results of the query execution.

 ResultConfiguration = [
        "OutputLocation" => <required> ::String,
        "EncryptionConfiguration" =>  [
            "EncryptionOption" => <required> "SSE_S3", "SSE_KMS" or "CSE_KMS",
            "KmsKey" =>  ::String
        ]
    ]

Returns

StartQueryExecutionOutput

Exceptions

InternalServerException, InvalidRequestException or TooManyRequestsException.

See also: AWS API Documentation

AWSSDK.Athena.stop_query_execution — Function.

using AWSSDK.Athena.stop_query_execution
stop_query_execution([::AWSConfig], arguments::Dict)
stop_query_execution([::AWSConfig]; QueryExecutionId=)

using AWSCore.Services.athena
athena([::AWSConfig], "StopQueryExecution", arguments::Dict)
athena([::AWSConfig], "StopQueryExecution", QueryExecutionId=)

StopQueryExecution Operation

Stops a query execution.

For code samples using the AWS SDK for Java, see Examples and Code Samples in the Amazon Athena User Guide.

Arguments

QueryExecutionId = ::String – Required

The unique ID of the query execution to stop.

Returns

StopQueryExecutionOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation