AWSSDK.Discovery.jl

AWSSDK.Discovery

AWS Application Discovery Service

AWS Application Discovery Service helps you plan application migration projects by automatically identifying servers, virtual machines (VMs), software, and software dependencies running in your on-premises data centers. Application Discovery Service also collects application performance data, which can help you assess the outcome of your migration. The data collected by Application Discovery Service is securely retained in an Amazon-hosted and managed database in the cloud. You can export the data as a CSV or XML file into your preferred visualization tool or cloud-migration solution to plan your migration. For more information, see the Application Discovery Service FAQ.

Application Discovery Service offers two modes of operation.

Agentless discovery mode is recommended for environments that use VMware vCenter Server. This mode doesn't require you to install an agent on each host. Agentless discovery gathers server information regardless of the operating systems, which minimizes the time required for initial on-premises infrastructure assessment. Agentless discovery doesn't collect information about software and software dependencies. It also doesn't work in non-VMware environments. We recommend that you use agent-based discovery for non-VMware environments and if you want to collect information about software and software dependencies. You can also run agent-based and agentless discovery simultaneously. Use agentless discovery to quickly complete the initial infrastructure assessment and then install agents on select hosts to gather information about software and software dependencies.

Agent-based discovery mode collects a richer set of data than agentless discovery by using Amazon software, the AWS Application Discovery Agent, which you install on one or more hosts in your data center. The agent captures infrastructure and application information, including an inventory of installed software applications, system and process performance, resource utilization, and network dependencies between workloads. The information collected by agents is secured at rest and in transit to the Application Discovery Service database in the cloud.

Application Discovery Service integrates with application discovery solutions from AWS Partner Network (APN) partners. Third-party application discovery tools can query Application Discovery Service and write to the Application Discovery Service database using a public API. You can then import the data into either a visualization tool or cloud-migration solution.

Important

Application Discovery Service doesn't gather sensitive information. All data is handled according to the AWS Privacy Policy. You can operate Application Discovery Service using offline mode to inspect collected data before it is shared with the service.

Your AWS account must be granted access to Application Discovery Service, a process called whitelisting. This is true for AWS partners and customers alike. To request access, sign up for AWS Application Discovery Service here. We send you information about how to get started.

This API reference provides descriptions, syntax, and usage examples for each of the actions and data types for Application Discovery Service. The topic for each action shows the API request parameters and the response. Alternatively, you can use one of the AWS SDKs to access an API that is tailored to the programming language or platform that you're using. For more information, see AWS SDKs.

This guide is intended for use with the AWS Application Discovery Service User Guide .

This document is generated from apis/discovery-2015-11-01.normal.json. See JuliaCloud/AWSCore.jl.

AWSSDK.Discovery.associate_configuration_items_to_application

AWSSDK.Discovery.create_application

AWSSDK.Discovery.create_tags

AWSSDK.Discovery.delete_applications

AWSSDK.Discovery.delete_tags

AWSSDK.Discovery.describe_agents

AWSSDK.Discovery.describe_configurations

AWSSDK.Discovery.describe_export_configurations

AWSSDK.Discovery.describe_export_tasks

AWSSDK.Discovery.describe_tags

AWSSDK.Discovery.disassociate_configuration_items_from_application

AWSSDK.Discovery.export_configurations

AWSSDK.Discovery.get_discovery_summary

AWSSDK.Discovery.list_configurations

AWSSDK.Discovery.list_server_neighbors

AWSSDK.Discovery.start_data_collection_by_agent_ids

AWSSDK.Discovery.start_export_task

AWSSDK.Discovery.stop_data_collection_by_agent_ids

AWSSDK.Discovery.update_application

AWSSDK.Discovery.associate_configuration_items_to_application — Function.

using AWSSDK.Discovery.associate_configuration_items_to_application
associate_configuration_items_to_application([::AWSConfig], arguments::Dict)
associate_configuration_items_to_application([::AWSConfig]; applicationConfigurationId=, configurationIds=)

using AWSCore.Services.discovery
discovery([::AWSConfig], "AssociateConfigurationItemsToApplication", arguments::Dict)
discovery([::AWSConfig], "AssociateConfigurationItemsToApplication", applicationConfigurationId=, configurationIds=)

AssociateConfigurationItemsToApplication Operation

Associates one or more configuration items with an application.

Arguments

applicationConfigurationId = ::String – Required

The configuration ID of an application with which items are to be associated.

configurationIds = [::String, ...] – Required

The ID of each configuration item to be associated with an application.

Returns

AssociateConfigurationItemsToApplicationResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.create_application — Function.

using AWSSDK.Discovery.create_application
create_application([::AWSConfig], arguments::Dict)
create_application([::AWSConfig]; name=, <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "CreateApplication", arguments::Dict)
discovery([::AWSConfig], "CreateApplication", name=, <keyword arguments>)

CreateApplication Operation

Creates an application with the given name and description.

Arguments

name = ::String – Required

Name of the application to be created.

description = ::String

Description of the application to be created.

Returns

CreateApplicationResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.create_tags — Function.

using AWSSDK.Discovery.create_tags
create_tags([::AWSConfig], arguments::Dict)
create_tags([::AWSConfig]; configurationIds=, tags=)

using AWSCore.Services.discovery
discovery([::AWSConfig], "CreateTags", arguments::Dict)
discovery([::AWSConfig], "CreateTags", configurationIds=, tags=)

CreateTags Operation

Creates one or more tags for configuration items. Tags are metadata that help you categorize IT assets. This API accepts a list of multiple configuration items.

Arguments

configurationIds = [::String, ...] – Required

A list of configuration items that you want to tag.

tags = [[ ... ], ...] – Required

Tags that you want to associate with one or more configuration items. Specify the tags that you want to create in a key-value format. For example:

{"key": "serverType", "value": "webServer"}

 tags = [[
        "key" => <required> ::String,
        "value" => <required> ::String
    ], ...]

Returns

CreateTagsResponse

Exceptions

AuthorizationErrorException, ResourceNotFoundException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.delete_applications — Function.

using AWSSDK.Discovery.delete_applications
delete_applications([::AWSConfig], arguments::Dict)
delete_applications([::AWSConfig]; configurationIds=)

using AWSCore.Services.discovery
discovery([::AWSConfig], "DeleteApplications", arguments::Dict)
discovery([::AWSConfig], "DeleteApplications", configurationIds=)

DeleteApplications Operation

Deletes a list of applications and their associations with configuration items.

Arguments

configurationIds = [::String, ...] – Required

Configuration ID of an application to be deleted.

Returns

DeleteApplicationsResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.delete_tags — Function.

using AWSSDK.Discovery.delete_tags
delete_tags([::AWSConfig], arguments::Dict)
delete_tags([::AWSConfig]; configurationIds=, <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "DeleteTags", arguments::Dict)
discovery([::AWSConfig], "DeleteTags", configurationIds=, <keyword arguments>)

DeleteTags Operation

Deletes the association between configuration items and one or more tags. This API accepts a list of multiple configuration items.

Arguments

configurationIds = [::String, ...] – Required

A list of configuration items with tags that you want to delete.

tags = [[ ... ], ...]

Tags that you want to delete from one or more configuration items. Specify the tags that you want to delete in a key-value format. For example:

{"key": "serverType", "value": "webServer"}

 tags = [[
        "key" => <required> ::String,
        "value" => <required> ::String
    ], ...]

Returns

DeleteTagsResponse

Exceptions

AuthorizationErrorException, ResourceNotFoundException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.describe_agents — Function.

using AWSSDK.Discovery.describe_agents
describe_agents([::AWSConfig], arguments::Dict)
describe_agents([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "DescribeAgents", arguments::Dict)
discovery([::AWSConfig], "DescribeAgents", <keyword arguments>)

DescribeAgents Operation

Lists agents or the Connector by ID or lists all agents/Connectors associated with your user account if you did not specify an ID.

Arguments

agentIds = [::String, ...]

The agent or the Connector IDs for which you want information. If you specify no IDs, the system returns information about all agents/Connectors associated with your AWS user account.

filters = [[ ... ], ...]

You can filter the request using various logical operators and a key-value format. For example:

{"key": "collectionStatus", "value": "STARTED"}

 filters = [[
        "name" => <required> ::String,
        "values" => <required> [::String, ...],
        "condition" => <required> ::String
    ], ...]

maxResults = ::Int

The total number of agents/Connectors to return in a single page of output. The maximum value is 100.

nextToken = ::String

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for DescribeAgentsRequest$agentIds but set DescribeAgentsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

Returns

DescribeAgentsResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.describe_configurations — Function.

using AWSSDK.Discovery.describe_configurations
describe_configurations([::AWSConfig], arguments::Dict)
describe_configurations([::AWSConfig]; configurationIds=)

using AWSCore.Services.discovery
discovery([::AWSConfig], "DescribeConfigurations", arguments::Dict)
discovery([::AWSConfig], "DescribeConfigurations", configurationIds=)

DescribeConfigurations Operation

Retrieves attributes for a list of configuration item IDs. All of the supplied IDs must be for the same asset type (server, application, process, or connection). Output fields are specific to the asset type selected. For example, the output for a server configuration item includes a list of attributes about the server, such as host name, operating system, and number of network cards.

For a complete list of outputs for each asset type, see Using the DescribeConfigurations Action.

Arguments

configurationIds = [::String, ...] – Required

One or more configuration IDs.

Returns

DescribeConfigurationsResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.describe_export_configurations — Function.

using AWSSDK.Discovery.describe_export_configurations
describe_export_configurations([::AWSConfig], arguments::Dict)
describe_export_configurations([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "DescribeExportConfigurations", arguments::Dict)
discovery([::AWSConfig], "DescribeExportConfigurations", <keyword arguments>)

DescribeExportConfigurations Operation

Deprecated. Use DescribeExportTasks instead.

Retrieves the status of a given export process. You can retrieve status from a maximum of 100 processes.

Arguments

exportIds = [::String, ...]

A unique identifier that you can use to query the export status.

maxResults = ::Int

The maximum number of results that you want to display as a part of the query.

nextToken = ::String

A token to get the next set of results. For example, if you specify 100 IDs for DescribeExportConfigurationsRequest$exportIds but set DescribeExportConfigurationsRequest$maxResults to 10, you get results in a set of 10. Use the token in the query to get the next set of 10.

Returns

DescribeExportConfigurationsResponse

Exceptions

AuthorizationErrorException, ResourceNotFoundException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.describe_export_tasks — Function.

using AWSSDK.Discovery.describe_export_tasks
describe_export_tasks([::AWSConfig], arguments::Dict)
describe_export_tasks([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "DescribeExportTasks", arguments::Dict)
discovery([::AWSConfig], "DescribeExportTasks", <keyword arguments>)

DescribeExportTasks Operation

Retrieve status of one or more export tasks. You can retrieve the status of up to 100 export tasks.

Arguments

exportIds = [::String, ...]

One or more unique identifiers used to query the status of an export request.

filters = [[ ... ], ...]

One or more filters.

AgentId - ID of the agent whose collected data will be exported

 filters = [[
        "name" => <required> ::String,
        "values" => <required> [::String, ...],
        "condition" => <required> ::String
    ], ...]

maxResults = ::Int

The maximum number of volume results returned by DescribeExportTasks in paginated output. When this parameter is used, DescribeExportTasks only returns maxResults results in a single page along with a nextToken response element.

nextToken = ::String

The nextToken value returned from a previous paginated DescribeExportTasks request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Returns

DescribeExportTasksResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.describe_tags — Function.

using AWSSDK.Discovery.describe_tags
describe_tags([::AWSConfig], arguments::Dict)
describe_tags([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "DescribeTags", arguments::Dict)
discovery([::AWSConfig], "DescribeTags", <keyword arguments>)

DescribeTags Operation

Retrieves a list of configuration items that are tagged with a specific tag. Or retrieves a list of all tags assigned to a specific configuration item.

Arguments

filters = [[ ... ], ...]

You can filter the list using a key-value format. You can separate these items by using logical operators. Allowed filters include tagKey, tagValue, and configurationId.

 filters = [[
        "name" => <required> ::String,
        "values" => <required> [::String, ...]
    ], ...]

maxResults = ::Int

The total number of items to return in a single page of output. The maximum value is 100.

nextToken = ::String

A token to start the list. Use this token to get the next set of results.

Returns

DescribeTagsResponse

Exceptions

AuthorizationErrorException, ResourceNotFoundException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.disassociate_configuration_items_from_application — Function.

using AWSSDK.Discovery.disassociate_configuration_items_from_application
disassociate_configuration_items_from_application([::AWSConfig], arguments::Dict)
disassociate_configuration_items_from_application([::AWSConfig]; applicationConfigurationId=, configurationIds=)

using AWSCore.Services.discovery
discovery([::AWSConfig], "DisassociateConfigurationItemsFromApplication", arguments::Dict)
discovery([::AWSConfig], "DisassociateConfigurationItemsFromApplication", applicationConfigurationId=, configurationIds=)

DisassociateConfigurationItemsFromApplication Operation

Disassociates one or more configuration items from an application.

Arguments

applicationConfigurationId = ::String – Required

Configuration ID of an application from which each item is disassociated.

configurationIds = [::String, ...] – Required

Configuration ID of each item to be disassociated from an application.

Returns

DisassociateConfigurationItemsFromApplicationResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.export_configurations — Function.

using AWSSDK.Discovery.export_configurations
export_configurations([::AWSConfig])

using AWSCore.Services.discovery
discovery([::AWSConfig], "ExportConfigurations",)

ExportConfigurations Operation

Deprecated. Use StartExportTask instead.

Exports all discovered configuration data to an Amazon S3 bucket or an application that enables you to view and evaluate the data. Data includes tags and tag associations, processes, connections, servers, and system performance. This API returns an export ID that you can query using the DescribeExportConfigurations API. The system imposes a limit of two configuration exports in six hours.

Returns

ExportConfigurationsResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException, ServerInternalErrorException or OperationNotPermittedException.

See also: AWS API Documentation

AWSSDK.Discovery.get_discovery_summary — Function.

using AWSSDK.Discovery.get_discovery_summary
get_discovery_summary([::AWSConfig], arguments::Dict)
get_discovery_summary([::AWSConfig]; )

using AWSCore.Services.discovery
discovery([::AWSConfig], "GetDiscoverySummary", arguments::Dict)
discovery([::AWSConfig], "GetDiscoverySummary", )

GetDiscoverySummary Operation

Retrieves a short summary of discovered assets.

Arguments

Returns

GetDiscoverySummaryResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.list_configurations — Function.

using AWSSDK.Discovery.list_configurations
list_configurations([::AWSConfig], arguments::Dict)
list_configurations([::AWSConfig]; configurationType=, <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "ListConfigurations", arguments::Dict)
discovery([::AWSConfig], "ListConfigurations", configurationType=, <keyword arguments>)

ListConfigurations Operation

Retrieves a list of configuration items according to criteria that you specify in a filter. The filter criteria identifies the relationship requirements.

Arguments

configurationType = "SERVER", "PROCESS", "CONNECTION" or "APPLICATION" – Required

A valid configuration identified by Application Discovery Service.

filters = [[ ... ], ...]

You can filter the request using various logical operators and a key-value format. For example:

{"key": "serverType", "value": "webServer"}

For a complete list of filter options and guidance about using them with this action, see Querying Discovered Configuration Items.

 filters = [[
        "name" => <required> ::String,
        "values" => <required> [::String, ...],
        "condition" => <required> ::String
    ], ...]

maxResults = ::Int

The total number of items to return. The maximum value is 100.

nextToken = ::String

Token to retrieve the next set of results. For example, if a previous call to ListConfigurations returned 100 items, but you set ListConfigurationsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

orderBy = [[ ... ], ...]

Certain filter criteria return output that can be sorted in ascending or descending order. For a list of output characteristics for each filter, see Using the ListConfigurations Action.

 orderBy = [[
        "fieldName" => <required> ::String,
        "sortOrder" =>  "ASC" or "DESC"
    ], ...]

Returns

ListConfigurationsResponse

Exceptions

AuthorizationErrorException, ResourceNotFoundException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.list_server_neighbors — Function.

using AWSSDK.Discovery.list_server_neighbors
list_server_neighbors([::AWSConfig], arguments::Dict)
list_server_neighbors([::AWSConfig]; configurationId=, <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "ListServerNeighbors", arguments::Dict)
discovery([::AWSConfig], "ListServerNeighbors", configurationId=, <keyword arguments>)

ListServerNeighbors Operation

Retrieves a list of servers that are one network hop away from a specified server.

Arguments

configurationId = ::String – Required

Configuration ID of the server for which neighbors are being listed.

portInformationNeeded = ::Bool

Flag to indicate if port and protocol information is needed as part of the response.

neighborConfigurationIds = [::String, ...]

List of configuration IDs to test for one-hop-away.

maxResults = ::Int

Maximum number of results to return in a single page of output.

nextToken = ::String

Token to retrieve the next set of results. For example, if you previously specified 100 IDs for ListServerNeighborsRequest$neighborConfigurationIds but set ListServerNeighborsRequest$maxResults to 10, you received a set of 10 results along with a token. Use that token in this query to get the next set of 10.

Returns

ListServerNeighborsResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.start_data_collection_by_agent_ids — Function.

using AWSSDK.Discovery.start_data_collection_by_agent_ids
start_data_collection_by_agent_ids([::AWSConfig], arguments::Dict)
start_data_collection_by_agent_ids([::AWSConfig]; agentIds=)

using AWSCore.Services.discovery
discovery([::AWSConfig], "StartDataCollectionByAgentIds", arguments::Dict)
discovery([::AWSConfig], "StartDataCollectionByAgentIds", agentIds=)

StartDataCollectionByAgentIds Operation

Instructs the specified agents or connectors to start collecting data.

Arguments

agentIds = [::String, ...] – Required

The IDs of the agents or connectors from which to start collecting data. If you send a request to an agent/connector ID that you do not have permission to contact, according to your AWS account, the service does not throw an exception. Instead, it returns the error in the Description field. If you send a request to multiple agents/connectors and you do not have permission to contact some of those agents/connectors, the system does not throw an exception. Instead, the system shows Failed in the Description field.

Returns

StartDataCollectionByAgentIdsResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.start_export_task — Function.

using AWSSDK.Discovery.start_export_task
start_export_task([::AWSConfig], arguments::Dict)
start_export_task([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "StartExportTask", arguments::Dict)
discovery([::AWSConfig], "StartExportTask", <keyword arguments>)

StartExportTask Operation

Begins the export of discovered data to an S3 bucket.

If you specify agentId in a filter, the task exports up to 72 hours of detailed data collected by the identified Application Discovery Agent, including network, process, and performance details. A time range for exported agent data may be set by using startTime and endTime. Export of detailed agent data is limited to five concurrently running exports.

If you do not include an agentId filter, summary data is exported that includes both AWS Agentless Discovery Connector data and summary data from AWS Discovery Agents. Export of summary data is limited to two exports per day.

Arguments

exportDataFormat = ["CSV" or "GRAPHML", ...]

The file format for the returned export data. Default value is CSV.

filters = [[ ... ], ...]

If a filter is present, it selects the single agentId of the Application Discovery Agent for which data is exported. The agentId can be found in the results of the DescribeAgents API or CLI. If no filter is present, startTime and endTime are ignored and exported data includes both Agentless Discovery Connector data and summary data from Application Discovery agents.

 filters = [[
        "name" => <required> ::String,
        "values" => <required> [::String, ...],
        "condition" => <required> ::String
    ], ...]

startTime = timestamp

The start timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, data is exported starting from the first data collected by the agent.

endTime = timestamp

The end timestamp for exported data from the single Application Discovery Agent selected in the filters. If no value is specified, exported data includes the most recent data collected by the agent.

Returns

StartExportTaskResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException, ServerInternalErrorException or OperationNotPermittedException.

See also: AWS API Documentation

AWSSDK.Discovery.stop_data_collection_by_agent_ids — Function.

using AWSSDK.Discovery.stop_data_collection_by_agent_ids
stop_data_collection_by_agent_ids([::AWSConfig], arguments::Dict)
stop_data_collection_by_agent_ids([::AWSConfig]; agentIds=)

using AWSCore.Services.discovery
discovery([::AWSConfig], "StopDataCollectionByAgentIds", arguments::Dict)
discovery([::AWSConfig], "StopDataCollectionByAgentIds", agentIds=)

StopDataCollectionByAgentIds Operation

Instructs the specified agents or connectors to stop collecting data.

Arguments

agentIds = [::String, ...] – Required

The IDs of the agents or connectors from which to stop collecting data.

Returns

StopDataCollectionByAgentIdsResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation

AWSSDK.Discovery.update_application — Function.

using AWSSDK.Discovery.update_application
update_application([::AWSConfig], arguments::Dict)
update_application([::AWSConfig]; configurationId=, <keyword arguments>)

using AWSCore.Services.discovery
discovery([::AWSConfig], "UpdateApplication", arguments::Dict)
discovery([::AWSConfig], "UpdateApplication", configurationId=, <keyword arguments>)

UpdateApplication Operation

Updates metadata about an application.

Arguments

configurationId = ::String – Required

Configuration ID of the application to be updated.

name = ::String

New name of the application to be updated.

description = ::String

New description of the application to be updated.

Returns

UpdateApplicationResponse

Exceptions

AuthorizationErrorException, InvalidParameterException, InvalidParameterValueException or ServerInternalErrorException.

See also: AWS API Documentation