AWSSDK.IotData.jl

AWSSDK.IotData

AWS IoT

AWS IoT-Data enables secure, bi-directional communication between Internet-connected things (such as sensors, actuators, embedded devices, or smart appliances) and the AWS cloud. It implements a broker for applications and things to publish messages over HTTP (Publish) and retrieve, update, and delete thing shadows. A thing shadow is a persistent representation of your things and their state in the AWS cloud.

This document is generated from apis/iot-data-2015-05-28.normal.json. See JuliaCloud/AWSCore.jl.

AWSSDK.IotData.delete_thing_shadow

AWSSDK.IotData.get_thing_shadow

AWSSDK.IotData.publish

AWSSDK.IotData.update_thing_shadow

AWSSDK.IotData.delete_thing_shadow — Function.

using AWSSDK.IotData.delete_thing_shadow
delete_thing_shadow([::AWSConfig], arguments::Dict)
delete_thing_shadow([::AWSConfig]; thingName=)

using AWSCore.Services.iot_data
iot_data([::AWSConfig], "DELETE", "/things/{thingName}/shadow", arguments::Dict)
iot_data([::AWSConfig], "DELETE", "/things/{thingName}/shadow", thingName=)

DeleteThingShadow Operation

Deletes the thing shadow for the specified thing.

For more information, see DeleteThingShadow in the AWS IoT Developer Guide.

Arguments

thingName = ::String – Required

The name of the thing.

Returns

DeleteThingShadowResponse

Exceptions

ResourceNotFoundException, InvalidRequestException, ThrottlingException, UnauthorizedException, ServiceUnavailableException, InternalFailureException, MethodNotAllowedException or UnsupportedDocumentEncodingException.

See also: AWS API Documentation

AWSSDK.IotData.get_thing_shadow — Function.

using AWSSDK.IotData.get_thing_shadow
get_thing_shadow([::AWSConfig], arguments::Dict)
get_thing_shadow([::AWSConfig]; thingName=)

using AWSCore.Services.iot_data
iot_data([::AWSConfig], "GET", "/things/{thingName}/shadow", arguments::Dict)
iot_data([::AWSConfig], "GET", "/things/{thingName}/shadow", thingName=)

GetThingShadow Operation

Gets the thing shadow for the specified thing.

For more information, see GetThingShadow in the AWS IoT Developer Guide.

Arguments

thingName = ::String – Required

The name of the thing.

Returns

GetThingShadowResponse

Exceptions

InvalidRequestException, ResourceNotFoundException, ThrottlingException, UnauthorizedException, ServiceUnavailableException, InternalFailureException, MethodNotAllowedException or UnsupportedDocumentEncodingException.

See also: AWS API Documentation

AWSSDK.IotData.publish — Function.

using AWSSDK.IotData.publish
publish([::AWSConfig], arguments::Dict)
publish([::AWSConfig]; topic=, <keyword arguments>)

using AWSCore.Services.iot_data
iot_data([::AWSConfig], "POST", "/topics/{topic}", arguments::Dict)
iot_data([::AWSConfig], "POST", "/topics/{topic}", topic=, <keyword arguments>)

Publish Operation

Publishes state information.

For more information, see HTTP Protocol in the AWS IoT Developer Guide.

Arguments

topic = ::String – Required

The name of the MQTT topic.

qos = ::Int

The Quality of Service (QoS) level.

payload = blob

The state information, in JSON format.

Exceptions

InternalFailureException, InvalidRequestException, UnauthorizedException or MethodNotAllowedException.

See also: AWS API Documentation

AWSSDK.IotData.update_thing_shadow — Function.

using AWSSDK.IotData.update_thing_shadow
update_thing_shadow([::AWSConfig], arguments::Dict)
update_thing_shadow([::AWSConfig]; thingName=, payload=)

using AWSCore.Services.iot_data
iot_data([::AWSConfig], "POST", "/things/{thingName}/shadow", arguments::Dict)
iot_data([::AWSConfig], "POST", "/things/{thingName}/shadow", thingName=, payload=)

UpdateThingShadow Operation

Updates the thing shadow for the specified thing.

For more information, see UpdateThingShadow in the AWS IoT Developer Guide.

Arguments

thingName = ::String – Required

The name of the thing.

payload = blob – Required

The state information, in JSON format.

Returns

UpdateThingShadowResponse

Exceptions

ConflictException, RequestEntityTooLargeException, InvalidRequestException, ThrottlingException, UnauthorizedException, ServiceUnavailableException, InternalFailureException, MethodNotAllowedException or UnsupportedDocumentEncodingException.

See also: AWS API Documentation