AWSSDK.KinesisAnalytics.jl

AWSSDK.KinesisAnalytics

This document is generated from apis/kinesisanalytics-2015-08-14.normal.json. See JuliaCloud/AWSCore.jl.

AWSSDK.KinesisAnalytics.add_application_cloud_watch_logging_option

AWSSDK.KinesisAnalytics.add_application_input

AWSSDK.KinesisAnalytics.add_application_output

AWSSDK.KinesisAnalytics.add_application_reference_data_source

AWSSDK.KinesisAnalytics.create_application

AWSSDK.KinesisAnalytics.delete_application

AWSSDK.KinesisAnalytics.delete_application_cloud_watch_logging_option

AWSSDK.KinesisAnalytics.delete_application_output

AWSSDK.KinesisAnalytics.delete_application_reference_data_source

AWSSDK.KinesisAnalytics.describe_application

AWSSDK.KinesisAnalytics.discover_input_schema

AWSSDK.KinesisAnalytics.list_applications

AWSSDK.KinesisAnalytics.start_application

AWSSDK.KinesisAnalytics.stop_application

AWSSDK.KinesisAnalytics.update_application

AWSSDK.KinesisAnalytics.add_application_cloud_watch_logging_option — Function.

using AWSSDK.KinesisAnalytics.add_application_cloud_watch_logging_option
add_application_cloud_watch_logging_option([::AWSConfig], arguments::Dict)
add_application_cloud_watch_logging_option([::AWSConfig]; ApplicationName=, CurrentApplicationVersionId=, CloudWatchLoggingOption=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "AddApplicationCloudWatchLoggingOption", arguments::Dict)
kinesisanalytics([::AWSConfig], "AddApplicationCloudWatchLoggingOption", ApplicationName=, CurrentApplicationVersionId=, CloudWatchLoggingOption=)

AddApplicationCloudWatchLoggingOption Operation

Adds a CloudWatch log stream to monitor application configuration errors. For more information about using CloudWatch log streams with Amazon Kinesis Analytics applications, see Working with Amazon CloudWatch Logs.

Arguments

ApplicationName = ::String – Required

The Kinesis Analytics application name.

CurrentApplicationVersionId = ::Int – Required

The version ID of the Kinesis Analytics application.

CloudWatchLoggingOption = [ ... ] – Required

Provides the CloudWatch log stream Amazon Resource Name (ARN) and the IAM role ARN. Note: To write application messages to CloudWatch, the IAM role that is used must have the PutLogEvents policy action enabled.

 CloudWatchLoggingOption = [
        "LogStreamARN" => <required> ::String,
        "RoleARN" => <required> ::String
    ]

Returns

AddApplicationCloudWatchLoggingOptionResponse

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or ConcurrentModificationException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.add_application_input — Function.

using AWSSDK.KinesisAnalytics.add_application_input
add_application_input([::AWSConfig], arguments::Dict)
add_application_input([::AWSConfig]; ApplicationName=, CurrentApplicationVersionId=, Input=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "AddApplicationInput", arguments::Dict)
kinesisanalytics([::AWSConfig], "AddApplicationInput", ApplicationName=, CurrentApplicationVersionId=, Input=)

AddApplicationInput Operation

Adds a streaming source to your Amazon Kinesis application. For conceptual information, see Configuring Application Input.

You can add a streaming source either when you create an application or you can use this operation to add a streaming source after you create an application. For more information, see CreateApplication.

Any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

This operation requires permissions to perform the kinesisanalytics:AddApplicationInput action.

Arguments

ApplicationName = ::String – Required

Name of your existing Amazon Kinesis Analytics application to which you want to add the streaming source.

CurrentApplicationVersionId = ::Int – Required

Current version of your Amazon Kinesis Analytics application. You can use the DescribeApplication operation to find the current application version.

Input = [ ... ] – Required

 Input = [
        "NamePrefix" => <required> ::String,
        "KinesisStreamsInput" =>  [
            "ResourceARN" => <required> ::String,
            "RoleARN" => <required> ::String
        ],
        "KinesisFirehoseInput" =>  [
            "ResourceARN" => <required> ::String,
            "RoleARN" => <required> ::String
        ],
        "InputParallelism" =>  ["Count" =>  ::Int],
        "InputSchema" => <required> [
            "RecordFormat" => <required> [
                "RecordFormatType" => <required> "JSON" or "CSV",
                "MappingParameters" =>  [
                    "JSONMappingParameters" =>  ["RecordRowPath" => <required> ::String],
                    "CSVMappingParameters" =>  [
                        "RecordRowDelimiter" => <required> ::String,
                        "RecordColumnDelimiter" => <required> ::String
                    ]
                ]
            ],
            "RecordEncoding" =>  ::String,
            "RecordColumns" => <required> [[
                "Name" => <required> ::String,
                "Mapping" =>  ::String,
                "SqlType" => <required> ::String
            ], ...]
        ]
    ]

Returns

AddApplicationInputResponse

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException, ConcurrentModificationException or CodeValidationException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.add_application_output — Function.

using AWSSDK.KinesisAnalytics.add_application_output
add_application_output([::AWSConfig], arguments::Dict)
add_application_output([::AWSConfig]; ApplicationName=, CurrentApplicationVersionId=, Output=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "AddApplicationOutput", arguments::Dict)
kinesisanalytics([::AWSConfig], "AddApplicationOutput", ApplicationName=, CurrentApplicationVersionId=, Output=)

AddApplicationOutput Operation

Adds an external destination to your Amazon Kinesis Analytics application.

If you want Amazon Kinesis Analytics to deliver data from an in-application stream within your application to an external destination (such as an Amazon Kinesis stream or a Firehose delivery stream), you add the relevant configuration to your application using this operation. You can configure one or more outputs for your application. Each output configuration maps an in-application stream and an external destination.

You can use one of the output configurations to deliver data from your in-application error stream to an external destination so that you can analyze the errors. For conceptual information, see Understanding Application Output (Destination).

Note that any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

For the limits on the number of application inputs and outputs you can configure, see Limits.

This operation requires permissions to perform the kinesisanalytics:AddApplicationOutput action.

Arguments

ApplicationName = ::String – Required

Name of the application to which you want to add the output configuration.

CurrentApplicationVersionId = ::Int – Required

Version of the application to which you want add the output configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

Output = [ ... ] – Required

An array of objects, each describing one output configuration. In the output configuration, you specify the name of an in-application stream, a destination (that is, an Amazon Kinesis stream or an Amazon Kinesis Firehose delivery stream), and record the formation to use when writing to the destination.

 Output = [
        "Name" => <required> ::String,
        "KinesisStreamsOutput" =>  [
            "ResourceARN" => <required> ::String,
            "RoleARN" => <required> ::String
        ],
        "KinesisFirehoseOutput" =>  [
            "ResourceARN" => <required> ::String,
            "RoleARN" => <required> ::String
        ],
        "DestinationSchema" => <required> ["RecordFormatType" =>  "JSON" or "CSV"]
    ]

Returns

AddApplicationOutputResponse

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or ConcurrentModificationException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.add_application_reference_data_source — Function.

using AWSSDK.KinesisAnalytics.add_application_reference_data_source
add_application_reference_data_source([::AWSConfig], arguments::Dict)
add_application_reference_data_source([::AWSConfig]; ApplicationName=, CurrentApplicationVersionId=, ReferenceDataSource=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "AddApplicationReferenceDataSource", arguments::Dict)
kinesisanalytics([::AWSConfig], "AddApplicationReferenceDataSource", ApplicationName=, CurrentApplicationVersionId=, ReferenceDataSource=)

AddApplicationReferenceDataSource Operation

Adds a reference data source to an existing application.

Amazon Kinesis Analytics reads reference data (that is, an Amazon S3 object) and creates an in-application table within your application. In the request, you provide the source (S3 bucket name and object key name), name of the in-application table to create, and the necessary mapping information that describes how data in Amazon S3 object maps to columns in the resulting in-application table.

For conceptual information, see Configuring Application Input. For the limits on data sources you can add to your application, see Limits.

This operation requires permissions to perform the kinesisanalytics:AddApplicationOutput action.

Arguments

ApplicationName = ::String – Required

Name of an existing application.

CurrentApplicationVersionId = ::Int – Required

Version of the application for which you are adding the reference data source. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

ReferenceDataSource = [ ... ] – Required

The reference data source can be an object in your Amazon S3 bucket. Amazon Kinesis Analytics reads the object and copies the data into the in-application table that is created. You provide an S3 bucket, object key name, and the resulting in-application table that is created. You must also provide an IAM role with the necessary permissions that Amazon Kinesis Analytics can assume to read the object from your S3 bucket on your behalf.

 ReferenceDataSource = [
        "TableName" => <required> ::String,
        "S3ReferenceDataSource" =>  [
            "BucketARN" => <required> ::String,
            "FileKey" => <required> ::String,
            "ReferenceRoleARN" => <required> ::String
        ],
        "ReferenceSchema" => <required> [
            "RecordFormat" => <required> [
                "RecordFormatType" => <required> "JSON" or "CSV",
                "MappingParameters" =>  [
                    "JSONMappingParameters" =>  ["RecordRowPath" => <required> ::String],
                    "CSVMappingParameters" =>  [
                        "RecordRowDelimiter" => <required> ::String,
                        "RecordColumnDelimiter" => <required> ::String
                    ]
                ]
            ],
            "RecordEncoding" =>  ::String,
            "RecordColumns" => <required> [[
                "Name" => <required> ::String,
                "Mapping" =>  ::String,
                "SqlType" => <required> ::String
            ], ...]
        ]
    ]

Returns

AddApplicationReferenceDataSourceResponse

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or ConcurrentModificationException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.create_application — Function.

using AWSSDK.KinesisAnalytics.create_application
create_application([::AWSConfig], arguments::Dict)
create_application([::AWSConfig]; ApplicationName=, <keyword arguments>)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "CreateApplication", arguments::Dict)
kinesisanalytics([::AWSConfig], "CreateApplication", ApplicationName=, <keyword arguments>)

CreateApplication Operation

Creates an Amazon Kinesis Analytics application. You can configure each application with one streaming source as input, application code to process the input, and up to five streaming destinations where you want Amazon Kinesis Analytics to write the output data from your application. For an overview, see How it Works.

In the input configuration, you map the streaming source to an in-application stream, which you can think of as a constantly updating table. In the mapping, you must provide a schema for the in-application stream and map each data column in the in-application stream to a data element in the streaming source.

Your application code is one or more SQL statements that read input data, transform it, and generate output. Your application code can create one or more SQL artifacts like SQL streams or pumps.

In the output configuration, you can configure the application to write data from in-application streams created in your applications to up to five streaming destinations.

To read data from your source stream or write data to destination streams, Amazon Kinesis Analytics needs your permissions. You grant these permissions by creating IAM roles. This operation requires permissions to perform the kinesisanalytics:CreateApplication action.

For introductory exercises to create an Amazon Kinesis Analytics application, see Getting Started.

Arguments

ApplicationName = ::String – Required

Name of your Amazon Kinesis Analytics application (for example, sample-app).

ApplicationDescription = ::String

Summary description of the application.

Inputs = [[ ... ], ...]

Use this parameter to configure the application input.

You can configure your application to receive input from a single streaming source. In this configuration, you map this streaming source to an in-application stream that is created. Your application code can then query the in-application stream like a table (you can think of it as a constantly updating table).

For the streaming source, you provide its Amazon Resource Name (ARN) and format of data on the stream (for example, JSON, CSV, etc). You also must provide an IAM role that Amazon Kinesis Analytics can assume to read this stream on your behalf.

To create the in-application stream, you need to specify a schema to transform your data into a schematized version used in SQL. In the schema, you provide the necessary mapping of the data elements in the streaming source to record columns in the in-app stream.

 Inputs = [[
        "NamePrefix" => <required> ::String,
        "KinesisStreamsInput" =>  [
            "ResourceARN" => <required> ::String,
            "RoleARN" => <required> ::String
        ],
        "KinesisFirehoseInput" =>  [
            "ResourceARN" => <required> ::String,
            "RoleARN" => <required> ::String
        ],
        "InputParallelism" =>  ["Count" =>  ::Int],
        "InputSchema" => <required> [
            "RecordFormat" => <required> [
                "RecordFormatType" => <required> "JSON" or "CSV",
                "MappingParameters" =>  [
                    "JSONMappingParameters" =>  ["RecordRowPath" => <required> ::String],
                    "CSVMappingParameters" =>  [
                        "RecordRowDelimiter" => <required> ::String,
                        "RecordColumnDelimiter" => <required> ::String
                    ]
                ]
            ],
            "RecordEncoding" =>  ::String,
            "RecordColumns" => <required> [[
                "Name" => <required> ::String,
                "Mapping" =>  ::String,
                "SqlType" => <required> ::String
            ], ...]
        ]
    ], ...]

Outputs = [[ ... ], ...]

You can configure application output to write data from any of the in-application streams to up to five destinations.

These destinations can be Amazon Kinesis streams, Amazon Kinesis Firehose delivery streams, or both.

In the configuration, you specify the in-application stream name, the destination stream Amazon Resource Name (ARN), and the format to use when writing data. You must also provide an IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf.

In the output configuration, you also provide the output stream Amazon Resource Name (ARN) and the format of data in the stream (for example, JSON, CSV). You also must provide an IAM role that Amazon Kinesis Analytics can assume to write to this stream on your behalf.

 Outputs = [[
        "Name" => <required> ::String,
        "KinesisStreamsOutput" =>  [
            "ResourceARN" => <required> ::String,
            "RoleARN" => <required> ::String
        ],
        "KinesisFirehoseOutput" =>  [
            "ResourceARN" => <required> ::String,
            "RoleARN" => <required> ::String
        ],
        "DestinationSchema" => <required> ["RecordFormatType" =>  "JSON" or "CSV"]
    ], ...]

CloudWatchLoggingOptions = [[ ... ], ...]

Use this parameter to configure a CloudWatch log stream to monitor application configuration errors. For more information, see Working with Amazon CloudWatch Logs.

 CloudWatchLoggingOptions = [[
        "LogStreamARN" => <required> ::String,
        "RoleARN" => <required> ::String
    ], ...]

ApplicationCode = ::String

One or more SQL statements that read input data, transform it, and generate output. For example, you can write a SQL statement that reads data from one in-application stream, generates a running average of the number of advertisement clicks by vendor, and insert resulting rows in another in-application stream using pumps. For more inforamtion about the typical pattern, see Application Code.

You can provide such series of SQL statements, where output of one statement can be used as the input for the next statement. You store intermediate results by creating in-application streams and pumps.

Note that the application code must create the streams with names specified in the Outputs. For example, if your Outputs defines output streams named ExampleOutputStream1 and ExampleOutputStream2, then your application code must create these streams.

Returns

CreateApplicationResponse

Exceptions

CodeValidationException, ResourceInUseException, LimitExceededException or InvalidArgumentException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.delete_application — Function.

using AWSSDK.KinesisAnalytics.delete_application
delete_application([::AWSConfig], arguments::Dict)
delete_application([::AWSConfig]; ApplicationName=, CreateTimestamp=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "DeleteApplication", arguments::Dict)
kinesisanalytics([::AWSConfig], "DeleteApplication", ApplicationName=, CreateTimestamp=)

DeleteApplication Operation

Deletes the specified application. Amazon Kinesis Analytics halts application execution and deletes the application, including any application artifacts (such as in-application streams, reference table, and application code).

This operation requires permissions to perform the kinesisanalytics:DeleteApplication action.

Arguments

ApplicationName = ::String – Required

Name of the Amazon Kinesis Analytics application to delete.

CreateTimestamp = timestamp – Required

You can use the DescribeApplication operation to get this value.

Returns

DeleteApplicationResponse

Exceptions

ConcurrentModificationException, ResourceNotFoundException or ResourceInUseException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.delete_application_cloud_watch_logging_option — Function.

using AWSSDK.KinesisAnalytics.delete_application_cloud_watch_logging_option
delete_application_cloud_watch_logging_option([::AWSConfig], arguments::Dict)
delete_application_cloud_watch_logging_option([::AWSConfig]; ApplicationName=, CurrentApplicationVersionId=, CloudWatchLoggingOptionId=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "DeleteApplicationCloudWatchLoggingOption", arguments::Dict)
kinesisanalytics([::AWSConfig], "DeleteApplicationCloudWatchLoggingOption", ApplicationName=, CurrentApplicationVersionId=, CloudWatchLoggingOptionId=)

DeleteApplicationCloudWatchLoggingOption Operation

Deletes a CloudWatch log stream from an application. For more information about using CloudWatch log streams with Amazon Kinesis Analytics applications, see Working with Amazon CloudWatch Logs.

Arguments

ApplicationName = ::String – Required

The Kinesis Analytics application name.

CurrentApplicationVersionId = ::Int – Required

The version ID of the Kinesis Analytics application.

CloudWatchLoggingOptionId = ::String – Required

The CloudWatchLoggingOptionId of the CloudWatch logging option to delete. You can use the DescribeApplication operation to get the CloudWatchLoggingOptionId.

Returns

DeleteApplicationCloudWatchLoggingOptionResponse

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or ConcurrentModificationException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.delete_application_output — Function.

using AWSSDK.KinesisAnalytics.delete_application_output
delete_application_output([::AWSConfig], arguments::Dict)
delete_application_output([::AWSConfig]; ApplicationName=, CurrentApplicationVersionId=, OutputId=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "DeleteApplicationOutput", arguments::Dict)
kinesisanalytics([::AWSConfig], "DeleteApplicationOutput", ApplicationName=, CurrentApplicationVersionId=, OutputId=)

DeleteApplicationOutput Operation

Deletes output destination configuration from your application configuration. Amazon Kinesis Analytics will no longer write data from the corresponding in-application stream to the external output destination.

This operation requires permissions to perform the kinesisanalytics:DeleteApplicationOutput action.

Arguments

ApplicationName = ::String – Required

Amazon Kinesis Analytics application name.

CurrentApplicationVersionId = ::Int – Required

Amazon Kinesis Analytics application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

OutputId = ::String – Required

The ID of the configuration to delete. Each output configuration that is added to the application, either when the application is created or later using the AddApplicationOutput operation, has a unique ID. You need to provide the ID to uniquely identify the output configuration that you want to delete from the application configuration. You can use the DescribeApplication operation to get the specific OutputId.

Returns

DeleteApplicationOutputResponse

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or ConcurrentModificationException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.delete_application_reference_data_source — Function.

using AWSSDK.KinesisAnalytics.delete_application_reference_data_source
delete_application_reference_data_source([::AWSConfig], arguments::Dict)
delete_application_reference_data_source([::AWSConfig]; ApplicationName=, CurrentApplicationVersionId=, ReferenceId=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "DeleteApplicationReferenceDataSource", arguments::Dict)
kinesisanalytics([::AWSConfig], "DeleteApplicationReferenceDataSource", ApplicationName=, CurrentApplicationVersionId=, ReferenceId=)

DeleteApplicationReferenceDataSource Operation

Deletes a reference data source configuration from the specified application configuration.

If the application is running, Amazon Kinesis Analytics immediately removes the in-application table that you created using the AddApplicationReferenceDataSource operation.

This operation requires permissions to perform the kinesisanalytics.DeleteApplicationReferenceDataSource action.

Arguments

ApplicationName = ::String – Required

Name of an existing application.

CurrentApplicationVersionId = ::Int – Required

Version of the application. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

ReferenceId = ::String – Required

ID of the reference data source. When you add a reference data source to your application using the AddApplicationReferenceDataSource, Amazon Kinesis Analytics assigns an ID. You can use the DescribeApplication operation to get the reference ID.

Returns

DeleteApplicationReferenceDataSourceResponse

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or ConcurrentModificationException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.describe_application — Function.

using AWSSDK.KinesisAnalytics.describe_application
describe_application([::AWSConfig], arguments::Dict)
describe_application([::AWSConfig]; ApplicationName=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "DescribeApplication", arguments::Dict)
kinesisanalytics([::AWSConfig], "DescribeApplication", ApplicationName=)

DescribeApplication Operation

Returns information about a specific Amazon Kinesis Analytics application.

If you want to retrieve a list of all applications in your account, use the ListApplications operation.

This operation requires permissions to perform the kinesisanalytics:DescribeApplication action. You can use DescribeApplication to get the current application versionId, which you need to call other operations such as Update.

Arguments

ApplicationName = ::String – Required

Name of the application.

Returns

DescribeApplicationResponse

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.discover_input_schema — Function.

using AWSSDK.KinesisAnalytics.discover_input_schema
discover_input_schema([::AWSConfig], arguments::Dict)
discover_input_schema([::AWSConfig]; ResourceARN=, RoleARN=, InputStartingPositionConfiguration=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "DiscoverInputSchema", arguments::Dict)
kinesisanalytics([::AWSConfig], "DiscoverInputSchema", ResourceARN=, RoleARN=, InputStartingPositionConfiguration=)

DiscoverInputSchema Operation

Infers a schema by evaluating sample records on the specified streaming source (Amazon Kinesis stream or Amazon Kinesis Firehose delivery stream). In the response, the operation returns the inferred schema and also the sample records that the operation used to infer the schema.

You can use the inferred schema when configuring a streaming source for your application. For conceptual information, see Configuring Application Input. Note that when you create an application using the Amazon Kinesis Analytics console, the console uses this operation to infer a schema and show it in the console user interface.

This operation requires permissions to perform the kinesisanalytics:DiscoverInputSchema action.

Arguments

ResourceARN = ::String – Required

Amazon Resource Name (ARN) of the streaming source.

RoleARN = ::String – Required

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf.

InputStartingPositionConfiguration = ["InputStartingPosition" => "NOW", "TRIM_HORIZON" or "LAST_STOPPED_POINT"] – Required

Point at which you want Amazon Kinesis Analytics to start reading records from the specified streaming source discovery purposes.

Returns

DiscoverInputSchemaResponse

Exceptions

InvalidArgumentException, UnableToDetectSchemaException, ResourceProvisionedThroughputExceededException or ServiceUnavailableException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.list_applications — Function.

using AWSSDK.KinesisAnalytics.list_applications
list_applications([::AWSConfig], arguments::Dict)
list_applications([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "ListApplications", arguments::Dict)
kinesisanalytics([::AWSConfig], "ListApplications", <keyword arguments>)

ListApplications Operation

Returns a list of Amazon Kinesis Analytics applications in your account. For each application, the response includes the application name, Amazon Resource Name (ARN), and status. If the response returns the HasMoreApplications value as true, you can send another request by adding the ExclusiveStartApplicationName in the request body, and set the value of this to the last application name from the previous response.

If you want detailed information about a specific application, use DescribeApplication.

This operation requires permissions to perform the kinesisanalytics:ListApplications action.

Arguments

Limit = ::Int

Maximum number of applications to list.

ExclusiveStartApplicationName = ::String

Name of the application to start the list with. When using pagination to retrieve the list, you don't need to specify this parameter in the first request. However, in subsequent requests, you add the last application name from the previous response to get the next page of applications.

Returns

ListApplicationsResponse

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.start_application — Function.

using AWSSDK.KinesisAnalytics.start_application
start_application([::AWSConfig], arguments::Dict)
start_application([::AWSConfig]; ApplicationName=, InputConfigurations=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "StartApplication", arguments::Dict)
kinesisanalytics([::AWSConfig], "StartApplication", ApplicationName=, InputConfigurations=)

StartApplication Operation

Starts the specified Amazon Kinesis Analytics application. After creating an application, you must exclusively call this operation to start your application.

After the application starts, it begins consuming the input data, processes it, and writes the output to the configured destination.

The application status must be READY for you to start an application. You can get the application status in the console or using the DescribeApplication operation.

After you start the application, you can stop the application from processing the input by calling the StopApplication operation.

This operation requires permissions to perform the kinesisanalytics:StartApplication action.

Arguments

ApplicationName = ::String – Required

Name of the application.

InputConfigurations = [[ ... ], ...] – Required

Identifies the specific input, by ID, that the application starts consuming. Amazon Kinesis Analytics starts reading the streaming source associated with the input. You can also specify where in the streaming source you want Amazon Kinesis Analytics to start reading.

 InputConfigurations = [[
        "Id" => <required> ::String,
        "InputStartingPositionConfiguration" => <required> ["InputStartingPosition" =>  "NOW", "TRIM_HORIZON" or "LAST_STOPPED_POINT"]
    ], ...]

Returns

StartApplicationResponse

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or InvalidApplicationConfigurationException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.stop_application — Function.

using AWSSDK.KinesisAnalytics.stop_application
stop_application([::AWSConfig], arguments::Dict)
stop_application([::AWSConfig]; ApplicationName=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "StopApplication", arguments::Dict)
kinesisanalytics([::AWSConfig], "StopApplication", ApplicationName=)

StopApplication Operation

Stops the application from processing input data. You can stop an application only if it is in the running state. You can use the DescribeApplication operation to find the application state. After the application is stopped, Amazon Kinesis Analytics stops reading data from the input, the application stops processing data, and there is no output written to the destination.

This operation requires permissions to perform the kinesisanalytics:StopApplication action.

Arguments

ApplicationName = ::String – Required

Name of the running application to stop.

Returns

StopApplicationResponse

Exceptions

ResourceNotFoundException or ResourceInUseException.

See also: AWS API Documentation

AWSSDK.KinesisAnalytics.update_application — Function.

using AWSSDK.KinesisAnalytics.update_application
update_application([::AWSConfig], arguments::Dict)
update_application([::AWSConfig]; ApplicationName=, CurrentApplicationVersionId=, ApplicationUpdate=)

using AWSCore.Services.kinesisanalytics
kinesisanalytics([::AWSConfig], "UpdateApplication", arguments::Dict)
kinesisanalytics([::AWSConfig], "UpdateApplication", ApplicationName=, CurrentApplicationVersionId=, ApplicationUpdate=)

UpdateApplication Operation

Updates an existing Amazon Kinesis Analytics application. Using this API, you can update application code, input configuration, and output configuration.

Note that Amazon Kinesis Analytics updates the CurrentApplicationVersionId each time you update your application.

This operation requires permission for the kinesisanalytics:UpdateApplication action.

Arguments

ApplicationName = ::String – Required

Name of the Amazon Kinesis Analytics application to update.

CurrentApplicationVersionId = ::Int – Required

The current application version ID. You can use the DescribeApplication operation to get this value.

ApplicationUpdate = [ ... ] – Required

Describes application updates.

 ApplicationUpdate = [
        "InputUpdates" =>  [[
            "InputId" => <required> ::String,
            "NamePrefixUpdate" =>  ::String,
            "KinesisStreamsInputUpdate" =>  [
                "ResourceARNUpdate" =>  ::String,
                "RoleARNUpdate" =>  ::String
            ],
            "KinesisFirehoseInputUpdate" =>  [
                "ResourceARNUpdate" =>  ::String,
                "RoleARNUpdate" =>  ::String
            ],
            "InputSchemaUpdate" =>  [
                "RecordFormatUpdate" =>  [
                    "RecordFormatType" => <required> "JSON" or "CSV",
                    "MappingParameters" =>  [
                        "JSONMappingParameters" =>  ["RecordRowPath" => <required> ::String],
                        "CSVMappingParameters" =>  [
                            "RecordRowDelimiter" => <required> ::String,
                            "RecordColumnDelimiter" => <required> ::String
                        ]
                    ]
                ],
                "RecordEncodingUpdate" =>  ::String,
                "RecordColumnUpdates" =>  [[
                    "Name" => <required> ::String,
                    "Mapping" =>  ::String,
                    "SqlType" => <required> ::String
                ], ...]
            ],
            "InputParallelismUpdate" =>  ["CountUpdate" =>  ::Int]
        ], ...],
        "ApplicationCodeUpdate" =>  ::String,
        "OutputUpdates" =>  [[
            "OutputId" => <required> ::String,
            "NameUpdate" =>  ::String,
            "KinesisStreamsOutputUpdate" =>  [
                "ResourceARNUpdate" =>  ::String,
                "RoleARNUpdate" =>  ::String
            ],
            "KinesisFirehoseOutputUpdate" =>  [
                "ResourceARNUpdate" =>  ::String,
                "RoleARNUpdate" =>  ::String
            ],
            "DestinationSchemaUpdate" =>  ["RecordFormatType" =>  "JSON" or "CSV"]
        ], ...],
        "ReferenceDataSourceUpdates" =>  [[
            "ReferenceId" => <required> ::String,
            "TableNameUpdate" =>  ::String,
            "S3ReferenceDataSourceUpdate" =>  [
                "BucketARNUpdate" =>  ::String,
                "FileKeyUpdate" =>  ::String,
                "ReferenceRoleARNUpdate" =>  ::String
            ],
            "ReferenceSchemaUpdate" =>  [
                "RecordFormat" => <required> [
                    "RecordFormatType" => <required> "JSON" or "CSV",
                    "MappingParameters" =>  [
                        "JSONMappingParameters" =>  ["RecordRowPath" => <required> ::String],
                        "CSVMappingParameters" =>  [
                            "RecordRowDelimiter" => <required> ::String,
                            "RecordColumnDelimiter" => <required> ::String
                        ]
                    ]
                ],
                "RecordEncoding" =>  ::String,
                "RecordColumns" => <required> [[
                    "Name" => <required> ::String,
                    "Mapping" =>  ::String,
                    "SqlType" => <required> ::String
                ], ...]
            ]
        ], ...],
        "CloudWatchLoggingOptionUpdates" =>  [[
            "CloudWatchLoggingOptionId" => <required> ::String,
            "LogStreamARNUpdate" =>  ::String,
            "RoleARNUpdate" =>  ::String
        ], ...]
    ]

Returns

UpdateApplicationResponse

Exceptions

CodeValidationException, ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or ConcurrentModificationException.

See also: AWS API Documentation