AWSSDK.MarketplaceEntitlementService.jl

AWSSDK.MarketplaceEntitlementService

AWS Marketplace Entitlement Service

This reference provides descriptions of the AWS Marketplace Entitlement Service API.

AWS Marketplace Entitlement Service is used to determine the entitlement of a customer to a given product. An entitlement represents capacity in a product owned by the customer. For example, a customer might own some number of users or seats in an SaaS application or some amount of data capacity in a multi-tenant database.

Getting Entitlement Records

GetEntitlements- Gets the entitlements for a Marketplace product.

This document is generated from apis/entitlement.marketplace-2017-01-11.normal.json. See JuliaCloud/AWSCore.jl.

AWSSDK.MarketplaceEntitlementService.get_entitlements

AWSSDK.MarketplaceEntitlementService.get_entitlements — Function.

using AWSSDK.MarketplaceEntitlementService.get_entitlements
get_entitlements([::AWSConfig], arguments::Dict)
get_entitlements([::AWSConfig]; ProductCode=, <keyword arguments>)

using AWSCore.Services.entitlement_marketplace
entitlement_marketplace([::AWSConfig], "GetEntitlements", arguments::Dict)
entitlement_marketplace([::AWSConfig], "GetEntitlements", ProductCode=, <keyword arguments>)

GetEntitlements Operation

GetEntitlements retrieves entitlement values for a given product. The results can be filtered based on customer identifier or product dimensions.

Arguments

ProductCode = ::String – Required

Product code is used to uniquely identify a product in AWS Marketplace. The product code will be provided by AWS Marketplace when the product listing is created.

Filter = ::Dict{String,String}

Filter is used to return entitlements for a specific customer or for a specific dimension. Filters are described as keys mapped to a lists of values. Filtered requests are unioned for each value in the value list, and then intersected for each filter key.

NextToken = ::String

For paginated calls to GetEntitlements, pass the NextToken from the previous GetEntitlementsResult.

MaxResults = ::Int

The maximum number of items to retrieve from the GetEntitlements operation. For pagination, use the NextToken field in subsequent calls to GetEntitlements.

Returns

GetEntitlementsResult

Exceptions

InvalidParameterException, ThrottlingException or InternalServiceErrorException.

See also: AWS API Documentation