AWSSDK.Mobile.jl

AWSSDK.Mobile

AWS Mobile Service provides mobile app and website developers with capabilities required to configure AWS resources and bootstrap their developer desktop projects with the necessary SDKs, constants, tools and samples to make use of those resources.

This document is generated from apis/mobile-2017-07-01.normal.json. See JuliaCloud/AWSCore.jl.

AWSSDK.Mobile.create_project

AWSSDK.Mobile.delete_project

AWSSDK.Mobile.describe_bundle

AWSSDK.Mobile.describe_project

AWSSDK.Mobile.export_bundle

AWSSDK.Mobile.export_project

AWSSDK.Mobile.list_bundles

AWSSDK.Mobile.list_projects

AWSSDK.Mobile.update_project

AWSSDK.Mobile.create_project — Function.

using AWSSDK.Mobile.create_project
create_project([::AWSConfig], arguments::Dict)
create_project([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.mobile
mobile([::AWSConfig], "POST", "/projects", arguments::Dict)
mobile([::AWSConfig], "POST", "/projects", <keyword arguments>)

CreateProject Operation

Creates an AWS Mobile Hub project.

Arguments

name = ::String

Name of the project.

region = ::String

Default region where project resources should be created.

contents = blob

ZIP or YAML file which contains configuration settings to be used when creating the project. This may be the contents of the file downloaded from the URL provided in an export project operation.

snapshotId = ::String

Unique identifier for an exported snapshot of project configuration. This snapshot identifier is included in the share URL when a project is exported.

Returns

CreateProjectResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException, BadRequestException, NotFoundException or LimitExceededException.

See also: AWS API Documentation

AWSSDK.Mobile.delete_project — Function.

using AWSSDK.Mobile.delete_project
delete_project([::AWSConfig], arguments::Dict)
delete_project([::AWSConfig]; projectId=)

using AWSCore.Services.mobile
mobile([::AWSConfig], "DELETE", "/projects/{projectId}", arguments::Dict)
mobile([::AWSConfig], "DELETE", "/projects/{projectId}", projectId=)

DeleteProject Operation

Delets a project in AWS Mobile Hub.

Arguments

projectId = ::String – Required

Unique project identifier.

Returns

DeleteProjectResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException or NotFoundException.

See also: AWS API Documentation

AWSSDK.Mobile.describe_bundle — Function.

using AWSSDK.Mobile.describe_bundle
describe_bundle([::AWSConfig], arguments::Dict)
describe_bundle([::AWSConfig]; bundleId=)

using AWSCore.Services.mobile
mobile([::AWSConfig], "GET", "/bundles/{bundleId}", arguments::Dict)
mobile([::AWSConfig], "GET", "/bundles/{bundleId}", bundleId=)

DescribeBundle Operation

Get the bundle details for the requested bundle id.

Arguments

bundleId = ::String – Required

Unique bundle identifier.

Returns

DescribeBundleResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException, BadRequestException or NotFoundException.

See also: AWS API Documentation

AWSSDK.Mobile.describe_project — Function.

using AWSSDK.Mobile.describe_project
describe_project([::AWSConfig], arguments::Dict)
describe_project([::AWSConfig]; projectId=, <keyword arguments>)

using AWSCore.Services.mobile
mobile([::AWSConfig], "GET", "/project", arguments::Dict)
mobile([::AWSConfig], "GET", "/project", projectId=, <keyword arguments>)

DescribeProject Operation

Gets details about a project in AWS Mobile Hub.

Arguments

projectId = ::String – Required

Unique project identifier.

syncFromResources = ::Bool

If set to true, causes AWS Mobile Hub to synchronize information from other services, e.g., update state of AWS CloudFormation stacks in the AWS Mobile Hub project.

Returns

DescribeProjectResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException, BadRequestException or NotFoundException.

See also: AWS API Documentation

AWSSDK.Mobile.export_bundle — Function.

using AWSSDK.Mobile.export_bundle
export_bundle([::AWSConfig], arguments::Dict)
export_bundle([::AWSConfig]; bundleId=, <keyword arguments>)

using AWSCore.Services.mobile
mobile([::AWSConfig], "POST", "/bundles/{bundleId}", arguments::Dict)
mobile([::AWSConfig], "POST", "/bundles/{bundleId}", bundleId=, <keyword arguments>)

ExportBundle Operation

Generates customized software development kit (SDK) and or tool packages used to integrate mobile web or mobile app clients with backend AWS resources.

Arguments

bundleId = ::String – Required

Unique bundle identifier.

projectId = ::String

Unique project identifier.

platform = "OSX", "WINDOWS", "LINUX", "OBJC", "SWIFT", "ANDROID" or "JAVASCRIPT"

Developer desktop or target application platform.

Returns

ExportBundleResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException, BadRequestException or NotFoundException.

See also: AWS API Documentation

AWSSDK.Mobile.export_project — Function.

using AWSSDK.Mobile.export_project
export_project([::AWSConfig], arguments::Dict)
export_project([::AWSConfig]; projectId=)

using AWSCore.Services.mobile
mobile([::AWSConfig], "POST", "/exports/{projectId}", arguments::Dict)
mobile([::AWSConfig], "POST", "/exports/{projectId}", projectId=)

ExportProject Operation

Exports project configuration to a snapshot which can be downloaded and shared. Note that mobile app push credentials are encrypted in exported projects, so they can only be shared successfully within the same AWS account.

Arguments

projectId = ::String – Required

Unique project identifier.

Returns

ExportProjectResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException, BadRequestException or NotFoundException.

See also: AWS API Documentation

AWSSDK.Mobile.list_bundles — Function.

using AWSSDK.Mobile.list_bundles
list_bundles([::AWSConfig], arguments::Dict)
list_bundles([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.mobile
mobile([::AWSConfig], "GET", "/bundles", arguments::Dict)
mobile([::AWSConfig], "GET", "/bundles", <keyword arguments>)

ListBundles Operation

List all available bundles.

Arguments

maxResults = ::Int

Maximum number of records to list in a single response.

nextToken = ::String

Pagination token. Set to null to start listing bundles from start. If non-null pagination token is returned in a result, then pass its value in here in another request to list more bundles.

Returns

ListBundlesResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException or BadRequestException.

See also: AWS API Documentation

AWSSDK.Mobile.list_projects — Function.

using AWSSDK.Mobile.list_projects
list_projects([::AWSConfig], arguments::Dict)
list_projects([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.mobile
mobile([::AWSConfig], "GET", "/projects", arguments::Dict)
mobile([::AWSConfig], "GET", "/projects", <keyword arguments>)

ListProjects Operation

Lists projects in AWS Mobile Hub.

Arguments

maxResults = ::Int

Maximum number of records to list in a single response.

nextToken = ::String

Pagination token. Set to null to start listing projects from start. If non-null pagination token is returned in a result, then pass its value in here in another request to list more projects.

Returns

ListProjectsResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException or BadRequestException.

See also: AWS API Documentation

AWSSDK.Mobile.update_project — Function.

using AWSSDK.Mobile.update_project
update_project([::AWSConfig], arguments::Dict)
update_project([::AWSConfig]; projectId=, <keyword arguments>)

using AWSCore.Services.mobile
mobile([::AWSConfig], "POST", "/update", arguments::Dict)
mobile([::AWSConfig], "POST", "/update", projectId=, <keyword arguments>)

UpdateProject Operation

Update an existing project.

Arguments

contents = blob

ZIP or YAML file which contains project configuration to be updated. This should be the contents of the file downloaded from the URL provided in an export project operation.

projectId = ::String – Required

Unique project identifier.

Returns

UpdateProjectResult

Exceptions

InternalFailureException, ServiceUnavailableException, UnauthorizedException, TooManyRequestsException, BadRequestException, NotFoundException, AccountActionRequiredException or LimitExceededException.

See also: AWS API Documentation