AWSSDK.StepFunctions.jl

AWSSDK.StepFunctions

AWS Step Functions

AWS Step Functions is a service that lets you coordinate the components of distributed applications and microservices using visual workflows.

You can use Step Functions to build applications from individual components, each of which performs a discrete function, or task, allowing you to scale and change applications quickly. Step Functions provides a console that helps visualize the components of your application as a series of steps. Step Functions automatically triggers and tracks each step, and retries steps when there are errors, so your application executes in order and as expected, every time. Step Functions logs the state of each step, so you can diagnose and debug problems quickly.

Step Functions manages operations and underlying infrastructure to ensure your application is available at any scale. You can run tasks on AWS, your own servers, or any system that has access to AWS. You can access and use Step Functions using the console, the AWS SDKs, or an HTTP API. For more information about Step Functions, see the AWS Step Functions Developer Guide .

This document is generated from apis/states-2016-11-23.normal.json. See JuliaCloud/AWSCore.jl.

AWSSDK.StepFunctions.create_activity

AWSSDK.StepFunctions.create_state_machine

AWSSDK.StepFunctions.delete_activity

AWSSDK.StepFunctions.delete_state_machine

AWSSDK.StepFunctions.describe_activity

AWSSDK.StepFunctions.describe_execution

AWSSDK.StepFunctions.describe_state_machine

AWSSDK.StepFunctions.get_activity_task

AWSSDK.StepFunctions.get_execution_history

AWSSDK.StepFunctions.list_activities

AWSSDK.StepFunctions.list_executions

AWSSDK.StepFunctions.list_state_machines

AWSSDK.StepFunctions.send_task_failure

AWSSDK.StepFunctions.send_task_heartbeat

AWSSDK.StepFunctions.send_task_success

AWSSDK.StepFunctions.start_execution

AWSSDK.StepFunctions.stop_execution

AWSSDK.StepFunctions.create_activity — Function.

using AWSSDK.StepFunctions.create_activity
create_activity([::AWSConfig], arguments::Dict)
create_activity([::AWSConfig]; name=)

using AWSCore.Services.states
states([::AWSConfig], "CreateActivity", arguments::Dict)
states([::AWSConfig], "CreateActivity", name=)

CreateActivity Operation

Creates an activity. An Activity is a task which you write, in any language and hosted on any machine which has access to AWS Step Functions. Activities must poll Step Functions using the GetActivityTask and respond using SendTask* API calls. This function lets Step Functions know the existence of your activity and returns an identifier for use in a state machine and when polling from the activity.

Arguments

name = ::String – Required

The name of the activity to create. This name must be unique for your AWS account and region for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

whitespace

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ $ & , ; : /`

control characters (U+0000-001F, U+007F-009F)

Returns

CreateActivityOutput

Exceptions

ActivityLimitExceeded or InvalidName.

See also: AWS API Documentation

AWSSDK.StepFunctions.create_state_machine — Function.

using AWSSDK.StepFunctions.create_state_machine
create_state_machine([::AWSConfig], arguments::Dict)
create_state_machine([::AWSConfig]; name=, definition=, roleArn=)

using AWSCore.Services.states
states([::AWSConfig], "CreateStateMachine", arguments::Dict)
states([::AWSConfig], "CreateStateMachine", name=, definition=, roleArn=)

CreateStateMachine Operation

Creates a state machine. A state machine consists of a collection of states that can do work (Task states), determine which states to transition to next (Choice states), stop an execution with an error (Fail states), and so on. State machines are specified using a JSON-based, structured language.

Arguments

name = ::String – Required

The name of the state machine. This name must be unique for your AWS account and region for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

whitespace

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ $ & , ; : /`

control characters (U+0000-001F, U+007F-009F)

definition = ::String – Required

The Amazon States Language definition of the state machine.

roleArn = ::String – Required

The Amazon Resource Name (ARN) of the IAM role to use for this state machine.

Returns

CreateStateMachineOutput

Exceptions

InvalidArn, InvalidDefinition, InvalidName, StateMachineAlreadyExists, StateMachineDeleting or StateMachineLimitExceeded.

See also: AWS API Documentation

AWSSDK.StepFunctions.delete_activity — Function.

using AWSSDK.StepFunctions.delete_activity
delete_activity([::AWSConfig], arguments::Dict)
delete_activity([::AWSConfig]; activityArn=)

using AWSCore.Services.states
states([::AWSConfig], "DeleteActivity", arguments::Dict)
states([::AWSConfig], "DeleteActivity", activityArn=)

DeleteActivity Operation

Deletes an activity.

Arguments

activityArn = ::String – Required

The Amazon Resource Name (ARN) of the activity to delete.

Returns

DeleteActivityOutput

Exceptions

InvalidArn.

See also: AWS API Documentation

AWSSDK.StepFunctions.delete_state_machine — Function.

using AWSSDK.StepFunctions.delete_state_machine
delete_state_machine([::AWSConfig], arguments::Dict)
delete_state_machine([::AWSConfig]; stateMachineArn=)

using AWSCore.Services.states
states([::AWSConfig], "DeleteStateMachine", arguments::Dict)
states([::AWSConfig], "DeleteStateMachine", stateMachineArn=)

DeleteStateMachine Operation

Deletes a state machine. This is an asynchronous operation– it sets the state machine's status to "DELETING" and begins the delete process. Each state machine execution will be deleted the next time it makes a state transition. After all executions have completed or been deleted, the state machine itself will be deleted.

Arguments

stateMachineArn = ::String – Required

The Amazon Resource Name (ARN) of the state machine to delete.

Returns

DeleteStateMachineOutput

Exceptions

InvalidArn.

See also: AWS API Documentation

AWSSDK.StepFunctions.describe_activity — Function.

using AWSSDK.StepFunctions.describe_activity
describe_activity([::AWSConfig], arguments::Dict)
describe_activity([::AWSConfig]; activityArn=)

using AWSCore.Services.states
states([::AWSConfig], "DescribeActivity", arguments::Dict)
states([::AWSConfig], "DescribeActivity", activityArn=)

DescribeActivity Operation

Describes an activity.

Arguments

activityArn = ::String – Required

The Amazon Resource Name (ARN) of the activity to describe.

Returns

DescribeActivityOutput

Exceptions

ActivityDoesNotExist or InvalidArn.

See also: AWS API Documentation

AWSSDK.StepFunctions.describe_execution — Function.

using AWSSDK.StepFunctions.describe_execution
describe_execution([::AWSConfig], arguments::Dict)
describe_execution([::AWSConfig]; executionArn=)

using AWSCore.Services.states
states([::AWSConfig], "DescribeExecution", arguments::Dict)
states([::AWSConfig], "DescribeExecution", executionArn=)

DescribeExecution Operation

Describes an execution.

Arguments

executionArn = ::String – Required

The Amazon Resource Name (ARN) of the execution to describe.

Returns

DescribeExecutionOutput

Exceptions

ExecutionDoesNotExist or InvalidArn.

See also: AWS API Documentation

AWSSDK.StepFunctions.describe_state_machine — Function.

using AWSSDK.StepFunctions.describe_state_machine
describe_state_machine([::AWSConfig], arguments::Dict)
describe_state_machine([::AWSConfig]; stateMachineArn=)

using AWSCore.Services.states
states([::AWSConfig], "DescribeStateMachine", arguments::Dict)
states([::AWSConfig], "DescribeStateMachine", stateMachineArn=)

DescribeStateMachine Operation

Describes a state machine.

Arguments

stateMachineArn = ::String – Required

The Amazon Resource Name (ARN) of the state machine to describe.

Returns

DescribeStateMachineOutput

Exceptions

InvalidArn or StateMachineDoesNotExist.

See also: AWS API Documentation

AWSSDK.StepFunctions.get_activity_task — Function.

using AWSSDK.StepFunctions.get_activity_task
get_activity_task([::AWSConfig], arguments::Dict)
get_activity_task([::AWSConfig]; activityArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "GetActivityTask", arguments::Dict)
states([::AWSConfig], "GetActivityTask", activityArn=, <keyword arguments>)

GetActivityTask Operation

Used by workers to retrieve a task (with the specified activity ARN) which has been scheduled for execution by a running state machine. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available (i.e. an execution of a task of this type is needed.) The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll will return a taskToken with a null string.

Important

Workers should set their client side socket timeout to at least 65 seconds (5 seconds higher than the maximum time the service may hold the poll request).

Arguments

activityArn = ::String – Required

The Amazon Resource Name (ARN) of the activity to retrieve tasks from (assigned when you create the task using CreateActivity.)

workerName = ::String

You can provide an arbitrary name in order to identify the worker that the task is assigned to. This name will be used when it is logged in the execution history.

Returns

GetActivityTaskOutput

Exceptions

ActivityDoesNotExist, ActivityWorkerLimitExceeded or InvalidArn.

See also: AWS API Documentation

AWSSDK.StepFunctions.get_execution_history — Function.

using AWSSDK.StepFunctions.get_execution_history
get_execution_history([::AWSConfig], arguments::Dict)
get_execution_history([::AWSConfig]; executionArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "GetExecutionHistory", arguments::Dict)
states([::AWSConfig], "GetExecutionHistory", executionArn=, <keyword arguments>)

GetExecutionHistory Operation

Returns the history of the specified execution as a list of events. By default, the results are returned in ascending order of the timeStamp of the events. Use the reverseOrder parameter to get the latest events first. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

Arguments

executionArn = ::String – Required

The Amazon Resource Name (ARN) of the execution.

maxResults = ::Int

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 100. A value of 0 means to use the default.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

reverseOrder = ::Bool

Lists events in descending order of their timeStamp.

nextToken = ::String

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken. Keep all other arguments unchanged.

The configured maxResults determines how many results can be returned in a single call.

Returns

GetExecutionHistoryOutput

Exceptions

ExecutionDoesNotExist, InvalidArn or InvalidToken.

See also: AWS API Documentation

AWSSDK.StepFunctions.list_activities — Function.

using AWSSDK.StepFunctions.list_activities
list_activities([::AWSConfig], arguments::Dict)
list_activities([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "ListActivities", arguments::Dict)
states([::AWSConfig], "ListActivities", <keyword arguments>)

ListActivities Operation

Lists the existing activities. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

Arguments

maxResults = ::Int

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 100. A value of 0 means to use the default.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

nextToken = ::String

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken. Keep all other arguments unchanged.

The configured maxResults determines how many results can be returned in a single call.

Returns

ListActivitiesOutput

Exceptions

InvalidToken.

See also: AWS API Documentation

AWSSDK.StepFunctions.list_executions — Function.

using AWSSDK.StepFunctions.list_executions
list_executions([::AWSConfig], arguments::Dict)
list_executions([::AWSConfig]; stateMachineArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "ListExecutions", arguments::Dict)
states([::AWSConfig], "ListExecutions", stateMachineArn=, <keyword arguments>)

ListExecutions Operation

Lists the executions of a state machine that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

Arguments

stateMachineArn = ::String – Required

The Amazon Resource Name (ARN) of the state machine whose executions will be listed.

statusFilter = "RUNNING", "SUCCEEDED", "FAILED", "TIMED_OUT" or "ABORTED"

If specified, only list the executions whose current execution status matches the given filter.

maxResults = ::Int

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 100. A value of 0 means to use the default.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

nextToken = ::String

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken. Keep all other arguments unchanged.

The configured maxResults determines how many results can be returned in a single call.

Returns

ListExecutionsOutput

Exceptions

InvalidArn, InvalidToken or StateMachineDoesNotExist.

See also: AWS API Documentation

AWSSDK.StepFunctions.list_state_machines — Function.

using AWSSDK.StepFunctions.list_state_machines
list_state_machines([::AWSConfig], arguments::Dict)
list_state_machines([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "ListStateMachines", arguments::Dict)
states([::AWSConfig], "ListStateMachines", <keyword arguments>)

ListStateMachines Operation

Lists the existing state machines. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

Arguments

maxResults = ::Int

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 100. A value of 0 means to use the default.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

nextToken = ::String

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken. Keep all other arguments unchanged.

The configured maxResults determines how many results can be returned in a single call.

Returns

ListStateMachinesOutput

Exceptions

InvalidToken.

See also: AWS API Documentation

AWSSDK.StepFunctions.send_task_failure — Function.

using AWSSDK.StepFunctions.send_task_failure
send_task_failure([::AWSConfig], arguments::Dict)
send_task_failure([::AWSConfig]; taskToken=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "SendTaskFailure", arguments::Dict)
states([::AWSConfig], "SendTaskFailure", taskToken=, <keyword arguments>)

SendTaskFailure Operation

Used by workers to report that the task identified by the taskToken failed.

Arguments

taskToken = ::String – Required

The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

error = ::String

An arbitrary error code that identifies the cause of the failure.

cause = ::String

A more detailed explanation of the cause of the failure.

Returns

SendTaskFailureOutput

Exceptions

TaskDoesNotExist, InvalidToken or TaskTimedOut.

See also: AWS API Documentation

AWSSDK.StepFunctions.send_task_heartbeat — Function.

using AWSSDK.StepFunctions.send_task_heartbeat
send_task_heartbeat([::AWSConfig], arguments::Dict)
send_task_heartbeat([::AWSConfig]; taskToken=)

using AWSCore.Services.states
states([::AWSConfig], "SendTaskHeartbeat", arguments::Dict)
states([::AWSConfig], "SendTaskHeartbeat", taskToken=)

SendTaskHeartbeat Operation

Used by workers to report to the service that the task represented by the specified taskToken is still making progress. This action resets the Heartbeat clock. The Heartbeat threshold is specified in the state machine's Amazon States Language definition. This action does not in itself create an event in the execution history. However, if the task times out, the execution history will contain an ActivityTimedOut event.

Note

The Timeout of a task, defined in the state machine's Amazon States Language definition, is its maximum allowed duration, regardless of the number of SendTaskHeartbeat requests received.

Note

This operation is only useful for long-lived tasks to report the liveliness of the task.

Arguments

taskToken = ::String – Required

The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

Returns

SendTaskHeartbeatOutput

Exceptions

TaskDoesNotExist, InvalidToken or TaskTimedOut.

See also: AWS API Documentation

AWSSDK.StepFunctions.send_task_success — Function.

using AWSSDK.StepFunctions.send_task_success
send_task_success([::AWSConfig], arguments::Dict)
send_task_success([::AWSConfig]; taskToken=, output=)

using AWSCore.Services.states
states([::AWSConfig], "SendTaskSuccess", arguments::Dict)
states([::AWSConfig], "SendTaskSuccess", taskToken=, output=)

SendTaskSuccess Operation

Used by workers to report that the task identified by the taskToken completed successfully.

Arguments

taskToken = ::String – Required

The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

output = ::String – Required

The JSON output of the task.

Returns

SendTaskSuccessOutput

Exceptions

TaskDoesNotExist, InvalidOutput, InvalidToken or TaskTimedOut.

See also: AWS API Documentation

AWSSDK.StepFunctions.start_execution — Function.

using AWSSDK.StepFunctions.start_execution
start_execution([::AWSConfig], arguments::Dict)
start_execution([::AWSConfig]; stateMachineArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "StartExecution", arguments::Dict)
states([::AWSConfig], "StartExecution", stateMachineArn=, <keyword arguments>)

StartExecution Operation

Starts a state machine execution.

Arguments

stateMachineArn = ::String – Required

The Amazon Resource Name (ARN) of the state machine to execute.

name = ::String

The name of the execution. This name must be unique for your AWS account and region for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide.

A name must not contain:

whitespace

brackets < > { } [ ]

wildcard characters ? *

special characters " # % \ ^ | ~ $ & , ; : /`

control characters (U+0000-001F, U+007F-009F)

input = ::String

The string that contains the JSON input data for the execution, for example:

"input": "{\"first_name\" : \"test\"}"

Note

If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}"

Returns

StartExecutionOutput

Exceptions

ExecutionLimitExceeded, ExecutionAlreadyExists, InvalidArn, InvalidExecutionInput, InvalidName, StateMachineDoesNotExist or StateMachineDeleting.

See also: AWS API Documentation

AWSSDK.StepFunctions.stop_execution — Function.

using AWSSDK.StepFunctions.stop_execution
stop_execution([::AWSConfig], arguments::Dict)
stop_execution([::AWSConfig]; executionArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "StopExecution", arguments::Dict)
states([::AWSConfig], "StopExecution", executionArn=, <keyword arguments>)

StopExecution Operation

Stops an execution.

Arguments

executionArn = ::String – Required

The Amazon Resource Name (ARN) of the execution to stop.

error = ::String

An arbitrary error code that identifies the cause of the termination.

cause = ::String

A more detailed explanation of the cause of the termination.

Returns

StopExecutionOutput

Exceptions

ExecutionDoesNotExist or InvalidArn.

See also: AWS API Documentation