AmazonCloudSearch.jl

AmazonCloudSearch

Amazon CloudSearch Configuration Service

You use the Amazon CloudSearch configuration service to create, configure, and manage search domains. Configuration service requests are submitted using the AWS Query protocol. AWS Query requests are HTTP or HTTPS requests submitted via HTTP GET or POST with a query parameter named Action.

The endpoint for configuration service requests is region-specific: cloudsearch.region.amazonaws.com. For example, cloudsearch.us-east-1.amazonaws.com. For a current list of supported regions and endpoints, see Regions and Endpoints.

This document is generated from apis/cloudsearch-2013-01-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCloudSearch.build_suggesters

AmazonCloudSearch.create_domain

AmazonCloudSearch.define_analysis_scheme

AmazonCloudSearch.define_expression

AmazonCloudSearch.define_index_field

AmazonCloudSearch.define_suggester

AmazonCloudSearch.delete_analysis_scheme

AmazonCloudSearch.delete_domain

AmazonCloudSearch.delete_expression

AmazonCloudSearch.delete_index_field

AmazonCloudSearch.delete_suggester

AmazonCloudSearch.describe_analysis_schemes

AmazonCloudSearch.describe_availability_options

AmazonCloudSearch.describe_domains

AmazonCloudSearch.describe_expressions

AmazonCloudSearch.describe_index_fields

AmazonCloudSearch.describe_scaling_parameters

AmazonCloudSearch.describe_service_access_policies

AmazonCloudSearch.describe_suggesters

AmazonCloudSearch.index_documents

AmazonCloudSearch.list_domain_names

AmazonCloudSearch.update_availability_options

AmazonCloudSearch.update_scaling_parameters

AmazonCloudSearch.update_service_access_policies

AmazonCloudSearch.build_suggesters — Function.

build_suggesters(::AWSConfig, arguments::Dict)
build_suggesters([::AWSConfig]; DomainName=)

BuildSuggesters Operation

Indexes the search suggestions. For more information, see Configuring Suggesters in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

Returns

BuildSuggestersResponse

Exceptions

BaseException, InternalException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.create_domain — Function.

create_domain(::AWSConfig, arguments::Dict)
create_domain([::AWSConfig]; DomainName=)

CreateDomain Operation

Creates a new search domain. For more information, see Creating a Search Domain in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

A name for the domain you are creating. Allowed characters are a-z (lower-case letters), 0-9, and hyphen (-). Domain names must start with a letter or number and be at least 3 and no more than 28 characters long.

Returns

CreateDomainResponse

Exceptions

BaseException, InternalException or LimitExceededException.

See also: AWS API Documentation

AmazonCloudSearch.define_analysis_scheme — Function.

define_analysis_scheme(::AWSConfig, arguments::Dict)
define_analysis_scheme([::AWSConfig]; DomainName=, AnalysisScheme=)

DefineAnalysisScheme Operation

Configures an analysis scheme that can be applied to a text or text-array field to define language-specific text processing options. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

AnalysisScheme = Dict( ... ) – Required

 AnalysisScheme = Dict(
        "AnalysisSchemeName" => <required> ::String,
        "AnalysisSchemeLanguage" => <required> "ar", "bg", "ca", "cs", "da", "de", "el", "en", "es", "eu", "fa", "fi", "fr", "ga", "gl", "he", "hi", "hu", "hy", "id", "it", "ja", "ko", "lv", "mul", "nl", "no", "pt", "ro", "ru", "sv", "th", "tr", "zh-Hans" or "zh-Hant",
        "AnalysisOptions" =>  Dict(
            "Synonyms" =>  ::String,
            "Stopwords" =>  ::String,
            "StemmingDictionary" =>  ::String,
            "JapaneseTokenizationDictionary" =>  ::String,
            "AlgorithmicStemming" =>  "none", "minimal", "light" or "full"
        )
    )

Returns

DefineAnalysisSchemeResponse

Exceptions

BaseException, InternalException, LimitExceededException, InvalidTypeException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.define_expression — Function.

define_expression(::AWSConfig, arguments::Dict)
define_expression([::AWSConfig]; DomainName=, Expression=)

DefineExpression Operation

Configures an [Expression](@ref) for the search domain. Used to create new expressions and modify existing ones. If the expression exists, the new configuration replaces the old one. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

Expression = Dict( ... ) – Required

 Expression = Dict(
        "ExpressionName" => <required> ::String,
        "ExpressionValue" => <required> ::String
    )

Returns

DefineExpressionResponse

Exceptions

BaseException, InternalException, LimitExceededException, InvalidTypeException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.define_index_field — Function.

define_index_field(::AWSConfig, arguments::Dict)
define_index_field([::AWSConfig]; DomainName=, IndexField=)

DefineIndexField Operation

Configures an [IndexField](@ref) for the search domain. Used to create new fields and modify existing ones. You must specify the name of the domain you are configuring and an index field configuration. The index field configuration specifies a unique name, the index field type, and the options you want to configure for the field. The options you can specify depend on the [IndexFieldType](@ref). If the field exists, the new configuration replaces the old one. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

IndexField = Dict( ... ) – Required

The index field and field options you want to configure.

 IndexField = Dict(
        "IndexFieldName" => <required> ::String,
        "IndexFieldType" => <required> "int", "double", "literal", "text", "date", "latlon", "int-array", "double-array", "literal-array", "text-array" or "date-array",
        "IntOptions" =>  Dict(
            "DefaultValue" =>  ::Int,
            "SourceField" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool,
            "SortEnabled" =>  ::Bool
        ),
        "DoubleOptions" =>  Dict(
            "DefaultValue" =>  double,
            "SourceField" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool,
            "SortEnabled" =>  ::Bool
        ),
        "LiteralOptions" =>  Dict(
            "DefaultValue" =>  ::String,
            "SourceField" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool,
            "SortEnabled" =>  ::Bool
        ),
        "TextOptions" =>  Dict(
            "DefaultValue" =>  ::String,
            "SourceField" =>  ::String,
            "ReturnEnabled" =>  ::Bool,
            "SortEnabled" =>  ::Bool,
            "HighlightEnabled" =>  ::Bool,
            "AnalysisScheme" =>  ::String
        ),
        "DateOptions" =>  Dict(
            "DefaultValue" =>  ::String,
            "SourceField" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool,
            "SortEnabled" =>  ::Bool
        ),
        "LatLonOptions" =>  Dict(
            "DefaultValue" =>  ::String,
            "SourceField" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool,
            "SortEnabled" =>  ::Bool
        ),
        "IntArrayOptions" =>  Dict(
            "DefaultValue" =>  ::Int,
            "SourceFields" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool
        ),
        "DoubleArrayOptions" =>  Dict(
            "DefaultValue" =>  double,
            "SourceFields" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool
        ),
        "LiteralArrayOptions" =>  Dict(
            "DefaultValue" =>  ::String,
            "SourceFields" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool
        ),
        "TextArrayOptions" =>  Dict(
            "DefaultValue" =>  ::String,
            "SourceFields" =>  ::String,
            "ReturnEnabled" =>  ::Bool,
            "HighlightEnabled" =>  ::Bool,
            "AnalysisScheme" =>  ::String
        ),
        "DateArrayOptions" =>  Dict(
            "DefaultValue" =>  ::String,
            "SourceFields" =>  ::String,
            "FacetEnabled" =>  ::Bool,
            "SearchEnabled" =>  ::Bool,
            "ReturnEnabled" =>  ::Bool
        )
    )

Returns

DefineIndexFieldResponse

Exceptions

BaseException, InternalException, LimitExceededException, InvalidTypeException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.define_suggester — Function.

define_suggester(::AWSConfig, arguments::Dict)
define_suggester([::AWSConfig]; DomainName=, Suggester=)

DefineSuggester Operation

Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

Suggester = Dict( ... ) – Required

 Suggester = Dict(
        "SuggesterName" => <required> ::String,
        "DocumentSuggesterOptions" => <required> Dict(
            "SourceField" => <required> ::String,
            "FuzzyMatching" =>  "none", "low" or "high",
            "SortExpression" =>  ::String
        )
    )

Returns

DefineSuggesterResponse

Exceptions

BaseException, InternalException, LimitExceededException, InvalidTypeException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.delete_analysis_scheme — Function.

delete_analysis_scheme(::AWSConfig, arguments::Dict)
delete_analysis_scheme([::AWSConfig]; DomainName=, AnalysisSchemeName=)

DeleteAnalysisScheme Operation

Deletes an analysis scheme. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

AnalysisSchemeName = ::String – Required

The name of the analysis scheme you want to delete.

Returns

DeleteAnalysisSchemeResponse

Exceptions

BaseException, InternalException, InvalidTypeException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.delete_domain — Function.

delete_domain(::AWSConfig, arguments::Dict)
delete_domain([::AWSConfig]; DomainName=)

DeleteDomain Operation

Permanently deletes a search domain and all of its data. Once a domain has been deleted, it cannot be recovered. For more information, see Deleting a Search Domain in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

The name of the domain you want to permanently delete.

Returns

DeleteDomainResponse

Exceptions

BaseException or InternalException.

See also: AWS API Documentation

AmazonCloudSearch.delete_expression — Function.

delete_expression(::AWSConfig, arguments::Dict)
delete_expression([::AWSConfig]; DomainName=, ExpressionName=)

DeleteExpression Operation

Removes an [Expression](@ref) from the search domain. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

ExpressionName = ::String – Required

The name of the [Expression](@ref) to delete.

Returns

DeleteExpressionResponse

Exceptions

BaseException, InternalException, InvalidTypeException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.delete_index_field — Function.

delete_index_field(::AWSConfig, arguments::Dict)
delete_index_field([::AWSConfig]; DomainName=, IndexFieldName=)

DeleteIndexField Operation

Removes an [IndexField](@ref) from the search domain. For more information, see Configuring Index Fields in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

IndexFieldName = ::String – Required

The name of the index field your want to remove from the domain's indexing options.

Returns

DeleteIndexFieldResponse

Exceptions

BaseException, InternalException, InvalidTypeException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.delete_suggester — Function.

delete_suggester(::AWSConfig, arguments::Dict)
delete_suggester([::AWSConfig]; DomainName=, SuggesterName=)

DeleteSuggester Operation

Deletes a suggester. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

SuggesterName = ::String – Required

Specifies the name of the suggester you want to delete.

Returns

DeleteSuggesterResponse

Exceptions

BaseException, InternalException, InvalidTypeException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.describe_analysis_schemes — Function.

describe_analysis_schemes(::AWSConfig, arguments::Dict)
describe_analysis_schemes([::AWSConfig]; DomainName=, <keyword arguments>)

DescribeAnalysisSchemes Operation

Gets the analysis schemes configured for a domain. An analysis scheme defines language-specific text processing options for a text field. Can be limited to specific analysis schemes by name. By default, shows all analysis schemes and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Analysis Schemes in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

The name of the domain you want to describe.

AnalysisSchemeNames = [::String...]

The analysis schemes you want to describe.

Deployed = ::Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Returns

DescribeAnalysisSchemesResponse

Exceptions

BaseException, InternalException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.describe_availability_options — Function.

describe_availability_options(::AWSConfig, arguments::Dict)
describe_availability_options([::AWSConfig]; DomainName=, <keyword arguments>)

DescribeAvailabilityOptions Operation

Gets the availability options configured for a domain. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

The name of the domain you want to describe.

Deployed = ::Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Returns

DescribeAvailabilityOptionsResponse

Exceptions

BaseException, InternalException, InvalidTypeException, LimitExceededException, ResourceNotFoundException or DisabledOperationException.

See also: AWS API Documentation

AmazonCloudSearch.describe_domains — Function.

describe_domains(::AWSConfig, arguments::Dict)
describe_domains([::AWSConfig]; <keyword arguments>)

DescribeDomains Operation

Gets information about the search domains owned by this account. Can be limited to specific domains. Shows all domains by default. To get the number of searchable documents in a domain, use the console or submit a matchall request to your domain's search endpoint: q=matchall&amp;q.parser=structured&amp;size=0. For more information, see Getting Information about a Search Domain in the Amazon CloudSearch Developer Guide.

Arguments

DomainNames = [::String...]

The names of the domains you want to include in the response.

Returns

DescribeDomainsResponse

Exceptions

BaseException or InternalException.

See also: AWS API Documentation

AmazonCloudSearch.describe_expressions — Function.

describe_expressions(::AWSConfig, arguments::Dict)
describe_expressions([::AWSConfig]; DomainName=, <keyword arguments>)

DescribeExpressions Operation

Gets the expressions configured for the search domain. Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Expressions in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

The name of the domain you want to describe.

ExpressionNames = [::String...]

Limits the [DescribeExpressions](@ref) response to the specified expressions. If not specified, all expressions are shown.

Deployed = ::Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Returns

DescribeExpressionsResponse

Exceptions

BaseException, InternalException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.describe_index_fields — Function.

describe_index_fields(::AWSConfig, arguments::Dict)
describe_index_fields([::AWSConfig]; DomainName=, <keyword arguments>)

DescribeIndexFields Operation

Gets information about the index fields configured for the search domain. Can be limited to specific fields by name. By default, shows all fields and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Domain Information in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

The name of the domain you want to describe.

FieldNames = [::String...]

A list of the index fields you want to describe. If not specified, information is returned for all configured index fields.

Deployed = ::Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Returns

DescribeIndexFieldsResponse

Exceptions

BaseException, InternalException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.describe_scaling_parameters — Function.

describe_scaling_parameters(::AWSConfig, arguments::Dict)
describe_scaling_parameters([::AWSConfig]; DomainName=)

DescribeScalingParameters Operation

Gets the scaling parameters configured for a domain. A domain's scaling parameters specify the desired search instance type and replication count. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

Returns

DescribeScalingParametersResponse

Exceptions

BaseException, InternalException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.describe_service_access_policies — Function.

describe_service_access_policies(::AWSConfig, arguments::Dict)
describe_service_access_policies([::AWSConfig]; DomainName=, <keyword arguments>)

DescribeServiceAccessPolicies Operation

Gets information about the access policies that control access to the domain's document and search endpoints. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Configuring Access for a Search Domain in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

The name of the domain you want to describe.

Deployed = ::Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Returns

DescribeServiceAccessPoliciesResponse

Exceptions

BaseException, InternalException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.describe_suggesters — Function.

describe_suggesters(::AWSConfig, arguments::Dict)
describe_suggesters([::AWSConfig]; DomainName=, <keyword arguments>)

DescribeSuggesters Operation

Gets the suggesters configured for a domain. A suggester enables you to display possible matches before users finish typing their queries. Can be limited to specific suggesters by name. By default, shows all suggesters and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see Getting Search Suggestions in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

The name of the domain you want to describe.

SuggesterNames = [::String...]

The suggesters you want to describe.

Deployed = ::Bool

Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Returns

DescribeSuggestersResponse

Exceptions

BaseException, InternalException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.index_documents — Function.

index_documents(::AWSConfig, arguments::Dict)
index_documents([::AWSConfig]; DomainName=)

IndexDocuments Operation

Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is RequiresIndexDocuments.

Arguments

DomainName = ::String – Required

Returns

IndexDocumentsResponse

Exceptions

BaseException, InternalException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCloudSearch.list_domain_names — Function.

list_domain_names([::AWSConfig])

ListDomainNames Operation

Lists all search domains owned by an account.

Returns

ListDomainNamesResponse

Exceptions

BaseException.

See also: AWS API Documentation

AmazonCloudSearch.update_availability_options — Function.

update_availability_options(::AWSConfig, arguments::Dict)
update_availability_options([::AWSConfig]; DomainName=, MultiAZ=)

UpdateAvailabilityOptions Operation

Configures the availability options for a domain. Enabling the Multi-AZ option expands an Amazon CloudSearch domain to an additional Availability Zone in the same Region to increase fault tolerance in the event of a service disruption. Changes to the Multi-AZ option can take about half an hour to become active. For more information, see Configuring Availability Options in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

MultiAZ = ::Bool – Required

You expand an existing search domain to a second Availability Zone by setting the Multi-AZ option to true. Similarly, you can turn off the Multi-AZ option to downgrade the domain to a single Availability Zone by setting the Multi-AZ option to false.

Returns

UpdateAvailabilityOptionsResponse

Exceptions

BaseException, InternalException, InvalidTypeException, LimitExceededException, ResourceNotFoundException or DisabledOperationException.

See also: AWS API Documentation

AmazonCloudSearch.update_scaling_parameters — Function.

update_scaling_parameters(::AWSConfig, arguments::Dict)
update_scaling_parameters([::AWSConfig]; DomainName=, ScalingParameters=)

UpdateScalingParameters Operation

Configures scaling parameters for a domain. A domain's scaling parameters specify the desired search instance type and replication count. Amazon CloudSearch will still automatically scale your domain based on the volume of data and traffic, but not below the desired instance type and replication count. If the Multi-AZ option is enabled, these values control the resources used per Availability Zone. For more information, see Configuring Scaling Options in the Amazon CloudSearch Developer Guide.

Arguments

DomainName = ::String – Required

ScalingParameters = Dict( ... ) – Required

 ScalingParameters = Dict(
        "DesiredInstanceType" =>  "search.m1.small", "search.m1.large", "search.m2.xlarge", "search.m2.2xlarge", "search.m3.medium", "search.m3.large", "search.m3.xlarge" or "search.m3.2xlarge",
        "DesiredReplicationCount" =>  ::Int,
        "DesiredPartitionCount" =>  ::Int
    )

Returns

UpdateScalingParametersResponse

Exceptions

BaseException, InternalException, LimitExceededException, ResourceNotFoundException or InvalidTypeException.

See also: AWS API Documentation

AmazonCloudSearch.update_service_access_policies — Function.

update_service_access_policies(::AWSConfig, arguments::Dict)
update_service_access_policies([::AWSConfig]; DomainName=, AccessPolicies=)

UpdateServiceAccessPolicies Operation

Configures the access rules that control access to the domain's document and search endpoints. For more information, see Configuring Access for an Amazon CloudSearch Domain.

Arguments

DomainName = ::String – Required

AccessPolicies = ::String – Required

The access rules you want to configure. These rules replace any existing rules.

Returns

UpdateServiceAccessPoliciesResponse

Exceptions

BaseException, InternalException, LimitExceededException, ResourceNotFoundException or InvalidTypeException.

See also: AWS API Documentation