AmazonCloudWatchLogs.jl

AmazonCloudWatchLogs

You can use Amazon CloudWatch Logs to monitor, store, and access your log files from EC2 instances, Amazon CloudTrail, or other sources. You can then retrieve the associated log data from CloudWatch Logs using the Amazon CloudWatch console, the CloudWatch Logs commands in the AWS CLI, the CloudWatch Logs API, or the CloudWatch Logs SDK.

You can use CloudWatch Logs to:

Monitor Logs from Amazon EC2 Instances in Real-time: You can use CloudWatch Logs to monitor applications and systems using log data. For example, CloudWatch Logs can track the number of errors that occur in your application logs and send you a notification whenever the rate of errors exceeds a threshold you specify. CloudWatch Logs uses your log data for monitoring; so, no code changes are required. For example, you can monitor application logs for specific literal terms (such as "NullReferenceException") or count the number of occurrences of a literal term at a particular position in log data (such as "404" status codes in an Apache access log). When the term you are searching for is found, CloudWatch Logs reports the data to a Amazon CloudWatch metric that you specify.

Monitor Amazon CloudTrail Logged Events: You can create alarms in Amazon CloudWatch and receive notifications of particular API activity as captured by CloudTrail and use the notification to perform troubleshooting.

Archive Log Data: You can use CloudWatch Logs to store your log data in highly durable storage. You can change the log retention setting so that any log events older than this setting are automatically deleted. The CloudWatch Logs agent makes it easy to quickly send both rotated and non-rotated log data off of a host and into the log service. You can then access the raw log data when you need it.

This document is generated from apis/logs-2014-03-28.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCloudWatchLogs.cancel_export_task

AmazonCloudWatchLogs.create_export_task

AmazonCloudWatchLogs.create_log_group

AmazonCloudWatchLogs.create_log_stream

AmazonCloudWatchLogs.delete_destination

AmazonCloudWatchLogs.delete_log_group

AmazonCloudWatchLogs.delete_log_stream

AmazonCloudWatchLogs.delete_metric_filter

AmazonCloudWatchLogs.delete_retention_policy

AmazonCloudWatchLogs.delete_subscription_filter

AmazonCloudWatchLogs.describe_destinations

AmazonCloudWatchLogs.describe_export_tasks

AmazonCloudWatchLogs.describe_log_groups

AmazonCloudWatchLogs.describe_log_streams

AmazonCloudWatchLogs.describe_metric_filters

AmazonCloudWatchLogs.describe_subscription_filters

AmazonCloudWatchLogs.filter_log_events

AmazonCloudWatchLogs.get_log_events

AmazonCloudWatchLogs.list_tags_log_group

AmazonCloudWatchLogs.put_destination

AmazonCloudWatchLogs.put_destination_policy

AmazonCloudWatchLogs.put_log_events

AmazonCloudWatchLogs.put_metric_filter

AmazonCloudWatchLogs.put_retention_policy

AmazonCloudWatchLogs.put_subscription_filter

AmazonCloudWatchLogs.tag_log_group

AmazonCloudWatchLogs.test_metric_filter

AmazonCloudWatchLogs.untag_log_group

AmazonCloudWatchLogs.cancel_export_task — Function.

cancel_export_task(::AWSConfig, arguments::Dict)
cancel_export_task([::AWSConfig]; taskId=)

CancelExportTask Operation

Cancels the specified export task.

The task must be in the PENDING or RUNNING state.

Arguments

taskId = ::String – Required

The ID of the export task.

Exceptions

InvalidParameterException, ResourceNotFoundException, InvalidOperationException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.create_export_task — Function.

create_export_task(::AWSConfig, arguments::Dict)
create_export_task([::AWSConfig]; logGroupName=, from=, to=, destination=, <keyword arguments>)

CreateExportTask Operation

Creates an export task, which allows you to efficiently export data from a log group to an Amazon S3 bucket.

This is an asynchronous call. If all the required information is provided, this operation initiates an export task and responds with the ID of the task. After the task has started, you can use DescribeExportTasks to get the status of the export task. Each account can only have one active (RUNNING or PENDING) export task at a time. To cancel an export task, use CancelExportTask.

You can export logs from multiple log groups or multiple time ranges to the same S3 bucket. To separate out log data for each export task, you can specify a prefix that will be used as the Amazon S3 key prefix for all exported objects.

Arguments

taskName = ::String

The name of the export task.

logGroupName = ::String – Required

The name of the log group.

logStreamNamePrefix = ::String

Export only log streams that match the provided prefix. If you don't specify a value, no prefix filter is applied.

from = ::Int – Required

The start time of the range for the request, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not exported.

to = ::Int – Required

The end time of the range for the request, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

destination = ::String – Required

The name of S3 bucket for the exported log data. The bucket must be in the same AWS region.

destinationPrefix = ::String

The prefix used as the start of the key for every object exported. If you don't specify a value, the default is exportedlogs.

Returns

CreateExportTaskResponse

Exceptions

InvalidParameterException, LimitExceededException, OperationAbortedException, ServiceUnavailableException, ResourceNotFoundException or ResourceAlreadyExistsException.

See also: AWS API Documentation

AmazonCloudWatchLogs.create_log_group — Function.

create_log_group(::AWSConfig, arguments::Dict)
create_log_group([::AWSConfig]; logGroupName=, <keyword arguments>)

CreateLogGroup Operation

Creates a log group with the specified name.

You can create up to 5000 log groups per account.

You must use the following guidelines when naming a log group:

Log group names must be unique within a region for an AWS account.

Log group names can be between 1 and 512 characters long.

Log group names consist of the following characters: a-z, A-Z, 0-9, '_' (underscore), '-' (hyphen), '/' (forward slash), and '.' (period).

Arguments

logGroupName = ::String – Required

The name of the log group.

tags = ::Dict{String,String}

The key-value pairs to use for the tags.

Exceptions

InvalidParameterException, ResourceAlreadyExistsException, LimitExceededException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.create_log_stream — Function.

create_log_stream(::AWSConfig, arguments::Dict)
create_log_stream([::AWSConfig]; logGroupName=, logStreamName=)

CreateLogStream Operation

Creates a log stream for the specified log group.

There is no limit on the number of log streams that you can create for a log group.

You must use the following guidelines when naming a log stream:

Log stream names must be unique within the log group.

Log stream names can be between 1 and 512 characters long.

The ':' (colon) and '*' (asterisk) characters are not allowed.

Arguments

logGroupName = ::String – Required

The name of the log group.

logStreamName = ::String – Required

The name of the log stream.

Exceptions

InvalidParameterException, ResourceAlreadyExistsException, ResourceNotFoundException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.delete_destination — Function.

delete_destination(::AWSConfig, arguments::Dict)
delete_destination([::AWSConfig]; destinationName=)

DeleteDestination Operation

Deletes the specified destination, and eventually disables all the subscription filters that publish to it. This operation does not delete the physical resource encapsulated by the destination.

Arguments

destinationName = ::String – Required

The name of the destination.

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.delete_log_group — Function.

delete_log_group(::AWSConfig, arguments::Dict)
delete_log_group([::AWSConfig]; logGroupName=)

DeleteLogGroup Operation

Deletes the specified log group and permanently deletes all the archived log events associated with the log group.

Arguments

logGroupName = ::String – Required

The name of the log group.

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.delete_log_stream — Function.

delete_log_stream(::AWSConfig, arguments::Dict)
delete_log_stream([::AWSConfig]; logGroupName=, logStreamName=)

DeleteLogStream Operation

Deletes the specified log stream and permanently deletes all the archived log events associated with the log stream.

Arguments

logGroupName = ::String – Required

The name of the log group.

logStreamName = ::String – Required

The name of the log stream.

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.delete_metric_filter — Function.

delete_metric_filter(::AWSConfig, arguments::Dict)
delete_metric_filter([::AWSConfig]; logGroupName=, filterName=)

DeleteMetricFilter Operation

Deletes the specified metric filter.

Arguments

logGroupName = ::String – Required

The name of the log group.

filterName = ::String – Required

The name of the metric filter.

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.delete_retention_policy — Function.

delete_retention_policy(::AWSConfig, arguments::Dict)
delete_retention_policy([::AWSConfig]; logGroupName=)

DeleteRetentionPolicy Operation

Deletes the specified retention policy.

Log events do not expire if they belong to log groups without a retention policy.

Arguments

logGroupName = ::String – Required

The name of the log group.

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.delete_subscription_filter — Function.

delete_subscription_filter(::AWSConfig, arguments::Dict)
delete_subscription_filter([::AWSConfig]; logGroupName=, filterName=)

DeleteSubscriptionFilter Operation

Deletes the specified subscription filter.

Arguments

logGroupName = ::String – Required

The name of the log group.

filterName = ::String – Required

The name of the subscription filter.

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.describe_destinations — Function.

describe_destinations(::AWSConfig, arguments::Dict)
describe_destinations([::AWSConfig]; <keyword arguments>)

DescribeDestinations Operation

Lists all your destinations. The results are ASCII-sorted by destination name.

Arguments

DestinationNamePrefix = ::String

The prefix to match. If you don't specify a value, no prefix filter is applied.

nextToken = ::String

The token for the next set of items to return. (You received this token from a previous call.)

limit = ::Int

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Returns

DescribeDestinationsResponse

Exceptions

InvalidParameterException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.describe_export_tasks — Function.

describe_export_tasks(::AWSConfig, arguments::Dict)
describe_export_tasks([::AWSConfig]; <keyword arguments>)

DescribeExportTasks Operation

Lists the specified export tasks. You can list all your export tasks or filter the results based on task ID or task status.

Arguments

taskId = ::String

The ID of the export task. Specifying a task ID filters the results to zero or one export tasks.

statusCode = "CANCELLED", "COMPLETED", "FAILED", "PENDING", "PENDING_CANCEL" or "RUNNING"

The status code of the export task. Specifying a status code filters the results to zero or more export tasks.

nextToken = ::String

The token for the next set of items to return. (You received this token from a previous call.)

limit = ::Int

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Returns

DescribeExportTasksResponse

Exceptions

InvalidParameterException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.describe_log_groups — Function.

describe_log_groups(::AWSConfig, arguments::Dict)
describe_log_groups([::AWSConfig]; <keyword arguments>)

DescribeLogGroups Operation

Lists the specified log groups. You can list all your log groups or filter the results by prefix. The results are ASCII-sorted by log group name.

Arguments

logGroupNamePrefix = ::String

The prefix to match.

nextToken = ::String

The token for the next set of items to return. (You received this token from a previous call.)

limit = ::Int

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Returns

DescribeLogGroupsResponse

Exceptions

InvalidParameterException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.describe_log_streams — Function.

describe_log_streams(::AWSConfig, arguments::Dict)
describe_log_streams([::AWSConfig]; logGroupName=, <keyword arguments>)

DescribeLogStreams Operation

Lists the log streams for the specified log group. You can list all the log streams or filter the results by prefix. You can also control how the results are ordered.

This operation has a limit of five transactions per second, after which transactions are throttled.

Arguments

logGroupName = ::String – Required

The name of the log group.

logStreamNamePrefix = ::String

The prefix to match.

You cannot specify this parameter if orderBy is LastEventTime.

orderBy = "LogStreamName" or "LastEventTime"

If the value is LogStreamName, the results are ordered by log stream name. If the value is LastEventTime, the results are ordered by the event time. The default value is LogStreamName.

If you order the results by event time, you cannot specify the logStreamNamePrefix parameter.

lastEventTimestamp represents the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. lastEventTimeStamp updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but may take longer in some rare situations.

descending = ::Bool

If the value is true, results are returned in descending order. If the value is to false, results are returned in ascending order. The default value is false.

nextToken = ::String

The token for the next set of items to return. (You received this token from a previous call.)

limit = ::Int

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Returns

DescribeLogStreamsResponse

Exceptions

InvalidParameterException, ResourceNotFoundException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.describe_metric_filters — Function.

describe_metric_filters(::AWSConfig, arguments::Dict)
describe_metric_filters([::AWSConfig]; <keyword arguments>)

DescribeMetricFilters Operation

Lists the specified metric filters. You can list all the metric filters or filter the results by log name, prefix, metric name, and metric namespace. The results are ASCII-sorted by filter name.

Arguments

logGroupName = ::String

The name of the log group.

filterNamePrefix = ::String

The prefix to match.

nextToken = ::String

The token for the next set of items to return. (You received this token from a previous call.)

limit = ::Int

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

metricName = ::String

The name of the CloudWatch metric.

metricNamespace = ::String

The namespace of the CloudWatch metric.

Returns

DescribeMetricFiltersResponse

Exceptions

InvalidParameterException, ResourceNotFoundException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.describe_subscription_filters — Function.

describe_subscription_filters(::AWSConfig, arguments::Dict)
describe_subscription_filters([::AWSConfig]; logGroupName=, <keyword arguments>)

DescribeSubscriptionFilters Operation

Lists the subscription filters for the specified log group. You can list all the subscription filters or filter the results by prefix. The results are ASCII-sorted by filter name.

Arguments

logGroupName = ::String – Required

The name of the log group.

filterNamePrefix = ::String

The prefix to match. If you don't specify a value, no prefix filter is applied.

nextToken = ::String

The token for the next set of items to return. (You received this token from a previous call.)

limit = ::Int

The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

Returns

DescribeSubscriptionFiltersResponse

Exceptions

InvalidParameterException, ResourceNotFoundException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.filter_log_events — Function.

filter_log_events(::AWSConfig, arguments::Dict)
filter_log_events([::AWSConfig]; logGroupName=, <keyword arguments>)

FilterLogEvents Operation

Lists log events from the specified log group. You can list all the log events or filter the results using a filter pattern, a time range, and the name of the log stream.

By default, this operation returns as many log events as can fit in 1MB (up to 10,000 log events), or all the events found within the time range that you specify. If the results include a token, then there are more log events available, and you can get additional results by specifying the token in a subsequent call.

Arguments

logGroupName = ::String – Required

The name of the log group.

logStreamNames = [::String...]

Optional list of log stream names.

startTime = ::Int

The start of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp prior to this time are not returned.

endTime = ::Int

The end of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not returned.

filterPattern = ::String

The filter pattern to use. If not provided, all the events are matched.

nextToken = ::String

The token for the next set of events to return. (You received this token from a previous call.)

limit = ::Int

The maximum number of events to return. The default is 10,000 events.

interleaved = ::Bool

If the value is true, the operation makes a best effort to provide responses that contain events from multiple log streams within the log group interleaved in a single response. If the value is false all the matched log events in the first log stream are searched first, then those in the next log stream, and so on. The default is false.

Returns

FilterLogEventsResponse

Exceptions

InvalidParameterException, ResourceNotFoundException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.get_log_events — Function.

get_log_events(::AWSConfig, arguments::Dict)
get_log_events([::AWSConfig]; logGroupName=, logStreamName=, <keyword arguments>)

GetLogEvents Operation

Lists log events from the specified log stream. You can list all the log events or filter using a time range.

By default, this operation returns as many log events as can fit in a response size of 1MB (up to 10,000 log events). If the results include tokens, there are more log events available. You can get additional log events by specifying one of the tokens in a subsequent call.

Arguments

logGroupName = ::String – Required

The name of the log group.

logStreamName = ::String – Required

The name of the log stream.

startTime = ::Int

The start of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not included.

endTime = ::Int

The end of the time range, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not included.

nextToken = ::String

The token for the next set of items to return. (You received this token from a previous call.)

limit = ::Int

The maximum number of log events returned. If you don't specify a value, the maximum is as many log events as can fit in a response size of 1MB, up to 10,000 log events.

startFromHead = ::Bool

If the value is true, the earliest log events are returned first. If the value is false, the latest log events are returned first. The default value is false.

Returns

GetLogEventsResponse

Exceptions

InvalidParameterException, ResourceNotFoundException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.list_tags_log_group — Function.

list_tags_log_group(::AWSConfig, arguments::Dict)
list_tags_log_group([::AWSConfig]; logGroupName=)

ListTagsLogGroup Operation

Lists the tags for the specified log group.

To add tags, use TagLogGroup. To remove tags, use UntagLogGroup.

Arguments

logGroupName = ::String – Required

The name of the log group.

Returns

ListTagsLogGroupResponse

Exceptions

ResourceNotFoundException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.put_destination — Function.

put_destination(::AWSConfig, arguments::Dict)
put_destination([::AWSConfig]; destinationName=, targetArn=, roleArn=)

PutDestination Operation

Creates or updates a destination. A destination encapsulates a physical resource (such as a Kinesis stream) and enables you to subscribe to a real-time stream of log events of a different account, ingested using PutLogEvents. Currently, the only supported physical resource is a Amazon Kinesis stream belonging to the same account as the destination.

A destination controls what is written to its Amazon Kinesis stream through an access policy. By default, PutDestination does not set any access policy with the destination, which means a cross-account user cannot call PutSubscriptionFilter against this destination. To enable this, the destination owner must call PutDestinationPolicy after PutDestination.

Arguments

destinationName = ::String – Required

A name for the destination.

targetArn = ::String – Required

The ARN of an Amazon Kinesis stream to deliver matching log events to.

roleArn = ::String – Required

The ARN of an IAM role that grants CloudWatch Logs permissions to call Amazon Kinesis PutRecord on the destination stream.

Returns

PutDestinationResponse

Exceptions

InvalidParameterException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.put_destination_policy — Function.

put_destination_policy(::AWSConfig, arguments::Dict)
put_destination_policy([::AWSConfig]; destinationName=, accessPolicy=)

PutDestinationPolicy Operation

Creates or updates an access policy associated with an existing destination. An access policy is an IAM policy document that is used to authorize claims to register a subscription filter against a given destination.

Arguments

destinationName = ::String – Required

A name for an existing destination.

accessPolicy = ::String – Required

An IAM policy document that authorizes cross-account users to deliver their log events to the associated destination.

Exceptions

InvalidParameterException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.put_log_events — Function.

put_log_events(::AWSConfig, arguments::Dict)
put_log_events([::AWSConfig]; logGroupName=, logStreamName=, logEvents=, <keyword arguments>)

PutLogEvents Operation

Uploads a batch of log events to the specified log stream.

You must include the sequence token obtained from the response of the previous call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token using DescribeLogStreams.

The batch of events must satisfy the following constraints:

The maximum batch size is 1,048,576 bytes, and this size is calculated as the sum of all event messages in UTF-8, plus 26 bytes for each log event.

None of the log events in the batch can be more than 2 hours in the future.

None of the log events in the batch can be older than 14 days or the retention period of the log group.

The log events in the batch must be in chronological ordered by their timestamp (the time the event occurred, expressed as the number of milliseconds since Jan 1, 1970 00:00:00 UTC).

The maximum number of log events in a batch is 10,000.

A batch of log events in a single request cannot span more than 24 hours. Otherwise, the operation fails.

Arguments

logGroupName = ::String – Required

The name of the log group.

logStreamName = ::String – Required

The name of the log stream.

logEvents = [Dict( ... )...] – Required

The log events.

 logEvents = [Dict(
        "timestamp" => <required> ::Int,
        "message" => <required> ::String
    )...]

sequenceToken = ::String

The sequence token.

Returns

PutLogEventsResponse

Exceptions

InvalidParameterException, InvalidSequenceTokenException, DataAlreadyAcceptedException, ResourceNotFoundException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.put_metric_filter — Function.

put_metric_filter(::AWSConfig, arguments::Dict)
put_metric_filter([::AWSConfig]; logGroupName=, filterName=, filterPattern=, metricTransformations=)

PutMetricFilter Operation

Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents.

The maximum number of metric filters that can be associated with a log group is 100.

Arguments

logGroupName = ::String – Required

The name of the log group.

filterName = ::String – Required

A name for the metric filter.

filterPattern = ::String – Required

A filter pattern for extracting metric data out of ingested log events.

metricTransformations = [Dict( ... )...] – Required

A collection of information needed to define how metric data gets emitted.

 metricTransformations = [Dict(
        "metricName" => <required> ::String,
        "metricNamespace" => <required> ::String,
        "metricValue" => <required> ::String,
        "defaultValue" =>  double
    )...]

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException, LimitExceededException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.put_retention_policy — Function.

put_retention_policy(::AWSConfig, arguments::Dict)
put_retention_policy([::AWSConfig]; logGroupName=, retentionInDays=)

PutRetentionPolicy Operation

Sets the retention of the specified log group. A retention policy allows you to configure the number of days you want to retain log events in the specified log group.

Arguments

logGroupName = ::String – Required

The name of the log group.

retentionInDays = ::Int – Required

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.put_subscription_filter — Function.

put_subscription_filter(::AWSConfig, arguments::Dict)
put_subscription_filter([::AWSConfig]; logGroupName=, filterName=, filterPattern=, destinationArn=, <keyword arguments>)

PutSubscriptionFilter Operation

Creates or updates a subscription filter and associates it with the specified log group. Subscription filters allow you to subscribe to a real-time stream of log events ingested through PutLogEvents and have them delivered to a specific destination. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination that belongs to a different account, for cross-account delivery.

An Amazon Kinesis Firehose stream that belongs to the same account as the subscription filter, for same-account delivery.

An AWS Lambda function that belongs to the same account as the subscription filter, for same-account delivery.

There can only be one subscription filter associated with a log group. If you are updating an existing filter, you must specify the correct name in filterName. Otherwise, the call will fail because you cannot associate a second filter with a log group.

Arguments

logGroupName = ::String – Required

The name of the log group.

filterName = ::String – Required

A name for the subscription filter. If you are updating an existing filter, you must specify the correct name in filterName. Otherwise, the call will fail because you cannot associate a second filter with a log group. To find the name of the filter currently associated with a log group, use DescribeSubscriptionFilters.

filterPattern = ::String – Required

A filter pattern for subscribing to a filtered stream of log events.

destinationArn = ::String – Required

The ARN of the destination to deliver matching log events to. Currently, the supported destinations are:

An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery.

A logical destination (specified using an ARN) belonging to a different account, for cross-account delivery.

An Amazon Kinesis Firehose stream belonging to the same account as the subscription filter, for same-account delivery.

An AWS Lambda function belonging to the same account as the subscription filter, for same-account delivery.

roleArn = ::String

The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination for cross-account delivery.

distribution = "Random" or "ByLogStream"

The method used to distribute log data to the destination, when the destination is an Amazon Kinesis stream. By default, log data is grouped by log stream. For a more even distribution, you can group log data randomly.

Exceptions

InvalidParameterException, ResourceNotFoundException, OperationAbortedException, LimitExceededException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.tag_log_group — Function.

tag_log_group(::AWSConfig, arguments::Dict)
tag_log_group([::AWSConfig]; logGroupName=, tags=)

TagLogGroup Operation

Adds or updates the specified tags for the specified log group.

To list the tags for a log group, use ListTagsLogGroup. To remove tags, use UntagLogGroup.

For more information about tags, see Tag Log Groups in Amazon CloudWatch Logs in the Amazon CloudWatch Logs User Guide.

Arguments

logGroupName = ::String – Required

The name of the log group.

tags = ::Dict{String,String} – Required

The key-value pairs to use for the tags.

Exceptions

ResourceNotFoundException or InvalidParameterException.

See also: AWS API Documentation

AmazonCloudWatchLogs.test_metric_filter — Function.

test_metric_filter(::AWSConfig, arguments::Dict)
test_metric_filter([::AWSConfig]; filterPattern=, logEventMessages=)

TestMetricFilter Operation

Tests the filter pattern of a metric filter against a sample of log event messages. You can use this operation to validate the correctness of a metric filter pattern.

Arguments

filterPattern = ::String – Required

logEventMessages = [::String...] – Required

The log event messages to test.

Returns

TestMetricFilterResponse

Exceptions

InvalidParameterException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonCloudWatchLogs.untag_log_group — Function.

untag_log_group(::AWSConfig, arguments::Dict)
untag_log_group([::AWSConfig]; logGroupName=, tags=)

UntagLogGroup Operation

Removes the specified tags from the specified log group.

To list the tags for a log group, use ListTagsLogGroup. To add tags, use UntagLogGroup.

Arguments

logGroupName = ::String – Required

The name of the log group.

tags = [::String...] – Required

The tag keys. The corresponding tags are removed from the log group.

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation