AmazonCodeBuild.jl

AmazonCodeBuild

AWS CodeBuild

AWS CodeBuild is a fully managed build service in the cloud. AWS CodeBuild compiles your source code, runs unit tests, and produces artifacts that are ready to deploy. AWS CodeBuild eliminates the need to provision, manage, and scale your own build servers. It provides prepackaged build environments for the most popular programming languages and build tools, such as Apach Maven, Gradle, and more. You can also fully customize build environments in AWS CodeBuild to use your own build tools. AWS CodeBuild scales automatically to meet peak build requests, and you pay only for the build time you consume. For more information about AWS CodeBuild, see the AWS CodeBuild User Guide.

AWS CodeBuild supports these operations:

BatchGetProjects: Gets information about one or more build projects. A build project defines how AWS CodeBuild will run a build. This includes information such as where to get the source code to build, the build environment to use, the build commands to run, and where to store the build output. A build environment represents a combination of operating system, programming language runtime, and tools that AWS CodeBuild will use to run a build. Also, you can add tags to build projects to help manage your resources and costs.

CreateProject: Creates a build project.

DeleteProject: Deletes a build project.

ListProjects: Gets a list of build project names, with each build project name representing a single build project.

UpdateProject: Changes the settings of an existing build project.

BatchGetBuilds: Gets information about one or more builds.

ListBuilds: Gets a list of build IDs, with each build ID representing a single build.

ListBuildsForProject: Gets a list of build IDs for the specified build project, with each build ID representing a single build.

StartBuild: Starts running a build.

StopBuild: Attempts to stop running a build.

ListCuratedEnvironmentImages: Gets information about Docker images that are managed by AWS CodeBuild.

This document is generated from apis/codebuild-2016-10-06.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCodeBuild.batch_get_builds

AmazonCodeBuild.batch_get_projects

AmazonCodeBuild.create_project

AmazonCodeBuild.delete_project

AmazonCodeBuild.list_builds

AmazonCodeBuild.list_builds_for_project

AmazonCodeBuild.list_curated_environment_images

AmazonCodeBuild.list_projects

AmazonCodeBuild.start_build

AmazonCodeBuild.stop_build

AmazonCodeBuild.update_project

AmazonCodeBuild.batch_get_builds — Function.

batch_get_builds(::AWSConfig, arguments::Dict)
batch_get_builds([::AWSConfig]; ids=)

BatchGetBuilds Operation

Gets information about builds.

Arguments

ids = [::String, ...] – Required

The IDs of the builds.

Returns

BatchGetBuildsOutput

Exceptions

InvalidInputException.

See also: AWS API Documentation

AmazonCodeBuild.batch_get_projects — Function.

batch_get_projects(::AWSConfig, arguments::Dict)
batch_get_projects([::AWSConfig]; names=)

BatchGetProjects Operation

Gets information about build projects.

Arguments

names = [::String, ...] – Required

The names of the build projects.

Returns

BatchGetProjectsOutput

Exceptions

InvalidInputException.

See also: AWS API Documentation

AmazonCodeBuild.create_project — Function.

create_project(::AWSConfig, arguments::Dict)
create_project([::AWSConfig]; name=, source=, artifacts=, environment=, <keyword arguments>)

CreateProject Operation

Creates a build project.

Arguments

name = ::String – Required

The name of the build project.

description = ::String

A description that makes the build project easy to identify.

source = [ ... ] – Required

Information about the build input source code for the build project.

 source = [
        "type" => <required> "CODECOMMIT", "CODEPIPELINE", "GITHUB" or "S3",
        "location" =>  ::String,
        "buildspec" =>  ::String,
        "auth" =>  [
            "type" => <required> "OAUTH",
            "resource" =>  ::String
        ]
    ]

artifacts = [ ... ] – Required

Information about the build output artifacts for the build project.

 artifacts = [
        "type" => <required> "CODEPIPELINE", "S3" or "NO_ARTIFACTS",
        "location" =>  ::String,
        "path" =>  ::String,
        "namespaceType" =>  "NONE" or "BUILD_ID",
        "name" =>  ::String,
        "packaging" =>  "NONE" or "ZIP"
    ]

environment = [ ... ] – Required

Information about the build environment for the build project.

 environment = [
        "type" => <required> "LINUX_CONTAINER",
        "image" => <required> ::String,
        "computeType" => <required> "BUILD_GENERAL1_SMALL", "BUILD_GENERAL1_MEDIUM" or "BUILD_GENERAL1_LARGE",
        "environmentVariables" =>  [[
            "name" => <required> ::String,
            "value" => <required> ::String
        ], ...],
        "privilegedMode" =>  ::Bool
    ]

serviceRole = ::String

The ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

timeoutInMinutes = ::Int

How long, in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait until timing out any build that has not been marked as completed. The default is 60 minutes.

encryptionKey = ::String

The AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts.

You can specify either the CMK's Amazon Resource Name (ARN) or, if available, the CMK's alias (using the format alias/*alias-name* ).

tags = [[ ... ], ...]

A set of tags for this build project.

These tags are available for use by AWS services that support AWS CodeBuild build project tags.

 tags = [[
        "key" =>  ::String,
        "value" =>  ::String
    ], ...]

Returns

CreateProjectOutput

Exceptions

InvalidInputException, ResourceAlreadyExistsException or AccountLimitExceededException.

See also: AWS API Documentation

AmazonCodeBuild.delete_project — Function.

delete_project(::AWSConfig, arguments::Dict)
delete_project([::AWSConfig]; name=)

DeleteProject Operation

Deletes a build project.

Arguments

name = ::String – Required

The name of the build project.

Returns

DeleteProjectOutput

Exceptions

InvalidInputException.

See also: AWS API Documentation

AmazonCodeBuild.list_builds — Function.

list_builds(::AWSConfig, arguments::Dict)
list_builds([::AWSConfig]; <keyword arguments>)

ListBuilds Operation

Gets a list of build IDs, with each build ID representing a single build.

Arguments

sortOrder = "ASCENDING" or "DESCENDING"

The order to list build IDs. Valid values include:

ASCENDING: List the build IDs in ascending order by build ID.

DESCENDING: List the build IDs in descending order by build ID.

nextToken = ::String

During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

Returns

ListBuildsOutput

Exceptions

InvalidInputException.

See also: AWS API Documentation

AmazonCodeBuild.list_builds_for_project — Function.

list_builds_for_project(::AWSConfig, arguments::Dict)
list_builds_for_project([::AWSConfig]; projectName=, <keyword arguments>)

ListBuildsForProject Operation

Gets a list of build IDs for the specified build project, with each build ID representing a single build.

Arguments

projectName = ::String – Required

The name of the build project.

sortOrder = "ASCENDING" or "DESCENDING"

The order to list build IDs. Valid values include:

ASCENDING: List the build IDs in ascending order by build ID.

DESCENDING: List the build IDs in descending order by build ID.

nextToken = ::String

During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

Returns

ListBuildsForProjectOutput

Exceptions

InvalidInputException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCodeBuild.list_curated_environment_images — Function.

list_curated_environment_images(::AWSConfig, arguments::Dict)
list_curated_environment_images([::AWSConfig]; )

ListCuratedEnvironmentImages Operation

Gets information about Docker images that are managed by AWS CodeBuild.

Arguments

Returns

ListCuratedEnvironmentImagesOutput

See also: AWS API Documentation

AmazonCodeBuild.list_projects — Function.

list_projects(::AWSConfig, arguments::Dict)
list_projects([::AWSConfig]; <keyword arguments>)

ListProjects Operation

Gets a list of build project names, with each build project name representing a single build project.

Arguments

sortBy = "NAME", "CREATED_TIME" or "LAST_MODIFIED_TIME"

The criterion to be used to list build project names. Valid values include:

CREATED_TIME: List the build project names based on when each build project was created.

LAST_MODIFIED_TIME: List the build project names based on when information about each build project was last changed.

NAME: List the build project names based on each build project's name.

Use sortOrder to specify in what order to list the build project names based on the preceding criteria.

sortOrder = "ASCENDING" or "DESCENDING"

The order in which to list build projects. Valid values include:

ASCENDING: List the build project names in ascending order.

DESCENDING: List the build project names in descending order.

Use sortBy to specify the criterion to be used to list build project names.

nextToken = ::String

During a previous call, if there are more than 100 items in the list, only the first 100 items are returned, along with a unique string called a next token. To get the next batch of items in the list, call this operation again, adding the next token to the call. To get all of the items in the list, keep calling this operation with each subsequent next token that is returned, until no more next tokens are returned.

Returns

ListProjectsOutput

Exceptions

InvalidInputException.

See also: AWS API Documentation

AmazonCodeBuild.start_build — Function.

start_build(::AWSConfig, arguments::Dict)
start_build([::AWSConfig]; projectName=, <keyword arguments>)

StartBuild Operation

Starts running a build.

Arguments

projectName = ::String – Required

The name of the build project to start running a build.

sourceVersion = ::String

A version of the build input to be built, for this build only. If not specified, the latest version will be used. If specified, must be one of:

For AWS CodeCommit or GitHub: the commit ID to use.

For Amazon Simple Storage Service (Amazon S3): the version ID of the object representing the build input ZIP file to use.

artifactsOverride = [ ... ]

Build output artifact settings that override, for this build only, the latest ones already defined in the build project.

 artifactsOverride = [
        "type" => <required> "CODEPIPELINE", "S3" or "NO_ARTIFACTS",
        "location" =>  ::String,
        "path" =>  ::String,
        "namespaceType" =>  "NONE" or "BUILD_ID",
        "name" =>  ::String,
        "packaging" =>  "NONE" or "ZIP"
    ]

environmentVariablesOverride = [[ ... ], ...]

A set of environment variables that overrides, for this build only, the latest ones already defined in the build project.

 environmentVariablesOverride = [[
        "name" => <required> ::String,
        "value" => <required> ::String
    ], ...]

buildspecOverride = ::String

A build spec declaration that overrides, for this build only, the latest one already defined in the build project.

timeoutInMinutesOverride = ::Int

The number of build timeout minutes, from 5 to 480 (8 hours), that overrides, for this build only, the latest setting already defined in the build project.

Returns

StartBuildOutput

Exceptions

InvalidInputException, ResourceNotFoundException or AccountLimitExceededException.

See also: AWS API Documentation

AmazonCodeBuild.stop_build — Function.

stop_build(::AWSConfig, arguments::Dict)
stop_build([::AWSConfig]; id=)

StopBuild Operation

Attempts to stop running a build.

Arguments

id = ::String – Required

The ID of the build.

Returns

StopBuildOutput

Exceptions

InvalidInputException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonCodeBuild.update_project — Function.

update_project(::AWSConfig, arguments::Dict)
update_project([::AWSConfig]; name=, <keyword arguments>)

UpdateProject Operation

Changes the settings of a build project.

Arguments

name = ::String – Required

The name of the build project.

Note

You cannot change a build project's name.

description = ::String

A new or replacement description of the build project.

source = [ ... ]

Information to be changed about the build input source code for the build project.

 source = [
        "type" => <required> "CODECOMMIT", "CODEPIPELINE", "GITHUB" or "S3",
        "location" =>  ::String,
        "buildspec" =>  ::String,
        "auth" =>  [
            "type" => <required> "OAUTH",
            "resource" =>  ::String
        ]
    ]

artifacts = [ ... ]

Information to be changed about the build output artifacts for the build project.

 artifacts = [
        "type" => <required> "CODEPIPELINE", "S3" or "NO_ARTIFACTS",
        "location" =>  ::String,
        "path" =>  ::String,
        "namespaceType" =>  "NONE" or "BUILD_ID",
        "name" =>  ::String,
        "packaging" =>  "NONE" or "ZIP"
    ]

environment = [ ... ]

Information to be changed about the build environment for the build project.

 environment = [
        "type" => <required> "LINUX_CONTAINER",
        "image" => <required> ::String,
        "computeType" => <required> "BUILD_GENERAL1_SMALL", "BUILD_GENERAL1_MEDIUM" or "BUILD_GENERAL1_LARGE",
        "environmentVariables" =>  [[
            "name" => <required> ::String,
            "value" => <required> ::String
        ], ...],
        "privilegedMode" =>  ::Bool
    ]

serviceRole = ::String

The replacement ARN of the AWS Identity and Access Management (IAM) role that enables AWS CodeBuild to interact with dependent AWS services on behalf of the AWS account.

timeoutInMinutes = ::Int

The replacement value in minutes, from 5 to 480 (8 hours), for AWS CodeBuild to wait before timing out any related build that did not get marked as completed.

encryptionKey = ::String

The replacement AWS Key Management Service (AWS KMS) customer master key (CMK) to be used for encrypting the build output artifacts.

You can specify either the CMK's Amazon Resource Name (ARN) or, if available, the CMK's alias (using the format alias/*alias-name* ).

tags = [[ ... ], ...]

The replacement set of tags for this build project.

These tags are available for use by AWS services that support AWS CodeBuild build project tags.

 tags = [[
        "key" =>  ::String,
        "value" =>  ::String
    ], ...]

Returns

UpdateProjectOutput

Exceptions

InvalidInputException or ResourceNotFoundException.

See also: AWS API Documentation