AmazonCodeCommit.jl

AmazonCodeCommit

AWS CodeCommit

This is the AWS CodeCommit API Reference. This reference provides descriptions of the operations and data types for AWS CodeCommit API along with usage examples.

You can use the AWS CodeCommit API to work with the following objects:

Repositories, by calling the following:

BatchGetRepositories, which returns information about one or more repositories associated with your AWS account

CreateRepository, which creates an AWS CodeCommit repository

DeleteRepository, which deletes an AWS CodeCommit repository

GetRepository, which returns information about a specified repository

ListRepositories, which lists all AWS CodeCommit repositories associated with your AWS account

UpdateRepositoryDescription, which sets or updates the description of the repository

UpdateRepositoryName, which changes the name of the repository. If you change the name of a repository, no other users of that repository will be able to access it until you send them the new HTTPS or SSH URL to use.

Branches, by calling the following:

CreateBranch, which creates a new branch in a specified repository

GetBranch, which returns information about a specified branch

ListBranches, which lists all branches for a specified repository

UpdateDefaultBranch, which changes the default branch for a repository

Information about committed code in a repository, by calling the following:

GetBlob, which returns the base-64 encoded content of an individual Git blob object within a repository

GetCommit, which returns information about a commit, including commit messages and author and committer information

GetDifferences, which returns information about the differences in a valid commit specifier (such as a branch, tag, HEAD, commit ID or other fully qualified reference)

Triggers, by calling the following:

GetRepositoryTriggers, which returns information about triggers configured for a repository

PutRepositoryTriggers, which replaces all triggers for a repository and can be used to create or delete triggers

TestRepositoryTriggers, which tests the functionality of a repository trigger by sending data to the trigger target

For information about how to use AWS CodeCommit, see the AWS CodeCommit User Guide.

This document is generated from apis/codecommit-2015-04-13.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCodeCommit.batch_get_repositories

AmazonCodeCommit.create_branch

AmazonCodeCommit.create_repository

AmazonCodeCommit.delete_repository

AmazonCodeCommit.get_blob

AmazonCodeCommit.get_branch

AmazonCodeCommit.get_commit

AmazonCodeCommit.get_differences

AmazonCodeCommit.get_repository

AmazonCodeCommit.get_repository_triggers

AmazonCodeCommit.list_branches

AmazonCodeCommit.list_repositories

AmazonCodeCommit.put_repository_triggers

AmazonCodeCommit.test_repository_triggers

AmazonCodeCommit.update_default_branch

AmazonCodeCommit.update_repository_description

AmazonCodeCommit.update_repository_name

AmazonCodeCommit.batch_get_repositories — Function.

batch_get_repositories(::AWSConfig, arguments::Dict)
batch_get_repositories([::AWSConfig]; repositoryNames=)

BatchGetRepositories Operation

Returns information about one or more repositories.

Note

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

Arguments

repositoryNames = [::String, ...] – Required

The names of the repositories to get information about.

Returns

BatchGetRepositoriesOutput

Exceptions

RepositoryNamesRequiredException, MaximumRepositoryNamesExceededException, InvalidRepositoryNameException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.create_branch — Function.

create_branch(::AWSConfig, arguments::Dict)
create_branch([::AWSConfig]; repositoryName=, branchName=, commitId=)

CreateBranch Operation

Creates a new branch in a repository and points the branch to a commit.

Note

Calling the create branch operation does not set a repository's default branch. To do this, call the update default branch operation.

Arguments

repositoryName = ::String – Required

The name of the repository in which you want to create the new branch.

branchName = ::String – Required

The name of the new branch to create.

commitId = ::String – Required

The ID of the commit to point the new branch to.

Exceptions

RepositoryNameRequiredException, InvalidRepositoryNameException, RepositoryDoesNotExistException, BranchNameRequiredException, BranchNameExistsException, InvalidBranchNameException, CommitIdRequiredException, CommitDoesNotExistException, InvalidCommitIdException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.create_repository — Function.

create_repository(::AWSConfig, arguments::Dict)
create_repository([::AWSConfig]; repositoryName=, <keyword arguments>)

CreateRepository Operation

Creates a new, empty repository.

Arguments

repositoryName = ::String – Required

The name of the new repository to be created.

Note

The repository name must be unique across the calling AWS account. In addition, repository names are limited to 100 alphanumeric, dash, and underscore characters, and cannot include certain characters. For a full description of the limits on repository names, see Limits in the AWS CodeCommit User Guide. The suffix ".git" is prohibited.

repositoryDescription = ::String

A comment or description about the new repository.

Note

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

Returns

CreateRepositoryOutput

Exceptions

RepositoryNameExistsException, RepositoryNameRequiredException, InvalidRepositoryNameException, InvalidRepositoryDescriptionException, RepositoryLimitExceededException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.delete_repository — Function.

delete_repository(::AWSConfig, arguments::Dict)
delete_repository([::AWSConfig]; repositoryName=)

DeleteRepository Operation

Deletes a repository. If a specified repository was already deleted, a null repository ID will be returned.

Important

Deleting a repository also deletes all associated objects and metadata. After a repository is deleted, all future push calls to the deleted repository will fail.

Arguments

repositoryName = ::String – Required

The name of the repository to delete.

Returns

DeleteRepositoryOutput

Exceptions

RepositoryNameRequiredException, InvalidRepositoryNameException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.get_blob — Function.

get_blob(::AWSConfig, arguments::Dict)
get_blob([::AWSConfig]; repositoryName=, blobId=)

GetBlob Operation

Returns the base-64 encoded content of an individual blob within a repository.

Arguments

repositoryName = ::String – Required

The name of the repository that contains the blob.

blobId = ::String – Required

The ID of the blob, which is its SHA-1 pointer.

Returns

GetBlobOutput

Exceptions

RepositoryNameRequiredException, InvalidRepositoryNameException, RepositoryDoesNotExistException, BlobIdRequiredException, InvalidBlobIdException, BlobIdDoesNotExistException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException, EncryptionKeyUnavailableException or FileTooLargeException.

See also: AWS API Documentation

AmazonCodeCommit.get_branch — Function.

get_branch(::AWSConfig, arguments::Dict)
get_branch([::AWSConfig]; <keyword arguments>)

GetBranch Operation

Returns information about a repository branch, including its name and the last commit ID.

Arguments

repositoryName = ::String

The name of the repository that contains the branch for which you want to retrieve information.

branchName = ::String

The name of the branch for which you want to retrieve information.

Returns

GetBranchOutput

Exceptions

RepositoryNameRequiredException, RepositoryDoesNotExistException, InvalidRepositoryNameException, BranchNameRequiredException, InvalidBranchNameException, BranchDoesNotExistException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.get_commit — Function.

get_commit(::AWSConfig, arguments::Dict)
get_commit([::AWSConfig]; repositoryName=, commitId=)

GetCommit Operation

Returns information about a commit, including commit message and committer information.

Arguments

repositoryName = ::String – Required

The name of the repository to which the commit was made.

commitId = ::String – Required

The commit ID.

Returns

GetCommitOutput

Exceptions

RepositoryNameRequiredException, InvalidRepositoryNameException, RepositoryDoesNotExistException, CommitIdRequiredException, InvalidCommitIdException, CommitIdDoesNotExistException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.get_differences — Function.

get_differences(::AWSConfig, arguments::Dict)
get_differences([::AWSConfig]; repositoryName=, afterCommitSpecifier=, <keyword arguments>)

GetDifferences Operation

Returns information about the differences in a valid commit specifier (such as a branch, tag, HEAD, commit ID or other fully qualified reference). Results can be limited to a specified path.

Arguments

repositoryName = ::String – Required

The name of the repository where you want to get differences.

beforeCommitSpecifier = ::String

The branch, tag, HEAD, or other fully qualified reference used to identify a commit. For example, the full commit ID. Optional. If not specified, all changes prior to the afterCommitSpecifier value will be shown. If you do not use beforeCommitSpecifier in your request, consider limiting the results with maxResults.

afterCommitSpecifier = ::String – Required

The branch, tag, HEAD, or other fully qualified reference used to identify a commit.

beforePath = ::String

The file path in which to check for differences. Limits the results to this path. Can also be used to specify the previous name of a directory or folder. If beforePath and afterPath are not specified, differences will be shown for all paths.

afterPath = ::String

The file path in which to check differences. Limits the results to this path. Can also be used to specify the changed name of a directory or folder, if it has changed. If not specified, differences will be shown for all paths.

MaxResults = ::Int

A non-negative integer used to limit the number of returned results.

NextToken = ::String

An enumeration token that when provided in a request, returns the next batch of the results.

Returns

GetDifferencesOutput

Exceptions

RepositoryNameRequiredException, RepositoryDoesNotExistException, InvalidRepositoryNameException, InvalidContinuationTokenException, InvalidMaxResultsException, InvalidCommitIdException, CommitRequiredException, InvalidCommitException, CommitDoesNotExistException, InvalidPathException, PathDoesNotExistException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.get_repository — Function.

get_repository(::AWSConfig, arguments::Dict)
get_repository([::AWSConfig]; repositoryName=)

GetRepository Operation

Returns information about a repository.

Note

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

Arguments

repositoryName = ::String – Required

The name of the repository to get information about.

Returns

GetRepositoryOutput

Exceptions

RepositoryNameRequiredException, RepositoryDoesNotExistException, InvalidRepositoryNameException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.get_repository_triggers — Function.

get_repository_triggers(::AWSConfig, arguments::Dict)
get_repository_triggers([::AWSConfig]; repositoryName=)

GetRepositoryTriggers Operation

Gets information about triggers configured for a repository.

Arguments

repositoryName = ::String – Required

The name of the repository for which the trigger is configured.

Returns

GetRepositoryTriggersOutput

Exceptions

RepositoryNameRequiredException, InvalidRepositoryNameException, RepositoryDoesNotExistException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.list_branches — Function.

list_branches(::AWSConfig, arguments::Dict)
list_branches([::AWSConfig]; repositoryName=, <keyword arguments>)

ListBranches Operation

Gets information about one or more branches in a repository.

Arguments

repositoryName = ::String – Required

The name of the repository that contains the branches.

nextToken = ::String

An enumeration token that allows the operation to batch the results.

Returns

ListBranchesOutput

Exceptions

RepositoryNameRequiredException, RepositoryDoesNotExistException, InvalidRepositoryNameException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException, EncryptionKeyUnavailableException or InvalidContinuationTokenException.

See also: AWS API Documentation

AmazonCodeCommit.list_repositories — Function.

list_repositories(::AWSConfig, arguments::Dict)
list_repositories([::AWSConfig]; <keyword arguments>)

ListRepositories Operation

Gets information about one or more repositories.

Arguments

nextToken = ::String

An enumeration token that allows the operation to batch the results of the operation. Batch sizes are 1,000 for list repository operations. When the client sends the token back to AWS CodeCommit, another page of 1,000 records is retrieved.

sortBy = "repositoryName" or "lastModifiedDate"

The criteria used to sort the results of a list repositories operation.

order = "ascending" or "descending"

The order in which to sort the results of a list repositories operation.

Returns

ListRepositoriesOutput

Exceptions

InvalidSortByException, InvalidOrderException or InvalidContinuationTokenException.

See also: AWS API Documentation

AmazonCodeCommit.put_repository_triggers — Function.

put_repository_triggers(::AWSConfig, arguments::Dict)
put_repository_triggers([::AWSConfig]; repositoryName=, triggers=)

PutRepositoryTriggers Operation

Replaces all triggers for a repository. This can be used to create or delete triggers.

Arguments

repositoryName = ::String – Required

The name of the repository where you want to create or update the trigger.

triggers = [[ ... ], ...] – Required

The JSON block of configuration information for each trigger.

 triggers = [[
        "name" => <required> ::String,
        "destinationArn" => <required> ::String,
        "customData" =>  ::String,
        "branches" =>  [::String, ...],
        "events" => <required> ["all", "updateReference", "createReference" or "deleteReference", ...]
    ], ...]

Returns

PutRepositoryTriggersOutput

Exceptions

RepositoryDoesNotExistException, RepositoryNameRequiredException, InvalidRepositoryNameException, RepositoryTriggersListRequiredException, MaximumRepositoryTriggersExceededException, InvalidRepositoryTriggerNameException, InvalidRepositoryTriggerDestinationArnException, InvalidRepositoryTriggerRegionException, InvalidRepositoryTriggerCustomDataException, MaximumBranchesExceededException, InvalidRepositoryTriggerBranchNameException, InvalidRepositoryTriggerEventsException, RepositoryTriggerNameRequiredException, RepositoryTriggerDestinationArnRequiredException, RepositoryTriggerBranchNameListRequiredException, RepositoryTriggerEventsListRequiredException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.test_repository_triggers — Function.

test_repository_triggers(::AWSConfig, arguments::Dict)
test_repository_triggers([::AWSConfig]; repositoryName=, triggers=)

TestRepositoryTriggers Operation

Tests the functionality of repository triggers by sending information to the trigger target. If real data is available in the repository, the test will send data from the last commit. If no data is available, sample data will be generated.

Arguments

repositoryName = ::String – Required

The name of the repository in which to test the triggers.

triggers = [[ ... ], ...] – Required

The list of triggers to test.

 triggers = [[
        "name" => <required> ::String,
        "destinationArn" => <required> ::String,
        "customData" =>  ::String,
        "branches" =>  [::String, ...],
        "events" => <required> ["all", "updateReference", "createReference" or "deleteReference", ...]
    ], ...]

Returns

TestRepositoryTriggersOutput

Exceptions

RepositoryDoesNotExistException, RepositoryNameRequiredException, InvalidRepositoryNameException, RepositoryTriggersListRequiredException, MaximumRepositoryTriggersExceededException, InvalidRepositoryTriggerNameException, InvalidRepositoryTriggerDestinationArnException, InvalidRepositoryTriggerRegionException, InvalidRepositoryTriggerCustomDataException, MaximumBranchesExceededException, InvalidRepositoryTriggerBranchNameException, InvalidRepositoryTriggerEventsException, RepositoryTriggerNameRequiredException, RepositoryTriggerDestinationArnRequiredException, RepositoryTriggerBranchNameListRequiredException, RepositoryTriggerEventsListRequiredException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.update_default_branch — Function.

update_default_branch(::AWSConfig, arguments::Dict)
update_default_branch([::AWSConfig]; repositoryName=, defaultBranchName=)

UpdateDefaultBranch Operation

Sets or changes the default branch name for the specified repository.

Note

If you use this operation to change the default branch name to the current default branch name, a success message is returned even though the default branch did not change.

Arguments

repositoryName = ::String – Required

The name of the repository to set or change the default branch for.

defaultBranchName = ::String – Required

The name of the branch to set as the default.

Exceptions

RepositoryNameRequiredException, RepositoryDoesNotExistException, InvalidRepositoryNameException, BranchNameRequiredException, InvalidBranchNameException, BranchDoesNotExistException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.update_repository_description — Function.

update_repository_description(::AWSConfig, arguments::Dict)
update_repository_description([::AWSConfig]; repositoryName=, <keyword arguments>)

UpdateRepositoryDescription Operation

Sets or changes the comment or description for a repository.

Note

The description field for a repository accepts all HTML characters and all valid Unicode characters. Applications that do not HTML-encode the description and display it in a web page could expose users to potentially malicious code. Make sure that you HTML-encode the description field in any application that uses this API to display the repository description on a web page.

Arguments

repositoryName = ::String – Required

The name of the repository to set or change the comment or description for.

repositoryDescription = ::String

The new comment or description for the specified repository. Repository descriptions are limited to 1,000 characters.

Exceptions

RepositoryNameRequiredException, RepositoryDoesNotExistException, InvalidRepositoryNameException, InvalidRepositoryDescriptionException, EncryptionIntegrityChecksFailedException, EncryptionKeyAccessDeniedException, EncryptionKeyDisabledException, EncryptionKeyNotFoundException or EncryptionKeyUnavailableException.

See also: AWS API Documentation

AmazonCodeCommit.update_repository_name — Function.

update_repository_name(::AWSConfig, arguments::Dict)
update_repository_name([::AWSConfig]; oldName=, newName=)

UpdateRepositoryName Operation

Renames a repository. The repository name must be unique across the calling AWS account. In addition, repository names are limited to 100 alphanumeric, dash, and underscore characters, and cannot include certain characters. The suffix ".git" is prohibited. For a full description of the limits on repository names, see Limits in the AWS CodeCommit User Guide.

Arguments

oldName = ::String – Required

The existing name of the repository.

newName = ::String – Required

The new name for the repository.

Exceptions

RepositoryDoesNotExistException, RepositoryNameExistsException, RepositoryNameRequiredException or InvalidRepositoryNameException.

See also: AWS API Documentation