AmazonCodeDeploy.jl

AmazonCodeDeploy

AWS CodeDeploy

AWS CodeDeploy is a deployment service that automates application deployments to Amazon EC2 instances or on-premises instances running in your own facility.

You can deploy a nearly unlimited variety of application content, such as code, web and configuration files, executables, packages, scripts, multimedia files, and so on. AWS CodeDeploy can deploy application content stored in Amazon S3 buckets, GitHub repositories, or Bitbucket repositories. You do not need to make changes to your existing code before you can use AWS CodeDeploy.

AWS CodeDeploy makes it easier for you to rapidly release new features, helps you avoid downtime during application deployment, and handles the complexity of updating your applications, without many of the risks associated with error-prone manual deployments.

AWS CodeDeploy Components

Use the information in this guide to help you work with the following AWS CodeDeploy components:

Application: A name that uniquely identifies the application you want to deploy. AWS CodeDeploy uses this name, which functions as a container, to ensure the correct combination of revision, deployment configuration, and deployment group are referenced during a deployment.

Deployment group: A set of individual instances. A deployment group contains individually tagged instances, Amazon EC2 instances in Auto Scaling groups, or both.

Deployment configuration: A set of deployment rules and deployment success and failure conditions used by AWS CodeDeploy during a deployment.

Deployment: The process, and the components involved in the process, of installing content on one or more instances.

Application revisions: An archive file containing source content—source code, web pages, executable files, and deployment scripts—along with an application specification file (AppSpec file). Revisions are stored in Amazon S3 buckets or GitHub repositories. For Amazon S3, a revision is uniquely identified by its Amazon S3 object key and its ETag, version, or both. For GitHub, a revision is uniquely identified by its commit ID.

This guide also contains information to help you get details about the instances in your deployments and to make on-premises instances available for AWS CodeDeploy deployments.

AWS CodeDeploy Information Resources

AWS CodeDeploy User Guide

AWS CodeDeploy API Reference Guide

AWS CLI Reference for AWS CodeDeploy

AWS CodeDeploy Developer Forum

This document is generated from apis/codedeploy-2014-10-06.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCodeDeploy.add_tags_to_on_premises_instances

AmazonCodeDeploy.batch_get_application_revisions

AmazonCodeDeploy.batch_get_applications

AmazonCodeDeploy.batch_get_deployment_groups

AmazonCodeDeploy.batch_get_deployment_instances

AmazonCodeDeploy.batch_get_deployments

AmazonCodeDeploy.batch_get_on_premises_instances

AmazonCodeDeploy.continue_deployment

AmazonCodeDeploy.create_application

AmazonCodeDeploy.create_deployment

AmazonCodeDeploy.create_deployment_config

AmazonCodeDeploy.create_deployment_group

AmazonCodeDeploy.delete_application

AmazonCodeDeploy.delete_deployment_config

AmazonCodeDeploy.delete_deployment_group

AmazonCodeDeploy.deregister_on_premises_instance

AmazonCodeDeploy.get_application

AmazonCodeDeploy.get_application_revision

AmazonCodeDeploy.get_deployment

AmazonCodeDeploy.get_deployment_config

AmazonCodeDeploy.get_deployment_group

AmazonCodeDeploy.get_deployment_instance

AmazonCodeDeploy.get_on_premises_instance

AmazonCodeDeploy.list_application_revisions

AmazonCodeDeploy.list_applications

AmazonCodeDeploy.list_deployment_configs

AmazonCodeDeploy.list_deployment_groups

AmazonCodeDeploy.list_deployment_instances

AmazonCodeDeploy.list_deployments

AmazonCodeDeploy.list_git_hub_account_token_names

AmazonCodeDeploy.list_on_premises_instances

AmazonCodeDeploy.register_application_revision

AmazonCodeDeploy.register_on_premises_instance

AmazonCodeDeploy.remove_tags_from_on_premises_instances

AmazonCodeDeploy.skip_wait_time_for_instance_termination

AmazonCodeDeploy.stop_deployment

AmazonCodeDeploy.update_application

AmazonCodeDeploy.update_deployment_group

AmazonCodeDeploy.add_tags_to_on_premises_instances — Function.

using AmazonCodeDeploy
add_tags_to_on_premises_instances([::AWSConfig], arguments::Dict)
add_tags_to_on_premises_instances([::AWSConfig]; tags=, instanceNames=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"AddTagsToOnPremisesInstances", args)

AddTagsToOnPremisesInstances Operation

Adds tags to on-premises instances.

Arguments

tags = [[ ... ], ...] – Required

The tag key-value pairs to add to the on-premises instances.

Keys and values are both required. Keys cannot be null or empty strings. Value-only tags are not allowed.

 tags = [[
        "Key" =>  ::String,
        "Value" =>  ::String
    ], ...]

instanceNames = [::String, ...] – Required

The names of the on-premises instances to which to add tags.

Exceptions

InstanceNameRequiredException, TagRequiredException, InvalidTagException, TagLimitExceededException, InstanceLimitExceededException or InstanceNotRegisteredException.

See also: AWS API Documentation

AmazonCodeDeploy.batch_get_application_revisions — Function.

using AmazonCodeDeploy
batch_get_application_revisions([::AWSConfig], arguments::Dict)
batch_get_application_revisions([::AWSConfig]; applicationName=, revisions=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"BatchGetApplicationRevisions", args)

BatchGetApplicationRevisions Operation

Gets information about one or more application revisions.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application about which to get revision information.

revisions = [[ ... ], ...] – Required

Information to get about the application revisions, including type and location.

 revisions = [[
        "revisionType" =>  "S3" or "GitHub",
        "s3Location" =>  [
            "bucket" =>  ::String,
            "key" =>  ::String,
            "bundleType" =>  "tar", "tgz" or "zip",
            "version" =>  ::String,
            "eTag" =>  ::String
        ],
        "gitHubLocation" =>  [
            "repository" =>  ::String,
            "commitId" =>  ::String
        ]
    ], ...]

Returns

BatchGetApplicationRevisionsOutput

Exceptions

ApplicationDoesNotExistException, ApplicationNameRequiredException, InvalidApplicationNameException, RevisionRequiredException, InvalidRevisionException or BatchLimitExceededException.

See also: AWS API Documentation

AmazonCodeDeploy.batch_get_applications — Function.

using AmazonCodeDeploy
batch_get_applications([::AWSConfig], arguments::Dict)
batch_get_applications([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"BatchGetApplications", args)

BatchGetApplications Operation

Gets information about one or more applications.

Arguments

applicationNames = [::String, ...]

A list of application names separated by spaces.

Returns

BatchGetApplicationsOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationDoesNotExistException or BatchLimitExceededException.

See also: AWS API Documentation

AmazonCodeDeploy.batch_get_deployment_groups — Function.

using AmazonCodeDeploy
batch_get_deployment_groups([::AWSConfig], arguments::Dict)
batch_get_deployment_groups([::AWSConfig]; applicationName=, deploymentGroupNames=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"BatchGetDeploymentGroups", args)

BatchGetDeploymentGroups Operation

Gets information about one or more deployment groups.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

deploymentGroupNames = [::String, ...] – Required

The deployment groups' names.

Returns

BatchGetDeploymentGroupsOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationDoesNotExistException, DeploymentGroupNameRequiredException, InvalidDeploymentGroupNameException or BatchLimitExceededException.

See also: AWS API Documentation

AmazonCodeDeploy.batch_get_deployment_instances — Function.

using AmazonCodeDeploy
batch_get_deployment_instances([::AWSConfig], arguments::Dict)
batch_get_deployment_instances([::AWSConfig]; deploymentId=, instanceIds=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"BatchGetDeploymentInstances", args)

BatchGetDeploymentInstances Operation

Gets information about one or more instance that are part of a deployment group.

Arguments

deploymentId = ::String – Required

The unique ID of a deployment.

instanceIds = [::String, ...] – Required

The unique IDs of instances in the deployment group.

Returns

BatchGetDeploymentInstancesOutput

Exceptions

DeploymentIdRequiredException, DeploymentDoesNotExistException, InstanceIdRequiredException, InvalidDeploymentIdException, InvalidInstanceNameException or BatchLimitExceededException.

See also: AWS API Documentation

AmazonCodeDeploy.batch_get_deployments — Function.

using AmazonCodeDeploy
batch_get_deployments([::AWSConfig], arguments::Dict)
batch_get_deployments([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"BatchGetDeployments", args)

BatchGetDeployments Operation

Gets information about one or more deployments.

Arguments

deploymentIds = [::String, ...]

A list of deployment IDs, separated by spaces.

Returns

BatchGetDeploymentsOutput

Exceptions

DeploymentIdRequiredException, InvalidDeploymentIdException or BatchLimitExceededException.

See also: AWS API Documentation

AmazonCodeDeploy.batch_get_on_premises_instances — Function.

using AmazonCodeDeploy
batch_get_on_premises_instances([::AWSConfig], arguments::Dict)
batch_get_on_premises_instances([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"BatchGetOnPremisesInstances", args)

BatchGetOnPremisesInstances Operation

Gets information about one or more on-premises instances.

Arguments

instanceNames = [::String, ...]

The names of the on-premises instances about which to get information.

Returns

BatchGetOnPremisesInstancesOutput

Exceptions

InstanceNameRequiredException, InvalidInstanceNameException or BatchLimitExceededException.

See also: AWS API Documentation

AmazonCodeDeploy.continue_deployment — Function.

using AmazonCodeDeploy
continue_deployment([::AWSConfig], arguments::Dict)
continue_deployment([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ContinueDeployment", args)

ContinueDeployment Operation

For a blue/green deployment, starts the process of rerouting traffic from instances in the original environment to instances in the replacement environment without waiting for a specified wait time to elapse. (Traffic rerouting, which is achieved by registering instances in the replacement environment with the load balancer, can start as soon as all instances have a status of Ready.)

Arguments

deploymentId = ::String

The deployment ID of the blue/green deployment for which you want to start rerouting traffic to the replacement environment.

Exceptions

DeploymentIdRequiredException, DeploymentDoesNotExistException, DeploymentAlreadyCompletedException, InvalidDeploymentIdException, DeploymentIsNotInReadyStateException or UnsupportedActionForDeploymentTypeException.

See also: AWS API Documentation

AmazonCodeDeploy.create_application — Function.

using AmazonCodeDeploy
create_application([::AWSConfig], arguments::Dict)
create_application([::AWSConfig]; applicationName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"CreateApplication", args)

CreateApplication Operation

Creates an application.

Arguments

applicationName = ::String – Required

The name of the application. This name must be unique with the applicable IAM user or AWS account.

Returns

CreateApplicationOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationAlreadyExistsException or ApplicationLimitExceededException.

See also: AWS API Documentation

AmazonCodeDeploy.create_deployment — Function.

using AmazonCodeDeploy
create_deployment([::AWSConfig], arguments::Dict)
create_deployment([::AWSConfig]; applicationName=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"CreateDeployment", args)

CreateDeployment Operation

Deploys an application revision through the specified deployment group.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

deploymentGroupName = ::String

The name of the deployment group.

revision = [ ... ]

The type and location of the revision to deploy.

 revision = [
        "revisionType" =>  "S3" or "GitHub",
        "s3Location" =>  [
            "bucket" =>  ::String,
            "key" =>  ::String,
            "bundleType" =>  "tar", "tgz" or "zip",
            "version" =>  ::String,
            "eTag" =>  ::String
        ],
        "gitHubLocation" =>  [
            "repository" =>  ::String,
            "commitId" =>  ::String
        ]
    ]

deploymentConfigName = ::String

The name of a deployment configuration associated with the applicable IAM user or AWS account.

If not specified, the value configured in the deployment group will be used as the default. If the deployment group does not have a deployment configuration associated with it, then CodeDeployDefault.OneAtATime will be used by default.

description = ::String

A comment about the deployment.

ignoreApplicationStopFailures = ::Bool

If set to true, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will not be considered to have failed at that point and will continue on to the BeforeInstall deployment lifecycle event.

If set to false or not specified, then if the deployment causes the ApplicationStop deployment lifecycle event to fail to an instance, the deployment to that instance will stop, and the deployment to that instance will be considered to have failed.

targetInstances = [ ... ]

Information about the instances that will belong to the replacement environment in a blue/green deployment.

 targetInstances = [
        "tagFilters" =>  [[
            "Key" =>  ::String,
            "Value" =>  ::String,
            "Type" =>  "KEY_ONLY", "VALUE_ONLY" or "KEY_AND_VALUE"
        ], ...],
        "autoScalingGroups" =>  [::String, ...]
    ]

autoRollbackConfiguration = [ ... ]

Configuration information for an automatic rollback that is added when a deployment is created.

 autoRollbackConfiguration = [
        "enabled" =>  ::Bool,
        "events" =>  ["DEPLOYMENT_FAILURE", "DEPLOYMENT_STOP_ON_ALARM" or "DEPLOYMENT_STOP_ON_REQUEST", ...]
    ]

updateOutdatedInstancesOnly = ::Bool

Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.

fileExistsBehavior = "DISALLOW", "OVERWRITE" or "RETAIN"

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

The fileExistsBehavior parameter takes any of the following values:

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

Returns

CreateDeploymentOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationDoesNotExistException, DeploymentGroupNameRequiredException, InvalidDeploymentGroupNameException, DeploymentGroupDoesNotExistException, RevisionRequiredException, RevisionDoesNotExistException, InvalidRevisionException, InvalidDeploymentConfigNameException, DeploymentConfigDoesNotExistException, DescriptionTooLongException, DeploymentLimitExceededException, InvalidTargetInstancesException, InvalidAutoRollbackConfigException, InvalidLoadBalancerInfoException or InvalidFileExistsBehaviorException.

See also: AWS API Documentation

AmazonCodeDeploy.create_deployment_config — Function.

using AmazonCodeDeploy
create_deployment_config([::AWSConfig], arguments::Dict)
create_deployment_config([::AWSConfig]; deploymentConfigName=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"CreateDeploymentConfig", args)

CreateDeploymentConfig Operation

Creates a deployment configuration.

Arguments

deploymentConfigName = ::String – Required

The name of the deployment configuration to create.

minimumHealthyHosts = [ ... ]

The minimum number of healthy instances that should be available at any time during the deployment. There are two parameters expected in the input: type and value.

The type parameter takes either of the following values:

HOST_COUNT: The value parameter represents the minimum number of healthy instances as an absolute value.

FLEET_PERCENT: The value parameter represents the minimum number of healthy instances as a percentage of the total number of instances in the deployment. If you specify FLEET_PERCENT, at the start of the deployment, AWS CodeDeploy converts the percentage to the equivalent number of instance and rounds up fractional instances.

The value parameter takes an integer.

For example, to set a minimum of 95% healthy instance, specify a type of FLEET_PERCENT and a value of 95.

 minimumHealthyHosts = [
        "value" =>  ::Int,
        "type" =>  "HOST_COUNT" or "FLEET_PERCENT"
    ]

Returns

CreateDeploymentConfigOutput

Exceptions

InvalidDeploymentConfigNameException, DeploymentConfigNameRequiredException, DeploymentConfigAlreadyExistsException, InvalidMinimumHealthyHostValueException or DeploymentConfigLimitExceededException.

See also: AWS API Documentation

AmazonCodeDeploy.create_deployment_group — Function.

using AmazonCodeDeploy
create_deployment_group([::AWSConfig], arguments::Dict)
create_deployment_group([::AWSConfig]; applicationName=, deploymentGroupName=, serviceRoleArn=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"CreateDeploymentGroup", args)

CreateDeploymentGroup Operation

Creates a deployment group to which application revisions will be deployed.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

deploymentGroupName = ::String – Required

The name of a new deployment group for the specified application.

deploymentConfigName = ::String

If specified, the deployment configuration name can be either one of the predefined configurations provided with AWS CodeDeploy or a custom deployment configuration that you create by calling the create deployment configuration operation.

CodeDeployDefault.OneAtATime is the default deployment configuration. It is used if a configuration isn't specified for the deployment or the deployment group.

For more information about the predefined deployment configurations in AWS CodeDeploy, see Working with Deployment Groups in AWS CodeDeploy in the AWS CodeDeploy User Guide.

ec2TagFilters = [[ ... ], ...]

The Amazon EC2 tags on which to filter. The deployment group will include EC2 instances with any of the specified tags.

 ec2TagFilters = [[
        "Key" =>  ::String,
        "Value" =>  ::String,
        "Type" =>  "KEY_ONLY", "VALUE_ONLY" or "KEY_AND_VALUE"
    ], ...]

onPremisesInstanceTagFilters = [[ ... ], ...]

The on-premises instance tags on which to filter. The deployment group will include on-premises instances with any of the specified tags.

 onPremisesInstanceTagFilters = [[
        "Key" =>  ::String,
        "Value" =>  ::String,
        "Type" =>  "KEY_ONLY", "VALUE_ONLY" or "KEY_AND_VALUE"
    ], ...]

autoScalingGroups = [::String, ...]

A list of associated Auto Scaling groups.

serviceRoleArn = ::String – Required

A service role ARN that allows AWS CodeDeploy to act on the user's behalf when interacting with AWS services.

triggerConfigurations = [[ ... ], ...]

Information about triggers to create when the deployment group is created. For examples, see Create a Trigger for an AWS CodeDeploy Event in the AWS CodeDeploy User Guide.

 triggerConfigurations = [[
        "triggerName" =>  ::String,
        "triggerTargetArn" =>  ::String,
        "triggerEvents" =>  ["DeploymentStart", "DeploymentSuccess", "DeploymentFailure", "DeploymentStop", "DeploymentRollback", "DeploymentReady", "InstanceStart", "InstanceSuccess", "InstanceFailure" or "InstanceReady", ...]
    ], ...]

alarmConfiguration = [ ... ]

Information to add about Amazon CloudWatch alarms when the deployment group is created.

 alarmConfiguration = [
        "enabled" =>  ::Bool,
        "ignorePollAlarmFailure" =>  ::Bool,
        "alarms" =>  [["name" =>  ::String], ...]
    ]

autoRollbackConfiguration = [ ... ]

Configuration information for an automatic rollback that is added when a deployment group is created.

 autoRollbackConfiguration = [
        "enabled" =>  ::Bool,
        "events" =>  ["DEPLOYMENT_FAILURE", "DEPLOYMENT_STOP_ON_ALARM" or "DEPLOYMENT_STOP_ON_REQUEST", ...]
    ]

deploymentStyle = [ ... ]

Information about the type of deployment, in-place or blue/green, that you want to run and whether to route deployment traffic behind a load balancer.

 deploymentStyle = [
        "deploymentType" =>  "IN_PLACE" or "BLUE_GREEN",
        "deploymentOption" =>  "WITH_TRAFFIC_CONTROL" or "WITHOUT_TRAFFIC_CONTROL"
    ]

blueGreenDeploymentConfiguration = [ ... ]

Information about blue/green deployment options for a deployment group.

 blueGreenDeploymentConfiguration = [
        "terminateBlueInstancesOnDeploymentSuccess" =>  [
            "action" =>  "TERMINATE" or "KEEP_ALIVE",
            "terminationWaitTimeInMinutes" =>  ::Int
        ],
        "deploymentReadyOption" =>  [
            "actionOnTimeout" =>  "CONTINUE_DEPLOYMENT" or "STOP_DEPLOYMENT",
            "waitTimeInMinutes" =>  ::Int
        ],
        "greenFleetProvisioningOption" =>  ["action" =>  "DISCOVER_EXISTING" or "COPY_AUTO_SCALING_GROUP"]
    ]

loadBalancerInfo = ["elbInfoList" => [["name" => ::String], ...]]

Information about the load balancer used in a deployment.

Returns

CreateDeploymentGroupOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationDoesNotExistException, DeploymentGroupNameRequiredException, InvalidDeploymentGroupNameException, DeploymentGroupAlreadyExistsException, InvalidEC2TagException, InvalidTagException, InvalidAutoScalingGroupException, InvalidDeploymentConfigNameException, DeploymentConfigDoesNotExistException, RoleRequiredException, InvalidRoleException, DeploymentGroupLimitExceededException, LifecycleHookLimitExceededException, InvalidTriggerConfigException, TriggerTargetsLimitExceededException, InvalidAlarmConfigException, AlarmsLimitExceededException, InvalidAutoRollbackConfigException, InvalidLoadBalancerInfoException, InvalidDeploymentStyleException or InvalidBlueGreenDeploymentConfigurationException.

See also: AWS API Documentation

AmazonCodeDeploy.delete_application — Function.

using AmazonCodeDeploy
delete_application([::AWSConfig], arguments::Dict)
delete_application([::AWSConfig]; applicationName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"DeleteApplication", args)

DeleteApplication Operation

Deletes an application.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

Exceptions

ApplicationNameRequiredException or InvalidApplicationNameException.

See also: AWS API Documentation

AmazonCodeDeploy.delete_deployment_config — Function.

using AmazonCodeDeploy
delete_deployment_config([::AWSConfig], arguments::Dict)
delete_deployment_config([::AWSConfig]; deploymentConfigName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"DeleteDeploymentConfig", args)

DeleteDeploymentConfig Operation

Deletes a deployment configuration.

Note

A deployment configuration cannot be deleted if it is currently in use. Predefined configurations cannot be deleted.

Arguments

deploymentConfigName = ::String – Required

The name of a deployment configuration associated with the applicable IAM user or AWS account.

Exceptions

InvalidDeploymentConfigNameException, DeploymentConfigNameRequiredException, DeploymentConfigInUseException or InvalidOperationException.

See also: AWS API Documentation

AmazonCodeDeploy.delete_deployment_group — Function.

using AmazonCodeDeploy
delete_deployment_group([::AWSConfig], arguments::Dict)
delete_deployment_group([::AWSConfig]; applicationName=, deploymentGroupName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"DeleteDeploymentGroup", args)

DeleteDeploymentGroup Operation

Deletes a deployment group.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

deploymentGroupName = ::String – Required

The name of an existing deployment group for the specified application.

Returns

DeleteDeploymentGroupOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, DeploymentGroupNameRequiredException, InvalidDeploymentGroupNameException or InvalidRoleException.

See also: AWS API Documentation

AmazonCodeDeploy.deregister_on_premises_instance — Function.

using AmazonCodeDeploy
deregister_on_premises_instance([::AWSConfig], arguments::Dict)
deregister_on_premises_instance([::AWSConfig]; instanceName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"DeregisterOnPremisesInstance", args)

DeregisterOnPremisesInstance Operation

Deregisters an on-premises instance.

Arguments

instanceName = ::String – Required

The name of the on-premises instance to deregister.

Exceptions

InstanceNameRequiredException or InvalidInstanceNameException.

See also: AWS API Documentation

AmazonCodeDeploy.get_application — Function.

using AmazonCodeDeploy
get_application([::AWSConfig], arguments::Dict)
get_application([::AWSConfig]; applicationName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"GetApplication", args)

GetApplication Operation

Gets information about an application.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

Returns

GetApplicationOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException or ApplicationDoesNotExistException.

See also: AWS API Documentation

AmazonCodeDeploy.get_application_revision — Function.

using AmazonCodeDeploy
get_application_revision([::AWSConfig], arguments::Dict)
get_application_revision([::AWSConfig]; applicationName=, revision=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"GetApplicationRevision", args)

GetApplicationRevision Operation

Gets information about an application revision.

Arguments

applicationName = ::String – Required

The name of the application that corresponds to the revision.

revision = [ ... ] – Required

Information about the application revision to get, including type and location.

 revision = [
        "revisionType" =>  "S3" or "GitHub",
        "s3Location" =>  [
            "bucket" =>  ::String,
            "key" =>  ::String,
            "bundleType" =>  "tar", "tgz" or "zip",
            "version" =>  ::String,
            "eTag" =>  ::String
        ],
        "gitHubLocation" =>  [
            "repository" =>  ::String,
            "commitId" =>  ::String
        ]
    ]

Returns

GetApplicationRevisionOutput

Exceptions

ApplicationDoesNotExistException, ApplicationNameRequiredException, InvalidApplicationNameException, RevisionDoesNotExistException, RevisionRequiredException or InvalidRevisionException.

See also: AWS API Documentation

AmazonCodeDeploy.get_deployment — Function.

using AmazonCodeDeploy
get_deployment([::AWSConfig], arguments::Dict)
get_deployment([::AWSConfig]; deploymentId=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"GetDeployment", args)

GetDeployment Operation

Gets information about a deployment.

Arguments

deploymentId = ::String – Required

A deployment ID associated with the applicable IAM user or AWS account.

Returns

GetDeploymentOutput

Exceptions

DeploymentIdRequiredException, InvalidDeploymentIdException or DeploymentDoesNotExistException.

See also: AWS API Documentation

AmazonCodeDeploy.get_deployment_config — Function.

using AmazonCodeDeploy
get_deployment_config([::AWSConfig], arguments::Dict)
get_deployment_config([::AWSConfig]; deploymentConfigName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"GetDeploymentConfig", args)

GetDeploymentConfig Operation

Gets information about a deployment configuration.

Arguments

deploymentConfigName = ::String – Required

The name of a deployment configuration associated with the applicable IAM user or AWS account.

Returns

GetDeploymentConfigOutput

Exceptions

InvalidDeploymentConfigNameException, DeploymentConfigNameRequiredException or DeploymentConfigDoesNotExistException.

See also: AWS API Documentation

AmazonCodeDeploy.get_deployment_group — Function.

using AmazonCodeDeploy
get_deployment_group([::AWSConfig], arguments::Dict)
get_deployment_group([::AWSConfig]; applicationName=, deploymentGroupName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"GetDeploymentGroup", args)

GetDeploymentGroup Operation

Gets information about a deployment group.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

deploymentGroupName = ::String – Required

The name of an existing deployment group for the specified application.

Returns

GetDeploymentGroupOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationDoesNotExistException, DeploymentGroupNameRequiredException, InvalidDeploymentGroupNameException or DeploymentGroupDoesNotExistException.

See also: AWS API Documentation

AmazonCodeDeploy.get_deployment_instance — Function.

using AmazonCodeDeploy
get_deployment_instance([::AWSConfig], arguments::Dict)
get_deployment_instance([::AWSConfig]; deploymentId=, instanceId=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"GetDeploymentInstance", args)

GetDeploymentInstance Operation

Gets information about an instance as part of a deployment.

Arguments

deploymentId = ::String – Required

The unique ID of a deployment.

instanceId = ::String – Required

The unique ID of an instance in the deployment group.

Returns

GetDeploymentInstanceOutput

Exceptions

DeploymentIdRequiredException, DeploymentDoesNotExistException, InstanceIdRequiredException, InvalidDeploymentIdException, InstanceDoesNotExistException or InvalidInstanceNameException.

See also: AWS API Documentation

AmazonCodeDeploy.get_on_premises_instance — Function.

using AmazonCodeDeploy
get_on_premises_instance([::AWSConfig], arguments::Dict)
get_on_premises_instance([::AWSConfig]; instanceName=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"GetOnPremisesInstance", args)

GetOnPremisesInstance Operation

Gets information about an on-premises instance.

Arguments

instanceName = ::String – Required

The name of the on-premises instance about which to get information.

Returns

GetOnPremisesInstanceOutput

Exceptions

InstanceNameRequiredException, InstanceNotRegisteredException or InvalidInstanceNameException.

See also: AWS API Documentation

AmazonCodeDeploy.list_application_revisions — Function.

using AmazonCodeDeploy
list_application_revisions([::AWSConfig], arguments::Dict)
list_application_revisions([::AWSConfig]; applicationName=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ListApplicationRevisions", args)

ListApplicationRevisions Operation

Lists information about revisions for an application.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

sortBy = "registerTime", "firstUsedTime" or "lastUsedTime"

The column name to use to sort the list results:

registerTime: Sort by the time the revisions were registered with AWS CodeDeploy.

firstUsedTime: Sort by the time the revisions were first used in a deployment.

lastUsedTime: Sort by the time the revisions were last used in a deployment.

If not specified or set to null, the results will be returned in an arbitrary order.

sortOrder = "ascending" or "descending"

The order in which to sort the list results:

ascending: ascending order.

descending: descending order.

If not specified, the results will be sorted in ascending order.

If set to null, the results will be sorted in an arbitrary order.

s3Bucket = ::String

An Amazon S3 bucket name to limit the search for revisions.

If set to null, all of the user's buckets will be searched.

s3KeyPrefix = ::String

A key prefix for the set of Amazon S3 objects to limit the search for revisions.

deployed = "include", "exclude" or "ignore"

Whether to list revisions based on whether the revision is the target revision of an deployment group:

include: List revisions that are target revisions of a deployment group.

exclude: Do not list revisions that are target revisions of a deployment group.

ignore: List all revisions.

nextToken = ::String

An identifier returned from the previous list application revisions call. It can be used to return the next set of applications in the list.

Returns

ListApplicationRevisionsOutput

Exceptions

ApplicationDoesNotExistException, ApplicationNameRequiredException, InvalidApplicationNameException, InvalidSortByException, InvalidSortOrderException, InvalidBucketNameFilterException, InvalidKeyPrefixFilterException, BucketNameFilterRequiredException, InvalidDeployedStateFilterException or InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodeDeploy.list_applications — Function.

using AmazonCodeDeploy
list_applications([::AWSConfig], arguments::Dict)
list_applications([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ListApplications", args)

ListApplications Operation

Lists the applications registered with the applicable IAM user or AWS account.

Arguments

nextToken = ::String

An identifier returned from the previous list applications call. It can be used to return the next set of applications in the list.

Returns

ListApplicationsOutput

Exceptions

InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodeDeploy.list_deployment_configs — Function.

using AmazonCodeDeploy
list_deployment_configs([::AWSConfig], arguments::Dict)
list_deployment_configs([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ListDeploymentConfigs", args)

ListDeploymentConfigs Operation

Lists the deployment configurations with the applicable IAM user or AWS account.

Arguments

nextToken = ::String

An identifier returned from the previous list deployment configurations call. It can be used to return the next set of deployment configurations in the list.

Returns

ListDeploymentConfigsOutput

Exceptions

InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodeDeploy.list_deployment_groups — Function.

using AmazonCodeDeploy
list_deployment_groups([::AWSConfig], arguments::Dict)
list_deployment_groups([::AWSConfig]; applicationName=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ListDeploymentGroups", args)

ListDeploymentGroups Operation

Lists the deployment groups for an application registered with the applicable IAM user or AWS account.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

nextToken = ::String

An identifier returned from the previous list deployment groups call. It can be used to return the next set of deployment groups in the list.

Returns

ListDeploymentGroupsOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationDoesNotExistException or InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodeDeploy.list_deployment_instances — Function.

using AmazonCodeDeploy
list_deployment_instances([::AWSConfig], arguments::Dict)
list_deployment_instances([::AWSConfig]; deploymentId=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ListDeploymentInstances", args)

ListDeploymentInstances Operation

Lists the instance for a deployment associated with the applicable IAM user or AWS account.

Arguments

deploymentId = ::String – Required

The unique ID of a deployment.

nextToken = ::String

An identifier returned from the previous list deployment instances call. It can be used to return the next set of deployment instances in the list.

instanceStatusFilter = ["Pending", "InProgress", "Succeeded", "Failed", "Skipped", "Unknown" or "Ready", ...]

A subset of instances to list by status:

Pending: Include those instance with pending deployments.

InProgress: Include those instance where deployments are still in progress.

Succeeded: Include those instances with successful deployments.

Failed: Include those instance with failed deployments.

Skipped: Include those instance with skipped deployments.

Unknown: Include those instance with deployments in an unknown state.

instanceTypeFilter = ["Blue" or "Green", ...]

The set of instances in a blue/green deployment, either those in the original environment ("BLUE") or those in the replacement environment ("GREEN"), for which you want to view instance information.

Returns

ListDeploymentInstancesOutput

Exceptions

DeploymentIdRequiredException, DeploymentDoesNotExistException, DeploymentNotStartedException, InvalidNextTokenException, InvalidDeploymentIdException, InvalidInstanceStatusException, InvalidInstanceTypeException or InvalidDeploymentInstanceTypeException.

See also: AWS API Documentation

AmazonCodeDeploy.list_deployments — Function.

using AmazonCodeDeploy
list_deployments([::AWSConfig], arguments::Dict)
list_deployments([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ListDeployments", args)

ListDeployments Operation

Lists the deployments in a deployment group for an application registered with the applicable IAM user or AWS account.

Arguments

applicationName = ::String

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

deploymentGroupName = ::String

The name of an existing deployment group for the specified application.

includeOnlyStatuses = ["Created", "Queued", "InProgress", "Succeeded", "Failed", "Stopped" or "Ready", ...]

A subset of deployments to list by status:

Created: Include created deployments in the resulting list.

Queued: Include queued deployments in the resulting list.

In Progress: Include in-progress deployments in the resulting list.

Succeeded: Include successful deployments in the resulting list.

Failed: Include failed deployments in the resulting list.

Stopped: Include stopped deployments in the resulting list.

createTimeRange = [ ... ]

A time range (start and end) for returning a subset of the list of deployments.

 createTimeRange = [
        "start" =>  timestamp,
        "end" =>  timestamp
    ]

nextToken = ::String

An identifier returned from the previous list deployments call. It can be used to return the next set of deployments in the list.

Returns

ListDeploymentsOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationDoesNotExistException, InvalidDeploymentGroupNameException, DeploymentGroupDoesNotExistException, DeploymentGroupNameRequiredException, InvalidTimeRangeException, InvalidDeploymentStatusException or InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodeDeploy.list_git_hub_account_token_names — Function.

using AmazonCodeDeploy
list_git_hub_account_token_names([::AWSConfig], arguments::Dict)
list_git_hub_account_token_names([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ListGitHubAccountTokenNames", args)

ListGitHubAccountTokenNames Operation

Lists the names of stored connections to GitHub accounts.

Arguments

nextToken = ::String

An identifier returned from the previous ListGitHubAccountTokenNames call. It can be used to return the next set of names in the list.

Returns

ListGitHubAccountTokenNamesOutput

Exceptions

InvalidNextTokenException or ResourceValidationException.

See also: AWS API Documentation

AmazonCodeDeploy.list_on_premises_instances — Function.

using AmazonCodeDeploy
list_on_premises_instances([::AWSConfig], arguments::Dict)
list_on_premises_instances([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"ListOnPremisesInstances", args)

ListOnPremisesInstances Operation

Gets a list of names for one or more on-premises instances.

Unless otherwise specified, both registered and deregistered on-premises instance names will be listed. To list only registered or deregistered on-premises instance names, use the registration status parameter.

Arguments

registrationStatus = "Registered" or "Deregistered"

The registration status of the on-premises instances:

Deregistered: Include deregistered on-premises instances in the resulting list.

Registered: Include registered on-premises instances in the resulting list.

tagFilters = [[ ... ], ...]

The on-premises instance tags that will be used to restrict the corresponding on-premises instance names returned.

 tagFilters = [[
        "Key" =>  ::String,
        "Value" =>  ::String,
        "Type" =>  "KEY_ONLY", "VALUE_ONLY" or "KEY_AND_VALUE"
    ], ...]

nextToken = ::String

An identifier returned from the previous list on-premises instances call. It can be used to return the next set of on-premises instances in the list.

Returns

ListOnPremisesInstancesOutput

Exceptions

InvalidRegistrationStatusException, InvalidTagFilterException or InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodeDeploy.register_application_revision — Function.

using AmazonCodeDeploy
register_application_revision([::AWSConfig], arguments::Dict)
register_application_revision([::AWSConfig]; applicationName=, revision=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"RegisterApplicationRevision", args)

RegisterApplicationRevision Operation

Registers with AWS CodeDeploy a revision for the specified application.

Arguments

applicationName = ::String – Required

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

description = ::String

A comment about the revision.

revision = [ ... ] – Required

Information about the application revision to register, including type and location.

 revision = [
        "revisionType" =>  "S3" or "GitHub",
        "s3Location" =>  [
            "bucket" =>  ::String,
            "key" =>  ::String,
            "bundleType" =>  "tar", "tgz" or "zip",
            "version" =>  ::String,
            "eTag" =>  ::String
        ],
        "gitHubLocation" =>  [
            "repository" =>  ::String,
            "commitId" =>  ::String
        ]
    ]

Exceptions

ApplicationDoesNotExistException, ApplicationNameRequiredException, InvalidApplicationNameException, DescriptionTooLongException, RevisionRequiredException or InvalidRevisionException.

See also: AWS API Documentation

AmazonCodeDeploy.register_on_premises_instance — Function.

using AmazonCodeDeploy
register_on_premises_instance([::AWSConfig], arguments::Dict)
register_on_premises_instance([::AWSConfig]; instanceName=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"RegisterOnPremisesInstance", args)

RegisterOnPremisesInstance Operation

Registers an on-premises instance.

Note

Only one IAM ARN (an IAM session ARN or IAM user ARN) is supported in the request. You cannot use both.

Arguments

instanceName = ::String – Required

The name of the on-premises instance to register.

iamSessionArn = ::String

The ARN of the IAM session to associate with the on-premises instance.

iamUserArn = ::String

The ARN of the IAM user to associate with the on-premises instance.

Exceptions

InstanceNameAlreadyRegisteredException, IamArnRequiredException, IamSessionArnAlreadyRegisteredException, IamUserArnAlreadyRegisteredException, InstanceNameRequiredException, IamUserArnRequiredException, InvalidInstanceNameException, InvalidIamSessionArnException, InvalidIamUserArnException or MultipleIamArnsProvidedException.

See also: AWS API Documentation

AmazonCodeDeploy.remove_tags_from_on_premises_instances — Function.

using AmazonCodeDeploy
remove_tags_from_on_premises_instances([::AWSConfig], arguments::Dict)
remove_tags_from_on_premises_instances([::AWSConfig]; tags=, instanceNames=)

using AWSCore.Services.codedeploy
request([::AWSConfig],"RemoveTagsFromOnPremisesInstances", args)

RemoveTagsFromOnPremisesInstances Operation

Removes one or more tags from one or more on-premises instances.

Arguments

tags = [[ ... ], ...] – Required

The tag key-value pairs to remove from the on-premises instances.

 tags = [[
        "Key" =>  ::String,
        "Value" =>  ::String
    ], ...]

instanceNames = [::String, ...] – Required

The names of the on-premises instances from which to remove tags.

Exceptions

InstanceNameRequiredException, TagRequiredException, InvalidTagException, TagLimitExceededException, InstanceLimitExceededException or InstanceNotRegisteredException.

See also: AWS API Documentation

AmazonCodeDeploy.skip_wait_time_for_instance_termination — Function.

using AmazonCodeDeploy
skip_wait_time_for_instance_termination([::AWSConfig], arguments::Dict)
skip_wait_time_for_instance_termination([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"SkipWaitTimeForInstanceTermination", args)

SkipWaitTimeForInstanceTermination Operation

In a blue/green deployment, overrides any specified wait time and starts terminating instances immediately after the traffic routing is completed.

Arguments

deploymentId = ::String

The ID of the blue/green deployment for which you want to skip the instance termination wait time.

Exceptions

DeploymentIdRequiredException, DeploymentDoesNotExistException, DeploymentAlreadyCompletedException, InvalidDeploymentIdException, DeploymentNotStartedException or UnsupportedActionForDeploymentTypeException.

See also: AWS API Documentation

AmazonCodeDeploy.stop_deployment — Function.

using AmazonCodeDeploy
stop_deployment([::AWSConfig], arguments::Dict)
stop_deployment([::AWSConfig]; deploymentId=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"StopDeployment", args)

StopDeployment Operation

Attempts to stop an ongoing deployment.

Arguments

deploymentId = ::String – Required

The unique ID of a deployment.

autoRollbackEnabled = ::Bool

Indicates, when a deployment is stopped, whether instances that have been updated should be rolled back to the previous version of the application revision.

Returns

StopDeploymentOutput

Exceptions

DeploymentIdRequiredException, DeploymentDoesNotExistException, DeploymentAlreadyCompletedException or InvalidDeploymentIdException.

See also: AWS API Documentation

AmazonCodeDeploy.update_application — Function.

using AmazonCodeDeploy
update_application([::AWSConfig], arguments::Dict)
update_application([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"UpdateApplication", args)

UpdateApplication Operation

Changes the name of an application.

Arguments

applicationName = ::String

The current name of the application you want to change.

newApplicationName = ::String

The new name to give the application.

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationAlreadyExistsException or ApplicationDoesNotExistException.

See also: AWS API Documentation

AmazonCodeDeploy.update_deployment_group — Function.

using AmazonCodeDeploy
update_deployment_group([::AWSConfig], arguments::Dict)
update_deployment_group([::AWSConfig]; applicationName=, currentDeploymentGroupName=, <keyword arguments>)

using AWSCore.Services.codedeploy
request([::AWSConfig],"UpdateDeploymentGroup", args)

UpdateDeploymentGroup Operation

Changes information about a deployment group.

Arguments

applicationName = ::String – Required

The application name corresponding to the deployment group to update.

currentDeploymentGroupName = ::String – Required

The current name of the deployment group.

newDeploymentGroupName = ::String

The new name of the deployment group, if you want to change it.

deploymentConfigName = ::String

The replacement deployment configuration name to use, if you want to change it.

ec2TagFilters = [[ ... ], ...]

The replacement set of Amazon EC2 tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

 ec2TagFilters = [[
        "Key" =>  ::String,
        "Value" =>  ::String,
        "Type" =>  "KEY_ONLY", "VALUE_ONLY" or "KEY_AND_VALUE"
    ], ...]

onPremisesInstanceTagFilters = [[ ... ], ...]

The replacement set of on-premises instance tags on which to filter, if you want to change them. To keep the existing tags, enter their names. To remove tags, do not enter any tag names.

 onPremisesInstanceTagFilters = [[
        "Key" =>  ::String,
        "Value" =>  ::String,
        "Type" =>  "KEY_ONLY", "VALUE_ONLY" or "KEY_AND_VALUE"
    ], ...]

autoScalingGroups = [::String, ...]

The replacement list of Auto Scaling groups to be included in the deployment group, if you want to change them. To keep the Auto Scaling groups, enter their names. To remove Auto Scaling groups, do not enter any Auto Scaling group names.

serviceRoleArn = ::String

A replacement ARN for the service role, if you want to change it.

triggerConfigurations = [[ ... ], ...]

Information about triggers to change when the deployment group is updated. For examples, see Modify Triggers in an AWS CodeDeploy Deployment Group in the AWS CodeDeploy User Guide.

 triggerConfigurations = [[
        "triggerName" =>  ::String,
        "triggerTargetArn" =>  ::String,
        "triggerEvents" =>  ["DeploymentStart", "DeploymentSuccess", "DeploymentFailure", "DeploymentStop", "DeploymentRollback", "DeploymentReady", "InstanceStart", "InstanceSuccess", "InstanceFailure" or "InstanceReady", ...]
    ], ...]

alarmConfiguration = [ ... ]

Information to add or change about Amazon CloudWatch alarms when the deployment group is updated.

 alarmConfiguration = [
        "enabled" =>  ::Bool,
        "ignorePollAlarmFailure" =>  ::Bool,
        "alarms" =>  [["name" =>  ::String], ...]
    ]

autoRollbackConfiguration = [ ... ]

Information for an automatic rollback configuration that is added or changed when a deployment group is updated.

 autoRollbackConfiguration = [
        "enabled" =>  ::Bool,
        "events" =>  ["DEPLOYMENT_FAILURE", "DEPLOYMENT_STOP_ON_ALARM" or "DEPLOYMENT_STOP_ON_REQUEST", ...]
    ]

deploymentStyle = [ ... ]

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

 deploymentStyle = [
        "deploymentType" =>  "IN_PLACE" or "BLUE_GREEN",
        "deploymentOption" =>  "WITH_TRAFFIC_CONTROL" or "WITHOUT_TRAFFIC_CONTROL"
    ]

blueGreenDeploymentConfiguration = [ ... ]

Information about blue/green deployment options for a deployment group.

 blueGreenDeploymentConfiguration = [
        "terminateBlueInstancesOnDeploymentSuccess" =>  [
            "action" =>  "TERMINATE" or "KEEP_ALIVE",
            "terminationWaitTimeInMinutes" =>  ::Int
        ],
        "deploymentReadyOption" =>  [
            "actionOnTimeout" =>  "CONTINUE_DEPLOYMENT" or "STOP_DEPLOYMENT",
            "waitTimeInMinutes" =>  ::Int
        ],
        "greenFleetProvisioningOption" =>  ["action" =>  "DISCOVER_EXISTING" or "COPY_AUTO_SCALING_GROUP"]
    ]

loadBalancerInfo = ["elbInfoList" => [["name" => ::String], ...]]

Information about the load balancer used in a deployment.

Returns

UpdateDeploymentGroupOutput

Exceptions

ApplicationNameRequiredException, InvalidApplicationNameException, ApplicationDoesNotExistException, InvalidDeploymentGroupNameException, DeploymentGroupAlreadyExistsException, DeploymentGroupNameRequiredException, DeploymentGroupDoesNotExistException, InvalidEC2TagException, InvalidTagException, InvalidAutoScalingGroupException, InvalidDeploymentConfigNameException, DeploymentConfigDoesNotExistException, InvalidRoleException, LifecycleHookLimitExceededException, InvalidTriggerConfigException, TriggerTargetsLimitExceededException, InvalidAlarmConfigException, AlarmsLimitExceededException, InvalidAutoRollbackConfigException, InvalidLoadBalancerInfoException, InvalidDeploymentStyleException or InvalidBlueGreenDeploymentConfigurationException.

See also: AWS API Documentation