AmazonCodePipeline.jl

AmazonCodePipeline

AWS CodePipeline

Overview

This is the AWS CodePipeline API Reference. This guide provides descriptions of the actions and data types for AWS CodePipeline. Some functionality for your pipeline is only configurable through the API. For additional information, see the AWS CodePipeline User Guide.

You can use the AWS CodePipeline API to work with pipelines, stages, actions, gates, and transitions, as described below.

Pipelines are models of automated release processes. Each pipeline is uniquely named, and consists of actions, gates, and stages.

You can work with pipelines by calling:

CreatePipeline, which creates a uniquely-named pipeline.

DeletePipeline, which deletes the specified pipeline.

GetPipeline, which returns information about a pipeline structure.

GetPipelineExecution, which returns information about a specific execution of a pipeline.

GetPipelineState, which returns information about the current state of the stages and actions of a pipeline.

ListPipelines, which gets a summary of all of the pipelines associated with your account.

StartPipelineExecution, which runs the the most recent revision of an artifact through the pipeline.

UpdatePipeline, which updates a pipeline with edits or changes to the structure of the pipeline.

Pipelines include stages, which are logical groupings of gates and actions. Each stage contains one or more actions that must complete before the next stage begins. A stage will result in success or failure. If a stage fails, then the pipeline stops at that stage and will remain stopped until either a new version of an artifact appears in the source location, or a user takes action to re-run the most recent artifact through the pipeline. You can call GetPipelineState, which displays the status of a pipeline, including the status of stages in the pipeline, or GetPipeline, which returns the entire structure of the pipeline, including the stages of that pipeline. For more information about the structure of stages and actions, also refer to the AWS CodePipeline Pipeline Structure Reference.

Pipeline stages include actions, which are categorized into categories such as source or build actions performed within a stage of a pipeline. For example, you can use a source action to import artifacts into a pipeline from a source such as Amazon S3. Like stages, you do not work with actions directly in most cases, but you do define and interact with actions when working with pipeline operations such as CreatePipeline and GetPipelineState.

Pipelines also include transitions, which allow the transition of artifacts from one stage to the next in a pipeline after the actions in one stage complete.

You can work with transitions by calling:

DisableStageTransition, which prevents artifacts from transitioning to the next stage in a pipeline.

EnableStageTransition, which enables transition of artifacts between stages in a pipeline.

Using the API to integrate with AWS CodePipeline

For third-party integrators or developers who want to create their own integrations with AWS CodePipeline, the expected sequence varies from the standard API user. In order to integrate with AWS CodePipeline, developers will need to work with the following items:

Jobs, which are instances of an action. For example, a job for a source action might import a revision of an artifact from a source.

You can work with jobs by calling:

AcknowledgeJob, which confirms whether a job worker has received the specified job,

GetJobDetails, which returns the details of a job,

PollForJobs, which determines whether there are any jobs to act upon,

PutJobFailureResult, which provides details of a job failure, and

PutJobSuccessResult, which provides details of a job success.

Third party jobs, which are instances of an action created by a partner action and integrated into AWS CodePipeline. Partner actions are created by members of the AWS Partner Network.

You can work with third party jobs by calling:

AcknowledgeThirdPartyJob, which confirms whether a job worker has received the specified job,

GetThirdPartyJobDetails, which requests the details of a job for a partner action,

PollForThirdPartyJobs, which determines whether there are any jobs to act upon,

PutThirdPartyJobFailureResult, which provides details of a job failure, and

PutThirdPartyJobSuccessResult, which provides details of a job success.

This document is generated from apis/codepipeline-2015-07-09.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCodePipeline.acknowledge_job

AmazonCodePipeline.acknowledge_third_party_job

AmazonCodePipeline.create_custom_action_type

AmazonCodePipeline.create_pipeline

AmazonCodePipeline.delete_custom_action_type

AmazonCodePipeline.delete_pipeline

AmazonCodePipeline.disable_stage_transition

AmazonCodePipeline.enable_stage_transition

AmazonCodePipeline.get_job_details

AmazonCodePipeline.get_pipeline

AmazonCodePipeline.get_pipeline_execution

AmazonCodePipeline.get_pipeline_state

AmazonCodePipeline.get_third_party_job_details

AmazonCodePipeline.list_action_types

AmazonCodePipeline.list_pipeline_executions

AmazonCodePipeline.list_pipelines

AmazonCodePipeline.poll_for_jobs

AmazonCodePipeline.poll_for_third_party_jobs

AmazonCodePipeline.put_action_revision

AmazonCodePipeline.put_approval_result

AmazonCodePipeline.put_job_failure_result

AmazonCodePipeline.put_job_success_result

AmazonCodePipeline.put_third_party_job_failure_result

AmazonCodePipeline.put_third_party_job_success_result

AmazonCodePipeline.retry_stage_execution

AmazonCodePipeline.start_pipeline_execution

AmazonCodePipeline.update_pipeline

AmazonCodePipeline.acknowledge_job — Function.

using AmazonCodePipeline
acknowledge_job([::AWSConfig], arguments::Dict)
acknowledge_job([::AWSConfig]; jobId=, nonce=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "AcknowledgeJob", arguments::Dict)
codepipeline([::AWSConfig], "AcknowledgeJob", jobId=, nonce=)

AcknowledgeJob Operation

Returns information about a specified job and whether that job has been received by the job worker. Only used for custom actions.

Arguments

jobId = ::String – Required

The unique system-generated ID of the job for which you want to confirm receipt.

nonce = ::String – Required

A system-generated random number that AWS CodePipeline uses to ensure that the job is being worked on by only one job worker. Get this number from the response of the PollForJobs request that returned this job.

Returns

AcknowledgeJobOutput

Exceptions

ValidationException, InvalidNonceException or JobNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.acknowledge_third_party_job — Function.

using AmazonCodePipeline
acknowledge_third_party_job([::AWSConfig], arguments::Dict)
acknowledge_third_party_job([::AWSConfig]; jobId=, nonce=, clientToken=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "AcknowledgeThirdPartyJob", arguments::Dict)
codepipeline([::AWSConfig], "AcknowledgeThirdPartyJob", jobId=, nonce=, clientToken=)

AcknowledgeThirdPartyJob Operation

Confirms a job worker has received the specified job. Only used for partner actions.

Arguments

jobId = ::String – Required

The unique system-generated ID of the job.

nonce = ::String – Required

A system-generated random number that AWS CodePipeline uses to ensure that the job is being worked on by only one job worker. Get this number from the response to a GetThirdPartyJobDetails request.

clientToken = ::String – Required

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

Returns

AcknowledgeThirdPartyJobOutput

Exceptions

ValidationException, InvalidNonceException, JobNotFoundException or InvalidClientTokenException.

See also: AWS API Documentation

AmazonCodePipeline.create_custom_action_type — Function.

using AmazonCodePipeline
create_custom_action_type([::AWSConfig], arguments::Dict)
create_custom_action_type([::AWSConfig]; category=, provider=, version=, inputArtifactDetails=, outputArtifactDetails=, <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "CreateCustomActionType", arguments::Dict)
codepipeline([::AWSConfig], "CreateCustomActionType", category=, provider=, version=, inputArtifactDetails=, outputArtifactDetails=, <keyword arguments>)

CreateCustomActionType Operation

Creates a new custom action that can be used in all pipelines associated with the AWS account. Only used for custom actions.

Arguments

category = "Source", "Build", "Deploy", "Test", "Invoke" or "Approval" – Required

The category of the custom action, such as a build action or a test action.

Note

Although Source and Approval are listed as valid values, they are not currently functional. These values are reserved for future use.

provider = ::String – Required

The provider of the service used in the custom action, such as AWS CodeDeploy.

version = ::String – Required

The version identifier of the custom action.

settings = [ ... ]

Returns information about the settings for an action type.

 settings = [
        "thirdPartyConfigurationUrl" =>  ::String,
        "entityUrlTemplate" =>  ::String,
        "executionUrlTemplate" =>  ::String,
        "revisionUrlTemplate" =>  ::String
    ]

configurationProperties = [[ ... ], ...]

The configuration properties for the custom action.

Note

You can refer to a name in the configuration properties of the custom action within the URL templates by following the format of {Config:name}, as long as the configuration property is both required and not secret. For more information, see Create a Custom Action for a Pipeline.

 configurationProperties = [[
        "name" => <required> ::String,
        "required" => <required> ::Bool,
        "key" => <required> ::Bool,
        "secret" => <required> ::Bool,
        "queryable" =>  ::Bool,
        "description" =>  ::String,
        "type" =>  "String", "Number" or "Boolean"
    ], ...]

inputArtifactDetails = [ ... ] – Required

The details of the input artifact for the action, such as its commit ID.

 inputArtifactDetails = [
        "minimumCount" => <required> ::Int,
        "maximumCount" => <required> ::Int
    ]

outputArtifactDetails = [ ... ] – Required

The details of the output artifact of the action, such as its commit ID.

 outputArtifactDetails = [
        "minimumCount" => <required> ::Int,
        "maximumCount" => <required> ::Int
    ]

Returns

CreateCustomActionTypeOutput

Exceptions

ValidationException or LimitExceededException.

See also: AWS API Documentation

AmazonCodePipeline.create_pipeline — Function.

using AmazonCodePipeline
create_pipeline([::AWSConfig], arguments::Dict)
create_pipeline([::AWSConfig]; pipeline=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "CreatePipeline", arguments::Dict)
codepipeline([::AWSConfig], "CreatePipeline", pipeline=)

CreatePipeline Operation

Creates a pipeline.

Arguments

pipeline = [ ... ] – Required

Represents the structure of actions and stages to be performed in the pipeline.

 pipeline = [
        "name" => <required> ::String,
        "roleArn" => <required> ::String,
        "artifactStore" => <required> [
            "type" => <required> "S3",
            "location" => <required> ::String,
            "encryptionKey" =>  [
                "id" => <required> ::String,
                "type" => <required> "KMS"
            ]
        ],
        "stages" => <required> [[
            "name" => <required> ::String,
            "blockers" =>  [[
                "name" => <required> ::String,
                "type" => <required> "Schedule"
            ], ...],
            "actions" => <required> [[
                "name" => <required> ::String,
                "actionTypeId" => <required> [
                    "category" => <required> "Source", "Build", "Deploy", "Test", "Invoke" or "Approval",
                    "owner" => <required> "AWS", "ThirdParty" or "Custom",
                    "provider" => <required> ::String,
                    "version" => <required> ::String
                ],
                "runOrder" =>  ::Int,
                "configuration" =>  ::Dict{String,String},
                "outputArtifacts" =>  [["name" => <required> ::String], ...],
                "inputArtifacts" =>  [["name" => <required> ::String], ...],
                "roleArn" =>  ::String
            ], ...]
        ], ...],
        "version" =>  ::Int
    ]

Returns

CreatePipelineOutput

Exceptions

ValidationException, PipelineNameInUseException, InvalidStageDeclarationException, InvalidActionDeclarationException, InvalidBlockerDeclarationException, InvalidStructureException or LimitExceededException.

See also: AWS API Documentation

AmazonCodePipeline.delete_custom_action_type — Function.

using AmazonCodePipeline
delete_custom_action_type([::AWSConfig], arguments::Dict)
delete_custom_action_type([::AWSConfig]; category=, provider=, version=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "DeleteCustomActionType", arguments::Dict)
codepipeline([::AWSConfig], "DeleteCustomActionType", category=, provider=, version=)

DeleteCustomActionType Operation

Marks a custom action as deleted. PollForJobs for the custom action will fail after the action is marked for deletion. Only used for custom actions.

Important

You cannot recreate a custom action after it has been deleted unless you increase the version number of the action.

Arguments

category = "Source", "Build", "Deploy", "Test", "Invoke" or "Approval" – Required

The category of the custom action that you want to delete, such as source or deploy.

provider = ::String – Required

The provider of the service used in the custom action, such as AWS CodeDeploy.

version = ::String – Required

The version of the custom action to delete.

Exceptions

ValidationException.

See also: AWS API Documentation

AmazonCodePipeline.delete_pipeline — Function.

using AmazonCodePipeline
delete_pipeline([::AWSConfig], arguments::Dict)
delete_pipeline([::AWSConfig]; name=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "DeletePipeline", arguments::Dict)
codepipeline([::AWSConfig], "DeletePipeline", name=)

DeletePipeline Operation

Deletes the specified pipeline.

Arguments

name = ::String – Required

The name of the pipeline to be deleted.

Exceptions

ValidationException.

See also: AWS API Documentation

AmazonCodePipeline.disable_stage_transition — Function.

using AmazonCodePipeline
disable_stage_transition([::AWSConfig], arguments::Dict)
disable_stage_transition([::AWSConfig]; pipelineName=, stageName=, transitionType=, reason=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "DisableStageTransition", arguments::Dict)
codepipeline([::AWSConfig], "DisableStageTransition", pipelineName=, stageName=, transitionType=, reason=)

DisableStageTransition Operation

Prevents artifacts in a pipeline from transitioning to the next stage in the pipeline.

Arguments

pipelineName = ::String – Required

The name of the pipeline in which you want to disable the flow of artifacts from one stage to another.

stageName = ::String – Required

The name of the stage where you want to disable the inbound or outbound transition of artifacts.

transitionType = "Inbound" or "Outbound" – Required

Specifies whether artifacts will be prevented from transitioning into the stage and being processed by the actions in that stage (inbound), or prevented from transitioning from the stage after they have been processed by the actions in that stage (outbound).

reason = ::String – Required

The reason given to the user why a stage is disabled, such as waiting for manual approval or manual tests. This message is displayed in the pipeline console UI.

Exceptions

ValidationException, PipelineNotFoundException or StageNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.enable_stage_transition — Function.

using AmazonCodePipeline
enable_stage_transition([::AWSConfig], arguments::Dict)
enable_stage_transition([::AWSConfig]; pipelineName=, stageName=, transitionType=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "EnableStageTransition", arguments::Dict)
codepipeline([::AWSConfig], "EnableStageTransition", pipelineName=, stageName=, transitionType=)

EnableStageTransition Operation

Enables artifacts in a pipeline to transition to a stage in a pipeline.

Arguments

pipelineName = ::String – Required

The name of the pipeline in which you want to enable the flow of artifacts from one stage to another.

stageName = ::String – Required

The name of the stage where you want to enable the transition of artifacts, either into the stage (inbound) or from that stage to the next stage (outbound).

transitionType = "Inbound" or "Outbound" – Required

Specifies whether artifacts will be allowed to enter the stage and be processed by the actions in that stage (inbound) or whether already-processed artifacts will be allowed to transition to the next stage (outbound).

Exceptions

ValidationException, PipelineNotFoundException or StageNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.get_job_details — Function.

using AmazonCodePipeline
get_job_details([::AWSConfig], arguments::Dict)
get_job_details([::AWSConfig]; jobId=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "GetJobDetails", arguments::Dict)
codepipeline([::AWSConfig], "GetJobDetails", jobId=)

GetJobDetails Operation

Returns information about a job. Only used for custom actions.

Important

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

Arguments

jobId = ::String – Required

The unique system-generated ID for the job.

Returns

GetJobDetailsOutput

Exceptions

ValidationException or JobNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.get_pipeline — Function.

using AmazonCodePipeline
get_pipeline([::AWSConfig], arguments::Dict)
get_pipeline([::AWSConfig]; name=, <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "GetPipeline", arguments::Dict)
codepipeline([::AWSConfig], "GetPipeline", name=, <keyword arguments>)

GetPipeline Operation

Returns the metadata, structure, stages, and actions of a pipeline. Can be used to return the entire structure of a pipeline in JSON format, which can then be modified and used to update the pipeline structure with UpdatePipeline.

Arguments

name = ::String – Required

The name of the pipeline for which you want to get information. Pipeline names must be unique under an Amazon Web Services (AWS) user account.

version = ::Int

The version number of the pipeline. If you do not specify a version, defaults to the most current version.

Returns

GetPipelineOutput

Exceptions

ValidationException, PipelineNotFoundException or PipelineVersionNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.get_pipeline_execution — Function.

using AmazonCodePipeline
get_pipeline_execution([::AWSConfig], arguments::Dict)
get_pipeline_execution([::AWSConfig]; pipelineName=, pipelineExecutionId=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "GetPipelineExecution", arguments::Dict)
codepipeline([::AWSConfig], "GetPipelineExecution", pipelineName=, pipelineExecutionId=)

GetPipelineExecution Operation

Returns information about an execution of a pipeline, including details about artifacts, the pipeline execution ID, and the name, version, and status of the pipeline.

Arguments

pipelineName = ::String – Required

The name of the pipeline about which you want to get execution details.

pipelineExecutionId = ::String – Required

The ID of the pipeline execution about which you want to get execution details.

Returns

GetPipelineExecutionOutput

Exceptions

ValidationException, PipelineNotFoundException or PipelineExecutionNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.get_pipeline_state — Function.

using AmazonCodePipeline
get_pipeline_state([::AWSConfig], arguments::Dict)
get_pipeline_state([::AWSConfig]; name=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "GetPipelineState", arguments::Dict)
codepipeline([::AWSConfig], "GetPipelineState", name=)

GetPipelineState Operation

Returns information about the state of a pipeline, including the stages and actions.

Arguments

name = ::String – Required

The name of the pipeline about which you want to get information.

Returns

GetPipelineStateOutput

Exceptions

ValidationException or PipelineNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.get_third_party_job_details — Function.

using AmazonCodePipeline
get_third_party_job_details([::AWSConfig], arguments::Dict)
get_third_party_job_details([::AWSConfig]; jobId=, clientToken=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "GetThirdPartyJobDetails", arguments::Dict)
codepipeline([::AWSConfig], "GetThirdPartyJobDetails", jobId=, clientToken=)

GetThirdPartyJobDetails Operation

Requests the details of a job for a third party action. Only used for partner actions.

Important

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

Arguments

jobId = ::String – Required

The unique system-generated ID used for identifying the job.

clientToken = ::String – Required

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

Returns

GetThirdPartyJobDetailsOutput

Exceptions

JobNotFoundException, ValidationException, InvalidClientTokenException or InvalidJobException.

See also: AWS API Documentation

AmazonCodePipeline.list_action_types — Function.

using AmazonCodePipeline
list_action_types([::AWSConfig], arguments::Dict)
list_action_types([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "ListActionTypes", arguments::Dict)
codepipeline([::AWSConfig], "ListActionTypes", <keyword arguments>)

ListActionTypes Operation

Gets a summary of all AWS CodePipeline action types associated with your account.

Arguments

actionOwnerFilter = "AWS", "ThirdParty" or "Custom"

Filters the list of action types to those created by a specified entity.

nextToken = ::String

An identifier that was returned from the previous list action types call, which can be used to return the next set of action types in the list.

Returns

ListActionTypesOutput

Exceptions

ValidationException or InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodePipeline.list_pipeline_executions — Function.

using AmazonCodePipeline
list_pipeline_executions([::AWSConfig], arguments::Dict)
list_pipeline_executions([::AWSConfig]; pipelineName=, <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "ListPipelineExecutions", arguments::Dict)
codepipeline([::AWSConfig], "ListPipelineExecutions", pipelineName=, <keyword arguments>)

ListPipelineExecutions Operation

Gets a summary of the most recent executions for a pipeline.

Arguments

pipelineName = ::String – Required

The name of the pipeline for which you want to get execution summary information.

maxResults = ::Int

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned nextToken value. The available pipeline execution history is limited to the most recent 12 months, based on pipeline execution start times. Default value is 100.

nextToken = ::String

The token that was returned from the previous list pipeline executions call, which can be used to return the next set of pipeline executions in the list.

Returns

ListPipelineExecutionsOutput

Exceptions

ValidationException, PipelineNotFoundException or InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodePipeline.list_pipelines — Function.

using AmazonCodePipeline
list_pipelines([::AWSConfig], arguments::Dict)
list_pipelines([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "ListPipelines", arguments::Dict)
codepipeline([::AWSConfig], "ListPipelines", <keyword arguments>)

ListPipelines Operation

Gets a summary of all of the pipelines associated with your account.

Arguments

nextToken = ::String

An identifier that was returned from the previous list pipelines call, which can be used to return the next set of pipelines in the list.

Returns

ListPipelinesOutput

Exceptions

InvalidNextTokenException.

See also: AWS API Documentation

AmazonCodePipeline.poll_for_jobs — Function.

using AmazonCodePipeline
poll_for_jobs([::AWSConfig], arguments::Dict)
poll_for_jobs([::AWSConfig]; actionTypeId=, <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "PollForJobs", arguments::Dict)
codepipeline([::AWSConfig], "PollForJobs", actionTypeId=, <keyword arguments>)

PollForJobs Operation

Returns information about any jobs for AWS CodePipeline to act upon.

Important

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts. Additionally, this API returns any secret values defined for the action.

Arguments

actionTypeId = [ ... ] – Required

Represents information about an action type.

 actionTypeId = [
        "category" => <required> "Source", "Build", "Deploy", "Test", "Invoke" or "Approval",
        "owner" => <required> "AWS", "ThirdParty" or "Custom",
        "provider" => <required> ::String,
        "version" => <required> ::String
    ]

maxBatchSize = ::Int

The maximum number of jobs to return in a poll for jobs call.

queryParam = ::Dict{String,String}

A map of property names and values. For an action type with no queryable properties, this value must be null or an empty map. For an action type with a queryable property, you must supply that property as a key in the map. Only jobs whose action configuration matches the mapped value will be returned.

Returns

PollForJobsOutput

Exceptions

ValidationException or ActionTypeNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.poll_for_third_party_jobs — Function.

using AmazonCodePipeline
poll_for_third_party_jobs([::AWSConfig], arguments::Dict)
poll_for_third_party_jobs([::AWSConfig]; actionTypeId=, <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "PollForThirdPartyJobs", arguments::Dict)
codepipeline([::AWSConfig], "PollForThirdPartyJobs", actionTypeId=, <keyword arguments>)

PollForThirdPartyJobs Operation

Determines whether there are any third party jobs for a job worker to act on. Only used for partner actions.

Important

When this API is called, AWS CodePipeline returns temporary credentials for the Amazon S3 bucket used to store artifacts for the pipeline, if the action requires access to that Amazon S3 bucket for input or output artifacts.

Arguments

actionTypeId = [ ... ] – Required

Represents information about an action type.

 actionTypeId = [
        "category" => <required> "Source", "Build", "Deploy", "Test", "Invoke" or "Approval",
        "owner" => <required> "AWS", "ThirdParty" or "Custom",
        "provider" => <required> ::String,
        "version" => <required> ::String
    ]

maxBatchSize = ::Int

The maximum number of jobs to return in a poll for jobs call.

Returns

PollForThirdPartyJobsOutput

Exceptions

ActionTypeNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonCodePipeline.put_action_revision — Function.

using AmazonCodePipeline
put_action_revision([::AWSConfig], arguments::Dict)
put_action_revision([::AWSConfig]; pipelineName=, stageName=, actionName=, actionRevision=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "PutActionRevision", arguments::Dict)
codepipeline([::AWSConfig], "PutActionRevision", pipelineName=, stageName=, actionName=, actionRevision=)

PutActionRevision Operation

Provides information to AWS CodePipeline about new revisions to a source.

Arguments

pipelineName = ::String – Required

The name of the pipeline that will start processing the revision to the source.

stageName = ::String – Required

The name of the stage that contains the action that will act upon the revision.

actionName = ::String – Required

The name of the action that will process the revision.

actionRevision = [ ... ] – Required

Represents information about the version (or revision) of an action.

 actionRevision = [
        "revisionId" => <required> ::String,
        "revisionChangeId" => <required> ::String,
        "created" => <required> timestamp
    ]

Returns

PutActionRevisionOutput

Exceptions

PipelineNotFoundException, StageNotFoundException, ActionNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonCodePipeline.put_approval_result — Function.

using AmazonCodePipeline
put_approval_result([::AWSConfig], arguments::Dict)
put_approval_result([::AWSConfig]; pipelineName=, stageName=, actionName=, result=, token=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "PutApprovalResult", arguments::Dict)
codepipeline([::AWSConfig], "PutApprovalResult", pipelineName=, stageName=, actionName=, result=, token=)

PutApprovalResult Operation

Provides the response to a manual approval request to AWS CodePipeline. Valid responses include Approved and Rejected.

Arguments

pipelineName = ::String – Required

The name of the pipeline that contains the action.

stageName = ::String – Required

The name of the stage that contains the action.

actionName = ::String – Required

The name of the action for which approval is requested.

result = [ ... ] – Required

Represents information about the result of the approval request.

 result = [
        "summary" => <required> ::String,
        "status" => <required> "Approved" or "Rejected"
    ]

token = ::String – Required

The system-generated token used to identify a unique approval request. The token for each open approval request can be obtained using the GetPipelineState action and is used to validate that the approval request corresponding to this token is still valid.

Returns

PutApprovalResultOutput

Exceptions

InvalidApprovalTokenException, ApprovalAlreadyCompletedException, PipelineNotFoundException, StageNotFoundException, ActionNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonCodePipeline.put_job_failure_result — Function.

using AmazonCodePipeline
put_job_failure_result([::AWSConfig], arguments::Dict)
put_job_failure_result([::AWSConfig]; jobId=, failureDetails=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "PutJobFailureResult", arguments::Dict)
codepipeline([::AWSConfig], "PutJobFailureResult", jobId=, failureDetails=)

PutJobFailureResult Operation

Represents the failure of a job as returned to the pipeline by a job worker. Only used for custom actions.

Arguments

jobId = ::String – Required

The unique system-generated ID of the job that failed. This is the same ID returned from PollForJobs.

failureDetails = [ ... ] – Required

The details about the failure of a job.

 failureDetails = [
        "type" => <required> "JobFailed", "ConfigurationError", "PermissionError", "RevisionOutOfSync", "RevisionUnavailable" or "SystemUnavailable",
        "message" => <required> ::String,
        "externalExecutionId" =>  ::String
    ]

Exceptions

ValidationException, JobNotFoundException or InvalidJobStateException.

See also: AWS API Documentation

AmazonCodePipeline.put_job_success_result — Function.

using AmazonCodePipeline
put_job_success_result([::AWSConfig], arguments::Dict)
put_job_success_result([::AWSConfig]; jobId=, <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "PutJobSuccessResult", arguments::Dict)
codepipeline([::AWSConfig], "PutJobSuccessResult", jobId=, <keyword arguments>)

PutJobSuccessResult Operation

Represents the success of a job as returned to the pipeline by a job worker. Only used for custom actions.

Arguments

jobId = ::String – Required

The unique system-generated ID of the job that succeeded. This is the same ID returned from PollForJobs.

currentRevision = [ ... ]

The ID of the current revision of the artifact successfully worked upon by the job.

 currentRevision = [
        "revision" => <required> ::String,
        "changeIdentifier" => <required> ::String,
        "created" =>  timestamp,
        "revisionSummary" =>  ::String
    ]

continuationToken = ::String

A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a custom action in progress. Future jobs will use this token in order to identify the running instance of the action. It can be reused to return additional information about the progress of the custom action. When the action is complete, no continuation token should be supplied.

executionDetails = [ ... ]

The execution details of the successful job, such as the actions taken by the job worker.

 executionDetails = [
        "summary" =>  ::String,
        "externalExecutionId" =>  ::String,
        "percentComplete" =>  ::Int
    ]

Exceptions

ValidationException, JobNotFoundException or InvalidJobStateException.

See also: AWS API Documentation

AmazonCodePipeline.put_third_party_job_failure_result — Function.

using AmazonCodePipeline
put_third_party_job_failure_result([::AWSConfig], arguments::Dict)
put_third_party_job_failure_result([::AWSConfig]; jobId=, clientToken=, failureDetails=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "PutThirdPartyJobFailureResult", arguments::Dict)
codepipeline([::AWSConfig], "PutThirdPartyJobFailureResult", jobId=, clientToken=, failureDetails=)

PutThirdPartyJobFailureResult Operation

Represents the failure of a third party job as returned to the pipeline by a job worker. Only used for partner actions.

Arguments

jobId = ::String – Required

The ID of the job that failed. This is the same ID returned from PollForThirdPartyJobs.

clientToken = ::String – Required

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

failureDetails = [ ... ] – Required

Represents information about failure details.

 failureDetails = [
        "type" => <required> "JobFailed", "ConfigurationError", "PermissionError", "RevisionOutOfSync", "RevisionUnavailable" or "SystemUnavailable",
        "message" => <required> ::String,
        "externalExecutionId" =>  ::String
    ]

Exceptions

ValidationException, JobNotFoundException, InvalidJobStateException or InvalidClientTokenException.

See also: AWS API Documentation

AmazonCodePipeline.put_third_party_job_success_result — Function.

using AmazonCodePipeline
put_third_party_job_success_result([::AWSConfig], arguments::Dict)
put_third_party_job_success_result([::AWSConfig]; jobId=, clientToken=, <keyword arguments>)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "PutThirdPartyJobSuccessResult", arguments::Dict)
codepipeline([::AWSConfig], "PutThirdPartyJobSuccessResult", jobId=, clientToken=, <keyword arguments>)

PutThirdPartyJobSuccessResult Operation

Represents the success of a third party job as returned to the pipeline by a job worker. Only used for partner actions.

Arguments

jobId = ::String – Required

The ID of the job that successfully completed. This is the same ID returned from PollForThirdPartyJobs.

clientToken = ::String – Required

The clientToken portion of the clientId and clientToken pair used to verify that the calling entity is allowed access to the job and its details.

currentRevision = [ ... ]

Represents information about a current revision.

 currentRevision = [
        "revision" => <required> ::String,
        "changeIdentifier" => <required> ::String,
        "created" =>  timestamp,
        "revisionSummary" =>  ::String
    ]

continuationToken = ::String

A token generated by a job worker, such as an AWS CodeDeploy deployment ID, that a successful job provides to identify a partner action in progress. Future jobs will use this token in order to identify the running instance of the action. It can be reused to return additional information about the progress of the partner action. When the action is complete, no continuation token should be supplied.

executionDetails = [ ... ]

The details of the actions taken and results produced on an artifact as it passes through stages in the pipeline.

 executionDetails = [
        "summary" =>  ::String,
        "externalExecutionId" =>  ::String,
        "percentComplete" =>  ::Int
    ]

Exceptions

ValidationException, JobNotFoundException, InvalidJobStateException or InvalidClientTokenException.

See also: AWS API Documentation

AmazonCodePipeline.retry_stage_execution — Function.

using AmazonCodePipeline
retry_stage_execution([::AWSConfig], arguments::Dict)
retry_stage_execution([::AWSConfig]; pipelineName=, stageName=, pipelineExecutionId=, retryMode=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "RetryStageExecution", arguments::Dict)
codepipeline([::AWSConfig], "RetryStageExecution", pipelineName=, stageName=, pipelineExecutionId=, retryMode=)

RetryStageExecution Operation

Resumes the pipeline execution by retrying the last failed actions in a stage.

Arguments

pipelineName = ::String – Required

The name of the pipeline that contains the failed stage.

stageName = ::String – Required

The name of the failed stage to be retried.

pipelineExecutionId = ::String – Required

The ID of the pipeline execution in the failed stage to be retried. Use the GetPipelineState action to retrieve the current pipelineExecutionId of the failed stage

retryMode = "FAILED_ACTIONS" – Required

The scope of the retry attempt. Currently, the only supported value is FAILED_ACTIONS.

Returns

RetryStageExecutionOutput

Exceptions

ValidationException, PipelineNotFoundException, StageNotFoundException, StageNotRetryableException or NotLatestPipelineExecutionException.

See also: AWS API Documentation

AmazonCodePipeline.start_pipeline_execution — Function.

using AmazonCodePipeline
start_pipeline_execution([::AWSConfig], arguments::Dict)
start_pipeline_execution([::AWSConfig]; name=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "StartPipelineExecution", arguments::Dict)
codepipeline([::AWSConfig], "StartPipelineExecution", name=)

StartPipelineExecution Operation

Starts the specified pipeline. Specifically, it begins processing the latest commit to the source location specified as part of the pipeline.

Arguments

name = ::String – Required

The name of the pipeline to start.

Returns

StartPipelineExecutionOutput

Exceptions

ValidationException or PipelineNotFoundException.

See also: AWS API Documentation

AmazonCodePipeline.update_pipeline — Function.

using AmazonCodePipeline
update_pipeline([::AWSConfig], arguments::Dict)
update_pipeline([::AWSConfig]; pipeline=)

using AWSCore.Services.codepipeline
codepipeline([::AWSConfig], "UpdatePipeline", arguments::Dict)
codepipeline([::AWSConfig], "UpdatePipeline", pipeline=)

UpdatePipeline Operation

Updates a specified pipeline with edits or changes to its structure. Use a JSON file with the pipeline structure in conjunction with UpdatePipeline to provide the full structure of the pipeline. Updating the pipeline increases the version number of the pipeline by 1.

Arguments

pipeline = [ ... ] – Required

The name of the pipeline to be updated.

 pipeline = [
        "name" => <required> ::String,
        "roleArn" => <required> ::String,
        "artifactStore" => <required> [
            "type" => <required> "S3",
            "location" => <required> ::String,
            "encryptionKey" =>  [
                "id" => <required> ::String,
                "type" => <required> "KMS"
            ]
        ],
        "stages" => <required> [[
            "name" => <required> ::String,
            "blockers" =>  [[
                "name" => <required> ::String,
                "type" => <required> "Schedule"
            ], ...],
            "actions" => <required> [[
                "name" => <required> ::String,
                "actionTypeId" => <required> [
                    "category" => <required> "Source", "Build", "Deploy", "Test", "Invoke" or "Approval",
                    "owner" => <required> "AWS", "ThirdParty" or "Custom",
                    "provider" => <required> ::String,
                    "version" => <required> ::String
                ],
                "runOrder" =>  ::Int,
                "configuration" =>  ::Dict{String,String},
                "outputArtifacts" =>  [["name" => <required> ::String], ...],
                "inputArtifacts" =>  [["name" => <required> ::String], ...],
                "roleArn" =>  ::String
            ], ...]
        ], ...],
        "version" =>  ::Int
    ]

Returns

UpdatePipelineOutput

Exceptions

ValidationException, InvalidStageDeclarationException, InvalidActionDeclarationException, InvalidBlockerDeclarationException or InvalidStructureException.

See also: AWS API Documentation