AmazonCodeStar.jl

AmazonCodeStar

AWS CodeStar

This is the API reference for AWS CodeStar. This reference provides descriptions of the operations and data types for the AWS CodeStar API along with usage examples.

You can use the AWS CodeStar API to work with:

Projects and their resources, by calling the following:

DeleteProject, which deletes a project in AWS CodeStar.

DescribeProject, which lists the attributes of a project.

ListProjects, which lists all AWS CodeStar projects associated with your AWS account.

ListResources, which lists the resources associated with an AWS CodeStar project.

UpdateProject, which updates the attributes of an AWS CodeStar project.

Teams and team members, by calling the following:

AssociateTeamMember, which adds an IAM user to the team for an AWS CodeStar project.

DisassociateTeamMember, which removes an IAM user from the team for an AWS CodeStar project.

ListTeamMembers, which lists all the IAM users in the team for an AWS CodeStar project, including their roles and attributes.

Users, by calling the following:

CreateUserProfile, which creates a user profile that contains data associated with the user across all AWS CodeStar projects.

DeleteUserProfile, which deletes all user profile information across all AWS CodeStar projects.

DescribeUserProfile, which describes the profile of a user.

ListUserProfiles, which lists all AWS CodeStar user profiles.

UpdateUserProfile, which updates the profile for an AWS CodeStar user.

This document is generated from apis/codestar-2017-04-19.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCodeStar.associate_team_member

AmazonCodeStar.create_project

AmazonCodeStar.create_user_profile

AmazonCodeStar.delete_project

AmazonCodeStar.delete_user_profile

AmazonCodeStar.describe_project

AmazonCodeStar.describe_user_profile

AmazonCodeStar.disassociate_team_member

AmazonCodeStar.list_projects

AmazonCodeStar.list_resources

AmazonCodeStar.list_team_members

AmazonCodeStar.list_user_profiles

AmazonCodeStar.update_project

AmazonCodeStar.update_team_member

AmazonCodeStar.update_user_profile

AmazonCodeStar.associate_team_member — Function.

using AmazonCodeStar
associate_team_member([::AWSConfig], arguments::Dict)
associate_team_member([::AWSConfig]; projectId=, userArn=, projectRole=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"AssociateTeamMember", args)

AssociateTeamMember Operation

Adds an IAM user to the team for an AWS CodeStar project.

Arguments

projectId = ::String – Required

The ID of the project to which you will add the IAM user.

clientRequestToken = ::String

A user- or system-generated token that identifies the entity that requested the team member association to the project. This token can be used to repeat the request.

userArn = ::String – Required

The Amazon Resource Name (ARN) for the IAM user you want to add to the DevHub project.

projectRole = ::String – Required

The AWS CodeStar project role that will apply to this user. This role determines what actions a user can take in an AWS CodeStar project.

remoteAccessAllowed = ::Bool

Whether the team member is allowed to use an SSH public/private key pair to remotely access project resources, for example Amazon EC2 instances.

Returns

AssociateTeamMemberResult

Exceptions

LimitExceededException, ProjectNotFoundException, TeamMemberAlreadyAssociatedException, ValidationException, InvalidServiceRoleException, ProjectConfigurationException or ConcurrentModificationException.

See also: AWS API Documentation

AmazonCodeStar.create_project — Function.

using AmazonCodeStar
create_project([::AWSConfig], arguments::Dict)
create_project([::AWSConfig]; name=, id=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"CreateProject", args)

CreateProject Operation

Reserved for future use. To create a project, use the AWS CodeStar console.

Arguments

name = ::String – Required

Reserved for future use.

id = ::String – Required

Reserved for future use.

description = ::String

Reserved for future use.

clientRequestToken = ::String

Reserved for future use.

Returns

CreateProjectResult

Exceptions

ProjectAlreadyExistsException, LimitExceededException, ValidationException, ProjectCreationFailedException, InvalidServiceRoleException, ProjectConfigurationException or ConcurrentModificationException.

See also: AWS API Documentation

AmazonCodeStar.create_user_profile — Function.

using AmazonCodeStar
create_user_profile([::AWSConfig], arguments::Dict)
create_user_profile([::AWSConfig]; userArn=, displayName=, emailAddress=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"CreateUserProfile", args)

CreateUserProfile Operation

Creates a profile for a user that includes user preferences, such as the display name and email address assocciated with the user, in AWS CodeStar. The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.

Arguments

userArn = ::String – Required

The Amazon Resource Name (ARN) of the user in IAM.

displayName = ::String – Required

The name that will be displayed as the friendly name for the user in AWS CodeStar.

emailAddress = ::String – Required

The email address that will be displayed as part of the user's profile in AWS CodeStar.

sshPublicKey = ::String

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

Returns

CreateUserProfileResult

Exceptions

UserProfileAlreadyExistsException or ValidationException.

See also: AWS API Documentation

AmazonCodeStar.delete_project — Function.

using AmazonCodeStar
delete_project([::AWSConfig], arguments::Dict)
delete_project([::AWSConfig]; id=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"DeleteProject", args)

DeleteProject Operation

Deletes a project, including project resources. Does not delete users associated with the project, but does delete the IAM roles that allowed access to the project.

Arguments

id = ::String – Required

The ID of the project to be deleted in AWS CodeStar.

clientRequestToken = ::String

A user- or system-generated token that identifies the entity that requested project deletion. This token can be used to repeat the request.

deleteStack = ::Bool

Whether to send a delete request for the primary stack in AWS CloudFormation originally used to generate the project and its resources. This option will delete all AWS resources for the project (except for any buckets in Amazon S3) as well as deleting the project itself. Recommended for most use cases.

Returns

DeleteProjectResult

Exceptions

ConcurrentModificationException, ValidationException or InvalidServiceRoleException.

See also: AWS API Documentation

AmazonCodeStar.delete_user_profile — Function.

using AmazonCodeStar
delete_user_profile([::AWSConfig], arguments::Dict)
delete_user_profile([::AWSConfig]; userArn=)

using AWSCore.Services.codestar
request([::AWSConfig],"DeleteUserProfile", args)

DeleteUserProfile Operation

Deletes a user profile in AWS CodeStar, including all personal preference data associated with that profile, such as display name and email address. It does not delete the history of that user, for example the history of commits made by that user.

Arguments

userArn = ::String – Required

The Amazon Resource Name (ARN) of the user to delete from AWS CodeStar.

Returns

DeleteUserProfileResult

Exceptions

ValidationException.

See also: AWS API Documentation

AmazonCodeStar.describe_project — Function.

using AmazonCodeStar
describe_project([::AWSConfig], arguments::Dict)
describe_project([::AWSConfig]; id=)

using AWSCore.Services.codestar
request([::AWSConfig],"DescribeProject", args)

DescribeProject Operation

Describes a project and its resources.

Arguments

id = ::String – Required

The ID of the project.

Returns

DescribeProjectResult

Exceptions

ProjectNotFoundException, ValidationException, InvalidServiceRoleException, ProjectConfigurationException or ConcurrentModificationException.

See also: AWS API Documentation

AmazonCodeStar.describe_user_profile — Function.

using AmazonCodeStar
describe_user_profile([::AWSConfig], arguments::Dict)
describe_user_profile([::AWSConfig]; userArn=)

using AWSCore.Services.codestar
request([::AWSConfig],"DescribeUserProfile", args)

DescribeUserProfile Operation

Describes a user in AWS CodeStar and the user attributes across all projects.

Arguments

userArn = ::String – Required

The Amazon Resource Name (ARN) of the user.

Returns

DescribeUserProfileResult

Exceptions

UserProfileNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonCodeStar.disassociate_team_member — Function.

using AmazonCodeStar
disassociate_team_member([::AWSConfig], arguments::Dict)
disassociate_team_member([::AWSConfig]; projectId=, userArn=)

using AWSCore.Services.codestar
request([::AWSConfig],"DisassociateTeamMember", args)

DisassociateTeamMember Operation

Removes a user from a project. Removing a user from a project also removes the IAM policies from that user that allowed access to the project and its resources. Disassociating a team member does not remove that user's profile from AWS CodeStar. It does not remove the user from IAM.

Arguments

projectId = ::String – Required

The ID of the AWS CodeStar project from which you want to remove a team member.

userArn = ::String – Required

The Amazon Resource Name (ARN) of the IAM user or group whom you want to remove from the project.

Returns

DisassociateTeamMemberResult

Exceptions

ProjectNotFoundException, ValidationException, InvalidServiceRoleException or ConcurrentModificationException.

See also: AWS API Documentation

AmazonCodeStar.list_projects — Function.

using AmazonCodeStar
list_projects([::AWSConfig], arguments::Dict)
list_projects([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"ListProjects", args)

ListProjects Operation

Lists all projects in AWS CodeStar associated with your AWS account.

Arguments

nextToken = ::String

The continuation token to be used to return the next set of results, if the results cannot be returned in one response.

maxResults = ::Int

The maximum amount of data that can be contained in a single set of results.

Returns

ListProjectsResult

Exceptions

InvalidNextTokenException or ValidationException.

See also: AWS API Documentation

AmazonCodeStar.list_resources — Function.

using AmazonCodeStar
list_resources([::AWSConfig], arguments::Dict)
list_resources([::AWSConfig]; projectId=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"ListResources", args)

ListResources Operation

Lists resources associated with a project in AWS CodeStar.

Arguments

projectId = ::String – Required

The ID of the project.

nextToken = ::String

The continuation token for the next set of results, if the results cannot be returned in one response.

maxResults = ::Int

he maximum amount of data that can be contained in a single set of results.

Returns

ListResourcesResult

Exceptions

ProjectNotFoundException, InvalidNextTokenException or ValidationException.

See also: AWS API Documentation

AmazonCodeStar.list_team_members — Function.

using AmazonCodeStar
list_team_members([::AWSConfig], arguments::Dict)
list_team_members([::AWSConfig]; projectId=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"ListTeamMembers", args)

ListTeamMembers Operation

Lists all team members associated with a project.

Arguments

projectId = ::String – Required

The ID of the project for which you want to list team members.

nextToken = ::String

The continuation token for the next set of results, if the results cannot be returned in one response.

maxResults = ::Int

The maximum number of team members you want returned in a response.

Returns

ListTeamMembersResult

Exceptions

ProjectNotFoundException, InvalidNextTokenException or ValidationException.

See also: AWS API Documentation

AmazonCodeStar.list_user_profiles — Function.

using AmazonCodeStar
list_user_profiles([::AWSConfig], arguments::Dict)
list_user_profiles([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"ListUserProfiles", args)

ListUserProfiles Operation

Lists all the user profiles configured for your AWS account in AWS CodeStar.

Arguments

nextToken = ::String

The continuation token for the next set of results, if the results cannot be returned in one response.

maxResults = ::Int

The maximum number of results to return in a response.

Returns

ListUserProfilesResult

Exceptions

InvalidNextTokenException or ValidationException.

See also: AWS API Documentation

AmazonCodeStar.update_project — Function.

using AmazonCodeStar
update_project([::AWSConfig], arguments::Dict)
update_project([::AWSConfig]; id=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"UpdateProject", args)

UpdateProject Operation

Updates a project in AWS CodeStar.

Arguments

id = ::String – Required

The ID of the project you want to update.

name = ::String

The name of the project you want to update.

description = ::String

The description of the project, if any.

Returns

UpdateProjectResult

Exceptions

ProjectNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonCodeStar.update_team_member — Function.

using AmazonCodeStar
update_team_member([::AWSConfig], arguments::Dict)
update_team_member([::AWSConfig]; projectId=, userArn=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"UpdateTeamMember", args)

UpdateTeamMember Operation

Updates a team member's attributes in an AWS CodeStar project. For example, you can change a team member's role in the project, or change whether they have remote access to project resources.

Arguments

projectId = ::String – Required

The ID of the project.

userArn = ::String – Required

The Amazon Resource Name (ARN) of the user for whom you want to change team membership attributes.

projectRole = ::String

The role assigned to the user in the project. Project roles have different levels of access. For more information, see Working with Teams in the AWS CodeStar User Guide.

remoteAccessAllowed = ::Bool

Whether a team member is allowed to remotely access project resources using the SSH public key associated with the user's profile. Even if this is set to True, the user must associate a public key with their profile before the user can access resources.

Returns

UpdateTeamMemberResult

Exceptions

LimitExceededException, ProjectNotFoundException, ValidationException, InvalidServiceRoleException, ProjectConfigurationException, ConcurrentModificationException or TeamMemberNotFoundException.

See also: AWS API Documentation

AmazonCodeStar.update_user_profile — Function.

using AmazonCodeStar
update_user_profile([::AWSConfig], arguments::Dict)
update_user_profile([::AWSConfig]; userArn=, <keyword arguments>)

using AWSCore.Services.codestar
request([::AWSConfig],"UpdateUserProfile", args)

UpdateUserProfile Operation

Updates a user's profile in AWS CodeStar. The user profile is not project-specific. Information in the user profile is displayed wherever the user's information appears to other users in AWS CodeStar.

Arguments

userArn = ::String – Required

The name that will be displayed as the friendly name for the user in AWS CodeStar.

displayName = ::String

The name that is displayed as the friendly name for the user in AWS CodeStar.

emailAddress = ::String

The email address that is displayed as part of the user's profile in AWS CodeStar.

sshPublicKey = ::String

The SSH public key associated with the user in AWS CodeStar. If a project owner allows the user remote access to project resources, this public key will be used along with the user's private key for SSH access.

Returns

UpdateUserProfileResult

Exceptions

UserProfileNotFoundException or ValidationException.

See also: AWS API Documentation