AmazonCognitoIdentity.jl

AmazonCognitoIdentity

Amazon Cognito

Amazon Cognito is a web service that delivers scoped temporary credentials to mobile devices and other untrusted environments. Amazon Cognito uniquely identifies a device and supplies the user with a consistent identity over the lifetime of an application.

Using Amazon Cognito, you can enable authentication with one or more third-party identity providers (Facebook, Google, or Login with Amazon), and you can also choose to support unauthenticated access from your app. Cognito delivers a unique identifier for each user and acts as an OpenID token provider trusted by AWS Security Token Service (STS) to access temporary, limited-privilege AWS credentials.

To provide end-user credentials, first make an unsigned call to GetId. If the end user is authenticated with one of the supported identity providers, set the Logins map with the identity provider token. GetId returns a unique identifier for the user.

Next, make an unsigned call to GetCredentialsForIdentity. This call expects the same Logins map as the GetId call, as well as the IdentityID originally returned by GetId. Assuming your identity pool has been configured via the SetIdentityPoolRoles operation, GetCredentialsForIdentity will return AWS credentials for your use. If your pool has not been configured with SetIdentityPoolRoles, or if you want to follow legacy flow, make an unsigned call to GetOpenIdToken, which returns the OpenID token necessary to call STS and retrieve AWS credentials. This call expects the same Logins map as the GetId call, as well as the IdentityID originally returned by GetId. The token returned by GetOpenIdToken can be passed to the STS operation AssumeRoleWithWebIdentity to retrieve AWS credentials.

If you want to use Amazon Cognito in an Android, iOS, or Unity application, you will probably want to make API calls via the AWS Mobile SDK. To learn more, see the AWS Mobile SDK Developer Guide.

This document is generated from apis/cognito-identity-2014-06-30.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCognitoIdentity.create_identity_pool

AmazonCognitoIdentity.delete_identities

AmazonCognitoIdentity.delete_identity_pool

AmazonCognitoIdentity.describe_identity

AmazonCognitoIdentity.describe_identity_pool

AmazonCognitoIdentity.get_credentials_for_identity

AmazonCognitoIdentity.get_id

AmazonCognitoIdentity.get_identity_pool_roles

AmazonCognitoIdentity.get_open_id_token

AmazonCognitoIdentity.get_open_id_token_for_developer_identity

AmazonCognitoIdentity.list_identities

AmazonCognitoIdentity.list_identity_pools

AmazonCognitoIdentity.lookup_developer_identity

AmazonCognitoIdentity.merge_developer_identities

AmazonCognitoIdentity.set_identity_pool_roles

AmazonCognitoIdentity.unlink_developer_identity

AmazonCognitoIdentity.unlink_identity

AmazonCognitoIdentity.update_identity_pool

AmazonCognitoIdentity.create_identity_pool — Function.

create_identity_pool(::AWSConfig, arguments::Dict)
create_identity_pool([::AWSConfig]; IdentityPoolName=, AllowUnauthenticatedIdentities=, <keyword arguments>)

CreateIdentityPool Operation

Creates a new identity pool. The identity pool is a store of user identity information that is specific to your AWS account. The limit on identity pools is 60 per account. The keys for SupportedLoginProviders are as follows:

Facebook: graph.facebook.com

Google: accounts.google.com

Amazon: www.amazon.com

Twitter: api.twitter.com

Digits: www.digits.com

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolName = ::String – Required

A string that you provide.

AllowUnauthenticatedIdentities = ::Bool – Required

TRUE if the identity pool supports unauthenticated logins.

SupportedLoginProviders = ::Dict{String,String}

Optional key:value pairs mapping provider names to provider app IDs.

DeveloperProviderName = ::String

The "domain" by which Cognito will refer to your users. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

Once you have set a developer provider name, you cannot change it. Please take care in setting this parameter.

OpenIdConnectProviderARNs = [::String, ...]

A list of OpendID Connect provider ARNs.

CognitoIdentityProviders = [[ ... ], ...]

An array of Amazon Cognito Identity user pools and their client IDs.

 CognitoIdentityProviders = [[
        "ProviderName" =>  ::String,
        "ClientId" =>  ::String,
        "ServerSideTokenCheck" =>  ::Bool
    ], ...]

SamlProviderARNs = [::String, ...]

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

Returns

IdentityPool

Exceptions

InvalidParameterException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException, InternalErrorException or LimitExceededException.

See also: AWS API Documentation

AmazonCognitoIdentity.delete_identities — Function.

delete_identities(::AWSConfig, arguments::Dict)
delete_identities([::AWSConfig]; IdentityIdsToDelete=)

DeleteIdentities Operation

Deletes identities from an identity pool. You can specify a list of 1-60 identities that you want to delete.

You must use AWS Developer credentials to call this API.

Arguments

IdentityIdsToDelete = [::String, ...] – Required

A list of 1-60 identities that you want to delete.

Returns

DeleteIdentitiesResponse

Exceptions

InvalidParameterException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.delete_identity_pool — Function.

delete_identity_pool(::AWSConfig, arguments::Dict)
delete_identity_pool([::AWSConfig]; IdentityPoolId=)

DeleteIdentityPool Operation

Deletes a user pool. Once a pool is deleted, users will not be able to authenticate with the pool.

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.describe_identity — Function.

describe_identity(::AWSConfig, arguments::Dict)
describe_identity([::AWSConfig]; IdentityId=)

DescribeIdentity Operation

Returns metadata related to the given identity, including when the identity was created and any associated linked logins.

You must use AWS Developer credentials to call this API.

Arguments

IdentityId = ::String – Required

A unique identifier in the format REGION:GUID.

Returns

IdentityDescription

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.describe_identity_pool — Function.

describe_identity_pool(::AWSConfig, arguments::Dict)
describe_identity_pool([::AWSConfig]; IdentityPoolId=)

DescribeIdentityPool Operation

Gets details about a particular identity pool, including the pool name, ID description, creation date, and current number of users.

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

Returns

IdentityPool

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.get_credentials_for_identity — Function.

get_credentials_for_identity(::AWSConfig, arguments::Dict)
get_credentials_for_identity([::AWSConfig]; IdentityId=, <keyword arguments>)

GetCredentialsForIdentity Operation

Returns credentials for the provided identity ID. Any provided logins will be validated against supported login providers. If the token is for cognito-identity.amazonaws.com, it will be passed through to AWS Security Token Service with the appropriate role for the token.

This is a public API. You do not need any credentials to call this API.

Arguments

IdentityId = ::String – Required

A unique identifier in the format REGION:GUID.

Logins = ::Dict{String,String}

A set of optional name-value pairs that map provider names to provider tokens.

CustomRoleArn = ::String

The Amazon Resource Name (ARN) of the role to be assumed when multiple roles were received in the token from the identity provider. For example, a SAML-based identity provider. This parameter is optional for identity providers that do not support role customization.

Returns

GetCredentialsForIdentityResponse

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException, InvalidIdentityPoolConfigurationException, InternalErrorException or ExternalServiceException.

See also: AWS API Documentation

AmazonCognitoIdentity.get_id — Function.

get_id(::AWSConfig, arguments::Dict)
get_id([::AWSConfig]; IdentityPoolId=, <keyword arguments>)

GetId Operation

Generates (or retrieves) a Cognito ID. Supplying multiple logins will create an implicit linked account.

This is a public API. You do not need any credentials to call this API.

Arguments

AccountId = ::String

A standard AWS account ID (9+ digits).

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

Logins = ::Dict{String,String}

A set of optional name-value pairs that map provider names to provider tokens. The available provider names for Logins are as follows:

Facebook: graph.facebook.com

Amazon Cognito Identity Provider: cognito-idp.us-east-1.amazonaws.com/us-east-1_123456789

Google: accounts.google.com

Amazon: www.amazon.com

Twitter: api.twitter.com

Digits: www.digits.com

Returns

GetIdResponse

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException, InternalErrorException, LimitExceededException or ExternalServiceException.

See also: AWS API Documentation

AmazonCognitoIdentity.get_identity_pool_roles — Function.

get_identity_pool_roles(::AWSConfig, arguments::Dict)
get_identity_pool_roles([::AWSConfig]; IdentityPoolId=)

GetIdentityPoolRoles Operation

Gets the roles for an identity pool.

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

Returns

GetIdentityPoolRolesResponse

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.get_open_id_token — Function.

get_open_id_token(::AWSConfig, arguments::Dict)
get_open_id_token([::AWSConfig]; IdentityId=, <keyword arguments>)

GetOpenIdToken Operation

Gets an OpenID token, using a known Cognito ID. This known Cognito ID is returned by GetId. You can optionally add additional logins for the identity. Supplying multiple logins creates an implicit link.

The OpenId token is valid for 15 minutes.

This is a public API. You do not need any credentials to call this API.

Arguments

IdentityId = ::String – Required

A unique identifier in the format REGION:GUID.

Logins = ::Dict{String,String}

A set of optional name-value pairs that map provider names to provider tokens. When using graph.facebook.com and www.amazon.com, supply the access_token returned from the provider's authflow. For accounts.google.com, an Amazon Cognito Identity Provider, or any other OpenId Connect provider, always include the id_token.

Returns

GetOpenIdTokenResponse

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException, InternalErrorException or ExternalServiceException.

See also: AWS API Documentation

AmazonCognitoIdentity.get_open_id_token_for_developer_identity — Function.

get_open_id_token_for_developer_identity(::AWSConfig, arguments::Dict)
get_open_id_token_for_developer_identity([::AWSConfig]; IdentityPoolId=, Logins=, <keyword arguments>)

GetOpenIdTokenForDeveloperIdentity Operation

Registers (or retrieves) a Cognito IdentityId and an OpenID Connect token for a user authenticated by your backend authentication process. Supplying multiple logins will create an implicit linked account. You can only specify one developer provider as part of the Logins map, which is linked to the identity pool. The developer provider is the "domain" by which Cognito will refer to your users.

You can use GetOpenIdTokenForDeveloperIdentity to create a new identity and to link new logins (that is, user credentials issued by a public provider or developer provider) to an existing identity. When you want to create a new identity, the IdentityId should be null. When you want to associate a new login with an existing authenticated/unauthenticated identity, you can do so by providing the existing IdentityId. This API will create the identity in the specified IdentityPoolId.

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

IdentityId = ::String

A unique identifier in the format REGION:GUID.

Logins = ::Dict{String,String} – Required

A set of optional name-value pairs that map provider names to provider tokens. Each name-value pair represents a user from a public provider or developer provider. If the user is from a developer provider, the name-value pair will follow the syntax "developer_provider_name": "developer_user_identifier". The developer provider is the "domain" by which Cognito will refer to your users; you provided this domain while creating/updating the identity pool. The developer user identifier is an identifier from your backend that uniquely identifies a user. When you create an identity pool, you can specify the supported logins.

TokenDuration = ::Int

The expiration time of the token, in seconds. You can specify a custom expiration time for the token so that you can cache it. If you don't provide an expiration time, the token is valid for 15 minutes. You can exchange the token with Amazon STS for temporary AWS credentials, which are valid for a maximum of one hour. The maximum token duration you can set is 24 hours. You should take care in setting the expiration time for a token, as there are significant security implications: an attacker could use a leaked token to access your AWS resources for the token's duration.

Returns

GetOpenIdTokenForDeveloperIdentityResponse

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException, InternalErrorException or DeveloperUserAlreadyRegisteredException.

See also: AWS API Documentation

AmazonCognitoIdentity.list_identities — Function.

list_identities(::AWSConfig, arguments::Dict)
list_identities([::AWSConfig]; IdentityPoolId=, MaxResults=, <keyword arguments>)

ListIdentities Operation

Lists the identities in a pool.

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

MaxResults = ::Int – Required

The maximum number of identities to return.

NextToken = ::String

A pagination token.

HideDisabled = ::Bool

An optional boolean parameter that allows you to hide disabled identities. If omitted, the ListIdentities API will include disabled identities in the response.

Returns

ListIdentitiesResponse

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.list_identity_pools — Function.

list_identity_pools(::AWSConfig, arguments::Dict)
list_identity_pools([::AWSConfig]; MaxResults=, <keyword arguments>)

ListIdentityPools Operation

Lists all of the Cognito identity pools registered for your account.

You must use AWS Developer credentials to call this API.

Arguments

MaxResults = ::Int – Required

The maximum number of identities to return.

NextToken = ::String

A pagination token.

Returns

ListIdentityPoolsResponse

Exceptions

InvalidParameterException, NotAuthorizedException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.lookup_developer_identity — Function.

lookup_developer_identity(::AWSConfig, arguments::Dict)
lookup_developer_identity([::AWSConfig]; IdentityPoolId=, <keyword arguments>)

LookupDeveloperIdentity Operation

Retrieves the IdentityID associated with a DeveloperUserIdentifier or the list of DeveloperUserIdentifiers associated with an IdentityId for an existing identity. Either IdentityID or DeveloperUserIdentifier must not be null. If you supply only one of these values, the other value will be searched in the database and returned as a part of the response. If you supply both, DeveloperUserIdentifier will be matched against IdentityID. If the values are verified against the database, the response returns both values and is the same as the request. Otherwise a ResourceConflictException is thrown.

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

IdentityId = ::String

A unique identifier in the format REGION:GUID.

DeveloperUserIdentifier = ::String

A unique ID used by your backend authentication process to identify a user. Typically, a developer identity provider would issue many developer user identifiers, in keeping with the number of users.

MaxResults = ::Int

The maximum number of identities to return.

NextToken = ::String

A pagination token. The first call you make will have NextToken set to null. After that the service will return NextToken values as needed. For example, let's say you make a request with MaxResults set to 10, and there are 20 matches in the database. The service will return a pagination token as a part of the response. This token can be used to call the API again and get results starting from the 11th match.

Returns

LookupDeveloperIdentityResponse

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.merge_developer_identities — Function.

merge_developer_identities(::AWSConfig, arguments::Dict)
merge_developer_identities([::AWSConfig]; SourceUserIdentifier=, DestinationUserIdentifier=, DeveloperProviderName=, IdentityPoolId=)

MergeDeveloperIdentities Operation

Merges two users having different IdentityIds, existing in the same identity pool, and identified by the same developer provider. You can use this action to request that discrete users be merged and identified as a single user in the Cognito environment. Cognito associates the given source user (SourceUserIdentifier) with the IdentityId of the DestinationUserIdentifier. Only developer-authenticated users can be merged. If the users to be merged are associated with the same public provider, but as two different users, an exception will be thrown.

You must use AWS Developer credentials to call this API.

Arguments

SourceUserIdentifier = ::String – Required

User identifier for the source user. The value should be a DeveloperUserIdentifier.

DestinationUserIdentifier = ::String – Required

User identifier for the destination user. The value should be a DeveloperUserIdentifier.

DeveloperProviderName = ::String – Required

The "domain" by which Cognito will refer to your users. This is a (pseudo) domain name that you provide while creating an identity pool. This name acts as a placeholder that allows your backend and the Cognito service to communicate about the developer provider. For the DeveloperProviderName, you can use letters as well as period (.), underscore (_), and dash (-).

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

Returns

MergeDeveloperIdentitiesResponse

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.set_identity_pool_roles — Function.

set_identity_pool_roles(::AWSConfig, arguments::Dict)
set_identity_pool_roles([::AWSConfig]; IdentityPoolId=, Roles=, <keyword arguments>)

SetIdentityPoolRoles Operation

Sets the roles for an identity pool. These roles are used when making calls to GetCredentialsForIdentity action.

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

Roles = ::Dict{String,String} – Required

The map of roles associated with this pool. For a given role, the key will be either "authenticated" or "unauthenticated" and the value will be the Role ARN.

RoleMappings = ::Dict{String,String}

How users for a specific identity provider are to mapped to roles. This is a string to RoleMapping object map. The string identifies the identity provider, for example, "graph.facebook.com" or "cognito-idp-east-1.amazonaws.com/us-east-1_abcdefghi:app_client_id".

Up to 25 rules can be specified per identity provider.

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException, InternalErrorException or ConcurrentModificationException.

See also: AWS API Documentation

AmazonCognitoIdentity.unlink_developer_identity — Function.

unlink_developer_identity(::AWSConfig, arguments::Dict)
unlink_developer_identity([::AWSConfig]; IdentityId=, IdentityPoolId=, DeveloperProviderName=, DeveloperUserIdentifier=)

UnlinkDeveloperIdentity Operation

Unlinks a DeveloperUserIdentifier from an existing identity. Unlinked developer users will be considered new identities next time they are seen. If, for a given Cognito identity, you remove all federated identities as well as the developer user identifier, the Cognito identity becomes inaccessible.

You must use AWS Developer credentials to call this API.

Arguments

IdentityId = ::String – Required

A unique identifier in the format REGION:GUID.

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

DeveloperProviderName = ::String – Required

The "domain" by which Cognito will refer to your users.

DeveloperUserIdentifier = ::String – Required

A unique ID used by your backend authentication process to identify a user.

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException or InternalErrorException.

See also: AWS API Documentation

AmazonCognitoIdentity.unlink_identity — Function.

unlink_identity(::AWSConfig, arguments::Dict)
unlink_identity([::AWSConfig]; IdentityId=, Logins=, LoginsToRemove=)

UnlinkIdentity Operation

Unlinks a federated identity from an existing account. Unlinked logins will be considered new identities next time they are seen. Removing the last linked login will make this identity inaccessible.

This is a public API. You do not need any credentials to call this API.

Arguments

IdentityId = ::String – Required

A unique identifier in the format REGION:GUID.

Logins = ::Dict{String,String} – Required

A set of optional name-value pairs that map provider names to provider tokens.

LoginsToRemove = [::String, ...] – Required

Provider names to unlink from this identity.

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException, InternalErrorException or ExternalServiceException.

See also: AWS API Documentation

AmazonCognitoIdentity.update_identity_pool — Function.

update_identity_pool(::AWSConfig, arguments::Dict)
update_identity_pool([::AWSConfig]; IdentityPoolId=, IdentityPoolName=, AllowUnauthenticatedIdentities=, <keyword arguments>)

UpdateIdentityPool Operation

Updates a user pool.

You must use AWS Developer credentials to call this API.

Arguments

IdentityPoolId = ::String – Required

An identity pool ID in the format REGION:GUID.

IdentityPoolName = ::String – Required

A string that you provide.

AllowUnauthenticatedIdentities = ::Bool – Required

TRUE if the identity pool supports unauthenticated logins.

SupportedLoginProviders = ::Dict{String,String}

Optional key:value pairs mapping provider names to provider app IDs.

DeveloperProviderName = ::String

The "domain" by which Cognito will refer to your users.

OpenIdConnectProviderARNs = [::String, ...]

A list of OpendID Connect provider ARNs.

CognitoIdentityProviders = [[ ... ], ...]

A list representing an Amazon Cognito Identity User Pool and its client ID.

 CognitoIdentityProviders = [[
        "ProviderName" =>  ::String,
        "ClientId" =>  ::String,
        "ServerSideTokenCheck" =>  ::Bool
    ], ...]

SamlProviderARNs = [::String, ...]

An array of Amazon Resource Names (ARNs) of the SAML provider for your identity pool.

Returns

IdentityPool

Exceptions

InvalidParameterException, ResourceNotFoundException, NotAuthorizedException, ResourceConflictException, TooManyRequestsException, InternalErrorException, ConcurrentModificationException or LimitExceededException.

See also: AWS API Documentation