AmazonDMS.jl

AmazonDMS

AWS Database Migration Service

AWS Database Migration Service (AWS DMS) can migrate your data to and from the most widely used commercial and open-source databases such as Oracle, PostgreSQL, Microsoft SQL Server, Amazon Redshift, MariaDB, Amazon Aurora, MySQL, and SAP Adaptive Server Enterprise (ASE). The service supports homogeneous migrations such as Oracle to Oracle, as well as heterogeneous migrations between different database platforms, such as Oracle to MySQL or SQL Server to PostgreSQL.

For more information about AWS DMS, see the AWS DMS user guide at What Is AWS Database Migration Service?

This document is generated from apis/dms-2016-01-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonDMS.add_tags_to_resource

AmazonDMS.create_endpoint

AmazonDMS.create_event_subscription

AmazonDMS.create_replication_instance

AmazonDMS.create_replication_subnet_group

AmazonDMS.create_replication_task

AmazonDMS.delete_certificate

AmazonDMS.delete_endpoint

AmazonDMS.delete_event_subscription

AmazonDMS.delete_replication_instance

AmazonDMS.delete_replication_subnet_group

AmazonDMS.delete_replication_task

AmazonDMS.describe_account_attributes

AmazonDMS.describe_certificates

AmazonDMS.describe_connections

AmazonDMS.describe_endpoint_types

AmazonDMS.describe_endpoints

AmazonDMS.describe_event_categories

AmazonDMS.describe_event_subscriptions

AmazonDMS.describe_events

AmazonDMS.describe_orderable_replication_instances

AmazonDMS.describe_refresh_schemas_status

AmazonDMS.describe_replication_instances

AmazonDMS.describe_replication_subnet_groups

AmazonDMS.describe_replication_tasks

AmazonDMS.describe_schemas

AmazonDMS.describe_table_statistics

AmazonDMS.import_certificate

AmazonDMS.list_tags_for_resource

AmazonDMS.modify_endpoint

AmazonDMS.modify_event_subscription

AmazonDMS.modify_replication_instance

AmazonDMS.modify_replication_subnet_group

AmazonDMS.modify_replication_task

AmazonDMS.refresh_schemas

AmazonDMS.reload_tables

AmazonDMS.remove_tags_from_resource

AmazonDMS.start_replication_task

AmazonDMS.stop_replication_task

AmazonDMS.test_connection

AmazonDMS.add_tags_to_resource — Method.

add_tags_to_resource(::AWSConfig, arguments::Dict)
add_tags_to_resource(::AWSConfig; ResourceArn=, Tags=)

AddTagsToResource Operation

Adds metadata tags to a DMS resource, including replication instance, endpoint, security group, and migration task. These tags can also be used with cost allocation reporting to track cost associated with DMS resources, or used in a Condition statement in an IAM policy for DMS.

Arguments

ResourceArn::String – Required

The Amazon Resource Name (ARN) of the AWS DMS resource the tag is to be added to. AWS DMS resources include a replication instance, endpoint, and a replication task.

Tags: – Required

The tag to be assigned to the DMS resource.

    [Dict(
        Key  => ::String,
        Value  => ::String
    )...]

Returns

AddTagsToResourceResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.create_endpoint — Method.

create_endpoint(::AWSConfig, arguments::Dict)
create_endpoint(::AWSConfig; EndpointIdentifier=, EndpointType=, EngineName=, <keyword arguments>)

CreateEndpoint Operation

Creates an endpoint using the provided settings.

Arguments

EndpointIdentifier::String – Required

The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

EndpointType: "source" or "target" – Required

The type of endpoint.

EngineName::String – Required

The type of engine for the endpoint. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, SYBASE, DYNAMODB, MONGODB, and SQLSERVER.

Username::String

The user name to be used to login to the endpoint database.

Password::String

The password to be used to login to the endpoint database.

ServerName::String

The name of the server where the endpoint database resides.

Port::Int

The port used by the endpoint database.

DatabaseName::String

The name of the endpoint database.

ExtraConnectionAttributes::String

Additional attributes associated with the connection.

KmsKeyId::String

The KMS key identifier that will be used to encrypt the connection parameters. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

Tags:

Tags to be added to the endpoint.

    [Dict(
        Key  => ::String,
        Value  => ::String
    )...]

CertificateArn::String

The Amazon Resource Number (ARN) for the certificate.

SslMode: "none", "require", "verify-ca" or "verify-full"

The SSL mode to use for the SSL connection.

SSL mode can be one of four values: none, require, verify-ca, verify-full.

The default value is none.

DynamoDbSettings::String *

Settings in JSON format for the target Amazon DynamoDB endpoint. For more information about the available settings, see the Using Object Mapping to Migrate Data to DynamoDB section at Using an Amazon DynamoDB Database as a Target for AWS Database Migration Service.

S3Settings:

Settings in JSON format for the target S3 endpoint. For more information about the available settings, see the Extra Connection Attributes section at Using Amazon S3 as a Target for AWS Database Migration Service.

    Dict(
        ServiceAccessRoleArn  => ::String,
        ExternalTableDefinition  => ::String,
        CsvRowDelimiter  => ::String,
        CsvDelimiter  => ::String,
        BucketFolder  => ::String,
        BucketName  => ::String,
        CompressionType  => : "none" or "gzip"
    )

MongoDbSettings:

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the Configuration Properties When Using MongoDB as a Source for AWS Database Migration Service section at Using Amazon S3 as a Target for AWS Database Migration Service.

    Dict(
        Username  => ::String,
        Password  => ::String,
        ServerName  => ::String,
        Port  => ::Int,
        DatabaseName  => ::String,
        AuthType  => : "no" or "password",
        AuthMechanism  => : "default", "mongodb_cr" or "scram_sha_1",
        NestingLevel  => : "none" or "one",
        ExtractDocId  => ::String,
        DocsToInvestigate  => ::String,
        AuthSource  => ::String
    )

Returns

CreateEndpointResponse

Exceptions

KMSKeyNotAccessibleFault, ResourceAlreadyExistsFault, ResourceQuotaExceededFault, InvalidResourceStateFault, ResourceNotFoundFault or AccessDeniedFault.

See also: AWS API Documentation

AmazonDMS.create_event_subscription — Method.

create_event_subscription(::AWSConfig, arguments::Dict)
create_event_subscription(::AWSConfig; SubscriptionName=, SnsTopicArn=, <keyword arguments>)

CreateEventSubscription Operation

Creates an AWS DMS event notification subscription.

You can specify the type of source (SourceType) you want to be notified of, provide a list of AWS DMS source IDs (SourceIds) that triggers the events, and provide a list of event categories (EventCategories) for events you want to be notified of. If you specify both the SourceType and SourceIds, such as SourceType = replication-instance and SourceIdentifier = my-replinstance, you will be notified of all the replication instance events for the specified source. If you specify a SourceType but don't specify a SourceIdentifier, you receive notice of the events for that source type for all your AWS DMS sources. If you don't specify either SourceType nor SourceIdentifier, you will be notified of events generated from all AWS DMS sources belonging to your customer account.

For more information about AWS DMS events, see Working with Events and Notifications in the AWS Database MIgration Service User Guide.

Arguments

SubscriptionName::String – Required

The name of the DMS event notification subscription.

Constraints: The name must be less than 255 characters.

SnsTopicArn::String – Required

The Amazon Resource Name (ARN) of the Amazon SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

SourceType::String

The type of AWS DMS resource that generates the events. For example, if you want to be notified of events generated by a replication instance, you set this parameter to replication-instance. If this value is not specified, all events are returned.

Valid values: replication-instance | migration-task

EventCategories: [::String...]

A list of event categories for a source type that you want to subscribe to. You can see a list of the categories for a given source type by calling the DescribeEventCategories action or in the topic Working with Events and Notifications in the AWS Database Migration Service User Guide.

SourceIds: [::String...]

The list of identifiers of the event sources for which events will be returned. If not specified, then all sources are included in the response. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it cannot end with a hyphen or contain two consecutive hyphens.

Enabled::Bool

A Boolean value; set to true to activate the subscription, or set to false to create the subscription but not activate it.

Tags:

A tag to be attached to the event subscription.

    [Dict(
        Key  => ::String,
        Value  => ::String
    )...]

Returns

CreateEventSubscriptionResponse

Exceptions

ResourceQuotaExceededFault, ResourceAlreadyExistsFault, SNSInvalidTopicFault, SNSNoAuthorizationFault or ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.create_replication_instance — Method.

create_replication_instance(::AWSConfig, arguments::Dict)
create_replication_instance(::AWSConfig; ReplicationInstanceIdentifier=, ReplicationInstanceClass=, <keyword arguments>)

CreateReplicationInstance Operation

Creates the replication instance using the specified parameters.

Arguments

ReplicationInstanceIdentifier::String – Required

The replication instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: myrepinstance

AllocatedStorage::Int

The amount of storage (in gigabytes) to be initially allocated for the replication instance.

ReplicationInstanceClass::String – Required

The compute and memory capacity of the replication instance as specified by the replication instance class.

Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

VpcSecurityGroupIds: [::String...]

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

AvailabilityZone::String

The EC2 Availability Zone that the replication instance will be created in.

Default: A random, system-chosen Availability Zone in the endpoint's region.

Example: us-east-1d

ReplicationSubnetGroupIdentifier::String

A subnet group to associate with the replication instance.

PreferredMaintenanceWindow::String

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun

Constraints: Minimum 30-minute window.

MultiAZ::Bool

Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

EngineVersion::String

The engine version number of the replication instance.

AutoMinorVersionUpgrade::Bool

Indicates that minor engine upgrades will be applied automatically to the replication instance during the maintenance window.

Default: true

Tags:

Tags to be associated with the replication instance.

    [Dict(
        Key  => ::String,
        Value  => ::String
    )...]

KmsKeyId::String

The KMS key identifier that will be used to encrypt the content on the replication instance. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

PubliclyAccessible::Bool

Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true.

Returns

CreateReplicationInstanceResponse

Exceptions

AccessDeniedFault, ResourceAlreadyExistsFault, InsufficientResourceCapacityFault, ResourceQuotaExceededFault, StorageQuotaExceededFault, ResourceNotFoundFault, ReplicationSubnetGroupDoesNotCoverEnoughAZs, InvalidResourceStateFault, InvalidSubnet or KMSKeyNotAccessibleFault.

See also: AWS API Documentation

AmazonDMS.create_replication_subnet_group — Method.

create_replication_subnet_group(::AWSConfig, arguments::Dict)
create_replication_subnet_group(::AWSConfig; ReplicationSubnetGroupIdentifier=, ReplicationSubnetGroupDescription=, SubnetIds=, <keyword arguments>)

CreateReplicationSubnetGroup Operation

Creates a replication subnet group given a list of the subnet IDs in a VPC.

Arguments

ReplicationSubnetGroupIdentifier::String – Required

The name for the replication subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 alphanumeric characters, periods, spaces, underscores, or hyphens. Must not be "default".

Example: mySubnetgroup

ReplicationSubnetGroupDescription::String – Required

The description for the subnet group.

SubnetIds: [::String...] – Required

The EC2 subnet IDs for the subnet group.

Tags:

The tag to be assigned to the subnet group.

    [Dict(
        Key  => ::String,
        Value  => ::String
    )...]

Returns

CreateReplicationSubnetGroupResponse

Exceptions

AccessDeniedFault, ResourceAlreadyExistsFault, ResourceNotFoundFault, ResourceQuotaExceededFault, ReplicationSubnetGroupDoesNotCoverEnoughAZs or InvalidSubnet.

See also: AWS API Documentation

AmazonDMS.create_replication_task — Method.

create_replication_task(::AWSConfig, arguments::Dict)
create_replication_task(::AWSConfig; ReplicationTaskIdentifier=, SourceEndpointArn=, TargetEndpointArn=, ReplicationInstanceArn=, MigrationType=, TableMappings=, <keyword arguments>)

CreateReplicationTask Operation

Creates a replication task using the specified parameters.

Arguments

ReplicationTaskIdentifier::String – Required

The replication task identifier.

Constraints:

Must contain from 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

SourceEndpointArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

TargetEndpointArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

ReplicationInstanceArn::String – Required

The Amazon Resource Name (ARN) of the replication instance.

MigrationType: "full-load", "cdc" or "full-load-and-cdc" – Required

The migration type.

TableMappings::String – Required

When using the AWS CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with "file://". When working with the DMS API, provide the JSON as the parameter value.

For example, –table-mappings file://mappingfile.json

ReplicationTaskSettings::String

Settings for the task, such as target metadata settings. For a complete list of task settings, see Task Settings for AWS Database Migration Service Tasks.

CdcStartTime: timestamp

The start time for the Change Data Capture (CDC) operation.

Tags:

Tags to be added to the replication instance.

    [Dict(
        Key  => ::String,
        Value  => ::String
    )...]

Returns

CreateReplicationTaskResponse

Exceptions

AccessDeniedFault, InvalidResourceStateFault, ResourceAlreadyExistsFault, ResourceNotFoundFault, KMSKeyNotAccessibleFault or ResourceQuotaExceededFault.

See also: AWS API Documentation

AmazonDMS.delete_certificate — Method.

delete_certificate(::AWSConfig, arguments::Dict)
delete_certificate(::AWSConfig; CertificateArn=)

DeleteCertificate Operation

Deletes the specified certificate.

Arguments

CertificateArn::String – Required

The Amazon Resource Name (ARN) of the deleted certificate.

Returns

DeleteCertificateResponse

Exceptions

ResourceNotFoundFault or InvalidResourceStateFault.

See also: AWS API Documentation

AmazonDMS.delete_endpoint — Method.

delete_endpoint(::AWSConfig, arguments::Dict)
delete_endpoint(::AWSConfig; EndpointArn=)

DeleteEndpoint Operation

Deletes the specified endpoint.

Note

All tasks associated with the endpoint must be deleted before you can delete the endpoint.

Arguments

EndpointArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Returns

DeleteEndpointResponse

Exceptions

ResourceNotFoundFault or InvalidResourceStateFault.

See also: AWS API Documentation

AmazonDMS.delete_event_subscription — Method.

delete_event_subscription(::AWSConfig, arguments::Dict)
delete_event_subscription(::AWSConfig; SubscriptionName=)

DeleteEventSubscription Operation

Deletes an AWS DMS event subscription.

Arguments

SubscriptionName::String – Required

The name of the DMS event notification subscription to be deleted.

Returns

DeleteEventSubscriptionResponse

Exceptions

ResourceNotFoundFault or InvalidResourceStateFault.

See also: AWS API Documentation

AmazonDMS.delete_replication_instance — Method.

delete_replication_instance(::AWSConfig, arguments::Dict)
delete_replication_instance(::AWSConfig; ReplicationInstanceArn=)

DeleteReplicationInstance Operation

Deletes the specified replication instance.

Note

You must delete any migration tasks that are associated with the replication instance before you can delete it.

Arguments

ReplicationInstanceArn::String – Required

The Amazon Resource Name (ARN) of the replication instance to be deleted.

Returns

DeleteReplicationInstanceResponse

Exceptions

InvalidResourceStateFault or ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.delete_replication_subnet_group — Method.

delete_replication_subnet_group(::AWSConfig, arguments::Dict)
delete_replication_subnet_group(::AWSConfig; ReplicationSubnetGroupIdentifier=)

DeleteReplicationSubnetGroup Operation

Deletes a subnet group.

Arguments

ReplicationSubnetGroupIdentifier::String – Required

The subnet group name of the replication instance.

Returns

DeleteReplicationSubnetGroupResponse

Exceptions

InvalidResourceStateFault or ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.delete_replication_task — Method.

delete_replication_task(::AWSConfig, arguments::Dict)
delete_replication_task(::AWSConfig; ReplicationTaskArn=)

DeleteReplicationTask Operation

Deletes the specified replication task.

Arguments

ReplicationTaskArn::String – Required

The Amazon Resource Name (ARN) of the replication task to be deleted.

Returns

DeleteReplicationTaskResponse

Exceptions

ResourceNotFoundFault or InvalidResourceStateFault.

See also: AWS API Documentation

AmazonDMS.describe_account_attributes — Method.

describe_account_attributes(::AWSConfig, arguments::Dict)
describe_account_attributes(::AWSConfig; )

DescribeAccountAttributes Operation

Lists all of the AWS DMS attributes for a customer account. The attributes include AWS DMS quotas for the account, such as the number of replication instances allowed. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value.

This command does not take any parameters.

Arguments

Returns

DescribeAccountAttributesResponse

See also: AWS API Documentation

AmazonDMS.describe_certificates — Method.

describe_certificates(::AWSConfig, arguments::Dict)
describe_certificates(::AWSConfig; <keyword arguments>)

DescribeCertificates Operation

Provides a description of the certificate.

Arguments

Filters:

Filters applied to the certificate described in the form of key-value pairs.

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 10

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeCertificatesResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_connections — Method.

describe_connections(::AWSConfig, arguments::Dict)
describe_connections(::AWSConfig; <keyword arguments>)

DescribeConnections Operation

Describes the status of the connections that have been made between the replication instance and an endpoint. Connections are created when you test an endpoint.

Arguments

Filters:

The filters applied to the connection.

Valid filter names: endpoint-arn | replication-instance-arn

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeConnectionsResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_endpoint_types — Method.

describe_endpoint_types(::AWSConfig, arguments::Dict)
describe_endpoint_types(::AWSConfig; <keyword arguments>)

DescribeEndpointTypes Operation

Returns information about the type of endpoints available.

Arguments

Filters:

Filters applied to the describe action.

Valid filter names: engine-name | endpoint-type

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeEndpointTypesResponse

See also: AWS API Documentation

AmazonDMS.describe_endpoints — Method.

describe_endpoints(::AWSConfig, arguments::Dict)
describe_endpoints(::AWSConfig; <keyword arguments>)

DescribeEndpoints Operation

Returns information about the endpoints for your account in the current region.

Arguments

Filters:

Filters applied to the describe action.

Valid filter names: endpoint-arn | endpoint-type | endpoint-id | engine-name

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeEndpointsResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_event_categories — Method.

describe_event_categories(::AWSConfig, arguments::Dict)
describe_event_categories(::AWSConfig; <keyword arguments>)

DescribeEventCategories Operation

Lists categories for all event source types, or, if specified, for a specified source type. You can see a list of the event categories and source types in Working with Events and Notifications in the AWS Database Migration Service User Guide.

Arguments

SourceType::String

The type of AWS DMS resource that generates events.

Valid values: replication-instance | migration-task

Filters:

Filters applied to the action.

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

Returns

DescribeEventCategoriesResponse

See also: AWS API Documentation

AmazonDMS.describe_event_subscriptions — Method.

describe_event_subscriptions(::AWSConfig, arguments::Dict)
describe_event_subscriptions(::AWSConfig; <keyword arguments>)

DescribeEventSubscriptions Operation

Lists all the event subscriptions for a customer account. The description of a subscription includes SubscriptionName, SNSTopicARN, CustomerID, SourceType, SourceID, CreationTime, and Status.

If you specify SubscriptionName, this action lists the description for that subscription.

Arguments

SubscriptionName::String

The name of the AWS DMS event subscription to be described.

Filters:

Filters applied to the action.

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeEventSubscriptionsResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_events — Method.

describe_events(::AWSConfig, arguments::Dict)
describe_events(::AWSConfig; <keyword arguments>)

DescribeEvents Operation

Lists events for a given source identifier and source type. You can also specify a start and end time. For more information on AWS DMS events, see Working with Events and Notifications .

Arguments

SourceIdentifier::String

The identifier of the event source. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens. It cannot end with a hyphen or contain two consecutive hyphens.

SourceType: "replication-instance"

The type of AWS DMS resource that generates events.

Valid values: replication-instance | migration-task

StartTime: timestamp

The start time for the events to be listed.

EndTime: timestamp

The end time for the events to be listed.

Duration::Int

The duration of the events to be listed.

EventCategories: [::String...]

A list of event categories for a source type that you want to subscribe to.

Filters:

Filters applied to the action.

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeEventsResponse

See also: AWS API Documentation

AmazonDMS.describe_orderable_replication_instances — Method.

describe_orderable_replication_instances(::AWSConfig, arguments::Dict)
describe_orderable_replication_instances(::AWSConfig; <keyword arguments>)

DescribeOrderableReplicationInstances Operation

Returns information about the replication instance types that can be created in the specified region.

Arguments

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeOrderableReplicationInstancesResponse

See also: AWS API Documentation

AmazonDMS.describe_refresh_schemas_status — Method.

describe_refresh_schemas_status(::AWSConfig, arguments::Dict)
describe_refresh_schemas_status(::AWSConfig; EndpointArn=)

DescribeRefreshSchemasStatus Operation

Returns the status of the RefreshSchemas operation.

Arguments

EndpointArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Returns

DescribeRefreshSchemasStatusResponse

Exceptions

InvalidResourceStateFault or ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_replication_instances — Method.

describe_replication_instances(::AWSConfig, arguments::Dict)
describe_replication_instances(::AWSConfig; <keyword arguments>)

DescribeReplicationInstances Operation

Returns information about replication instances for your account in the current region.

Arguments

Filters:

Filters applied to the describe action.

Valid filter names: replication-instance-arn | replication-instance-id | replication-instance-class | engine-version

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeReplicationInstancesResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_replication_subnet_groups — Method.

describe_replication_subnet_groups(::AWSConfig, arguments::Dict)
describe_replication_subnet_groups(::AWSConfig; <keyword arguments>)

DescribeReplicationSubnetGroups Operation

Returns information about the replication subnet groups.

Arguments

Filters:

Filters applied to the describe action.

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeReplicationSubnetGroupsResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_replication_tasks — Method.

describe_replication_tasks(::AWSConfig, arguments::Dict)
describe_replication_tasks(::AWSConfig; <keyword arguments>)

DescribeReplicationTasks Operation

Returns information about replication tasks for your account in the current region.

Arguments

Filters:

Filters applied to the describe action.

Valid filter names: replication-task-arn | replication-task-id | migration-type | endpoint-arn | replication-instance-arn

    [Dict(
        Name * => ::String,
        Values * => [::String...]
    )...]

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeReplicationTasksResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_schemas — Method.

describe_schemas(::AWSConfig, arguments::Dict)
describe_schemas(::AWSConfig; EndpointArn=, <keyword arguments>)

DescribeSchemas Operation

Returns information about the schema for the specified endpoint.

Arguments

EndpointArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeSchemasResponse

Exceptions

InvalidResourceStateFault or ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.describe_table_statistics — Method.

describe_table_statistics(::AWSConfig, arguments::Dict)
describe_table_statistics(::AWSConfig; ReplicationTaskArn=, <keyword arguments>)

DescribeTableStatistics Operation

Returns table statistics on the database migration task, including table name, rows inserted, rows updated, and rows deleted.

Arguments

ReplicationTaskArn::String – Required

The Amazon Resource Name (ARN) of the replication task.

MaxRecords::Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

Marker::String

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Returns

DescribeTableStatisticsResponse

Exceptions

ResourceNotFoundFault or InvalidResourceStateFault.

See also: AWS API Documentation

AmazonDMS.import_certificate — Method.

import_certificate(::AWSConfig, arguments::Dict)
import_certificate(::AWSConfig; CertificateIdentifier=, <keyword arguments>)

ImportCertificate Operation

Uploads the specified certificate.

Arguments

CertificateIdentifier::String – Required

The customer-assigned name of the certificate. Valid characters are A-z and 0-9.

CertificatePem::String

The contents of the .pem X.509 certificate file for the certificate.

CertificateWallet: blob

The location of the imported Oracle Wallet certificate for use with SSL.

Tags:

The tags associated with the certificate.

    [Dict(
        Key  => ::String,
        Value  => ::String
    )...]

Returns

ImportCertificateResponse

Exceptions

ResourceAlreadyExistsFault or InvalidCertificateFault.

See also: AWS API Documentation

AmazonDMS.list_tags_for_resource — Method.

list_tags_for_resource(::AWSConfig, arguments::Dict)
list_tags_for_resource(::AWSConfig; ResourceArn=)

ListTagsForResource Operation

Lists all tags for an AWS DMS resource.

Arguments

ResourceArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the AWS DMS resource.

Returns

ListTagsForResourceResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.modify_endpoint — Method.

modify_endpoint(::AWSConfig, arguments::Dict)
modify_endpoint(::AWSConfig; EndpointArn=, <keyword arguments>)

ModifyEndpoint Operation

Modifies the specified endpoint.

Arguments

EndpointArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

EndpointIdentifier::String

The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

EndpointType: "source" or "target"

The type of endpoint.

EngineName::String

The type of engine for the endpoint. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, DYNAMODB, MONGODB, SYBASE, and SQLSERVER.

Username::String

The user name to be used to login to the endpoint database.

Password::String

The password to be used to login to the endpoint database.

ServerName::String

The name of the server where the endpoint database resides.

Port::Int

The port used by the endpoint database.

DatabaseName::String

The name of the endpoint database.

ExtraConnectionAttributes::String

Additional attributes associated with the connection.

CertificateArn::String

The Amazon Resource Name (ARN) of the certificate used for SSL connection.

SslMode: "none", "require", "verify-ca" or "verify-full"

The SSL mode to be used.

SSL mode can be one of four values: none, require, verify-ca, verify-full.

The default value is none.

DynamoDbSettings::String *

Settings in JSON format for the target Amazon DynamoDB endpoint. For more information about the available settings, see the Using Object Mapping to Migrate Data to DynamoDB section at Using an Amazon DynamoDB Database as a Target for AWS Database Migration Service.

S3Settings:

Settings in JSON format for the target S3 endpoint. For more information about the available settings, see the Extra Connection Attributes section at Using Amazon S3 as a Target for AWS Database Migration Service.

    Dict(
        ServiceAccessRoleArn  => ::String,
        ExternalTableDefinition  => ::String,
        CsvRowDelimiter  => ::String,
        CsvDelimiter  => ::String,
        BucketFolder  => ::String,
        BucketName  => ::String,
        CompressionType  => : "none" or "gzip"
    )

MongoDbSettings:

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the Configuration Properties When Using MongoDB as a Source for AWS Database Migration Service section at Using Amazon S3 as a Target for AWS Database Migration Service.

    Dict(
        Username  => ::String,
        Password  => ::String,
        ServerName  => ::String,
        Port  => ::Int,
        DatabaseName  => ::String,
        AuthType  => : "no" or "password",
        AuthMechanism  => : "default", "mongodb_cr" or "scram_sha_1",
        NestingLevel  => : "none" or "one",
        ExtractDocId  => ::String,
        DocsToInvestigate  => ::String,
        AuthSource  => ::String
    )

Returns

ModifyEndpointResponse

Exceptions

InvalidResourceStateFault, ResourceNotFoundFault, ResourceAlreadyExistsFault, KMSKeyNotAccessibleFault or AccessDeniedFault.

See also: AWS API Documentation

AmazonDMS.modify_event_subscription — Method.

modify_event_subscription(::AWSConfig, arguments::Dict)
modify_event_subscription(::AWSConfig; SubscriptionName=, <keyword arguments>)

ModifyEventSubscription Operation

Modifies an existing AWS DMS event notification subscription.

Arguments

SubscriptionName::String – Required

The name of the AWS DMS event notification subscription to be modified.

SnsTopicArn::String

The Amazon Resource Name (ARN) of the Amazon SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

SourceType::String

The type of AWS DMS resource that generates the events you want to subscribe to.

Valid values: replication-instance | migration-task

EventCategories: [::String...]

A list of event categories for a source type that you want to subscribe to. Use the DescribeEventCategories action to see a list of event categories.

Enabled::Bool

A Boolean value; set to true to activate the subscription.

Returns

ModifyEventSubscriptionResponse

Exceptions

ResourceQuotaExceededFault, ResourceNotFoundFault, SNSInvalidTopicFault or SNSNoAuthorizationFault.

See also: AWS API Documentation

AmazonDMS.modify_replication_instance — Method.

modify_replication_instance(::AWSConfig, arguments::Dict)
modify_replication_instance(::AWSConfig; ReplicationInstanceArn=, <keyword arguments>)

ModifyReplicationInstance Operation

Modifies the replication instance to apply new settings. You can change one or more parameters by specifying these parameters and the new values in the request.

Some settings are applied during the maintenance window.

Arguments

ReplicationInstanceArn::String – Required

The Amazon Resource Name (ARN) of the replication instance.

AllocatedStorage::Int

The amount of storage (in gigabytes) to be allocated for the replication instance.

ApplyImmediately::Bool

Indicates whether the changes should be applied immediately or during the next maintenance window.

ReplicationInstanceClass::String

The compute and memory capacity of the replication instance.

Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

VpcSecurityGroupIds: [::String...]

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

PreferredMaintenanceWindow::String

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter does not result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

MultiAZ::Bool

Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

EngineVersion::String

The engine version number of the replication instance.

AllowMajorVersionUpgrade::Bool

Indicates that major version upgrades are allowed. Changing this parameter does not result in an outage and the change is asynchronously applied as soon as possible.

Constraints: This parameter must be set to true when specifying a value for the EngineVersion parameter that is a different major version than the replication instance's current version.

AutoMinorVersionUpgrade::Bool

Indicates that minor version upgrades will be applied automatically to the replication instance during the maintenance window. Changing this parameter does not result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and AWS DMS has enabled auto patching for that engine version.

ReplicationInstanceIdentifier::String

The replication instance identifier. This parameter is stored as a lowercase string.

Returns

ModifyReplicationInstanceResponse

Exceptions

InvalidResourceStateFault, ResourceAlreadyExistsFault, ResourceNotFoundFault, InsufficientResourceCapacityFault, StorageQuotaExceededFault or UpgradeDependencyFailureFault.

See also: AWS API Documentation

AmazonDMS.modify_replication_subnet_group — Method.

modify_replication_subnet_group(::AWSConfig, arguments::Dict)
modify_replication_subnet_group(::AWSConfig; ReplicationSubnetGroupIdentifier=, SubnetIds=, <keyword arguments>)

ModifyReplicationSubnetGroup Operation

Modifies the settings for the specified replication subnet group.

Arguments

ReplicationSubnetGroupIdentifier::String – Required

The name of the replication instance subnet group.

ReplicationSubnetGroupDescription::String

The description of the replication instance subnet group.

SubnetIds: [::String...] – Required

A list of subnet IDs.

Returns

ModifyReplicationSubnetGroupResponse

Exceptions

AccessDeniedFault, ResourceNotFoundFault, ResourceQuotaExceededFault, SubnetAlreadyInUse, ReplicationSubnetGroupDoesNotCoverEnoughAZs or InvalidSubnet.

See also: AWS API Documentation

AmazonDMS.modify_replication_task — Method.

modify_replication_task(::AWSConfig, arguments::Dict)
modify_replication_task(::AWSConfig; ReplicationTaskArn=, <keyword arguments>)

ModifyReplicationTask Operation

Modifies the specified replication task.

You can't modify the task endpoints. The task must be stopped before you can modify it.

For more information about AWS DMS tasks, see the AWS DMS user guide at Working with Migration Tasks

Arguments

ReplicationTaskArn::String – Required

The Amazon Resource Name (ARN) of the replication task.

ReplicationTaskIdentifier::String

The replication task identifier.

Constraints:

Must contain from 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

MigrationType: "full-load", "cdc" or "full-load-and-cdc"

The migration type.

Valid values: full-load | cdc | full-load-and-cdc

TableMappings::String

When using the AWS CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with "file://". When working with the DMS API, provide the JSON as the parameter value.

For example, –table-mappings file://mappingfile.json

ReplicationTaskSettings::String

JSON file that contains settings for the task, such as target metadata settings.

CdcStartTime: timestamp

The start time for the Change Data Capture (CDC) operation.

Returns

ModifyReplicationTaskResponse

Exceptions

InvalidResourceStateFault, ResourceNotFoundFault, ResourceAlreadyExistsFault or KMSKeyNotAccessibleFault.

See also: AWS API Documentation

AmazonDMS.refresh_schemas — Method.

refresh_schemas(::AWSConfig, arguments::Dict)
refresh_schemas(::AWSConfig; EndpointArn=, ReplicationInstanceArn=)

RefreshSchemas Operation

Populates the schema for the specified endpoint. This is an asynchronous operation and can take several minutes. You can check the status of this operation by calling the DescribeRefreshSchemasStatus operation.

Arguments

EndpointArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

ReplicationInstanceArn::String – Required

The Amazon Resource Name (ARN) of the replication instance.

Returns

RefreshSchemasResponse

Exceptions

InvalidResourceStateFault, ResourceNotFoundFault, KMSKeyNotAccessibleFault or ResourceQuotaExceededFault.

See also: AWS API Documentation

AmazonDMS.reload_tables — Method.

reload_tables(::AWSConfig, arguments::Dict)
reload_tables(::AWSConfig; ReplicationTaskArn=, TablesToReload=)

ReloadTables Operation

Reloads the target database table with the source data.

Arguments

ReplicationTaskArn::String – Required

The Amazon Resource Name (ARN) of the replication instance.

TablesToReload: – Required

The name and schema of the table to be reloaded.

    [Dict(
        SchemaName  => ::String,
        TableName  => ::String
    )...]

Returns

ReloadTablesResponse

Exceptions

ResourceNotFoundFault or InvalidResourceStateFault.

See also: AWS API Documentation

AmazonDMS.remove_tags_from_resource — Method.

remove_tags_from_resource(::AWSConfig, arguments::Dict)
remove_tags_from_resource(::AWSConfig; ResourceArn=, TagKeys=)

RemoveTagsFromResource Operation

Removes metadata tags from a DMS resource.

Arguments

ResourceArn::String – Required

The Amazon Resource Name (ARN) of the AWS DMS resource the tag is to be removed from.

TagKeys: [::String...] – Required

The tag key (name) of the tag to be removed.

Returns

RemoveTagsFromResourceResponse

Exceptions

ResourceNotFoundFault.

See also: AWS API Documentation

AmazonDMS.start_replication_task — Method.

start_replication_task(::AWSConfig, arguments::Dict)
start_replication_task(::AWSConfig; ReplicationTaskArn=, StartReplicationTaskType=, <keyword arguments>)

StartReplicationTask Operation

Starts the replication task.

For more information about AWS DMS tasks, see the AWS DMS user guide at Working with Migration Tasks

Arguments

ReplicationTaskArn::String – Required

The Amazon Resource Number (ARN) of the replication task to be started.

StartReplicationTaskType: "start-replication", "resume-processing" or "reload-target" – Required

The type of replication task.

CdcStartTime: timestamp

The start time for the Change Data Capture (CDC) operation.

Returns

StartReplicationTaskResponse

Exceptions

ResourceNotFoundFault or InvalidResourceStateFault.

See also: AWS API Documentation

AmazonDMS.stop_replication_task — Method.

stop_replication_task(::AWSConfig, arguments::Dict)
stop_replication_task(::AWSConfig; ReplicationTaskArn=)

StopReplicationTask Operation

Stops the replication task.

Arguments

ReplicationTaskArn::String – Required

The Amazon Resource Number(ARN) of the replication task to be stopped.

Returns

StopReplicationTaskResponse

Exceptions

ResourceNotFoundFault or InvalidResourceStateFault.

See also: AWS API Documentation

AmazonDMS.test_connection — Method.

test_connection(::AWSConfig, arguments::Dict)
test_connection(::AWSConfig; ReplicationInstanceArn=, EndpointArn=)

TestConnection Operation

Tests the connection between the replication instance and the endpoint.

Arguments

ReplicationInstanceArn::String – Required

The Amazon Resource Name (ARN) of the replication instance.

EndpointArn::String – Required

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Returns

TestConnectionResponse

Exceptions

ResourceNotFoundFault, InvalidResourceStateFault, KMSKeyNotAccessibleFault or ResourceQuotaExceededFault.

See also: AWS API Documentation