AmazonDataPipeline.jl

AmazonDataPipeline

AWS Data Pipeline configures and manages a data-driven workflow called a pipeline. AWS Data Pipeline handles the details of scheduling and ensuring that data dependencies are met so that your application can focus on processing the data.

AWS Data Pipeline provides a JAR implementation of a task runner called AWS Data Pipeline Task Runner. AWS Data Pipeline Task Runner provides logic for common data management scenarios, such as performing database queries and running data analysis using Amazon Elastic MapReduce (Amazon EMR). You can use AWS Data Pipeline Task Runner as your task runner, or you can write your own task runner to provide custom data management.

AWS Data Pipeline implements two main sets of functionality. Use the first set to create a pipeline and define data sources, schedules, dependencies, and the transforms to be performed on the data. Use the second set in your task runner application to receive the next task ready for processing. The logic for performing the task, such as querying the data, running data analysis, or converting the data from one format to another, is contained within the task runner. The task runner performs the task assigned to it by the web service, reporting progress to the web service as it does so. When the task is done, the task runner reports the final success or failure of the task to the web service.

This document is generated from apis/datapipeline-2012-10-29.normal.json. See JuliaCloud/AWSCore.jl.

AmazonDataPipeline.activate_pipeline

AmazonDataPipeline.add_tags

AmazonDataPipeline.create_pipeline

AmazonDataPipeline.deactivate_pipeline

AmazonDataPipeline.delete_pipeline

AmazonDataPipeline.describe_objects

AmazonDataPipeline.describe_pipelines

AmazonDataPipeline.evaluate_expression

AmazonDataPipeline.get_pipeline_definition

AmazonDataPipeline.list_pipelines

AmazonDataPipeline.poll_for_task

AmazonDataPipeline.put_pipeline_definition

AmazonDataPipeline.query_objects

AmazonDataPipeline.remove_tags

AmazonDataPipeline.report_task_progress

AmazonDataPipeline.report_task_runner_heartbeat

AmazonDataPipeline.set_status

AmazonDataPipeline.set_task_status

AmazonDataPipeline.validate_pipeline_definition

AmazonDataPipeline.activate_pipeline — Function.

activate_pipeline(::AWSConfig, arguments::Dict)
activate_pipeline([::AWSConfig]; pipelineId=, <keyword arguments>)

ActivatePipeline Operation

Validates the specified pipeline and starts processing pipeline tasks. If the pipeline does not pass validation, activation fails.

If you need to pause the pipeline to investigate an issue with a component, such as a data source or script, call DeactivatePipeline.

To activate a finished pipeline, modify the end date for the pipeline and then activate it.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

parameterValues = [[ ... ], ...]

A list of parameter values to pass to the pipeline at activation.

 parameterValues = [[
        "id" => <required> ::String,
        "stringValue" => <required> ::String
    ], ...]

startTimestamp = timestamp

The date and time to resume the pipeline. By default, the pipeline resumes from the last completed execution.

Returns

ActivatePipelineOutput

Exceptions

PipelineNotFoundException, PipelineDeletedException, InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.add_tags — Function.

add_tags(::AWSConfig, arguments::Dict)
add_tags([::AWSConfig]; pipelineId=, tags=)

AddTags Operation

Adds or modifies tags for the specified pipeline.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

tags = [[ ... ], ...] – Required

The tags to add, as key/value pairs.

 tags = [[
        "key" => <required> ::String,
        "value" => <required> ::String
    ], ...]

Returns

AddTagsOutput

Exceptions

InternalServiceError, InvalidRequestException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation

AmazonDataPipeline.create_pipeline — Function.

create_pipeline(::AWSConfig, arguments::Dict)
create_pipeline([::AWSConfig]; name=, uniqueId=, <keyword arguments>)

CreatePipeline Operation

Creates a new, empty pipeline. Use PutPipelineDefinition to populate the pipeline.

Arguments

name = ::String – Required

The name for the pipeline. You can use the same name for multiple pipelines associated with your AWS account, because AWS Data Pipeline assigns each pipeline a unique pipeline identifier.

uniqueId = ::String – Required

A unique identifier. This identifier is not the same as the pipeline identifier assigned by AWS Data Pipeline. You are responsible for defining the format and ensuring the uniqueness of this identifier. You use this parameter to ensure idempotency during repeated calls to CreatePipeline. For example, if the first call to CreatePipeline does not succeed, you can pass in the same unique identifier and pipeline name combination on a subsequent call to CreatePipeline. CreatePipeline ensures that if a pipeline already exists with the same name and unique identifier, a new pipeline is not created. Instead, you'll receive the pipeline identifier from the previous attempt. The uniqueness of the name and unique identifier combination is scoped to the AWS account or IAM user credentials.

description = ::String

The description for the pipeline.

tags = [[ ... ], ...]

A list of tags to associate with the pipeline at creation. Tags let you control access to pipelines. For more information, see Controlling User Access to Pipelines in the AWS Data Pipeline Developer Guide.

 tags = [[
        "key" => <required> ::String,
        "value" => <required> ::String
    ], ...]

Returns

CreatePipelineOutput

Exceptions

InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.deactivate_pipeline — Function.

deactivate_pipeline(::AWSConfig, arguments::Dict)
deactivate_pipeline([::AWSConfig]; pipelineId=, <keyword arguments>)

DeactivatePipeline Operation

Deactivates the specified running pipeline. The pipeline is set to the DEACTIVATING state until the deactivation process completes.

To resume a deactivated pipeline, use ActivatePipeline. By default, the pipeline resumes from the last completed execution. Optionally, you can specify the date and time to resume the pipeline.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

cancelActive = ::Bool

Indicates whether to cancel any running objects. The default is true, which sets the state of any running objects to CANCELED. If this value is false, the pipeline is deactivated after all running objects finish.

Returns

DeactivatePipelineOutput

Exceptions

PipelineNotFoundException, PipelineDeletedException, InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.delete_pipeline — Function.

delete_pipeline(::AWSConfig, arguments::Dict)
delete_pipeline([::AWSConfig]; pipelineId=)

DeletePipeline Operation

Deletes a pipeline, its pipeline definition, and its run history. AWS Data Pipeline attempts to cancel instances associated with the pipeline that are currently being processed by task runners.

Deleting a pipeline cannot be undone. You cannot query or restore a deleted pipeline. To temporarily pause a pipeline instead of deleting it, call SetStatus with the status set to PAUSE on individual components. Components that are paused by SetStatus can be resumed.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

Exceptions

PipelineNotFoundException, InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.describe_objects — Function.

describe_objects(::AWSConfig, arguments::Dict)
describe_objects([::AWSConfig]; pipelineId=, objectIds=, <keyword arguments>)

DescribeObjects Operation

Gets the object definitions for a set of objects associated with the pipeline. Object definitions are composed of a set of fields that define the properties of the object.

Arguments

pipelineId = ::String – Required

The ID of the pipeline that contains the object definitions.

objectIds = [::String, ...] – Required

The IDs of the pipeline objects that contain the definitions to be described. You can pass as many as 25 identifiers in a single call to DescribeObjects.

evaluateExpressions = ::Bool

Indicates whether any expressions in the object should be evaluated when the object descriptions are returned.

marker = ::String

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call DescribeObjects with the marker value from the previous call to retrieve the next set of results.

Returns

DescribeObjectsOutput

Exceptions

InternalServiceError, InvalidRequestException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation

AmazonDataPipeline.describe_pipelines — Function.

describe_pipelines(::AWSConfig, arguments::Dict)
describe_pipelines([::AWSConfig]; pipelineIds=)

DescribePipelines Operation

Retrieves metadata about one or more pipelines. The information retrieved includes the name of the pipeline, the pipeline identifier, its current state, and the user account that owns the pipeline. Using account credentials, you can retrieve metadata about pipelines that you or your IAM users have created. If you are using an IAM user account, you can retrieve metadata about only those pipelines for which you have read permissions.

To retrieve the full pipeline definition instead of metadata about the pipeline, call GetPipelineDefinition.

Arguments

pipelineIds = [::String, ...] – Required

The IDs of the pipelines to describe. You can pass as many as 25 identifiers in a single call. To obtain pipeline IDs, call ListPipelines.

Returns

DescribePipelinesOutput

Exceptions

PipelineNotFoundException, PipelineDeletedException, InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.evaluate_expression — Function.

evaluate_expression(::AWSConfig, arguments::Dict)
evaluate_expression([::AWSConfig]; pipelineId=, objectId=, expression=)

EvaluateExpression Operation

Task runners call EvaluateExpression to evaluate a string in the context of the specified object. For example, a task runner can evaluate SQL queries stored in Amazon S3.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

objectId = ::String – Required

The ID of the object.

expression = ::String – Required

The expression to evaluate.

Returns

EvaluateExpressionOutput

Exceptions

InternalServiceError, TaskNotFoundException, InvalidRequestException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation

AmazonDataPipeline.get_pipeline_definition — Function.

get_pipeline_definition(::AWSConfig, arguments::Dict)
get_pipeline_definition([::AWSConfig]; pipelineId=, <keyword arguments>)

GetPipelineDefinition Operation

Gets the definition of the specified pipeline. You can call GetPipelineDefinition to retrieve the pipeline definition that you provided using PutPipelineDefinition.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

version = ::String

The version of the pipeline definition to retrieve. Set this parameter to latest (default) to use the last definition saved to the pipeline or active to use the last definition that was activated.

Returns

GetPipelineDefinitionOutput

Exceptions

InternalServiceError, InvalidRequestException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation

AmazonDataPipeline.list_pipelines — Function.

list_pipelines(::AWSConfig, arguments::Dict)
list_pipelines([::AWSConfig]; <keyword arguments>)

ListPipelines Operation

Lists the pipeline identifiers for all active pipelines that you have permission to access.

Arguments

marker = ::String

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call ListPipelines with the marker value from the previous call to retrieve the next set of results.

Returns

ListPipelinesOutput

Exceptions

InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.poll_for_task — Function.

poll_for_task(::AWSConfig, arguments::Dict)
poll_for_task([::AWSConfig]; workerGroup=, <keyword arguments>)

PollForTask Operation

Task runners call PollForTask to receive a task to perform from AWS Data Pipeline. The task runner specifies which tasks it can perform by setting a value for the workerGroup parameter. The task returned can come from any of the pipelines that match the workerGroup value passed in by the task runner and that was launched using the IAM user credentials specified by the task runner.

If tasks are ready in the work queue, PollForTask returns a response immediately. If no tasks are available in the queue, PollForTask uses long-polling and holds on to a poll connection for up to a 90 seconds, during which time the first newly scheduled task is handed to the task runner. To accomodate this, set the socket timeout in your task runner to 90 seconds. The task runner should not call PollForTask again on the same workerGroup until it receives a response, and this can take up to 90 seconds.

Arguments

workerGroup = ::String – Required

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup in the call to PollForTask. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

hostname = ::String

The public DNS name of the calling task runner.

instanceIdentity = [ ... ]

Identity information for the EC2 instance that is hosting the task runner. You can get this value from the instance using http://169.254.169.254/latest/meta-data/instance-id. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. Passing in this value proves that your task runner is running on an EC2 instance, and ensures the proper AWS Data Pipeline service charges are applied to your pipeline.

 instanceIdentity = [
        "document" =>  ::String,
        "signature" =>  ::String
    ]

Returns

PollForTaskOutput

Exceptions

InternalServiceError, InvalidRequestException or TaskNotFoundException.

See also: AWS API Documentation

AmazonDataPipeline.put_pipeline_definition — Function.

put_pipeline_definition(::AWSConfig, arguments::Dict)
put_pipeline_definition([::AWSConfig]; pipelineId=, pipelineObjects=, <keyword arguments>)

PutPipelineDefinition Operation

Adds tasks, schedules, and preconditions to the specified pipeline. You can use PutPipelineDefinition to populate a new pipeline.

PutPipelineDefinition also validates the configuration as it adds it to the pipeline. Changes to the pipeline are saved unless one of the following three validation errors exists in the pipeline.

An object is missing a name or identifier field.

A string or reference field is empty.

The number of objects in the pipeline exceeds the maximum allowed objects.

The pipeline is in a FINISHED state.

Pipeline object definitions are passed to the PutPipelineDefinition action and returned by the GetPipelineDefinition action.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

pipelineObjects = [[ ... ], ...] – Required

The objects that define the pipeline. These objects overwrite the existing pipeline definition.

 pipelineObjects = [[
        "id" => <required> ::String,
        "name" => <required> ::String,
        "fields" => <required> [[
            "key" => <required> ::String,
            "stringValue" =>  ::String,
            "refValue" =>  ::String
        ], ...]
    ], ...]

parameterObjects = [[ ... ], ...]

The parameter objects used with the pipeline.

 parameterObjects = [[
        "id" => <required> ::String,
        "attributes" => <required> [[
            "key" => <required> ::String,
            "stringValue" => <required> ::String
        ], ...]
    ], ...]

parameterValues = [[ ... ], ...]

The parameter values used with the pipeline.

 parameterValues = [[
        "id" => <required> ::String,
        "stringValue" => <required> ::String
    ], ...]

Returns

PutPipelineDefinitionOutput

Exceptions

InternalServiceError, InvalidRequestException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation

AmazonDataPipeline.query_objects — Function.

query_objects(::AWSConfig, arguments::Dict)
query_objects([::AWSConfig]; pipelineId=, sphere=, <keyword arguments>)

QueryObjects Operation

Queries the specified pipeline for the names of objects that match the specified set of conditions.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

query = ["selectors" => [[ ... ], ...]]

The query that defines the objects to be returned. The Query object can contain a maximum of ten selectors. The conditions in the query are limited to top-level String fields in the object. These filters can be applied to components, instances, and attempts.

 query = ["selectors" =>  [[
            "fieldName" =>  ::String,
            "operator" =>  [
                "type" =>  "EQ", "REF_EQ", "LE", "GE" or "BETWEEN",
                "values" =>  [::String, ...]
            ]
        ], ...]]

sphere = ::String – Required

Indicates whether the query applies to components or instances. The possible values are: COMPONENT, INSTANCE, and ATTEMPT.

marker = ::String

The starting point for the results to be returned. For the first call, this value should be empty. As long as there are more results, continue to call QueryObjects with the marker value from the previous call to retrieve the next set of results.

limit = ::Int

The maximum number of object names that QueryObjects will return in a single call. The default value is 100.

Returns

QueryObjectsOutput

Exceptions

PipelineNotFoundException, PipelineDeletedException, InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.remove_tags — Function.

remove_tags(::AWSConfig, arguments::Dict)
remove_tags([::AWSConfig]; pipelineId=, tagKeys=)

RemoveTags Operation

Removes existing tags from the specified pipeline.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

tagKeys = [::String, ...] – Required

The keys of the tags to remove.

Returns

RemoveTagsOutput

Exceptions

InternalServiceError, InvalidRequestException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation

AmazonDataPipeline.report_task_progress — Function.

report_task_progress(::AWSConfig, arguments::Dict)
report_task_progress([::AWSConfig]; taskId=, <keyword arguments>)

ReportTaskProgress Operation

Task runners call ReportTaskProgress when assigned a task to acknowledge that it has the task. If the web service does not receive this acknowledgement within 2 minutes, it assigns the task in a subsequent PollForTask call. After this initial acknowledgement, the task runner only needs to report progress every 15 minutes to maintain its ownership of the task. You can change this reporting time from 15 minutes by specifying a reportProgressTimeout field in your pipeline.

If a task runner does not report its status after 5 minutes, AWS Data Pipeline assumes that the task runner is unable to process the task and reassigns the task in a subsequent response to PollForTask. Task runners should call ReportTaskProgress every 60 seconds.

Arguments

taskId = ::String – Required

The ID of the task assigned to the task runner. This value is provided in the response for PollForTask.

fields = [[ ... ], ...]

Key-value pairs that define the properties of the ReportTaskProgressInput object.

 fields = [[
        "key" => <required> ::String,
        "stringValue" =>  ::String,
        "refValue" =>  ::String
    ], ...]

Returns

ReportTaskProgressOutput

Exceptions

InternalServiceError, InvalidRequestException, TaskNotFoundException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation

AmazonDataPipeline.report_task_runner_heartbeat — Function.

report_task_runner_heartbeat(::AWSConfig, arguments::Dict)
report_task_runner_heartbeat([::AWSConfig]; taskrunnerId=, <keyword arguments>)

ReportTaskRunnerHeartbeat Operation

Task runners call ReportTaskRunnerHeartbeat every 15 minutes to indicate that they are operational. If the AWS Data Pipeline Task Runner is launched on a resource managed by AWS Data Pipeline, the web service can use this call to detect when the task runner application has failed and restart a new instance.

Arguments

taskrunnerId = ::String – Required

The ID of the task runner. This value should be unique across your AWS account. In the case of AWS Data Pipeline Task Runner launched on a resource managed by AWS Data Pipeline, the web service provides a unique identifier when it launches the application. If you have written a custom task runner, you should assign a unique identifier for the task runner.

workerGroup = ::String

The type of task the task runner is configured to accept and process. The worker group is set as a field on objects in the pipeline when they are created. You can only specify a single value for workerGroup. There are no wildcard values permitted in workerGroup; the string must be an exact, case-sensitive, match.

hostname = ::String

The public DNS name of the task runner.

Returns

ReportTaskRunnerHeartbeatOutput

Exceptions

InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.set_status — Function.

set_status(::AWSConfig, arguments::Dict)
set_status([::AWSConfig]; pipelineId=, objectIds=, status=)

SetStatus Operation

Requests that the status of the specified physical or logical pipeline objects be updated in the specified pipeline. This update might not occur immediately, but is eventually consistent. The status that can be set depends on the type of object (for example, DataNode or Activity). You cannot perform this operation on FINISHED pipelines and attempting to do so returns InvalidRequestException.

Arguments

pipelineId = ::String – Required

The ID of the pipeline that contains the objects.

objectIds = [::String, ...] – Required

The IDs of the objects. The corresponding objects can be either physical or components, but not a mix of both types.

status = ::String – Required

The status to be set on all the objects specified in objectIds. For components, use PAUSE or RESUME. For instances, use TRY_CANCEL, RERUN, or MARK_FINISHED.

Exceptions

PipelineNotFoundException, PipelineDeletedException, InternalServiceError or InvalidRequestException.

See also: AWS API Documentation

AmazonDataPipeline.set_task_status — Function.

set_task_status(::AWSConfig, arguments::Dict)
set_task_status([::AWSConfig]; taskId=, taskStatus=, <keyword arguments>)

SetTaskStatus Operation

Task runners call SetTaskStatus to notify AWS Data Pipeline that a task is completed and provide information about the final status. A task runner makes this call regardless of whether the task was sucessful. A task runner does not need to call SetTaskStatus for tasks that are canceled by the web service during a call to ReportTaskProgress.

Arguments

taskId = ::String – Required

The ID of the task assigned to the task runner. This value is provided in the response for PollForTask.

taskStatus = "FINISHED", "FAILED" or "FALSE" – Required

If FINISHED, the task successfully completed. If FAILED, the task ended unsuccessfully. Preconditions use false.

errorId = ::String

If an error occurred during the task, this value specifies the error code. This value is set on the physical attempt object. It is used to display error information to the user. It should not start with string "Service_" which is reserved by the system.

errorMessage = ::String

If an error occurred during the task, this value specifies a text description of the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

errorStackTrace = ::String

If an error occurred during the task, this value specifies the stack trace associated with the error. This value is set on the physical attempt object. It is used to display error information to the user. The web service does not parse this value.

Returns

SetTaskStatusOutput

Exceptions

InternalServiceError, TaskNotFoundException, InvalidRequestException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation

AmazonDataPipeline.validate_pipeline_definition — Function.

validate_pipeline_definition(::AWSConfig, arguments::Dict)
validate_pipeline_definition([::AWSConfig]; pipelineId=, pipelineObjects=, <keyword arguments>)

ValidatePipelineDefinition Operation

Validates the specified pipeline definition to ensure that it is well formed and can be run without error.

Arguments

pipelineId = ::String – Required

The ID of the pipeline.

pipelineObjects = [[ ... ], ...] – Required

The objects that define the pipeline changes to validate against the pipeline.

 pipelineObjects = [[
        "id" => <required> ::String,
        "name" => <required> ::String,
        "fields" => <required> [[
            "key" => <required> ::String,
            "stringValue" =>  ::String,
            "refValue" =>  ::String
        ], ...]
    ], ...]

parameterObjects = [[ ... ], ...]

The parameter objects used with the pipeline.

 parameterObjects = [[
        "id" => <required> ::String,
        "attributes" => <required> [[
            "key" => <required> ::String,
            "stringValue" => <required> ::String
        ], ...]
    ], ...]

parameterValues = [[ ... ], ...]

The parameter values used with the pipeline.

 parameterValues = [[
        "id" => <required> ::String,
        "stringValue" => <required> ::String
    ], ...]

Returns

ValidatePipelineDefinitionOutput

Exceptions

InternalServiceError, InvalidRequestException, PipelineNotFoundException or PipelineDeletedException.

See also: AWS API Documentation