AmazonDirectConnect.jl

AmazonDirectConnect

AWS Direct Connect links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. With this connection in place, you can create virtual interfaces directly to the AWS cloud (for example, to Amazon Elastic Compute Cloud (Amazon EC2) and Amazon Simple Storage Service (Amazon S3)) and to Amazon Virtual Private Cloud (Amazon VPC), bypassing Internet service providers in your network path. An AWS Direct Connect location provides access to AWS in the region it is associated with, as well as access to other US regions. For example, you can provision a single connection to any AWS Direct Connect location in the US and use it to access public AWS services in all US Regions and AWS GovCloud (US).

This document is generated from apis/directconnect-2012-10-25.normal.json. See JuliaCloud/AWSCore.jl.

AmazonDirectConnect.allocate_connection_on_interconnect

AmazonDirectConnect.allocate_hosted_connection

AmazonDirectConnect.allocate_private_virtual_interface

AmazonDirectConnect.allocate_public_virtual_interface

AmazonDirectConnect.associate_connection_with_lag

AmazonDirectConnect.associate_hosted_connection

AmazonDirectConnect.associate_virtual_interface

AmazonDirectConnect.confirm_connection

AmazonDirectConnect.confirm_private_virtual_interface

AmazonDirectConnect.confirm_public_virtual_interface

AmazonDirectConnect.create_bgppeer

AmazonDirectConnect.create_connection

AmazonDirectConnect.create_interconnect

AmazonDirectConnect.create_lag

AmazonDirectConnect.create_private_virtual_interface

AmazonDirectConnect.create_public_virtual_interface

AmazonDirectConnect.delete_bgppeer

AmazonDirectConnect.delete_connection

AmazonDirectConnect.delete_interconnect

AmazonDirectConnect.delete_lag

AmazonDirectConnect.delete_virtual_interface

AmazonDirectConnect.describe_connection_loa

AmazonDirectConnect.describe_connections

AmazonDirectConnect.describe_connections_on_interconnect

AmazonDirectConnect.describe_hosted_connections

AmazonDirectConnect.describe_interconnect_loa

AmazonDirectConnect.describe_interconnects

AmazonDirectConnect.describe_lags

AmazonDirectConnect.describe_loa

AmazonDirectConnect.describe_locations

AmazonDirectConnect.describe_tags

AmazonDirectConnect.describe_virtual_gateways

AmazonDirectConnect.describe_virtual_interfaces

AmazonDirectConnect.disassociate_connection_from_lag

AmazonDirectConnect.tag_resource

AmazonDirectConnect.untag_resource

AmazonDirectConnect.update_lag

AmazonDirectConnect.allocate_connection_on_interconnect — Function.

allocate_connection_on_interconnect(::AWSConfig, arguments::Dict)
allocate_connection_on_interconnect([::AWSConfig]; bandwidth=, connectionName=, ownerAccount=, interconnectId=, vlan=)

AllocateConnectionOnInterconnect Operation

Deprecated in favor of AllocateHostedConnection.

Creates a hosted connection on an interconnect.

Allocates a VLAN number and a specified amount of bandwidth for use by a hosted connection on the given interconnect.

Note

This is intended for use by AWS Direct Connect partners only.

Arguments

bandwidth = ::String – Required

Bandwidth of the connection.

Example: "500Mbps"

Default: None

Values: 50Mbps, 100Mbps, 200Mbps, 300Mbps, 400Mbps, or 500Mbps

connectionName = ::String – Required

Name of the provisioned connection.

Example: "500M Connection to AWS"

Default: None

ownerAccount = ::String – Required

Numeric account Id of the customer for whom the connection will be provisioned.

Example: 123443215678

Default: None

interconnectId = ::String – Required

ID of the interconnect on which the connection will be provisioned.

Example: dxcon-456abc78

Default: None

vlan = ::Int – Required

The dedicated VLAN provisioned to the connection.

Example: 101

Default: None

Returns

Connection

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.allocate_hosted_connection — Function.

allocate_hosted_connection(::AWSConfig, arguments::Dict)
allocate_hosted_connection([::AWSConfig]; connectionId=, ownerAccount=, bandwidth=, connectionName=, vlan=)

AllocateHostedConnection Operation

Creates a hosted connection on an interconnect or a link aggregation group (LAG).

Allocates a VLAN number and a specified amount of bandwidth for use by a hosted connection on the given interconnect or LAG.

Note

This is intended for use by AWS Direct Connect partners only.

Arguments

connectionId = ::String – Required

The ID of the interconnect or LAG on which the connection will be provisioned.

Example: dxcon-456abc78 or dxlag-abc123

Default: None

ownerAccount = ::String – Required

The numeric account ID of the customer for whom the connection will be provisioned.

Example: 123443215678

Default: None

bandwidth = ::String – Required

The bandwidth of the connection.

Example: 500Mbps

Default: None

Values: 50Mbps, 100Mbps, 200Mbps, 300Mbps, 400Mbps, or 500Mbps

connectionName = ::String – Required

The name of the provisioned connection.

Example: "500M Connection to AWS"

Default: None

vlan = ::Int – Required

The dedicated VLAN provisioned to the hosted connection.

Example: 101

Default: None

Returns

Connection

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.allocate_private_virtual_interface — Function.

allocate_private_virtual_interface(::AWSConfig, arguments::Dict)
allocate_private_virtual_interface([::AWSConfig]; connectionId=, ownerAccount=, newPrivateVirtualInterfaceAllocation=)

AllocatePrivateVirtualInterface Operation

Provisions a private virtual interface to be owned by another AWS customer.

Virtual interfaces created using this action must be confirmed by the virtual interface owner by using the ConfirmPrivateVirtualInterface action. Until then, the virtual interface will be in 'Confirming' state, and will not be available for handling traffic.

Arguments

connectionId = ::String – Required

The connection ID on which the private virtual interface is provisioned.

Default: None

ownerAccount = ::String – Required

The AWS account that will own the new private virtual interface.

Default: None

newPrivateVirtualInterfaceAllocation = Dict( ... ) – Required

Detailed information for the private virtual interface to be provisioned.

Default: None

 newPrivateVirtualInterfaceAllocation = Dict(
        "virtualInterfaceName" => <required> ::String,
        "vlan" => <required> ::Int,
        "asn" => <required> ::Int,
        "authKey" =>  ::String,
        "amazonAddress" =>  ::String,
        "addressFamily" =>  "ipv4" or "ipv6",
        "customerAddress" =>  ::String
    )

Returns

VirtualInterface

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.allocate_public_virtual_interface — Function.

allocate_public_virtual_interface(::AWSConfig, arguments::Dict)
allocate_public_virtual_interface([::AWSConfig]; connectionId=, ownerAccount=, newPublicVirtualInterfaceAllocation=)

AllocatePublicVirtualInterface Operation

Provisions a public virtual interface to be owned by a different customer.

The owner of a connection calls this function to provision a public virtual interface which will be owned by another AWS customer.

Virtual interfaces created using this function must be confirmed by the virtual interface owner by calling ConfirmPublicVirtualInterface. Until this step has been completed, the virtual interface will be in 'Confirming' state, and will not be available for handling traffic.

When creating an IPv6 public virtual interface (addressFamily is 'ipv6'), the customer and amazon address fields should be left blank to use auto-assigned IPv6 space. Custom IPv6 Addresses are currently not supported.

Arguments

connectionId = ::String – Required

The connection ID on which the public virtual interface is provisioned.

Default: None

ownerAccount = ::String – Required

The AWS account that will own the new public virtual interface.

Default: None

newPublicVirtualInterfaceAllocation = Dict( ... ) – Required

Detailed information for the public virtual interface to be provisioned.

Default: None

 newPublicVirtualInterfaceAllocation = Dict(
        "virtualInterfaceName" => <required> ::String,
        "vlan" => <required> ::Int,
        "asn" => <required> ::Int,
        "authKey" =>  ::String,
        "amazonAddress" =>  ::String,
        "customerAddress" =>  ::String,
        "addressFamily" =>  "ipv4" or "ipv6",
        "routeFilterPrefixes" =>  [Dict("cidr" =>  ::String)...]
    )

Returns

VirtualInterface

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.associate_connection_with_lag — Function.

associate_connection_with_lag(::AWSConfig, arguments::Dict)
associate_connection_with_lag([::AWSConfig]; connectionId=, lagId=)

AssociateConnectionWithLag Operation

Associates an existing connection with a link aggregation group (LAG). The connection is interrupted and re-established as a member of the LAG (connectivity to AWS will be interrupted). The connection must be hosted on the same AWS Direct Connect endpoint as the LAG, and its bandwidth must match the bandwidth for the LAG. You can reassociate a connection that's currently associated with a different LAG; however, if removing the connection will cause the original LAG to fall below its setting for minimum number of operational connections, the request fails.

Any virtual interfaces that are directly associated with the connection are automatically re-associated with the LAG. If the connection was originally associated with a different LAG, the virtual interfaces remain associated with the original LAG.

For interconnects, any hosted connections are automatically re-associated with the LAG. If the interconnect was originally associated with a different LAG, the hosted connections remain associated with the original LAG.

Arguments

connectionId = ::String – Required

The ID of the connection.

Example: dxcon-abc123

Default: None

lagId = ::String – Required

The ID of the LAG with which to associate the connection.

Example: dxlag-abc123

Default: None

Returns

Connection

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.associate_hosted_connection — Function.

associate_hosted_connection(::AWSConfig, arguments::Dict)
associate_hosted_connection([::AWSConfig]; connectionId=, parentConnectionId=)

AssociateHostedConnection Operation

Associates a hosted connection and its virtual interfaces with a link aggregation group (LAG) or interconnect. If the target interconnect or LAG has an existing hosted connection with a conflicting VLAN number or IP address, the operation fails. This action temporarily interrupts the hosted connection's connectivity to AWS as it is being migrated.

Note

This is intended for use by AWS Direct Connect partners only.

Arguments

connectionId = ::String – Required

The ID of the hosted connection.

Example: dxcon-abc123

Default: None

parentConnectionId = ::String – Required

The ID of the interconnect or the LAG.

Example: dxcon-abc123 or dxlag-abc123

Default: None

Returns

Connection

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.associate_virtual_interface — Function.

associate_virtual_interface(::AWSConfig, arguments::Dict)
associate_virtual_interface([::AWSConfig]; virtualInterfaceId=, connectionId=)

AssociateVirtualInterface Operation

Associates a virtual interface with a specified link aggregation group (LAG) or connection. Connectivity to AWS is temporarily interrupted as the virtual interface is being migrated. If the target connection or LAG has an associated virtual interface with a conflicting VLAN number or a conflicting IP address, the operation fails.

Virtual interfaces associated with a hosted connection cannot be associated with a LAG; hosted connections must be migrated along with their virtual interfaces using AssociateHostedConnection.

Hosted virtual interfaces (an interface for which the owner of the connection is not the owner of physical connection) can only be reassociated by the owner of the physical connection.

Arguments

virtualInterfaceId = ::String – Required

The ID of the virtual interface.

Example: dxvif-123dfg56

Default: None

connectionId = ::String – Required

The ID of the LAG or connection with which to associate the virtual interface.

Example: dxlag-abc123 or dxcon-abc123

Default: None

Returns

VirtualInterface

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.confirm_connection — Function.

confirm_connection(::AWSConfig, arguments::Dict)
confirm_connection([::AWSConfig]; connectionId=)

ConfirmConnection Operation

Confirm the creation of a hosted connection on an interconnect.

Upon creation, the hosted connection is initially in the 'Ordering' state, and will remain in this state until the owner calls ConfirmConnection to confirm creation of the hosted connection.

Arguments

connectionId = ::String – Required

Returns

ConfirmConnectionResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.confirm_private_virtual_interface — Function.

confirm_private_virtual_interface(::AWSConfig, arguments::Dict)
confirm_private_virtual_interface([::AWSConfig]; virtualInterfaceId=, virtualGatewayId=)

ConfirmPrivateVirtualInterface Operation

Accept ownership of a private virtual interface created by another customer.

After the virtual interface owner calls this function, the virtual interface will be created and attached to the given virtual private gateway, and will be available for handling traffic.

Arguments

virtualInterfaceId = ::String – Required

virtualGatewayId = ::String – Required

ID of the virtual private gateway that will be attached to the virtual interface.

A virtual private gateway can be managed via the Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action.

Default: None

Returns

ConfirmPrivateVirtualInterfaceResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.confirm_public_virtual_interface — Function.

confirm_public_virtual_interface(::AWSConfig, arguments::Dict)
confirm_public_virtual_interface([::AWSConfig]; virtualInterfaceId=)

ConfirmPublicVirtualInterface Operation

Accept ownership of a public virtual interface created by another customer.

After the virtual interface owner calls this function, the specified virtual interface will be created and made available for handling traffic.

Arguments

virtualInterfaceId = ::String – Required

Returns

ConfirmPublicVirtualInterfaceResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.create_bgppeer — Function.

create_bgppeer(::AWSConfig, arguments::Dict)
create_bgppeer([::AWSConfig]; <keyword arguments>)

CreateBGPPeer Operation

Creates a new BGP peer on a specified virtual interface. The BGP peer cannot be in the same address family (IPv4/IPv6) of an existing BGP peer on the virtual interface.

You must create a BGP peer for the corresponding address family in order to access AWS resources that also use that address family.

When creating a IPv6 BGP peer, the Amazon address and customer address fields must be left blank. IPv6 addresses are automatically assigned from Amazon's pool of IPv6 addresses; you cannot specify custom IPv6 addresses.

For a public virtual interface, the Autonomous System Number (ASN) must be private or already whitelisted for the virtual interface.

Arguments

virtualInterfaceId = ::String

The ID of the virtual interface on which the BGP peer will be provisioned.

Example: dxvif-456abc78

Default: None

newBGPPeer = Dict( ... )

Detailed information for the BGP peer to be created.

Default: None

 newBGPPeer = Dict(
        "asn" =>  ::Int,
        "authKey" =>  ::String,
        "addressFamily" =>  "ipv4" or "ipv6",
        "amazonAddress" =>  ::String,
        "customerAddress" =>  ::String
    )

Returns

CreateBGPPeerResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.create_connection — Function.

create_connection(::AWSConfig, arguments::Dict)
create_connection([::AWSConfig]; location=, bandwidth=, connectionName=, <keyword arguments>)

CreateConnection Operation

Creates a new connection between the customer network and a specific AWS Direct Connect location.

A connection links your internal network to an AWS Direct Connect location over a standard 1 gigabit or 10 gigabit Ethernet fiber-optic cable. One end of the cable is connected to your router, the other to an AWS Direct Connect router. An AWS Direct Connect location provides access to Amazon Web Services in the region it is associated with. You can establish connections with AWS Direct Connect locations in multiple regions, but a connection in one region does not provide connectivity to other regions.

You can automatically add the new connection to a link aggregation group (LAG) by specifying a LAG ID in the request. This ensures that the new connection is allocated on the same AWS Direct Connect endpoint that hosts the specified LAG. If there are no available ports on the endpoint, the request fails and no connection will be created.

Arguments

location = ::String – Required

bandwidth = ::String – Required

connectionName = ::String – Required

lagId = ::String

Returns

Connection

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.create_interconnect — Function.

create_interconnect(::AWSConfig, arguments::Dict)
create_interconnect([::AWSConfig]; interconnectName=, bandwidth=, location=, <keyword arguments>)

CreateInterconnect Operation

Creates a new interconnect between a AWS Direct Connect partner's network and a specific AWS Direct Connect location.

An interconnect is a connection which is capable of hosting other connections. The AWS Direct Connect partner can use an interconnect to provide sub-1Gbps AWS Direct Connect service to tier 2 customers who do not have their own connections. Like a standard connection, an interconnect links the AWS Direct Connect partner's network to an AWS Direct Connect location over a standard 1 Gbps or 10 Gbps Ethernet fiber-optic cable. One end is connected to the partner's router, the other to an AWS Direct Connect router.

You can automatically add the new interconnect to a link aggregation group (LAG) by specifying a LAG ID in the request. This ensures that the new interconnect is allocated on the same AWS Direct Connect endpoint that hosts the specified LAG. If there are no available ports on the endpoint, the request fails and no interconnect will be created.

For each end customer, the AWS Direct Connect partner provisions a connection on their interconnect by calling AllocateConnectionOnInterconnect. The end customer can then connect to AWS resources by creating a virtual interface on their connection, using the VLAN assigned to them by the AWS Direct Connect partner.

Note

This is intended for use by AWS Direct Connect partners only.

Arguments

interconnectName = ::String – Required

The name of the interconnect.

Example: "1G Interconnect to AWS"

Default: None

bandwidth = ::String – Required

The port bandwidth

Example: 1Gbps

Default: None

Available values: 1Gbps,10Gbps

location = ::String – Required

Where the interconnect is located

Example: EqSV5

Default: None

lagId = ::String

Returns

Interconnect

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.create_lag — Function.

create_lag(::AWSConfig, arguments::Dict)
create_lag([::AWSConfig]; numberOfConnections=, location=, connectionsBandwidth=, lagName=, <keyword arguments>)

CreateLag Operation

Creates a new link aggregation group (LAG) with the specified number of bundled physical connections between the customer network and a specific AWS Direct Connect location. A LAG is a logical interface that uses the Link Aggregation Control Protocol (LACP) to aggregate multiple 1 gigabit or 10 gigabit interfaces, allowing you to treat them as a single interface.

All connections in a LAG must use the same bandwidth (for example, 10 Gbps), and must terminate at the same AWS Direct Connect endpoint.

You can have up to 10 connections per LAG. Regardless of this limit, if you request more connections for the LAG than AWS Direct Connect can allocate on a single endpoint, no LAG is created.

You can specify an existing physical connection or interconnect to include in the LAG (which counts towards the total number of connections). Doing so interrupts the current physical connection or hosted connections, and re-establishes them as a member of the LAG. The LAG will be created on the same AWS Direct Connect endpoint to which the connection terminates. Any virtual interfaces associated with the connection are automatically disassociated and re-associated with the LAG. The connection ID does not change.

If the AWS account used to create a LAG is a registered AWS Direct Connect partner, the LAG is automatically enabled to host sub-connections. For a LAG owned by a partner, any associated virtual interfaces cannot be directly configured.

Arguments

numberOfConnections = ::Int – Required

The number of physical connections initially provisioned and bundled by the LAG.

Default: None

location = ::String – Required

The AWS Direct Connect location in which the LAG should be allocated.

Example: EqSV5

Default: None

connectionsBandwidth = ::String – Required

The bandwidth of the individual physical connections bundled by the LAG.

Default: None

Available values: 1Gbps, 10Gbps

lagName = ::String – Required

The name of the LAG.

Example: "3x10G LAG to AWS"

Default: None

connectionId = ::String

The ID of an existing connection to migrate to the LAG.

Default: None

Returns

Lag

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.create_private_virtual_interface — Function.

create_private_virtual_interface(::AWSConfig, arguments::Dict)
create_private_virtual_interface([::AWSConfig]; connectionId=, newPrivateVirtualInterface=)

CreatePrivateVirtualInterface Operation

Creates a new private virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A private virtual interface supports sending traffic to a single virtual private cloud (VPC).

Arguments

connectionId = ::String – Required

newPrivateVirtualInterface = Dict( ... ) – Required

Detailed information for the private virtual interface to be created.

Default: None

 newPrivateVirtualInterface = Dict(
        "virtualInterfaceName" => <required> ::String,
        "vlan" => <required> ::Int,
        "asn" => <required> ::Int,
        "authKey" =>  ::String,
        "amazonAddress" =>  ::String,
        "customerAddress" =>  ::String,
        "addressFamily" =>  "ipv4" or "ipv6",
        "virtualGatewayId" => <required> ::String
    )

Returns

VirtualInterface

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.create_public_virtual_interface — Function.

create_public_virtual_interface(::AWSConfig, arguments::Dict)
create_public_virtual_interface([::AWSConfig]; connectionId=, newPublicVirtualInterface=)

CreatePublicVirtualInterface Operation

Creates a new public virtual interface. A virtual interface is the VLAN that transports AWS Direct Connect traffic. A public virtual interface supports sending traffic to public services of AWS such as Amazon Simple Storage Service (Amazon S3).

When creating an IPv6 public virtual interface (addressFamily is 'ipv6'), the customer and amazon address fields should be left blank to use auto-assigned IPv6 space. Custom IPv6 Addresses are currently not supported.

Arguments

connectionId = ::String – Required

newPublicVirtualInterface = Dict( ... ) – Required

Detailed information for the public virtual interface to be created.

Default: None

 newPublicVirtualInterface = Dict(
        "virtualInterfaceName" => <required> ::String,
        "vlan" => <required> ::Int,
        "asn" => <required> ::Int,
        "authKey" =>  ::String,
        "amazonAddress" =>  ::String,
        "customerAddress" =>  ::String,
        "addressFamily" =>  "ipv4" or "ipv6",
        "routeFilterPrefixes" =>  [Dict("cidr" =>  ::String)...]
    )

Returns

VirtualInterface

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.delete_bgppeer — Function.

delete_bgppeer(::AWSConfig, arguments::Dict)
delete_bgppeer([::AWSConfig]; <keyword arguments>)

DeleteBGPPeer Operation

Deletes a BGP peer on the specified virtual interface that matches the specified customer address and ASN. You cannot delete the last BGP peer from a virtual interface.

Arguments

virtualInterfaceId = ::String

The ID of the virtual interface from which the BGP peer will be deleted.

Example: dxvif-456abc78

Default: None

asn = ::Int

customerAddress = ::String

Returns

DeleteBGPPeerResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.delete_connection — Function.

delete_connection(::AWSConfig, arguments::Dict)
delete_connection([::AWSConfig]; connectionId=)

DeleteConnection Operation

Deletes the connection.

Deleting a connection only stops the AWS Direct Connect port hour and data transfer charges. You need to cancel separately with the providers any services or charges for cross-connects or network circuits that connect you to the AWS Direct Connect location.

Arguments

connectionId = ::String – Required

Returns

Connection

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.delete_interconnect — Function.

delete_interconnect(::AWSConfig, arguments::Dict)
delete_interconnect([::AWSConfig]; interconnectId=)

DeleteInterconnect Operation

Deletes the specified interconnect.

Note

This is intended for use by AWS Direct Connect partners only.

Arguments

interconnectId = ::String – Required

Returns

DeleteInterconnectResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.delete_lag — Function.

delete_lag(::AWSConfig, arguments::Dict)
delete_lag([::AWSConfig]; lagId=)

DeleteLag Operation

Deletes a link aggregation group (LAG). You cannot delete a LAG if it has active virtual interfaces or hosted connections.

Arguments

lagId = ::String – Required

The ID of the LAG to delete.

Example: dxlag-abc123

Default: None

Returns

Lag

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.delete_virtual_interface — Function.

delete_virtual_interface(::AWSConfig, arguments::Dict)
delete_virtual_interface([::AWSConfig]; virtualInterfaceId=)

DeleteVirtualInterface Operation

Deletes a virtual interface.

Arguments

virtualInterfaceId = ::String – Required

Returns

DeleteVirtualInterfaceResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_connection_loa — Function.

describe_connection_loa(::AWSConfig, arguments::Dict)
describe_connection_loa([::AWSConfig]; connectionId=, <keyword arguments>)

DescribeConnectionLoa Operation

Deprecated in favor of DescribeLoa.

Returns the LOA-CFA for a Connection.

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA) is a document that your APN partner or service provider uses when establishing your cross connect to AWS at the colocation facility. For more information, see Requesting Cross Connects at AWS Direct Connect Locations in the AWS Direct Connect user guide.

Arguments

connectionId = ::String – Required

providerName = ::String

The name of the APN partner or service provider who establishes connectivity on your behalf. If you supply this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect.

Default: None

loaContentType = "application/pdf"

Returns

DescribeConnectionLoaResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_connections — Function.

describe_connections(::AWSConfig, arguments::Dict)
describe_connections([::AWSConfig]; <keyword arguments>)

DescribeConnections Operation

Displays all connections in this region.

If a connection ID is provided, the call returns only that particular connection.

Arguments

connectionId = ::String

Returns

Connections

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_connections_on_interconnect — Function.

describe_connections_on_interconnect(::AWSConfig, arguments::Dict)
describe_connections_on_interconnect([::AWSConfig]; interconnectId=)

DescribeConnectionsOnInterconnect Operation

Deprecated in favor of DescribeHostedConnections.

Returns a list of connections that have been provisioned on the given interconnect.

Note

This is intended for use by AWS Direct Connect partners only.

Arguments

interconnectId = ::String – Required

ID of the interconnect on which a list of connection is provisioned.

Example: dxcon-abc123

Default: None

Returns

Connections

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_hosted_connections — Function.

describe_hosted_connections(::AWSConfig, arguments::Dict)
describe_hosted_connections([::AWSConfig]; connectionId=)

DescribeHostedConnections Operation

Returns a list of hosted connections that have been provisioned on the given interconnect or link aggregation group (LAG).

Note

This is intended for use by AWS Direct Connect partners only.

Arguments

connectionId = ::String – Required

The ID of the interconnect or LAG on which the hosted connections are provisioned.

Example: dxcon-abc123 or dxlag-abc123

Default: None

Returns

Connections

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_interconnect_loa — Function.

describe_interconnect_loa(::AWSConfig, arguments::Dict)
describe_interconnect_loa([::AWSConfig]; interconnectId=, <keyword arguments>)

DescribeInterconnectLoa Operation

Deprecated in favor of DescribeLoa.

Returns the LOA-CFA for an Interconnect.

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA) is a document that is used when establishing your cross connect to AWS at the colocation facility. For more information, see Requesting Cross Connects at AWS Direct Connect Locations in the AWS Direct Connect user guide.

Arguments

interconnectId = ::String – Required

providerName = ::String

The name of the service provider who establishes connectivity on your behalf. If you supply this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect.

Default: None

loaContentType = "application/pdf"

Returns

DescribeInterconnectLoaResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_interconnects — Function.

describe_interconnects(::AWSConfig, arguments::Dict)
describe_interconnects([::AWSConfig]; <keyword arguments>)

DescribeInterconnects Operation

Returns a list of interconnects owned by the AWS account.

If an interconnect ID is provided, it will only return this particular interconnect.

Arguments

interconnectId = ::String

Returns

Interconnects

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_lags — Function.

describe_lags(::AWSConfig, arguments::Dict)
describe_lags([::AWSConfig]; <keyword arguments>)

DescribeLags Operation

Describes the link aggregation groups (LAGs) in your account.

If a LAG ID is provided, only information about the specified LAG is returned.

Arguments

lagId = ::String

The ID of the LAG.

Example: dxlag-abc123

Default: None

Returns

Lags

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_loa — Function.

describe_loa(::AWSConfig, arguments::Dict)
describe_loa([::AWSConfig]; connectionId=, <keyword arguments>)

DescribeLoa Operation

Returns the LOA-CFA for a connection, interconnect, or link aggregation group (LAG).

The Letter of Authorization - Connecting Facility Assignment (LOA-CFA) is a document that is used when establishing your cross connect to AWS at the colocation facility. For more information, see Requesting Cross Connects at AWS Direct Connect Locations in the AWS Direct Connect user guide.

Arguments

connectionId = ::String – Required

The ID of a connection, LAG, or interconnect for which to get the LOA-CFA information.

Example: dxcon-abc123 or dxlag-abc123

Default: None

providerName = ::String

The name of the service provider who establishes connectivity on your behalf. If you supply this parameter, the LOA-CFA lists the provider name alongside your company name as the requester of the cross connect.

Default: None

loaContentType = "application/pdf"

A standard media type indicating the content type of the LOA-CFA document. Currently, the only supported value is "application/pdf".

Default: application/pdf

Returns

Loa

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_locations — Function.

describe_locations([::AWSConfig])

DescribeLocations Operation

Returns the list of AWS Direct Connect locations in the current AWS region. These are the locations that may be selected when calling CreateConnection or CreateInterconnect.

Returns

Locations

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_tags — Function.

describe_tags(::AWSConfig, arguments::Dict)
describe_tags([::AWSConfig]; resourceArns=)

DescribeTags Operation

Describes the tags associated with the specified Direct Connect resources.

Arguments

resourceArns = [::String...] – Required

The Amazon Resource Names (ARNs) of the Direct Connect resources.

Returns

DescribeTagsResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_virtual_gateways — Function.

describe_virtual_gateways([::AWSConfig])

DescribeVirtualGateways Operation

Returns a list of virtual private gateways owned by the AWS account.

You can create one or more AWS Direct Connect private virtual interfaces linking to a virtual private gateway. A virtual private gateway can be managed via Amazon Virtual Private Cloud (VPC) console or the EC2 CreateVpnGateway action.

Returns

VirtualGateways

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.describe_virtual_interfaces — Function.

describe_virtual_interfaces(::AWSConfig, arguments::Dict)
describe_virtual_interfaces([::AWSConfig]; <keyword arguments>)

DescribeVirtualInterfaces Operation

Displays all virtual interfaces for an AWS account. Virtual interfaces deleted fewer than 15 minutes before you make the request are also returned. If you specify a connection ID, only the virtual interfaces associated with the connection are returned. If you specify a virtual interface ID, then only a single virtual interface is returned.

A virtual interface (VLAN) transmits the traffic between the AWS Direct Connect location and the customer.

Arguments

connectionId = ::String

virtualInterfaceId = ::String

Returns

VirtualInterfaces

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.disassociate_connection_from_lag — Function.

disassociate_connection_from_lag(::AWSConfig, arguments::Dict)
disassociate_connection_from_lag([::AWSConfig]; connectionId=, lagId=)

DisassociateConnectionFromLag Operation

Disassociates a connection from a link aggregation group (LAG). The connection is interrupted and re-established as a standalone connection (the connection is not deleted; to delete the connection, use the DeleteConnection request). If the LAG has associated virtual interfaces or hosted connections, they remain associated with the LAG. A disassociated connection owned by an AWS Direct Connect partner is automatically converted to an interconnect.

If disassociating the connection will cause the LAG to fall below its setting for minimum number of operational connections, the request fails, except when it's the last member of the LAG. If all connections are disassociated, the LAG continues to exist as an empty LAG with no physical connections.

Arguments

connectionId = ::String – Required

The ID of the connection to disassociate from the LAG.

Example: dxcon-abc123

Default: None

lagId = ::String – Required

The ID of the LAG.

Example: dxlag-abc123

Default: None

Returns

Connection

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.tag_resource — Function.

tag_resource(::AWSConfig, arguments::Dict)
tag_resource([::AWSConfig]; resourceArn=, tags=)

TagResource Operation

Adds the specified tags to the specified Direct Connect resource. Each Direct Connect resource can have a maximum of 50 tags.

Each tag consists of a key and an optional value. If a tag with the same key is already associated with the Direct Connect resource, this action updates its value.

Arguments

resourceArn = ::String – Required

The Amazon Resource Name (ARN) of the Direct Connect resource.

Example: arn:aws:directconnect:us-east-1:123456789012:dxcon/dxcon-fg5678gh

tags = [Dict( ... )...] – Required

The list of tags to add.

 tags = [Dict(
        "key" => <required> ::String,
        "value" =>  ::String
    )...]

Returns

TagResourceResponse

Exceptions

DuplicateTagKeysException, TooManyTagsException, DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.untag_resource — Function.

untag_resource(::AWSConfig, arguments::Dict)
untag_resource([::AWSConfig]; resourceArn=, tagKeys=)

UntagResource Operation

Removes one or more tags from the specified Direct Connect resource.

Arguments

resourceArn = ::String – Required

The Amazon Resource Name (ARN) of the Direct Connect resource.

tagKeys = [::String...] – Required

The list of tag keys to remove.

Returns

UntagResourceResponse

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation

AmazonDirectConnect.update_lag — Function.

update_lag(::AWSConfig, arguments::Dict)
update_lag([::AWSConfig]; lagId=, <keyword arguments>)

UpdateLag Operation

Updates the attributes of a link aggregation group (LAG).

You can update the following attributes:

The name of the LAG.

The value for the minimum number of connections that must be operational for the LAG itself to be operational.

When you create a LAG, the default value for the minimum number of operational connections is zero (0). If you update this value, and the number of operational connections falls below the specified value, the LAG will automatically go down to avoid overutilization of the remaining connections. Adjusting this value should be done with care as it could force the LAG down if the value is set higher than the current number of operational connections.

Arguments

lagId = ::String – Required

The ID of the LAG to update.

Example: dxlag-abc123

Default: None

lagName = ::String

The name for the LAG.

Example: "3x10G LAG to AWS"

Default: None

minimumLinks = ::Int

The minimum number of physical connections that must be operational for the LAG itself to be operational.

Default: None

Returns

Lag

Exceptions

DirectConnectServerException or DirectConnectClientException.

See also: AWS API Documentation