AmazonELB.jl

AmazonELB

Elastic Load Balancing

A load balancer distributes incoming traffic across your EC2 instances. This enables you to increase the availability of your application. The load balancer also monitors the health of its registered instances and ensures that it routes traffic only to healthy instances. You configure your load balancer to accept incoming traffic by specifying one or more listeners, which are configured with a protocol and port number for connections from clients to the load balancer and a protocol and port number for connections from the load balancer to the instances.

Elastic Load Balancing supports two types of load balancers: Classic Load Balancers and Application Load Balancers (new). A Classic Load Balancer makes routing and load balancing decisions either at the transport layer (TCP/SSL) or the application layer (HTTP/HTTPS), and supports either EC2-Classic or a VPC. An Application Load Balancer makes routing and load balancing decisions at the application layer (HTTP/HTTPS), supports path-based routing, and can route requests to one or more ports on each EC2 instance or container instance in your virtual private cloud (VPC). For more information, see the Elastic Load Balancing User Guide.

This reference covers the 2012-06-01 API, which supports Classic Load Balancers. The 2015-12-01 API supports Application Load Balancers.

To get started, create a load balancer with one or more listeners using CreateLoadBalancer. Register your instances with the load balancer using RegisterInstancesWithLoadBalancer.

All Elastic Load Balancing operations are idempotent, which means that they complete at most one time. If you repeat an operation, it succeeds with a 200 OK response code.

This document is generated from apis/elasticloadbalancing-2012-06-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonELB.add_tags

AmazonELB.apply_security_groups_to_load_balancer

AmazonELB.attach_load_balancer_to_subnets

AmazonELB.configure_health_check

AmazonELB.create_app_cookie_stickiness_policy

AmazonELB.create_lbcookie_stickiness_policy

AmazonELB.create_load_balancer

AmazonELB.create_load_balancer_listeners

AmazonELB.create_load_balancer_policy

AmazonELB.delete_load_balancer

AmazonELB.delete_load_balancer_listeners

AmazonELB.delete_load_balancer_policy

AmazonELB.deregister_instances_from_load_balancer

AmazonELB.describe_account_limits

AmazonELB.describe_instance_health

AmazonELB.describe_load_balancer_attributes

AmazonELB.describe_load_balancer_policies

AmazonELB.describe_load_balancer_policy_types

AmazonELB.describe_load_balancers

AmazonELB.describe_tags

AmazonELB.detach_load_balancer_from_subnets

AmazonELB.disable_availability_zones_for_load_balancer

AmazonELB.enable_availability_zones_for_load_balancer

AmazonELB.modify_load_balancer_attributes

AmazonELB.register_instances_with_load_balancer

AmazonELB.remove_tags

AmazonELB.set_load_balancer_listener_sslcertificate

AmazonELB.set_load_balancer_policies_for_backend_server

AmazonELB.set_load_balancer_policies_of_listener

AmazonELB.add_tags — Function.

add_tags(::AWSConfig, arguments::Dict)
add_tags([::AWSConfig]; LoadBalancerNames=, Tags=)

AddTags Operation

Adds the specified tags to the specified load balancer. Each load balancer can have a maximum of 10 tags.

Each tag consists of a key and an optional value. If a tag with the same key is already associated with the load balancer, AddTags updates its value.

For more information, see Tag Your Classic Load Balancer in the Classic Load Balancer Guide.

Arguments

LoadBalancerNames = [::String, ...] – Required

The name of the load balancer. You can specify one load balancer only.

Tags = [[ ... ], ...] – Required

The tags.

 Tags = [[
        "Key" => <required> ::String,
        "Value" =>  ::String
    ], ...]

Returns

AddTagsOutput

Exceptions

AccessPointNotFoundException, TooManyTagsException or DuplicateTagKeysException.

See also: AWS API Documentation

AmazonELB.apply_security_groups_to_load_balancer — Function.

apply_security_groups_to_load_balancer(::AWSConfig, arguments::Dict)
apply_security_groups_to_load_balancer([::AWSConfig]; LoadBalancerName=, SecurityGroups=)

ApplySecurityGroupsToLoadBalancer Operation

Associates one or more security groups with your load balancer in a virtual private cloud (VPC). The specified security groups override the previously associated security groups.

For more information, see Security Groups for Load Balancers in a VPC in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

SecurityGroups = [::String, ...] – Required

The IDs of the security groups to associate with the load balancer. Note that you cannot specify the name of the security group.

Returns

ApplySecurityGroupsToLoadBalancerOutput

Exceptions

AccessPointNotFoundException, InvalidConfigurationRequestException or InvalidSecurityGroupException.

See also: AWS API Documentation

AmazonELB.attach_load_balancer_to_subnets — Function.

attach_load_balancer_to_subnets(::AWSConfig, arguments::Dict)
attach_load_balancer_to_subnets([::AWSConfig]; LoadBalancerName=, Subnets=)

AttachLoadBalancerToSubnets Operation

Adds one or more subnets to the set of configured subnets for the specified load balancer.

The load balancer evenly distributes requests across all registered subnets. For more information, see Add or Remove Subnets for Your Load Balancer in a VPC in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

Subnets = [::String, ...] – Required

The IDs of the subnets to add. You can add only one subnet per Availability Zone.

Returns

AttachLoadBalancerToSubnetsOutput

Exceptions

AccessPointNotFoundException, InvalidConfigurationRequestException, SubnetNotFoundException or InvalidSubnetException.

See also: AWS API Documentation

AmazonELB.configure_health_check — Function.

configure_health_check(::AWSConfig, arguments::Dict)
configure_health_check([::AWSConfig]; LoadBalancerName=, HealthCheck=)

ConfigureHealthCheck Operation

Specifies the health check settings to use when evaluating the health state of your EC2 instances.

For more information, see Configure Health Checks for Your Load Balancer in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

HealthCheck = [ ... ] – Required

The configuration information.

 HealthCheck = [
        "Target" => <required> ::String,
        "Interval" => <required> ::Int,
        "Timeout" => <required> ::Int,
        "UnhealthyThreshold" => <required> ::Int,
        "HealthyThreshold" => <required> ::Int
    ]

Returns

ConfigureHealthCheckOutput

Exceptions

AccessPointNotFoundException.

See also: AWS API Documentation

AmazonELB.create_app_cookie_stickiness_policy — Function.

create_app_cookie_stickiness_policy(::AWSConfig, arguments::Dict)
create_app_cookie_stickiness_policy([::AWSConfig]; LoadBalancerName=, PolicyName=, CookieName=)

CreateAppCookieStickinessPolicy Operation

Generates a stickiness policy with sticky session lifetimes that follow that of an application-generated cookie. This policy can be associated only with HTTP/HTTPS listeners.

This policy is similar to the policy created by CreateLBCookieStickinessPolicy, except that the lifetime of the special Elastic Load Balancing cookie, AWSELB, follows the lifetime of the application-generated cookie specified in the policy configuration. The load balancer only inserts a new stickiness cookie when the application response includes a new application cookie.

If the application cookie is explicitly removed or expires, the session stops being sticky until a new application cookie is issued.

For more information, see Application-Controlled Session Stickiness in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

PolicyName = ::String – Required

The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

CookieName = ::String – Required

The name of the application cookie used for stickiness.

Returns

CreateAppCookieStickinessPolicyOutput

Exceptions

AccessPointNotFoundException, DuplicatePolicyNameException, TooManyPoliciesException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELB.create_lbcookie_stickiness_policy — Function.

create_lbcookie_stickiness_policy(::AWSConfig, arguments::Dict)
create_lbcookie_stickiness_policy([::AWSConfig]; LoadBalancerName=, PolicyName=, <keyword arguments>)

CreateLBCookieStickinessPolicy Operation

Generates a stickiness policy with sticky session lifetimes controlled by the lifetime of the browser (user-agent) or a specified expiration period. This policy can be associated only with HTTP/HTTPS listeners.

When a load balancer implements this policy, the load balancer uses a special cookie to track the instance for each request. When the load balancer receives a request, it first checks to see if this cookie is present in the request. If so, the load balancer sends the request to the application server specified in the cookie. If not, the load balancer sends the request to a server that is chosen based on the existing load-balancing algorithm.

A cookie is inserted into the response for binding subsequent requests from the same user to that server. The validity of the cookie is based on the cookie expiration time, which is specified in the policy configuration.

For more information, see Duration-Based Session Stickiness in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

PolicyName = ::String – Required

The name of the policy being created. Policy names must consist of alphanumeric characters and dashes (-). This name must be unique within the set of policies for this load balancer.

CookieExpirationPeriod = ::Int

The time period, in seconds, after which the cookie should be considered stale. If you do not specify this parameter, the default value is 0, which indicates that the sticky session should last for the duration of the browser session.

Returns

CreateLBCookieStickinessPolicyOutput

Exceptions

AccessPointNotFoundException, DuplicatePolicyNameException, TooManyPoliciesException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELB.create_load_balancer — Function.

create_load_balancer(::AWSConfig, arguments::Dict)
create_load_balancer([::AWSConfig]; LoadBalancerName=, Listeners=, <keyword arguments>)

CreateLoadBalancer Operation

Creates a Classic Load Balancer.

You can add listeners, security groups, subnets, and tags when you create your load balancer, or you can add them later using CreateLoadBalancerListeners, ApplySecurityGroupsToLoadBalancer, AttachLoadBalancerToSubnets, and AddTags.

To describe your current load balancers, see DescribeLoadBalancers. When you are finished with a load balancer, you can delete it using DeleteLoadBalancer.

You can create up to 20 load balancers per region per account. You can request an increase for the number of load balancers for your account. For more information, see Limits for Your Classic Load Balancer in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

This name must be unique within your set of load balancers for the region, must have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and cannot begin or end with a hyphen.

Listeners = [[ ... ], ...] – Required

The listeners.

For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancer Guide.

 Listeners = [[
        "Protocol" => <required> ::String,
        "LoadBalancerPort" => <required> ::Int,
        "InstanceProtocol" =>  ::String,
        "InstancePort" => <required> ::Int,
        "SSLCertificateId" =>  ::String
    ], ...]

AvailabilityZones = [::String, ...]

One or more Availability Zones from the same region as the load balancer.

You must specify at least one Availability Zone.

You can add more Availability Zones after you create the load balancer using EnableAvailabilityZonesForLoadBalancer.

Subnets = [::String, ...]

The IDs of the subnets in your VPC to attach to the load balancer. Specify one subnet per Availability Zone specified in AvailabilityZones.

SecurityGroups = [::String, ...]

The IDs of the security groups to assign to the load balancer.

Scheme = ::String

The type of a load balancer. Valid only for load balancers in a VPC.

By default, Elastic Load Balancing creates an Internet-facing load balancer with a DNS name that resolves to public IP addresses. For more information about Internet-facing and Internal load balancers, see Load Balancer Scheme in the Elastic Load Balancing User Guide.

Specify internal to create a load balancer with a DNS name that resolves to private IP addresses.

Tags = [[ ... ], ...]

A list of tags to assign to the load balancer.

For more information about tagging your load balancer, see Tag Your Classic Load Balancer in the Classic Load Balancer Guide.

 Tags = [[
        "Key" => <required> ::String,
        "Value" =>  ::String
    ], ...]

Returns

CreateAccessPointOutput

Exceptions

DuplicateAccessPointNameException, TooManyAccessPointsException, CertificateNotFoundException, InvalidConfigurationRequestException, SubnetNotFoundException, InvalidSubnetException, InvalidSecurityGroupException, InvalidSchemeException, TooManyTagsException, DuplicateTagKeysException or UnsupportedProtocolException.

See also: AWS API Documentation

AmazonELB.create_load_balancer_listeners — Function.

create_load_balancer_listeners(::AWSConfig, arguments::Dict)
create_load_balancer_listeners([::AWSConfig]; LoadBalancerName=, Listeners=)

CreateLoadBalancerListeners Operation

Creates one or more listeners for the specified load balancer. If a listener with the specified port does not already exist, it is created; otherwise, the properties of the new listener must match the properties of the existing listener.

For more information, see Listeners for Your Classic Load Balancer in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

Listeners = [[ ... ], ...] – Required

The listeners.

 Listeners = [[
        "Protocol" => <required> ::String,
        "LoadBalancerPort" => <required> ::Int,
        "InstanceProtocol" =>  ::String,
        "InstancePort" => <required> ::Int,
        "SSLCertificateId" =>  ::String
    ], ...]

Returns

CreateLoadBalancerListenerOutput

Exceptions

AccessPointNotFoundException, DuplicateListenerException, CertificateNotFoundException, InvalidConfigurationRequestException or UnsupportedProtocolException.

See also: AWS API Documentation

AmazonELB.create_load_balancer_policy — Function.

create_load_balancer_policy(::AWSConfig, arguments::Dict)
create_load_balancer_policy([::AWSConfig]; LoadBalancerName=, PolicyName=, PolicyTypeName=, <keyword arguments>)

CreateLoadBalancerPolicy Operation

Creates a policy with the specified attributes for the specified load balancer.

Policies are settings that are saved for your load balancer and that can be applied to the listener or the application server, depending on the policy type.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

PolicyName = ::String – Required

The name of the load balancer policy to be created. This name must be unique within the set of policies for this load balancer.

PolicyTypeName = ::String – Required

The name of the base policy type. To get the list of policy types, use DescribeLoadBalancerPolicyTypes.

PolicyAttributes = [[ ... ], ...]

The policy attributes.

 PolicyAttributes = [[
        "AttributeName" =>  ::String,
        "AttributeValue" =>  ::String
    ], ...]

Returns

CreateLoadBalancerPolicyOutput

Exceptions

AccessPointNotFoundException, PolicyTypeNotFoundException, DuplicatePolicyNameException, TooManyPoliciesException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELB.delete_load_balancer — Function.

delete_load_balancer(::AWSConfig, arguments::Dict)
delete_load_balancer([::AWSConfig]; LoadBalancerName=)

DeleteLoadBalancer Operation

Deletes the specified load balancer.

If you are attempting to recreate a load balancer, you must reconfigure all settings. The DNS name associated with a deleted load balancer are no longer usable. The name and associated DNS record of the deleted load balancer no longer exist and traffic sent to any of its IP addresses is no longer delivered to your instances.

If the load balancer does not exist or has already been deleted, the call to DeleteLoadBalancer still succeeds.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

Returns

DeleteAccessPointOutput

See also: AWS API Documentation

AmazonELB.delete_load_balancer_listeners — Function.

delete_load_balancer_listeners(::AWSConfig, arguments::Dict)
delete_load_balancer_listeners([::AWSConfig]; LoadBalancerName=, LoadBalancerPorts=)

DeleteLoadBalancerListeners Operation

Deletes the specified listeners from the specified load balancer.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

LoadBalancerPorts = [::Int, ...] – Required

The client port numbers of the listeners.

Returns

DeleteLoadBalancerListenerOutput

Exceptions

AccessPointNotFoundException.

See also: AWS API Documentation

AmazonELB.delete_load_balancer_policy — Function.

delete_load_balancer_policy(::AWSConfig, arguments::Dict)
delete_load_balancer_policy([::AWSConfig]; LoadBalancerName=, PolicyName=)

DeleteLoadBalancerPolicy Operation

Deletes the specified policy from the specified load balancer. This policy must not be enabled for any listeners.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

PolicyName = ::String – Required

The name of the policy.

Returns

DeleteLoadBalancerPolicyOutput

Exceptions

AccessPointNotFoundException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELB.deregister_instances_from_load_balancer — Function.

deregister_instances_from_load_balancer(::AWSConfig, arguments::Dict)
deregister_instances_from_load_balancer([::AWSConfig]; LoadBalancerName=, Instances=)

DeregisterInstancesFromLoadBalancer Operation

Deregisters the specified instances from the specified load balancer. After the instance is deregistered, it no longer receives traffic from the load balancer.

You can use DescribeLoadBalancers to verify that the instance is deregistered from the load balancer.

For more information, see Register or De-Register EC2 Instances in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

Instances = [["InstanceId" => ::String], ...] – Required

The IDs of the instances.

Returns

DeregisterEndPointsOutput

Exceptions

AccessPointNotFoundException or InvalidEndPointException.

See also: AWS API Documentation

AmazonELB.describe_account_limits — Function.

describe_account_limits(::AWSConfig, arguments::Dict)
describe_account_limits([::AWSConfig]; <keyword arguments>)

DescribeAccountLimits Operation

Describes the current Elastic Load Balancing resource limits for your AWS account.

For more information, see Limits for Your Classic Load Balancer in the Classic Load Balancer Guide.

Arguments

Marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

PageSize = ::Int

The maximum number of results to return with this call.

Returns

DescribeAccountLimitsOutput

See also: AWS API Documentation

AmazonELB.describe_instance_health — Function.

describe_instance_health(::AWSConfig, arguments::Dict)
describe_instance_health([::AWSConfig]; LoadBalancerName=, <keyword arguments>)

DescribeInstanceHealth Operation

Describes the state of the specified instances with respect to the specified load balancer. If no instances are specified, the call describes the state of all instances that are currently registered with the load balancer. If instances are specified, their state is returned even if they are no longer registered with the load balancer. The state of terminated instances is not returned.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

Instances = [["InstanceId" => ::String], ...]

The IDs of the instances.

Returns

DescribeEndPointStateOutput

Exceptions

AccessPointNotFoundException or InvalidEndPointException.

See also: AWS API Documentation

AmazonELB.describe_load_balancer_attributes — Function.

describe_load_balancer_attributes(::AWSConfig, arguments::Dict)
describe_load_balancer_attributes([::AWSConfig]; LoadBalancerName=)

DescribeLoadBalancerAttributes Operation

Describes the attributes for the specified load balancer.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

Returns

DescribeLoadBalancerAttributesOutput

Exceptions

AccessPointNotFoundException or LoadBalancerAttributeNotFoundException.

See also: AWS API Documentation

AmazonELB.describe_load_balancer_policies — Function.

describe_load_balancer_policies(::AWSConfig, arguments::Dict)
describe_load_balancer_policies([::AWSConfig]; <keyword arguments>)

DescribeLoadBalancerPolicies Operation

Describes the specified policies.

If you specify a load balancer name, the action returns the descriptions of all policies created for the load balancer. If you specify a policy name associated with your load balancer, the action returns the description of that policy. If you don't specify a load balancer name, the action returns descriptions of the specified sample policies, or descriptions of all sample policies. The names of the sample policies have the ELBSample- prefix.

Arguments

LoadBalancerName = ::String

The name of the load balancer.

PolicyNames = [::String, ...]

The names of the policies.

Returns

DescribeLoadBalancerPoliciesOutput

Exceptions

AccessPointNotFoundException or PolicyNotFoundException.

See also: AWS API Documentation

AmazonELB.describe_load_balancer_policy_types — Function.

describe_load_balancer_policy_types(::AWSConfig, arguments::Dict)
describe_load_balancer_policy_types([::AWSConfig]; <keyword arguments>)

DescribeLoadBalancerPolicyTypes Operation

Describes the specified load balancer policy types or all load balancer policy types.

The description of each type indicates how it can be used. For example, some policies can be used only with layer 7 listeners, some policies can be used only with layer 4 listeners, and some policies can be used only with your EC2 instances.

You can use CreateLoadBalancerPolicy to create a policy configuration for any of these policy types. Then, depending on the policy type, use either SetLoadBalancerPoliciesOfListener or SetLoadBalancerPoliciesForBackendServer to set the policy.

Arguments

PolicyTypeNames = [::String, ...]

The names of the policy types. If no names are specified, describes all policy types defined by Elastic Load Balancing.

Returns

DescribeLoadBalancerPolicyTypesOutput

Exceptions

PolicyTypeNotFoundException.

See also: AWS API Documentation

AmazonELB.describe_load_balancers — Function.

describe_load_balancers(::AWSConfig, arguments::Dict)
describe_load_balancers([::AWSConfig]; <keyword arguments>)

DescribeLoadBalancers Operation

Describes the specified the load balancers. If no load balancers are specified, the call describes all of your load balancers.

Arguments

LoadBalancerNames = [::String, ...]

The names of the load balancers.

Marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

PageSize = ::Int

The maximum number of results to return with this call (a number from 1 to 400). The default is 400.

Returns

DescribeAccessPointsOutput

Exceptions

AccessPointNotFoundException or DependencyThrottleException.

See also: AWS API Documentation

AmazonELB.describe_tags — Function.

describe_tags(::AWSConfig, arguments::Dict)
describe_tags([::AWSConfig]; LoadBalancerNames=)

DescribeTags Operation

Describes the tags associated with the specified load balancers.

Arguments

LoadBalancerNames = [::String, ...] – Required

The names of the load balancers.

Returns

DescribeTagsOutput

Exceptions

AccessPointNotFoundException.

See also: AWS API Documentation

AmazonELB.detach_load_balancer_from_subnets — Function.

detach_load_balancer_from_subnets(::AWSConfig, arguments::Dict)
detach_load_balancer_from_subnets([::AWSConfig]; LoadBalancerName=, Subnets=)

DetachLoadBalancerFromSubnets Operation

Removes the specified subnets from the set of configured subnets for the load balancer.

After a subnet is removed, all EC2 instances registered with the load balancer in the removed subnet go into the OutOfService state. Then, the load balancer balances the traffic among the remaining routable subnets.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

Subnets = [::String, ...] – Required

The IDs of the subnets.

Returns

DetachLoadBalancerFromSubnetsOutput

Exceptions

AccessPointNotFoundException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELB.disable_availability_zones_for_load_balancer — Function.

disable_availability_zones_for_load_balancer(::AWSConfig, arguments::Dict)
disable_availability_zones_for_load_balancer([::AWSConfig]; LoadBalancerName=, AvailabilityZones=)

DisableAvailabilityZonesForLoadBalancer Operation

Removes the specified Availability Zones from the set of Availability Zones for the specified load balancer.

There must be at least one Availability Zone registered with a load balancer at all times. After an Availability Zone is removed, all instances registered with the load balancer that are in the removed Availability Zone go into the OutOfService state. Then, the load balancer attempts to equally balance the traffic among its remaining Availability Zones.

For more information, see Add or Remove Availability Zones in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

AvailabilityZones = [::String, ...] – Required

The Availability Zones.

Returns

RemoveAvailabilityZonesOutput

Exceptions

AccessPointNotFoundException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELB.enable_availability_zones_for_load_balancer — Function.

enable_availability_zones_for_load_balancer(::AWSConfig, arguments::Dict)
enable_availability_zones_for_load_balancer([::AWSConfig]; LoadBalancerName=, AvailabilityZones=)

EnableAvailabilityZonesForLoadBalancer Operation

Adds the specified Availability Zones to the set of Availability Zones for the specified load balancer.

The load balancer evenly distributes requests across all its registered Availability Zones that contain instances.

For more information, see Add or Remove Availability Zones in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

AvailabilityZones = [::String, ...] – Required

The Availability Zones. These must be in the same region as the load balancer.

Returns

AddAvailabilityZonesOutput

Exceptions

AccessPointNotFoundException.

See also: AWS API Documentation

AmazonELB.modify_load_balancer_attributes — Function.

modify_load_balancer_attributes(::AWSConfig, arguments::Dict)
modify_load_balancer_attributes([::AWSConfig]; LoadBalancerName=, LoadBalancerAttributes=)

ModifyLoadBalancerAttributes Operation

Modifies the attributes of the specified load balancer.

You can modify the load balancer attributes, such as AccessLogs, ConnectionDraining, and CrossZoneLoadBalancing by either enabling or disabling them. Or, you can modify the load balancer attribute ConnectionSettings by specifying an idle connection timeout value for your load balancer.

For more information, see the following in the Classic Load Balancer Guide:

Cross-Zone Load Balancing

Connection Draining

Access Logs

Idle Connection Timeout

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

LoadBalancerAttributes = [ ... ] – Required

The attributes for the load balancer.

 LoadBalancerAttributes = [
        "CrossZoneLoadBalancing" =>  ["Enabled" => <required> ::Bool],
        "AccessLog" =>  [
            "Enabled" => <required> ::Bool,
            "S3BucketName" =>  ::String,
            "EmitInterval" =>  ::Int,
            "S3BucketPrefix" =>  ::String
        ],
        "ConnectionDraining" =>  [
            "Enabled" => <required> ::Bool,
            "Timeout" =>  ::Int
        ],
        "ConnectionSettings" =>  ["IdleTimeout" => <required> ::Int],
        "AdditionalAttributes" =>  [[
            "Key" =>  ::String,
            "Value" =>  ::String
        ], ...]
    ]

Returns

ModifyLoadBalancerAttributesOutput

Exceptions

AccessPointNotFoundException, LoadBalancerAttributeNotFoundException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELB.register_instances_with_load_balancer — Function.

register_instances_with_load_balancer(::AWSConfig, arguments::Dict)
register_instances_with_load_balancer([::AWSConfig]; LoadBalancerName=, Instances=)

RegisterInstancesWithLoadBalancer Operation

Adds the specified instances to the specified load balancer.

The instance must be a running instance in the same network as the load balancer (EC2-Classic or the same VPC). If you have EC2-Classic instances and a load balancer in a VPC with ClassicLink enabled, you can link the EC2-Classic instances to that VPC and then register the linked EC2-Classic instances with the load balancer in the VPC.

Note that RegisterInstanceWithLoadBalancer completes when the request has been registered. Instance registration takes a little time to complete. To check the state of the registered instances, use DescribeLoadBalancers or DescribeInstanceHealth.

After the instance is registered, it starts receiving traffic and requests from the load balancer. Any instance that is not in one of the Availability Zones registered for the load balancer is moved to the OutOfService state. If an Availability Zone is added to the load balancer later, any instances registered with the load balancer move to the InService state.

To deregister instances from a load balancer, use DeregisterInstancesFromLoadBalancer.

For more information, see Register or De-Register EC2 Instances in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

Instances = [["InstanceId" => ::String], ...] – Required

The IDs of the instances.

Returns

RegisterEndPointsOutput

Exceptions

AccessPointNotFoundException or InvalidEndPointException.

See also: AWS API Documentation

AmazonELB.remove_tags — Function.

remove_tags(::AWSConfig, arguments::Dict)
remove_tags([::AWSConfig]; LoadBalancerNames=, Tags=)

RemoveTags Operation

Removes one or more tags from the specified load balancer.

Arguments

LoadBalancerNames = [::String, ...] – Required

The name of the load balancer. You can specify a maximum of one load balancer name.

Tags = [["Key" => ::String], ...] – Required

The list of tag keys to remove.

Returns

RemoveTagsOutput

Exceptions

AccessPointNotFoundException.

See also: AWS API Documentation

AmazonELB.set_load_balancer_listener_sslcertificate — Function.

set_load_balancer_listener_sslcertificate(::AWSConfig, arguments::Dict)
set_load_balancer_listener_sslcertificate([::AWSConfig]; LoadBalancerName=, LoadBalancerPort=, SSLCertificateId=)

SetLoadBalancerListenerSSLCertificate Operation

Sets the certificate that terminates the specified listener's SSL connections. The specified certificate replaces any prior certificate that was used on the same load balancer and port.

For more information about updating your SSL certificate, see Replace the SSL Certificate for Your Load Balancer in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

LoadBalancerPort = ::Int – Required

The port that uses the specified SSL certificate.

SSLCertificateId = ::String – Required

The Amazon Resource Name (ARN) of the SSL certificate.

Returns

SetLoadBalancerListenerSSLCertificateOutput

Exceptions

CertificateNotFoundException, AccessPointNotFoundException, ListenerNotFoundException, InvalidConfigurationRequestException or UnsupportedProtocolException.

See also: AWS API Documentation

AmazonELB.set_load_balancer_policies_for_backend_server — Function.

set_load_balancer_policies_for_backend_server(::AWSConfig, arguments::Dict)
set_load_balancer_policies_for_backend_server([::AWSConfig]; LoadBalancerName=, InstancePort=, PolicyNames=)

SetLoadBalancerPoliciesForBackendServer Operation

Replaces the set of policies associated with the specified port on which the EC2 instance is listening with a new set of policies. At this time, only the back-end server authentication policy type can be applied to the instance ports; this policy type is composed of multiple public key policies.

Each time you use SetLoadBalancerPoliciesForBackendServer to enable the policies, use the PolicyNames parameter to list the policies that you want to enable.

You can use DescribeLoadBalancers or DescribeLoadBalancerPolicies to verify that the policy is associated with the EC2 instance.

For more information about enabling back-end instance authentication, see Configure Back-end Instance Authentication in the Classic Load Balancer Guide. For more information about Proxy Protocol, see Configure Proxy Protocol Support in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

InstancePort = ::Int – Required

The port number associated with the EC2 instance.

PolicyNames = [::String, ...] – Required

The names of the policies. If the list is empty, then all current polices are removed from the EC2 instance.

Returns

SetLoadBalancerPoliciesForBackendServerOutput

Exceptions

AccessPointNotFoundException, PolicyNotFoundException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELB.set_load_balancer_policies_of_listener — Function.

set_load_balancer_policies_of_listener(::AWSConfig, arguments::Dict)
set_load_balancer_policies_of_listener([::AWSConfig]; LoadBalancerName=, LoadBalancerPort=, PolicyNames=)

SetLoadBalancerPoliciesOfListener Operation

Replaces the current set of policies for the specified load balancer port with the specified set of policies.

To enable back-end server authentication, use SetLoadBalancerPoliciesForBackendServer.

For more information about setting policies, see Update the SSL Negotiation Configuration, Duration-Based Session Stickiness, and Application-Controlled Session Stickiness in the Classic Load Balancer Guide.

Arguments

LoadBalancerName = ::String – Required

The name of the load balancer.

LoadBalancerPort = ::Int – Required

The external port of the load balancer.

PolicyNames = [::String, ...] – Required

The names of the policies. This list must include all policies to be enabled. If you omit a policy that is currently enabled, it is disabled. If the list is empty, all current policies are disabled.

Returns

SetLoadBalancerPoliciesOfListenerOutput

Exceptions

AccessPointNotFoundException, PolicyNotFoundException, ListenerNotFoundException or InvalidConfigurationRequestException.

See also: AWS API Documentation