AmazonELBv2.jl

AmazonELBv2

Elastic Load Balancing

A load balancer distributes incoming traffic across targets, such as your EC2 instances. This enables you to increase the availability of your application. The load balancer also monitors the health of its registered targets and ensures that it routes traffic only to healthy targets. You configure your load balancer to accept incoming traffic by specifying one or more listeners, which are configured with a protocol and port number for connections from clients to the load balancer. You configure a target group with a protocol and port number for connections from the load balancer to the targets, and with health check settings to be used when checking the health status of the targets.

Elastic Load Balancing supports two types of load balancers: Classic Load Balancers and Application Load Balancers. A Classic Load Balancer makes routing and load balancing decisions either at the transport layer (TCP/SSL) or the application layer (HTTP/HTTPS), and supports either EC2-Classic or a VPC. An Application Load Balancer makes routing and load balancing decisions at the application layer (HTTP/HTTPS), supports path-based routing, and can route requests to one or more ports on each EC2 instance or container instance in your virtual private cloud (VPC). For more information, see the Elastic Load Balancing User Guide.

This reference covers the 2015-12-01 API, which supports Application Load Balancers. The 2012-06-01 API supports Classic Load Balancers.

To get started, complete the following tasks:

Create an Application Load Balancer using CreateLoadBalancer.

Create a target group using CreateTargetGroup.

Register targets for the target group using RegisterTargets.

Create one or more listeners for your load balancer using CreateListener.

(Optional) Create one or more rules for content routing based on URL using CreateRule.

To delete an Application Load Balancer and its related resources, complete the following tasks:

Delete the load balancer using DeleteLoadBalancer.

Delete the target group using DeleteTargetGroup.

All Elastic Load Balancing operations are idempotent, which means that they complete at most one time. If you repeat an operation, it succeeds.

This document is generated from apis/elasticloadbalancingv2-2015-12-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonELBv2.add_tags

AmazonELBv2.create_listener

AmazonELBv2.create_load_balancer

AmazonELBv2.create_rule

AmazonELBv2.create_target_group

AmazonELBv2.delete_listener

AmazonELBv2.delete_load_balancer

AmazonELBv2.delete_rule

AmazonELBv2.delete_target_group

AmazonELBv2.deregister_targets

AmazonELBv2.describe_account_limits

AmazonELBv2.describe_listeners

AmazonELBv2.describe_load_balancer_attributes

AmazonELBv2.describe_load_balancers

AmazonELBv2.describe_rules

AmazonELBv2.describe_sslpolicies

AmazonELBv2.describe_tags

AmazonELBv2.describe_target_group_attributes

AmazonELBv2.describe_target_groups

AmazonELBv2.describe_target_health

AmazonELBv2.modify_listener

AmazonELBv2.modify_load_balancer_attributes

AmazonELBv2.modify_rule

AmazonELBv2.modify_target_group

AmazonELBv2.modify_target_group_attributes

AmazonELBv2.register_targets

AmazonELBv2.remove_tags

AmazonELBv2.set_ip_address_type

AmazonELBv2.set_rule_priorities

AmazonELBv2.set_security_groups

AmazonELBv2.set_subnets

AmazonELBv2.add_tags — Function.

add_tags(::AWSConfig, arguments::Dict)
add_tags([::AWSConfig]; ResourceArns=, Tags=)

AddTags Operation

Adds the specified tags to the specified resource. You can tag your Application Load Balancers and your target groups.

Each tag consists of a key and an optional value. If a resource already has a tag with the same key, AddTags updates its value.

To list the current tags for your resources, use DescribeTags. To remove tags from your resources, use RemoveTags.

Arguments

ResourceArns = [::String...] – Required

The Amazon Resource Name (ARN) of the resource.

Tags = [Dict( ... )...] – Required

The tags. Each resource can have a maximum of 10 tags.

 Tags = [Dict(
        "Key" => <required> ::String,
        "Value" =>  ::String
    )...]

Returns

AddTagsOutput

Exceptions

DuplicateTagKeysException, TooManyTagsException, LoadBalancerNotFoundException or TargetGroupNotFoundException.

See also: AWS API Documentation

AmazonELBv2.create_listener — Function.

create_listener(::AWSConfig, arguments::Dict)
create_listener([::AWSConfig]; LoadBalancerArn=, Protocol=, Port=, DefaultActions=, <keyword arguments>)

CreateListener Operation

Creates a listener for the specified Application Load Balancer.

You can create up to 10 listeners per load balancer.

To update a listener, use ModifyListener. When you are finished with a listener, you can delete it using DeleteListener. If you are finished with both the listener and the load balancer, you can delete them both using DeleteLoadBalancer.

For more information, see Listeners for Your Application Load Balancers in the Application Load Balancers Guide.

Arguments

LoadBalancerArn = ::String – Required

The Amazon Resource Name (ARN) of the load balancer.

Protocol = "HTTP" or "HTTPS" – Required

The protocol for connections from clients to the load balancer.

Port = ::Int – Required

The port on which the load balancer is listening.

SslPolicy = ::String

The security policy that defines which ciphers and protocols are supported. The default is the current predefined security policy.

Certificates = [Dict("CertificateArn" => ::String)...]

The SSL server certificate. You must provide exactly one certificate if the protocol is HTTPS.

DefaultActions = [Dict( ... )...] – Required

The default action for the listener.

 DefaultActions = [Dict(
        "Type" => <required> "forward",
        "TargetGroupArn" => <required> ::String
    )...]

Returns

CreateListenerOutput

Exceptions

DuplicateListenerException, TooManyListenersException, TooManyCertificatesException, LoadBalancerNotFoundException, TargetGroupNotFoundException, TargetGroupAssociationLimitException, InvalidConfigurationRequestException, IncompatibleProtocolsException, SSLPolicyNotFoundException, CertificateNotFoundException, UnsupportedProtocolException or TooManyRegistrationsForTargetIdException.

See also: AWS API Documentation

AmazonELBv2.create_load_balancer — Function.

create_load_balancer(::AWSConfig, arguments::Dict)
create_load_balancer([::AWSConfig]; Name=, Subnets=, <keyword arguments>)

CreateLoadBalancer Operation

Creates an Application Load Balancer.

When you create a load balancer, you can specify security groups, subnets, IP address type, and tags. Otherwise, you could do so later using SetSecurityGroups, SetSubnets, SetIpAddressType, and AddTags.

To create listeners for your load balancer, use CreateListener. To describe your current load balancers, see DescribeLoadBalancers. When you are finished with a load balancer, you can delete it using DeleteLoadBalancer.

You can create up to 20 load balancers per region per account. You can request an increase for the number of load balancers for your account. For more information, see Limits for Your Application Load Balancer in the Application Load Balancers Guide.

For more information, see Application Load Balancers in the Application Load Balancers Guide.

Arguments

Name = ::String – Required

The name of the load balancer.

This name must be unique per region per account, can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and must not begin or end with a hyphen.

Subnets = [::String...] – Required

The IDs of the subnets to attach to the load balancer. You can specify only one subnet per Availability Zone. You must specify subnets from at least two Availability Zones.

SecurityGroups = [::String...]

The IDs of the security groups to assign to the load balancer.

Scheme = "internet-facing" or "internal"

The nodes of an Internet-facing load balancer have public IP addresses. The DNS name of an Internet-facing load balancer is publicly resolvable to the public IP addresses of the nodes. Therefore, Internet-facing load balancers can route requests from clients over the Internet.

The nodes of an internal load balancer have only private IP addresses. The DNS name of an internal load balancer is publicly resolvable to the private IP addresses of the nodes. Therefore, internal load balancers can only route requests from clients with access to the VPC for the load balancer.

The default is an Internet-facing load balancer.

Tags = [Dict( ... )...]

One or more tags to assign to the load balancer.

 Tags = [Dict(
        "Key" => <required> ::String,
        "Value" =>  ::String
    )...]

IpAddressType = "ipv4" or "dualstack"

The type of IP addresses used by the subnets for your load balancer. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). Internal load balancers must use ipv4.

Returns

CreateLoadBalancerOutput

Exceptions

DuplicateLoadBalancerNameException, TooManyLoadBalancersException, InvalidConfigurationRequestException, SubnetNotFoundException, InvalidSubnetException, InvalidSecurityGroupException, InvalidSchemeException, TooManyTagsException or DuplicateTagKeysException.

See also: AWS API Documentation

AmazonELBv2.create_rule — Function.

create_rule(::AWSConfig, arguments::Dict)
create_rule([::AWSConfig]; ListenerArn=, Conditions=, Priority=, Actions=)

CreateRule Operation

Creates a rule for the specified listener.

Each rule can have one action and one condition. Rules are evaluated in priority order, from the lowest value to the highest value. When the condition for a rule is met, the specified action is taken. If no conditions are met, the default action for the default rule is taken. For more information, see Listener Rules in the Application Load Balancers Guide.

To view your current rules, use DescribeRules. To update a rule, use ModifyRule. To set the priorities of your rules, use SetRulePriorities. To delete a rule, use DeleteRule.

Arguments

ListenerArn = ::String – Required

The Amazon Resource Name (ARN) of the listener.

Conditions = [Dict( ... )...] – Required

A condition. Each condition specifies a field name and a single value.

If the field name is host-header, you can specify a single host name (for example, my.example.com). A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters.

A-Z, a-z, 0-9

.

(matches 0 or more characters)

? (matches exactly 1 character)

If the field name is path-pattern, you can specify a single path pattern. A path pattern is case sensitive, can be up to 128 characters in length, and can contain any of the following characters. Note that you can include up to three wildcard characters.

A-Z, a-z, 0-9

_ - . $ / ~ " ' @ : +

& (using &amp;)

(matches 0 or more characters)

? (matches exactly 1 character)

 Conditions = [Dict(
        "Field" =>  ::String,
        "Values" =>  [::String...]
    )...]

Priority = ::Int – Required

The priority for the rule. A listener can't have multiple rules with the same priority.

Actions = [Dict( ... )...] – Required

An action. Each action has the type forward and specifies a target group.

 Actions = [Dict(
        "Type" => <required> "forward",
        "TargetGroupArn" => <required> ::String
    )...]

Returns

CreateRuleOutput

Exceptions

PriorityInUseException, TooManyTargetGroupsException, TooManyRulesException, TargetGroupAssociationLimitException, ListenerNotFoundException, TargetGroupNotFoundException, InvalidConfigurationRequestException or TooManyRegistrationsForTargetIdException.

See also: AWS API Documentation

AmazonELBv2.create_target_group — Function.

create_target_group(::AWSConfig, arguments::Dict)
create_target_group([::AWSConfig]; Name=, Protocol=, Port=, VpcId=, <keyword arguments>)

CreateTargetGroup Operation

Creates a target group.

To register targets with the target group, use RegisterTargets. To update the health check settings for the target group, use ModifyTargetGroup. To monitor the health of targets in the target group, use DescribeTargetHealth.

To route traffic to the targets in a target group, specify the target group in an action using CreateListener or CreateRule.

To delete a target group, use DeleteTargetGroup.

For more information, see Target Groups for Your Application Load Balancers in the Application Load Balancers Guide.

Arguments

Name = ::String – Required

The name of the target group.

This name must be unique per region per account, can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens, and must not begin or end with a hyphen.

Protocol = "HTTP" or "HTTPS" – Required

The protocol to use for routing traffic to the targets.

Port = ::Int – Required

The port on which the targets receive traffic. This port is used unless you specify a port override when registering the target.

VpcId = ::String – Required

The identifier of the virtual private cloud (VPC).

HealthCheckProtocol = "HTTP" or "HTTPS"

The protocol the load balancer uses when performing health checks on targets. The default is the HTTP protocol.

HealthCheckPort = ::String

The port the load balancer uses when performing health checks on targets. The default is traffic-port, which indicates the port on which each target receives traffic from the load balancer.

HealthCheckPath = ::String

The ping path that is the destination on the targets for health checks. The default is /.

HealthCheckIntervalSeconds = ::Int

The approximate amount of time, in seconds, between health checks of an individual target. The default is 30 seconds.

HealthCheckTimeoutSeconds = ::Int

The amount of time, in seconds, during which no response from a target means a failed health check. The default is 5 seconds.

HealthyThresholdCount = ::Int

The number of consecutive health checks successes required before considering an unhealthy target healthy. The default is 5.

UnhealthyThresholdCount = ::Int

The number of consecutive health check failures required before considering a target unhealthy. The default is 2.

Matcher = Dict("HttpCode" => <required> ::String)

The HTTP codes to use when checking for a successful response from a target. The default is 200.

Returns

CreateTargetGroupOutput

Exceptions

DuplicateTargetGroupNameException or TooManyTargetGroupsException.

See also: AWS API Documentation

AmazonELBv2.delete_listener — Function.

delete_listener(::AWSConfig, arguments::Dict)
delete_listener([::AWSConfig]; ListenerArn=)

DeleteListener Operation

Deletes the specified listener.

Alternatively, your listener is deleted when you delete the load balancer it is attached to using DeleteLoadBalancer.

Arguments

ListenerArn = ::String – Required

The Amazon Resource Name (ARN) of the listener.

Returns

DeleteListenerOutput

Exceptions

ListenerNotFoundException.

See also: AWS API Documentation

AmazonELBv2.delete_load_balancer — Function.

delete_load_balancer(::AWSConfig, arguments::Dict)
delete_load_balancer([::AWSConfig]; LoadBalancerArn=)

DeleteLoadBalancer Operation

Deletes the specified Application Load Balancer and its attached listeners.

You can't delete a load balancer if deletion protection is enabled. If the load balancer does not exist or has already been deleted, the call succeeds.

Deleting a load balancer does not affect its registered targets. For example, your EC2 instances continue to run and are still registered to their target groups. If you no longer need these EC2 instances, you can stop or terminate them.

Arguments

LoadBalancerArn = ::String – Required

The Amazon Resource Name (ARN) of the load balancer.

Returns

DeleteLoadBalancerOutput

Exceptions

LoadBalancerNotFoundException or OperationNotPermittedException.

See also: AWS API Documentation

AmazonELBv2.delete_rule — Function.

delete_rule(::AWSConfig, arguments::Dict)
delete_rule([::AWSConfig]; RuleArn=)

DeleteRule Operation

Deletes the specified rule.

Arguments

RuleArn = ::String – Required

The Amazon Resource Name (ARN) of the rule.

Returns

DeleteRuleOutput

Exceptions

RuleNotFoundException or OperationNotPermittedException.

See also: AWS API Documentation

AmazonELBv2.delete_target_group — Function.

delete_target_group(::AWSConfig, arguments::Dict)
delete_target_group([::AWSConfig]; TargetGroupArn=)

DeleteTargetGroup Operation

Deletes the specified target group.

You can delete a target group if it is not referenced by any actions. Deleting a target group also deletes any associated health checks.

Arguments

TargetGroupArn = ::String – Required

The Amazon Resource Name (ARN) of the target group.

Returns

DeleteTargetGroupOutput

Exceptions

ResourceInUseException.

See also: AWS API Documentation

AmazonELBv2.deregister_targets — Function.

deregister_targets(::AWSConfig, arguments::Dict)
deregister_targets([::AWSConfig]; TargetGroupArn=, Targets=)

DeregisterTargets Operation

Deregisters the specified targets from the specified target group. After the targets are deregistered, they no longer receive traffic from the load balancer.

Arguments

TargetGroupArn = ::String – Required

The Amazon Resource Name (ARN) of the target group.

Targets = [Dict( ... )...] – Required

The targets. If you specified a port override when you registered a target, you must specify both the target ID and the port when you deregister it.

 Targets = [Dict(
        "Id" => <required> ::String,
        "Port" =>  ::Int
    )...]

Returns

DeregisterTargetsOutput

Exceptions

TargetGroupNotFoundException or InvalidTargetException.

See also: AWS API Documentation

AmazonELBv2.describe_account_limits — Function.

describe_account_limits(::AWSConfig, arguments::Dict)
describe_account_limits([::AWSConfig]; <keyword arguments>)

DescribeAccountLimits Operation

Describes the current Elastic Load Balancing resource limits for your AWS account.

For more information, see Limits for Your Application Load Balancer in the Application Load Balancer Guide.

Arguments

Marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

PageSize = ::Int

The maximum number of results to return with this call.

Returns

DescribeAccountLimitsOutput

See also: AWS API Documentation

AmazonELBv2.describe_listeners — Function.

describe_listeners(::AWSConfig, arguments::Dict)
describe_listeners([::AWSConfig]; <keyword arguments>)

DescribeListeners Operation

Describes the specified listeners or the listeners for the specified Application Load Balancer. You must specify either a load balancer or one or more listeners.

Arguments

LoadBalancerArn = ::String

The Amazon Resource Name (ARN) of the load balancer.

ListenerArns = [::String...]

The Amazon Resource Names (ARN) of the listeners.

Marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

PageSize = ::Int

The maximum number of results to return with this call.

Returns

DescribeListenersOutput

Exceptions

ListenerNotFoundException or LoadBalancerNotFoundException.

See also: AWS API Documentation

AmazonELBv2.describe_load_balancer_attributes — Function.

describe_load_balancer_attributes(::AWSConfig, arguments::Dict)
describe_load_balancer_attributes([::AWSConfig]; LoadBalancerArn=)

DescribeLoadBalancerAttributes Operation

Describes the attributes for the specified Application Load Balancer.

Arguments

LoadBalancerArn = ::String – Required

The Amazon Resource Name (ARN) of the load balancer.

Returns

DescribeLoadBalancerAttributesOutput

Exceptions

LoadBalancerNotFoundException.

See also: AWS API Documentation

AmazonELBv2.describe_load_balancers — Function.

describe_load_balancers(::AWSConfig, arguments::Dict)
describe_load_balancers([::AWSConfig]; <keyword arguments>)

DescribeLoadBalancers Operation

Describes the specified Application Load Balancers or all of your Application Load Balancers.

To describe the listeners for a load balancer, use DescribeListeners. To describe the attributes for a load balancer, use DescribeLoadBalancerAttributes.

Arguments

LoadBalancerArns = [::String...]

The Amazon Resource Names (ARN) of the load balancers. You can specify up to 20 load balancers in a single call.

Names = [::String...]

The names of the load balancers.

Marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

PageSize = ::Int

The maximum number of results to return with this call.

Returns

DescribeLoadBalancersOutput

Exceptions

LoadBalancerNotFoundException.

See also: AWS API Documentation

AmazonELBv2.describe_rules — Function.

describe_rules(::AWSConfig, arguments::Dict)
describe_rules([::AWSConfig]; <keyword arguments>)

DescribeRules Operation

Describes the specified rules or the rules for the specified listener. You must specify either a listener or one or more rules.

Arguments

ListenerArn = ::String

The Amazon Resource Name (ARN) of the listener.

RuleArns = [::String...]

The Amazon Resource Names (ARN) of the rules.

Marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

PageSize = ::Int

The maximum number of results to return with this call.

Returns

DescribeRulesOutput

Exceptions

ListenerNotFoundException or RuleNotFoundException.

See also: AWS API Documentation

AmazonELBv2.describe_sslpolicies — Function.

describe_sslpolicies(::AWSConfig, arguments::Dict)
describe_sslpolicies([::AWSConfig]; <keyword arguments>)

DescribeSSLPolicies Operation

Describes the specified policies or all policies used for SSL negotiation.

For more information, see Security Policies in the Application Load Balancers Guide.

Arguments

Names = [::String...]

The names of the policies.

Marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

PageSize = ::Int

The maximum number of results to return with this call.

Returns

DescribeSSLPoliciesOutput

Exceptions

SSLPolicyNotFoundException.

See also: AWS API Documentation

AmazonELBv2.describe_tags — Function.

describe_tags(::AWSConfig, arguments::Dict)
describe_tags([::AWSConfig]; ResourceArns=)

DescribeTags Operation

Describes the tags for the specified resources. You can describe the tags for one or more Application Load Balancers and target groups.

Arguments

ResourceArns = [::String...] – Required

The Amazon Resource Names (ARN) of the resources.

Returns

DescribeTagsOutput

Exceptions

LoadBalancerNotFoundException, TargetGroupNotFoundException, ListenerNotFoundException or RuleNotFoundException.

See also: AWS API Documentation

AmazonELBv2.describe_target_group_attributes — Function.

describe_target_group_attributes(::AWSConfig, arguments::Dict)
describe_target_group_attributes([::AWSConfig]; TargetGroupArn=)

DescribeTargetGroupAttributes Operation

Describes the attributes for the specified target group.

Arguments

TargetGroupArn = ::String – Required

The Amazon Resource Name (ARN) of the target group.

Returns

DescribeTargetGroupAttributesOutput

Exceptions

TargetGroupNotFoundException.

See also: AWS API Documentation

AmazonELBv2.describe_target_groups — Function.

describe_target_groups(::AWSConfig, arguments::Dict)
describe_target_groups([::AWSConfig]; <keyword arguments>)

DescribeTargetGroups Operation

Describes the specified target groups or all of your target groups. By default, all target groups are described. Alternatively, you can specify one of the following to filter the results: the ARN of the load balancer, the names of one or more target groups, or the ARNs of one or more target groups.

To describe the targets for a target group, use DescribeTargetHealth. To describe the attributes of a target group, use DescribeTargetGroupAttributes.

Arguments

LoadBalancerArn = ::String

The Amazon Resource Name (ARN) of the load balancer.

TargetGroupArns = [::String...]

The Amazon Resource Names (ARN) of the target groups.

Names = [::String...]

The names of the target groups.

Marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

PageSize = ::Int

The maximum number of results to return with this call.

Returns

DescribeTargetGroupsOutput

Exceptions

LoadBalancerNotFoundException or TargetGroupNotFoundException.

See also: AWS API Documentation

AmazonELBv2.describe_target_health — Function.

describe_target_health(::AWSConfig, arguments::Dict)
describe_target_health([::AWSConfig]; TargetGroupArn=, <keyword arguments>)

DescribeTargetHealth Operation

Describes the health of the specified targets or all of your targets.

Arguments

TargetGroupArn = ::String – Required

The Amazon Resource Name (ARN) of the target group.

Targets = [Dict( ... )...]

The targets.

 Targets = [Dict(
        "Id" => <required> ::String,
        "Port" =>  ::Int
    )...]

Returns

DescribeTargetHealthOutput

Exceptions

InvalidTargetException, TargetGroupNotFoundException or HealthUnavailableException.

See also: AWS API Documentation

AmazonELBv2.modify_listener — Function.

modify_listener(::AWSConfig, arguments::Dict)
modify_listener([::AWSConfig]; ListenerArn=, <keyword arguments>)

ModifyListener Operation

Modifies the specified properties of the specified listener.

Any properties that you do not specify retain their current values. However, changing the protocol from HTTPS to HTTP removes the security policy and SSL certificate properties. If you change the protocol from HTTP to HTTPS, you must add the security policy and server certificate.

Arguments

ListenerArn = ::String – Required

The Amazon Resource Name (ARN) of the listener.

Port = ::Int

The port for connections from clients to the load balancer.

Protocol = "HTTP" or "HTTPS"

The protocol for connections from clients to the load balancer.

SslPolicy = ::String

The security policy that defines which protocols and ciphers are supported. For more information, see Security Policies in the Application Load Balancers Guide.

Certificates = [Dict("CertificateArn" => ::String)...]

The SSL server certificate.

DefaultActions = [Dict( ... )...]

The default actions.

 DefaultActions = [Dict(
        "Type" => <required> "forward",
        "TargetGroupArn" => <required> ::String
    )...]

Returns

ModifyListenerOutput

Exceptions

DuplicateListenerException, TooManyListenersException, TooManyCertificatesException, ListenerNotFoundException, TargetGroupNotFoundException, TargetGroupAssociationLimitException, IncompatibleProtocolsException, SSLPolicyNotFoundException, CertificateNotFoundException, InvalidConfigurationRequestException, UnsupportedProtocolException or TooManyRegistrationsForTargetIdException.

See also: AWS API Documentation

AmazonELBv2.modify_load_balancer_attributes — Function.

modify_load_balancer_attributes(::AWSConfig, arguments::Dict)
modify_load_balancer_attributes([::AWSConfig]; LoadBalancerArn=, Attributes=)

ModifyLoadBalancerAttributes Operation

Modifies the specified attributes of the specified Application Load Balancer.

If any of the specified attributes can't be modified as requested, the call fails. Any existing attributes that you do not modify retain their current values.

Arguments

LoadBalancerArn = ::String – Required

The Amazon Resource Name (ARN) of the load balancer.

Attributes = [Dict( ... )...] – Required

The load balancer attributes.

 Attributes = [Dict(
        "Key" =>  ::String,
        "Value" =>  ::String
    )...]

Returns

ModifyLoadBalancerAttributesOutput

Exceptions

LoadBalancerNotFoundException or InvalidConfigurationRequestException.

See also: AWS API Documentation

AmazonELBv2.modify_rule — Function.

modify_rule(::AWSConfig, arguments::Dict)
modify_rule([::AWSConfig]; RuleArn=, <keyword arguments>)

ModifyRule Operation

Modifies the specified rule.

Any existing properties that you do not modify retain their current values.

To modify the default action, use ModifyListener.

Arguments

RuleArn = ::String – Required

The Amazon Resource Name (ARN) of the rule.

Conditions = [Dict( ... )...]

The conditions.

 Conditions = [Dict(
        "Field" =>  ::String,
        "Values" =>  [::String...]
    )...]

Actions = [Dict( ... )...]

The actions.

 Actions = [Dict(
        "Type" => <required> "forward",
        "TargetGroupArn" => <required> ::String
    )...]

Returns

ModifyRuleOutput

Exceptions

TargetGroupAssociationLimitException, RuleNotFoundException, OperationNotPermittedException, TooManyRegistrationsForTargetIdException, TooManyTargetsException or TargetGroupNotFoundException.

See also: AWS API Documentation

AmazonELBv2.modify_target_group — Function.

modify_target_group(::AWSConfig, arguments::Dict)
modify_target_group([::AWSConfig]; TargetGroupArn=, <keyword arguments>)

ModifyTargetGroup Operation

Modifies the health checks used when evaluating the health state of the targets in the specified target group.

To monitor the health of the targets, use DescribeTargetHealth.

Arguments

TargetGroupArn = ::String – Required

The Amazon Resource Name (ARN) of the target group.

HealthCheckProtocol = "HTTP" or "HTTPS"

The protocol to use to connect with the target.

HealthCheckPort = ::String

The port to use to connect with the target.

HealthCheckPath = ::String

The ping path that is the destination for the health check request.

HealthCheckIntervalSeconds = ::Int

The approximate amount of time, in seconds, between health checks of an individual target.

HealthCheckTimeoutSeconds = ::Int

The amount of time, in seconds, during which no response means a failed health check.

HealthyThresholdCount = ::Int

The number of consecutive health checks successes required before considering an unhealthy target healthy.

UnhealthyThresholdCount = ::Int

The number of consecutive health check failures required before considering the target unhealthy.

Matcher = Dict("HttpCode" => <required> ::String)

The HTTP codes to use when checking for a successful response from a target.

Returns

ModifyTargetGroupOutput

Exceptions

TargetGroupNotFoundException.

See also: AWS API Documentation

AmazonELBv2.modify_target_group_attributes — Function.

modify_target_group_attributes(::AWSConfig, arguments::Dict)
modify_target_group_attributes([::AWSConfig]; TargetGroupArn=, Attributes=)

ModifyTargetGroupAttributes Operation

Modifies the specified attributes of the specified target group.

Arguments

TargetGroupArn = ::String – Required

The Amazon Resource Name (ARN) of the target group.

Attributes = [Dict( ... )...] – Required

The attributes.

 Attributes = [Dict(
        "Key" =>  ::String,
        "Value" =>  ::String
    )...]

Returns

ModifyTargetGroupAttributesOutput

Exceptions

TargetGroupNotFoundException.

See also: AWS API Documentation

AmazonELBv2.register_targets — Function.

register_targets(::AWSConfig, arguments::Dict)
register_targets([::AWSConfig]; TargetGroupArn=, Targets=)

RegisterTargets Operation

Registers the specified targets with the specified target group.

By default, the load balancer routes requests to registered targets using the protocol and port number for the target group. Alternatively, you can override the port for a target when you register it.

The target must be in the virtual private cloud (VPC) that you specified for the target group. If the target is an EC2 instance, it must be in the running state when you register it.

To remove a target from a target group, use DeregisterTargets.

Arguments

TargetGroupArn = ::String – Required

The Amazon Resource Name (ARN) of the target group.

Targets = [Dict( ... )...] – Required

The targets. The default port for a target is the port for the target group. You can specify a port override. If a target is already registered, you can register it again using a different port.

 Targets = [Dict(
        "Id" => <required> ::String,
        "Port" =>  ::Int
    )...]

Returns

RegisterTargetsOutput

Exceptions

TargetGroupNotFoundException, TooManyTargetsException, InvalidTargetException or TooManyRegistrationsForTargetIdException.

See also: AWS API Documentation

AmazonELBv2.remove_tags — Function.

remove_tags(::AWSConfig, arguments::Dict)
remove_tags([::AWSConfig]; ResourceArns=, TagKeys=)

RemoveTags Operation

Removes the specified tags from the specified resource.

To list the current tags for your resources, use DescribeTags.

Arguments

ResourceArns = [::String...] – Required

The Amazon Resource Name (ARN) of the resource.

TagKeys = [::String...] – Required

The tag keys for the tags to remove.

Returns

RemoveTagsOutput

Exceptions

LoadBalancerNotFoundException, TargetGroupNotFoundException, ListenerNotFoundException, RuleNotFoundException or TooManyTagsException.

See also: AWS API Documentation

AmazonELBv2.set_ip_address_type — Function.

set_ip_address_type(::AWSConfig, arguments::Dict)
set_ip_address_type([::AWSConfig]; LoadBalancerArn=, IpAddressType=)

SetIpAddressType Operation

Sets the type of IP addresses used by the subnets of the specified Application Load Balancer.

Arguments

LoadBalancerArn = ::String – Required

The Amazon Resource Name (ARN) of the load balancer.

IpAddressType = "ipv4" or "dualstack" – Required

The IP address type. The possible values are ipv4 (for IPv4 addresses) and dualstack (for IPv4 and IPv6 addresses). Internal load balancers must use ipv4.

Returns

SetIpAddressTypeOutput

Exceptions

LoadBalancerNotFoundException, InvalidConfigurationRequestException or InvalidSubnetException.

See also: AWS API Documentation

AmazonELBv2.set_rule_priorities — Function.

set_rule_priorities(::AWSConfig, arguments::Dict)
set_rule_priorities([::AWSConfig]; RulePriorities=)

SetRulePriorities Operation

Sets the priorities of the specified rules.

You can reorder the rules as long as there are no priority conflicts in the new order. Any existing rules that you do not specify retain their current priority.

Arguments

RulePriorities = [Dict( ... )...] – Required

The rule priorities.

 RulePriorities = [Dict(
        "RuleArn" =>  ::String,
        "Priority" =>  ::Int
    )...]

Returns

SetRulePrioritiesOutput

Exceptions

RuleNotFoundException, PriorityInUseException or OperationNotPermittedException.

See also: AWS API Documentation

AmazonELBv2.set_security_groups — Function.

set_security_groups(::AWSConfig, arguments::Dict)
set_security_groups([::AWSConfig]; LoadBalancerArn=, SecurityGroups=)

SetSecurityGroups Operation

Associates the specified security groups with the specified load balancer. The specified security groups override the previously associated security groups.

Arguments

LoadBalancerArn = ::String – Required

The Amazon Resource Name (ARN) of the load balancer.

SecurityGroups = [::String...] – Required

The IDs of the security groups.

Returns

SetSecurityGroupsOutput

Exceptions

LoadBalancerNotFoundException, InvalidConfigurationRequestException or InvalidSecurityGroupException.

See also: AWS API Documentation

AmazonELBv2.set_subnets — Function.

set_subnets(::AWSConfig, arguments::Dict)
set_subnets([::AWSConfig]; LoadBalancerArn=, Subnets=)

SetSubnets Operation

Enables the Availability Zone for the specified subnets for the specified load balancer. The specified subnets replace the previously enabled subnets.

Arguments

LoadBalancerArn = ::String – Required

The Amazon Resource Name (ARN) of the load balancer.

Subnets = [::String...] – Required

The IDs of the subnets. You must specify at least two subnets. You can add only one subnet per Availability Zone.

Returns

SetSubnetsOutput

Exceptions

LoadBalancerNotFoundException, InvalidConfigurationRequestException, SubnetNotFoundException or InvalidSubnetException.

See also: AWS API Documentation