AmazonElasticBeanstalk.jl

AmazonElasticBeanstalk

AWS Elastic Beanstalk

AWS Elastic Beanstalk makes it easy for you to create, deploy, and manage scalable, fault-tolerant applications running on the Amazon Web Services cloud.

For more information about this product, go to the AWS Elastic Beanstalk details page. The location of the latest AWS Elastic Beanstalk WSDL is http://elasticbeanstalk.s3.amazonaws.com/doc/2010-12-01/AWSElasticBeanstalk.wsdl. To install the Software Development Kits (SDKs), Integrated Development Environment (IDE) Toolkits, and command line tools that enable you to access the API, go to Tools for Amazon Web Services.

Endpoints

For a list of region-specific endpoints that AWS Elastic Beanstalk supports, go to Regions and Endpoints in the Amazon Web Services Glossary.

This document is generated from apis/elasticbeanstalk-2010-12-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonElasticBeanstalk.abort_environment_update

AmazonElasticBeanstalk.apply_environment_managed_action

AmazonElasticBeanstalk.check_dnsavailability

AmazonElasticBeanstalk.compose_environments

AmazonElasticBeanstalk.create_application

AmazonElasticBeanstalk.create_application_version

AmazonElasticBeanstalk.create_configuration_template

AmazonElasticBeanstalk.create_environment

AmazonElasticBeanstalk.create_platform_version

AmazonElasticBeanstalk.create_storage_location

AmazonElasticBeanstalk.delete_application

AmazonElasticBeanstalk.delete_application_version

AmazonElasticBeanstalk.delete_configuration_template

AmazonElasticBeanstalk.delete_environment_configuration

AmazonElasticBeanstalk.delete_platform_version

AmazonElasticBeanstalk.describe_application_versions

AmazonElasticBeanstalk.describe_applications

AmazonElasticBeanstalk.describe_configuration_options

AmazonElasticBeanstalk.describe_configuration_settings

AmazonElasticBeanstalk.describe_environment_health

AmazonElasticBeanstalk.describe_environment_managed_action_history

AmazonElasticBeanstalk.describe_environment_managed_actions

AmazonElasticBeanstalk.describe_environment_resources

AmazonElasticBeanstalk.describe_environments

AmazonElasticBeanstalk.describe_events

AmazonElasticBeanstalk.describe_instances_health

AmazonElasticBeanstalk.describe_platform_version

AmazonElasticBeanstalk.list_available_solution_stacks

AmazonElasticBeanstalk.list_platform_versions

AmazonElasticBeanstalk.rebuild_environment

AmazonElasticBeanstalk.request_environment_info

AmazonElasticBeanstalk.restart_app_server

AmazonElasticBeanstalk.retrieve_environment_info

AmazonElasticBeanstalk.swap_environment_cnames

AmazonElasticBeanstalk.terminate_environment

AmazonElasticBeanstalk.update_application

AmazonElasticBeanstalk.update_application_resource_lifecycle

AmazonElasticBeanstalk.update_application_version

AmazonElasticBeanstalk.update_configuration_template

AmazonElasticBeanstalk.update_environment

AmazonElasticBeanstalk.validate_configuration_settings

AmazonElasticBeanstalk.abort_environment_update — Function.

abort_environment_update(::AWSConfig, arguments::Dict)
abort_environment_update([::AWSConfig]; <keyword arguments>)

AbortEnvironmentUpdate Operation

Cancels in-progress environment configuration update or application version deployment.

Arguments

EnvironmentId = ::String

This specifies the ID of the environment with the in-progress update that you want to cancel.

EnvironmentName = ::String

This specifies the name of the environment with the in-progress update that you want to cancel.

Exceptions

InsufficientPrivilegesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.apply_environment_managed_action — Function.

apply_environment_managed_action(::AWSConfig, arguments::Dict)
apply_environment_managed_action([::AWSConfig]; ActionId=, <keyword arguments>)

ApplyEnvironmentManagedAction Operation

Applies a scheduled managed action immediately. A managed action can be applied only if its status is Scheduled. Get the status and action ID of a managed action with DescribeEnvironmentManagedActions.

Arguments

EnvironmentName = ::String

The name of the target environment.

EnvironmentId = ::String

The environment ID of the target environment.

ActionId = ::String – Required

The action ID of the scheduled managed action to execute.

Returns

ApplyEnvironmentManagedActionResult

Exceptions

ElasticBeanstalkServiceException or ManagedActionInvalidStateException.

See also: AWS API Documentation

AmazonElasticBeanstalk.check_dnsavailability — Function.

check_dnsavailability(::AWSConfig, arguments::Dict)
check_dnsavailability([::AWSConfig]; CNAMEPrefix=)

CheckDNSAvailability Operation

Checks if the specified CNAME is available.

Arguments

CNAMEPrefix = ::String – Required

The prefix used when this CNAME is reserved.

Returns

CheckDNSAvailabilityResultMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.compose_environments — Function.

compose_environments(::AWSConfig, arguments::Dict)
compose_environments([::AWSConfig]; <keyword arguments>)

ComposeEnvironments Operation

Create or update a group of environments that each run a separate component of a single application. Takes a list of version labels that specify application source bundles for each of the environments to create or update. The name of each environment and other required information must be included in the source bundles in an environment manifest named env.yaml. See Compose Environments for details.

Arguments

ApplicationName = ::String

The name of the application to which the specified source bundles belong.

GroupName = ::String

The name of the group to which the target environments belong. Specify a group name only if the environment name defined in each target environment's manifest ends with a + (plus) character. See Environment Manifest (env.yaml) for details.

VersionLabels = [::String...]

A list of version labels, specifying one or more application source bundles that belong to the target application. Each source bundle must include an environment manifest that specifies the name of the environment and the name of the solution stack to use, and optionally can specify environment links to create.

Returns

EnvironmentDescriptionsMessage

Exceptions

TooManyEnvironmentsException or InsufficientPrivilegesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.create_application — Function.

create_application(::AWSConfig, arguments::Dict)
create_application([::AWSConfig]; ApplicationName=, <keyword arguments>)

CreateApplication Operation

Creates an application that has one configuration template named default and no application versions.

Arguments

ApplicationName = ::String – Required

The name of the application.

Constraint: This name must be unique within your account. If the specified name already exists, the action returns an InvalidParameterValue error.

Description = ::String

Describes the application.

ResourceLifecycleConfig = Dict( ... )

Specify an application resource lifecycle configuration to prevent your application from accumulating too many versions.

 ResourceLifecycleConfig = Dict(
        "ServiceRole" =>  ::String,
        "VersionLifecycleConfig" =>  Dict(
            "MaxCountRule" =>  Dict(
                "Enabled" => <required> ::Bool,
                "MaxCount" =>  ::Int,
                "DeleteSourceFromS3" =>  ::Bool
            ),
            "MaxAgeRule" =>  Dict(
                "Enabled" => <required> ::Bool,
                "MaxAgeInDays" =>  ::Int,
                "DeleteSourceFromS3" =>  ::Bool
            )
        )
    )

Returns

ApplicationDescriptionMessage

Exceptions

TooManyApplicationsException.

See also: AWS API Documentation

AmazonElasticBeanstalk.create_application_version — Function.

create_application_version(::AWSConfig, arguments::Dict)
create_application_version([::AWSConfig]; ApplicationName=, VersionLabel=, <keyword arguments>)

CreateApplicationVersion Operation

Creates an application version for the specified application. You can create an application version from a source bundle in Amazon S3, a commit in AWS CodeCommit, or the output of an AWS CodeBuild build as follows:

Specify a commit in an AWS CodeCommit repository with SourceBuildInformation.

Specify a build in an AWS CodeBuild with SourceBuildInformation and BuildConfiguration.

Specify a source bundle in S3 with SourceBundle

Omit both SourceBuildInformation and SourceBundle to use the default sample application.

Note

Once you create an application version with a specified Amazon S3 bucket and key location, you cannot change that Amazon S3 location. If you change the Amazon S3 location, you receive an exception when you attempt to launch an environment from the application version.

Arguments

ApplicationName = ::String – Required

The name of the application. If no application is found with this name, and AutoCreateApplication is false, returns an InvalidParameterValue error.

VersionLabel = ::String – Required

A label identifying this version.

Constraint: Must be unique per application. If an application version already exists with this label for the specified application, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Description = ::String

Describes this version.

SourceBuildInformation = Dict( ... )

Specify a commit in an AWS CodeCommit Git repository to use as the source code for the application version.

 SourceBuildInformation = Dict(
        "SourceType" => <required> "Git" or "Zip",
        "SourceRepository" => <required> "CodeCommit" or "S3",
        "SourceLocation" => <required> ::String
    )

SourceBundle = Dict( ... )

The Amazon S3 bucket and key that identify the location of the source bundle for this version.

Note

The Amazon S3 bucket must be in the same region as the environment.

Specify a source bundle in S3 or a commit in an AWS CodeCommit repository (with SourceBuildInformation), but not both. If neither SourceBundle nor SourceBuildInformation are provided, Elastic Beanstalk uses a sample application.

 SourceBundle = Dict(
        "S3Bucket" =>  ::String,
        "S3Key" =>  ::String
    )

BuildConfiguration = Dict( ... )

Settings for an AWS CodeBuild build.

 BuildConfiguration = Dict(
        "ArtifactName" =>  ::String,
        "CodeBuildServiceRole" => <required> ::String,
        "ComputeType" =>  "BUILD_GENERAL1_SMALL", "BUILD_GENERAL1_MEDIUM" or "BUILD_GENERAL1_LARGE",
        "Image" => <required> ::String,
        "TimeoutInMinutes" =>  ::Int
    )

AutoCreateApplication = ::Bool

Set to true to create an application with the specified name if it doesn't already exist.

Process = ::Bool

Preprocesses and validates the environment manifest and configuration files in the source bundle. Validating configuration files can identify issues prior to deploying the application version to an environment.

Returns

ApplicationVersionDescriptionMessage

Exceptions

TooManyApplicationsException, TooManyApplicationVersionsException, InsufficientPrivilegesException, S3LocationNotInServiceRegionException or CodeBuildNotInServiceRegionException.

See also: AWS API Documentation

AmazonElasticBeanstalk.create_configuration_template — Function.

create_configuration_template(::AWSConfig, arguments::Dict)
create_configuration_template([::AWSConfig]; ApplicationName=, TemplateName=, <keyword arguments>)

CreateConfigurationTemplate Operation

Creates a configuration template. Templates are associated with a specific application and are used to deploy different versions of the application with the same configuration settings.

Related Topics

DescribeConfigurationOptions

DescribeConfigurationSettings

ListAvailableSolutionStacks

Arguments

ApplicationName = ::String – Required

The name of the application to associate with this configuration template. If no application is found with this name, AWS Elastic Beanstalk returns an InvalidParameterValue error.

TemplateName = ::String – Required

The name of the configuration template.

Constraint: This name must be unique per application.

Default: If a configuration template already exists with this name, AWS Elastic Beanstalk returns an InvalidParameterValue error.

SolutionStackName = ::String

The name of the solution stack used by this configuration. The solution stack specifies the operating system, architecture, and application server for a configuration template. It determines the set of configuration options as well as the possible and default values.

Use ListAvailableSolutionStacks to obtain a list of available solution stacks.

A solution stack name or a source configuration parameter must be specified, otherwise AWS Elastic Beanstalk returns an InvalidParameterValue error.

If a solution stack name is not specified and the source configuration parameter is specified, AWS Elastic Beanstalk uses the same solution stack as the source configuration template.

PlatformArn = ::String

The ARN of the custome platform.

SourceConfiguration = Dict( ... )

If specified, AWS Elastic Beanstalk uses the configuration values from the specified configuration template to create a new configuration.

Values specified in the OptionSettings parameter of this call overrides any values obtained from the SourceConfiguration.

If no configuration template is found, returns an InvalidParameterValue error.

Constraint: If both the solution stack name parameter and the source configuration parameters are specified, the solution stack of the source configuration template must match the specified solution stack name or else AWS Elastic Beanstalk returns an InvalidParameterCombination error.

 SourceConfiguration = Dict(
        "ApplicationName" =>  ::String,
        "TemplateName" =>  ::String
    )

EnvironmentId = ::String

The ID of the environment used with this configuration template.

Description = ::String

Describes this configuration.

OptionSettings = [Dict( ... )...]

If specified, AWS Elastic Beanstalk sets the specified configuration option to the requested value. The new value overrides the value obtained from the solution stack or the source configuration template.

 OptionSettings = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String,
        "Value" =>  ::String
    )...]

Returns

ConfigurationSettingsDescription

Exceptions

InsufficientPrivilegesException, TooManyBucketsException or TooManyConfigurationTemplatesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.create_environment — Function.

create_environment(::AWSConfig, arguments::Dict)
create_environment([::AWSConfig]; ApplicationName=, <keyword arguments>)

CreateEnvironment Operation

Launches an environment for the specified application using the specified configuration.

Arguments

ApplicationName = ::String – Required

The name of the application that contains the version to be deployed.

If no application is found with this name, CreateEnvironment returns an InvalidParameterValue error.

EnvironmentName = ::String

A unique name for the deployment environment. Used in the application URL.

Constraint: Must be from 4 to 40 characters in length. The name can contain only letters, numbers, and hyphens. It cannot start or end with a hyphen. This name must be unique in your account. If the specified name already exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Default: If the CNAME parameter is not specified, the environment name becomes part of the CNAME, and therefore part of the visible URL for your application.

GroupName = ::String

The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name parameter. See Environment Manifest (env.yaml) for details.

Description = ::String

Describes this environment.

CNAMEPrefix = ::String

If specified, the environment attempts to use this value as the prefix for the CNAME. If not specified, the CNAME is generated automatically by appending a random alphanumeric string to the environment name.

Tier = Dict( ... )

This specifies the tier to use for creating this environment.

 Tier = Dict(
        "Name" =>  ::String,
        "Type" =>  ::String,
        "Version" =>  ::String
    )

Tags = [Dict( ... )...]

This specifies the tags applied to resources in the environment.

 Tags = [Dict(
        "Key" =>  ::String,
        "Value" =>  ::String
    )...]

VersionLabel = ::String

The name of the application version to deploy.

If the specified application has no associated application versions, AWS Elastic Beanstalk UpdateEnvironment returns an InvalidParameterValue error.

Default: If not specified, AWS Elastic Beanstalk attempts to launch the sample application in the container.

TemplateName = ::String

The name of the configuration template to use in deployment. If no configuration template is found with this name, AWS Elastic Beanstalk returns an InvalidParameterValue error.

SolutionStackName = ::String

This is an alternative to specifying a template name. If specified, AWS Elastic Beanstalk sets the configuration values to the default values associated with the specified solution stack.

PlatformArn = ::String

The ARN of the custom platform.

OptionSettings = [Dict( ... )...]

If specified, AWS Elastic Beanstalk sets the specified configuration options to the requested value in the configuration set for the new environment. These override the values obtained from the solution stack or the configuration template.

 OptionSettings = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String,
        "Value" =>  ::String
    )...]

OptionsToRemove = [Dict( ... )...]

A list of custom user-defined configuration options to remove from the configuration set for this new environment.

 OptionsToRemove = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String
    )...]

Returns

EnvironmentDescription

Exceptions

TooManyEnvironmentsException or InsufficientPrivilegesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.create_platform_version — Function.

create_platform_version(::AWSConfig, arguments::Dict)
create_platform_version([::AWSConfig]; PlatformName=, PlatformVersion=, PlatformDefinitionBundle=, <keyword arguments>)

CreatePlatformVersion Operation

Create a new version of your custom platform.

Arguments

PlatformName = ::String – Required

The name of your custom platform.

PlatformVersion = ::String – Required

The number, such as 1.0.2, for the new platform version.

PlatformDefinitionBundle = Dict( ... ) – Required

The location of the platform definition archive in Amazon S3.

 PlatformDefinitionBundle = Dict(
        "S3Bucket" =>  ::String,
        "S3Key" =>  ::String
    )

EnvironmentName = ::String

The name of the builder environment.

OptionSettings = [Dict( ... )...]

The configuration option settings to apply to the builder environment.

 OptionSettings = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String,
        "Value" =>  ::String
    )...]

Returns

CreatePlatformVersionResult

Exceptions

InsufficientPrivilegesException, ElasticBeanstalkServiceException or TooManyPlatformsException.

See also: AWS API Documentation

AmazonElasticBeanstalk.create_storage_location — Function.

create_storage_location([::AWSConfig])

CreateStorageLocation Operation

Creates the Amazon S3 storage location for the account.

This location is used to store user log files.

Returns

CreateStorageLocationResultMessage

Exceptions

TooManyBucketsException, S3SubscriptionRequiredException or InsufficientPrivilegesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.delete_application — Function.

delete_application(::AWSConfig, arguments::Dict)
delete_application([::AWSConfig]; ApplicationName=, <keyword arguments>)

DeleteApplication Operation

Deletes the specified application along with all associated versions and configurations. The application versions will not be deleted from your Amazon S3 bucket.

Note

You cannot delete an application that has a running environment.

Arguments

ApplicationName = ::String – Required

The name of the application to delete.

TerminateEnvByForce = ::Bool

When set to true, running environments will be terminated before deleting the application.

Exceptions

OperationInProgressException.

See also: AWS API Documentation

AmazonElasticBeanstalk.delete_application_version — Function.

delete_application_version(::AWSConfig, arguments::Dict)
delete_application_version([::AWSConfig]; ApplicationName=, VersionLabel=, <keyword arguments>)

DeleteApplicationVersion Operation

Deletes the specified version from the specified application.

Note

You cannot delete an application version that is associated with a running environment.

Arguments

ApplicationName = ::String – Required

The name of the application to which the version belongs.

VersionLabel = ::String – Required

The label of the version to delete.

DeleteSourceBundle = ::Bool

Set to true to delete the source bundle from your storage bucket. Otherwise, the application version is deleted only from Elastic Beanstalk and the source bundle remains in Amazon S3.

Exceptions

SourceBundleDeletionException, InsufficientPrivilegesException, OperationInProgressException or S3LocationNotInServiceRegionException.

See also: AWS API Documentation

AmazonElasticBeanstalk.delete_configuration_template — Function.

delete_configuration_template(::AWSConfig, arguments::Dict)
delete_configuration_template([::AWSConfig]; ApplicationName=, TemplateName=)

DeleteConfigurationTemplate Operation

Deletes the specified configuration template.

Note

When you launch an environment using a configuration template, the environment gets a copy of the template. You can delete or modify the environment's copy of the template without affecting the running environment.

Arguments

ApplicationName = ::String – Required

The name of the application to delete the configuration template from.

TemplateName = ::String – Required

The name of the configuration template to delete.

Exceptions

OperationInProgressException.

See also: AWS API Documentation

AmazonElasticBeanstalk.delete_environment_configuration — Function.

delete_environment_configuration(::AWSConfig, arguments::Dict)
delete_environment_configuration([::AWSConfig]; ApplicationName=, EnvironmentName=)

DeleteEnvironmentConfiguration Operation

Deletes the draft configuration associated with the running environment.

Updating a running environment with any configuration changes creates a draft configuration set. You can get the draft configuration using DescribeConfigurationSettings while the update is in progress or if the update fails. The DeploymentStatus for the draft configuration indicates whether the deployment is in process or has failed. The draft configuration remains in existence until it is deleted with this action.

Arguments

ApplicationName = ::String – Required

The name of the application the environment is associated with.

EnvironmentName = ::String – Required

The name of the environment to delete the draft configuration from.

See also: AWS API Documentation

AmazonElasticBeanstalk.delete_platform_version — Function.

delete_platform_version(::AWSConfig, arguments::Dict)
delete_platform_version([::AWSConfig]; <keyword arguments>)

DeletePlatformVersion Operation

Deletes the specified version of a custom platform.

Arguments

PlatformArn = ::String

The ARN of the version of the custom platform.

Returns

DeletePlatformVersionResult

Exceptions

OperationInProgressException, InsufficientPrivilegesException, ElasticBeanstalkServiceException or PlatformVersionStillReferencedException.

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_application_versions — Function.

describe_application_versions(::AWSConfig, arguments::Dict)
describe_application_versions([::AWSConfig]; <keyword arguments>)

DescribeApplicationVersions Operation

Retrieve a list of application versions.

Arguments

ApplicationName = ::String

Specify an application name to show only application versions for that application.

VersionLabels = [::String...]

Specify a version label to show a specific application version.

MaxRecords = ::Int

Specify a maximum number of application versions to paginate in the request.

NextToken = ::String

Specify a next token to retrieve the next page in a paginated request.

Returns

ApplicationVersionDescriptionsMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_applications — Function.

describe_applications(::AWSConfig, arguments::Dict)
describe_applications([::AWSConfig]; <keyword arguments>)

DescribeApplications Operation

Returns the descriptions of existing applications.

Arguments

ApplicationNames = [::String...]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to only include those with the specified names.

Returns

ApplicationDescriptionsMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_configuration_options — Function.

describe_configuration_options(::AWSConfig, arguments::Dict)
describe_configuration_options([::AWSConfig]; <keyword arguments>)

DescribeConfigurationOptions Operation

Describes the configuration options that are used in a particular configuration template or environment, or that a specified solution stack defines. The description includes the values the options, their default values, and an indication of the required action on a running environment if an option value is changed.

Arguments

ApplicationName = ::String

The name of the application associated with the configuration template or environment. Only needed if you want to describe the configuration options associated with either the configuration template or environment.

TemplateName = ::String

The name of the configuration template whose configuration options you want to describe.

EnvironmentName = ::String

The name of the environment whose configuration options you want to describe.

SolutionStackName = ::String

The name of the solution stack whose configuration options you want to describe.

PlatformArn = ::String

The ARN of the custom platform.

Options = [Dict( ... )...]

If specified, restricts the descriptions to only the specified options.

 Options = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String
    )...]

Returns

ConfigurationOptionsDescription

Exceptions

TooManyBucketsException.

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_configuration_settings — Function.

describe_configuration_settings(::AWSConfig, arguments::Dict)
describe_configuration_settings([::AWSConfig]; ApplicationName=, <keyword arguments>)

DescribeConfigurationSettings Operation

Returns a description of the settings for the specified configuration set, that is, either a configuration template or the configuration set associated with a running environment.

When describing the settings for the configuration set associated with a running environment, it is possible to receive two sets of setting descriptions. One is the deployed configuration set, and the other is a draft configuration of an environment that is either in the process of deployment or that failed to deploy.

Related Topics

DeleteEnvironmentConfiguration

Arguments

ApplicationName = ::String – Required

The application for the environment or configuration template.

TemplateName = ::String

The name of the configuration template to describe.

Conditional: You must specify either this parameter or an EnvironmentName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns a MissingRequiredParameter error.

EnvironmentName = ::String

The name of the environment to describe.

Condition: You must specify either this or a TemplateName, but not both. If you specify both, AWS Elastic Beanstalk returns an InvalidParameterCombination error. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Returns

ConfigurationSettingsDescriptions

Exceptions

TooManyBucketsException.

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_environment_health — Function.

describe_environment_health(::AWSConfig, arguments::Dict)
describe_environment_health([::AWSConfig]; <keyword arguments>)

DescribeEnvironmentHealth Operation

Returns information about the overall health of the specified environment. The DescribeEnvironmentHealth operation is only available with AWS Elastic Beanstalk Enhanced Health.

Arguments

EnvironmentName = ::String

Specify the environment by name.

You must specify either this or an EnvironmentName, or both.

EnvironmentId = ::String

Specify the environment by ID.

You must specify either this or an EnvironmentName, or both.

AttributeNames = ["Status", "Color", "Causes", "ApplicationMetrics", "InstancesHealth", "All", "HealthStatus" or "RefreshedAt"...]

Specify the response elements to return. To retrieve all attributes, set to All. If no attribute names are specified, returns the name of the environment.

Returns

DescribeEnvironmentHealthResult

Exceptions

InvalidRequestException or ElasticBeanstalkServiceException.

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_environment_managed_action_history — Function.

describe_environment_managed_action_history(::AWSConfig, arguments::Dict)
describe_environment_managed_action_history([::AWSConfig]; <keyword arguments>)

DescribeEnvironmentManagedActionHistory Operation

Lists an environment's completed and failed managed actions.

Arguments

EnvironmentId = ::String

The environment ID of the target environment.

EnvironmentName = ::String

The name of the target environment.

NextToken = ::String

The pagination token returned by a previous request.

MaxItems = ::Int

The maximum number of items to return for a single request.

Returns

DescribeEnvironmentManagedActionHistoryResult

Exceptions

ElasticBeanstalkServiceException.

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_environment_managed_actions — Function.

describe_environment_managed_actions(::AWSConfig, arguments::Dict)
describe_environment_managed_actions([::AWSConfig]; <keyword arguments>)

DescribeEnvironmentManagedActions Operation

Lists an environment's upcoming and in-progress managed actions.

Arguments

EnvironmentName = ::String

The name of the target environment.

EnvironmentId = ::String

The environment ID of the target environment.

Status = "Scheduled", "Pending", "Running" or "Unknown"

To show only actions with a particular status, specify a status.

Returns

DescribeEnvironmentManagedActionsResult

Exceptions

ElasticBeanstalkServiceException.

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_environment_resources — Function.

describe_environment_resources(::AWSConfig, arguments::Dict)
describe_environment_resources([::AWSConfig]; <keyword arguments>)

DescribeEnvironmentResources Operation

Returns AWS resources for this environment.

Arguments

EnvironmentId = ::String

The ID of the environment to retrieve AWS resource usage data.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

EnvironmentName = ::String

The name of the environment to retrieve AWS resource usage data.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Returns

EnvironmentResourceDescriptionsMessage

Exceptions

InsufficientPrivilegesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_environments — Function.

describe_environments(::AWSConfig, arguments::Dict)
describe_environments([::AWSConfig]; <keyword arguments>)

DescribeEnvironments Operation

Returns descriptions for existing environments.

Arguments

ApplicationName = ::String

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application.

VersionLabel = ::String

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that are associated with this application version.

EnvironmentIds = [::String...]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified IDs.

EnvironmentNames = [::String...]

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those that have the specified names.

IncludeDeleted = ::Bool

Indicates whether to include deleted environments:

true: Environments that have been deleted after IncludedDeletedBackTo are displayed.

false: Do not include deleted environments.

IncludedDeletedBackTo = timestamp

If specified when IncludeDeleted is set to true, then environments deleted after this date are displayed.

Returns

EnvironmentDescriptionsMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_events — Function.

describe_events(::AWSConfig, arguments::Dict)
describe_events([::AWSConfig]; <keyword arguments>)

DescribeEvents Operation

Returns list of event descriptions matching criteria up to the last 6 weeks.

Note

This action returns the most recent 1,000 events from the specified NextToken.

Arguments

ApplicationName = ::String

If specified, AWS Elastic Beanstalk restricts the returned descriptions to include only those associated with this application.

VersionLabel = ::String

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this application version.

TemplateName = ::String

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that are associated with this environment configuration.

EnvironmentId = ::String

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

EnvironmentName = ::String

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those associated with this environment.

PlatformArn = ::String

The ARN of the version of the custom platform.

RequestId = ::String

If specified, AWS Elastic Beanstalk restricts the described events to include only those associated with this request ID.

Severity = "TRACE", "DEBUG", "INFO", "WARN", "ERROR" or "FATAL"

If specified, limits the events returned from this call to include only those with the specified severity or higher.

StartTime = timestamp

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur on or after this time.

EndTime = timestamp

If specified, AWS Elastic Beanstalk restricts the returned descriptions to those that occur up to, but not including, the EndTime.

MaxRecords = ::Int

Specifies the maximum number of events that can be returned, beginning with the most recent event.

NextToken = ::String

Pagination token. If specified, the events return the next batch of results.

Returns

EventDescriptionsMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_instances_health — Function.

describe_instances_health(::AWSConfig, arguments::Dict)
describe_instances_health([::AWSConfig]; <keyword arguments>)

DescribeInstancesHealth Operation

Retrives detailed information about the health of instances in your AWS Elastic Beanstalk. This operation requires enhanced health reporting.

Arguments

EnvironmentName = ::String

Specify the AWS Elastic Beanstalk environment by name.

EnvironmentId = ::String

Specify the AWS Elastic Beanstalk environment by ID.

AttributeNames = ["HealthStatus", "Color", "Causes", "ApplicationMetrics", "RefreshedAt", "LaunchedAt", "System", "Deployment", "AvailabilityZone", "InstanceType" or "All"...]

Specifies the response elements you wish to receive. To retrieve all attributes, set to All. If no attribute names are specified, returns a list of instances.

NextToken = ::String

Specify the pagination token returned by a previous call.

Returns

DescribeInstancesHealthResult

Exceptions

InvalidRequestException or ElasticBeanstalkServiceException.

See also: AWS API Documentation

AmazonElasticBeanstalk.describe_platform_version — Function.

describe_platform_version(::AWSConfig, arguments::Dict)
describe_platform_version([::AWSConfig]; <keyword arguments>)

DescribePlatformVersion Operation

Describes the version of the platform.

Arguments

PlatformArn = ::String

The ARN of the version of the platform.

Returns

DescribePlatformVersionResult

Exceptions

InsufficientPrivilegesException or ElasticBeanstalkServiceException.

See also: AWS API Documentation

AmazonElasticBeanstalk.list_available_solution_stacks — Function.

list_available_solution_stacks([::AWSConfig])

ListAvailableSolutionStacks Operation

Returns a list of the available solution stack names, with the public version first and then in reverse chronological order.

Returns

ListAvailableSolutionStacksResultMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.list_platform_versions — Function.

list_platform_versions(::AWSConfig, arguments::Dict)
list_platform_versions([::AWSConfig]; <keyword arguments>)

ListPlatformVersions Operation

Lists the available platforms.

Arguments

Filters = [Dict( ... )...]

List only the platforms where the platform member value relates to one of the supplied values.

 Filters = [Dict(
        "Type" =>  ::String,
        "Operator" =>  ::String,
        "Values" =>  [::String...]
    )...]

MaxRecords = ::Int

The maximum number of platform values returned in one call.

NextToken = ::String

The starting index into the remaining list of platforms. Use the NextToken value from a previous ListPlatformVersion call.

Returns

ListPlatformVersionsResult

Exceptions

InsufficientPrivilegesException or ElasticBeanstalkServiceException.

See also: AWS API Documentation

AmazonElasticBeanstalk.rebuild_environment — Function.

rebuild_environment(::AWSConfig, arguments::Dict)
rebuild_environment([::AWSConfig]; <keyword arguments>)

RebuildEnvironment Operation

Deletes and recreates all of the AWS resources (for example: the Auto Scaling group, load balancer, etc.) for a specified environment and forces a restart.

Arguments

EnvironmentId = ::String

The ID of the environment to rebuild.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

EnvironmentName = ::String

The name of the environment to rebuild.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

Exceptions

InsufficientPrivilegesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.request_environment_info — Function.

request_environment_info(::AWSConfig, arguments::Dict)
request_environment_info([::AWSConfig]; InfoType=, <keyword arguments>)

RequestEnvironmentInfo Operation

Initiates a request to compile the specified type of information of the deployed environment.

Setting the InfoType to tail compiles the last lines from the application server log files of every Amazon EC2 instance in your environment.

Setting the InfoType to bundle compresses the application server log files for every Amazon EC2 instance into a .zip file. Legacy and .NET containers do not support bundle logs.

Use RetrieveEnvironmentInfo to obtain the set of logs.

Related Topics

RetrieveEnvironmentInfo

Arguments

EnvironmentId = ::String

The ID of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

EnvironmentName = ::String

The name of the environment of the requested data.

If no such environment is found, RequestEnvironmentInfo returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

InfoType = "tail" or "bundle" – Required

The type of information to request.

See also: AWS API Documentation

AmazonElasticBeanstalk.restart_app_server — Function.

restart_app_server(::AWSConfig, arguments::Dict)
restart_app_server([::AWSConfig]; <keyword arguments>)

RestartAppServer Operation

Causes the environment to restart the application container server running on each Amazon EC2 instance.

Arguments

EnvironmentId = ::String

The ID of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

EnvironmentName = ::String

The name of the environment to restart the server for.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

See also: AWS API Documentation

AmazonElasticBeanstalk.retrieve_environment_info — Function.

retrieve_environment_info(::AWSConfig, arguments::Dict)
retrieve_environment_info([::AWSConfig]; InfoType=, <keyword arguments>)

RetrieveEnvironmentInfo Operation

Retrieves the compiled information from a RequestEnvironmentInfo request.

Related Topics

RequestEnvironmentInfo

Arguments

EnvironmentId = ::String

The ID of the data's environment.

If no such environment is found, returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

EnvironmentName = ::String

The name of the data's environment.

If no such environment is found, returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

InfoType = "tail" or "bundle" – Required

The type of information to retrieve.

Returns

RetrieveEnvironmentInfoResultMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.swap_environment_cnames — Function.

swap_environment_cnames(::AWSConfig, arguments::Dict)
swap_environment_cnames([::AWSConfig]; <keyword arguments>)

SwapEnvironmentCNAMEs Operation

Swaps the CNAMEs of two environments.

Arguments

SourceEnvironmentId = ::String

The ID of the source environment.

Condition: You must specify at least the SourceEnvironmentID or the SourceEnvironmentName. You may also specify both. If you specify the SourceEnvironmentId, you must specify the DestinationEnvironmentId.

SourceEnvironmentName = ::String

The name of the source environment.

Condition: You must specify at least the SourceEnvironmentID or the SourceEnvironmentName. You may also specify both. If you specify the SourceEnvironmentName, you must specify the DestinationEnvironmentName.

DestinationEnvironmentId = ::String

The ID of the destination environment.

Condition: You must specify at least the DestinationEnvironmentID or the DestinationEnvironmentName. You may also specify both. You must specify the SourceEnvironmentId with the DestinationEnvironmentId.

DestinationEnvironmentName = ::String

The name of the destination environment.

Condition: You must specify at least the DestinationEnvironmentID or the DestinationEnvironmentName. You may also specify both. You must specify the SourceEnvironmentName with the DestinationEnvironmentName.

See also: AWS API Documentation

AmazonElasticBeanstalk.terminate_environment — Function.

terminate_environment(::AWSConfig, arguments::Dict)
terminate_environment([::AWSConfig]; <keyword arguments>)

TerminateEnvironment Operation

Terminates the specified environment.

Arguments

EnvironmentId = ::String

The ID of the environment to terminate.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

EnvironmentName = ::String

The name of the environment to terminate.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

TerminateResources = ::Bool

Indicates whether the associated AWS resources should shut down when the environment is terminated:

true: The specified environment as well as the associated AWS resources, such as Auto Scaling group and LoadBalancer, are terminated.

false: AWS Elastic Beanstalk resource management is removed from the environment, but the AWS resources continue to operate.

For more information, see the AWS Elastic Beanstalk User Guide.

Default: true

Valid Values: true | false

ForceTerminate = ::Bool

Terminates the target environment even if another environment in the same group is dependent on it.

Returns

EnvironmentDescription

Exceptions

InsufficientPrivilegesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.update_application — Function.

update_application(::AWSConfig, arguments::Dict)
update_application([::AWSConfig]; ApplicationName=, <keyword arguments>)

UpdateApplication Operation

Updates the specified application to have the specified properties.

Note

If a property (for example, description) is not provided, the value remains unchanged. To clear these properties, specify an empty string.

Arguments

ApplicationName = ::String – Required

The name of the application to update. If no such application is found, UpdateApplication returns an InvalidParameterValue error.

Description = ::String

A new description for the application.

Default: If not specified, AWS Elastic Beanstalk does not update the description.

Returns

ApplicationDescriptionMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.update_application_resource_lifecycle — Function.

update_application_resource_lifecycle(::AWSConfig, arguments::Dict)
update_application_resource_lifecycle([::AWSConfig]; ApplicationName=, ResourceLifecycleConfig=)

UpdateApplicationResourceLifecycle Operation

Modifies lifecycle settings for an application.

Arguments

ApplicationName = ::String – Required

The name of the application.

ResourceLifecycleConfig = Dict( ... ) – Required

The lifecycle configuration.

 ResourceLifecycleConfig = Dict(
        "ServiceRole" =>  ::String,
        "VersionLifecycleConfig" =>  Dict(
            "MaxCountRule" =>  Dict(
                "Enabled" => <required> ::Bool,
                "MaxCount" =>  ::Int,
                "DeleteSourceFromS3" =>  ::Bool
            ),
            "MaxAgeRule" =>  Dict(
                "Enabled" => <required> ::Bool,
                "MaxAgeInDays" =>  ::Int,
                "DeleteSourceFromS3" =>  ::Bool
            )
        )
    )

Returns

ApplicationResourceLifecycleDescriptionMessage

Exceptions

InsufficientPrivilegesException.

See also: AWS API Documentation

AmazonElasticBeanstalk.update_application_version — Function.

update_application_version(::AWSConfig, arguments::Dict)
update_application_version([::AWSConfig]; ApplicationName=, VersionLabel=, <keyword arguments>)

UpdateApplicationVersion Operation

Updates the specified application version to have the specified properties.

Note

If a property (for example, description) is not provided, the value remains unchanged. To clear properties, specify an empty string.

Arguments

ApplicationName = ::String – Required

The name of the application associated with this version.

If no application is found with this name, UpdateApplication returns an InvalidParameterValue error.

VersionLabel = ::String – Required

The name of the version to update.

If no application version is found with this label, UpdateApplication returns an InvalidParameterValue error.

Description = ::String

A new description for this version.

Returns

ApplicationVersionDescriptionMessage

See also: AWS API Documentation

AmazonElasticBeanstalk.update_configuration_template — Function.

update_configuration_template(::AWSConfig, arguments::Dict)
update_configuration_template([::AWSConfig]; ApplicationName=, TemplateName=, <keyword arguments>)

UpdateConfigurationTemplate Operation

Updates the specified configuration template to have the specified properties or configuration option values.

Note

If a property (for example, ApplicationName) is not provided, its value remains unchanged. To clear such properties, specify an empty string.

Related Topics

DescribeConfigurationOptions

Arguments

ApplicationName = ::String – Required

The name of the application associated with the configuration template to update.

If no application is found with this name, UpdateConfigurationTemplate returns an InvalidParameterValue error.

TemplateName = ::String – Required

The name of the configuration template to update.

If no configuration template is found with this name, UpdateConfigurationTemplate returns an InvalidParameterValue error.

Description = ::String

A new description for the configuration.

OptionSettings = [Dict( ... )...]

A list of configuration option settings to update with the new specified option value.

 OptionSettings = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String,
        "Value" =>  ::String
    )...]

OptionsToRemove = [Dict( ... )...]

A list of configuration options to remove from the configuration set.

Constraint: You can remove only UserDefined configuration options.

 OptionsToRemove = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String
    )...]

Returns

ConfigurationSettingsDescription

Exceptions

InsufficientPrivilegesException or TooManyBucketsException.

See also: AWS API Documentation

AmazonElasticBeanstalk.update_environment — Function.

update_environment(::AWSConfig, arguments::Dict)
update_environment([::AWSConfig]; <keyword arguments>)

UpdateEnvironment Operation

Updates the environment description, deploys a new application version, updates the configuration settings to an entirely new configuration template, or updates select configuration option values in the running environment.

Attempting to update both the release and configuration is not allowed and AWS Elastic Beanstalk returns an InvalidParameterCombination error.

When updating the configuration settings to a new template or individual settings, a draft configuration is created and DescribeConfigurationSettings for this environment returns two setting descriptions with different DeploymentStatus values.

Arguments

ApplicationName = ::String

The name of the application with which the environment is associated.

EnvironmentId = ::String

The ID of the environment to update.

If no environment with this ID exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentName, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

EnvironmentName = ::String

The name of the environment to update. If no environment with this name exists, AWS Elastic Beanstalk returns an InvalidParameterValue error.

Condition: You must specify either this or an EnvironmentId, or both. If you do not specify either, AWS Elastic Beanstalk returns MissingRequiredParameter error.

GroupName = ::String

The name of the group to which the target environment belongs. Specify a group name only if the environment's name is specified in an environment manifest and not with the environment name or environment ID parameters. See Environment Manifest (env.yaml) for details.

Description = ::String

If this parameter is specified, AWS Elastic Beanstalk updates the description of this environment.

Tier = Dict( ... )

This specifies the tier to use to update the environment.

Condition: At this time, if you change the tier version, name, or type, AWS Elastic Beanstalk returns InvalidParameterValue error.

 Tier = Dict(
        "Name" =>  ::String,
        "Type" =>  ::String,
        "Version" =>  ::String
    )

VersionLabel = ::String

If this parameter is specified, AWS Elastic Beanstalk deploys the named application version to the environment. If no such application version is found, returns an InvalidParameterValue error.

TemplateName = ::String

If this parameter is specified, AWS Elastic Beanstalk deploys this configuration template to the environment. If no such configuration template is found, AWS Elastic Beanstalk returns an InvalidParameterValue error.

SolutionStackName = ::String

This specifies the platform version that the environment will run after the environment is updated.

PlatformArn = ::String

The ARN of the platform, if used.

OptionSettings = [Dict( ... )...]

If specified, AWS Elastic Beanstalk updates the configuration set associated with the running environment and sets the specified configuration options to the requested value.

 OptionSettings = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String,
        "Value" =>  ::String
    )...]

OptionsToRemove = [Dict( ... )...]

A list of custom user-defined configuration options to remove from the configuration set for this environment.

 OptionsToRemove = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String
    )...]

Returns

EnvironmentDescription

Exceptions

InsufficientPrivilegesException or TooManyBucketsException.

See also: AWS API Documentation

AmazonElasticBeanstalk.validate_configuration_settings — Function.

validate_configuration_settings(::AWSConfig, arguments::Dict)
validate_configuration_settings([::AWSConfig]; ApplicationName=, OptionSettings=, <keyword arguments>)

ValidateConfigurationSettings Operation

Takes a set of configuration settings and either a configuration template or environment, and determines whether those values are valid.

This action returns a list of messages indicating any errors or warnings associated with the selection of option values.

Arguments

ApplicationName = ::String – Required

The name of the application that the configuration template or environment belongs to.

TemplateName = ::String

The name of the configuration template to validate the settings against.

Condition: You cannot specify both this and an environment name.

EnvironmentName = ::String

The name of the environment to validate the settings against.

Condition: You cannot specify both this and a configuration template name.

OptionSettings = [Dict( ... )...] – Required

A list of the options and desired values to evaluate.

 OptionSettings = [Dict(
        "ResourceName" =>  ::String,
        "Namespace" =>  ::String,
        "OptionName" =>  ::String,
        "Value" =>  ::String
    )...]

Returns

ConfigurationSettingsValidationMessages

Exceptions

InsufficientPrivilegesException or TooManyBucketsException.

See also: AWS API Documentation