AmazonGreengrass.jl

AmazonGreengrass

AWS Greengrass seamlessly extends AWS onto physical devices so they can act locally on the data they generate, while still using the cloud for management, analytics, and durable storage. AWS Greengrass ensures your devices can respond quickly to local events and operate with intermittent connectivity. AWS Greengrass minimizes the cost of transmitting data to the cloud by allowing you to author AWS Lambda functions that execute locally.

This document is generated from apis/greengrass-2017-06-07.normal.json. See JuliaCloud/AWSCore.jl.

AmazonGreengrass.associate_role_to_group

AmazonGreengrass.associate_service_role_to_account

AmazonGreengrass.create_core_definition

AmazonGreengrass.create_core_definition_version

AmazonGreengrass.create_deployment

AmazonGreengrass.create_device_definition

AmazonGreengrass.create_device_definition_version

AmazonGreengrass.create_function_definition

AmazonGreengrass.create_function_definition_version

AmazonGreengrass.create_group

AmazonGreengrass.create_group_certificate_authority

AmazonGreengrass.create_group_version

AmazonGreengrass.create_logger_definition

AmazonGreengrass.create_logger_definition_version

AmazonGreengrass.create_subscription_definition

AmazonGreengrass.create_subscription_definition_version

AmazonGreengrass.delete_core_definition

AmazonGreengrass.delete_device_definition

AmazonGreengrass.delete_function_definition

AmazonGreengrass.delete_group

AmazonGreengrass.delete_logger_definition

AmazonGreengrass.delete_subscription_definition

AmazonGreengrass.disassociate_role_from_group

AmazonGreengrass.disassociate_service_role_from_account

AmazonGreengrass.get_associated_role

AmazonGreengrass.get_connectivity_info

AmazonGreengrass.get_core_definition

AmazonGreengrass.get_core_definition_version

AmazonGreengrass.get_deployment_status

AmazonGreengrass.get_device_definition

AmazonGreengrass.get_device_definition_version

AmazonGreengrass.get_function_definition

AmazonGreengrass.get_function_definition_version

AmazonGreengrass.get_group

AmazonGreengrass.get_group_certificate_authority

AmazonGreengrass.get_group_certificate_configuration

AmazonGreengrass.get_group_version

AmazonGreengrass.get_logger_definition

AmazonGreengrass.get_logger_definition_version

AmazonGreengrass.get_service_role_for_account

AmazonGreengrass.get_subscription_definition

AmazonGreengrass.get_subscription_definition_version

AmazonGreengrass.list_core_definition_versions

AmazonGreengrass.list_core_definitions

AmazonGreengrass.list_deployments

AmazonGreengrass.list_device_definition_versions

AmazonGreengrass.list_device_definitions

AmazonGreengrass.list_function_definition_versions

AmazonGreengrass.list_function_definitions

AmazonGreengrass.list_group_certificate_authorities

AmazonGreengrass.list_group_versions

AmazonGreengrass.list_groups

AmazonGreengrass.list_logger_definition_versions

AmazonGreengrass.list_logger_definitions

AmazonGreengrass.list_subscription_definition_versions

AmazonGreengrass.list_subscription_definitions

AmazonGreengrass.update_connectivity_info

AmazonGreengrass.update_core_definition

AmazonGreengrass.update_device_definition

AmazonGreengrass.update_function_definition

AmazonGreengrass.update_group

AmazonGreengrass.update_group_certificate_configuration

AmazonGreengrass.update_logger_definition

AmazonGreengrass.update_subscription_definition

AmazonGreengrass.associate_role_to_group — Function.

using AmazonGreengrass
associate_role_to_group([::AWSConfig], arguments::Dict)
associate_role_to_group([::AWSConfig]; GroupId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/groups/{GroupId}/role", args)

AssociateRoleToGroup Operation

Associates a role with a group. The role will be used by the AWS Greengrass core in order to access AWS cloud services. The role's permissions will allow Greengrass core Lambda functions to perform actions against the cloud.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

RoleArn = ::String

Role arn you wish to associate with this group.

Returns

AssociateRoleToGroupResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.associate_service_role_to_account — Function.

using AmazonGreengrass
associate_service_role_to_account([::AWSConfig], arguments::Dict)
associate_service_role_to_account([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/servicerole", args)

AssociateServiceRoleToAccount Operation

Associates a role which is used by AWS Greengrass. AWS Greengrass uses the role to access your Lambda functions and AWS IoT resources. This is necessary for deployments to succeed. It needs to have minimum permissions in policy $AWSGreengrassResourceAccessRolePolicy$

Arguments

RoleArn = ::String

Role arn you wish to associate with this account.

Returns

AssociateServiceRoleToAccountResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.create_core_definition — Function.

using AmazonGreengrass
create_core_definition([::AWSConfig], arguments::Dict)
create_core_definition([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/cores", args)

CreateCoreDefinition Operation

Creates a core definition. You may optionally provide the initial version of the core definition or use ''CreateCoreDefinitionVersion'' at a later time. AWS Greengrass Groups must each contain exactly 1 AWS Greengrass Core.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

InitialVersion = ["Cores" => [[ ... ], ...]]

Information on the initial version

 InitialVersion = ["Cores" =>  [[
            "CertificateArn" =>  ::String,
            "Id" =>  ::String,
            "SyncShadow" =>  ::Bool,
            "ThingArn" =>  ::String
        ], ...]]

Name = ::String

name of the core definition

Returns

CreateCoreDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_core_definition_version — Function.

using AmazonGreengrass
create_core_definition_version([::AWSConfig], arguments::Dict)
create_core_definition_version([::AWSConfig]; CoreDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/cores/{CoreDefinitionId}/versions", args)

CreateCoreDefinitionVersion Operation

Creates a version of a core definition that has already been defined. AWS Greengrass Groups must each contain exactly 1 AWS Greengrass Core.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

CoreDefinitionId = ::String – Required

core definition Id

Cores = [[ ... ], ...]

Cores in the definition version.

 Cores = [[
        "CertificateArn" =>  ::String,
        "Id" =>  ::String,
        "SyncShadow" =>  ::Bool,
        "ThingArn" =>  ::String
    ], ...]

Returns

CreateCoreDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_deployment — Function.

using AmazonGreengrass
create_deployment([::AWSConfig], arguments::Dict)
create_deployment([::AWSConfig]; GroupId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/groups/{GroupId}/deployments", args)

CreateDeployment Operation

Creates a deployment.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

DeploymentId = ::String

Id of the deployment if you wish to redeploy a previous deployment.

DeploymentType = "NewDeployment" or "Redeployment"

Type of deployment

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

GroupVersionId = ::String

Group Version you wish to deploy.

Returns

CreateDeploymentResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_device_definition — Function.

using AmazonGreengrass
create_device_definition([::AWSConfig], arguments::Dict)
create_device_definition([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/devices", args)

CreateDeviceDefinition Operation

Creates a device definition. You may optinally provide the initial version of the device definition or use $CreateDeviceDefinitionVersion$ at a later time.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

InitialVersion = ["Devices" => [[ ... ], ...]]

Information on the initial version

 InitialVersion = ["Devices" =>  [[
            "CertificateArn" =>  ::String,
            "Id" =>  ::String,
            "SyncShadow" =>  ::Bool,
            "ThingArn" =>  ::String
        ], ...]]

Name = ::String

name of the device definition

Returns

CreateDeviceDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_device_definition_version — Function.

using AmazonGreengrass
create_device_definition_version([::AWSConfig], arguments::Dict)
create_device_definition_version([::AWSConfig]; DeviceDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/devices/{DeviceDefinitionId}/versions", args)

CreateDeviceDefinitionVersion Operation

Creates a version of a device definition that has already been defined.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

DeviceDefinitionId = ::String – Required

device definition Id

Devices = [[ ... ], ...]

Devices in the definition version.

 Devices = [[
        "CertificateArn" =>  ::String,
        "Id" =>  ::String,
        "SyncShadow" =>  ::Bool,
        "ThingArn" =>  ::String
    ], ...]

Returns

CreateDeviceDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_function_definition — Function.

using AmazonGreengrass
create_function_definition([::AWSConfig], arguments::Dict)
create_function_definition([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/functions", args)

CreateFunctionDefinition Operation

Creates a Lambda function definition which contains a list of Lambda functions and their configurations to be used in a group. You can create an initial version of the definition by providing a list of Lambda functions and their configurations now, or use $CreateFunctionDefinitionVersion$ later.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

InitialVersion = ["Functions" => [[ ... ], ...]]

Information on the initial version

 InitialVersion = ["Functions" =>  [[
            "FunctionArn" =>  ::String,
            "FunctionConfiguration" =>  [
                "Environment" =>  ["Variables" =>  ::Dict{String,String}],
                "ExecArgs" =>  ::String,
                "Executable" =>  ::String,
                "MemorySize" =>  ::Int,
                "Pinned" =>  ::Bool,
                "Timeout" =>  ::Int
            ],
            "Id" =>  ::String
        ], ...]]

Name = ::String

name of the function definition

Returns

CreateFunctionDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_function_definition_version — Function.

using AmazonGreengrass
create_function_definition_version([::AWSConfig], arguments::Dict)
create_function_definition_version([::AWSConfig]; FunctionDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/functions/{FunctionDefinitionId}/versions", args)

CreateFunctionDefinitionVersion Operation

Create a version of a Lambda function definition that has already been defined.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

FunctionDefinitionId = ::String – Required

the unique Id of the lambda definition

Functions = [[ ... ], ...]

Lambda functions in this function definition version.

 Functions = [[
        "FunctionArn" =>  ::String,
        "FunctionConfiguration" =>  [
            "Environment" =>  ["Variables" =>  ::Dict{String,String}],
            "ExecArgs" =>  ::String,
            "Executable" =>  ::String,
            "MemorySize" =>  ::Int,
            "Pinned" =>  ::Bool,
            "Timeout" =>  ::Int
        ],
        "Id" =>  ::String
    ], ...]

Returns

CreateFunctionDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_group — Function.

using AmazonGreengrass
create_group([::AWSConfig], arguments::Dict)
create_group([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/groups", args)

CreateGroup Operation

Creates a group. You may optionally provide the initial version of the group or use ''CreateGroupVersion'' at a later time.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

InitialVersion = [ ... ]

Information on the initial version

 InitialVersion = [
        "CoreDefinitionVersionArn" =>  ::String,
        "DeviceDefinitionVersionArn" =>  ::String,
        "FunctionDefinitionVersionArn" =>  ::String,
        "LoggerDefinitionVersionArn" =>  ::String,
        "SubscriptionDefinitionVersionArn" =>  ::String
    ]

Name = ::String

name of the group

Returns

CreateGroupResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_group_certificate_authority — Function.

using AmazonGreengrass
create_group_certificate_authority([::AWSConfig], arguments::Dict)
create_group_certificate_authority([::AWSConfig]; GroupId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/groups/{GroupId}/certificateauthorities", args)

CreateGroupCertificateAuthority Operation

Creates a CA for the group. If a CA already exists, it will rotate the existing CA.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

CreateGroupCertificateAuthorityResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.create_group_version — Function.

using AmazonGreengrass
create_group_version([::AWSConfig], arguments::Dict)
create_group_version([::AWSConfig]; GroupId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/groups/{GroupId}/versions", args)

CreateGroupVersion Operation

Creates a version of a group which has already been defined.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

CoreDefinitionVersionArn = ::String

Core definition version arn for this group.

DeviceDefinitionVersionArn = ::String

Device definition version arn for this group.

FunctionDefinitionVersionArn = ::String

Function definition version arn for this group.

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

LoggerDefinitionVersionArn = ::String

Logger definitionv ersion arn for this group.

SubscriptionDefinitionVersionArn = ::String

Subscription definition version arn for this group.

Returns

CreateGroupVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_logger_definition — Function.

using AmazonGreengrass
create_logger_definition([::AWSConfig], arguments::Dict)
create_logger_definition([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/loggers", args)

CreateLoggerDefinition Operation

Creates a logger definition. You may optionally provide the initial version of the logger definition or use $CreateLoggerDefinitionVersion$ at a later time.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

InitialVersion = ["Loggers" => [[ ... ], ...]]

Information on the initial version

 InitialVersion = ["Loggers" =>  [[
            "Component" =>  "GreengrassSystem" or "Lambda",
            "Id" =>  ::String,
            "Level" =>  "DEBUG", "INFO", "WARN", "ERROR" or "FATAL",
            "Space" =>  ::Int,
            "Type" =>  "FileSystem" or "AWSCloudWatch"
        ], ...]]

Name = ::String

name of the logger definition

Returns

CreateLoggerDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_logger_definition_version — Function.

using AmazonGreengrass
create_logger_definition_version([::AWSConfig], arguments::Dict)
create_logger_definition_version([::AWSConfig]; LoggerDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/loggers/{LoggerDefinitionId}/versions", args)

CreateLoggerDefinitionVersion Operation

Creates a version of a logger definition that has already been defined.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

LoggerDefinitionId = ::String – Required

logger definition Id

Loggers = [[ ... ], ...]

List of loggers.

 Loggers = [[
        "Component" =>  "GreengrassSystem" or "Lambda",
        "Id" =>  ::String,
        "Level" =>  "DEBUG", "INFO", "WARN", "ERROR" or "FATAL",
        "Space" =>  ::Int,
        "Type" =>  "FileSystem" or "AWSCloudWatch"
    ], ...]

Returns

CreateLoggerDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_subscription_definition — Function.

using AmazonGreengrass
create_subscription_definition([::AWSConfig], arguments::Dict)
create_subscription_definition([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/subscriptions", args)

CreateSubscriptionDefinition Operation

Creates a subscription definition. You may optionally provide the initial version of the subscription definition or use $CreateSubscriptionDefinitionVersion$ at a later time.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

InitialVersion = ["Subscriptions" => [[ ... ], ...]]

Information on the initial version

 InitialVersion = ["Subscriptions" =>  [[
            "Id" =>  ::String,
            "Source" =>  ::String,
            "Subject" =>  ::String,
            "Target" =>  ::String
        ], ...]]

Name = ::String

name of the subscription definition

Returns

CreateSubscriptionDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.create_subscription_definition_version — Function.

using AmazonGreengrass
create_subscription_definition_version([::AWSConfig], arguments::Dict)
create_subscription_definition_version([::AWSConfig]; SubscriptionDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "POST", "/greengrass/definition/subscriptions/{SubscriptionDefinitionId}/versions", args)

CreateSubscriptionDefinitionVersion Operation

Creates a version of a subscription definition which has already been defined.

Arguments

X-Amzn-Client-Token = ::String

The client token used to request idempotent operations.

SubscriptionDefinitionId = ::String – Required

subscription definition Id

Subscriptions = [[ ... ], ...]

Subscriptions in the version.

 Subscriptions = [[
        "Id" =>  ::String,
        "Source" =>  ::String,
        "Subject" =>  ::String,
        "Target" =>  ::String
    ], ...]

Returns

CreateSubscriptionDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.delete_core_definition — Function.

using AmazonGreengrass
delete_core_definition([::AWSConfig], arguments::Dict)
delete_core_definition([::AWSConfig]; CoreDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "DELETE", "/greengrass/definition/cores/{CoreDefinitionId}", args)

DeleteCoreDefinition Operation

Deletes a core definition. The core definition must not have been used in a deployment.

Arguments

CoreDefinitionId = ::String – Required

core definition Id

Returns

DeleteCoreDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.delete_device_definition — Function.

using AmazonGreengrass
delete_device_definition([::AWSConfig], arguments::Dict)
delete_device_definition([::AWSConfig]; DeviceDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "DELETE", "/greengrass/definition/devices/{DeviceDefinitionId}", args)

DeleteDeviceDefinition Operation

Deletes a device definition. The device definition must not have been used in a deployment.

Arguments

DeviceDefinitionId = ::String – Required

device definition Id

Returns

DeleteDeviceDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.delete_function_definition — Function.

using AmazonGreengrass
delete_function_definition([::AWSConfig], arguments::Dict)
delete_function_definition([::AWSConfig]; FunctionDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "DELETE", "/greengrass/definition/functions/{FunctionDefinitionId}", args)

DeleteFunctionDefinition Operation

Deletes a Lambda function definition. The Lambda function definition must not have been used in a deployment.

Arguments

FunctionDefinitionId = ::String – Required

the unique Id of the lambda definition

Returns

DeleteFunctionDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.delete_group — Function.

using AmazonGreengrass
delete_group([::AWSConfig], arguments::Dict)
delete_group([::AWSConfig]; GroupId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "DELETE", "/greengrass/groups/{GroupId}", args)

DeleteGroup Operation

Deletes a group. The group must not have been used in deployment.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

DeleteGroupResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.delete_logger_definition — Function.

using AmazonGreengrass
delete_logger_definition([::AWSConfig], arguments::Dict)
delete_logger_definition([::AWSConfig]; LoggerDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "DELETE", "/greengrass/definition/loggers/{LoggerDefinitionId}", args)

DeleteLoggerDefinition Operation

Deletes a logger definition. The logger definition must not have been used in a deployment.

Arguments

LoggerDefinitionId = ::String – Required

logger definition Id

Returns

DeleteLoggerDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.delete_subscription_definition — Function.

using AmazonGreengrass
delete_subscription_definition([::AWSConfig], arguments::Dict)
delete_subscription_definition([::AWSConfig]; SubscriptionDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "DELETE", "/greengrass/definition/subscriptions/{SubscriptionDefinitionId}", args)

DeleteSubscriptionDefinition Operation

Deletes a subscription definition. The subscription definition must not have been used in a deployment.

Arguments

SubscriptionDefinitionId = ::String – Required

subscription definition Id

Returns

DeleteSubscriptionDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.disassociate_role_from_group — Function.

using AmazonGreengrass
disassociate_role_from_group([::AWSConfig], arguments::Dict)
disassociate_role_from_group([::AWSConfig]; GroupId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "DELETE", "/greengrass/groups/{GroupId}/role", args)

DisassociateRoleFromGroup Operation

Disassociates the role from a group.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

DisassociateRoleFromGroupResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.disassociate_service_role_from_account — Function.

using AmazonGreengrass
disassociate_service_role_from_account([::AWSConfig], arguments::Dict)
disassociate_service_role_from_account([::AWSConfig]; )

using AWSCore.Services.greengrass
request([::AWSConfig], "DELETE", "/greengrass/servicerole", args)

DisassociateServiceRoleFromAccount Operation

Disassociates the service role from the account. Without a service role, deployments will not work.

Arguments

Returns

DisassociateServiceRoleFromAccountResponse

Exceptions

InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.get_associated_role — Function.

using AmazonGreengrass
get_associated_role([::AWSConfig], arguments::Dict)
get_associated_role([::AWSConfig]; GroupId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}/role", args)

GetAssociatedRole Operation

Retrieves the role associated with a particular group.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

GetAssociatedRoleResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.get_connectivity_info — Function.

using AmazonGreengrass
get_connectivity_info([::AWSConfig], arguments::Dict)
get_connectivity_info([::AWSConfig]; ThingName=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/things/{ThingName}/connectivityInfo", args)

GetConnectivityInfo Operation

Retrieves the connectivity information for a core.

Arguments

ThingName = ::String – Required

Thing Name

Returns

GetConnectivityInfoResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.get_core_definition — Function.

using AmazonGreengrass
get_core_definition([::AWSConfig], arguments::Dict)
get_core_definition([::AWSConfig]; CoreDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/cores/{CoreDefinitionId}", args)

GetCoreDefinition Operation

Retrieves information about a core definition version.

Arguments

CoreDefinitionId = ::String – Required

core definition Id

Returns

GetCoreDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_core_definition_version — Function.

using AmazonGreengrass
get_core_definition_version([::AWSConfig], arguments::Dict)
get_core_definition_version([::AWSConfig]; CoreDefinitionId=, CoreDefinitionVersionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/cores/{CoreDefinitionId}/versions/{CoreDefinitionVersionId}", args)

GetCoreDefinitionVersion Operation

Retrieves information about a core definition version.

Arguments

CoreDefinitionId = ::String – Required

core definition Id

CoreDefinitionVersionId = ::String – Required

core definition version Id

Returns

GetCoreDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_deployment_status — Function.

using AmazonGreengrass
get_deployment_status([::AWSConfig], arguments::Dict)
get_deployment_status([::AWSConfig]; DeploymentId=, GroupId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}/deployments/{DeploymentId}/status", args)

GetDeploymentStatus Operation

Returns the status of a deployment.

Arguments

DeploymentId = ::String – Required

the deployment Id

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

GetDeploymentStatusResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_device_definition — Function.

using AmazonGreengrass
get_device_definition([::AWSConfig], arguments::Dict)
get_device_definition([::AWSConfig]; DeviceDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/devices/{DeviceDefinitionId}", args)

GetDeviceDefinition Operation

Retrieves information about a device definition.

Arguments

DeviceDefinitionId = ::String – Required

device definition Id

Returns

GetDeviceDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_device_definition_version — Function.

using AmazonGreengrass
get_device_definition_version([::AWSConfig], arguments::Dict)
get_device_definition_version([::AWSConfig]; DeviceDefinitionId=, DeviceDefinitionVersionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/devices/{DeviceDefinitionId}/versions/{DeviceDefinitionVersionId}", args)

GetDeviceDefinitionVersion Operation

Retrieves information about a device definition version.

Arguments

DeviceDefinitionId = ::String – Required

device definition Id

DeviceDefinitionVersionId = ::String – Required

device definition version Id

Returns

GetDeviceDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_function_definition — Function.

using AmazonGreengrass
get_function_definition([::AWSConfig], arguments::Dict)
get_function_definition([::AWSConfig]; FunctionDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/functions/{FunctionDefinitionId}", args)

GetFunctionDefinition Operation

Retrieves information about a Lambda function definition, such as its creation time and latest version.

Arguments

FunctionDefinitionId = ::String – Required

the unique Id of the lambda definition

Returns

GetFunctionDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_function_definition_version — Function.

using AmazonGreengrass
get_function_definition_version([::AWSConfig], arguments::Dict)
get_function_definition_version([::AWSConfig]; FunctionDefinitionId=, FunctionDefinitionVersionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/functions/{FunctionDefinitionId}/versions/{FunctionDefinitionVersionId}", args)

GetFunctionDefinitionVersion Operation

Retrieves information about a Lambda function definition version, such as which Lambda functions are included in the version and their configurations.

Arguments

FunctionDefinitionId = ::String – Required

the unique Id of the lambda definition

FunctionDefinitionVersionId = ::String – Required

Function definition version Id

Returns

GetFunctionDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_group — Function.

using AmazonGreengrass
get_group([::AWSConfig], arguments::Dict)
get_group([::AWSConfig]; GroupId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}", args)

GetGroup Operation

Retrieves information about a group.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

GetGroupResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_group_certificate_authority — Function.

using AmazonGreengrass
get_group_certificate_authority([::AWSConfig], arguments::Dict)
get_group_certificate_authority([::AWSConfig]; CertificateAuthorityId=, GroupId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}/certificateauthorities/{CertificateAuthorityId}", args)

GetGroupCertificateAuthority Operation

Retreives the CA associated with a group. Returns the public key of the CA.

Arguments

CertificateAuthorityId = ::String – Required

certificate authority Id

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

GetGroupCertificateAuthorityResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.get_group_certificate_configuration — Function.

using AmazonGreengrass
get_group_certificate_configuration([::AWSConfig], arguments::Dict)
get_group_certificate_configuration([::AWSConfig]; GroupId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}/certificateauthorities/configuration/expiry", args)

GetGroupCertificateConfiguration Operation

Retrieves the current configuration for the CA used by the group.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

GetGroupCertificateConfigurationResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.get_group_version — Function.

using AmazonGreengrass
get_group_version([::AWSConfig], arguments::Dict)
get_group_version([::AWSConfig]; GroupId=, GroupVersionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}/versions/{GroupVersionId}", args)

GetGroupVersion Operation

Retrieves information about a group version.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

GroupVersionId = ::String – Required

Group version Id

Returns

GetGroupVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_logger_definition — Function.

using AmazonGreengrass
get_logger_definition([::AWSConfig], arguments::Dict)
get_logger_definition([::AWSConfig]; LoggerDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/loggers/{LoggerDefinitionId}", args)

GetLoggerDefinition Operation

Retrieves information about a logger definition.

Arguments

LoggerDefinitionId = ::String – Required

logger definition Id

Returns

GetLoggerDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_logger_definition_version — Function.

using AmazonGreengrass
get_logger_definition_version([::AWSConfig], arguments::Dict)
get_logger_definition_version([::AWSConfig]; LoggerDefinitionId=, LoggerDefinitionVersionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/loggers/{LoggerDefinitionId}/versions/{LoggerDefinitionVersionId}", args)

GetLoggerDefinitionVersion Operation

Retrieves information about a logger definition version.

Arguments

LoggerDefinitionId = ::String – Required

logger definition Id

LoggerDefinitionVersionId = ::String – Required

logger definition version Id

Returns

GetLoggerDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_service_role_for_account — Function.

using AmazonGreengrass
get_service_role_for_account([::AWSConfig], arguments::Dict)
get_service_role_for_account([::AWSConfig]; )

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/servicerole", args)

GetServiceRoleForAccount Operation

Retrieves the service role that is attached to the account.

Arguments

Returns

GetServiceRoleForAccountResponse

Exceptions

InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.get_subscription_definition — Function.

using AmazonGreengrass
get_subscription_definition([::AWSConfig], arguments::Dict)
get_subscription_definition([::AWSConfig]; SubscriptionDefinitionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/subscriptions/{SubscriptionDefinitionId}", args)

GetSubscriptionDefinition Operation

Retrieves information about a subscription definition.

Arguments

SubscriptionDefinitionId = ::String – Required

subscription definition Id

Returns

GetSubscriptionDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.get_subscription_definition_version — Function.

using AmazonGreengrass
get_subscription_definition_version([::AWSConfig], arguments::Dict)
get_subscription_definition_version([::AWSConfig]; SubscriptionDefinitionId=, SubscriptionDefinitionVersionId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/subscriptions/{SubscriptionDefinitionId}/versions/{SubscriptionDefinitionVersionId}", args)

GetSubscriptionDefinitionVersion Operation

Retrieves information about a subscription definition version.

Arguments

SubscriptionDefinitionId = ::String – Required

subscription definition Id

SubscriptionDefinitionVersionId = ::String – Required

subscription definition version Id

Returns

GetSubscriptionDefinitionVersionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.list_core_definition_versions — Function.

using AmazonGreengrass
list_core_definition_versions([::AWSConfig], arguments::Dict)
list_core_definition_versions([::AWSConfig]; CoreDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/cores/{CoreDefinitionId}/versions", args)

ListCoreDefinitionVersions Operation

Lists versions of a core definition.

Arguments

CoreDefinitionId = ::String – Required

core definition Id

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListCoreDefinitionVersionsResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.list_core_definitions — Function.

using AmazonGreengrass
list_core_definitions([::AWSConfig], arguments::Dict)
list_core_definitions([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/cores", args)

ListCoreDefinitions Operation

Retrieves a list of core definitions.

Arguments

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListCoreDefinitionsResponse

Exceptions

.

See also: AWS API Documentation

AmazonGreengrass.list_deployments — Function.

using AmazonGreengrass
list_deployments([::AWSConfig], arguments::Dict)
list_deployments([::AWSConfig]; GroupId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}/deployments", args)

ListDeployments Operation

Returns a history of deployments for the group.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListDeploymentsResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.list_device_definition_versions — Function.

using AmazonGreengrass
list_device_definition_versions([::AWSConfig], arguments::Dict)
list_device_definition_versions([::AWSConfig]; DeviceDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/devices/{DeviceDefinitionId}/versions", args)

ListDeviceDefinitionVersions Operation

Lists the versions of a device definition.

Arguments

DeviceDefinitionId = ::String – Required

device definition Id

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListDeviceDefinitionVersionsResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.list_device_definitions — Function.

using AmazonGreengrass
list_device_definitions([::AWSConfig], arguments::Dict)
list_device_definitions([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/devices", args)

ListDeviceDefinitions Operation

Retrieves a list of device definitions.

Arguments

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListDeviceDefinitionsResponse

Exceptions

.

See also: AWS API Documentation

AmazonGreengrass.list_function_definition_versions — Function.

using AmazonGreengrass
list_function_definition_versions([::AWSConfig], arguments::Dict)
list_function_definition_versions([::AWSConfig]; FunctionDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/functions/{FunctionDefinitionId}/versions", args)

ListFunctionDefinitionVersions Operation

Lists the versions of a Lambda function definition.

Arguments

FunctionDefinitionId = ::String – Required

the unique Id of the lambda definition

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListFunctionDefinitionVersionsResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.list_function_definitions — Function.

using AmazonGreengrass
list_function_definitions([::AWSConfig], arguments::Dict)
list_function_definitions([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/functions", args)

ListFunctionDefinitions Operation

Retrieves a list of Lambda function definitions.

Arguments

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListFunctionDefinitionsResponse

Exceptions

.

See also: AWS API Documentation

AmazonGreengrass.list_group_certificate_authorities — Function.

using AmazonGreengrass
list_group_certificate_authorities([::AWSConfig], arguments::Dict)
list_group_certificate_authorities([::AWSConfig]; GroupId=)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}/certificateauthorities", args)

ListGroupCertificateAuthorities Operation

Retrieves the current CAs for a group.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

ListGroupCertificateAuthoritiesResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.list_group_versions — Function.

using AmazonGreengrass
list_group_versions([::AWSConfig], arguments::Dict)
list_group_versions([::AWSConfig]; GroupId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups/{GroupId}/versions", args)

ListGroupVersions Operation

List the versions of a group.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListGroupVersionsResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.list_groups — Function.

using AmazonGreengrass
list_groups([::AWSConfig], arguments::Dict)
list_groups([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/groups", args)

ListGroups Operation

Retrieves a list of groups.

Arguments

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListGroupsResponse

Exceptions

.

See also: AWS API Documentation

AmazonGreengrass.list_logger_definition_versions — Function.

using AmazonGreengrass
list_logger_definition_versions([::AWSConfig], arguments::Dict)
list_logger_definition_versions([::AWSConfig]; LoggerDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/loggers/{LoggerDefinitionId}/versions", args)

ListLoggerDefinitionVersions Operation

Lists the versions of a logger definition.

Arguments

LoggerDefinitionId = ::String – Required

logger definition Id

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListLoggerDefinitionVersionsResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.list_logger_definitions — Function.

using AmazonGreengrass
list_logger_definitions([::AWSConfig], arguments::Dict)
list_logger_definitions([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/loggers", args)

ListLoggerDefinitions Operation

Retrieves a list of logger definitions.

Arguments

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListLoggerDefinitionsResponse

Exceptions

.

See also: AWS API Documentation

AmazonGreengrass.list_subscription_definition_versions — Function.

using AmazonGreengrass
list_subscription_definition_versions([::AWSConfig], arguments::Dict)
list_subscription_definition_versions([::AWSConfig]; SubscriptionDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/subscriptions/{SubscriptionDefinitionId}/versions", args)

ListSubscriptionDefinitionVersions Operation

Lists the versions of a subscription definition.

Arguments

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

SubscriptionDefinitionId = ::String – Required

subscription definition Id

Returns

ListSubscriptionDefinitionVersionsResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.list_subscription_definitions — Function.

using AmazonGreengrass
list_subscription_definitions([::AWSConfig], arguments::Dict)
list_subscription_definitions([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "GET", "/greengrass/definition/subscriptions", args)

ListSubscriptionDefinitions Operation

Retrieves a list of subscription definitions.

Arguments

MaxResults = ::String

Specifies the maximum number of list results to be returned in this page

NextToken = ::String

Specifies the pagination token used when iterating through a paginated request

Returns

ListSubscriptionDefinitionsResponse

Exceptions

.

See also: AWS API Documentation

AmazonGreengrass.update_connectivity_info — Function.

using AmazonGreengrass
update_connectivity_info([::AWSConfig], arguments::Dict)
update_connectivity_info([::AWSConfig]; ThingName=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/things/{ThingName}/connectivityInfo", args)

UpdateConnectivityInfo Operation

Updates the connectivity information for the core. Any devices that belong to the group which has this core will receive this information in order to find the location of the core and connect to it.

Arguments

ConnectivityInfo = [[ ... ], ...]

Connectivity info array

 ConnectivityInfo = [[
        "HostAddress" =>  ::String,
        "Id" =>  ::String,
        "Metadata" =>  ::String,
        "PortNumber" =>  ::Int
    ], ...]

ThingName = ::String – Required

Thing Name

Returns

UpdateConnectivityInfoResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.update_core_definition — Function.

using AmazonGreengrass
update_core_definition([::AWSConfig], arguments::Dict)
update_core_definition([::AWSConfig]; CoreDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/definition/cores/{CoreDefinitionId}", args)

UpdateCoreDefinition Operation

Updates a core definition.

Arguments

CoreDefinitionId = ::String – Required

core definition Id

Name = ::String

name of the definition

Returns

UpdateCoreDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.update_device_definition — Function.

using AmazonGreengrass
update_device_definition([::AWSConfig], arguments::Dict)
update_device_definition([::AWSConfig]; DeviceDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/definition/devices/{DeviceDefinitionId}", args)

UpdateDeviceDefinition Operation

Updates a device definition.

Arguments

DeviceDefinitionId = ::String – Required

device definition Id

Name = ::String

name of the definition

Returns

UpdateDeviceDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.update_function_definition — Function.

using AmazonGreengrass
update_function_definition([::AWSConfig], arguments::Dict)
update_function_definition([::AWSConfig]; FunctionDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/definition/functions/{FunctionDefinitionId}", args)

UpdateFunctionDefinition Operation

Updates a Lambda function definition.

Arguments

FunctionDefinitionId = ::String – Required

the unique Id of the lambda definition

Name = ::String

name of the definition

Returns

UpdateFunctionDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.update_group — Function.

using AmazonGreengrass
update_group([::AWSConfig], arguments::Dict)
update_group([::AWSConfig]; GroupId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/groups/{GroupId}", args)

UpdateGroup Operation

Updates a group.

Arguments

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Name = ::String

name of the definition

Returns

UpdateGroupResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.update_group_certificate_configuration — Function.

using AmazonGreengrass
update_group_certificate_configuration([::AWSConfig], arguments::Dict)
update_group_certificate_configuration([::AWSConfig]; GroupId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/groups/{GroupId}/certificateauthorities/configuration/expiry", args)

UpdateGroupCertificateConfiguration Operation

Updates the Cert expiry time for a group.

Arguments

CertificateExpiryInMilliseconds = ::String

Amount of time when the certificate expires in milliseconds.

GroupId = ::String – Required

The unique Id of the AWS Greengrass Group

Returns

UpdateGroupCertificateConfigurationResponse

Exceptions

BadRequestException or InternalServerErrorException.

See also: AWS API Documentation

AmazonGreengrass.update_logger_definition — Function.

using AmazonGreengrass
update_logger_definition([::AWSConfig], arguments::Dict)
update_logger_definition([::AWSConfig]; LoggerDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/definition/loggers/{LoggerDefinitionId}", args)

UpdateLoggerDefinition Operation

Updates a logger definition.

Arguments

LoggerDefinitionId = ::String – Required

logger definition Id

Name = ::String

name of the definition

Returns

UpdateLoggerDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation

AmazonGreengrass.update_subscription_definition — Function.

using AmazonGreengrass
update_subscription_definition([::AWSConfig], arguments::Dict)
update_subscription_definition([::AWSConfig]; SubscriptionDefinitionId=, <keyword arguments>)

using AWSCore.Services.greengrass
request([::AWSConfig], "PUT", "/greengrass/definition/subscriptions/{SubscriptionDefinitionId}", args)

UpdateSubscriptionDefinition Operation

Updates a subscription definition.

Arguments

Name = ::String

name of the definition

SubscriptionDefinitionId = ::String – Required

subscription definition Id

Returns

UpdateSubscriptionDefinitionResponse

Exceptions

BadRequestException.

See also: AWS API Documentation