AmazonHealth.jl

AmazonHealth

AWS Health

The AWS Health API provides programmatic access to the AWS Health information that is presented in the AWS Personal Health Dashboard. You can get information about events that affect your AWS resources:

DescribeEvents: Summary information about events.

DescribeEventDetails: Detailed information about one or more events.

DescribeAffectedEntities: Information about AWS resources that are affected by one or more events.

In addition, these operations provide information about event types and summary counts of events or affected entities:

DescribeEventTypes: Information about the kinds of events that AWS Health tracks.

DescribeEventAggregates: A count of the number of events that meet specified criteria.

DescribeEntityAggregates: A count of the number of affected entities that meet specified criteria.

The Health API requires a Business or Enterprise support plan from AWS Support. Calling the Health API from an account that does not have a Business or Enterprise support plan causes a SubscriptionRequiredException.

For authentication of requests, AWS Health uses the Signature Version 4 Signing Process.

See the AWS Health User Guide for information about how to use the API.

Service Endpoint

The HTTP endpoint for the AWS Health API is:

https://health.us-east-1.amazonaws.com

This document is generated from apis/health-2016-08-04.normal.json. See JuliaCloud/AWSCore.jl.

AmazonHealth.describe_affected_entities

AmazonHealth.describe_entity_aggregates

AmazonHealth.describe_event_aggregates

AmazonHealth.describe_event_details

AmazonHealth.describe_event_types

AmazonHealth.describe_events

AmazonHealth.describe_affected_entities — Function.

using AmazonHealth
describe_affected_entities([::AWSConfig], arguments::Dict)
describe_affected_entities([::AWSConfig]; filter=, <keyword arguments>)

using AWSCore.Services.health
request([::AWSConfig],"DescribeAffectedEntities", args)

DescribeAffectedEntities Operation

Returns a list of entities that have been affected by the specified events, based on the specified filter criteria. Entities can refer to individual customer resources, groups of customer resources, or any other construct, depending on the AWS service. Events that have impact beyond that of the affected entities, or where the extent of impact is unknown, include at least one entity indicating this.

At least one event ARN is required. Results are sorted by the lastUpdatedTime of the entity, starting with the most recent.

Arguments

filter = [ ... ] – Required

Values to narrow the results returned. At least one event ARN is required.

 filter = [
        "eventArns" => <required> [::String, ...],
        "entityArns" =>  [::String, ...],
        "entityValues" =>  [::String, ...],
        "lastUpdatedTimes" =>  [[
            "from" =>  timestamp,
            "to" =>  timestamp
        ], ...],
        "tags" =>  [::Dict{String,String}, ...],
        "statusCodes" =>  ["IMPAIRED", "UNIMPAIRED" or "UNKNOWN", ...]
    ]

locale = ::String

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

nextToken = ::String

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

maxResults = ::Int

The maximum number of items to return in one batch, between 10 and 100, inclusive.

Returns

DescribeAffectedEntitiesResponse

Exceptions

InvalidPaginationToken or UnsupportedLocale.

See also: AWS API Documentation

AmazonHealth.describe_entity_aggregates — Function.

using AmazonHealth
describe_entity_aggregates([::AWSConfig], arguments::Dict)
describe_entity_aggregates([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.health
request([::AWSConfig],"DescribeEntityAggregates", args)

DescribeEntityAggregates Operation

Returns the number of entities that are affected by each of the specified events. If no events are specified, the counts of all affected entities are returned.

Arguments

eventArns = [::String, ...]

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::event/AWS_EBS_LOST_VOLUME_xyz"

Returns

DescribeEntityAggregatesResponse

See also: AWS API Documentation

AmazonHealth.describe_event_aggregates — Function.

using AmazonHealth
describe_event_aggregates([::AWSConfig], arguments::Dict)
describe_event_aggregates([::AWSConfig]; aggregateField=, <keyword arguments>)

using AWSCore.Services.health
request([::AWSConfig],"DescribeEventAggregates", args)

DescribeEventAggregates Operation

Returns the number of events of each event type (issue, scheduled change, and account notification). If no filter is specified, the counts of all events in each category are returned.

Arguments

filter = [ ... ]

Values to narrow the results returned.

 filter = [
        "eventArns" =>  [::String, ...],
        "eventTypeCodes" =>  [::String, ...],
        "services" =>  [::String, ...],
        "regions" =>  [::String, ...],
        "availabilityZones" =>  [::String, ...],
        "startTimes" =>  [[
            "from" =>  timestamp,
            "to" =>  timestamp
        ], ...],
        "endTimes" =>  [[
            "from" =>  timestamp,
            "to" =>  timestamp
        ], ...],
        "lastUpdatedTimes" =>  [[
            "from" =>  timestamp,
            "to" =>  timestamp
        ], ...],
        "entityArns" =>  [::String, ...],
        "entityValues" =>  [::String, ...],
        "eventTypeCategories" =>  ["issue", "accountNotification" or "scheduledChange", ...],
        "tags" =>  [::Dict{String,String}, ...],
        "eventStatusCodes" =>  ["open", "closed" or "upcoming", ...]
    ]

aggregateField = "eventTypeCategory" – Required

The only currently supported value is eventTypeCategory.

maxResults = ::Int

The maximum number of items to return in one batch, between 10 and 100, inclusive.

nextToken = ::String

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

Returns

DescribeEventAggregatesResponse

Exceptions

InvalidPaginationToken.

See also: AWS API Documentation

AmazonHealth.describe_event_details — Function.

using AmazonHealth
describe_event_details([::AWSConfig], arguments::Dict)
describe_event_details([::AWSConfig]; eventArns=, <keyword arguments>)

using AWSCore.Services.health
request([::AWSConfig],"DescribeEventDetails", args)

DescribeEventDetails Operation

Returns detailed information about one or more specified events. Information includes standard event data (region, service, etc., as returned by DescribeEvents), a detailed event description, and possible additional metadata that depends upon the nature of the event. Affected entities are not included; to retrieve those, use the DescribeAffectedEntities operation.

If a specified event cannot be retrieved, an error message is returned for that event.

Arguments

eventArns = [::String, ...] – Required

A list of event ARNs (unique identifiers). For example: "arn:aws:health:us-east-1::event/AWS_EC2_MAINTENANCE_5331", "arn:aws:health:us-west-1::event/AWS_EBS_LOST_VOLUME_xyz"

locale = ::String

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

Returns

DescribeEventDetailsResponse

Exceptions

UnsupportedLocale.

See also: AWS API Documentation

AmazonHealth.describe_event_types — Function.

using AmazonHealth
describe_event_types([::AWSConfig], arguments::Dict)
describe_event_types([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.health
request([::AWSConfig],"DescribeEventTypes", args)

DescribeEventTypes Operation

Returns the event types that meet the specified filter criteria. If no filter criteria are specified, all event types are returned, in no particular order.

Arguments

filter = [ ... ]

Values to narrow the results returned.

 filter = [
        "eventTypeCodes" =>  [::String, ...],
        "services" =>  [::String, ...],
        "eventTypeCategories" =>  ["issue", "accountNotification" or "scheduledChange", ...]
    ]

locale = ::String

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

nextToken = ::String

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

maxResults = ::Int

The maximum number of items to return in one batch, between 10 and 100, inclusive.

Returns

DescribeEventTypesResponse

Exceptions

InvalidPaginationToken or UnsupportedLocale.

See also: AWS API Documentation

AmazonHealth.describe_events — Function.

using AmazonHealth
describe_events([::AWSConfig], arguments::Dict)
describe_events([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.health
request([::AWSConfig],"DescribeEvents", args)

DescribeEvents Operation

Returns information about events that meet the specified filter criteria. Events are returned in a summary form and do not include the detailed description, any additional metadata that depends on the event type, or any affected resources. To retrieve that information, use the DescribeEventDetails and DescribeAffectedEntities operations.

If no filter criteria are specified, all events are returned. Results are sorted by lastModifiedTime, starting with the most recent.

Arguments

filter = [ ... ]

Values to narrow the results returned.

 filter = [
        "eventArns" =>  [::String, ...],
        "eventTypeCodes" =>  [::String, ...],
        "services" =>  [::String, ...],
        "regions" =>  [::String, ...],
        "availabilityZones" =>  [::String, ...],
        "startTimes" =>  [[
            "from" =>  timestamp,
            "to" =>  timestamp
        ], ...],
        "endTimes" =>  [[
            "from" =>  timestamp,
            "to" =>  timestamp
        ], ...],
        "lastUpdatedTimes" =>  [[
            "from" =>  timestamp,
            "to" =>  timestamp
        ], ...],
        "entityArns" =>  [::String, ...],
        "entityValues" =>  [::String, ...],
        "eventTypeCategories" =>  ["issue", "accountNotification" or "scheduledChange", ...],
        "tags" =>  [::Dict{String,String}, ...],
        "eventStatusCodes" =>  ["open", "closed" or "upcoming", ...]
    ]

nextToken = ::String

If the results of a search are large, only a portion of the results are returned, and a nextToken pagination token is returned in the response. To retrieve the next batch of results, reissue the search request and include the returned token. When all results have been returned, the response does not contain a pagination token value.

maxResults = ::Int

The maximum number of items to return in one batch, between 10 and 100, inclusive.

locale = ::String

The locale (language) to return information in. English (en) is the default and the only supported value at this time.

Returns

DescribeEventsResponse

Exceptions

InvalidPaginationToken or UnsupportedLocale.

See also: AWS API Documentation