AmazonInspector.jl

AmazonInspector

Amazon Inspector

Amazon Inspector enables you to analyze the behavior of your AWS resources and to identify potential security issues. For more information, see Amazon Inspector User Guide.

This document is generated from apis/inspector-2016-02-16.normal.json. See JuliaCloud/AWSCore.jl.

AmazonInspector.add_attributes_to_findings

AmazonInspector.create_assessment_target

AmazonInspector.create_assessment_template

AmazonInspector.create_resource_group

AmazonInspector.delete_assessment_run

AmazonInspector.delete_assessment_target

AmazonInspector.delete_assessment_template

AmazonInspector.describe_assessment_runs

AmazonInspector.describe_assessment_targets

AmazonInspector.describe_assessment_templates

AmazonInspector.describe_cross_account_access_role

AmazonInspector.describe_findings

AmazonInspector.describe_resource_groups

AmazonInspector.describe_rules_packages

AmazonInspector.get_assessment_report

AmazonInspector.get_telemetry_metadata

AmazonInspector.list_assessment_run_agents

AmazonInspector.list_assessment_runs

AmazonInspector.list_assessment_targets

AmazonInspector.list_assessment_templates

AmazonInspector.list_event_subscriptions

AmazonInspector.list_findings

AmazonInspector.list_rules_packages

AmazonInspector.list_tags_for_resource

AmazonInspector.preview_agents

AmazonInspector.register_cross_account_access_role

AmazonInspector.remove_attributes_from_findings

AmazonInspector.set_tags_for_resource

AmazonInspector.start_assessment_run

AmazonInspector.stop_assessment_run

AmazonInspector.subscribe_to_event

AmazonInspector.unsubscribe_from_event

AmazonInspector.update_assessment_target

AmazonInspector.add_attributes_to_findings — Function.

add_attributes_to_findings(::AWSConfig, arguments::Dict)
add_attributes_to_findings([::AWSConfig]; findingArns=, attributes=)

AddAttributesToFindings Operation

Assigns attributes (key and value pairs) to the findings that are specified by the ARNs of the findings.

Arguments

findingArns = [::String...] – Required

The ARNs that specify the findings that you want to assign attributes to.

attributes = [Dict( ... )...] – Required

The array of attributes that you want to assign to specified findings.

 attributes = [Dict(
        "key" => <required> ::String,
        "value" =>  ::String
    )...]

Returns

AddAttributesToFindingsResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.create_assessment_target — Function.

create_assessment_target(::AWSConfig, arguments::Dict)
create_assessment_target([::AWSConfig]; assessmentTargetName=, resourceGroupArn=)

CreateAssessmentTarget Operation

Creates a new assessment target using the ARN of the resource group that is generated by CreateResourceGroup. You can create up to 50 assessment targets per AWS account. You can run up to 500 concurrent agents per AWS account. For more information, see Amazon Inspector Assessment Targets.

Arguments

assessmentTargetName = ::String – Required

The user-defined name that identifies the assessment target that you want to create. The name must be unique within the AWS account.

resourceGroupArn = ::String – Required

The ARN that specifies the resource group that is used to create the assessment target.

Returns

CreateAssessmentTargetResponse

Exceptions

InternalException, InvalidInputException, LimitExceededException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.create_assessment_template — Function.

create_assessment_template(::AWSConfig, arguments::Dict)
create_assessment_template([::AWSConfig]; assessmentTargetArn=, assessmentTemplateName=, durationInSeconds=, rulesPackageArns=, <keyword arguments>)

CreateAssessmentTemplate Operation

Creates an assessment template for the assessment target that is specified by the ARN of the assessment target.

Arguments

assessmentTargetArn = ::String – Required

The ARN that specifies the assessment target for which you want to create the assessment template.

assessmentTemplateName = ::String – Required

The user-defined name that identifies the assessment template that you want to create. You can create several assessment templates for an assessment target. The names of the assessment templates that correspond to a particular assessment target must be unique.

durationInSeconds = ::Int – Required

The duration of the assessment run in seconds. The default value is 3600 seconds (one hour).

rulesPackageArns = [::String...] – Required

The ARNs that specify the rules packages that you want to attach to the assessment template.

userAttributesForFindings = [Dict( ... )...]

The user-defined attributes that are assigned to every finding that is generated by the assessment run that uses this assessment template.

 userAttributesForFindings = [Dict(
        "key" => <required> ::String,
        "value" =>  ::String
    )...]

Returns

CreateAssessmentTemplateResponse

Exceptions

InternalException, InvalidInputException, LimitExceededException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.create_resource_group — Function.

create_resource_group(::AWSConfig, arguments::Dict)
create_resource_group([::AWSConfig]; resourceGroupTags=)

CreateResourceGroup Operation

Creates a resource group using the specified set of tags (key and value pairs) that are used to select the EC2 instances to be included in an Amazon Inspector assessment target. The created resource group is then used to create an Amazon Inspector assessment target. For more information, see CreateAssessmentTarget.

Arguments

resourceGroupTags = [Dict( ... )...] – Required

A collection of keys and an array of possible values, '[{"key":"key1","values":["Value1","Value2"]},{"key":"Key2","values":["Value3"]}]'.

For example,'[{"key":"Name","values":["TestEC2Instance"]}]'.

 resourceGroupTags = [Dict(
        "key" => <required> ::String,
        "value" =>  ::String
    )...]

Returns

CreateResourceGroupResponse

Exceptions

InternalException, InvalidInputException, LimitExceededException or AccessDeniedException.

See also: AWS API Documentation

AmazonInspector.delete_assessment_run — Function.

delete_assessment_run(::AWSConfig, arguments::Dict)
delete_assessment_run([::AWSConfig]; assessmentRunArn=)

DeleteAssessmentRun Operation

Deletes the assessment run that is specified by the ARN of the assessment run.

Arguments

assessmentRunArn = ::String – Required

The ARN that specifies the assessment run that you want to delete.

Exceptions

InternalException, InvalidInputException, AssessmentRunInProgressException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.delete_assessment_target — Function.

delete_assessment_target(::AWSConfig, arguments::Dict)
delete_assessment_target([::AWSConfig]; assessmentTargetArn=)

DeleteAssessmentTarget Operation

Deletes the assessment target that is specified by the ARN of the assessment target.

Arguments

assessmentTargetArn = ::String – Required

The ARN that specifies the assessment target that you want to delete.

Exceptions

InternalException, InvalidInputException, AssessmentRunInProgressException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.delete_assessment_template — Function.

delete_assessment_template(::AWSConfig, arguments::Dict)
delete_assessment_template([::AWSConfig]; assessmentTemplateArn=)

DeleteAssessmentTemplate Operation

Deletes the assessment template that is specified by the ARN of the assessment template.

Arguments

assessmentTemplateArn = ::String – Required

The ARN that specifies the assessment template that you want to delete.

Exceptions

InternalException, InvalidInputException, AssessmentRunInProgressException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.describe_assessment_runs — Function.

describe_assessment_runs(::AWSConfig, arguments::Dict)
describe_assessment_runs([::AWSConfig]; assessmentRunArns=)

DescribeAssessmentRuns Operation

Describes the assessment runs that are specified by the ARNs of the assessment runs.

Arguments

assessmentRunArns = [::String...] – Required

The ARN that specifies the assessment run that you want to describe.

Returns

DescribeAssessmentRunsResponse

Exceptions

InternalException or InvalidInputException.

See also: AWS API Documentation

AmazonInspector.describe_assessment_targets — Function.

describe_assessment_targets(::AWSConfig, arguments::Dict)
describe_assessment_targets([::AWSConfig]; assessmentTargetArns=)

DescribeAssessmentTargets Operation

Describes the assessment targets that are specified by the ARNs of the assessment targets.

Arguments

assessmentTargetArns = [::String...] – Required

The ARNs that specifies the assessment targets that you want to describe.

Returns

DescribeAssessmentTargetsResponse

Exceptions

InternalException or InvalidInputException.

See also: AWS API Documentation

AmazonInspector.describe_assessment_templates — Function.

describe_assessment_templates(::AWSConfig, arguments::Dict)
describe_assessment_templates([::AWSConfig]; assessmentTemplateArns=)

DescribeAssessmentTemplates Operation

Describes the assessment templates that are specified by the ARNs of the assessment templates.

Arguments

assessmentTemplateArns = [::String...] – Required

Returns

DescribeAssessmentTemplatesResponse

Exceptions

InternalException or InvalidInputException.

See also: AWS API Documentation

AmazonInspector.describe_cross_account_access_role — Function.

describe_cross_account_access_role([::AWSConfig])

DescribeCrossAccountAccessRole Operation

Describes the IAM role that enables Amazon Inspector to access your AWS account.

Returns

DescribeCrossAccountAccessRoleResponse

Exceptions

InternalException.

See also: AWS API Documentation

AmazonInspector.describe_findings — Function.

describe_findings(::AWSConfig, arguments::Dict)
describe_findings([::AWSConfig]; findingArns=, <keyword arguments>)

DescribeFindings Operation

Describes the findings that are specified by the ARNs of the findings.

Arguments

findingArns = [::String...] – Required

The ARN that specifies the finding that you want to describe.

locale = "EN_US"

The locale into which you want to translate a finding description, recommendation, and the short description that identifies the finding.

Returns

DescribeFindingsResponse

Exceptions

InternalException or InvalidInputException.

See also: AWS API Documentation

AmazonInspector.describe_resource_groups — Function.

describe_resource_groups(::AWSConfig, arguments::Dict)
describe_resource_groups([::AWSConfig]; resourceGroupArns=)

DescribeResourceGroups Operation

Describes the resource groups that are specified by the ARNs of the resource groups.

Arguments

resourceGroupArns = [::String...] – Required

The ARN that specifies the resource group that you want to describe.

Returns

DescribeResourceGroupsResponse

Exceptions

InternalException or InvalidInputException.

See also: AWS API Documentation

AmazonInspector.describe_rules_packages — Function.

describe_rules_packages(::AWSConfig, arguments::Dict)
describe_rules_packages([::AWSConfig]; rulesPackageArns=, <keyword arguments>)

DescribeRulesPackages Operation

Describes the rules packages that are specified by the ARNs of the rules packages.

Arguments

rulesPackageArns = [::String...] – Required

The ARN that specifies the rules package that you want to describe.

locale = "EN_US"

The locale that you want to translate a rules package description into.

Returns

DescribeRulesPackagesResponse

Exceptions

InternalException or InvalidInputException.

See also: AWS API Documentation

AmazonInspector.get_assessment_report — Function.

get_assessment_report(::AWSConfig, arguments::Dict)
get_assessment_report([::AWSConfig]; assessmentRunArn=, reportFileFormat=, reportType=)

GetAssessmentReport Operation

Produces an assessment report that includes detailed and comprehensive results of a specified assessment run.

Arguments

assessmentRunArn = ::String – Required

The ARN that specifies the assessment run for which you want to generate a report.

reportFileFormat = "HTML" or "PDF" – Required

Specifies the file format (html or pdf) of the assessment report that you want to generate.

reportType = "FINDING" or "FULL" – Required

Specifies the type of the assessment report that you want to generate. There are two types of assessment reports: a finding report and a full report. For more information, see Assessment Reports.

Returns

GetAssessmentReportResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException, NoSuchEntityException, AssessmentRunInProgressException or UnsupportedFeatureException.

See also: AWS API Documentation

AmazonInspector.get_telemetry_metadata — Function.

get_telemetry_metadata(::AWSConfig, arguments::Dict)
get_telemetry_metadata([::AWSConfig]; assessmentRunArn=)

GetTelemetryMetadata Operation

Information about the data that is collected for the specified assessment run.

Arguments

assessmentRunArn = ::String – Required

The ARN that specifies the assessment run that has the telemetry data that you want to obtain.

Returns

GetTelemetryMetadataResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.list_assessment_run_agents — Function.

list_assessment_run_agents(::AWSConfig, arguments::Dict)
list_assessment_run_agents([::AWSConfig]; assessmentRunArn=, <keyword arguments>)

ListAssessmentRunAgents Operation

Lists the agents of the assessment runs that are specified by the ARNs of the assessment runs.

Arguments

assessmentRunArn = ::String – Required

The ARN that specifies the assessment run whose agents you want to list.

filter = Dict( ... )

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

 filter = Dict(
        "agentHealths" => <required> ["HEALTHY" or "UNHEALTHY"...],
        "agentHealthCodes" => <required> ["IDLE", "RUNNING", "SHUTDOWN", "UNHEALTHY", "THROTTLED" or "UNKNOWN"...]
    )

nextToken = ::String

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRunAgents action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults = ::Int

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

Returns

ListAssessmentRunAgentsResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.list_assessment_runs — Function.

list_assessment_runs(::AWSConfig, arguments::Dict)
list_assessment_runs([::AWSConfig]; <keyword arguments>)

ListAssessmentRuns Operation

Lists the assessment runs that correspond to the assessment templates that are specified by the ARNs of the assessment templates.

Arguments

assessmentTemplateArns = [::String...]

The ARNs that specify the assessment templates whose assessment runs you want to list.

filter = Dict( ... )

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

 filter = Dict(
        "namePattern" =>  ::String,
        "states" =>  ["CREATED", "START_DATA_COLLECTION_PENDING", "START_DATA_COLLECTION_IN_PROGRESS", "COLLECTING_DATA", "STOP_DATA_COLLECTION_PENDING", "DATA_COLLECTED", "START_EVALUATING_RULES_PENDING", "EVALUATING_RULES", "FAILED", "ERROR", "COMPLETED" or "COMPLETED_WITH_ERRORS"...],
        "durationRange" =>  Dict(
            "minSeconds" =>  ::Int,
            "maxSeconds" =>  ::Int
        ),
        "rulesPackageArns" =>  [::String...],
        "startTimeRange" =>  Dict(
            "beginDate" =>  timestamp,
            "endDate" =>  timestamp
        ),
        "completionTimeRange" =>  Dict(
            "beginDate" =>  timestamp,
            "endDate" =>  timestamp
        ),
        "stateChangeTimeRange" =>  Dict(
            "beginDate" =>  timestamp,
            "endDate" =>  timestamp
        )
    )

nextToken = ::String

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRuns action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults = ::Int

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

Returns

ListAssessmentRunsResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.list_assessment_targets — Function.

list_assessment_targets(::AWSConfig, arguments::Dict)
list_assessment_targets([::AWSConfig]; <keyword arguments>)

ListAssessmentTargets Operation

Lists the ARNs of the assessment targets within this AWS account. For more information about assessment targets, see Amazon Inspector Assessment Targets.

Arguments

filter = Dict("assessmentTargetNamePattern" => ::String)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

nextToken = ::String

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTargets action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults = ::Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Returns

ListAssessmentTargetsResponse

Exceptions

InternalException, InvalidInputException or AccessDeniedException.

See also: AWS API Documentation

AmazonInspector.list_assessment_templates — Function.

list_assessment_templates(::AWSConfig, arguments::Dict)
list_assessment_templates([::AWSConfig]; <keyword arguments>)

ListAssessmentTemplates Operation

Lists the assessment templates that correspond to the assessment targets that are specified by the ARNs of the assessment targets.

Arguments

assessmentTargetArns = [::String...]

A list of ARNs that specifies the assessment targets whose assessment templates you want to list.

filter = Dict( ... )

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

 filter = Dict(
        "namePattern" =>  ::String,
        "durationRange" =>  Dict(
            "minSeconds" =>  ::Int,
            "maxSeconds" =>  ::Int
        ),
        "rulesPackageArns" =>  [::String...]
    )

nextToken = ::String

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTemplates action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults = ::Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Returns

ListAssessmentTemplatesResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.list_event_subscriptions — Function.

list_event_subscriptions(::AWSConfig, arguments::Dict)
list_event_subscriptions([::AWSConfig]; <keyword arguments>)

ListEventSubscriptions Operation

Lists all the event subscriptions for the assessment template that is specified by the ARN of the assessment template. For more information, see SubscribeToEvent and UnsubscribeFromEvent.

Arguments

resourceArn = ::String

The ARN of the assessment template for which you want to list the existing event subscriptions.

nextToken = ::String

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListEventSubscriptions action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults = ::Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Returns

ListEventSubscriptionsResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.list_findings — Function.

list_findings(::AWSConfig, arguments::Dict)
list_findings([::AWSConfig]; <keyword arguments>)

ListFindings Operation

Lists findings that are generated by the assessment runs that are specified by the ARNs of the assessment runs.

Arguments

assessmentRunArns = [::String...]

The ARNs of the assessment runs that generate the findings that you want to list.

filter = Dict( ... )

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

 filter = Dict(
        "agentIds" =>  [::String...],
        "autoScalingGroups" =>  [::String...],
        "ruleNames" =>  [::String...],
        "severities" =>  ["Low", "Medium", "High", "Informational" or "Undefined"...],
        "rulesPackageArns" =>  [::String...],
        "attributes" =>  [Dict(
            "key" => <required> ::String,
            "value" =>  ::String
        )...],
        "userAttributes" =>  [Dict(
            "key" => <required> ::String,
            "value" =>  ::String
        )...],
        "creationTimeRange" =>  Dict(
            "beginDate" =>  timestamp,
            "endDate" =>  timestamp
        )
    )

nextToken = ::String

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListFindings action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults = ::Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Returns

ListFindingsResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.list_rules_packages — Function.

list_rules_packages(::AWSConfig, arguments::Dict)
list_rules_packages([::AWSConfig]; <keyword arguments>)

ListRulesPackages Operation

Lists all available Amazon Inspector rules packages.

Arguments

nextToken = ::String

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListRulesPackages action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults = ::Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Returns

ListRulesPackagesResponse

Exceptions

InternalException, InvalidInputException or AccessDeniedException.

See also: AWS API Documentation

AmazonInspector.list_tags_for_resource — Function.

list_tags_for_resource(::AWSConfig, arguments::Dict)
list_tags_for_resource([::AWSConfig]; resourceArn=)

ListTagsForResource Operation

Lists all tags associated with an assessment template.

Arguments

resourceArn = ::String – Required

The ARN that specifies the assessment template whose tags you want to list.

Returns

ListTagsForResourceResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.preview_agents — Function.

preview_agents(::AWSConfig, arguments::Dict)
preview_agents([::AWSConfig]; previewAgentsArn=, <keyword arguments>)

PreviewAgents Operation

Previews the agents installed on the EC2 instances that are part of the specified assessment target.

Arguments

previewAgentsArn = ::String – Required

The ARN of the assessment target whose agents you want to preview.

nextToken = ::String

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the PreviewAgents action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults = ::Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Returns

PreviewAgentsResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException, NoSuchEntityException or InvalidCrossAccountRoleException.

See also: AWS API Documentation

AmazonInspector.register_cross_account_access_role — Function.

register_cross_account_access_role(::AWSConfig, arguments::Dict)
register_cross_account_access_role([::AWSConfig]; roleArn=)

RegisterCrossAccountAccessRole Operation

Registers the IAM role that Amazon Inspector uses to list your EC2 instances at the start of the assessment run or when you call the PreviewAgents action.

Arguments

roleArn = ::String – Required

The ARN of the IAM role that Amazon Inspector uses to list your EC2 instances during the assessment run or when you call the PreviewAgents action.

Exceptions

InternalException, InvalidInputException, AccessDeniedException or InvalidCrossAccountRoleException.

See also: AWS API Documentation

AmazonInspector.remove_attributes_from_findings — Function.

remove_attributes_from_findings(::AWSConfig, arguments::Dict)
remove_attributes_from_findings([::AWSConfig]; findingArns=, attributeKeys=)

RemoveAttributesFromFindings Operation

Removes entire attributes (key and value pairs) from the findings that are specified by the ARNs of the findings where an attribute with the specified key exists.

Arguments

findingArns = [::String...] – Required

The ARNs that specify the findings that you want to remove attributes from.

attributeKeys = [::String...] – Required

The array of attribute keys that you want to remove from specified findings.

Returns

RemoveAttributesFromFindingsResponse

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.set_tags_for_resource — Function.

set_tags_for_resource(::AWSConfig, arguments::Dict)
set_tags_for_resource([::AWSConfig]; resourceArn=, <keyword arguments>)

SetTagsForResource Operation

Sets tags (key and value pairs) to the assessment template that is specified by the ARN of the assessment template.

Arguments

resourceArn = ::String – Required

The ARN of the assessment template that you want to set tags to.

tags = [Dict( ... )...]

A collection of key and value pairs that you want to set to the assessment template.

 tags = [Dict(
        "key" => <required> ::String,
        "value" =>  ::String
    )...]

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.start_assessment_run — Function.

start_assessment_run(::AWSConfig, arguments::Dict)
start_assessment_run([::AWSConfig]; assessmentTemplateArn=, <keyword arguments>)

StartAssessmentRun Operation

Starts the assessment run specified by the ARN of the assessment template. For this API to function properly, you must not exceed the limit of running up to 500 concurrent agents per AWS account.

Arguments

assessmentTemplateArn = ::String – Required

The ARN of the assessment template of the assessment run that you want to start.

assessmentRunName = ::String

You can specify the name for the assessment run. The name must be unique for the assessment template whose ARN is used to start the assessment run.

Returns

StartAssessmentRunResponse

Exceptions

InternalException, InvalidInputException, LimitExceededException, AccessDeniedException, NoSuchEntityException, InvalidCrossAccountRoleException or AgentsAlreadyRunningAssessmentException.

See also: AWS API Documentation

AmazonInspector.stop_assessment_run — Function.

stop_assessment_run(::AWSConfig, arguments::Dict)
stop_assessment_run([::AWSConfig]; assessmentRunArn=)

StopAssessmentRun Operation

Stops the assessment run that is specified by the ARN of the assessment run.

Arguments

assessmentRunArn = ::String – Required

The ARN of the assessment run that you want to stop.

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.subscribe_to_event — Function.

subscribe_to_event(::AWSConfig, arguments::Dict)
subscribe_to_event([::AWSConfig]; resourceArn=, event=, topicArn=)

SubscribeToEvent Operation

Enables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

Arguments

resourceArn = ::String – Required

The ARN of the assessment template that is used during the event for which you want to receive SNS notifications.

event = "ASSESSMENT_RUN_STARTED", "ASSESSMENT_RUN_COMPLETED", "ASSESSMENT_RUN_STATE_CHANGED", "FINDING_REPORTED" or "OTHER" – Required

The event for which you want to receive SNS notifications.

topicArn = ::String – Required

The ARN of the SNS topic to which the SNS notifications are sent.

Exceptions

InternalException, InvalidInputException, LimitExceededException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.unsubscribe_from_event — Function.

unsubscribe_from_event(::AWSConfig, arguments::Dict)
unsubscribe_from_event([::AWSConfig]; resourceArn=, event=, topicArn=)

UnsubscribeFromEvent Operation

Disables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

Arguments

resourceArn = ::String – Required

The ARN of the assessment template that is used during the event for which you want to stop receiving SNS notifications.

event = "ASSESSMENT_RUN_STARTED", "ASSESSMENT_RUN_COMPLETED", "ASSESSMENT_RUN_STATE_CHANGED", "FINDING_REPORTED" or "OTHER" – Required

The event for which you want to stop receiving SNS notifications.

topicArn = ::String – Required

The ARN of the SNS topic to which SNS notifications are sent.

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation

AmazonInspector.update_assessment_target — Function.

update_assessment_target(::AWSConfig, arguments::Dict)
update_assessment_target([::AWSConfig]; assessmentTargetArn=, assessmentTargetName=, resourceGroupArn=)

UpdateAssessmentTarget Operation

Updates the assessment target that is specified by the ARN of the assessment target.

Arguments

assessmentTargetArn = ::String – Required

The ARN of the assessment target that you want to update.

assessmentTargetName = ::String – Required

The name of the assessment target that you want to update.

resourceGroupArn = ::String – Required

The ARN of the resource group that is used to specify the new resource group to associate with the assessment target.

Exceptions

InternalException, InvalidInputException, AccessDeniedException or NoSuchEntityException.

See also: AWS API Documentation