AmazonKMS.jl

AmazonKMS

AWS Key Management Service

AWS Key Management Service (AWS KMS) is an encryption and key management web service. This guide describes the AWS KMS operations that you can call programmatically. For general information about AWS KMS, see the AWS Key Management Service Developer Guide.

Note

AWS provides SDKs that consist of libraries and sample code for various programming languages and platforms (Java, Ruby, .Net, iOS, Android, etc.). The SDKs provide a convenient way to create programmatic access to AWS KMS and other AWS services. For example, the SDKs take care of tasks such as signing requests (see below), managing errors, and retrying requests automatically. For more information about the AWS SDKs, including how to download and install them, see Tools for Amazon Web Services.

We recommend that you use the AWS SDKs to make programmatic API calls to AWS KMS.

Clients must support TLS (Transport Layer Security) 1.0. We recommend TLS 1.2. Clients must also support cipher suites with Perfect Forward Secrecy (PFS) such as Ephemeral Diffie-Hellman (DHE) or Elliptic Curve Ephemeral Diffie-Hellman (ECDHE). Most modern systems such as Java 7 and later support these modes.

Signing Requests

Requests must be signed by using an access key ID and a secret access key. We strongly recommend that you do not use your AWS account (root) access key ID and secret key for everyday work with AWS KMS. Instead, use the access key ID and secret access key for an IAM user, or you can use the AWS Security Token Service to generate temporary security credentials that you can use to sign requests.

All AWS KMS operations require Signature Version 4.

Logging API Requests

AWS KMS supports AWS CloudTrail, a service that logs AWS API calls and related events for your AWS account and delivers them to an Amazon S3 bucket that you specify. By using the information collected by CloudTrail, you can determine what requests were made to AWS KMS, who made the request, when it was made, and so on. To learn more about CloudTrail, including how to turn it on and find your log files, see the AWS CloudTrail User Guide.

Additional Resources

For more information about credentials and request signing, see the following:

AWS Security Credentials - This topic provides general information about the types of credentials used for accessing AWS.

Temporary Security Credentials - This section of the IAM User Guide describes how to create and use temporary security credentials.

Signature Version 4 Signing Process - This set of topics walks you through the process of signing a request using an access key ID and a secret access key.

Commonly Used APIs

Of the APIs discussed in this guide, the following will prove the most useful for most applications. You will likely perform actions other than these, such as creating keys and assigning policies, by using the console.

Encrypt

Decrypt

GenerateDataKey

GenerateDataKeyWithoutPlaintext

This document is generated from apis/kms-2014-11-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonKMS.cancel_key_deletion

AmazonKMS.create_alias

AmazonKMS.create_grant

AmazonKMS.create_key

AmazonKMS.decrypt

AmazonKMS.delete_alias

AmazonKMS.delete_imported_key_material

AmazonKMS.describe_key

AmazonKMS.disable_key

AmazonKMS.disable_key_rotation

AmazonKMS.enable_key

AmazonKMS.enable_key_rotation

AmazonKMS.encrypt

AmazonKMS.generate_data_key

AmazonKMS.generate_data_key_without_plaintext

AmazonKMS.generate_random

AmazonKMS.get_key_policy

AmazonKMS.get_key_rotation_status

AmazonKMS.get_parameters_for_import

AmazonKMS.import_key_material

AmazonKMS.list_aliases

AmazonKMS.list_grants

AmazonKMS.list_key_policies

AmazonKMS.list_keys

AmazonKMS.list_resource_tags

AmazonKMS.list_retirable_grants

AmazonKMS.put_key_policy

AmazonKMS.re_encrypt

AmazonKMS.retire_grant

AmazonKMS.revoke_grant

AmazonKMS.schedule_key_deletion

AmazonKMS.tag_resource

AmazonKMS.untag_resource

AmazonKMS.update_alias

AmazonKMS.update_key_description

AmazonKMS.cancel_key_deletion — Method.

cancel_key_deletion(::AWSConfig, arguments::Dict)
cancel_key_deletion(::AWSConfig; KeyId=)

CancelKeyDeletion Operation

Cancels the deletion of a customer master key (CMK). When this operation is successful, the CMK is set to the Disabled state. To enable a CMK, use EnableKey.

For more information about scheduling and canceling deletion of a CMK, see Deleting Customer Master Keys in the AWS Key Management Service Developer Guide.

Arguments

KeyId::String – Required

The unique identifier for the customer master key (CMK) for which to cancel deletion.

To specify this value, use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To obtain the unique key ID and key ARN for a given CMK, use ListKeys or DescribeKey.

Returns

CancelKeyDeletionResponse

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.create_alias — Method.

create_alias(::AWSConfig, arguments::Dict)
create_alias(::AWSConfig; AliasName=, TargetKeyId=)

CreateAlias Operation

Creates a display name for a customer master key. An alias can be used to identify a key and should be unique. The console enforces a one-to-one mapping between the alias and a key. An alias name can contain only alphanumeric characters, forward slashes (/), underscores (_), and dashes (-). An alias must start with the word "alias" followed by a forward slash (alias/). An alias that begins with "aws" after the forward slash (alias/aws...) is reserved by Amazon Web Services (AWS).

The alias and the key it is mapped to must be in the same AWS account and the same region.

To map an alias to a different key, call UpdateAlias.

Arguments

AliasName::String – Required

String that contains the display name. The name must start with the word "alias" followed by a forward slash (alias/). Aliases that begin with "alias/AWS" are reserved.

TargetKeyId::String – Required

An identifier of the key for which you are creating the alias. This value cannot be another alias but can be a globally unique identifier or a fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Exceptions

DependencyTimeoutException, AlreadyExistsException, NotFoundException, InvalidAliasNameException, KMSInternalException, LimitExceededException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.create_grant — Method.

create_grant(::AWSConfig, arguments::Dict)
create_grant(::AWSConfig; KeyId=, GranteePrincipal=, <keyword arguments>)

CreateGrant Operation

Adds a grant to a key to specify who can use the key and under what conditions. Grants are alternate permission mechanisms to key policies.

For more information about grants, see Grants in the AWS Key Management Service Developer Guide.

Arguments

KeyId::String – Required

The unique identifier for the customer master key (CMK) that the grant applies to.

To specify this value, use the globally unique key ID or the Amazon Resource Name (ARN) of the key. Examples:

Globally unique key ID: 12345678-1234-1234-1234-123456789012

Key ARN: arn:aws:kms:us-west-2:123456789012:key/12345678-1234-1234-1234-123456789012

GranteePrincipal::String – Required

The principal that is given permission to perform the operations that the grant permits.

To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, IAM roles, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference.

RetiringPrincipal::String

The principal that is given permission to retire the grant by using RetireGrant operation.

To specify the principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, federated users, and assumed role users. For examples of the ARN syntax to use for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the AWS General Reference.

Operations: [: "Decrypt", "Encrypt", "GenerateDataKey", "GenerateDataKeyWithoutPlaintext", "ReEncryptFrom", "ReEncryptTo", "CreateGrant", "RetireGrant" or "DescribeKey"...]

A list of operations that the grant permits.

Constraints:

A structure that you can use to allow certain operations in the grant only when the desired encryption context is present. For more information about encryption context, see Encryption Context in the AWS Key Management Service Developer Guide.

    Dict(
        EncryptionContextSubset  => ::Dict{String,String},
        EncryptionContextEquals  => ::Dict{String,String}
    )

GrantTokens: [::String...]

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

Name::String

A friendly name for identifying the grant. Use this value to prevent unintended creation of duplicate grants when retrying this request.

When this value is absent, all CreateGrant requests result in a new grant with a unique GrantId even if all the supplied parameters are identical. This can result in unintended duplicates when you retry the CreateGrant request.

When this value is present, you can retry a CreateGrant request with identical parameters; if the grant already exists, the original GrantId is returned without creating a new grant. Note that the returned grant token is unique with every CreateGrant request, even when a duplicate GrantId is returned. All grant tokens obtained in this way can be used interchangeably.

Returns

CreateGrantResponse

Exceptions

NotFoundException, DisabledException, DependencyTimeoutException, InvalidArnException, KMSInternalException, InvalidGrantTokenException, LimitExceededException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.create_key — Method.

create_key(::AWSConfig, arguments::Dict)
create_key(::AWSConfig; <keyword arguments>)

CreateKey Operation

Creates a customer master key (CMK).

You can use a CMK to encrypt small amounts of data (4 KiB or less) directly, but CMKs are more commonly used to encrypt data encryption keys (DEKs), which are used to encrypt raw data. For more information about DEKs and the difference between CMKs and DEKs, see the following:

The GenerateDataKey operation

AWS Key Management Service Concepts in the AWS Key Management Service Developer Guide

Arguments

Policy::String

The key policy to attach to the CMK.

If you specify a policy and do not set BypassPolicyLockoutSafetyCheck to true, the policy must meet the following criteria:

It must allow the principal that is making the CreateKey request to make a subsequent PutKeyPolicy request on the CMK. This reduces the likelihood that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide.

The principals that are specified in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before specifying the new principal in a key policy because the new principal might not immediately be visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the IAM User Guide.

If you do not specify a policy, AWS KMS attaches a default key policy to the CMK. For more information, see Default Key Policy in the AWS Key Management Service Developer Guide.

The policy size limit is 32 KiB (32768 bytes).

Description::String

A description of the CMK.

Use a description that helps you decide whether the CMK is appropriate for a task.

KeyUsage: "ENCRYPT_DECRYPT"

The intended use of the CMK.

You can use CMKs only for symmetric encryption and decryption.

Origin: "AWS_KMS" or "EXTERNAL"

The source of the CMK's key material.

The default is AWS_KMS, which means AWS KMS creates the key material. When this parameter is set to EXTERNAL, the request creates a CMK without key material so that you can import key material from your existing key management infrastructure. For more information about importing key material into AWS KMS, see Importing Key Material in the AWS Key Management Service Developer Guide.

The CMK's Origin is immutable and is set when the CMK is created.

BypassPolicyLockoutSafetyCheck::Bool

A flag to indicate whether to bypass the key policy lockout safety check.

Important

Setting this value to true increases the likelihood that the CMK becomes unmanageable. Do not set this value to true indiscriminately.

For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide.

Use this parameter only when you include a policy in the request and you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK.

The default value is false.

Tags:

One or more tags. Each tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty (null) strings.

Use this parameter to tag the CMK when it is created. Alternately, you can omit this parameter and instead tag the CMK after it is created using TagResource.

    [Dict(
        TagKey * => ::String,
        TagValue * => ::String
    )...]

Returns

CreateKeyResponse

Exceptions

MalformedPolicyDocumentException, DependencyTimeoutException, InvalidArnException, UnsupportedOperationException, KMSInternalException, LimitExceededException or TagException.

See also: AWS API Documentation

AmazonKMS.decrypt — Method.

decrypt(::AWSConfig, arguments::Dict)
decrypt(::AWSConfig; CiphertextBlob=, <keyword arguments>)

Decrypt Operation

Decrypts ciphertext. Ciphertext is plaintext that has been previously encrypted by using any of the following functions:

GenerateDataKey

GenerateDataKeyWithoutPlaintext

Encrypt

Note that if a caller has been granted access permissions to all keys (through, for example, IAM user policies that grant Decrypt permission on all resources), then ciphertext encrypted by using keys in other accounts where the key grants access to the caller can be decrypted. To remedy this, we recommend that you do not grant Decrypt access in an IAM user policy. Instead grant Decrypt access only in key policies. If you must grant Decrypt access in an IAM user policy, you should scope the resource to specific keys or to specific trusted accounts.

Arguments

CiphertextBlob: blob – Required

Ciphertext to be decrypted. The blob includes metadata.

EncryptionContext::Dict{String,String}

The encryption context. If this was specified in the Encrypt function, it must be specified here or the decryption operation will fail. For more information, see Encryption Context.

GrantTokens: [::String...]

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

Returns

DecryptResponse

Exceptions

NotFoundException, DisabledException, InvalidCiphertextException, KeyUnavailableException, DependencyTimeoutException, InvalidGrantTokenException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.delete_alias — Method.

delete_alias(::AWSConfig, arguments::Dict)
delete_alias(::AWSConfig; AliasName=)

DeleteAlias Operation

Deletes the specified alias. To map an alias to a different key, call UpdateAlias.

Arguments

AliasName::String – Required

The alias to be deleted. The name must start with the word "alias" followed by a forward slash (alias/). Aliases that begin with "alias/AWS" are reserved.

Exceptions

DependencyTimeoutException, NotFoundException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.delete_imported_key_material — Method.

delete_imported_key_material(::AWSConfig, arguments::Dict)
delete_imported_key_material(::AWSConfig; KeyId=)

DeleteImportedKeyMaterial Operation

Deletes key material that you previously imported and makes the specified customer master key (CMK) unusable. For more information about importing key material into AWS KMS, see Importing Key Material in the AWS Key Management Service Developer Guide.

When the specified CMK is in the PendingDeletion state, this operation does not change the CMK's state. Otherwise, it changes the CMK's state to PendingImport.

After you delete key material, you can use ImportKeyMaterial to reimport the same key material into the CMK.

Arguments

KeyId::String – Required

The identifier of the CMK whose key material to delete. The CMK's Origin must be EXTERNAL.

A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Exceptions

InvalidArnException, UnsupportedOperationException, DependencyTimeoutException, NotFoundException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.describe_key — Method.

describe_key(::AWSConfig, arguments::Dict)
describe_key(::AWSConfig; KeyId=, <keyword arguments>)

DescribeKey Operation

Provides detailed information about the specified customer master key.

Arguments

KeyId::String – Required

A unique identifier for the customer master key. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN Example - arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Alias Name Example - alias/MyAliasName

GrantTokens: [::String...]

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

Returns

DescribeKeyResponse

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException or KMSInternalException.

See also: AWS API Documentation

AmazonKMS.disable_key — Method.

disable_key(::AWSConfig, arguments::Dict)
disable_key(::AWSConfig; KeyId=)

DisableKey Operation

Sets the state of a customer master key (CMK) to disabled, thereby preventing its use for cryptographic operations. For more information about how key state affects the use of a CMK, see How Key State Affects the Use of a Customer Master Key in the AWS Key Management Service Developer Guide.

Arguments

KeyId::String – Required

A unique identifier for the CMK.

Use the CMK's unique identifier or its Amazon Resource Name (ARN). For example:

Unique ID: 1234abcd-12ab-34cd-56ef-1234567890ab

ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.disable_key_rotation — Method.

disable_key_rotation(::AWSConfig, arguments::Dict)
disable_key_rotation(::AWSConfig; KeyId=)

DisableKeyRotation Operation

Disables rotation of the specified key.

Arguments

KeyId::String – Required

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Exceptions

NotFoundException, DisabledException, InvalidArnException, DependencyTimeoutException, KMSInternalException, KMSInvalidStateException or UnsupportedOperationException.

See also: AWS API Documentation

AmazonKMS.enable_key — Method.

enable_key(::AWSConfig, arguments::Dict)
enable_key(::AWSConfig; KeyId=)

EnableKey Operation

Marks a key as enabled, thereby permitting its use.

Arguments

KeyId::String – Required

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException, KMSInternalException, LimitExceededException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.enable_key_rotation — Method.

enable_key_rotation(::AWSConfig, arguments::Dict)
enable_key_rotation(::AWSConfig; KeyId=)

EnableKeyRotation Operation

Enables rotation of the specified customer master key.

Arguments

KeyId::String – Required

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Exceptions

NotFoundException, DisabledException, InvalidArnException, DependencyTimeoutException, KMSInternalException, KMSInvalidStateException or UnsupportedOperationException.

See also: AWS API Documentation

AmazonKMS.encrypt — Method.

encrypt(::AWSConfig, arguments::Dict)
encrypt(::AWSConfig; KeyId=, Plaintext=, <keyword arguments>)

Encrypt Operation

Encrypts plaintext into ciphertext by using a customer master key. The Encrypt function has two primary use cases:

You can encrypt up to 4 KB of arbitrary data such as an RSA key, a database password, or other sensitive customer information.

If you are moving encrypted data from one region to another, you can use this API to encrypt in the new region the plaintext data key that was used to encrypt the data in the original region. This provides you with an encrypted copy of the data key that can be decrypted in the new region and used there to decrypt the encrypted data.

Unless you are moving encrypted data from one region to another, you don't use this function to encrypt a generated data key within a region. You retrieve data keys already encrypted by calling the GenerateDataKey or GenerateDataKeyWithoutPlaintext function. Data keys don't need to be encrypted again by calling Encrypt.

If you want to encrypt data locally in your application, you can use the GenerateDataKey function to return a plaintext data encryption key and a copy of the key encrypted under the customer master key (CMK) of your choosing.

Arguments

KeyId::String – Required

A unique identifier for the customer master key. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN Example - arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Alias Name Example - alias/MyAliasName

Plaintext: blob – Required

Data to be encrypted.

EncryptionContext::Dict{String,String}

Name-value pair that specifies the encryption context to be used for authenticated encryption. If used here, the same value must be supplied to the Decrypt API or decryption will fail. For more information, see Encryption Context.

GrantTokens: [::String...]

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

Returns

EncryptResponse

Exceptions

NotFoundException, DisabledException, KeyUnavailableException, DependencyTimeoutException, InvalidKeyUsageException, InvalidGrantTokenException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.generate_data_key — Method.

generate_data_key(::AWSConfig, arguments::Dict)
generate_data_key(::AWSConfig; KeyId=, <keyword arguments>)

GenerateDataKey Operation

Returns a data encryption key that you can use in your application to encrypt data locally.

You must specify the customer master key (CMK) under which to generate the data key. You must also specify the length of the data key using either the KeySpec or NumberOfBytes field. You must specify one field or the other, but not both. For common key lengths (128-bit and 256-bit symmetric keys), we recommend that you use KeySpec.

This operation returns a plaintext copy of the data key in the Plaintext field of the response, and an encrypted copy of the data key in the CiphertextBlob field. The data key is encrypted under the CMK specified in the KeyId field of the request.

We recommend that you use the following pattern to encrypt data locally in your application:

Use this operation (GenerateDataKey) to retrieve a data encryption key.

Use the plaintext data encryption key (returned in the Plaintext field of the response) to encrypt data locally, then erase the plaintext data key from memory.

Store the encrypted data key (returned in the CiphertextBlob field of the response) alongside the locally encrypted data.

To decrypt data locally:

Use the Decrypt operation to decrypt the encrypted data key into a plaintext copy of the data key.

Use the plaintext data key to decrypt data locally, then erase the plaintext data key from memory.

To return only an encrypted copy of the data key, use GenerateDataKeyWithoutPlaintext. To return a random byte string that is cryptographically secure, use GenerateRandom.

If you use the optional EncryptionContext field, you must store at least enough information to be able to reconstruct the full encryption context when you later send the ciphertext to the Decrypt operation. It is a good practice to choose an encryption context that you can reconstruct on the fly to better secure the ciphertext. For more information, see Encryption Context in the AWS Key Management Service Developer Guide.

Arguments

KeyId::String – Required

The identifier of the CMK under which to generate and encrypt the data encryption key.

A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK, or the alias name or ARN of an alias that refers to the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

CMK ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

EncryptionContext::Dict{String,String}

A set of key-value pairs that represents additional authenticated data.

For more information, see Encryption Context in the AWS Key Management Service Developer Guide.

NumberOfBytes::Int

The length of the data encryption key in bytes. For example, use the value 64 to generate a 512-bit data key (64 bytes is 512 bits). For common key lengths (128-bit and 256-bit symmetric keys), we recommend that you use the KeySpec field instead of this one.

KeySpec: "AES_256" or "AES_128"

The length of the data encryption key. Use AES_128 to generate a 128-bit symmetric key, or AES_256 to generate a 256-bit symmetric key.

GrantTokens: [::String...]

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

Returns

GenerateDataKeyResponse

Exceptions

NotFoundException, DisabledException, KeyUnavailableException, DependencyTimeoutException, InvalidKeyUsageException, InvalidGrantTokenException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.generate_data_key_without_plaintext — Method.

generate_data_key_without_plaintext(::AWSConfig, arguments::Dict)
generate_data_key_without_plaintext(::AWSConfig; KeyId=, <keyword arguments>)

GenerateDataKeyWithoutPlaintext Operation

Returns a data encryption key encrypted under a customer master key (CMK). This operation is identical to GenerateDataKey but returns only the encrypted copy of the data key.

This operation is useful in a system that has multiple components with different degrees of trust. For example, consider a system that stores encrypted data in containers. Each container stores the encrypted data and an encrypted copy of the data key. One component of the system, called the control plane, creates new containers. When it creates a new container, it uses this operation (GenerateDataKeyWithoutPlaintext) to get an encrypted data key and then stores it in the container. Later, a different component of the system, called the data plane, puts encrypted data into the containers. To do this, it passes the encrypted data key to the Decrypt operation, then uses the returned plaintext data key to encrypt data, and finally stores the encrypted data in the container. In this system, the control plane never sees the plaintext data key.

Arguments

KeyId::String – Required

The identifier of the CMK under which to generate and encrypt the data encryption key.

A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK, or the alias name or ARN of an alias that refers to the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

CMK ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

EncryptionContext::Dict{String,String}

A set of key-value pairs that represents additional authenticated data.

For more information, see Encryption Context in the AWS Key Management Service Developer Guide.

KeySpec: "AES_256" or "AES_128"

The length of the data encryption key. Use AES_128 to generate a 128-bit symmetric key, or AES_256 to generate a 256-bit symmetric key.

NumberOfBytes::Int

The length of the data encryption key in bytes. For example, use the value 64 to generate a 512-bit data key (64 bytes is 512 bits). For common key lengths (128-bit and 256-bit symmetric keys), we recommend that you use the KeySpec field instead of this one.

GrantTokens: [::String...]

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

Returns

GenerateDataKeyWithoutPlaintextResponse

Exceptions

NotFoundException, DisabledException, KeyUnavailableException, DependencyTimeoutException, InvalidKeyUsageException, InvalidGrantTokenException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.generate_random — Method.

generate_random(::AWSConfig, arguments::Dict)
generate_random(::AWSConfig; <keyword arguments>)

GenerateRandom Operation

Returns a random byte string that is cryptographically secure.

For more information about entropy and random number generation, see the AWS Key Management Service Cryptographic Details whitepaper.

Arguments

NumberOfBytes::Int

The length of the byte string.

Returns

GenerateRandomResponse

Exceptions

DependencyTimeoutException or KMSInternalException.

See also: AWS API Documentation

AmazonKMS.get_key_policy — Method.

get_key_policy(::AWSConfig, arguments::Dict)
get_key_policy(::AWSConfig; KeyId=, PolicyName=)

GetKeyPolicy Operation

Retrieves a policy attached to the specified key.

Arguments

KeyId::String – Required

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

PolicyName::String – Required

String that contains the name of the policy. Currently, this must be "default". Policy names can be discovered by calling ListKeyPolicies.

Returns

GetKeyPolicyResponse

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.get_key_rotation_status — Method.

get_key_rotation_status(::AWSConfig, arguments::Dict)
get_key_rotation_status(::AWSConfig; KeyId=)

GetKeyRotationStatus Operation

Retrieves a Boolean value that indicates whether key rotation is enabled for the specified key.

Arguments

KeyId::String – Required

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Returns

GetKeyRotationStatusResponse

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException, KMSInternalException, KMSInvalidStateException or UnsupportedOperationException.

See also: AWS API Documentation

AmazonKMS.get_parameters_for_import — Method.

get_parameters_for_import(::AWSConfig, arguments::Dict)
get_parameters_for_import(::AWSConfig; KeyId=, WrappingAlgorithm=, WrappingKeySpec=)

GetParametersForImport Operation

Returns the items you need in order to import key material into AWS KMS from your existing key management infrastructure. For more information about importing key material into AWS KMS, see Importing Key Material in the AWS Key Management Service Developer Guide.

You must specify the key ID of the customer master key (CMK) into which you will import key material. This CMK's Origin must be EXTERNAL. You must also specify the wrapping algorithm and type of wrapping key (public key) that you will use to encrypt the key material.

This operation returns a public key and an import token. Use the public key to encrypt the key material. Store the import token to send with a subsequent ImportKeyMaterial request. The public key and import token from the same response must be used together. These items are valid for 24 hours, after which they cannot be used for a subsequent ImportKeyMaterial request. To retrieve new ones, send another GetParametersForImport request.

Arguments

KeyId::String – Required

The identifier of the CMK into which you will import key material. The CMK's Origin must be EXTERNAL.

A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

WrappingAlgorithm: "RSAES_PKCS1_V1_5", "RSAES_OAEP_SHA_1" or "RSAES_OAEP_SHA_256" – Required

The algorithm you will use to encrypt the key material before importing it with ImportKeyMaterial. For more information, see Encrypt the Key Material in the AWS Key Management Service Developer Guide.

WrappingKeySpec: "RSA_2048" – Required

The type of wrapping key (public key) to return in the response. Only 2048-bit RSA public keys are supported.

Returns

GetParametersForImportResponse

Exceptions

InvalidArnException, UnsupportedOperationException, DependencyTimeoutException, NotFoundException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.import_key_material — Method.

import_key_material(::AWSConfig, arguments::Dict)
import_key_material(::AWSConfig; KeyId=, ImportToken=, EncryptedKeyMaterial=, <keyword arguments>)

ImportKeyMaterial Operation

Imports key material into an AWS KMS customer master key (CMK) from your existing key management infrastructure. For more information about importing key material into AWS KMS, see Importing Key Material in the AWS Key Management Service Developer Guide.

You must specify the key ID of the CMK to import the key material into. This CMK's Origin must be EXTERNAL. You must also send an import token and the encrypted key material. Send the import token that you received in the same GetParametersForImport response that contained the public key that you used to encrypt the key material. You must also specify whether the key material expires and if so, when. When the key material expires, AWS KMS deletes the key material and the CMK becomes unusable. To use the CMK again, you can reimport the same key material. If you set an expiration date, you can change it only by reimporting the same key material and specifying a new expiration date.

When this operation is successful, the specified CMK's key state changes to Enabled, and you can use the CMK.

After you successfully import key material into a CMK, you can reimport the same key material into that CMK, but you cannot import different key material.

Arguments

KeyId::String – Required

The identifier of the CMK to import the key material into. The CMK's Origin must be EXTERNAL.

A valid identifier is the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

ImportToken: blob – Required

The import token that you received in the response to a previous GetParametersForImport request. It must be from the same response that contained the public key that you used to encrypt the key material.

EncryptedKeyMaterial: blob – Required

The encrypted key material to import. It must be encrypted with the public key that you received in the response to a previous GetParametersForImport request, using the wrapping algorithm that you specified in that request.

ValidTo: timestamp

The time at which the imported key material expires. When the key material expires, AWS KMS deletes the key material and the CMK becomes unusable. You must omit this parameter when the ExpirationModel parameter is set to KEY_MATERIAL_DOES_NOT_EXPIRE. Otherwise it is required.

ExpirationModel: "KEY_MATERIAL_EXPIRES" or "KEY_MATERIAL_DOES_NOT_EXPIRE"

Specifies whether the key material expires. The default is KEY_MATERIAL_EXPIRES, in which case you must include the ValidTo parameter. When this parameter is set to KEY_MATERIAL_DOES_NOT_EXPIRE, you must omit the ValidTo parameter.

Returns

ImportKeyMaterialResponse

Exceptions

InvalidArnException, UnsupportedOperationException, DependencyTimeoutException, NotFoundException, KMSInternalException, KMSInvalidStateException, InvalidCiphertextException, IncorrectKeyMaterialException, ExpiredImportTokenException or InvalidImportTokenException.

See also: AWS API Documentation

AmazonKMS.list_aliases — Method.

list_aliases(::AWSConfig, arguments::Dict)
list_aliases(::AWSConfig; <keyword arguments>)

ListAliases Operation

Lists all of the key aliases in the account.

Arguments

Limit::Int

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 100, inclusive. If you do not include a value, it defaults to 50.

Marker::String

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

Returns

ListAliasesResponse

Exceptions

DependencyTimeoutException, InvalidMarkerException or KMSInternalException.

See also: AWS API Documentation

AmazonKMS.list_grants — Method.

list_grants(::AWSConfig, arguments::Dict)
list_grants(::AWSConfig; KeyId=, <keyword arguments>)

ListGrants Operation

List the grants for a specified key.

Arguments

Limit::Int

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 100, inclusive. If you do not include a value, it defaults to 50.

Marker::String

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

KeyId::String – Required

A unique identifier for the customer master key. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Returns

ListGrantsResponse

Exceptions

NotFoundException, DependencyTimeoutException, InvalidMarkerException, InvalidArnException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.list_key_policies — Method.

list_key_policies(::AWSConfig, arguments::Dict)
list_key_policies(::AWSConfig; KeyId=, <keyword arguments>)

ListKeyPolicies Operation

Retrieves a list of policies attached to a key.

Arguments

KeyId::String – Required

A unique identifier for the customer master key (CMK). You can use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Limit::Int

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 1000, inclusive. If you do not include a value, it defaults to 100.

Currently only 1 policy can be attached to a key.

Marker::String

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

Returns

ListKeyPoliciesResponse

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.list_keys — Method.

list_keys(::AWSConfig, arguments::Dict)
list_keys(::AWSConfig; <keyword arguments>)

ListKeys Operation

Lists the customer master keys.

Arguments

Limit::Int

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 1000, inclusive. If you do not include a value, it defaults to 100.

Marker::String

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

Returns

ListKeysResponse

Exceptions

DependencyTimeoutException, KMSInternalException or InvalidMarkerException.

See also: AWS API Documentation

AmazonKMS.list_resource_tags — Method.

list_resource_tags(::AWSConfig, arguments::Dict)
list_resource_tags(::AWSConfig; KeyId=, <keyword arguments>)

ListResourceTags Operation

Returns a list of all tags for the specified customer master key (CMK).

Arguments

KeyId::String – Required

A unique identifier for the CMK whose tags you are listing. You can use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Limit::Int

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 50, inclusive. If you do not include a value, it defaults to 50.

Marker::String

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

Do not attempt to construct this value. Use only the value of NextMarker from the truncated response you just received.

Returns

ListResourceTagsResponse

Exceptions

KMSInternalException, NotFoundException, InvalidArnException or InvalidMarkerException.

See also: AWS API Documentation

AmazonKMS.list_retirable_grants — Method.

list_retirable_grants(::AWSConfig, arguments::Dict)
list_retirable_grants(::AWSConfig; RetiringPrincipal=, <keyword arguments>)

ListRetirableGrants Operation

Returns a list of all grants for which the grant's RetiringPrincipal matches the one specified.

A typical use is to list all grants that you are able to retire. To retire a grant, use RetireGrant.

Arguments

Limit::Int

Use this parameter to specify the maximum number of items to return. When this value is present, AWS KMS does not return more than the specified number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and 100, inclusive. If you do not include a value, it defaults to 50.

Marker::String

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

RetiringPrincipal::String – Required

The retiring principal for which to list grants.

To specify the retiring principal, use the Amazon Resource Name (ARN) of an AWS principal. Valid AWS principals include AWS accounts (root), IAM users, federated users, and assumed role users. For examples of the ARN syntax for specifying a principal, see AWS Identity and Access Management (IAM) in the Example ARNs section of the Amazon Web Services General Reference.

Returns

ListGrantsResponse

Exceptions

DependencyTimeoutException, InvalidMarkerException, InvalidArnException, NotFoundException or KMSInternalException.

See also: AWS API Documentation

AmazonKMS.put_key_policy — Method.

put_key_policy(::AWSConfig, arguments::Dict)
put_key_policy(::AWSConfig; KeyId=, PolicyName=, Policy=, <keyword arguments>)

PutKeyPolicy Operation

Attaches a key policy to the specified customer master key (CMK).

For more information about key policies, see Key Policies in the AWS Key Management Service Developer Guide.

Arguments

KeyId::String – Required

A unique identifier for the CMK.

Use the CMK's unique identifier or its Amazon Resource Name (ARN). For example:

Unique ID: 1234abcd-12ab-34cd-56ef-1234567890ab

ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

PolicyName::String – Required

The name of the key policy.

This value must be default.

Policy::String – Required

The key policy to attach to the CMK.

If you do not set BypassPolicyLockoutSafetyCheck to true, the policy must meet the following criteria:

It must allow the principal that is making the PutKeyPolicy request to make a subsequent PutKeyPolicy request on the CMK. This reduces the likelihood that the CMK becomes unmanageable. For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide.

The principals that are specified in the key policy must exist and be visible to AWS KMS. When you create a new AWS principal (for example, an IAM user or role), you might need to enforce a delay before specifying the new principal in a key policy because the new principal might not immediately be visible to AWS KMS. For more information, see Changes that I make are not always immediately visible in the IAM User Guide.

The policy size limit is 32 KiB (32768 bytes).

BypassPolicyLockoutSafetyCheck::Bool

A flag to indicate whether to bypass the key policy lockout safety check.

Important

Setting this value to true increases the likelihood that the CMK becomes unmanageable. Do not set this value to true indiscriminately.

For more information, refer to the scenario in the Default Key Policy section in the AWS Key Management Service Developer Guide.

Use this parameter only when you intend to prevent the principal that is making the request from making a subsequent PutKeyPolicy request on the CMK.

The default value is false.

Exceptions

NotFoundException, InvalidArnException, MalformedPolicyDocumentException, DependencyTimeoutException, UnsupportedOperationException, KMSInternalException, LimitExceededException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.re_encrypt — Method.

re_encrypt(::AWSConfig, arguments::Dict)
re_encrypt(::AWSConfig; CiphertextBlob=, DestinationKeyId=, <keyword arguments>)

ReEncrypt Operation

Encrypts data on the server side with a new customer master key (CMK) without exposing the plaintext of the data on the client side. The data is first decrypted and then reencrypted. You can also use this operation to change the encryption context of a ciphertext.

Unlike other operations, ReEncrypt is authorized twice, once as ReEncryptFrom on the source CMK and once as ReEncryptTo on the destination CMK. We recommend that you include the "kms:ReEncrypt*" permission in your key policies to permit reencryption from or to the CMK. This permission is automatically included in the key policy when you create a CMK through the console, but you must include it manually when you create a CMK programmatically or when you set a key policy with the PutKeyPolicy operation.

Arguments

CiphertextBlob: blob – Required

Ciphertext of the data to reencrypt.

SourceEncryptionContext::Dict{String,String}

Encryption context used to encrypt and decrypt the data specified in the CiphertextBlob parameter.

DestinationKeyId::String – Required

A unique identifier for the CMK to use to reencrypt the data. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN Example - arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Alias Name Example - alias/MyAliasName

DestinationEncryptionContext::Dict{String,String}

Encryption context to use when the data is reencrypted.

GrantTokens: [::String...]

A list of grant tokens.

For more information, see Grant Tokens in the AWS Key Management Service Developer Guide.

Returns

ReEncryptResponse

Exceptions

NotFoundException, DisabledException, InvalidCiphertextException, KeyUnavailableException, DependencyTimeoutException, InvalidKeyUsageException, InvalidGrantTokenException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.retire_grant — Method.

retire_grant(::AWSConfig, arguments::Dict)
retire_grant(::AWSConfig; <keyword arguments>)

RetireGrant Operation

Retires a grant. To clean up, you can retire a grant when you're done using it. You should revoke a grant when you intend to actively deny operations that depend on it. The following are permitted to call this API:

The AWS account (root user) under which the grant was created

The RetiringPrincipal, if present in the grant

The GranteePrincipal, if RetireGrant is an operation specified in the grant

You must identify the grant to retire by its grant token or by a combination of the grant ID and the Amazon Resource Name (ARN) of the customer master key (CMK). A grant token is a unique variable-length base64-encoded string. A grant ID is a 64 character unique identifier of a grant. The CreateGrant operation returns both.

Arguments

GrantToken::String

Token that identifies the grant to be retired.

KeyId::String

The Amazon Resource Name of the CMK associated with the grant. Example:

arn:aws:kms:us-east-2:444455556666:key/1234abcd-12ab-34cd-56ef-1234567890ab

GrantId::String

Unique identifier of the grant to retire. The grant ID is returned in the response to a CreateGrant operation.

Grant ID Example - 0123456789012345678901234567890123456789012345678901234567890123

Exceptions

InvalidGrantTokenException, InvalidGrantIdException, NotFoundException, DependencyTimeoutException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.revoke_grant — Method.

revoke_grant(::AWSConfig, arguments::Dict)
revoke_grant(::AWSConfig; KeyId=, GrantId=)

RevokeGrant Operation

Revokes a grant. You can revoke a grant to actively deny operations that depend on it.

Arguments

KeyId::String – Required

A unique identifier for the customer master key associated with the grant. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

GrantId::String – Required

Identifier of the grant to be revoked.

Exceptions

NotFoundException, DependencyTimeoutException, InvalidArnException, InvalidGrantIdException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.schedule_key_deletion — Method.

schedule_key_deletion(::AWSConfig, arguments::Dict)
schedule_key_deletion(::AWSConfig; KeyId=, <keyword arguments>)

ScheduleKeyDeletion Operation

Schedules the deletion of a customer master key (CMK). You may provide a waiting period, specified in days, before deletion occurs. If you do not provide a waiting period, the default period of 30 days is used. When this operation is successful, the state of the CMK changes to PendingDeletion. Before the waiting period ends, you can use CancelKeyDeletion to cancel the deletion of the CMK. After the waiting period ends, AWS KMS deletes the CMK and all AWS KMS data associated with it, including all aliases that refer to it.

Important

Deleting a CMK is a destructive and potentially dangerous operation. When a CMK is deleted, all data that was encrypted under the CMK is rendered unrecoverable. To restrict the use of a CMK without deleting it, use DisableKey.

For more information about scheduling a CMK for deletion, see Deleting Customer Master Keys in the AWS Key Management Service Developer Guide.

Arguments

KeyId::String – Required

The unique identifier for the customer master key (CMK) to delete.

To specify this value, use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To obtain the unique key ID and key ARN for a given CMK, use ListKeys or DescribeKey.

PendingWindowInDays::Int

The waiting period, specified in number of days. After the waiting period ends, AWS KMS deletes the customer master key (CMK).

This value is optional. If you include a value, it must be between 7 and 30, inclusive. If you do not include a value, it defaults to 30.

Returns

ScheduleKeyDeletionResponse

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.tag_resource — Method.

tag_resource(::AWSConfig, arguments::Dict)
tag_resource(::AWSConfig; KeyId=, Tags=)

TagResource Operation

Adds or overwrites one or more tags for the specified customer master key (CMK).

Each tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty (null) strings.

You cannot use the same tag key more than once per CMK. For example, consider a CMK with one tag whose tag key is Purpose and tag value is Test. If you send a TagResource request for this CMK with a tag key of Purpose and a tag value of Prod, it does not create a second tag. Instead, the original tag is overwritten with the new tag value.

Arguments

KeyId::String – Required

A unique identifier for the CMK you are tagging. You can use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Tags: – Required

One or more tags. Each tag consists of a tag key and a tag value.

    [Dict(
        TagKey * => ::String,
        TagValue * => ::String
    )...]

Exceptions

KMSInternalException, NotFoundException, InvalidArnException, KMSInvalidStateException, LimitExceededException or TagException.

See also: AWS API Documentation

AmazonKMS.untag_resource — Method.

untag_resource(::AWSConfig, arguments::Dict)
untag_resource(::AWSConfig; KeyId=, TagKeys=)

UntagResource Operation

Removes the specified tag or tags from the specified customer master key (CMK).

To remove a tag, you specify the tag key for each tag to remove. You do not specify the tag value. To overwrite the tag value for an existing tag, use TagResource.

Arguments

KeyId::String – Required

A unique identifier for the CMK from which you are removing tags. You can use the unique key ID or the Amazon Resource Name (ARN) of the CMK. Examples:

Unique key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

TagKeys: [::String...] – Required

One or more tag keys. Specify only the tag keys, not the tag values.

Exceptions

KMSInternalException, NotFoundException, InvalidArnException, KMSInvalidStateException or TagException.

See also: AWS API Documentation

AmazonKMS.update_alias — Method.

update_alias(::AWSConfig, arguments::Dict)
update_alias(::AWSConfig; AliasName=, TargetKeyId=)

UpdateAlias Operation

Updates an alias to map it to a different key.

An alias is not a property of a key. Therefore, an alias can be mapped to and unmapped from an existing key without changing the properties of the key.

An alias name can contain only alphanumeric characters, forward slashes (/), underscores (_), and dashes (-). An alias must start with the word "alias" followed by a forward slash (alias/). An alias that begins with "aws" after the forward slash (alias/aws...) is reserved by Amazon Web Services (AWS).

The alias and the key it is mapped to must be in the same AWS account and the same region.

Arguments

AliasName::String – Required

String that contains the name of the alias to be modified. The name must start with the word "alias" followed by a forward slash (alias/). Aliases that begin with "alias/aws" are reserved.

TargetKeyId::String – Required

Unique identifier of the customer master key to be mapped to the alias. This value can be a globally unique identifier or the fully specified ARN of a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

You can call ListAliases to verify that the alias is mapped to the correct TargetKeyId.

Exceptions

DependencyTimeoutException, NotFoundException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation

AmazonKMS.update_key_description — Method.

update_key_description(::AWSConfig, arguments::Dict)
update_key_description(::AWSConfig; KeyId=, Description=)

UpdateKeyDescription Operation

Updates the description of a customer master key (CMK).

Arguments

KeyId::String – Required

A unique identifier for the CMK. This value can be a globally unique identifier or the fully specified ARN to a key.

Key ARN Example - arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Globally Unique Key ID Example - 12345678-1234-1234-1234-123456789012

Description::String – Required

New description for the CMK.

Exceptions

NotFoundException, InvalidArnException, DependencyTimeoutException, KMSInternalException or KMSInvalidStateException.

See also: AWS API Documentation