AmazonKinesis.jl

AmazonKinesis

Amazon Kinesis Streams Service API Reference

Amazon Kinesis Streams is a managed service that scales elastically for real time processing of streaming big data.

This document is generated from apis/kinesis-2013-12-02.normal.json. See JuliaCloud/AWSCore.jl.

AmazonKinesis.add_tags_to_stream

AmazonKinesis.create_stream

AmazonKinesis.decrease_stream_retention_period

AmazonKinesis.delete_stream

AmazonKinesis.describe_limits

AmazonKinesis.describe_stream

AmazonKinesis.disable_enhanced_monitoring

AmazonKinesis.enable_enhanced_monitoring

AmazonKinesis.get_records

AmazonKinesis.get_shard_iterator

AmazonKinesis.increase_stream_retention_period

AmazonKinesis.list_streams

AmazonKinesis.list_tags_for_stream

AmazonKinesis.merge_shards

AmazonKinesis.put_record

AmazonKinesis.put_records

AmazonKinesis.remove_tags_from_stream

AmazonKinesis.split_shard

AmazonKinesis.start_stream_encryption

AmazonKinesis.stop_stream_encryption

AmazonKinesis.update_shard_count

AmazonKinesis.add_tags_to_stream — Function.

add_tags_to_stream(::AWSConfig, arguments::Dict)
add_tags_to_stream([::AWSConfig]; StreamName=, Tags=)

AddTagsToStream Operation

Adds or updates tags for the specified Amazon Kinesis stream. Each stream can have up to 10 tags.

If tags have already been assigned to the stream, AddTagsToStream overwrites any existing tags that correspond to the specified tag keys.

Arguments

StreamName = ::String – Required

The name of the stream.

Tags = ::Dict{String,String} – Required

The set of key-value pairs to use to create the tags.

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.create_stream — Function.

create_stream(::AWSConfig, arguments::Dict)
create_stream([::AWSConfig]; StreamName=, ShardCount=)

CreateStream Operation

Creates an Amazon Kinesis stream. A stream captures and transports data records that are continuously emitted from different data sources or producers. Scale-out within a stream is explicitly supported by means of shards, which are uniquely identified groups of data records in a stream.

You specify and control the number of shards that a stream is composed of. Each shard can support reads up to 5 transactions per second, up to a maximum data read total of 2 MB per second. Each shard can support writes up to 1,000 records per second, up to a maximum data write total of 1 MB per second. You can add shards to a stream if the amount of data input increases and you can remove shards if the amount of data input decreases.

The stream name identifies the stream. The name is scoped to the AWS account used by the application. It is also scoped by region. That is, two streams in two different accounts can have the same name, and two streams in the same account, but in two different regions, can have the same name.

CreateStream is an asynchronous operation. Upon receiving a CreateStream request, Amazon Kinesis immediately returns and sets the stream status to CREATING. After the stream is created, Amazon Kinesis sets the stream status to ACTIVE. You should perform read and write operations only on an ACTIVE stream.

You receive a LimitExceededException when making a CreateStream request if you try to do one of the following:

Have more than five streams in the CREATING state at any point in time.

Create more shards than are authorized for your account.

For the default shard limit for an AWS account, see Streams Limits in the Amazon Kinesis Streams Developer Guide. If you need to increase this limit, contact AWS Support.

You can use DescribeStream to check the stream status, which is returned in StreamStatus.

CreateStream has a limit of 5 transactions per second per account.

Arguments

StreamName = ::String – Required

A name to identify the stream. The stream name is scoped to the AWS account used by the application that creates the stream. It is also scoped by region. That is, two streams in two different AWS accounts can have the same name, and two streams in the same AWS account but in two different regions can have the same name.

ShardCount = ::Int – Required

The number of shards that the stream will use. The throughput of the stream is a function of the number of shards; more shards are required for greater provisioned throughput.

DefaultShardLimit;

Exceptions

ResourceInUseException, LimitExceededException or InvalidArgumentException.

See also: AWS API Documentation

AmazonKinesis.decrease_stream_retention_period — Function.

decrease_stream_retention_period(::AWSConfig, arguments::Dict)
decrease_stream_retention_period([::AWSConfig]; StreamName=, RetentionPeriodHours=)

DecreaseStreamRetentionPeriod Operation

Decreases the Amazon Kinesis stream's retention period, which is the length of time data records are accessible after they are added to the stream. The minimum value of a stream's retention period is 24 hours.

This operation may result in lost data. For example, if the stream's retention period is 48 hours and is decreased to 24 hours, any data already in the stream that is older than 24 hours is inaccessible.

Arguments

StreamName = ::String – Required

The name of the stream to modify.

RetentionPeriodHours = ::Int – Required

The new retention period of the stream, in hours. Must be less than the current retention period.

Exceptions

ResourceInUseException, ResourceNotFoundException or InvalidArgumentException.

See also: AWS API Documentation

AmazonKinesis.delete_stream — Function.

delete_stream(::AWSConfig, arguments::Dict)
delete_stream([::AWSConfig]; StreamName=)

DeleteStream Operation

Deletes an Amazon Kinesis stream and all its shards and data. You must shut down any applications that are operating on the stream before you delete the stream. If an application attempts to operate on a deleted stream, it will receive the exception ResourceNotFoundException.

If the stream is in the ACTIVE state, you can delete it. After a DeleteStream request, the specified stream is in the DELETING state until Amazon Kinesis completes the deletion.

Note: Amazon Kinesis might continue to accept data read and write operations, such as PutRecord, PutRecords, and GetRecords, on a stream in the DELETING state until the stream deletion is complete.

When you delete a stream, any shards in that stream are also deleted, and any tags are dissociated from the stream.

You can use the DescribeStream operation to check the state of the stream, which is returned in StreamStatus.

DeleteStream has a limit of 5 transactions per second per account.

Arguments

StreamName = ::String – Required

The name of the stream to delete.

Exceptions

ResourceNotFoundException or LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.describe_limits — Function.

describe_limits(::AWSConfig, arguments::Dict)
describe_limits([::AWSConfig]; )

DescribeLimits Operation

Describes the shard limits and usage for the account.

If you update your account limits, the old limits might be returned for a few minutes.

This operation has a limit of 1 transaction per second per account.

Arguments

Returns

DescribeLimitsOutput

Exceptions

LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.describe_stream — Function.

describe_stream(::AWSConfig, arguments::Dict)
describe_stream([::AWSConfig]; StreamName=, <keyword arguments>)

DescribeStream Operation

Describes the specified Amazon Kinesis stream.

The information returned includes the stream name, Amazon Resource Name (ARN), creation time, enhanced metric configuration, and shard map. The shard map is an array of shard objects. For each shard object, there is the hash key and sequence number ranges that the shard spans, and the IDs of any earlier shards that played in a role in creating the shard. Every record ingested in the stream is identified by a sequence number, which is assigned when the record is put into the stream.

You can limit the number of shards returned by each call. For more information, see Retrieving Shards from a Stream in the Amazon Kinesis Streams Developer Guide.

There are no guarantees about the chronological order shards returned. To process shards in chronological order, use the ID of the parent shard to track the lineage to the oldest shard.

This operation has a limit of 10 transactions per second per account.

Arguments

StreamName = ::String – Required

The name of the stream to describe.

Limit = ::Int

The maximum number of shards to return in a single call. The default value is 100. If you specify a value greater than 100, at most 100 shards are returned.

ExclusiveStartShardId = ::String

The shard ID of the shard to start with.

Returns

DescribeStreamOutput

Exceptions

ResourceNotFoundException or LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.disable_enhanced_monitoring — Function.

disable_enhanced_monitoring(::AWSConfig, arguments::Dict)
disable_enhanced_monitoring([::AWSConfig]; StreamName=, ShardLevelMetrics=)

DisableEnhancedMonitoring Operation

Disables enhanced monitoring.

Arguments

StreamName = ::String – Required

The name of the Amazon Kinesis stream for which to disable enhanced monitoring.

ShardLevelMetrics = ["IncomingBytes", "IncomingRecords", "OutgoingBytes", "OutgoingRecords", "WriteProvisionedThroughputExceeded", "ReadProvisionedThroughputExceeded", "IteratorAgeMilliseconds" or "ALL", ...] – Required

List of shard-level metrics to disable.

The following are the valid shard-level metrics. The value "ALL" disables every metric.

IncomingBytes

IncomingRecords

OutgoingBytes

OutgoingRecords

WriteProvisionedThroughputExceeded

ReadProvisionedThroughputExceeded

IteratorAgeMilliseconds

ALL

For more information, see Monitoring the Amazon Kinesis Streams Service with Amazon CloudWatch in the Amazon Kinesis Streams Developer Guide.

Returns

EnhancedMonitoringOutput

Exceptions

InvalidArgumentException, LimitExceededException, ResourceInUseException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonKinesis.enable_enhanced_monitoring — Function.

enable_enhanced_monitoring(::AWSConfig, arguments::Dict)
enable_enhanced_monitoring([::AWSConfig]; StreamName=, ShardLevelMetrics=)

EnableEnhancedMonitoring Operation

Enables enhanced Amazon Kinesis stream monitoring for shard-level metrics.

Arguments

StreamName = ::String – Required

The name of the stream for which to enable enhanced monitoring.

ShardLevelMetrics = ["IncomingBytes", "IncomingRecords", "OutgoingBytes", "OutgoingRecords", "WriteProvisionedThroughputExceeded", "ReadProvisionedThroughputExceeded", "IteratorAgeMilliseconds" or "ALL", ...] – Required

List of shard-level metrics to enable.

The following are the valid shard-level metrics. The value "ALL" enables every metric.

IncomingBytes

IncomingRecords

OutgoingBytes

OutgoingRecords

WriteProvisionedThroughputExceeded

ReadProvisionedThroughputExceeded

IteratorAgeMilliseconds

ALL

For more information, see Monitoring the Amazon Kinesis Streams Service with Amazon CloudWatch in the Amazon Kinesis Streams Developer Guide.

Returns

EnhancedMonitoringOutput

Exceptions

InvalidArgumentException, LimitExceededException, ResourceInUseException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonKinesis.get_records — Function.

get_records(::AWSConfig, arguments::Dict)
get_records([::AWSConfig]; ShardIterator=, <keyword arguments>)

GetRecords Operation

Gets data records from an Amazon Kinesis stream's shard.

Specify a shard iterator using the ShardIterator parameter. The shard iterator specifies the position in the shard from which you want to start reading data records sequentially. If there are no records available in the portion of the shard that the iterator points to, GetRecords returns an empty list. Note that it might take multiple calls to get to a portion of the shard that contains records.

You can scale by provisioning multiple shards per stream while considering service limits (for more information, see Streams Limits in the Amazon Kinesis Streams Developer Guide). Your application should have one thread per shard, each reading continuously from its stream. To read from a stream continually, call GetRecords in a loop. Use GetShardIterator to get the shard iterator to specify in the first GetRecords call. GetRecords returns a new shard iterator in NextShardIterator. Specify the shard iterator returned in NextShardIterator in subsequent calls to GetRecords. Note that if the shard has been closed, the shard iterator can't return more data and GetRecords returns null in NextShardIterator. You can terminate the loop when the shard is closed, or when the shard iterator reaches the record with the sequence number or other attribute that marks it as the last record to process.

Each data record can be up to 1 MB in size, and each shard can read up to 2 MB per second. You can ensure that your calls don't exceed the maximum supported size or throughput by using the Limit parameter to specify the maximum number of records that GetRecords can return. Consider your average record size when determining this limit.

The size of the data returned by GetRecords varies depending on the utilization of the shard. The maximum size of data that GetRecords can return is 10 MB. If a call returns this amount of data, subsequent calls made within the next 5 seconds throw ProvisionedThroughputExceededException. If there is insufficient provisioned throughput on the shard, subsequent calls made within the next 1 second throw ProvisionedThroughputExceededException. Note that GetRecords won't return any data when it throws an exception. For this reason, we recommend that you wait one second between calls to GetRecords; however, it's possible that the application will get exceptions for longer than 1 second.

To detect whether the application is falling behind in processing, you can use the MillisBehindLatest response attribute. You can also monitor the stream using CloudWatch metrics and other mechanisms (see Monitoring in the Amazon Kinesis Streams Developer Guide).

Each Amazon Kinesis record includes a value, ApproximateArrivalTimestamp, that is set when a stream successfully receives and stores a record. This is commonly referred to as a server-side timestamp, whereas a client-side timestamp is set when a data producer creates or sends the record to a stream (a data producer is any data source putting data records into a stream, for example with PutRecords). The timestamp has millisecond precision. There are no guarantees about the timestamp accuracy, or that the timestamp is always increasing. For example, records in a shard or across a stream might have timestamps that are out of order.

Arguments

ShardIterator = ::String – Required

The position in the shard from which you want to start sequentially reading data records. A shard iterator specifies this position using the sequence number of a data record in the shard.

Limit = ::Int

The maximum number of records to return. Specify a value of up to 10,000. If you specify a value that is greater than 10,000, GetRecords throws InvalidArgumentException.

Returns

GetRecordsOutput

Exceptions

ResourceNotFoundException, InvalidArgumentException, ProvisionedThroughputExceededException, ExpiredIteratorException, KMSDisabledException, KMSInvalidStateException, KMSAccessDeniedException, KMSNotFoundException, KMSOptInRequired or KMSThrottlingException.

See also: AWS API Documentation

AmazonKinesis.get_shard_iterator — Function.

get_shard_iterator(::AWSConfig, arguments::Dict)
get_shard_iterator([::AWSConfig]; StreamName=, ShardId=, ShardIteratorType=, <keyword arguments>)

GetShardIterator Operation

Gets an Amazon Kinesis shard iterator. A shard iterator expires five minutes after it is returned to the requester.

A shard iterator specifies the shard position from which to start reading data records sequentially. The position is specified using the sequence number of a data record in a shard. A sequence number is the identifier associated with every record ingested in the stream, and is assigned when a record is put into the stream. Each stream has one or more shards.

You must specify the shard iterator type. For example, you can set the ShardIteratorType parameter to read exactly from the position denoted by a specific sequence number by using the AT_SEQUENCE_NUMBER shard iterator type, or right after the sequence number by using the AFTER_SEQUENCE_NUMBER shard iterator type, using sequence numbers returned by earlier calls to PutRecord, PutRecords, GetRecords, or DescribeStream. In the request, you can specify the shard iterator type AT_TIMESTAMP to read records from an arbitrary point in time, TRIM_HORIZON to cause ShardIterator to point to the last untrimmed record in the shard in the system (the oldest data record in the shard), or LATEST so that you always read the most recent data in the shard.

When you read repeatedly from a stream, use a GetShardIterator request to get the first shard iterator for use in your first GetRecords request and for subsequent reads use the shard iterator returned by the GetRecords request in NextShardIterator. A new shard iterator is returned by every GetRecords request in NextShardIterator, which you use in the ShardIterator parameter of the next GetRecords request.

If a GetShardIterator request is made too often, you receive a ProvisionedThroughputExceededException. For more information about throughput limits, see GetRecords, and Streams Limits in the Amazon Kinesis Streams Developer Guide.

If the shard is closed, GetShardIterator returns a valid iterator for the last sequence number of the shard. Note that a shard can be closed as a result of using SplitShard or MergeShards.

GetShardIterator has a limit of 5 transactions per second per account per open shard.

Arguments

StreamName = ::String – Required

The name of the Amazon Kinesis stream.

ShardId = ::String – Required

The shard ID of the Amazon Kinesis shard to get the iterator for.

ShardIteratorType = "AT_SEQUENCE_NUMBER", "AFTER_SEQUENCE_NUMBER", "TRIM_HORIZON", "LATEST" or "AT_TIMESTAMP" – Required

Determines how the shard iterator is used to start reading data records from the shard.

The following are the valid Amazon Kinesis shard iterator types:

AT_SEQUENCE_NUMBER - Start reading from the position denoted by a specific sequence number, provided in the value StartingSequenceNumber.

AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number, provided in the value StartingSequenceNumber.

AT_TIMESTAMP - Start reading from the position denoted by a specific timestamp, provided in the value Timestamp.

TRIM_HORIZON - Start reading at the last untrimmed record in the shard in the system, which is the oldest data record in the shard.

LATEST - Start reading just after the most recent record in the shard, so that you always read the most recent data in the shard.

StartingSequenceNumber = ::String

The sequence number of the data record in the shard from which to start reading. Used with shard iterator type AT_SEQUENCE_NUMBER and AFTER_SEQUENCE_NUMBER.

Timestamp = timestamp

The timestamp of the data record from which to start reading. Used with shard iterator type AT_TIMESTAMP. A timestamp is the Unix epoch date with precision in milliseconds. For example, 2016-04-04T19:58:46.480-00:00 or 1459799926.480. If a record with this exact timestamp does not exist, the iterator returned is for the next (later) record. If the timestamp is older than the current trim horizon, the iterator returned is for the oldest untrimmed data record (TRIM_HORIZON).

Returns

GetShardIteratorOutput

Exceptions

ResourceNotFoundException, InvalidArgumentException or ProvisionedThroughputExceededException.

See also: AWS API Documentation

AmazonKinesis.increase_stream_retention_period — Function.

increase_stream_retention_period(::AWSConfig, arguments::Dict)
increase_stream_retention_period([::AWSConfig]; StreamName=, RetentionPeriodHours=)

IncreaseStreamRetentionPeriod Operation

Increases the Amazon Kinesis stream's retention period, which is the length of time data records are accessible after they are added to the stream. The maximum value of a stream's retention period is 168 hours (7 days).

Upon choosing a longer stream retention period, this operation will increase the time period records are accessible that have not yet expired. However, it will not make previous data that has expired (older than the stream's previous retention period) accessible after the operation has been called. For example, if a stream's retention period is set to 24 hours and is increased to 168 hours, any data that is older than 24 hours will remain inaccessible to consumer applications.

Arguments

StreamName = ::String – Required

The name of the stream to modify.

RetentionPeriodHours = ::Int – Required

The new retention period of the stream, in hours. Must be more than the current retention period.

Exceptions

ResourceInUseException, ResourceNotFoundException or InvalidArgumentException.

See also: AWS API Documentation

AmazonKinesis.list_streams — Function.

list_streams(::AWSConfig, arguments::Dict)
list_streams([::AWSConfig]; <keyword arguments>)

ListStreams Operation

Lists your Amazon Kinesis streams.

The number of streams may be too large to return from a single call to ListStreams. You can limit the number of returned streams using the Limit parameter. If you do not specify a value for the Limit parameter, Amazon Kinesis uses the default limit, which is currently 10.

You can detect if there are more streams available to list by using the HasMoreStreams flag from the returned output. If there are more streams available, you can request more streams by using the name of the last stream returned by the ListStreams request in the ExclusiveStartStreamName parameter in a subsequent request to ListStreams. The group of stream names returned by the subsequent request is then added to the list. You can continue this process until all the stream names have been collected in the list.

ListStreams has a limit of 5 transactions per second per account.

Arguments

Limit = ::Int

The maximum number of streams to list.

ExclusiveStartStreamName = ::String

The name of the stream to start the list with.

Returns

ListStreamsOutput

Exceptions

LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.list_tags_for_stream — Function.

list_tags_for_stream(::AWSConfig, arguments::Dict)
list_tags_for_stream([::AWSConfig]; StreamName=, <keyword arguments>)

ListTagsForStream Operation

Lists the tags for the specified Amazon Kinesis stream.

Arguments

StreamName = ::String – Required

The name of the stream.

ExclusiveStartTagKey = ::String

The key to use as the starting point for the list of tags. If this parameter is set, ListTagsForStream gets all tags that occur after ExclusiveStartTagKey.

Limit = ::Int

The number of tags to return. If this number is less than the total number of tags associated with the stream, HasMoreTags is set to true. To list additional tags, set ExclusiveStartTagKey to the last key in the response.

Returns

ListTagsForStreamOutput

Exceptions

ResourceNotFoundException, InvalidArgumentException or LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.merge_shards — Function.

merge_shards(::AWSConfig, arguments::Dict)
merge_shards([::AWSConfig]; StreamName=, ShardToMerge=, AdjacentShardToMerge=)

MergeShards Operation

Merges two adjacent shards in an Amazon Kinesis stream and combines them into a single shard to reduce the stream's capacity to ingest and transport data. Two shards are considered adjacent if the union of the hash key ranges for the two shards form a contiguous set with no gaps. For example, if you have two shards, one with a hash key range of 276...381 and the other with a hash key range of 382...454, then you could merge these two shards into a single shard that would have a hash key range of 276...454. After the merge, the single child shard receives data for all hash key values covered by the two parent shards.

MergeShards is called when there is a need to reduce the overall capacity of a stream because of excess capacity that is not being used. You must specify the shard to be merged and the adjacent shard for a stream. For more information about merging shards, see Merge Two Shards in the Amazon Kinesis Streams Developer Guide.

If the stream is in the ACTIVE state, you can call MergeShards. If a stream is in the CREATING, UPDATING, or DELETING state, MergeShards returns a ResourceInUseException. If the specified stream does not exist, MergeShards returns a ResourceNotFoundException.

You can use DescribeStream to check the state of the stream, which is returned in StreamStatus.

MergeShards is an asynchronous operation. Upon receiving a MergeShards request, Amazon Kinesis immediately returns a response and sets the StreamStatus to UPDATING. After the operation is completed, Amazon Kinesis sets the StreamStatus to ACTIVE. Read and write operations continue to work while the stream is in the UPDATING state.

You use DescribeStream to determine the shard IDs that are specified in the MergeShards request.

If you try to operate on too many streams in parallel using CreateStream, DeleteStream, MergeShards or SplitShard, you will receive a LimitExceededException.

MergeShards has limit of 5 transactions per second per account.

Arguments

StreamName = ::String – Required

The name of the stream for the merge.

ShardToMerge = ::String – Required

The shard ID of the shard to combine with the adjacent shard for the merge.

AdjacentShardToMerge = ::String – Required

The shard ID of the adjacent shard for the merge.

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.put_record — Function.

put_record(::AWSConfig, arguments::Dict)
put_record([::AWSConfig]; StreamName=, Data=, PartitionKey=, <keyword arguments>)

PutRecord Operation

Writes a single data record into an Amazon Kinesis stream. Call PutRecord to send data into the stream for real-time ingestion and subsequent processing, one record at a time. Each shard can support writes up to 1,000 records per second, up to a maximum data write total of 1 MB per second.

You must specify the name of the stream that captures, stores, and transports the data; a partition key; and the data blob itself.

The data blob can be any type of data; for example, a segment from a log file, geographic/location data, website clickstream data, and so on.

The partition key is used by Amazon Kinesis to distribute data across shards. Amazon Kinesis segregates the data records that belong to a stream into multiple shards, using the partition key associated with each data record to determine which shard a given data record belongs to.

Partition keys are Unicode strings, with a maximum length limit of 256 characters for each key. An MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards using the hash key ranges of the shards. You can override hashing the partition key to determine the shard by explicitly specifying a hash value using the ExplicitHashKey parameter. For more information, see Adding Data to a Stream in the Amazon Kinesis Streams Developer Guide.

PutRecord returns the shard ID of where the data record was placed and the sequence number that was assigned to the data record.

Sequence numbers increase over time and are specific to a shard within a stream, not across all shards within a stream. To guarantee strictly increasing ordering, write serially to a shard and use the SequenceNumberForOrdering parameter. For more information, see Adding Data to a Stream in the Amazon Kinesis Streams Developer Guide.

If a PutRecord request cannot be processed because of insufficient provisioned throughput on the shard involved in the request, PutRecord throws ProvisionedThroughputExceededException.

By default, data records are accessible for 24 hours from the time that they are added to a stream. You can use IncreaseStreamRetentionPeriod or DecreaseStreamRetentionPeriod to modify this retention period.

Arguments

StreamName = ::String – Required

The name of the stream to put the data record into.

Data = blob – Required

The data blob to put into the record, which is base64-encoded when the blob is serialized. When the data blob (the payload before base64-encoding) is added to the partition key size, the total size must not exceed the maximum record size (1 MB).

PartitionKey = ::String – Required

Determines which shard in the stream the data record is assigned to. Partition keys are Unicode strings with a maximum length limit of 256 characters for each key. Amazon Kinesis uses the partition key as input to a hash function that maps the partition key and associated data to a specific shard. Specifically, an MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key map to the same shard within the stream.

ExplicitHashKey = ::String

The hash value used to explicitly determine the shard the data record is assigned to by overriding the partition key hash.

SequenceNumberForOrdering = ::String

Guarantees strictly increasing sequence numbers, for puts from the same client and to the same partition key. Usage: set the SequenceNumberForOrdering of record n to the sequence number of record n-1 (as returned in the result when putting record n-1). If this parameter is not set, records will be coarsely ordered based on arrival time.

Returns

PutRecordOutput

Exceptions

ResourceNotFoundException, InvalidArgumentException, ProvisionedThroughputExceededException, KMSDisabledException, KMSInvalidStateException, KMSAccessDeniedException, KMSNotFoundException, KMSOptInRequired or KMSThrottlingException.

See also: AWS API Documentation

AmazonKinesis.put_records — Function.

put_records(::AWSConfig, arguments::Dict)
put_records([::AWSConfig]; Records=, StreamName=)

PutRecords Operation

Writes multiple data records into an Amazon Kinesis stream in a single call (also referred to as a PutRecords request). Use this operation to send data into the stream for data ingestion and processing.

Each PutRecords request can support up to 500 records. Each record in the request can be as large as 1 MB, up to a limit of 5 MB for the entire request, including partition keys. Each shard can support writes up to 1,000 records per second, up to a maximum data write total of 1 MB per second.

You must specify the name of the stream that captures, stores, and transports the data; and an array of request Records, with each record in the array requiring a partition key and data blob. The record size limit applies to the total size of the partition key and data blob.

The data blob can be any type of data; for example, a segment from a log file, geographic/location data, website clickstream data, and so on.

The partition key is used by Amazon Kinesis as input to a hash function that maps the partition key and associated data to a specific shard. An MD5 hash function is used to map partition keys to 128-bit integer values and to map associated data records to shards. As a result of this hashing mechanism, all data records with the same partition key map to the same shard within the stream. For more information, see Adding Data to a Stream in the Amazon Kinesis Streams Developer Guide.

Each record in the Records array may include an optional parameter, ExplicitHashKey, which overrides the partition key to shard mapping. This parameter allows a data producer to determine explicitly the shard where the record is stored. For more information, see Adding Multiple Records with PutRecords in the Amazon Kinesis Streams Developer Guide.

The PutRecords response includes an array of response Records. Each record in the response array directly correlates with a record in the request array using natural ordering, from the top to the bottom of the request and response. The response Records array always includes the same number of records as the request array.

The response Records array includes both successfully and unsuccessfully processed records. Amazon Kinesis attempts to process all records in each PutRecords request. A single record failure does not stop the processing of subsequent records.

A successfully-processed record includes ShardId and SequenceNumber values. The ShardId parameter identifies the shard in the stream where the record is stored. The SequenceNumber parameter is an identifier assigned to the put record, unique to all records in the stream.

An unsuccessfully-processed record includes ErrorCode and ErrorMessage values. ErrorCode reflects the type of error and can be one of the following values: ProvisionedThroughputExceededException or InternalFailure. ErrorMessage provides more detailed information about the ProvisionedThroughputExceededException exception including the account ID, stream name, and shard ID of the record that was throttled. For more information about partially successful responses, see Adding Multiple Records with PutRecords in the Amazon Kinesis Streams Developer Guide.

By default, data records are accessible for 24 hours from the time that they are added to a stream. You can use IncreaseStreamRetentionPeriod or DecreaseStreamRetentionPeriod to modify this retention period.

Arguments

Records = [[ ... ], ...] – Required

The records associated with the request.

 Records = [[
        "Data" => <required> blob,
        "ExplicitHashKey" =>  ::String,
        "PartitionKey" => <required> ::String
    ], ...]

StreamName = ::String – Required

The stream name associated with the request.

Returns

PutRecordsOutput

Exceptions

ResourceNotFoundException, InvalidArgumentException, ProvisionedThroughputExceededException, KMSDisabledException, KMSInvalidStateException, KMSAccessDeniedException, KMSNotFoundException, KMSOptInRequired or KMSThrottlingException.

See also: AWS API Documentation

AmazonKinesis.remove_tags_from_stream — Function.

remove_tags_from_stream(::AWSConfig, arguments::Dict)
remove_tags_from_stream([::AWSConfig]; StreamName=, TagKeys=)

RemoveTagsFromStream Operation

Removes tags from the specified Amazon Kinesis stream. Removed tags are deleted and cannot be recovered after this operation successfully completes.

If you specify a tag that does not exist, it is ignored.

Arguments

StreamName = ::String – Required

The name of the stream.

TagKeys = [::String, ...] – Required

A list of tag keys. Each corresponding tag is removed from the stream.

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.split_shard — Function.

split_shard(::AWSConfig, arguments::Dict)
split_shard([::AWSConfig]; StreamName=, ShardToSplit=, NewStartingHashKey=)

SplitShard Operation

Splits a shard into two new shards in the Amazon Kinesis stream to increase the stream's capacity to ingest and transport data. SplitShard is called when there is a need to increase the overall capacity of a stream because of an expected increase in the volume of data records being ingested.

You can also use SplitShard when a shard appears to be approaching its maximum utilization; for example, the producers sending data into the specific shard are suddenly sending more than previously anticipated. You can also call SplitShard to increase stream capacity, so that more Amazon Kinesis applications can simultaneously read data from the stream for real-time processing.

You must specify the shard to be split and the new hash key, which is the position in the shard where the shard gets split in two. In many cases, the new hash key might simply be the average of the beginning and ending hash key, but it can be any hash key value in the range being mapped into the shard. For more information about splitting shards, see Split a Shard in the Amazon Kinesis Streams Developer Guide.

You can use DescribeStream to determine the shard ID and hash key values for the ShardToSplit and NewStartingHashKey parameters that are specified in the SplitShard request.

SplitShard is an asynchronous operation. Upon receiving a SplitShard request, Amazon Kinesis immediately returns a response and sets the stream status to UPDATING. After the operation is completed, Amazon Kinesis sets the stream status to ACTIVE. Read and write operations continue to work while the stream is in the UPDATING state.

You can use DescribeStream to check the status of the stream, which is returned in StreamStatus. If the stream is in the ACTIVE state, you can call SplitShard. If a stream is in CREATING or UPDATING or DELETING states, DescribeStream returns a ResourceInUseException.

If the specified stream does not exist, DescribeStream returns a ResourceNotFoundException. If you try to create more shards than are authorized for your account, you receive a LimitExceededException.

For the default shard limit for an AWS account, see Streams Limits in the Amazon Kinesis Streams Developer Guide. If you need to increase this limit, contact AWS Support.

If you try to operate on too many streams simultaneously using CreateStream, DeleteStream, MergeShards, and/or SplitShard, you receive a LimitExceededException.

SplitShard has limit of 5 transactions per second per account.

Arguments

StreamName = ::String – Required

The name of the stream for the shard split.

ShardToSplit = ::String – Required

The shard ID of the shard to split.

NewStartingHashKey = ::String – Required

A hash key value for the starting hash key of one of the child shards created by the split. The hash key range for a given shard constitutes a set of ordered contiguous positive integers. The value for NewStartingHashKey must be in the range of hash keys being mapped into the shard. The NewStartingHashKey hash key value and all higher hash key values in hash key range are distributed to one of the child shards. All the lower hash key values in the range are distributed to the other child shard.

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidArgumentException or LimitExceededException.

See also: AWS API Documentation

AmazonKinesis.start_stream_encryption — Function.

start_stream_encryption(::AWSConfig, arguments::Dict)
start_stream_encryption([::AWSConfig]; StreamName=, EncryptionType=, KeyId=)

StartStreamEncryption Operation

Enables or updates server-side encryption using an AWS KMS key for a specified stream.

Starting encryption is an asynchronous operation. Upon receiving the request, Amazon Kinesis returns immediately and sets the status of the stream to UPDATING. After the update is complete, Amazon Kinesis sets the status of the stream back to ACTIVE. Updating or applying encryption normally takes a few seconds to complete but it can take minutes. You can continue to read and write data to your stream while its status is UPDATING. Once the status of the stream is ACTIVE, records written to the stream will begin to be encrypted.

API Limits: You can successfully apply a new AWS KMS key for server-side encryption 25 times in a rolling 24 hour period.

Note: It can take up to 5 seconds after the stream is in an ACTIVE status before all records written to the stream are encrypted. After you’ve enabled encryption, you can verify encryption was applied by inspecting the API response from PutRecord or PutRecords.

Arguments

StreamName = ::String – Required

The name of the stream for which to start encrypting records.

EncryptionType = "NONE" or "KMS" – Required

The encryption type to use. This parameter can be one of the following values:

NONE: Not valid for this operation. An InvalidOperationException will be thrown.

KMS: Use server-side encryption on the records in the stream using a customer-managed KMS key.

KeyId = ::String – Required

The GUID for the customer-managed KMS key to use for encryption. You can also use a Kinesis-owned master key by specifying the alias aws/kinesis.

Exceptions

InvalidArgumentException, LimitExceededException, ResourceInUseException, ResourceNotFoundException, KMSDisabledException, KMSInvalidStateException, KMSAccessDeniedException, KMSNotFoundException, KMSOptInRequired or KMSThrottlingException.

See also: AWS API Documentation

AmazonKinesis.stop_stream_encryption — Function.

stop_stream_encryption(::AWSConfig, arguments::Dict)
stop_stream_encryption([::AWSConfig]; StreamName=, EncryptionType=, KeyId=)

StopStreamEncryption Operation

Disables server-side encryption for a specified stream.

Stopping encryption is an asynchronous operation. Upon receiving the request, Amazon Kinesis returns immediately and sets the status of the stream to UPDATING. After the update is complete, Amazon Kinesis sets the status of the stream back to ACTIVE. Stopping encryption normally takes a few seconds to complete but it can take minutes. You can continue to read and write data to your stream while its status is UPDATING. Once the status of the stream is ACTIVE records written to the stream will no longer be encrypted by the Amazon Kinesis Streams service.

API Limits: You can successfully disable server-side encryption 25 times in a rolling 24 hour period.

Note: It can take up to 5 seconds after the stream is in an ACTIVE status before all records written to the stream are no longer subject to encryption. After you’ve disabled encryption, you can verify encryption was not applied by inspecting the API response from PutRecord or PutRecords.

Arguments

StreamName = ::String – Required

The name of the stream on which to stop encrypting records.

EncryptionType = "NONE" or "KMS" – Required

The encryption type. This parameter can be one of the following values:

NONE: Not valid for this operation. An InvalidOperationException will be thrown.

KMS: Use server-side encryption on the records in the stream using a customer-managed KMS key.

KeyId = ::String – Required

The GUID for the customer-managed key that was used for encryption.

Exceptions

InvalidArgumentException, LimitExceededException, ResourceInUseException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonKinesis.update_shard_count — Function.

update_shard_count(::AWSConfig, arguments::Dict)
update_shard_count([::AWSConfig]; StreamName=, TargetShardCount=, ScalingType=)

UpdateShardCount Operation

Updates the shard count of the specified stream to the specified number of shards.

Updating the shard count is an asynchronous operation. Upon receiving the request, Amazon Kinesis returns immediately and sets the status of the stream to UPDATING. After the update is complete, Amazon Kinesis sets the status of the stream back to ACTIVE. Depending on the size of the stream, the scaling action could take a few minutes to complete. You can continue to read and write data to your stream while its status is UPDATING.

To update the shard count, Amazon Kinesis performs splits or merges on individual shards. This can cause short-lived shards to be created, in addition to the final shards. We recommend that you double or halve the shard count, as this results in the fewest number of splits or merges.

This operation has the following limits, which are per region per account unless otherwise noted:

scale more than twice per rolling 24 hour period

scale up above double your current shard count

scale down below half your current shard count

scale up above 200 shards in a stream

scale a stream with more than 200 shards down unless the result is less than 200 shards

scale up above the shard limits for your account

For the default limits for an AWS account, see Streams Limits in the Amazon Kinesis Streams Developer Guide. If you need to increase a limit, contact AWS Support.

Arguments

StreamName = ::String – Required

The name of the stream.

TargetShardCount = ::Int – Required

The new number of shards.

ScalingType = "UNIFORM_SCALING" – Required

The scaling type. Uniform scaling creates shards of equal size.

Returns

UpdateShardCountOutput

Exceptions

InvalidArgumentException, LimitExceededException, ResourceInUseException or ResourceNotFoundException.

See also: AWS API Documentation