AmazonLexModelBuildingService.jl

AmazonLexModelBuildingService

Amazon Lex Build-Time Actions

Amazon Lex is an AWS service for building conversational voice and text interfaces. Use these actions to create, update, and delete conversational bots for new and existing client applications.

This document is generated from apis/lex-models-2017-04-19.normal.json. See JuliaCloud/AWSCore.jl.

AmazonLexModelBuildingService.create_bot_version

AmazonLexModelBuildingService.create_intent_version

AmazonLexModelBuildingService.create_slot_type_version

AmazonLexModelBuildingService.delete_bot

AmazonLexModelBuildingService.delete_bot_alias

AmazonLexModelBuildingService.delete_bot_channel_association

AmazonLexModelBuildingService.delete_bot_version

AmazonLexModelBuildingService.delete_intent

AmazonLexModelBuildingService.delete_intent_version

AmazonLexModelBuildingService.delete_slot_type

AmazonLexModelBuildingService.delete_slot_type_version

AmazonLexModelBuildingService.delete_utterances

AmazonLexModelBuildingService.get_bot

AmazonLexModelBuildingService.get_bot_alias

AmazonLexModelBuildingService.get_bot_aliases

AmazonLexModelBuildingService.get_bot_channel_association

AmazonLexModelBuildingService.get_bot_channel_associations

AmazonLexModelBuildingService.get_bot_versions

AmazonLexModelBuildingService.get_bots

AmazonLexModelBuildingService.get_builtin_intent

AmazonLexModelBuildingService.get_builtin_intents

AmazonLexModelBuildingService.get_builtin_slot_types

AmazonLexModelBuildingService.get_intent

AmazonLexModelBuildingService.get_intent_versions

AmazonLexModelBuildingService.get_intents

AmazonLexModelBuildingService.get_slot_type

AmazonLexModelBuildingService.get_slot_type_versions

AmazonLexModelBuildingService.get_slot_types

AmazonLexModelBuildingService.get_utterances_view

AmazonLexModelBuildingService.put_bot

AmazonLexModelBuildingService.put_bot_alias

AmazonLexModelBuildingService.put_intent

AmazonLexModelBuildingService.put_slot_type

AmazonLexModelBuildingService.create_bot_version — Function.

create_bot_version(::AWSConfig, arguments::Dict)
create_bot_version([::AWSConfig]; name=, <keyword arguments>)

CreateBotVersion Operation

Creates a new version of the bot based on the $LATEST version. If the $LATEST version of this resource hasn't changed since you created the last version, Amazon Lex doesn't create a new version. It returns the last created version.

Note

You can update only the $LATEST version of the bot. You can't update the numbered versions that you create with the CreateBotVersion operation.

When you create the first version of a bot, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see versioning-intro.

This operation requires permission for the lex:CreateBotVersion action.

Arguments

name = ::String – Required

The name of the bot that you want to create a new version of. The name is case sensitive.

checksum = ::String

Identifies a specific revision of the $LATEST version of the bot. If you specify a checksum and the $LATEST version of the bot has a different checksum, a PreconditionFailedException exception is returned and Amazon Lex doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

Returns

CreateBotVersionResponse

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or PreconditionFailedException.

See also: AWS API Documentation

AmazonLexModelBuildingService.create_intent_version — Function.

create_intent_version(::AWSConfig, arguments::Dict)
create_intent_version([::AWSConfig]; name=, <keyword arguments>)

CreateIntentVersion Operation

Creates a new version of an intent based on the $LATEST version of the intent. If the $LATEST version of this intent hasn't changed since you last updated it, Amazon Lex doesn't create a new version. It returns the last version you created.

Note

You can update only the $LATEST version of the intent. You can't update the numbered versions that you create with the CreateIntentVersion operation.

When you create a version of an intent, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see versioning-intro.

This operation requires permissions to perform the lex:CreateIntentVersion action.

Arguments

name = ::String – Required

The name of the intent that you want to create a new version of. The name is case sensitive.

checksum = ::String

Checksum of the $LATEST version of the intent that should be used to create the new version. If you specify a checksum and the $LATEST version of the intent has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

Returns

CreateIntentVersionResponse

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or PreconditionFailedException.

See also: AWS API Documentation

AmazonLexModelBuildingService.create_slot_type_version — Function.

create_slot_type_version(::AWSConfig, arguments::Dict)
create_slot_type_version([::AWSConfig]; name=, <keyword arguments>)

CreateSlotTypeVersion Operation

Creates a new version of a slot type based on the $LATEST version of the specified slot type. If the $LATEST version of this resource has not changed since the last version that you created, Amazon Lex doesn't create a new version. It returns the last version that you created.

Note

You can update only the $LATEST version of a slot type. You can't update the numbered versions that you create with the CreateSlotTypeVersion operation.

When you create a version of a slot type, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see versioning-intro.

This operation requires permissions for the lex:CreateSlotTypeVersion action.

Arguments

name = ::String – Required

The name of the slot type that you want to create a new version for. The name is case sensitive.

checksum = ::String

Checksum for the $LATEST version of the slot type that you want to publish. If you specify a checksum and the $LATEST version of the slot type has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish the new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

Returns

CreateSlotTypeVersionResponse

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or PreconditionFailedException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_bot — Function.

delete_bot(::AWSConfig, arguments::Dict)
delete_bot([::AWSConfig]; name=)

DeleteBot Operation

Deletes all versions of the bot, including the $LATEST version. To delete a specific version of the bot, use the DeleteBotVersion operation.

If a bot has an alias, you can't delete it. Instead, the DeleteBot operation returns a ResourceInUseException exception that includes a reference to the alias that refers to the bot. To remove the reference to the bot, delete the alias. If you get the same exception again, delete the referring alias until the DeleteBot operation is successful.

This operation requires permissions for the lex:DeleteBot action.

Arguments

name = ::String – Required

The name of the bot. The name is case sensitive.

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or ResourceInUseException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_bot_alias — Function.

delete_bot_alias(::AWSConfig, arguments::Dict)
delete_bot_alias([::AWSConfig]; name=, botName=)

DeleteBotAlias Operation

Deletes an alias for the specified bot.

You can't delete an alias that is used in the association between a bot and a messaging channel. If an alias is used in a channel association, the DeleteBot operation returns a ResourceInUseException exception that includes a reference to the channel association that refers to the bot. You can remove the reference to the alias by deleting the channel association. If you get the same exception again, delete the referring association until the DeleteBotAlias operation is successful.

Arguments

name = ::String – Required

The name of the alias to delete. The name is case sensitive.

botName = ::String – Required

The name of the bot that the alias points to.

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or ResourceInUseException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_bot_channel_association — Function.

delete_bot_channel_association(::AWSConfig, arguments::Dict)
delete_bot_channel_association([::AWSConfig]; name=, botName=, aliasName=)

DeleteBotChannelAssociation Operation

Deletes the association between an Amazon Lex bot and a messaging platform.

This operation requires permission for the lex:DeleteBotChannelAssociation action.

Arguments

name = ::String – Required

The name of the association. The name is case sensitive.

botName = ::String – Required

The name of the Amazon Lex bot.

aliasName = ::String – Required

An alias that points to the specific version of the Amazon Lex bot to which this association is being made.

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_bot_version — Function.

delete_bot_version(::AWSConfig, arguments::Dict)
delete_bot_version([::AWSConfig]; name=, version=)

DeleteBotVersion Operation

Deletes a specific version of a bot. To delete all versions of a bot, use the DeleteBot operation.

This operation requires permissions for the lex:DeleteBotVersion action.

Arguments

name = ::String – Required

The name of the bot.

version = ::String – Required

The version of the bot to delete. You cannot delete the $LATEST version of the bot. To delete the $LATEST version, use the DeleteBot operation.

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or ResourceInUseException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_intent — Function.

delete_intent(::AWSConfig, arguments::Dict)
delete_intent([::AWSConfig]; name=)

DeleteIntent Operation

Deletes all versions of the intent, including the $LATEST version. To delete a specific version of the intent, use the DeleteIntentVersion operation.

You can delete a version of an intent only if it is not referenced. To delete an intent that is referred to in one or more bots (see how-it-works), you must remove those references first.

Note

If you get the ResourceInUseException exception, it provides an example reference that shows where the intent is referenced. To remove the reference to the intent, either update the bot or delete it. If you get the same exception when you attempt to delete the intent again, repeat until the intent has no references and the call to DeleteIntent is successful.

This operation requires permission for the lex:DeleteIntent action.

Arguments

name = ::String – Required

The name of the intent. The name is case sensitive.

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or ResourceInUseException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_intent_version — Function.

delete_intent_version(::AWSConfig, arguments::Dict)
delete_intent_version([::AWSConfig]; name=, version=)

DeleteIntentVersion Operation

Deletes a specific version of an intent. To delete all versions of a intent, use the DeleteIntent operation.

This operation requires permissions for the lex:DeleteIntentVersion action.

Arguments

name = ::String – Required

The name of the intent.

version = ::String – Required

The version of the intent to delete. You cannot delete the $LATEST version of the intent. To delete the $LATEST version, use the DeleteIntent operation.

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or ResourceInUseException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_slot_type — Function.

delete_slot_type(::AWSConfig, arguments::Dict)
delete_slot_type([::AWSConfig]; name=)

DeleteSlotType Operation

Deletes all versions of the slot type, including the $LATEST version. To delete a specific version of the slot type, use the DeleteSlotTypeVersion operation.

You can delete a version of a slot type only if it is not referenced. To delete a slot type that is referred to in one or more intents, you must remove those references first.

Note

If you get the ResourceInUseException exception, the exception provides an example reference that shows the intent where the slot type is referenced. To remove the reference to the slot type, either update the intent or delete it. If you get the same exception when you attempt to delete the slot type again, repeat until the slot type has no references and the DeleteSlotType call is successful.

This operation requires permission for the lex:DeleteSlotType action.

Arguments

name = ::String – Required

The name of the slot type. The name is case sensitive.

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or ResourceInUseException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_slot_type_version — Function.

delete_slot_type_version(::AWSConfig, arguments::Dict)
delete_slot_type_version([::AWSConfig]; name=, version=)

DeleteSlotTypeVersion Operation

Deletes a specific version of a slot type. To delete all versions of a slot type, use the DeleteSlotType operation.

This operation requires permissions for the lex:DeleteSlotTypeVersion action.

Arguments

name = ::String – Required

The name of the slot type.

version = ::String – Required

The version of the slot type to delete. You cannot delete the $LATEST version of the slot type. To delete the $LATEST version, use the DeleteSlotType operation.

Exceptions

NotFoundException, ConflictException, LimitExceededException, InternalFailureException, BadRequestException or ResourceInUseException.

See also: AWS API Documentation

AmazonLexModelBuildingService.delete_utterances — Function.

delete_utterances(::AWSConfig, arguments::Dict)
delete_utterances([::AWSConfig]; botName=, userId=)

DeleteUtterances Operation

Deletes stored utterances.

Amazon Lex stores the utterances that users send to your bot unless the childDirected field in the bot is set to true. Utterances are stored for 15 days for use with the GetUtterancesView operation, and then stored indefinately for use in improving the ability of your bot to respond to user input.

Use the DeleteStoredUtterances operation to manually delete stored utterances for a specific user.

This operation requires permissions for the lex:DeleteUtterances action.

Arguments

botName = ::String – Required

The name of the bot that stored the utterances.

userId = ::String – Required

The unique identifier for the user that made the utterances. This is the user ID that was sent in the PostContent or PostText operation request that contained the utterance.

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_bot — Function.

get_bot(::AWSConfig, arguments::Dict)
get_bot([::AWSConfig]; name=, versionoralias=)

GetBot Operation

Returns metadata information for a specific bot. You must provide the bot name and the bot version or alias.

The GetBot operation requires permissions for the lex:GetBot action.

Arguments

name = ::String – Required

The name of the bot. The name is case sensitive.

versionoralias = ::String – Required

The version or alias of the bot.

Returns

GetBotResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_bot_alias — Function.

get_bot_alias(::AWSConfig, arguments::Dict)
get_bot_alias([::AWSConfig]; name=, botName=)

GetBotAlias Operation

Returns information about an Amazon Lex bot alias. For more information about aliases, see versioning-aliases.

This operation requires permissions for the lex:GetBotAlias action.

Arguments

name = ::String – Required

The name of the bot alias. The name is case sensitive.

botName = ::String – Required

The name of the bot.

Returns

GetBotAliasResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_bot_aliases — Function.

get_bot_aliases(::AWSConfig, arguments::Dict)
get_bot_aliases([::AWSConfig]; botName=, <keyword arguments>)

GetBotAliases Operation

Returns a list of aliases for a specified Amazon Lex bot.

This operation requires permissions for the lex:GetBotAliases action.

Arguments

botName = ::String – Required

The name of the bot.

nextToken = ::String

A pagination token for fetching the next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of aliases to return in the response. The default is 50. .

nameContains = ::String

Substring to match in bot alias names. An alias will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

Returns

GetBotAliasesResponse

Exceptions

LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_bot_channel_association — Function.

get_bot_channel_association(::AWSConfig, arguments::Dict)
get_bot_channel_association([::AWSConfig]; name=, botName=, aliasName=)

GetBotChannelAssociation Operation

Returns information about the association between an Amazon Lex bot and a messaging platform.

This operation requires permissions for the lex:GetBotChannelAssociation action.

Arguments

name = ::String – Required

The name of the association between the bot and the channel. The name is case sensitive.

botName = ::String – Required

The name of the Amazon Lex bot.

aliasName = ::String – Required

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

Returns

GetBotChannelAssociationResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_bot_channel_associations — Function.

get_bot_channel_associations(::AWSConfig, arguments::Dict)
get_bot_channel_associations([::AWSConfig]; botName=, aliasName=, <keyword arguments>)

GetBotChannelAssociations Operation

Returns a list of all of the channels associated with the specified bot.

The GetBotChannelAssociations operation requires permissions for the lex:GetBotChannelAssociations action.

Arguments

botName = ::String – Required

The name of the Amazon Lex bot in the association.

aliasName = ::String – Required

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

nextToken = ::String

A pagination token for fetching the next page of associations. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of associations, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of associations to return in the response. The default is 50.

nameContains = ::String

Substring to match in channel association names. An association will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz." To return all bot channel associations, use a hyphen ("-") as the nameContains parameter.

Returns

GetBotChannelAssociationsResponse

Exceptions

LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_bot_versions — Function.

get_bot_versions(::AWSConfig, arguments::Dict)
get_bot_versions([::AWSConfig]; name=, <keyword arguments>)

GetBotVersions Operation

Gets information about all of the versions of a bot.

The GetBotVersions operation returns a BotMetadata object for each version of a bot. For example, if a bot has three numbered versions, the GetBotVersions operation returns four BotMetadata objects in the response, one for each numbered version and one for the $LATEST version.

The GetBotVersions operation always returns at least one version, the $LATEST version.

This operation requires permissions for the lex:GetBotVersions action.

Arguments

name = ::String – Required

The name of the bot for which versions should be returned.

nextToken = ::String

A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of bot versions to return in the response. The default is 10.

Returns

GetBotVersionsResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_bots — Function.

get_bots(::AWSConfig, arguments::Dict)
get_bots([::AWSConfig]; <keyword arguments>)

GetBots Operation

Returns bot information as follows:

If you provide the nameContains field, the response includes information for the $LATEST version of all bots whose name contains the specified string.

If you don't specify the nameContains field, the operation returns information about the $LATEST version of all of your bots.

This operation requires permission for the lex:GetBots action.

Arguments

nextToken = ::String

A pagination token that fetches the next page of bots. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of bots, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of bots to return in the response that the request will return. The default is 10.

nameContains = ::String

Substring to match in bot names. A bot will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

Returns

GetBotsResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_builtin_intent — Function.

get_builtin_intent(::AWSConfig, arguments::Dict)
get_builtin_intent([::AWSConfig]; signature=)

GetBuiltinIntent Operation

Returns information about a built-in intent.

This operation requires permission for the lex:GetBuiltinIntent action.

Arguments

signature = ::String – Required

The unique identifier for a built-in intent. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

Returns

GetBuiltinIntentResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_builtin_intents — Function.

get_builtin_intents(::AWSConfig, arguments::Dict)
get_builtin_intents([::AWSConfig]; <keyword arguments>)

GetBuiltinIntents Operation

Gets a list of built-in intents that meet the specified criteria.

This operation requires permission for the lex:GetBuiltinIntents action.

Arguments

locale = "en-US"

A list of locales that the intent supports.

signatureContains = ::String

Substring to match in built-in intent signatures. An intent will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz." To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

nextToken = ::String

A pagination token that fetches the next page of intents. If this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, use the pagination token in the next request.

maxResults = ::Int

The maximum number of intents to return in the response. The default is 10.

Returns

GetBuiltinIntentsResponse

Exceptions

LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_builtin_slot_types — Function.

get_builtin_slot_types(::AWSConfig, arguments::Dict)
get_builtin_slot_types([::AWSConfig]; <keyword arguments>)

GetBuiltinSlotTypes Operation

Gets a list of built-in slot types that meet the specified criteria.

For a list of built-in slot types, see Slot Type Reference in the Alexa Skills Kit.

This operation requires permission for the lex:GetBuiltInSlotTypes action.

Arguments

locale = "en-US"

A list of locales that the slot type supports.

signatureContains = ::String

Substring to match in built-in slot type signatures. A slot type will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

nextToken = ::String

A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of slot types, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of slot types to return in the response. The default is 10.

Returns

GetBuiltinSlotTypesResponse

Exceptions

LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_intent — Function.

get_intent(::AWSConfig, arguments::Dict)
get_intent([::AWSConfig]; name=, version=)

GetIntent Operation

Returns information about an intent. In addition to the intent name, you must specify the intent version.

This operation requires permissions to perform the lex:GetIntent action.

Arguments

name = ::String – Required

The name of the intent. The name is case sensitive.

version = ::String – Required

The version of the intent.

Returns

GetIntentResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_intent_versions — Function.

get_intent_versions(::AWSConfig, arguments::Dict)
get_intent_versions([::AWSConfig]; name=, <keyword arguments>)

GetIntentVersions Operation

Gets information about all of the versions of an intent.

The GetIntentVersions operation returns an IntentMetadata object for each version of an intent. For example, if an intent has three numbered versions, the GetIntentVersions operation returns four IntentMetadata objects in the response, one for each numbered version and one for the $LATEST version.

The GetIntentVersions operation always returns at least one version, the $LATEST version.

This operation requires permissions for the lex:GetIntentVersions action.

Arguments

name = ::String – Required

The name of the intent for which versions should be returned.

nextToken = ::String

A pagination token for fetching the next page of intent versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of intent versions to return in the response. The default is 10.

Returns

GetIntentVersionsResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_intents — Function.

get_intents(::AWSConfig, arguments::Dict)
get_intents([::AWSConfig]; <keyword arguments>)

GetIntents Operation

Returns intent information as follows:

If you specify the nameContains field, returns the $LATEST version of all intents that contain the specified string.

If you don't specify the nameContains field, returns information about the $LATEST version of all intents.

The operation requires permission for the lex:GetIntents action.

Arguments

nextToken = ::String

A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of intents to return in the response. The default is 10.

nameContains = ::String

Substring to match in intent names. An intent will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

Returns

GetIntentsResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_slot_type — Function.

get_slot_type(::AWSConfig, arguments::Dict)
get_slot_type([::AWSConfig]; name=, version=)

GetSlotType Operation

Returns information about a specific version of a slot type. In addition to specifying the slot type name, you must specify the slot type version.

This operation requires permissions for the lex:GetSlotType action.

Arguments

name = ::String – Required

The name of the slot type. The name is case sensitive.

version = ::String – Required

The version of the slot type.

Returns

GetSlotTypeResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_slot_type_versions — Function.

get_slot_type_versions(::AWSConfig, arguments::Dict)
get_slot_type_versions([::AWSConfig]; name=, <keyword arguments>)

GetSlotTypeVersions Operation

Gets information about all versions of a slot type.

The GetSlotTypeVersions operation returns a SlotTypeMetadata object for each version of a slot type. For example, if a slot type has three numbered versions, the GetSlotTypeVersions operation returns four SlotTypeMetadata objects in the response, one for each numbered version and one for the $LATEST version.

The GetSlotTypeVersions operation always returns at least one version, the $LATEST version.

This operation requires permissions for the lex:GetSlotTypeVersions action.

Arguments

name = ::String – Required

The name of the slot type for which versions should be returned.

nextToken = ::String

A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of slot type versions to return in the response. The default is 10.

Returns

GetSlotTypeVersionsResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_slot_types — Function.

get_slot_types(::AWSConfig, arguments::Dict)
get_slot_types([::AWSConfig]; <keyword arguments>)

GetSlotTypes Operation

Returns slot type information as follows:

If you specify the nameContains field, returns the $LATEST version of all slot types that contain the specified string.

If you don't specify the nameContains field, returns information about the $LATEST version of all slot types.

The operation requires permission for the lex:GetSlotTypes action.

Arguments

nextToken = ::String

A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch next page of slot types, specify the pagination token in the next request.

maxResults = ::Int

The maximum number of slot types to return in the response. The default is 10.

nameContains = ::String

Substring to match in slot type names. A slot type will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

Returns

GetSlotTypesResponse

Exceptions

NotFoundException, LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.get_utterances_view — Function.

get_utterances_view(::AWSConfig, arguments::Dict)
get_utterances_view([::AWSConfig]; botname=, bot_versions=, status_type=)

GetUtterancesView Operation

Use the GetUtterancesView operation to get information about the utterances that your users have made to your bot. You can use this list to tune the utterances that your bot responds to.

For example, say that you have created a bot to order flowers. After your users have used your bot for a while, use the GetUtterancesView operation to see the requests that they have made and whether they have been successful. You might find that the utterance "I want flowers" is not being recognized. You could add this utterance to the OrderFlowers intent so that your bot recognizes that utterance.

After you publish a new version of a bot, you can get information about the old version and the new so that you can compare the performance across the two versions.

Data is available for the last 15 days. You can request information for up to 5 versions in each request. The response contains information about a maximum of 100 utterances for each version.

If the bot's childDirected field is set to true, utterances for the bot are not stored and cannot be retrieved with the GetUtterancesView operation. For more information, see PutBot.

This operation requires permissions for the lex:GetUtterancesView action.

Arguments

botname = ::String – Required

The name of the bot for which utterance information should be returned.

bot_versions = [::String, ...] – Required

An array of bot versions for which utterance information should be returned. The limit is 5 versions per request.

status_type = "Detected" or "Missed" – Required

To return utterances that were recognized and handled, useDetected. To return utterances that were not recognized, use Missed.

Returns

GetUtterancesViewResponse

Exceptions

LimitExceededException, InternalFailureException or BadRequestException.

See also: AWS API Documentation

AmazonLexModelBuildingService.put_bot — Function.

put_bot(::AWSConfig, arguments::Dict)
put_bot([::AWSConfig]; name=, locale=, childDirected=, <keyword arguments>)

PutBot Operation

Creates an Amazon Lex conversational bot or replaces an existing bot. When you create or update a bot you are only required to specify a name. You can use this to add intents later, or to remove intents from an existing bot. When you create a bot with a name only, the bot is created or updated but Amazon Lex returns the responseFAILED. You can build the bot after you add one or more intents. For more information about Amazon Lex bots, see [how-it-works](@ref).

If you specify the name of an existing bot, the fields in the request replace the existing values in theLATESTversion of the bot. Amazon Lex removes any fields that you don't provide values for in the request, except for theidleTTLInSecondsandprivacySettings` fields, which are set to their default values. If you don't specify values for required fields, Amazon Lex throws an exception.

This operation requires permissions for the lex:PutBot action. For more information, see auth-and-access-control.`

Arguments

name = ::String – Required

The name of the bot. The name is not case sensitive.

description = ::String

A description of the bot.

intents = [[ ... ], ...]

An array of Intent objects. Each intent represents a command that a user can express. For example, a pizza ordering bot might support an OrderPizza intent. For more information, see how-it-works.

 intents = [[
        "intentName" => <required> ::String,
        "intentVersion" => <required> ::String
    ], ...]

clarificationPrompt = [ ... ]

When Amazon Lex doesn't understand the user's intent, it uses this message to get clarification. To specify how many times Amazon Lex should repeate the clarification prompt, use the maxAttempts field. If Amazon Lex still doesn't understand, it sends the message in the abortStatement field.

When you create a clarification prompt, make sure that it suggests the correct response from the user. for example, for a bot that orders pizza and drinks, you might create this clarification prompt: "What would you like to do? You can say 'Order a pizza' or 'Order a drink.'"

 clarificationPrompt = [
        "messages" => <required> [[
            "contentType" => <required> "PlainText" or "SSML",
            "content" => <required> ::String
        ], ...],
        "maxAttempts" => <required> ::Int,
        "responseCard" =>  ::String
    ]

abortStatement = [ ... ]

When Amazon Lex can't understand the user's input in context, it tries to elicit the information a few times. After that, Amazon Lex sends the message defined in abortStatement to the user, and then aborts the conversation. To set the number of retries, use the valueElicitationPrompt field for the slot type.

For example, in a pizza ordering bot, Amazon Lex might ask a user "What type of crust would you like?" If the user's response is not one of the expected responses (for example, "thin crust, "deep dish," etc.), Amazon Lex tries to elicit a correct response a few more times.

For example, in a pizza ordering application, OrderPizza might be one of the intents. This intent might require the CrustType slot. You specify the valueElicitationPrompt field when you create the CrustType slot.

 abortStatement = [
        "messages" => <required> [[
            "contentType" => <required> "PlainText" or "SSML",
            "content" => <required> ::String
        ], ...],
        "responseCard" =>  ::String
    ]

idleSessionTTLInSeconds = ::Int

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation.

A user interaction session remains active for the amount of time specified. If no conversation occurs during this time, the session expires and Amazon Lex deletes any data provided before the timeout.

For example, suppose that a user chooses the OrderPizza intent, but gets sidetracked halfway through placing an order. If the user doesn't complete the order within the specified time, Amazon Lex discards the slot information that it gathered, and the user must start over.

If you don't include the idleSessionTTLInSeconds element in a PutBot operation request, Amazon Lex uses the default value. This is also true if the request replaces an existing bot.

The default is 300 seconds (5 minutes).

voiceId = ::String

The Amazon Polly voice ID that you want Amazon Lex to use for voice interactions with the user. The locale configured for the voice must match the locale of the bot. For more information, see Available Voices in the Amazon Polly Developer Guide.

checksum = ::String

Identifies a specific revision of the $LATEST version.

When you create a new bot, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

processBehavior = "SAVE" or "BUILD"

If you set the processBehavior element to Build, Amazon Lex builds the bot so that it can be run. If you set the element to SaveAmazon Lex saves the bot, but doesn't build it.

If you don't specify this value, the default value is Save.

locale = "en-US" – Required

Specifies the target locale for the bot. Any intent used in the bot must be compatible with the locale of the bot.

The default is en-US.

childDirected = ::Bool – Required

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

Returns

PutBotResponse

Exceptions

ConflictException, LimitExceededException, InternalFailureException, BadRequestException or PreconditionFailedException.

See also: AWS API Documentation

AmazonLexModelBuildingService.put_bot_alias — Function.

put_bot_alias(::AWSConfig, arguments::Dict)
put_bot_alias([::AWSConfig]; name=, botVersion=, botName=, <keyword arguments>)

PutBotAlias Operation

Creates an alias for the specified version of the bot or replaces an alias for the specified bot. To change the version of the bot that the alias points to, replace the alias. For more information about aliases, see versioning-aliases.

This operation requires permissions for the lex:PutBotAlias action.

Arguments

name = ::String – Required

The name of the alias. The name is not case sensitive.

description = ::String

A description of the alias.

botVersion = ::String – Required

The version of the bot.

botName = ::String – Required

The name of the bot.

checksum = ::String

Identifies a specific revision of the $LATEST version.

When you create a new bot alias, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot alias, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

Returns

PutBotAliasResponse

Exceptions

ConflictException, LimitExceededException, InternalFailureException, BadRequestException or PreconditionFailedException.

See also: AWS API Documentation

AmazonLexModelBuildingService.put_intent — Function.

put_intent(::AWSConfig, arguments::Dict)
put_intent([::AWSConfig]; name=, <keyword arguments>)

PutIntent Operation

Creates an intent or replaces an existing intent.

To define the interaction between the user and your bot, you use one or more intents. For a pizza ordering bot, for example, you would create an OrderPizza intent.

To create an intent or replace an existing intent, you must provide the following:

Intent name. For example, OrderPizza.

Sample utterances. For example, "Can I order a pizza, please." and "I want to order a pizza."

Information to be gathered. You specify slot types for the information that your bot will request from the user. You can specify standard slot types, such as a date or a time, or custom slot types such as the size and crust of a pizza.

How the intent will be fulfilled. You can provide a Lambda function or configure the intent to return the intent information to the client application. If you use a Lambda function, when all of the intent information is available, Amazon Lex invokes your Lambda function. If you configure your intent to return the intent information to the client application.

You can specify other optional information in the request, such as:

A confirmation prompt to ask the user to confirm an intent. For example, "Shall I order your pizza?"

A conclusion statement to send to the user after the intent has been fulfilled. For example, "I placed your pizza order."

A follow-up prompt that asks the user for additional activity. For example, asking "Do you want to order a drink with your pizza?"

If you specify an existing intent name to update the intent, Amazon Lex replaces the values in the $LATEST version of the slot type with the values in the request. Amazon Lex removes fields that you don't provide in the request. If you don't specify the required fields, Amazon Lex throws an exception.

For more information, see how-it-works.

This operation requires permissions for the lex:PutIntent action.

Arguments

name = ::String – Required

The name of the intent. The name is not case sensitive.

The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent, you can't create a custom intent called HelpIntent.

For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit.

description = ::String

A description of the intent.

slots = [[ ... ], ...]

An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see <xref linkend="how-it-works"/>.

 slots = [[
        "name" => <required> ::String,
        "description" =>  ::String,
        "slotConstraint" => <required> "Required" or "Optional",
        "slotType" =>  ::String,
        "slotTypeVersion" =>  ::String,
        "valueElicitationPrompt" =>  [
            "messages" => <required> [[
                "contentType" => <required> "PlainText" or "SSML",
                "content" => <required> ::String
            ], ...],
            "maxAttempts" => <required> ::Int,
            "responseCard" =>  ::String
        ],
        "priority" =>  ::Int,
        "sampleUtterances" =>  [::String, ...],
        "responseCard" =>  ::String
    ], ...]

sampleUtterances = [::String, ...]

An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas".

In each utterance, a slot name is enclosed in curly braces.

confirmationPrompt = [ ... ]

Prompts the user to confirm the intent. This question should have a yes or no answer.

Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

Note

You you must provide both the rejectionStatement and the confirmationPrompt, or neither.

 confirmationPrompt = [
        "messages" => <required> [[
            "contentType" => <required> "PlainText" or "SSML",
            "content" => <required> ::String
        ], ...],
        "maxAttempts" => <required> ::Int,
        "responseCard" =>  ::String
    ]

rejectionStatement = [ ... ]

When the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

Note

You must provide both the rejectionStatement and the confirmationPrompt, or neither.

 rejectionStatement = [
        "messages" => <required> [[
            "contentType" => <required> "PlainText" or "SSML",
            "content" => <required> ::String
        ], ...],
        "responseCard" =>  ::String
    ]

followUpPrompt = [ ... ]

Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink.

The action that Amazon Lex takes depends on the user's response, as follows:

If the user says "Yes" it responds with the clarification prompt that is configured for the bot.

if the user says "Yes" and continues with an utterance that triggers an intent it starts a conversation for the intent.

If the user says "No" it responds with the rejection statement configured for the the follow-up prompt.

If it doesn't recognize the utterance it repeats the follow-up prompt again.

The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

 followUpPrompt = [
        "prompt" => <required> [
            "messages" => <required> [[
                "contentType" => <required> "PlainText" or "SSML",
                "content" => <required> ::String
            ], ...],
            "maxAttempts" => <required> ::Int,
            "responseCard" =>  ::String
        ],
        "rejectionStatement" => <required> [
            "messages" => <required> [[
                "contentType" => <required> "PlainText" or "SSML",
                "content" => <required> ::String
            ], ...],
            "responseCard" =>  ::String
        ]
    ]

conclusionStatement = [ ... ]

The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function.

This element is relevant only if you provide a Lambda function in the fulfillmentActivity. If you return the intent to the client application, you can't specify this element.

Note

The followUpPrompt and conclusionStatement are mutually exclusive. You can specify only one.

 conclusionStatement = [
        "messages" => <required> [[
            "contentType" => <required> "PlainText" or "SSML",
            "content" => <required> ::String
        ], ...],
        "responseCard" =>  ::String
    ]

dialogCodeHook = [ ... ]

Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction.

For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant, to true. You might find John's phone number and set the corresponding session attribute.

 dialogCodeHook = [
        "uri" => <required> ::String,
        "messageVersion" => <required> ::String
    ]

fulfillmentActivity = [ ... ]

Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store.

You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

 fulfillmentActivity = [
        "type" => <required> "ReturnIntent" or "CodeHook",
        "codeHook" =>  [
            "uri" => <required> ::String,
            "messageVersion" => <required> ::String
        ]
    ]

parentIntentSignature = ::String

A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

checksum = ::String

Identifies a specific revision of the $LATEST version.

When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a intent, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

Returns

PutIntentResponse

Exceptions

ConflictException, LimitExceededException, InternalFailureException, BadRequestException or PreconditionFailedException.

See also: AWS API Documentation

AmazonLexModelBuildingService.put_slot_type — Function.

put_slot_type(::AWSConfig, arguments::Dict)
put_slot_type([::AWSConfig]; name=, <keyword arguments>)

PutSlotType Operation

Creates a custom slot type or replaces an existing custom slot type.

To create a custom slot type, specify a name for the slot type and a set of enumeration values, which are the values that a slot of this type can assume. For more information, see how-it-works.

If you specify the name of an existing slot type, the fields in the request replace the existing values in the $LATEST version of the slot type. Amazon Lex removes the fields that you don't provide in the request. If you don't specify required fields, Amazon Lex throws an exception.

This operation requires permissions for the lex:PutSlotType action.

Arguments

name = ::String – Required

The name of the slot type. The name is not case sensitive.

The name can't match a built-in slot type name, or a built-in slot type name with "AMAZON." removed. For example, because there is a built-in slot type called AMAZON.DATE, you can't create a custom slot type called DATE.

For a list of built-in slot types, see Slot Type Reference in the Alexa Skills Kit.

description = ::String

A description of the slot type.

enumerationValues = [["value" => <required> ::String], ...]

A list of EnumerationValue objects that defines the values that the slot type can take.

checksum = ::String

Identifies a specific revision of the $LATEST version.

When you create a new slot type, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a slot type, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

Returns

PutSlotTypeResponse

Exceptions

ConflictException, LimitExceededException, InternalFailureException, BadRequestException or PreconditionFailedException.

See also: AWS API Documentation