AmazonLightsail.jl

AmazonLightsail

Amazon Lightsail is the easiest way to get started with AWS for developers who just need virtual private servers. Lightsail includes everything you need to launch your project quickly - a virtual machine, SSD-based storage, data transfer, DNS management, and a static IP - for a low, predictable price. You manage those Lightsail servers through the Lightsail console or by using the API or command-line interface (CLI).

For more information about Lightsail concepts and tasks, see the Lightsail Dev Guide.

To use the Lightsail API or the CLI, you will need to use AWS Identity and Access Management (IAM) to generate access keys. For details about how to set this up, see the Lightsail Dev Guide.

This document is generated from apis/lightsail-2016-11-28.normal.json. See JuliaCloud/AWSCore.jl.

AmazonLightsail.allocate_static_ip

AmazonLightsail.attach_static_ip

AmazonLightsail.close_instance_public_ports

AmazonLightsail.create_domain

AmazonLightsail.create_domain_entry

AmazonLightsail.create_instance_snapshot

AmazonLightsail.create_instances

AmazonLightsail.create_instances_from_snapshot

AmazonLightsail.create_key_pair

AmazonLightsail.delete_domain

AmazonLightsail.delete_domain_entry

AmazonLightsail.delete_instance

AmazonLightsail.delete_instance_snapshot

AmazonLightsail.delete_key_pair

AmazonLightsail.detach_static_ip

AmazonLightsail.download_default_key_pair

AmazonLightsail.get_active_names

AmazonLightsail.get_blueprints

AmazonLightsail.get_bundles

AmazonLightsail.get_domain

AmazonLightsail.get_domains

AmazonLightsail.get_instance

AmazonLightsail.get_instance_access_details

AmazonLightsail.get_instance_metric_data

AmazonLightsail.get_instance_port_states

AmazonLightsail.get_instance_snapshot

AmazonLightsail.get_instance_snapshots

AmazonLightsail.get_instance_state

AmazonLightsail.get_instances

AmazonLightsail.get_key_pair

AmazonLightsail.get_key_pairs

AmazonLightsail.get_operation

AmazonLightsail.get_operations

AmazonLightsail.get_operations_for_resource

AmazonLightsail.get_regions

AmazonLightsail.get_static_ip

AmazonLightsail.get_static_ips

AmazonLightsail.import_key_pair

AmazonLightsail.is_vpc_peered

AmazonLightsail.open_instance_public_ports

AmazonLightsail.peer_vpc

AmazonLightsail.put_instance_public_ports

AmazonLightsail.reboot_instance

AmazonLightsail.release_static_ip

AmazonLightsail.start_instance

AmazonLightsail.stop_instance

AmazonLightsail.unpeer_vpc

AmazonLightsail.update_domain_entry

AmazonLightsail.allocate_static_ip — Method.

allocate_static_ip(::AWSConfig, arguments::Dict)
allocate_static_ip(::AWSConfig; staticIpName=)

AllocateStaticIp Operation

Allocates a static IP address.

Arguments

staticIpName::String – Required

The name of the static IP address.

Returns

AllocateStaticIpResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.attach_static_ip — Method.

attach_static_ip(::AWSConfig, arguments::Dict)
attach_static_ip(::AWSConfig; staticIpName=, instanceName=)

AttachStaticIp Operation

Attaches a static IP address to a specific Amazon Lightsail instance.

Arguments

staticIpName::String – Required

The name of the static IP.

instanceName::String – Required

The instance name to which you want to attach the static IP address.

Returns

AttachStaticIpResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.close_instance_public_ports — Method.

close_instance_public_ports(::AWSConfig, arguments::Dict)
close_instance_public_ports(::AWSConfig; portInfo=, instanceName=)

CloseInstancePublicPorts Operation

Closes the public ports on a specific Amazon Lightsail instance.

Arguments

portInfo: – Required

Information about the public port you are trying to close.

    Dict(
        fromPort  => ::Int,
        toPort  => ::Int,
        protocol  => : "tcp", "all" or "udp"
    )

instanceName::String – Required

The name of the instance on which you're attempting to close the public ports.

Returns

CloseInstancePublicPortsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.create_domain — Method.

create_domain(::AWSConfig, arguments::Dict)
create_domain(::AWSConfig; domainName=)

CreateDomain Operation

Creates a domain resource for the specified domain (e.g., example.com).

Arguments

domainName::String – Required

The domain name to manage (e.g., example.com).

Note

You cannot register a new domain name using Lightsail. You must register a domain name using Amazon Route 53 or another domain name registrar. If you have already registered your domain, you can enter its name in this parameter to manage the DNS records for that domain.

Returns

CreateDomainResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.create_domain_entry — Method.

create_domain_entry(::AWSConfig, arguments::Dict)
create_domain_entry(::AWSConfig; domainName=, domainEntry=)

CreateDomainEntry Operation

Creates one of the following entry records associated with the domain: A record, CNAME record, TXT record, or MX record.

Arguments

domainName::String – Required

The domain name (e.g., example.com) for which you want to create the domain entry.

domainEntry: – Required

An array of key-value pairs containing information about the domain entry request.

    Dict(
        id  => ::String,
        name  => ::String,
        target  => ::String,
        type  => ::String,
        options  => ::Dict{String,String}
    )

Returns

CreateDomainEntryResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.create_instance_snapshot — Method.

create_instance_snapshot(::AWSConfig, arguments::Dict)
create_instance_snapshot(::AWSConfig; instanceSnapshotName=, instanceName=)

CreateInstanceSnapshot Operation

Creates a snapshot of a specific virtual private server, or instance. You can use a snapshot to create a new instance that is based on that snapshot.

Arguments

instanceSnapshotName::String – Required

The name for your new snapshot.

instanceName::String – Required

The Lightsail instance on which to base your snapshot.

Returns

CreateInstanceSnapshotResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.create_instances — Method.

create_instances(::AWSConfig, arguments::Dict)
create_instances(::AWSConfig; instanceNames=, availabilityZone=, blueprintId=, bundleId=, <keyword arguments>)

CreateInstances Operation

Creates one or more Amazon Lightsail virtual private servers, or instances.

Arguments

instanceNames: [::String...] – Required

The names to use for your new Lightsail instances. Separate multiple values using quotation marks and commas, for example: ["MyFirstInstance","MySecondInstance"]

availabilityZone::String – Required

The Availability Zone in which to create your instance. Use the following format: us-east-1a (case sensitive). You can get a list of availability zones by using the get regions operation. Be sure to add the include availability zones parameter to your request.

customImageName::String

(Deprecated) The name for your custom image.

Note

In releases prior to June 12, 2017, this parameter was ignored by the API. It is now deprecated.

blueprintId::String – Required

The ID for a virtual private server image (e.g., app_wordpress_4_4 or app_lamp_7_0). Use the get blueprints operation to return a list of available images (or blueprints).

bundleId::String – Required

The bundle of specification information for your virtual private server (or instance), including the pricing plan (e.g., micro_1_0).

userData::String

A launch script you can create that configures a server with additional user data. For example, you might want to run apt-get –y update.

Note

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg. For a complete list, see the Dev Guide.

keyPairName::String

The name of your key pair.

Returns

CreateInstancesResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.create_instances_from_snapshot — Method.

create_instances_from_snapshot(::AWSConfig, arguments::Dict)
create_instances_from_snapshot(::AWSConfig; instanceNames=, availabilityZone=, instanceSnapshotName=, bundleId=, <keyword arguments>)

CreateInstancesFromSnapshot Operation

Uses a specific snapshot as a blueprint for creating one or more new instances that are based on that identical configuration.

Arguments

instanceNames: [::String...] – Required

The names for your new instances.

availabilityZone::String – Required

The Availability Zone where you want to create your instances. Use the following formatting: us-east-1a (case sensitive). You can get a list of availability zones by using the get regions operation. Be sure to add the include availability zones parameter to your request.

instanceSnapshotName::String – Required

The name of the instance snapshot on which you are basing your new instances. Use the get instance snapshots operation to return information about your existing snapshots.

bundleId::String – Required

The bundle of specification information for your virtual private server (or instance), including the pricing plan (e.g., micro_1_0).

userData::String

You can create a launch script that configures a server with additional user data. For example, apt-get –y update.

Note

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg. For a complete list, see the Dev Guide.

keyPairName::String

The name for your key pair.

Returns

CreateInstancesFromSnapshotResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.create_key_pair — Method.

create_key_pair(::AWSConfig, arguments::Dict)
create_key_pair(::AWSConfig; keyPairName=)

CreateKeyPair Operation

Creates sn SSH key pair.

Arguments

keyPairName::String – Required

The name for your new key pair.

Returns

CreateKeyPairResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.delete_domain — Method.

delete_domain(::AWSConfig, arguments::Dict)
delete_domain(::AWSConfig; domainName=)

DeleteDomain Operation

Deletes the specified domain recordset and all of its domain records.

Arguments

domainName::String – Required

The specific domain name to delete.

Returns

DeleteDomainResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.delete_domain_entry — Method.

delete_domain_entry(::AWSConfig, arguments::Dict)
delete_domain_entry(::AWSConfig; domainName=, domainEntry=)

DeleteDomainEntry Operation

Deletes a specific domain entry.

Arguments

domainName::String – Required

The name of the domain entry to delete.

domainEntry: – Required

An array of key-value pairs containing information about your domain entries.

    Dict(
        id  => ::String,
        name  => ::String,
        target  => ::String,
        type  => ::String,
        options  => ::Dict{String,String}
    )

Returns

DeleteDomainEntryResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.delete_instance — Method.

delete_instance(::AWSConfig, arguments::Dict)
delete_instance(::AWSConfig; instanceName=)

DeleteInstance Operation

Deletes a specific Amazon Lightsail virtual private server, or instance.

Arguments

instanceName::String – Required

The name of the instance to delete.

Returns

DeleteInstanceResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.delete_instance_snapshot — Method.

delete_instance_snapshot(::AWSConfig, arguments::Dict)
delete_instance_snapshot(::AWSConfig; instanceSnapshotName=)

DeleteInstanceSnapshot Operation

Deletes a specific snapshot of a virtual private server (or instance).

Arguments

instanceSnapshotName::String – Required

The name of the snapshot to delete.

Returns

DeleteInstanceSnapshotResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.delete_key_pair — Method.

delete_key_pair(::AWSConfig, arguments::Dict)
delete_key_pair(::AWSConfig; keyPairName=)

DeleteKeyPair Operation

Deletes a specific SSH key pair.

Arguments

keyPairName::String – Required

The name of the key pair to delete.

Returns

DeleteKeyPairResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.detach_static_ip — Method.

detach_static_ip(::AWSConfig, arguments::Dict)
detach_static_ip(::AWSConfig; staticIpName=)

DetachStaticIp Operation

Detaches a static IP from the Amazon Lightsail instance to which it is attached.

Arguments

staticIpName::String – Required

The name of the static IP to detach from the instance.

Returns

DetachStaticIpResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.download_default_key_pair — Method.

download_default_key_pair(::AWSConfig, arguments::Dict)
download_default_key_pair(::AWSConfig; )

DownloadDefaultKeyPair Operation

Downloads the default SSH key pair from the user's account.

Arguments

Returns

DownloadDefaultKeyPairResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_active_names — Method.

get_active_names(::AWSConfig, arguments::Dict)
get_active_names(::AWSConfig; <keyword arguments>)

GetActiveNames Operation

Returns the names of all active (not deleted) resources.

Arguments

pageToken::String

A token used for paginating results from your get active names request.

Returns

GetActiveNamesResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_blueprints — Method.

get_blueprints(::AWSConfig, arguments::Dict)
get_blueprints(::AWSConfig; <keyword arguments>)

GetBlueprints Operation

Returns the list of available instance images, or blueprints. You can use a blueprint to create a new virtual private server already running a specific operating system, as well as a preinstalled app or development stack. The software each instance is running depends on the blueprint image you choose.

Arguments

includeInactive::Bool

A Boolean value indicating whether to include inactive results in your request.

pageToken::String

A token used for advancing to the next page of results from your get blueprints request.

Returns

GetBlueprintsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_bundles — Method.

get_bundles(::AWSConfig, arguments::Dict)
get_bundles(::AWSConfig; <keyword arguments>)

GetBundles Operation

Returns the list of bundles that are available for purchase. A bundle describes the specs for your virtual private server (or instance).

Arguments

includeInactive::Bool

A Boolean value that indicates whether to include inactive bundle results in your request.

pageToken::String

A token used for advancing to the next page of results from your get bundles request.

Returns

GetBundlesResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_domain — Method.

get_domain(::AWSConfig, arguments::Dict)
get_domain(::AWSConfig; domainName=)

GetDomain Operation

Returns information about a specific domain recordset.

Arguments

domainName::String – Required

The domain name for which your want to return information about.

Returns

GetDomainResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_domains — Method.

get_domains(::AWSConfig, arguments::Dict)
get_domains(::AWSConfig; <keyword arguments>)

GetDomains Operation

Returns a list of all domains in the user's account.

Arguments

pageToken::String

A token used for advancing to the next page of results from your get domains request.

Returns

GetDomainsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_instance — Method.

get_instance(::AWSConfig, arguments::Dict)
get_instance(::AWSConfig; instanceName=)

GetInstance Operation

Returns information about a specific Amazon Lightsail instance, which is a virtual private server.

Arguments

instanceName::String – Required

The name of the instance.

Returns

GetInstanceResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_instance_access_details — Method.

get_instance_access_details(::AWSConfig, arguments::Dict)
get_instance_access_details(::AWSConfig; instanceName=, <keyword arguments>)

GetInstanceAccessDetails Operation

Returns temporary SSH keys you can use to connect to a specific virtual private server, or instance.

Arguments

instanceName::String – Required

The name of the instance to access.

protocol: "ssh" or "rdp"

The protocol to use to connect to your instance. Defaults to ssh.

Returns

GetInstanceAccessDetailsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_instance_metric_data — Method.

get_instance_metric_data(::AWSConfig, arguments::Dict)
get_instance_metric_data(::AWSConfig; instanceName=, metricName=, period=, startTime=, endTime=, unit=, statistics=)

GetInstanceMetricData Operation

Returns the data points for the specified Amazon Lightsail instance metric, given an instance name.

Arguments

instanceName::String – Required

The name of the instance for which you want to get metrics data.

metricName: "CPUUtilization", "NetworkIn", "NetworkOut", "StatusCheckFailed", "StatusCheckFailed_Instance" or "StatusCheckFailed_System" – Required

The metric name to get data about.

period::Int – Required

The time period for which you are requesting data.

startTime: timestamp – Required

The start time of the time period.

endTime: timestamp – Required

The end time of the time period.

unit: "Seconds", "Microseconds", "Milliseconds", "Bytes", "Kilobytes", "Megabytes", "Gigabytes", "Terabytes", "Bits", "Kilobits", "Megabits", "Gigabits", "Terabits", "Percent", "Count", "Bytes/Second", "Kilobytes/Second", "Megabytes/Second", "Gigabytes/Second", "Terabytes/Second", "Bits/Second", "Kilobits/Second", "Megabits/Second", "Gigabits/Second", "Terabits/Second", "Count/Second" or "None" – Required

The unit. The list of valid values is below.

statistics: [: "Minimum", "Maximum", "Sum", "Average" or "SampleCount"...] – Required

The instance statistics.

Returns

GetInstanceMetricDataResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_instance_port_states — Method.

get_instance_port_states(::AWSConfig, arguments::Dict)
get_instance_port_states(::AWSConfig; instanceName=)

GetInstancePortStates Operation

Returns the port states for a specific virtual private server, or instance.

Arguments

instanceName::String – Required

The name of the instance.

Returns

GetInstancePortStatesResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_instance_snapshot — Method.

get_instance_snapshot(::AWSConfig, arguments::Dict)
get_instance_snapshot(::AWSConfig; instanceSnapshotName=)

GetInstanceSnapshot Operation

Returns information about a specific instance snapshot.

Arguments

instanceSnapshotName::String – Required

The name of the snapshot for which you are requesting information.

Returns

GetInstanceSnapshotResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_instance_snapshots — Method.

get_instance_snapshots(::AWSConfig, arguments::Dict)
get_instance_snapshots(::AWSConfig; <keyword arguments>)

GetInstanceSnapshots Operation

Returns all instance snapshots for the user's account.

Arguments

pageToken::String

A token used for advancing to the next page of results from your get instance snapshots request.

Returns

GetInstanceSnapshotsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_instance_state — Method.

get_instance_state(::AWSConfig, arguments::Dict)
get_instance_state(::AWSConfig; instanceName=)

GetInstanceState Operation

Returns the state of a specific instance. Works on one instance at a time.

Arguments

instanceName::String – Required

The name of the instance to get state information about.

Returns

GetInstanceStateResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_instances — Method.

get_instances(::AWSConfig, arguments::Dict)
get_instances(::AWSConfig; <keyword arguments>)

GetInstances Operation

Returns information about all Amazon Lightsail virtual private servers, or instances.

Arguments

pageToken::String

A token used for advancing to the next page of results from your get instances request.

Returns

GetInstancesResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_key_pair — Method.

get_key_pair(::AWSConfig, arguments::Dict)
get_key_pair(::AWSConfig; keyPairName=)

GetKeyPair Operation

Returns information about a specific key pair.

Arguments

keyPairName::String – Required

The name of the key pair for which you are requesting information.

Returns

GetKeyPairResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_key_pairs — Method.

get_key_pairs(::AWSConfig, arguments::Dict)
get_key_pairs(::AWSConfig; <keyword arguments>)

GetKeyPairs Operation

Returns information about all key pairs in the user's account.

Arguments

pageToken::String

A token used for advancing to the next page of results from your get key pairs request.

Returns

GetKeyPairsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_operation — Method.

get_operation(::AWSConfig, arguments::Dict)
get_operation(::AWSConfig; operationId=)

GetOperation Operation

Returns information about a specific operation. Operations include events such as when you create an instance, allocate a static IP, attach a static IP, and so on.

Arguments

operationId::String – Required

A GUID used to identify the operation.

Returns

GetOperationResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_operations — Method.

get_operations(::AWSConfig, arguments::Dict)
get_operations(::AWSConfig; <keyword arguments>)

GetOperations Operation

Returns information about all operations.

Results are returned from oldest to newest, up to a maximum of 200. Results can be paged by making each subsequent call to GetOperations use the maximum (last) statusChangedAt value from the previous request.

Arguments

pageToken::String

A token used for advancing to the next page of results from your get operations request.

Returns

GetOperationsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_operations_for_resource — Method.

get_operations_for_resource(::AWSConfig, arguments::Dict)
get_operations_for_resource(::AWSConfig; resourceName=, <keyword arguments>)

GetOperationsForResource Operation

Gets operations for a specific resource (e.g., an instance or a static IP).

Arguments

resourceName::String – Required

The name of the resource for which you are requesting information.

pageToken::String

A token used for advancing to the next page of results from your get operations for resource request.

Returns

GetOperationsForResourceResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_regions — Method.

get_regions(::AWSConfig, arguments::Dict)
get_regions(::AWSConfig; <keyword arguments>)

GetRegions Operation

Returns a list of all valid regions for Amazon Lightsail. Use the include availability zones parameter to also return the availability zones in a region.

Arguments

includeAvailabilityZones::Bool

A Boolean value indicating whether to also include Availability Zones in your get regions request. Availability Zones are indicated with a letter: e.g., us-east-1a.

Returns

GetRegionsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_static_ip — Method.

get_static_ip(::AWSConfig, arguments::Dict)
get_static_ip(::AWSConfig; staticIpName=)

GetStaticIp Operation

Returns information about a specific static IP.

Arguments

staticIpName::String – Required

The name of the static IP in Lightsail.

Returns

GetStaticIpResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.get_static_ips — Method.

get_static_ips(::AWSConfig, arguments::Dict)
get_static_ips(::AWSConfig; <keyword arguments>)

GetStaticIps Operation

Returns information about all static IPs in the user's account.

Arguments

pageToken::String

A token used for advancing to the next page of results from your get static IPs request.

Returns

GetStaticIpsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.import_key_pair — Method.

import_key_pair(::AWSConfig, arguments::Dict)
import_key_pair(::AWSConfig; keyPairName=, publicKeyBase64=)

ImportKeyPair Operation

Imports a public SSH key from a specific key pair.

Arguments

keyPairName::String – Required

The name of the key pair for which you want to import the public key.

publicKeyBase64::String – Required

A base64-encoded public key of the ssh-rsa type.

Returns

ImportKeyPairResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.is_vpc_peered — Method.

is_vpc_peered(::AWSConfig, arguments::Dict)
is_vpc_peered(::AWSConfig; )

IsVpcPeered Operation

Returns a Boolean value indicating whether your Lightsail VPC is peered.

Arguments

Returns

IsVpcPeeredResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.open_instance_public_ports — Method.

open_instance_public_ports(::AWSConfig, arguments::Dict)
open_instance_public_ports(::AWSConfig; portInfo=, instanceName=)

OpenInstancePublicPorts Operation

Adds public ports to an Amazon Lightsail instance.

Arguments

portInfo: – Required

An array of key-value pairs containing information about the port mappings.

    Dict(
        fromPort  => ::Int,
        toPort  => ::Int,
        protocol  => : "tcp", "all" or "udp"
    )

instanceName::String – Required

The name of the instance for which you want to open the public ports.

Returns

OpenInstancePublicPortsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.peer_vpc — Method.

peer_vpc(::AWSConfig, arguments::Dict)
peer_vpc(::AWSConfig; )

PeerVpc Operation

Tries to peer the Lightsail VPC with the user's default VPC.

Arguments

Returns

PeerVpcResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.put_instance_public_ports — Method.

put_instance_public_ports(::AWSConfig, arguments::Dict)
put_instance_public_ports(::AWSConfig; portInfos=, instanceName=)

PutInstancePublicPorts Operation

Sets the specified open ports for an Amazon Lightsail instance, and closes all ports for every protocol not included in the current request.

Arguments

portInfos: – Required

Specifies information about the public port(s).

    [Dict(
        fromPort  => ::Int,
        toPort  => ::Int,
        protocol  => : "tcp", "all" or "udp"
    )...]

instanceName::String – Required

The Lightsail instance name of the public port(s) you are setting.

Returns

PutInstancePublicPortsResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.reboot_instance — Method.

reboot_instance(::AWSConfig, arguments::Dict)
reboot_instance(::AWSConfig; instanceName=)

RebootInstance Operation

Restarts a specific instance. When your Amazon Lightsail instance is finished rebooting, Lightsail assigns a new public IP address. To use the same IP address after restarting, create a static IP address and attach it to the instance.

Arguments

instanceName::String – Required

The name of the instance to reboot.

Returns

RebootInstanceResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.release_static_ip — Method.

release_static_ip(::AWSConfig, arguments::Dict)
release_static_ip(::AWSConfig; staticIpName=)

ReleaseStaticIp Operation

Deletes a specific static IP from your account.

Arguments

staticIpName::String – Required

The name of the static IP to delete.

Returns

ReleaseStaticIpResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.start_instance — Method.

start_instance(::AWSConfig, arguments::Dict)
start_instance(::AWSConfig; instanceName=)

StartInstance Operation

Starts a specific Amazon Lightsail instance from a stopped state. To restart an instance, use the reboot instance operation.

Arguments

instanceName::String – Required

The name of the instance (a virtual private server) to start.

Returns

StartInstanceResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.stop_instance — Method.

stop_instance(::AWSConfig, arguments::Dict)
stop_instance(::AWSConfig; instanceName=)

StopInstance Operation

Stops a specific Amazon Lightsail instance that is currently running.

Arguments

instanceName::String – Required

The name of the instance (a virtual private server) to stop.

Returns

StopInstanceResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.unpeer_vpc — Method.

unpeer_vpc(::AWSConfig, arguments::Dict)
unpeer_vpc(::AWSConfig; )

UnpeerVpc Operation

Attempts to unpeer the Lightsail VPC from the user's default VPC.

Arguments

Returns

UnpeerVpcResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation

AmazonLightsail.update_domain_entry — Method.

update_domain_entry(::AWSConfig, arguments::Dict)
update_domain_entry(::AWSConfig; domainName=, domainEntry=)

UpdateDomainEntry Operation

Updates a domain recordset after it is created.

Arguments

domainName::String – Required

The name of the domain recordset to update.

domainEntry: – Required

An array of key-value pairs containing information about the domain entry.

    Dict(
        id  => ::String,
        name  => ::String,
        target  => ::String,
        type  => ::String,
        options  => ::Dict{String,String}
    )

Returns

UpdateDomainEntryResult

Exceptions

ServiceException, InvalidInputException, NotFoundException, OperationFailureException, AccessDeniedException, AccountSetupInProgressException or UnauthenticatedException.

See also: AWS API Documentation