AmazonMarketplaceMetering.jl

AmazonMarketplaceMetering

AWS Marketplace Metering Service

This reference provides descriptions of the low-level AWS Marketplace Metering Service API.

AWS Marketplace sellers can use this API to submit usage data for custom usage dimensions.

Submitting Metering Records

MeterUsage- Submits the metering record for a Marketplace product. MeterUsage is called from an EC2 instance.

BatchMeterUsage- Submits the metering record for a set of customers. BatchMeterUsage is called from a software-as-a-service (SaaS) application.

Accepting New Customers

ResolveCustomer- Called by a SaaS application during the registration process. When a buyer visits your website during the registration process, the buyer submits a Registration Token through the browser. The Registration Token is resolved through this API to obtain a CustomerIdentifier and Product Code.

This document is generated from apis/meteringmarketplace-2016-01-14.normal.json. See JuliaCloud/AWSCore.jl.

AmazonMarketplaceMetering.batch_meter_usage

AmazonMarketplaceMetering.meter_usage

AmazonMarketplaceMetering.resolve_customer

AmazonMarketplaceMetering.batch_meter_usage — Method.

batch_meter_usage(::AWSConfig, arguments::Dict)
batch_meter_usage(::AWSConfig; UsageRecords=, ProductCode=)

BatchMeterUsage Operation

BatchMeterUsage is called from a SaaS application listed on the AWS Marketplace to post metering records for a set of customers.

For identical requests, the API is idempotent; requests can be retried with the same records or a subset of the input records.

Every request to BatchMeterUsage is for one product. If you need to meter usage for multiple products, you must make multiple calls to BatchMeterUsage.

BatchMeterUsage can process up to 25 UsageRecords at a time.

Arguments

UsageRecords: – Required

The set of UsageRecords to submit. BatchMeterUsage accepts up to 25 UsageRecords at a time.

    [Dict(
        Timestamp * => timestamp,
        CustomerIdentifier * => ::String,
        Dimension * => ::String,
        Quantity * => ::Int
    )...]

ProductCode::String – Required

Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

Returns

BatchMeterUsageResult

Exceptions

InternalServiceErrorException, InvalidProductCodeException, InvalidUsageDimensionException, InvalidCustomerIdentifierException, TimestampOutOfBoundsException or ThrottlingException.

See also: AWS API Documentation

AmazonMarketplaceMetering.meter_usage — Method.

meter_usage(::AWSConfig, arguments::Dict)
meter_usage(::AWSConfig; ProductCode=, Timestamp=, UsageDimension=, UsageQuantity=, DryRun=)

MeterUsage Operation

API to emit metering records. For identical requests, the API is idempotent. It simply returns the metering record ID.

MeterUsage is authenticated on the buyer's AWS account, generally when running from an EC2 instance on the AWS Marketplace.

Arguments

ProductCode::String – Required

Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

Timestamp: timestamp – Required

Timestamp of the hour, recorded in UTC. The seconds and milliseconds portions of the timestamp will be ignored.

UsageDimension::String – Required

It will be one of the fcp dimension name provided during the publishing of the product.

UsageQuantity::Int – Required

Consumption value for the hour.

DryRun::Bool – Required

Checks whether you have the permissions required for the action, but does not make the request. If you have the permissions, the request returns DryRunOperation; otherwise, it returns UnauthorizedException.

Returns

MeterUsageResult

Exceptions

InternalServiceErrorException, InvalidProductCodeException, InvalidUsageDimensionException, InvalidEndpointRegionException, TimestampOutOfBoundsException, DuplicateRequestException or ThrottlingException.

See also: AWS API Documentation

AmazonMarketplaceMetering.resolve_customer — Method.

resolve_customer(::AWSConfig, arguments::Dict)
resolve_customer(::AWSConfig; RegistrationToken=)

ResolveCustomer Operation

ResolveCustomer is called by a SaaS application during the registration process. When a buyer visits your website during the registration process, the buyer submits a registration token through their browser. The registration token is resolved through this API to obtain a CustomerIdentifier and product code.

Arguments

RegistrationToken::String – Required

When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a CustomerIdentifier and product code.

Returns

ResolveCustomerResult

Exceptions

InvalidTokenException, ExpiredTokenException, ThrottlingException or InternalServiceErrorException.

See also: AWS API Documentation