AmazonPinpoint.jl

AmazonPinpoint

This document is generated from apis/pinpoint-2016-12-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonPinpoint.create_campaign

AmazonPinpoint.create_import_job

AmazonPinpoint.create_segment

AmazonPinpoint.delete_apns_channel

AmazonPinpoint.delete_apns_sandbox_channel

AmazonPinpoint.delete_campaign

AmazonPinpoint.delete_email_channel

AmazonPinpoint.delete_event_stream

AmazonPinpoint.delete_gcm_channel

AmazonPinpoint.delete_segment

AmazonPinpoint.delete_sms_channel

AmazonPinpoint.get_apns_channel

AmazonPinpoint.get_apns_sandbox_channel

AmazonPinpoint.get_application_settings

AmazonPinpoint.get_campaign

AmazonPinpoint.get_campaign_activities

AmazonPinpoint.get_campaign_version

AmazonPinpoint.get_campaign_versions

AmazonPinpoint.get_campaigns

AmazonPinpoint.get_email_channel

AmazonPinpoint.get_endpoint

AmazonPinpoint.get_event_stream

AmazonPinpoint.get_gcm_channel

AmazonPinpoint.get_import_job

AmazonPinpoint.get_import_jobs

AmazonPinpoint.get_segment

AmazonPinpoint.get_segment_import_jobs

AmazonPinpoint.get_segment_version

AmazonPinpoint.get_segment_versions

AmazonPinpoint.get_segments

AmazonPinpoint.get_sms_channel

AmazonPinpoint.put_event_stream

AmazonPinpoint.send_messages

AmazonPinpoint.update_apns_channel

AmazonPinpoint.update_apns_sandbox_channel

AmazonPinpoint.update_application_settings

AmazonPinpoint.update_campaign

AmazonPinpoint.update_email_channel

AmazonPinpoint.update_endpoint

AmazonPinpoint.update_endpoints_batch

AmazonPinpoint.update_gcm_channel

AmazonPinpoint.update_segment

AmazonPinpoint.update_sms_channel

AmazonPinpoint.create_campaign — Method.

create_campaign(::AWSConfig, arguments::Dict)
create_campaign(::AWSConfig; ApplicationId=, WriteCampaignRequest=)

CreateCampaign Operation

Creates or updates a campaign.

Arguments

ApplicationId::String – Required

WriteCampaignRequest: – Required

    Dict(
        AdditionalTreatments  => [Dict(
            MessageConfiguration  => Dict(
                APNSMessage  => Dict(
                    Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                    Body  => ::String,
                    ImageIconUrl  => ::String,
                    ImageSmallIconUrl  => ::String,
                    ImageUrl  => ::String,
                    JsonBody  => ::String,
                    MediaUrl  => ::String,
                    SilentPush  => ::Bool,
                    Title  => ::String,
                    Url  => ::String
                ),
                DefaultMessage  => Dict(
                    Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                    Body  => ::String,
                    ImageIconUrl  => ::String,
                    ImageSmallIconUrl  => ::String,
                    ImageUrl  => ::String,
                    JsonBody  => ::String,
                    MediaUrl  => ::String,
                    SilentPush  => ::Bool,
                    Title  => ::String,
                    Url  => ::String
                ),
                EmailMessage  => Dict(
                    Body  => ::String,
                    HtmlBody  => ::String,
                    Title  => ::String
                ),
                GCMMessage  => Dict(
                    Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                    Body  => ::String,
                    ImageIconUrl  => ::String,
                    ImageSmallIconUrl  => ::String,
                    ImageUrl  => ::String,
                    JsonBody  => ::String,
                    MediaUrl  => ::String,
                    SilentPush  => ::Bool,
                    Title  => ::String,
                    Url  => ::String
                ),
                SMSMessage  => Dict(
                    Body  => ::String,
                    MessageType  => : "TRANSACTIONAL" or "PROMOTIONAL",
                    SenderId  => ::String
                )
            ),
            Schedule  => Dict(
                EndTime  => ::String,
                Frequency  => : "ONCE", "HOURLY", "DAILY", "WEEKLY" or "MONTHLY",
                IsLocalTime  => ::Bool,
                QuietTime  => Dict(
                    End  => ::String,
                    Start  => ::String
                ),
                StartTime  => ::String,
                Timezone  => ::String
            ),
            SizePercent  => ::Int,
            TreatmentDescription  => ::String,
            TreatmentName  => ::String
        )...],
        Description  => ::String,
        HoldoutPercent  => ::Int,
        IsPaused  => ::Bool,
        Limits  => Dict(
            Daily  => ::Int,
            Total  => ::Int
        ),
        MessageConfiguration  => Dict(
            APNSMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Body  => ::String,
                ImageIconUrl  => ::String,
                ImageSmallIconUrl  => ::String,
                ImageUrl  => ::String,
                JsonBody  => ::String,
                MediaUrl  => ::String,
                SilentPush  => ::Bool,
                Title  => ::String,
                Url  => ::String
            ),
            DefaultMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Body  => ::String,
                ImageIconUrl  => ::String,
                ImageSmallIconUrl  => ::String,
                ImageUrl  => ::String,
                JsonBody  => ::String,
                MediaUrl  => ::String,
                SilentPush  => ::Bool,
                Title  => ::String,
                Url  => ::String
            ),
            EmailMessage  => Dict(
                Body  => ::String,
                HtmlBody  => ::String,
                Title  => ::String
            ),
            GCMMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Body  => ::String,
                ImageIconUrl  => ::String,
                ImageSmallIconUrl  => ::String,
                ImageUrl  => ::String,
                JsonBody  => ::String,
                MediaUrl  => ::String,
                SilentPush  => ::Bool,
                Title  => ::String,
                Url  => ::String
            ),
            SMSMessage  => Dict(
                Body  => ::String,
                MessageType  => : "TRANSACTIONAL" or "PROMOTIONAL",
                SenderId  => ::String
            )
        ),
        Name  => ::String,
        Schedule  => Dict(
            EndTime  => ::String,
            Frequency  => : "ONCE", "HOURLY", "DAILY", "WEEKLY" or "MONTHLY",
            IsLocalTime  => ::Bool,
            QuietTime  => Dict(
                End  => ::String,
                Start  => ::String
            ),
            StartTime  => ::String,
            Timezone  => ::String
        ),
        SegmentId  => ::String,
        SegmentVersion  => ::Int,
        TreatmentDescription  => ::String,
        TreatmentName  => ::String
    )

Returns

CreateCampaignResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.create_import_job — Method.

create_import_job(::AWSConfig, arguments::Dict)
create_import_job(::AWSConfig; ApplicationId=, ImportJobRequest=)

CreateImportJob Operation

Creates or updates an import job.

Arguments

ApplicationId::String – Required

ImportJobRequest: – Required

    Dict(
        DefineSegment  => ::Bool,
        ExternalId  => ::String,
        Format  => : "CSV" or "JSON",
        RegisterEndpoints  => ::Bool,
        RoleArn  => ::String,
        S3Url  => ::String,
        SegmentId  => ::String,
        SegmentName  => ::String
    )

Returns

CreateImportJobResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.create_segment — Method.

create_segment(::AWSConfig, arguments::Dict)
create_segment(::AWSConfig; ApplicationId=, WriteSegmentRequest=)

CreateSegment Operation

Used to create or update a segment.

Arguments

ApplicationId::String – Required

WriteSegmentRequest: – Required

    Dict(
        Dimensions  => Dict(
            Attributes  => ::Dict{String,String},
            Behavior  => Dict(
                Duration  => : "HR_24", "DAY_7", "DAY_14" or "DAY_30",
                RecencyType  => : "ACTIVE" or "INACTIVE"
            ),
            Demographic  => Dict(
                AppVersion  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                Channel  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                DeviceType  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                Make  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                Model  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                Platform  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                )
            ),
            Location  => Dict(
                DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                Values  => [::String...]
            ),
            UserAttributes  => ::Dict{String,String}
        ),
        Name  => ::String
    )

Returns

CreateSegmentResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_apns_channel — Method.

delete_apns_channel(::AWSConfig, arguments::Dict)
delete_apns_channel(::AWSConfig; ApplicationId=)

DeleteApnsChannel Operation

Deletes the APNs channel for an app.

Arguments

ApplicationId::String – Required

Returns

DeleteApnsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_apns_sandbox_channel — Method.

delete_apns_sandbox_channel(::AWSConfig, arguments::Dict)
delete_apns_sandbox_channel(::AWSConfig; ApplicationId=)

DeleteApnsSandboxChannel Operation

Delete an APNS sandbox channel

Arguments

ApplicationId::String – Required

Returns

DeleteApnsSandboxChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_campaign — Method.

delete_campaign(::AWSConfig, arguments::Dict)
delete_campaign(::AWSConfig; ApplicationId=, CampaignId=)

DeleteCampaign Operation

Deletes a campaign.

Arguments

ApplicationId::String – Required

CampaignId::String – Required

Returns

DeleteCampaignResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_email_channel — Method.

delete_email_channel(::AWSConfig, arguments::Dict)
delete_email_channel(::AWSConfig; ApplicationId=)

DeleteEmailChannel Operation

Delete an email channel

Arguments

ApplicationId::String – Required

Returns

DeleteEmailChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_event_stream — Method.

delete_event_stream(::AWSConfig, arguments::Dict)
delete_event_stream(::AWSConfig; ApplicationId=)

DeleteEventStream Operation

Deletes the event stream for an app.

Arguments

ApplicationId::String – Required

ApplicationId

Returns

DeleteEventStreamResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_gcm_channel — Method.

delete_gcm_channel(::AWSConfig, arguments::Dict)
delete_gcm_channel(::AWSConfig; ApplicationId=)

DeleteGcmChannel Operation

Deletes the GCM channel for an app.

Arguments

ApplicationId::String – Required

Returns

DeleteGcmChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_segment — Method.

delete_segment(::AWSConfig, arguments::Dict)
delete_segment(::AWSConfig; ApplicationId=, SegmentId=)

DeleteSegment Operation

Deletes a segment.

Arguments

ApplicationId::String – Required

SegmentId::String – Required

Returns

DeleteSegmentResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_sms_channel — Method.

delete_sms_channel(::AWSConfig, arguments::Dict)
delete_sms_channel(::AWSConfig; ApplicationId=)

DeleteSmsChannel Operation

Delete an SMS channel

Arguments

ApplicationId::String – Required

Returns

DeleteSmsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_apns_channel — Method.

get_apns_channel(::AWSConfig, arguments::Dict)
get_apns_channel(::AWSConfig; ApplicationId=)

GetApnsChannel Operation

Returns information about the APNs channel for an app.

Arguments

ApplicationId::String – Required

Returns

GetApnsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_apns_sandbox_channel — Method.

get_apns_sandbox_channel(::AWSConfig, arguments::Dict)
get_apns_sandbox_channel(::AWSConfig; ApplicationId=)

GetApnsSandboxChannel Operation

Get an APNS sandbox channel

Arguments

ApplicationId::String – Required

Returns

GetApnsSandboxChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_application_settings — Method.

get_application_settings(::AWSConfig, arguments::Dict)
get_application_settings(::AWSConfig; ApplicationId=)

GetApplicationSettings Operation

Used to request the settings for an app.

Arguments

ApplicationId::String – Required

Returns

GetApplicationSettingsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaign — Method.

get_campaign(::AWSConfig, arguments::Dict)
get_campaign(::AWSConfig; ApplicationId=, CampaignId=)

GetCampaign Operation

Returns information about a campaign.

Arguments

ApplicationId::String – Required

CampaignId::String – Required

Returns

GetCampaignResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaign_activities — Method.

get_campaign_activities(::AWSConfig, arguments::Dict)
get_campaign_activities(::AWSConfig; ApplicationId=, CampaignId=, <keyword arguments>)

GetCampaignActivities Operation

Returns information about the activity performed by a campaign.

Arguments

ApplicationId::String – Required

CampaignId::String – Required

PageSize::String

Token::String

Returns

GetCampaignActivitiesResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaign_version — Method.

get_campaign_version(::AWSConfig, arguments::Dict)
get_campaign_version(::AWSConfig; ApplicationId=, CampaignId=, Version=)

GetCampaignVersion Operation

Returns information about a specific version of a campaign.

Arguments

ApplicationId::String – Required

CampaignId::String – Required

Version::String – Required

Returns

GetCampaignVersionResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaign_versions — Method.

get_campaign_versions(::AWSConfig, arguments::Dict)
get_campaign_versions(::AWSConfig; ApplicationId=, CampaignId=, <keyword arguments>)

GetCampaignVersions Operation

Returns information about your campaign versions.

Arguments

ApplicationId::String – Required

CampaignId::String – Required

PageSize::String

Token::String

Returns

GetCampaignVersionsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaigns — Method.

get_campaigns(::AWSConfig, arguments::Dict)
get_campaigns(::AWSConfig; ApplicationId=, <keyword arguments>)

GetCampaigns Operation

Returns information about your campaigns.

Arguments

ApplicationId::String – Required

PageSize::String

Token::String

Returns

GetCampaignsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_email_channel — Method.

get_email_channel(::AWSConfig, arguments::Dict)
get_email_channel(::AWSConfig; ApplicationId=)

GetEmailChannel Operation

Get an email channel

Arguments

ApplicationId::String – Required

Returns

GetEmailChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_endpoint — Method.

get_endpoint(::AWSConfig, arguments::Dict)
get_endpoint(::AWSConfig; ApplicationId=, EndpointId=)

GetEndpoint Operation

Returns information about an endpoint.

Arguments

ApplicationId::String – Required

EndpointId::String – Required

Returns

GetEndpointResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_event_stream — Method.

get_event_stream(::AWSConfig, arguments::Dict)
get_event_stream(::AWSConfig; ApplicationId=)

GetEventStream Operation

Returns the event stream for an app.

Arguments

ApplicationId::String – Required

ApplicationId

Returns

GetEventStreamResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_gcm_channel — Method.

get_gcm_channel(::AWSConfig, arguments::Dict)
get_gcm_channel(::AWSConfig; ApplicationId=)

GetGcmChannel Operation

Returns information about the GCM channel for an app.

Arguments

ApplicationId::String – Required

Returns

GetGcmChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_import_job — Method.

get_import_job(::AWSConfig, arguments::Dict)
get_import_job(::AWSConfig; ApplicationId=, JobId=)

GetImportJob Operation

Returns information about an import job.

Arguments

ApplicationId::String – Required

JobId::String – Required

Returns

GetImportJobResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_import_jobs — Method.

get_import_jobs(::AWSConfig, arguments::Dict)
get_import_jobs(::AWSConfig; ApplicationId=, <keyword arguments>)

GetImportJobs Operation

Returns information about your import jobs.

Arguments

ApplicationId::String – Required

PageSize::String

Token::String

Returns

GetImportJobsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segment — Method.

get_segment(::AWSConfig, arguments::Dict)
get_segment(::AWSConfig; ApplicationId=, SegmentId=)

GetSegment Operation

Returns information about a segment.

Arguments

ApplicationId::String – Required

SegmentId::String – Required

Returns

GetSegmentResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segment_import_jobs — Method.

get_segment_import_jobs(::AWSConfig, arguments::Dict)
get_segment_import_jobs(::AWSConfig; ApplicationId=, SegmentId=, <keyword arguments>)

GetSegmentImportJobs Operation

Returns a list of import jobs for a specific segment.

Arguments

ApplicationId::String – Required

PageSize::String

SegmentId::String – Required

Token::String

Returns

GetSegmentImportJobsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segment_version — Method.

get_segment_version(::AWSConfig, arguments::Dict)
get_segment_version(::AWSConfig; ApplicationId=, SegmentId=, Version=)

GetSegmentVersion Operation

Returns information about a segment version.

Arguments

ApplicationId::String – Required

SegmentId::String – Required

Version::String – Required

Returns

GetSegmentVersionResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segment_versions — Method.

get_segment_versions(::AWSConfig, arguments::Dict)
get_segment_versions(::AWSConfig; ApplicationId=, SegmentId=, <keyword arguments>)

GetSegmentVersions Operation

Returns information about your segment versions.

Arguments

ApplicationId::String – Required

PageSize::String

SegmentId::String – Required

Token::String

Returns

GetSegmentVersionsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segments — Method.

get_segments(::AWSConfig, arguments::Dict)
get_segments(::AWSConfig; ApplicationId=, <keyword arguments>)

GetSegments Operation

Used to get information about your segments.

Arguments

ApplicationId::String – Required

PageSize::String

Token::String

Returns

GetSegmentsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_sms_channel — Method.

get_sms_channel(::AWSConfig, arguments::Dict)
get_sms_channel(::AWSConfig; ApplicationId=)

GetSmsChannel Operation

Get an SMS channel

Arguments

ApplicationId::String – Required

Returns

GetSmsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.put_event_stream — Method.

put_event_stream(::AWSConfig, arguments::Dict)
put_event_stream(::AWSConfig; ApplicationId=, WriteEventStream=)

PutEventStream Operation

Use to create or update the event stream for an app.

Arguments

ApplicationId::String – Required

ApplicationId

WriteEventStream: – Required

EventStream to write.

    Dict(
        DestinationStreamArn  => ::String,
        ExternalId  => ::String,
        RoleArn  => ::String
    )

Returns

PutEventStreamResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.send_messages — Method.

send_messages(::AWSConfig, arguments::Dict)
send_messages(::AWSConfig; ApplicationId=, MessageRequest=)

SendMessages Operation

Send a batch of messages

Arguments

ApplicationId::String – Required

MessageRequest: – Required

    Dict(
        Addresses  => ::Dict{String,String},
        Context  => ::Dict{String,String},
        MessageConfiguration  => Dict(
            APNSMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Badge  => ::Int,
                Body  => ::String,
                Category  => ::String,
                Data  => ::Dict{String,String},
                MediaUrl  => ::String,
                RawContent  => ::String,
                SilentPush  => ::Bool,
                Sound  => ::String,
                Substitutions  => ::Dict{String,String},
                ThreadId  => ::String,
                Title  => ::String,
                Url  => ::String
            ),
            DefaultMessage  => Dict(
                Body  => ::String,
                Substitutions  => ::Dict{String,String}
            ),
            DefaultPushNotificationMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Body  => ::String,
                Data  => ::Dict{String,String},
                SilentPush  => ::Bool,
                Substitutions  => ::Dict{String,String},
                Title  => ::String,
                Url  => ::String
            ),
            GCMMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Body  => ::String,
                CollapseKey  => ::String,
                Data  => ::Dict{String,String},
                IconReference  => ::String,
                ImageIconUrl  => ::String,
                ImageUrl  => ::String,
                RawContent  => ::String,
                RestrictedPackageName  => ::String,
                SilentPush  => ::Bool,
                SmallImageIconUrl  => ::String,
                Sound  => ::String,
                Substitutions  => ::Dict{String,String},
                Title  => ::String,
                Url  => ::String
            ),
            SMSMessage  => Dict(
                Body  => ::String,
                MessageType  => : "TRANSACTIONAL" or "PROMOTIONAL",
                SenderId  => ::String,
                Substitutions  => ::Dict{String,String}
            )
        )
    )

Returns

SendMessagesResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_apns_channel — Method.

update_apns_channel(::AWSConfig, arguments::Dict)
update_apns_channel(::AWSConfig; APNSChannelRequest=, ApplicationId=)

UpdateApnsChannel Operation

Use to update the APNs channel for an app.

Arguments

APNSChannelRequest: – Required

    Dict(
        Certificate  => ::String,
        Enabled  => ::Bool,
        PrivateKey  => ::String
    )

ApplicationId::String – Required

Returns

UpdateApnsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_apns_sandbox_channel — Method.

update_apns_sandbox_channel(::AWSConfig, arguments::Dict)
update_apns_sandbox_channel(::AWSConfig; APNSSandboxChannelRequest=, ApplicationId=)

UpdateApnsSandboxChannel Operation

Update an APNS sandbox channel

Arguments

APNSSandboxChannelRequest: – Required

    Dict(
        Certificate  => ::String,
        Enabled  => ::Bool,
        PrivateKey  => ::String
    )

ApplicationId::String – Required

Returns

UpdateApnsSandboxChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_application_settings — Method.

update_application_settings(::AWSConfig, arguments::Dict)
update_application_settings(::AWSConfig; ApplicationId=, WriteApplicationSettingsRequest=)

UpdateApplicationSettings Operation

Used to update the settings for an app.

Arguments

ApplicationId::String – Required

WriteApplicationSettingsRequest: – Required

    Dict(
        Limits  => Dict(
            Daily  => ::Int,
            Total  => ::Int
        ),
        QuietTime  => Dict(
            End  => ::String,
            Start  => ::String
        )
    )

Returns

UpdateApplicationSettingsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_campaign — Method.

update_campaign(::AWSConfig, arguments::Dict)
update_campaign(::AWSConfig; ApplicationId=, CampaignId=, WriteCampaignRequest=)

UpdateCampaign Operation

Use to update a campaign.

Arguments

ApplicationId::String – Required

CampaignId::String – Required

WriteCampaignRequest: – Required

    Dict(
        AdditionalTreatments  => [Dict(
            MessageConfiguration  => Dict(
                APNSMessage  => Dict(
                    Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                    Body  => ::String,
                    ImageIconUrl  => ::String,
                    ImageSmallIconUrl  => ::String,
                    ImageUrl  => ::String,
                    JsonBody  => ::String,
                    MediaUrl  => ::String,
                    SilentPush  => ::Bool,
                    Title  => ::String,
                    Url  => ::String
                ),
                DefaultMessage  => Dict(
                    Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                    Body  => ::String,
                    ImageIconUrl  => ::String,
                    ImageSmallIconUrl  => ::String,
                    ImageUrl  => ::String,
                    JsonBody  => ::String,
                    MediaUrl  => ::String,
                    SilentPush  => ::Bool,
                    Title  => ::String,
                    Url  => ::String
                ),
                EmailMessage  => Dict(
                    Body  => ::String,
                    HtmlBody  => ::String,
                    Title  => ::String
                ),
                GCMMessage  => Dict(
                    Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                    Body  => ::String,
                    ImageIconUrl  => ::String,
                    ImageSmallIconUrl  => ::String,
                    ImageUrl  => ::String,
                    JsonBody  => ::String,
                    MediaUrl  => ::String,
                    SilentPush  => ::Bool,
                    Title  => ::String,
                    Url  => ::String
                ),
                SMSMessage  => Dict(
                    Body  => ::String,
                    MessageType  => : "TRANSACTIONAL" or "PROMOTIONAL",
                    SenderId  => ::String
                )
            ),
            Schedule  => Dict(
                EndTime  => ::String,
                Frequency  => : "ONCE", "HOURLY", "DAILY", "WEEKLY" or "MONTHLY",
                IsLocalTime  => ::Bool,
                QuietTime  => Dict(
                    End  => ::String,
                    Start  => ::String
                ),
                StartTime  => ::String,
                Timezone  => ::String
            ),
            SizePercent  => ::Int,
            TreatmentDescription  => ::String,
            TreatmentName  => ::String
        )...],
        Description  => ::String,
        HoldoutPercent  => ::Int,
        IsPaused  => ::Bool,
        Limits  => Dict(
            Daily  => ::Int,
            Total  => ::Int
        ),
        MessageConfiguration  => Dict(
            APNSMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Body  => ::String,
                ImageIconUrl  => ::String,
                ImageSmallIconUrl  => ::String,
                ImageUrl  => ::String,
                JsonBody  => ::String,
                MediaUrl  => ::String,
                SilentPush  => ::Bool,
                Title  => ::String,
                Url  => ::String
            ),
            DefaultMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Body  => ::String,
                ImageIconUrl  => ::String,
                ImageSmallIconUrl  => ::String,
                ImageUrl  => ::String,
                JsonBody  => ::String,
                MediaUrl  => ::String,
                SilentPush  => ::Bool,
                Title  => ::String,
                Url  => ::String
            ),
            EmailMessage  => Dict(
                Body  => ::String,
                HtmlBody  => ::String,
                Title  => ::String
            ),
            GCMMessage  => Dict(
                Action  => : "OPEN_APP", "DEEP_LINK" or "URL",
                Body  => ::String,
                ImageIconUrl  => ::String,
                ImageSmallIconUrl  => ::String,
                ImageUrl  => ::String,
                JsonBody  => ::String,
                MediaUrl  => ::String,
                SilentPush  => ::Bool,
                Title  => ::String,
                Url  => ::String
            ),
            SMSMessage  => Dict(
                Body  => ::String,
                MessageType  => : "TRANSACTIONAL" or "PROMOTIONAL",
                SenderId  => ::String
            )
        ),
        Name  => ::String,
        Schedule  => Dict(
            EndTime  => ::String,
            Frequency  => : "ONCE", "HOURLY", "DAILY", "WEEKLY" or "MONTHLY",
            IsLocalTime  => ::Bool,
            QuietTime  => Dict(
                End  => ::String,
                Start  => ::String
            ),
            StartTime  => ::String,
            Timezone  => ::String
        ),
        SegmentId  => ::String,
        SegmentVersion  => ::Int,
        TreatmentDescription  => ::String,
        TreatmentName  => ::String
    )

Returns

UpdateCampaignResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_email_channel — Method.

update_email_channel(::AWSConfig, arguments::Dict)
update_email_channel(::AWSConfig; ApplicationId=, EmailChannelRequest=)

UpdateEmailChannel Operation

Update an email channel

Arguments

ApplicationId::String – Required

EmailChannelRequest: – Required

    Dict(
        Enabled  => ::Bool,
        FromAddress  => ::String,
        Identity  => ::String,
        RoleArn  => ::String
    )

Returns

UpdateEmailChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_endpoint — Method.

update_endpoint(::AWSConfig, arguments::Dict)
update_endpoint(::AWSConfig; ApplicationId=, EndpointId=, EndpointRequest=)

UpdateEndpoint Operation

Use to update an endpoint.

Arguments

ApplicationId::String – Required

EndpointId::String – Required

EndpointRequest: – Required

    Dict(
        Address  => ::String,
        Attributes  => ::Dict{String,String},
        ChannelType  => : "GCM", "APNS", "APNS_SANDBOX", "ADM", "SMS" or "EMAIL",
        Demographic  => Dict(
            AppVersion  => ::String,
            Locale  => ::String,
            Make  => ::String,
            Model  => ::String,
            ModelVersion  => ::String,
            Platform  => ::String,
            PlatformVersion  => ::String,
            Timezone  => ::String
        ),
        EffectiveDate  => ::String,
        EndpointStatus  => ::String,
        Location  => Dict(
            City  => ::String,
            Country  => ::String,
            Latitude  => double,
            Longitude  => double,
            PostalCode  => ::String,
            Region  => ::String
        ),
        Metrics  => ::Dict{String,String},
        OptOut  => ::String,
        RequestId  => ::String,
        User  => Dict(
            UserAttributes  => ::Dict{String,String},
            UserId  => ::String
        )
    )

Returns

UpdateEndpointResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_endpoints_batch — Method.

update_endpoints_batch(::AWSConfig, arguments::Dict)
update_endpoints_batch(::AWSConfig; ApplicationId=, EndpointBatchRequest=)

UpdateEndpointsBatch Operation

Use to update a batch of endpoints.

Arguments

ApplicationId::String – Required

EndpointBatchRequest: – Required

    [Dict(
        Address  => ::String,
        Attributes  => ::Dict{String,String},
        ChannelType  => : "GCM", "APNS", "APNS_SANDBOX", "ADM", "SMS" or "EMAIL",
        Demographic  => Dict(
            AppVersion  => ::String,
            Locale  => ::String,
            Make  => ::String,
            Model  => ::String,
            ModelVersion  => ::String,
            Platform  => ::String,
            PlatformVersion  => ::String,
            Timezone  => ::String
        ),
        EffectiveDate  => ::String,
        EndpointStatus  => ::String,
        Id  => ::String,
        Location  => Dict(
            City  => ::String,
            Country  => ::String,
            Latitude  => double,
            Longitude  => double,
            PostalCode  => ::String,
            Region  => ::String
        ),
        Metrics  => ::Dict{String,String},
        OptOut  => ::String,
        RequestId  => ::String,
        User  => Dict(
            UserAttributes  => ::Dict{String,String},
            UserId  => ::String
        )
    )...]

Returns

UpdateEndpointsBatchResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_gcm_channel — Method.

update_gcm_channel(::AWSConfig, arguments::Dict)
update_gcm_channel(::AWSConfig; ApplicationId=, GCMChannelRequest=)

UpdateGcmChannel Operation

Use to update the GCM channel for an app.

Arguments

ApplicationId::String – Required

GCMChannelRequest: – Required

    Dict(
        ApiKey  => ::String,
        Enabled  => ::Bool
    )

Returns

UpdateGcmChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_segment — Method.

update_segment(::AWSConfig, arguments::Dict)
update_segment(::AWSConfig; ApplicationId=, SegmentId=, WriteSegmentRequest=)

UpdateSegment Operation

Use to update a segment.

Arguments

ApplicationId::String – Required

SegmentId::String – Required

WriteSegmentRequest: – Required

    Dict(
        Dimensions  => Dict(
            Attributes  => ::Dict{String,String},
            Behavior  => Dict(
                Duration  => : "HR_24", "DAY_7", "DAY_14" or "DAY_30",
                RecencyType  => : "ACTIVE" or "INACTIVE"
            ),
            Demographic  => Dict(
                AppVersion  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                Channel  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                DeviceType  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                Make  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                Model  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                ),
                Platform  => Dict(
                    DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                    Values  => [::String...]
                )
            ),
            Location  => Dict(
                DimensionType  => : "INCLUSIVE" or "EXCLUSIVE",
                Values  => [::String...]
            ),
            UserAttributes  => ::Dict{String,String}
        ),
        Name  => ::String
    )

Returns

UpdateSegmentResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_sms_channel — Method.

update_sms_channel(::AWSConfig, arguments::Dict)
update_sms_channel(::AWSConfig; ApplicationId=, SMSChannelRequest=)

UpdateSmsChannel Operation

Update an SMS channel

Arguments

ApplicationId::String – Required

SMSChannelRequest: – Required

    Dict(
        Enabled  => ::Bool,
        SenderId  => ::String
    )

Returns

UpdateSmsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation