AmazonRoute53Domains.jl

AmazonRoute53Domains

Amazon Route 53 API actions let you register domain names and perform related operations.

This document is generated from apis/route53domains-2014-05-15.normal.json. See JuliaCloud/AWSCore.jl.

AmazonRoute53Domains.check_domain_availability

AmazonRoute53Domains.delete_tags_for_domain

AmazonRoute53Domains.disable_domain_auto_renew

AmazonRoute53Domains.disable_domain_transfer_lock

AmazonRoute53Domains.enable_domain_auto_renew

AmazonRoute53Domains.enable_domain_transfer_lock

AmazonRoute53Domains.get_contact_reachability_status

AmazonRoute53Domains.get_domain_detail

AmazonRoute53Domains.get_domain_suggestions

AmazonRoute53Domains.get_operation_detail

AmazonRoute53Domains.list_domains

AmazonRoute53Domains.list_operations

AmazonRoute53Domains.list_tags_for_domain

AmazonRoute53Domains.register_domain

AmazonRoute53Domains.renew_domain

AmazonRoute53Domains.resend_contact_reachability_email

AmazonRoute53Domains.retrieve_domain_auth_code

AmazonRoute53Domains.transfer_domain

AmazonRoute53Domains.update_domain_contact

AmazonRoute53Domains.update_domain_contact_privacy

AmazonRoute53Domains.update_domain_nameservers

AmazonRoute53Domains.update_tags_for_domain

AmazonRoute53Domains.view_billing

AmazonRoute53Domains.check_domain_availability — Function.

using AmazonRoute53Domains
check_domain_availability([::AWSConfig], arguments::Dict)
check_domain_availability([::AWSConfig]; DomainName=, <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "CheckDomainAvailability", arguments::Dict)
route53domains([::AWSConfig], "CheckDomainAvailability", DomainName=, <keyword arguments>)

CheckDomainAvailability Operation

This operation checks the availability of one domain name. Note that if the availability status of a domain is pending, you must submit another request to determine the availability of the domain name.

Arguments

DomainName = ::String – Required

The name of the domain that you want to get availability for.

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

IdnLangCode = ::String

Reserved for future use.

Returns

CheckDomainAvailabilityResponse

Exceptions

InvalidInput or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.delete_tags_for_domain — Function.

using AmazonRoute53Domains
delete_tags_for_domain([::AWSConfig], arguments::Dict)
delete_tags_for_domain([::AWSConfig]; DomainName=, TagsToDelete=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "DeleteTagsForDomain", arguments::Dict)
route53domains([::AWSConfig], "DeleteTagsForDomain", DomainName=, TagsToDelete=)

DeleteTagsForDomain Operation

This operation deletes the specified tags for a domain.

All tag operations are eventually consistent; subsequent operations may not immediately represent all issued operations.

Arguments

DomainName = ::String – Required

The domain for which you want to delete one or more tags.

TagsToDelete = [::String, ...] – Required

A list of tag keys to delete.

Returns

DeleteTagsForDomainResponse

Exceptions

InvalidInput, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.disable_domain_auto_renew — Function.

using AmazonRoute53Domains
disable_domain_auto_renew([::AWSConfig], arguments::Dict)
disable_domain_auto_renew([::AWSConfig]; DomainName=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "DisableDomainAutoRenew", arguments::Dict)
route53domains([::AWSConfig], "DisableDomainAutoRenew", DomainName=)

DisableDomainAutoRenew Operation

This operation disables automatic renewal of domain registration for the specified domain.

Arguments

DomainName = ::String – Required

The name of the domain that you want to disable automatic renewal for.

Returns

DisableDomainAutoRenewResponse

Exceptions

InvalidInput or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.disable_domain_transfer_lock — Function.

using AmazonRoute53Domains
disable_domain_transfer_lock([::AWSConfig], arguments::Dict)
disable_domain_transfer_lock([::AWSConfig]; DomainName=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "DisableDomainTransferLock", arguments::Dict)
route53domains([::AWSConfig], "DisableDomainTransferLock", DomainName=)

DisableDomainTransferLock Operation

This operation removes the transfer lock on the domain (specifically the clientTransferProhibited status) to allow domain transfers. We recommend you refrain from performing this action unless you intend to transfer the domain to a different registrar. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Arguments

DomainName = ::String – Required

The name of the domain that you want to remove the transfer lock for.

Returns

DisableDomainTransferLockResponse

Exceptions

InvalidInput, DuplicateRequest, TLDRulesViolation, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.enable_domain_auto_renew — Function.

using AmazonRoute53Domains
enable_domain_auto_renew([::AWSConfig], arguments::Dict)
enable_domain_auto_renew([::AWSConfig]; DomainName=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "EnableDomainAutoRenew", arguments::Dict)
route53domains([::AWSConfig], "EnableDomainAutoRenew", DomainName=)

EnableDomainAutoRenew Operation

This operation configures Amazon Route 53 to automatically renew the specified domain before the domain registration expires. The cost of renewing your domain registration is billed to your AWS account.

The period during which you can renew a domain name varies by TLD. For a list of TLDs and their renewal policies, see "Renewal, restoration, and deletion times" on the website for our registrar partner, Gandi. Route 53 requires that you renew before the end of the renewal period that is listed on the Gandi website so we can complete processing before the deadline.

Arguments

DomainName = ::String – Required

The name of the domain that you want to enable automatic renewal for.

Returns

EnableDomainAutoRenewResponse

Exceptions

InvalidInput, UnsupportedTLD or TLDRulesViolation.

See also: AWS API Documentation

AmazonRoute53Domains.enable_domain_transfer_lock — Function.

using AmazonRoute53Domains
enable_domain_transfer_lock([::AWSConfig], arguments::Dict)
enable_domain_transfer_lock([::AWSConfig]; DomainName=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "EnableDomainTransferLock", arguments::Dict)
route53domains([::AWSConfig], "EnableDomainTransferLock", DomainName=)

EnableDomainTransferLock Operation

This operation sets the transfer lock on the domain (specifically the clientTransferProhibited status) to prevent domain transfers. Successful submission returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Arguments

DomainName = ::String – Required

The name of the domain that you want to set the transfer lock for.

Returns

EnableDomainTransferLockResponse

Exceptions

InvalidInput, DuplicateRequest, TLDRulesViolation, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.get_contact_reachability_status — Function.

using AmazonRoute53Domains
get_contact_reachability_status([::AWSConfig], arguments::Dict)
get_contact_reachability_status([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "GetContactReachabilityStatus", arguments::Dict)
route53domains([::AWSConfig], "GetContactReachabilityStatus", <keyword arguments>)

GetContactReachabilityStatus Operation

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation returns information about whether the registrant contact has responded.

If you want us to resend the email, use the ResendContactReachabilityEmail operation.

Arguments

domainName = ::String

The name of the domain for which you want to know whether the registrant contact has confirmed that the email address is valid.

Returns

GetContactReachabilityStatusResponse

Exceptions

InvalidInput, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.get_domain_detail — Function.

using AmazonRoute53Domains
get_domain_detail([::AWSConfig], arguments::Dict)
get_domain_detail([::AWSConfig]; DomainName=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "GetDomainDetail", arguments::Dict)
route53domains([::AWSConfig], "GetDomainDetail", DomainName=)

GetDomainDetail Operation

This operation returns detailed information about a specified domain that is associated with the current AWS account. Contact information for the domain is also returned as part of the output.

Arguments

DomainName = ::String – Required

The name of the domain that you want to get detailed information about.

Returns

GetDomainDetailResponse

Exceptions

InvalidInput or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.get_domain_suggestions — Function.

using AmazonRoute53Domains
get_domain_suggestions([::AWSConfig], arguments::Dict)
get_domain_suggestions([::AWSConfig]; DomainName=, SuggestionCount=, OnlyAvailable=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "GetDomainSuggestions", arguments::Dict)
route53domains([::AWSConfig], "GetDomainSuggestions", DomainName=, SuggestionCount=, OnlyAvailable=)

GetDomainSuggestions Operation

The GetDomainSuggestions operation returns a list of suggested domain names given a string, which can either be a domain name or simply a word or phrase (without spaces).

Arguments

DomainName = ::String – Required

A domain name that you want to use as the basis for a list of possible domain names. The domain name must contain a top-level domain (TLD), such as .com, that Amazon Route 53 supports. For a list of TLDs, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

SuggestionCount = ::Int – Required

The number of suggested domain names that you want Amazon Route 53 to return.

OnlyAvailable = ::Bool – Required

If OnlyAvailable is true, Amazon Route 53 returns only domain names that are available. If OnlyAvailable is false, Amazon Route 53 returns domain names without checking whether they're available to be registered. To determine whether the domain is available, you can call checkDomainAvailability for each suggestion.

Returns

GetDomainSuggestionsResponse

Exceptions

InvalidInput or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.get_operation_detail — Function.

using AmazonRoute53Domains
get_operation_detail([::AWSConfig], arguments::Dict)
get_operation_detail([::AWSConfig]; OperationId=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "GetOperationDetail", arguments::Dict)
route53domains([::AWSConfig], "GetOperationDetail", OperationId=)

GetOperationDetail Operation

This operation returns the current status of an operation that is not completed.

Arguments

OperationId = ::String – Required

The identifier for the operation for which you want to get the status. Amazon Route 53 returned the identifier in the response to the original request.

Returns

GetOperationDetailResponse

Exceptions

InvalidInput.

See also: AWS API Documentation

AmazonRoute53Domains.list_domains — Function.

using AmazonRoute53Domains
list_domains([::AWSConfig], arguments::Dict)
list_domains([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "ListDomains", arguments::Dict)
route53domains([::AWSConfig], "ListDomains", <keyword arguments>)

ListDomains Operation

This operation returns all the domain names registered with Amazon Route 53 for the current AWS account.

Arguments

Marker = ::String

For an initial request for a list of domains, omit this element. If the number of domains that are associated with the current AWS account is greater than the value that you specified for MaxItems, you can use Marker to return additional domains. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element.

Constraints: The marker must match the value specified in the previous request.

MaxItems = ::Int

Number of domains to be returned.

Default: 20

Returns

ListDomainsResponse

Exceptions

InvalidInput.

See also: AWS API Documentation

AmazonRoute53Domains.list_operations — Function.

using AmazonRoute53Domains
list_operations([::AWSConfig], arguments::Dict)
list_operations([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "ListOperations", arguments::Dict)
route53domains([::AWSConfig], "ListOperations", <keyword arguments>)

ListOperations Operation

This operation returns the operation IDs of operations that are not yet complete.

Arguments

Marker = ::String

For an initial request for a list of operations, omit this element. If the number of operations that are not yet complete is greater than the value that you specified for MaxItems, you can use Marker to return additional operations. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element.

MaxItems = ::Int

Number of domains to be returned.

Default: 20

Returns

ListOperationsResponse

Exceptions

InvalidInput.

See also: AWS API Documentation

AmazonRoute53Domains.list_tags_for_domain — Function.

using AmazonRoute53Domains
list_tags_for_domain([::AWSConfig], arguments::Dict)
list_tags_for_domain([::AWSConfig]; DomainName=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "ListTagsForDomain", arguments::Dict)
route53domains([::AWSConfig], "ListTagsForDomain", DomainName=)

ListTagsForDomain Operation

This operation returns all of the tags that are associated with the specified domain.

All tag operations are eventually consistent; subsequent operations may not immediately represent all issued operations.

Arguments

DomainName = ::String – Required

The domain for which you want to get a list of tags.

Returns

ListTagsForDomainResponse

Exceptions

InvalidInput, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.register_domain — Function.

using AmazonRoute53Domains
register_domain([::AWSConfig], arguments::Dict)
register_domain([::AWSConfig]; DomainName=, DurationInYears=, AdminContact=, RegistrantContact=, TechContact=, <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "RegisterDomain", arguments::Dict)
route53domains([::AWSConfig], "RegisterDomain", DomainName=, DurationInYears=, AdminContact=, RegistrantContact=, TechContact=, <keyword arguments>)

RegisterDomain Operation

This operation registers a domain. Domains are registered by the AWS registrar partner, Gandi. For some top-level domains (TLDs), this operation requires extra parameters.

When you register a domain, Amazon Route 53 does the following:

Creates a Amazon Route 53 hosted zone that has the same name as the domain. Amazon Route 53 assigns four name servers to your hosted zone and automatically updates your domain registration with the names of these name servers.

Enables autorenew, so your domain registration will renew automatically each year. We'll notify you in advance of the renewal date so you can choose whether to renew the registration.

Optionally enables privacy protection, so WHOIS queries return contact information for our registrar partner, Gandi, instead of the information you entered for registrant, admin, and tech contacts.

If registration is successful, returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant is notified by email.

Charges your AWS account an amount based on the top-level domain. For more information, see Amazon Route 53 Pricing.

Arguments

DomainName = ::String – Required

The domain name that you want to register.

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

IdnLangCode = ::String

Reserved for future use.

DurationInYears = ::Int – Required

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

AutoRenew = ::Bool

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

AdminContact = [ ... ] – Required

Provides detailed contact information.

 AdminContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

RegistrantContact = [ ... ] – Required

Provides detailed contact information.

 RegistrantContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

TechContact = [ ... ] – Required

Provides detailed contact information.

 TechContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

PrivacyProtectAdminContact = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

PrivacyProtectRegistrantContact = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

PrivacyProtectTechContact = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

Returns

RegisterDomainResponse

Exceptions

InvalidInput, UnsupportedTLD, DuplicateRequest, TLDRulesViolation, DomainLimitExceeded or OperationLimitExceeded.

See also: AWS API Documentation

AmazonRoute53Domains.renew_domain — Function.

using AmazonRoute53Domains
renew_domain([::AWSConfig], arguments::Dict)
renew_domain([::AWSConfig]; DomainName=, CurrentExpiryYear=, <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "RenewDomain", arguments::Dict)
route53domains([::AWSConfig], "RenewDomain", DomainName=, CurrentExpiryYear=, <keyword arguments>)

RenewDomain Operation

This operation renews a domain for the specified number of years. The cost of renewing your domain is billed to your AWS account.

We recommend that you renew your domain several weeks before the expiration date. Some TLD registries delete domains before the expiration date if you haven't renewed far enough in advance. For more information about renewing domain registration, see Renewing Registration for a Domain in the Amazon Route 53 Developer Guide.

Arguments

DomainName = ::String – Required

The name of the domain that you want to renew.

DurationInYears = ::Int

The number of years that you want to renew the domain for. The maximum number of years depends on the top-level domain. For the range of valid values for your domain, see Domains that You Can Register with Amazon Route 53 in the Amazon Route 53 Developer Guide.

Default: 1

CurrentExpiryYear = ::Int – Required

The year when the registration for the domain is set to expire. This value must match the current expiration date for the domain.

Returns

RenewDomainResponse

Exceptions

InvalidInput, UnsupportedTLD, DuplicateRequest, TLDRulesViolation or OperationLimitExceeded.

See also: AWS API Documentation

AmazonRoute53Domains.resend_contact_reachability_email — Function.

using AmazonRoute53Domains
resend_contact_reachability_email([::AWSConfig], arguments::Dict)
resend_contact_reachability_email([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "ResendContactReachabilityEmail", arguments::Dict)
route53domains([::AWSConfig], "ResendContactReachabilityEmail", <keyword arguments>)

ResendContactReachabilityEmail Operation

For operations that require confirmation that the email address for the registrant contact is valid, such as registering a new domain, this operation resends the confirmation email to the current email address for the registrant contact.

Arguments

domainName = ::String

The name of the domain for which you want Amazon Route 53 to resend a confirmation email to the registrant contact.

Returns

ResendContactReachabilityEmailResponse

Exceptions

InvalidInput, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.retrieve_domain_auth_code — Function.

using AmazonRoute53Domains
retrieve_domain_auth_code([::AWSConfig], arguments::Dict)
retrieve_domain_auth_code([::AWSConfig]; DomainName=)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "RetrieveDomainAuthCode", arguments::Dict)
route53domains([::AWSConfig], "RetrieveDomainAuthCode", DomainName=)

RetrieveDomainAuthCode Operation

This operation returns the AuthCode for the domain. To transfer a domain to another registrar, you provide this value to the new registrar.

Arguments

DomainName = ::String – Required

The name of the domain that you want to get an authorization code for.

Returns

RetrieveDomainAuthCodeResponse

Exceptions

InvalidInput or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.transfer_domain — Function.

using AmazonRoute53Domains
transfer_domain([::AWSConfig], arguments::Dict)
transfer_domain([::AWSConfig]; DomainName=, DurationInYears=, AdminContact=, RegistrantContact=, TechContact=, <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "TransferDomain", arguments::Dict)
route53domains([::AWSConfig], "TransferDomain", DomainName=, DurationInYears=, AdminContact=, RegistrantContact=, TechContact=, <keyword arguments>)

TransferDomain Operation

This operation transfers a domain from another registrar to Amazon Route 53. When the transfer is complete, the domain is registered with the AWS registrar partner, Gandi.

For transfer requirements, a detailed procedure, and information about viewing the status of a domain transfer, see Transferring Registration for a Domain to Amazon Route 53 in the Amazon Route 53 Developer Guide.

If the registrar for your domain is also the DNS service provider for the domain, we highly recommend that you consider transferring your DNS service to Amazon Route 53 or to another DNS service provider before you transfer your registration. Some registrars provide free DNS service when you purchase a domain registration. When you transfer the registration, the previous registrar will not renew your domain registration and could end your DNS service at any time.

Important

If the registrar for your domain is also the DNS service provider for the domain and you don't transfer DNS service to another provider, your website, email, and the web applications associated with the domain might become unavailable.

If the transfer is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the transfer doesn't complete successfully, the domain registrant will be notified by email.

Arguments

DomainName = ::String – Required

The name of the domain that you want to transfer to Amazon Route 53.

Constraints: The domain name can contain only the letters a through z, the numbers 0 through 9, and hyphen (-). Internationalized Domain Names are not supported.

IdnLangCode = ::String

Reserved for future use.

DurationInYears = ::Int – Required

The number of years that you want to register the domain for. Domains are registered for a minimum of one year. The maximum period depends on the top-level domain.

Default: 1

Nameservers = [[ ... ], ...]

Contains details for the host and glue IP addresses.

 Nameservers = [[
        "Name" => <required> ::String,
        "GlueIps" =>  [::String, ...]
    ], ...]

AuthCode = ::String

The authorization code for the domain. You get this value from the current registrar.

AutoRenew = ::Bool

Indicates whether the domain will be automatically renewed (true) or not (false). Autorenewal only takes effect after the account is charged.

Default: true

AdminContact = [ ... ] – Required

Provides detailed contact information.

 AdminContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

RegistrantContact = [ ... ] – Required

Provides detailed contact information.

 RegistrantContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

TechContact = [ ... ] – Required

Provides detailed contact information.

 TechContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

PrivacyProtectAdminContact = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

PrivacyProtectRegistrantContact = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

PrivacyProtectTechContact = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Default: true

Returns

TransferDomainResponse

Exceptions

InvalidInput, UnsupportedTLD, DuplicateRequest, TLDRulesViolation, DomainLimitExceeded or OperationLimitExceeded.

See also: AWS API Documentation

AmazonRoute53Domains.update_domain_contact — Function.

using AmazonRoute53Domains
update_domain_contact([::AWSConfig], arguments::Dict)
update_domain_contact([::AWSConfig]; DomainName=, <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "UpdateDomainContact", arguments::Dict)
route53domains([::AWSConfig], "UpdateDomainContact", DomainName=, <keyword arguments>)

UpdateDomainContact Operation

This operation updates the contact information for a particular domain. Information for at least one contact (registrant, administrator, or technical) must be supplied for update.

If the update is successful, this method returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Arguments

DomainName = ::String – Required

The name of the domain that you want to update contact information for.

AdminContact = [ ... ]

Provides detailed contact information.

 AdminContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

RegistrantContact = [ ... ]

Provides detailed contact information.

 RegistrantContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

TechContact = [ ... ]

Provides detailed contact information.

 TechContact = [
        "FirstName" =>  ::String,
        "LastName" =>  ::String,
        "ContactType" =>  "PERSON", "COMPANY", "ASSOCIATION", "PUBLIC_BODY" or "RESELLER",
        "OrganizationName" =>  ::String,
        "AddressLine1" =>  ::String,
        "AddressLine2" =>  ::String,
        "City" =>  ::String,
        "State" =>  ::String,
        "CountryCode" =>  "AD", "AE", "AF", "AG", "AI", "AL", "AM", "AN", "AO", "AQ", "AR", "AS", "AT", "AU", "AW", "AZ", "BA", "BB", "BD", "BE", "BF", "BG", "BH", "BI", "BJ", "BL", "BM", "BN", "BO", "BR", "BS", "BT", "BW", "BY", "BZ", "CA", "CC", "CD", "CF", "CG", "CH", "CI", "CK", "CL", "CM", "CN", "CO", "CR", "CU", "CV", "CX", "CY", "CZ", "DE", "DJ", "DK", "DM", "DO", "DZ", "EC", "EE", "EG", "ER", "ES", "ET", "FI", "FJ", "FK", "FM", "FO", "FR", "GA", "GB", "GD", "GE", "GH", "GI", "GL", "GM", "GN", "GQ", "GR", "GT", "GU", "GW", "GY", "HK", "HN", "HR", "HT", "HU", "ID", "IE", "IL", "IM", "IN", "IQ", "IR", "IS", "IT", "JM", "JO", "JP", "KE", "KG", "KH", "KI", "KM", "KN", "KP", "KR", "KW", "KY", "KZ", "LA", "LB", "LC", "LI", "LK", "LR", "LS", "LT", "LU", "LV", "LY", "MA", "MC", "MD", "ME", "MF", "MG", "MH", "MK", "ML", "MM", "MN", "MO", "MP", "MR", "MS", "MT", "MU", "MV", "MW", "MX", "MY", "MZ", "NA", "NC", "NE", "NG", "NI", "NL", "NO", "NP", "NR", "NU", "NZ", "OM", "PA", "PE", "PF", "PG", "PH", "PK", "PL", "PM", "PN", "PR", "PT", "PW", "PY", "QA", "RO", "RS", "RU", "RW", "SA", "SB", "SC", "SD", "SE", "SG", "SH", "SI", "SK", "SL", "SM", "SN", "SO", "SR", "ST", "SV", "SY", "SZ", "TC", "TD", "TG", "TH", "TJ", "TK", "TL", "TM", "TN", "TO", "TR", "TT", "TV", "TW", "TZ", "UA", "UG", "US", "UY", "UZ", "VA", "VC", "VE", "VG", "VI", "VN", "VU", "WF", "WS", "YE", "YT", "ZA", "ZM" or "ZW",
        "ZipCode" =>  ::String,
        "PhoneNumber" =>  ::String,
        "Email" =>  ::String,
        "Fax" =>  ::String,
        "ExtraParams" =>  [[
            "Name" => <required> "DUNS_NUMBER", "BRAND_NUMBER", "BIRTH_DEPARTMENT", "BIRTH_DATE_IN_YYYY_MM_DD", "BIRTH_COUNTRY", "BIRTH_CITY", "DOCUMENT_NUMBER", "AU_ID_NUMBER", "AU_ID_TYPE", "CA_LEGAL_TYPE", "CA_BUSINESS_ENTITY_TYPE", "ES_IDENTIFICATION", "ES_IDENTIFICATION_TYPE", "ES_LEGAL_FORM", "FI_BUSINESS_NUMBER", "FI_ID_NUMBER", "IT_PIN", "RU_PASSPORT_DATA", "SE_ID_NUMBER", "SG_ID_NUMBER" or "VAT_NUMBER",
            "Value" => <required> ::String
        ], ...]
    ]

Returns

UpdateDomainContactResponse

Exceptions

InvalidInput, DuplicateRequest, TLDRulesViolation, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.update_domain_contact_privacy — Function.

using AmazonRoute53Domains
update_domain_contact_privacy([::AWSConfig], arguments::Dict)
update_domain_contact_privacy([::AWSConfig]; DomainName=, <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "UpdateDomainContactPrivacy", arguments::Dict)
route53domains([::AWSConfig], "UpdateDomainContactPrivacy", DomainName=, <keyword arguments>)

UpdateDomainContactPrivacy Operation

This operation updates the specified domain contact's privacy setting. When the privacy option is enabled, personal information such as postal or email address is hidden from the results of a public WHOIS query. The privacy services are provided by the AWS registrar, Gandi. For more information, see the Gandi privacy features.

This operation only affects the privacy of the specified contact type (registrant, administrator, or tech). Successful acceptance returns an operation ID that you can use with GetOperationDetail to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Arguments

DomainName = ::String – Required

The name of the domain that you want to update the privacy setting for.

AdminPrivacy = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

RegistrantPrivacy = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

TechPrivacy = ::Bool

Whether you want to conceal contact information from WHOIS queries. If you specify true, WHOIS ("who is") queries will return contact information for our registrar partner, Gandi, instead of the contact information that you enter.

Returns

UpdateDomainContactPrivacyResponse

Exceptions

InvalidInput, DuplicateRequest, TLDRulesViolation, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.update_domain_nameservers — Function.

using AmazonRoute53Domains
update_domain_nameservers([::AWSConfig], arguments::Dict)
update_domain_nameservers([::AWSConfig]; DomainName=, Nameservers=, <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "UpdateDomainNameservers", arguments::Dict)
route53domains([::AWSConfig], "UpdateDomainNameservers", DomainName=, Nameservers=, <keyword arguments>)

UpdateDomainNameservers Operation

This operation replaces the current set of name servers for the domain with the specified set of name servers. If you use Amazon Route 53 as your DNS service, specify the four name servers in the delegation set for the hosted zone for the domain.

If successful, this operation returns an operation ID that you can use to track the progress and completion of the action. If the request is not completed successfully, the domain registrant will be notified by email.

Arguments

DomainName = ::String – Required

The name of the domain that you want to change name servers for.

FIAuthKey = ::String

The authorization key for .fi domains

Nameservers = [[ ... ], ...] – Required

A list of new name servers for the domain.

 Nameservers = [[
        "Name" => <required> ::String,
        "GlueIps" =>  [::String, ...]
    ], ...]

Returns

UpdateDomainNameserversResponse

Exceptions

InvalidInput, DuplicateRequest, TLDRulesViolation, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.update_tags_for_domain — Function.

using AmazonRoute53Domains
update_tags_for_domain([::AWSConfig], arguments::Dict)
update_tags_for_domain([::AWSConfig]; DomainName=, <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "UpdateTagsForDomain", arguments::Dict)
route53domains([::AWSConfig], "UpdateTagsForDomain", DomainName=, <keyword arguments>)

UpdateTagsForDomain Operation

This operation adds or updates tags for a specified domain.

All tag operations are eventually consistent; subsequent operations may not immediately represent all issued operations.

Arguments

DomainName = ::String – Required

The domain for which you want to add or update tags.

TagsToUpdate = [[ ... ], ...]

A list of the tag keys and values that you want to add or update. If you specify a key that already exists, the corresponding value will be replaced.

 TagsToUpdate = [[
        "Key" =>  ::String,
        "Value" =>  ::String
    ], ...]

Returns

UpdateTagsForDomainResponse

Exceptions

InvalidInput, OperationLimitExceeded or UnsupportedTLD.

See also: AWS API Documentation

AmazonRoute53Domains.view_billing — Function.

using AmazonRoute53Domains
view_billing([::AWSConfig], arguments::Dict)
view_billing([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.route53domains
route53domains([::AWSConfig], "ViewBilling", arguments::Dict)
route53domains([::AWSConfig], "ViewBilling", <keyword arguments>)

ViewBilling Operation

Returns all the domain-related billing records for the current AWS account for a specified period

Arguments

Start = timestamp

The beginning date and time for the time period for which you want a list of billing records. Specify the date in Unix time format.

End = timestamp

The end date and time for the time period for which you want a list of billing records. Specify the date in Unix time format.

Marker = ::String

For an initial request for a list of billing records, omit this element. If the number of billing records that are associated with the current AWS account during the specified period is greater than the value that you specified for MaxItems, you can use Marker to return additional billing records. Get the value of NextPageMarker from the previous response, and submit another request that includes the value of NextPageMarker in the Marker element.

Constraints: The marker must match the value of NextPageMarker that was returned in the previous response.

MaxItems = ::Int

The number of billing records to be returned.

Default: 20

Returns

ViewBillingResponse

Exceptions

InvalidInput.

See also: AWS API Documentation