AmazonSNS.jl

AmazonSNS

Amazon Simple Notification Service

Amazon Simple Notification Service (Amazon SNS) is a web service that enables you to build distributed web-enabled applications. Applications can use Amazon SNS to easily push real-time notification messages to interested subscribers over multiple delivery protocols. For more information about this product see http://aws.amazon.com/sns. For detailed information about Amazon SNS features and their associated API calls, see the Amazon SNS Developer Guide.

We also provide SDKs that enable you to access Amazon SNS from your preferred programming language. The SDKs contain functionality that automatically takes care of tasks such as: cryptographically signing your service requests, retrying requests, and handling error responses. For a list of available SDKs, go to Tools for Amazon Web Services.

This document is generated from apis/sns-2010-03-31.normal.json. See JuliaCloud/AWSCore.jl.

AmazonSNS.add_permission

AmazonSNS.check_if_phone_number_is_opted_out

AmazonSNS.confirm_subscription

AmazonSNS.create_platform_application

AmazonSNS.create_platform_endpoint

AmazonSNS.create_topic

AmazonSNS.delete_endpoint

AmazonSNS.delete_platform_application

AmazonSNS.delete_topic

AmazonSNS.get_endpoint_attributes

AmazonSNS.get_platform_application_attributes

AmazonSNS.get_smsattributes

AmazonSNS.get_subscription_attributes

AmazonSNS.get_topic_attributes

AmazonSNS.list_endpoints_by_platform_application

AmazonSNS.list_phone_numbers_opted_out

AmazonSNS.list_platform_applications

AmazonSNS.list_subscriptions

AmazonSNS.list_subscriptions_by_topic

AmazonSNS.list_topics

AmazonSNS.opt_in_phone_number

AmazonSNS.publish

AmazonSNS.remove_permission

AmazonSNS.set_endpoint_attributes

AmazonSNS.set_platform_application_attributes

AmazonSNS.set_smsattributes

AmazonSNS.set_subscription_attributes

AmazonSNS.set_topic_attributes

AmazonSNS.subscribe

AmazonSNS.unsubscribe

AmazonSNS.add_permission — Function.

add_permission(::AWSConfig, arguments::Dict)
add_permission([::AWSConfig]; TopicArn=, Label=, AWSAccountId=, ActionName=)

AddPermission Operation

Adds a statement to a topic's access control policy, granting access for the specified AWS accounts to the specified actions.

Arguments

TopicArn = ::String – Required

The ARN of the topic whose access control policy you wish to modify.

Label = ::String – Required

A unique identifier for the new policy statement.

AWSAccountId = [::String, ...] – Required

The AWS account IDs of the users (principals) who will be given access to the specified actions. The users must have AWS accounts, but do not need to be signed up for this service.

ActionName = [::String, ...] – Required

The action you want to allow for the specified principal(s).

Valid values: any Amazon SNS action name.

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.check_if_phone_number_is_opted_out — Function.

check_if_phone_number_is_opted_out(::AWSConfig, arguments::Dict)
check_if_phone_number_is_opted_out([::AWSConfig]; phoneNumber=)

CheckIfPhoneNumberIsOptedOut Operation

Accepts a phone number and indicates whether the phone holder has opted out of receiving SMS messages from your account. You cannot send SMS messages to a number that is opted out.

To resume sending messages, you can opt in the number by using the OptInPhoneNumber action.

Arguments

phoneNumber = ::String – Required

The phone number for which you want to check the opt out status.

Returns

CheckIfPhoneNumberIsOptedOutResponse

Exceptions

ThrottledException, InternalErrorException, AuthorizationErrorException or InvalidParameterException.

See also: AWS API Documentation

AmazonSNS.confirm_subscription — Function.

confirm_subscription(::AWSConfig, arguments::Dict)
confirm_subscription([::AWSConfig]; TopicArn=, Token=, <keyword arguments>)

ConfirmSubscription Operation

Verifies an endpoint owner's intent to receive messages by validating the token sent to the endpoint by an earlier Subscribe action. If the token is valid, the action creates a new subscription and returns its Amazon Resource Name (ARN). This call requires an AWS signature only when the AuthenticateOnUnsubscribe flag is set to "true".

Arguments

TopicArn = ::String – Required

The ARN of the topic for which you wish to confirm a subscription.

Token = ::String – Required

Short-lived token sent to an endpoint during the Subscribe action.

AuthenticateOnUnsubscribe = ::String

Disallows unauthenticated unsubscribes of the subscription. If the value of this parameter is true and the request has an AWS signature, then only the topic owner and the subscription owner can unsubscribe the endpoint. The unsubscribe action requires AWS authentication.

Returns

ConfirmSubscriptionResponse

Exceptions

SubscriptionLimitExceededException, InvalidParameterException, NotFoundException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.create_platform_application — Function.

create_platform_application(::AWSConfig, arguments::Dict)
create_platform_application([::AWSConfig]; Name=, Platform=, Attributes=)

CreatePlatformApplication Operation

Creates a platform application object for one of the supported push notification services, such as APNS and GCM, to which devices and mobile apps may register. You must specify PlatformPrincipal and PlatformCredential attributes when using the CreatePlatformApplication action. The PlatformPrincipal is received from the notification service. For APNS/APNS_SANDBOX, PlatformPrincipal is "SSL certificate". For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id". The PlatformCredential is also received from the notification service. For WNS, PlatformPrincipal is "Package Security Identifier". For MPNS, PlatformPrincipal is "TLS certificate". For Baidu, PlatformPrincipal is "API key".

For APNS/APNS_SANDBOX, PlatformCredential is "private key". For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret". For WNS, PlatformCredential is "secret key". For MPNS, PlatformCredential is "private key". For Baidu, PlatformCredential is "secret key". The PlatformApplicationArn that is returned when using CreatePlatformApplication is then used as an attribute for the CreatePlatformEndpoint action. For more information, see Using Amazon SNS Mobile Push Notifications. For more information about obtaining the PlatformPrincipal and PlatformCredential for each of the supported push notification services, see Getting Started with Apple Push Notification Service, Getting Started with Amazon Device Messaging, Getting Started with Baidu Cloud Push, Getting Started with Google Cloud Messaging for Android, Getting Started with MPNS, or Getting Started with WNS.

Arguments

Name = ::String – Required

Application names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, hyphens, and periods, and must be between 1 and 256 characters long.

Platform = ::String – Required

The following platforms are supported: ADM (Amazon Device Messaging), APNS (Apple Push Notification Service), APNS_SANDBOX, and GCM (Google Cloud Messaging).

Attributes = ::Dict{String,String} – Required

For a list of attributes, see SetPlatformApplicationAttributes

Returns

CreatePlatformApplicationResponse

Exceptions

InvalidParameterException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.create_platform_endpoint — Function.

create_platform_endpoint(::AWSConfig, arguments::Dict)
create_platform_endpoint([::AWSConfig]; PlatformApplicationArn=, Token=, <keyword arguments>)

CreatePlatformEndpoint Operation

Creates an endpoint for a device and mobile app on one of the supported push notification services, such as GCM and APNS. CreatePlatformEndpoint requires the PlatformApplicationArn that is returned from CreatePlatformApplication. The EndpointArn that is returned when using CreatePlatformEndpoint can then be used by the Publish action to send a message to a mobile app or by the Subscribe action for subscription to a topic. The CreatePlatformEndpoint action is idempotent, so if the requester already owns an endpoint with the same device token and attributes, that endpoint's ARN is returned without creating a new endpoint. For more information, see Using Amazon SNS Mobile Push Notifications.

When using CreatePlatformEndpoint with Baidu, two attributes must be provided: ChannelId and UserId. The token field must also contain the ChannelId. For more information, see Creating an Amazon SNS Endpoint for Baidu.

Arguments

PlatformApplicationArn = ::String – Required

PlatformApplicationArn returned from CreatePlatformApplication is used to create a an endpoint.

Token = ::String – Required

Unique identifier created by the notification service for an app on a device. The specific name for Token will vary, depending on which notification service is being used. For example, when using APNS as the notification service, you need the device token. Alternatively, when using GCM or ADM, the device token equivalent is called the registration ID.

CustomUserData = ::String

Arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

Attributes = ::Dict{String,String}

For a list of attributes, see SetEndpointAttributes.

Returns

CreateEndpointResponse

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.create_topic — Function.

create_topic(::AWSConfig, arguments::Dict)
create_topic([::AWSConfig]; Name=)

CreateTopic Operation

Creates a topic to which notifications can be published. Users can create at most 100,000 topics. For more information, see http://aws.amazon.com/sns. This action is idempotent, so if the requester already owns a topic with the specified name, that topic's ARN is returned without creating a new topic.

Arguments

Name = ::String – Required

The name of the topic you want to create.

Constraints: Topic names must be made up of only uppercase and lowercase ASCII letters, numbers, underscores, and hyphens, and must be between 1 and 256 characters long.

Returns

CreateTopicResponse

Exceptions

InvalidParameterException, TopicLimitExceededException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.delete_endpoint — Function.

delete_endpoint(::AWSConfig, arguments::Dict)
delete_endpoint([::AWSConfig]; EndpointArn=)

DeleteEndpoint Operation

Deletes the endpoint for a device and mobile app from Amazon SNS. This action is idempotent. For more information, see Using Amazon SNS Mobile Push Notifications.

When you delete an endpoint that is also subscribed to a topic, then you must also unsubscribe the endpoint from the topic.

Arguments

EndpointArn = ::String – Required

EndpointArn of endpoint to delete.

Exceptions

InvalidParameterException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.delete_platform_application — Function.

delete_platform_application(::AWSConfig, arguments::Dict)
delete_platform_application([::AWSConfig]; PlatformApplicationArn=)

DeletePlatformApplication Operation

Deletes a platform application object for one of the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications.

Arguments

PlatformApplicationArn = ::String – Required

PlatformApplicationArn of platform application object to delete.

Exceptions

InvalidParameterException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.delete_topic — Function.

delete_topic(::AWSConfig, arguments::Dict)
delete_topic([::AWSConfig]; TopicArn=)

DeleteTopic Operation

Deletes a topic and all its subscriptions. Deleting a topic might prevent some messages previously sent to the topic from being delivered to subscribers. This action is idempotent, so deleting a topic that does not exist does not result in an error.

Arguments

TopicArn = ::String – Required

The ARN of the topic you want to delete.

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.get_endpoint_attributes — Function.

get_endpoint_attributes(::AWSConfig, arguments::Dict)
get_endpoint_attributes([::AWSConfig]; EndpointArn=)

GetEndpointAttributes Operation

Retrieves the endpoint attributes for a device on one of the supported push notification services, such as GCM and APNS. For more information, see Using Amazon SNS Mobile Push Notifications.

Arguments

EndpointArn = ::String – Required

EndpointArn for GetEndpointAttributes input.

Returns

GetEndpointAttributesResponse

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.get_platform_application_attributes — Function.

get_platform_application_attributes(::AWSConfig, arguments::Dict)
get_platform_application_attributes([::AWSConfig]; PlatformApplicationArn=)

GetPlatformApplicationAttributes Operation

Retrieves the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications.

Arguments

PlatformApplicationArn = ::String – Required

PlatformApplicationArn for GetPlatformApplicationAttributesInput.

Returns

GetPlatformApplicationAttributesResponse

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.get_smsattributes — Function.

get_smsattributes(::AWSConfig, arguments::Dict)
get_smsattributes([::AWSConfig]; <keyword arguments>)

GetSMSAttributes Operation

Returns the settings for sending SMS messages from your account.

These settings are set with the SetSMSAttributes action.

Arguments

attributes = [::String, ...]

A list of the individual attribute names, such as MonthlySpendLimit, for which you want values.

For all attribute names, see SetSMSAttributes.

If you don't use this parameter, Amazon SNS returns all SMS attributes.

Returns

GetSMSAttributesResponse

Exceptions

ThrottledException, InternalErrorException, AuthorizationErrorException or InvalidParameterException.

See also: AWS API Documentation

AmazonSNS.get_subscription_attributes — Function.

get_subscription_attributes(::AWSConfig, arguments::Dict)
get_subscription_attributes([::AWSConfig]; SubscriptionArn=)

GetSubscriptionAttributes Operation

Returns all of the properties of a subscription.

Arguments

SubscriptionArn = ::String – Required

The ARN of the subscription whose properties you want to get.

Returns

GetSubscriptionAttributesResponse

Exceptions

InvalidParameterException, InternalErrorException, NotFoundException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.get_topic_attributes — Function.

get_topic_attributes(::AWSConfig, arguments::Dict)
get_topic_attributes([::AWSConfig]; TopicArn=)

GetTopicAttributes Operation

Returns all of the properties of a topic. Topic properties returned might differ based on the authorization of the user.

Arguments

TopicArn = ::String – Required

The ARN of the topic whose properties you want to get.

Returns

GetTopicAttributesResponse

Exceptions

InvalidParameterException, InternalErrorException, NotFoundException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.list_endpoints_by_platform_application — Function.

list_endpoints_by_platform_application(::AWSConfig, arguments::Dict)
list_endpoints_by_platform_application([::AWSConfig]; PlatformApplicationArn=, <keyword arguments>)

ListEndpointsByPlatformApplication Operation

Lists the endpoints and endpoint attributes for devices in a supported push notification service, such as GCM and APNS. The results for ListEndpointsByPlatformApplication are paginated and return a limited list of endpoints, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListEndpointsByPlatformApplication again using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications.

Arguments

PlatformApplicationArn = ::String – Required

PlatformApplicationArn for ListEndpointsByPlatformApplicationInput action.

NextToken = ::String

NextToken string is used when calling ListEndpointsByPlatformApplication action to retrieve additional records that are available after the first page results.

Returns

ListEndpointsByPlatformApplicationResponse

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.list_phone_numbers_opted_out — Function.

list_phone_numbers_opted_out(::AWSConfig, arguments::Dict)
list_phone_numbers_opted_out([::AWSConfig]; <keyword arguments>)

ListPhoneNumbersOptedOut Operation

Returns a list of phone numbers that are opted out, meaning you cannot send SMS messages to them.

The results for ListPhoneNumbersOptedOut are paginated, and each page returns up to 100 phone numbers. If additional phone numbers are available after the first page of results, then a NextToken string will be returned. To receive the next page, you call ListPhoneNumbersOptedOut again using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null.

Arguments

nextToken = ::String

A NextToken string is used when you call the ListPhoneNumbersOptedOut action to retrieve additional records that are available after the first page of results.

Returns

ListPhoneNumbersOptedOutResponse

Exceptions

ThrottledException, InternalErrorException, AuthorizationErrorException or InvalidParameterException.

See also: AWS API Documentation

AmazonSNS.list_platform_applications — Function.

list_platform_applications(::AWSConfig, arguments::Dict)
list_platform_applications([::AWSConfig]; <keyword arguments>)

ListPlatformApplications Operation

Lists the platform application objects for the supported push notification services, such as APNS and GCM. The results for ListPlatformApplications are paginated and return a limited list of applications, up to 100. If additional records are available after the first page results, then a NextToken string will be returned. To receive the next page, you call ListPlatformApplications using the NextToken string received from the previous call. When there are no more records to return, NextToken will be null. For more information, see Using Amazon SNS Mobile Push Notifications.

Arguments

NextToken = ::String

NextToken string is used when calling ListPlatformApplications action to retrieve additional records that are available after the first page results.

Returns

ListPlatformApplicationsResponse

Exceptions

InvalidParameterException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.list_subscriptions — Function.

list_subscriptions(::AWSConfig, arguments::Dict)
list_subscriptions([::AWSConfig]; <keyword arguments>)

ListSubscriptions Operation

Returns a list of the requester's subscriptions. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptions call to get further results.

Arguments

NextToken = ::String

Token returned by the previous ListSubscriptions request.

Returns

ListSubscriptionsResponse

Exceptions

InvalidParameterException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.list_subscriptions_by_topic — Function.

list_subscriptions_by_topic(::AWSConfig, arguments::Dict)
list_subscriptions_by_topic([::AWSConfig]; TopicArn=, <keyword arguments>)

ListSubscriptionsByTopic Operation

Returns a list of the subscriptions to a specific topic. Each call returns a limited list of subscriptions, up to 100. If there are more subscriptions, a NextToken is also returned. Use the NextToken parameter in a new ListSubscriptionsByTopic call to get further results.

Arguments

TopicArn = ::String – Required

The ARN of the topic for which you wish to find subscriptions.

NextToken = ::String

Token returned by the previous ListSubscriptionsByTopic request.

Returns

ListSubscriptionsByTopicResponse

Exceptions

InvalidParameterException, InternalErrorException, NotFoundException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.list_topics — Function.

list_topics(::AWSConfig, arguments::Dict)
list_topics([::AWSConfig]; <keyword arguments>)

ListTopics Operation

Returns a list of the requester's topics. Each call returns a limited list of topics, up to 100. If there are more topics, a NextToken is also returned. Use the NextToken parameter in a new ListTopics call to get further results.

Arguments

NextToken = ::String

Token returned by the previous ListTopics request.

Returns

ListTopicsResponse

Exceptions

InvalidParameterException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.opt_in_phone_number — Function.

opt_in_phone_number(::AWSConfig, arguments::Dict)
opt_in_phone_number([::AWSConfig]; phoneNumber=)

OptInPhoneNumber Operation

Use this request to opt in a phone number that is opted out, which enables you to resume sending SMS messages to the number.

You can opt in a phone number only once every 30 days.

Arguments

phoneNumber = ::String – Required

The phone number to opt in.

Returns

OptInPhoneNumberResponse

Exceptions

ThrottledException, InternalErrorException, AuthorizationErrorException or InvalidParameterException.

See also: AWS API Documentation

AmazonSNS.publish — Function.

publish(::AWSConfig, arguments::Dict)
publish([::AWSConfig]; Message=, <keyword arguments>)

Publish Operation

Sends a message to all of a topic's subscribed endpoints. When a messageId is returned, the message has been saved and Amazon SNS will attempt to deliver it to the topic's subscribers shortly. The format of the outgoing message to each subscribed endpoint depends on the notification protocol.

To use the Publish action for sending a message to a mobile endpoint, such as an app on a Kindle device or mobile phone, you must specify the EndpointArn for the TargetArn parameter. The EndpointArn is returned when making a call with the CreatePlatformEndpoint action.

For more information about formatting messages, see Send Custom Platform-Specific Payloads in Messages to Mobile Devices.

Arguments

TopicArn = ::String

The topic you want to publish to.

If you don't specify a value for the TopicArn parameter, you must specify a value for the PhoneNumber or TargetArn parameters.

TargetArn = ::String

Either TopicArn or EndpointArn, but not both.

If you don't specify a value for the TargetArn parameter, you must specify a value for the PhoneNumber or TopicArn parameters.

PhoneNumber = ::String

The phone number to which you want to deliver an SMS message. Use E.164 format.

If you don't specify a value for the PhoneNumber parameter, you must specify a value for the TargetArn or TopicArn parameters.

Message = ::String – Required

The message you want to send to the topic.

If you want to send the same message to all transport protocols, include the text of the message as a String value.

If you want to send different messages for each transport protocol, set the value of the MessageStructure parameter to json and use a JSON object for the Message parameter.

Constraints: Messages must be UTF-8 encoded strings at most 256 KB in size (262144 bytes, not 262144 characters).

JSON-specific constraints:

Keys in the JSON object that correspond to supported transport protocols must have simple JSON string values.

The values will be parsed (unescaped) before they are used in outgoing messages.

Outbound notifications are JSON encoded (meaning that the characters will be reescaped for sending).

Values have a minimum length of 0 (the empty string, "", is allowed).

Values have a maximum length bounded by the overall message size (so, including multiple protocols may limit message sizes).

Non-string values will cause the key to be ignored.

Keys that do not correspond to supported transport protocols are ignored.

Duplicate keys are not allowed.

Failure to parse or validate any key or value in the message will cause the Publish call to return an error (no partial delivery).

Subject = ::String

Optional parameter to be used as the "Subject" line when the message is delivered to email endpoints. This field will also be included, if present, in the standard JSON messages delivered to other endpoints.

Constraints: Subjects must be ASCII text that begins with a letter, number, or punctuation mark; must not include line breaks or control characters; and must be less than 100 characters long.

MessageStructure = ::String

Set MessageStructure to json if you want to send a different message for each protocol. For example, using one publish action, you can send a short message to your SMS subscribers and a longer message to your email subscribers. If you set MessageStructure to json, the value of the Message parameter must:

be a syntactically valid JSON object; and

contain at least a top-level JSON key of "default" with a value that is a string.

You can define other top-level keys that define the message you want to send to a specific transport protocol (e.g., "http").

For information about sending different messages for each protocol using the AWS Management Console, go to Create Different Messages for Each Protocol in the Amazon Simple Notification Service Getting Started Guide.

Valid value: json

MessageAttributes = ::Dict{String,String}

Message attributes for Publish action.

Returns

PublishResponse

Exceptions

InvalidParameterException, InvalidParameterValueException, InternalErrorException, NotFoundException, EndpointDisabledException, PlatformApplicationDisabledException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.remove_permission — Function.

remove_permission(::AWSConfig, arguments::Dict)
remove_permission([::AWSConfig]; TopicArn=, Label=)

RemovePermission Operation

Removes a statement from a topic's access control policy.

Arguments

TopicArn = ::String – Required

The ARN of the topic whose access control policy you wish to modify.

Label = ::String – Required

The unique label of the statement you want to remove.

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.set_endpoint_attributes — Function.

set_endpoint_attributes(::AWSConfig, arguments::Dict)
set_endpoint_attributes([::AWSConfig]; EndpointArn=, Attributes=)

SetEndpointAttributes Operation

Sets the attributes for an endpoint for a device on one of the supported push notification services, such as GCM and APNS. For more information, see Using Amazon SNS Mobile Push Notifications.

Arguments

EndpointArn = ::String – Required

EndpointArn used for SetEndpointAttributes action.

Attributes = ::Dict{String,String} – Required

A map of the endpoint attributes. Attributes in this map include the following:

CustomUserData – arbitrary user data to associate with the endpoint. Amazon SNS does not use this data. The data must be in UTF-8 format and less than 2KB.

Enabled – flag that enables/disables delivery to the endpoint. Amazon SNS will set this to false when a notification service indicates to Amazon SNS that the endpoint is invalid. Users can set it back to true, typically after updating Token.

Token – device token, also referred to as a registration id, for an app and mobile device. This is returned from the notification service when an app and mobile device are registered with the notification service.

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.set_platform_application_attributes — Function.

set_platform_application_attributes(::AWSConfig, arguments::Dict)
set_platform_application_attributes([::AWSConfig]; PlatformApplicationArn=, Attributes=)

SetPlatformApplicationAttributes Operation

Sets the attributes of the platform application object for the supported push notification services, such as APNS and GCM. For more information, see Using Amazon SNS Mobile Push Notifications. For information on configuring attributes for message delivery status, see Using Amazon SNS Application Attributes for Message Delivery Status.

Arguments

PlatformApplicationArn = ::String – Required

PlatformApplicationArn for SetPlatformApplicationAttributes action.

Attributes = ::Dict{String,String} – Required

A map of the platform application attributes. Attributes in this map include the following:

PlatformCredential – The credential received from the notification service. For APNS/APNS_SANDBOX, PlatformCredential is private key. For GCM, PlatformCredential is "API key". For ADM, PlatformCredential is "client secret".

PlatformPrincipal – The principal received from the notification service. For APNS/APNS_SANDBOX, PlatformPrincipal is SSL certificate. For GCM, PlatformPrincipal is not applicable. For ADM, PlatformPrincipal is "client id".

EventEndpointCreated – Topic ARN to which EndpointCreated event notifications should be sent.

EventEndpointDeleted – Topic ARN to which EndpointDeleted event notifications should be sent.

EventEndpointUpdated – Topic ARN to which EndpointUpdate event notifications should be sent.

EventDeliveryFailure – Topic ARN to which DeliveryFailure event notifications should be sent upon Direct Publish delivery failure (permanent) to one of the application's endpoints.

SuccessFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

FailureFeedbackRoleArn – IAM role ARN used to give Amazon SNS write access to use CloudWatch Logs on your behalf.

SuccessFeedbackSampleRate – Sample rate percentage (0-100) of successfully delivered messages.

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation

AmazonSNS.set_smsattributes — Function.

set_smsattributes(::AWSConfig, arguments::Dict)
set_smsattributes([::AWSConfig]; attributes=)

SetSMSAttributes Operation

Use this request to set the default settings for sending SMS messages and receiving daily SMS usage reports.

You can override some of these settings for a single message when you use the Publish action with the MessageAttributes.entry.N parameter. For more information, see Sending an SMS Message in the Amazon SNS Developer Guide.

Arguments

attributes = ::Dict{String,String} – Required

The default settings for sending SMS messages from your account. You can set values for the following attribute names:

MonthlySpendLimit – The maximum amount in USD that you are willing to spend each month to send SMS messages. When Amazon SNS determines that sending an SMS message would incur a cost that exceeds this limit, it stops sending SMS messages within minutes.

Important

Amazon SNS stops sending SMS messages within minutes of the limit being crossed. During that interval, if you continue to send SMS messages, you will incur costs that exceed your limit.

By default, the spend limit is set to the maximum allowed by Amazon SNS. If you want to exceed the maximum, contact AWS Support or your AWS sales representative for a service limit increase.

DeliveryStatusIAMRole – The ARN of the IAM role that allows Amazon SNS to write logs about SMS deliveries in CloudWatch Logs. For each SMS message that you send, Amazon SNS writes a log that includes the message price, the success or failure status, the reason for failure (if the message failed), the message dwell time, and other information.

DeliveryStatusSuccessSamplingRate – The percentage of successful SMS deliveries for which Amazon SNS will write logs in CloudWatch Logs. The value can be an integer from 0 - 100. For example, to write logs only for failed deliveries, set this value to 0. To write logs for 10% of your successful deliveries, set it to 10.

DefaultSenderID – A string, such as your business brand, that is displayed as the sender on the receiving device. Support for sender IDs varies by country. The sender ID can be 1 - 11 alphanumeric characters, and it must contain at least one letter.

DefaultSMSType – The type of SMS message that you will send by default. You can assign the following values:

Promotional – (Default) Noncritical messages, such as marketing messages. Amazon SNS optimizes the message delivery to incur the lowest cost.

Transactional – Critical messages that support customer transactions, such as one-time passcodes for multi-factor authentication. Amazon SNS optimizes the message delivery to achieve the highest reliability.

UsageReportS3Bucket – The name of the Amazon S3 bucket to receive daily SMS usage reports from Amazon SNS. Each day, Amazon SNS will deliver a usage report as a CSV file to the bucket. The report includes the following information for each SMS message that was successfully delivered by your account:

Time that the message was published (in UTC)

Message ID

Destination phone number

Message type

Delivery status

Message price (in USD)

Part number (a message is split into multiple parts if it is too long for a single message)

Total number of parts

To receive the report, the bucket must have a policy that allows the Amazon SNS service principle to perform the s3:PutObject and s3:GetBucketLocation actions.

For an example bucket policy and usage report, see Monitoring SMS Activity in the Amazon SNS Developer Guide.

Returns

SetSMSAttributesResponse

Exceptions

InvalidParameterException, ThrottledException, InternalErrorException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.set_subscription_attributes — Function.

set_subscription_attributes(::AWSConfig, arguments::Dict)
set_subscription_attributes([::AWSConfig]; SubscriptionArn=, AttributeName=, <keyword arguments>)

SetSubscriptionAttributes Operation

Allows a subscription owner to set an attribute of the topic to a new value.

Arguments

SubscriptionArn = ::String – Required

The ARN of the subscription to modify.

AttributeName = ::String – Required

The name of the attribute you want to set. Only a subset of the subscriptions attributes are mutable.

Valid values: DeliveryPolicy | RawMessageDelivery

AttributeValue = ::String

The new value for the attribute in JSON format.

Exceptions

InvalidParameterException, InternalErrorException, NotFoundException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.set_topic_attributes — Function.

set_topic_attributes(::AWSConfig, arguments::Dict)
set_topic_attributes([::AWSConfig]; TopicArn=, AttributeName=, <keyword arguments>)

SetTopicAttributes Operation

Allows a topic owner to set an attribute of the topic to a new value.

Arguments

TopicArn = ::String – Required

The ARN of the topic to modify.

AttributeName = ::String – Required

The name of the attribute you want to set. Only a subset of the topic's attributes are mutable.

Valid values: Policy | DisplayName | DeliveryPolicy

AttributeValue = ::String

The new value for the attribute.

Exceptions

InvalidParameterException, InternalErrorException, NotFoundException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.subscribe — Function.

subscribe(::AWSConfig, arguments::Dict)
subscribe([::AWSConfig]; TopicArn=, Protocol=, <keyword arguments>)

Subscribe Operation

Prepares to subscribe an endpoint by sending the endpoint a confirmation message. To actually create a subscription, the endpoint owner must call the ConfirmSubscription action with the token from the confirmation message. Confirmation tokens are valid for three days.

Arguments

TopicArn = ::String – Required

The ARN of the topic you want to subscribe to.

Protocol = ::String – Required

The protocol you want to use. Supported protocols include:

http – delivery of JSON-encoded message via HTTP POST

https – delivery of JSON-encoded message via HTTPS POST

email – delivery of message via SMTP

email-json – delivery of JSON-encoded message via SMTP

sms – delivery of message via SMS

sqs – delivery of JSON-encoded message to an Amazon SQS queue

application – delivery of JSON-encoded message to an EndpointArn for a mobile app and device.

lambda – delivery of JSON-encoded message to an AWS Lambda function.

Endpoint = ::String

The endpoint that you want to receive notifications. Endpoints vary by protocol:

For the http protocol, the endpoint is an URL beginning with "http://"

For the https protocol, the endpoint is a URL beginning with "https://"

For the email protocol, the endpoint is an email address

For the email-json protocol, the endpoint is an email address

For the sms protocol, the endpoint is a phone number of an SMS-enabled device

For the sqs protocol, the endpoint is the ARN of an Amazon SQS queue

For the application protocol, the endpoint is the EndpointArn of a mobile app and device.

For the lambda protocol, the endpoint is the ARN of an AWS Lambda function.

Returns

SubscribeResponse

Exceptions

SubscriptionLimitExceededException, InvalidParameterException, InternalErrorException, NotFoundException or AuthorizationErrorException.

See also: AWS API Documentation

AmazonSNS.unsubscribe — Function.

unsubscribe(::AWSConfig, arguments::Dict)
unsubscribe([::AWSConfig]; SubscriptionArn=)

Unsubscribe Operation

Deletes a subscription. If the subscription requires authentication for deletion, only the owner of the subscription or the topic's owner can unsubscribe, and an AWS signature is required. If the Unsubscribe call does not require authentication and the requester is not the subscription owner, a final cancellation message is delivered to the endpoint, so that the endpoint owner can easily resubscribe to the topic if the Unsubscribe request was unintended.

Arguments

SubscriptionArn = ::String – Required

The ARN of the subscription to be deleted.

Exceptions

InvalidParameterException, InternalErrorException, AuthorizationErrorException or NotFoundException.

See also: AWS API Documentation