AmazonStepFunctions.jl

AmazonStepFunctions

AWS Step Functions

AWS Step Functions is a web service that enables you to coordinate the components of distributed applications and microservices using visual workflows. You build applications from individual components that each perform a discrete function, or task, allowing you to scale and change applications quickly. Step Functions provides a graphical console to visualize the components of your application as a series of steps. It automatically triggers and tracks each step, and retries when there are errors, so your application executes in order and as expected, every time. Step Functions logs the state of each step, so when things do go wrong, you can diagnose and debug problems quickly.

Step Functions manages the operations and underlying infrastructure for you to ensure your application is available at any scale. You can run tasks on the AWS cloud, on your own servers, or an any system that has access to AWS. Step Functions can be accessed and used with the Step Functions console, the AWS SDKs (included with your Beta release invitation email), or an HTTP API (the subject of this document).

This document is generated from apis/states-2016-11-23.normal.json. See JuliaCloud/AWSCore.jl.

AmazonStepFunctions.create_activity

AmazonStepFunctions.create_state_machine

AmazonStepFunctions.delete_activity

AmazonStepFunctions.delete_state_machine

AmazonStepFunctions.describe_activity

AmazonStepFunctions.describe_execution

AmazonStepFunctions.describe_state_machine

AmazonStepFunctions.get_activity_task

AmazonStepFunctions.get_execution_history

AmazonStepFunctions.list_activities

AmazonStepFunctions.list_executions

AmazonStepFunctions.list_state_machines

AmazonStepFunctions.send_task_failure

AmazonStepFunctions.send_task_heartbeat

AmazonStepFunctions.send_task_success

AmazonStepFunctions.start_execution

AmazonStepFunctions.stop_execution

AmazonStepFunctions.create_activity — Function.

using AmazonStepFunctions
create_activity([::AWSConfig], arguments::Dict)
create_activity([::AWSConfig]; name=)

using AWSCore.Services.states
states([::AWSConfig], "CreateActivity", arguments::Dict)
states([::AWSConfig], "CreateActivity", name=)

CreateActivity Operation

Creates an activity.

Arguments

name = ::String – Required

The name of the activity to create. This name must be unique for your AWS account and region.

Returns

CreateActivityOutput

Exceptions

ActivityLimitExceeded or InvalidName.

See also: AWS API Documentation

AmazonStepFunctions.create_state_machine — Function.

using AmazonStepFunctions
create_state_machine([::AWSConfig], arguments::Dict)
create_state_machine([::AWSConfig]; name=, definition=, roleArn=)

using AWSCore.Services.states
states([::AWSConfig], "CreateStateMachine", arguments::Dict)
states([::AWSConfig], "CreateStateMachine", name=, definition=, roleArn=)

CreateStateMachine Operation

Creates a state machine.

Arguments

name = ::String – Required

The name of the state machine. This name must be unique for your AWS account and region.

definition = ::String – Required

The Amazon States Language definition of the state machine.

roleArn = ::String – Required

The Amazon Resource Name (ARN) of the IAM role to use for this state machine.

Returns

CreateStateMachineOutput

Exceptions

InvalidArn, InvalidDefinition, InvalidName, StateMachineAlreadyExists, StateMachineDeleting or StateMachineLimitExceeded.

See also: AWS API Documentation

AmazonStepFunctions.delete_activity — Function.

using AmazonStepFunctions
delete_activity([::AWSConfig], arguments::Dict)
delete_activity([::AWSConfig]; activityArn=)

using AWSCore.Services.states
states([::AWSConfig], "DeleteActivity", arguments::Dict)
states([::AWSConfig], "DeleteActivity", activityArn=)

DeleteActivity Operation

Deletes an activity.

Arguments

activityArn = ::String – Required

The Amazon Resource Name (ARN) of the activity to delete.

Returns

DeleteActivityOutput

Exceptions

InvalidArn.

See also: AWS API Documentation

AmazonStepFunctions.delete_state_machine — Function.

using AmazonStepFunctions
delete_state_machine([::AWSConfig], arguments::Dict)
delete_state_machine([::AWSConfig]; stateMachineArn=)

using AWSCore.Services.states
states([::AWSConfig], "DeleteStateMachine", arguments::Dict)
states([::AWSConfig], "DeleteStateMachine", stateMachineArn=)

DeleteStateMachine Operation

Deletes a state machine. This is an asynchronous operation– it sets the state machine's status to "DELETING" and begins the delete process.

Arguments

stateMachineArn = ::String – Required

The Amazon Resource Name (ARN) of the state machine to delete.

Returns

DeleteStateMachineOutput

Exceptions

InvalidArn.

See also: AWS API Documentation

AmazonStepFunctions.describe_activity — Function.

using AmazonStepFunctions
describe_activity([::AWSConfig], arguments::Dict)
describe_activity([::AWSConfig]; activityArn=)

using AWSCore.Services.states
states([::AWSConfig], "DescribeActivity", arguments::Dict)
states([::AWSConfig], "DescribeActivity", activityArn=)

DescribeActivity Operation

Describes an activity.

Arguments

activityArn = ::String – Required

The Amazon Resource Name (ARN) of the activity to describe.

Returns

DescribeActivityOutput

Exceptions

ActivityDoesNotExist or InvalidArn.

See also: AWS API Documentation

AmazonStepFunctions.describe_execution — Function.

using AmazonStepFunctions
describe_execution([::AWSConfig], arguments::Dict)
describe_execution([::AWSConfig]; executionArn=)

using AWSCore.Services.states
states([::AWSConfig], "DescribeExecution", arguments::Dict)
states([::AWSConfig], "DescribeExecution", executionArn=)

DescribeExecution Operation

Describes an execution.

Arguments

executionArn = ::String – Required

The Amazon Resource Name (ARN) of the execution to describe.

Returns

DescribeExecutionOutput

Exceptions

ExecutionDoesNotExist or InvalidArn.

See also: AWS API Documentation

AmazonStepFunctions.describe_state_machine — Function.

using AmazonStepFunctions
describe_state_machine([::AWSConfig], arguments::Dict)
describe_state_machine([::AWSConfig]; stateMachineArn=)

using AWSCore.Services.states
states([::AWSConfig], "DescribeStateMachine", arguments::Dict)
states([::AWSConfig], "DescribeStateMachine", stateMachineArn=)

DescribeStateMachine Operation

Describes a state machine.

Arguments

stateMachineArn = ::String – Required

The Amazon Resource Name (ARN) of the state machine to describe.

Returns

DescribeStateMachineOutput

Exceptions

InvalidArn or StateMachineDoesNotExist.

See also: AWS API Documentation

AmazonStepFunctions.get_activity_task — Function.

using AmazonStepFunctions
get_activity_task([::AWSConfig], arguments::Dict)
get_activity_task([::AWSConfig]; activityArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "GetActivityTask", arguments::Dict)
states([::AWSConfig], "GetActivityTask", activityArn=, <keyword arguments>)

GetActivityTask Operation

Used by workers to retrieve a task (with the specified activity ARN) scheduled for execution by a running state machine. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available (i.e. an execution of a task of this type is needed.) The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll will return an empty result, that is, the taskToken returned is an empty string.

Important

Workers should set their client side socket timeout to at least 65 seconds (5 seconds higher than the maximum time the service may hold the poll request).

Arguments

activityArn = ::String – Required

The Amazon Resource Name (ARN) of the activity to retrieve tasks from.

workerName = ::String

An arbitrary name may be provided in order to identify the worker that the task is assigned to. This name will be used when it is logged in the execution history.

Returns

GetActivityTaskOutput

Exceptions

ActivityDoesNotExist, ActivityWorkerLimitExceeded or InvalidArn.

See also: AWS API Documentation

AmazonStepFunctions.get_execution_history — Function.

using AmazonStepFunctions
get_execution_history([::AWSConfig], arguments::Dict)
get_execution_history([::AWSConfig]; executionArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "GetExecutionHistory", arguments::Dict)
states([::AWSConfig], "GetExecutionHistory", executionArn=, <keyword arguments>)

GetExecutionHistory Operation

Returns the history of the specified execution as a list of events. By default, the results are returned in ascending order of the timeStamp of the events. Use the reverseOrder parameter to get the latest events first. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

Arguments

executionArn = ::String – Required

The Amazon Resource Name (ARN) of the execution.

maxResults = ::Int

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

reverseOrder = ::Bool

Lists events in descending order of their timeStamp.

nextToken = ::String

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken. Keep all other arguments unchanged.

The configured maxResults determines how many results can be returned in a single call.

Returns

GetExecutionHistoryOutput

Exceptions

ExecutionDoesNotExist, InvalidArn or InvalidToken.

See also: AWS API Documentation

AmazonStepFunctions.list_activities — Function.

using AmazonStepFunctions
list_activities([::AWSConfig], arguments::Dict)
list_activities([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "ListActivities", arguments::Dict)
states([::AWSConfig], "ListActivities", <keyword arguments>)

ListActivities Operation

Lists the existing activities. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

Arguments

maxResults = ::Int

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

nextToken = ::String

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken. Keep all other arguments unchanged.

The configured maxResults determines how many results can be returned in a single call.

Returns

ListActivitiesOutput

Exceptions

InvalidToken.

See also: AWS API Documentation

AmazonStepFunctions.list_executions — Function.

using AmazonStepFunctions
list_executions([::AWSConfig], arguments::Dict)
list_executions([::AWSConfig]; stateMachineArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "ListExecutions", arguments::Dict)
states([::AWSConfig], "ListExecutions", stateMachineArn=, <keyword arguments>)

ListExecutions Operation

Lists the executions of a state machine that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

Arguments

stateMachineArn = ::String – Required

The Amazon Resource Name (ARN) of the state machine whose executions will be listed.

statusFilter = "RUNNING", "SUCCEEDED", "FAILED", "TIMED_OUT" or "ABORTED"

If specified, only list the executions whose current execution status matches the given filter.

maxResults = ::Int

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

nextToken = ::String

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken. Keep all other arguments unchanged.

The configured maxResults determines how many results can be returned in a single call.

Returns

ListExecutionsOutput

Exceptions

InvalidArn, InvalidToken or StateMachineDoesNotExist.

See also: AWS API Documentation

AmazonStepFunctions.list_state_machines — Function.

using AmazonStepFunctions
list_state_machines([::AWSConfig], arguments::Dict)
list_state_machines([::AWSConfig]; <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "ListStateMachines", arguments::Dict)
states([::AWSConfig], "ListStateMachines", <keyword arguments>)

ListStateMachines Operation

Lists the existing state machines. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

Arguments

maxResults = ::Int

The maximum number of results that will be returned per call. nextToken can be used to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000.

This is an upper limit only; the actual number of results returned per call may be fewer than the specified maximum.

nextToken = ::String

If a nextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in nextToken. Keep all other arguments unchanged.

The configured maxResults determines how many results can be returned in a single call.

Returns

ListStateMachinesOutput

Exceptions

InvalidToken.

See also: AWS API Documentation

AmazonStepFunctions.send_task_failure — Function.

using AmazonStepFunctions
send_task_failure([::AWSConfig], arguments::Dict)
send_task_failure([::AWSConfig]; taskToken=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "SendTaskFailure", arguments::Dict)
states([::AWSConfig], "SendTaskFailure", taskToken=, <keyword arguments>)

SendTaskFailure Operation

Used by workers to report that the task identified by the taskToken failed.

Arguments

taskToken = ::String – Required

The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

error = ::String

An arbitrary error code that identifies the cause of the failure.

cause = ::String

A more detailed explanation of the cause of the failure.

Returns

SendTaskFailureOutput

Exceptions

TaskDoesNotExist, InvalidToken or TaskTimedOut.

See also: AWS API Documentation

AmazonStepFunctions.send_task_heartbeat — Function.

using AmazonStepFunctions
send_task_heartbeat([::AWSConfig], arguments::Dict)
send_task_heartbeat([::AWSConfig]; taskToken=)

using AWSCore.Services.states
states([::AWSConfig], "SendTaskHeartbeat", arguments::Dict)
states([::AWSConfig], "SendTaskHeartbeat", taskToken=)

SendTaskHeartbeat Operation

Used by workers to report to the service that the task represented by the specified taskToken is still making progress. This action resets the Heartbeat clock. The Heartbeat threshold is specified in the state machine's Amazon States Language definition. This action does not in itself create an event in the execution history. However, if the task times out, the execution history will contain an ActivityTimedOut event.

Note

The Timeout of a task, defined in the state machine's Amazon States Language definition, is its maximum allowed duration, regardless of the number of SendTaskHeartbeat requests received.

Note

This operation is only useful for long-lived tasks to report the liveliness of the task.

Arguments

taskToken = ::String – Required

The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

Returns

SendTaskHeartbeatOutput

Exceptions

TaskDoesNotExist, InvalidToken or TaskTimedOut.

See also: AWS API Documentation

AmazonStepFunctions.send_task_success — Function.

using AmazonStepFunctions
send_task_success([::AWSConfig], arguments::Dict)
send_task_success([::AWSConfig]; taskToken=, output=)

using AWSCore.Services.states
states([::AWSConfig], "SendTaskSuccess", arguments::Dict)
states([::AWSConfig], "SendTaskSuccess", taskToken=, output=)

SendTaskSuccess Operation

Used by workers to report that the task identified by the taskToken completed successfully.

Arguments

taskToken = ::String – Required

The token that represents this task. Task tokens are generated by the service when the tasks are assigned to a worker (see GetActivityTask::taskToken).

output = ::String – Required

The JSON output of the task.

Returns

SendTaskSuccessOutput

Exceptions

TaskDoesNotExist, InvalidOutput, InvalidToken or TaskTimedOut.

See also: AWS API Documentation

AmazonStepFunctions.start_execution — Function.

using AmazonStepFunctions
start_execution([::AWSConfig], arguments::Dict)
start_execution([::AWSConfig]; stateMachineArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "StartExecution", arguments::Dict)
states([::AWSConfig], "StartExecution", stateMachineArn=, <keyword arguments>)

StartExecution Operation

Starts a state machine execution.

Arguments

stateMachineArn = ::String – Required

The Amazon Resource Name (ARN) of the state machine to execute.

name = ::String

The name of the execution. This name must be unique for your AWS account and region.

input = ::String

The JSON input data for the execution.

Returns

StartExecutionOutput

Exceptions

ExecutionLimitExceeded, ExecutionAlreadyExists, InvalidArn, InvalidExecutionInput, InvalidName, StateMachineDoesNotExist or StateMachineDeleting.

See also: AWS API Documentation

AmazonStepFunctions.stop_execution — Function.

using AmazonStepFunctions
stop_execution([::AWSConfig], arguments::Dict)
stop_execution([::AWSConfig]; executionArn=, <keyword arguments>)

using AWSCore.Services.states
states([::AWSConfig], "StopExecution", arguments::Dict)
states([::AWSConfig], "StopExecution", executionArn=, <keyword arguments>)

StopExecution Operation

Stops an execution.

Arguments

executionArn = ::String – Required

The Amazon Resource Name (ARN) of the execution to stop.

error = ::String

An arbitrary error code that identifies the cause of the termination.

cause = ::String

A more detailed explanation of the cause of the termination.

Returns

StopExecutionOutput

Exceptions

ExecutionDoesNotExist or InvalidArn.

See also: AWS API Documentation