AmazonWorkDocs.jl

AmazonWorkDocs

The WorkDocs API is designed for the following use cases:

File Migration: File migration applications are supported for users who want to migrate their files from an on-premise or off-premise file system or service. Users can insert files into a user directory structure, as well as allow for basic metadata changes, such as modifications to the permissions of files.

Security: Support security applications are supported for users who have additional security needs, such as anti-virus or data loss prevention. The APIs, in conjunction with Amazon CloudTrail, allow these applications to detect when changes occur in Amazon WorkDocs, so the application can take the necessary actions and replace the target file. The application can also choose to email the user if the target file violates the policy.

eDiscovery/Analytics: General administrative applications are supported, such as eDiscovery and analytics. These applications can choose to mimic and/or record the actions in an Amazon WorkDocs site, in conjunction with Amazon CloudTrails, to replicate data for eDiscovery, backup, or analytical applications.

All Amazon WorkDocs APIs are Amazon authenticated, certificate-signed APIs. They not only require the use of the AWS SDK, but also allow for the exclusive use of IAM users and roles to help facilitate access, trust, and permission policies. By creating a role and allowing an IAM user to access the Amazon WorkDocs site, the IAM user gains full administrative visibility into the entire Amazon WorkDocs site (or as set in the IAM policy). This includes, but is not limited to, the ability to modify file permissions and upload any file to any user. This allows developers to perform the three use cases above, as well as give users the ability to grant access on a selective basis using the IAM model.

This document is generated from apis/workdocs-2016-05-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonWorkDocs.abort_document_version_upload

AmazonWorkDocs.activate_user

AmazonWorkDocs.add_resource_permissions

AmazonWorkDocs.create_comment

AmazonWorkDocs.create_custom_metadata

AmazonWorkDocs.create_folder

AmazonWorkDocs.create_labels

AmazonWorkDocs.create_notification_subscription

AmazonWorkDocs.create_user

AmazonWorkDocs.deactivate_user

AmazonWorkDocs.delete_comment

AmazonWorkDocs.delete_custom_metadata

AmazonWorkDocs.delete_document

AmazonWorkDocs.delete_folder

AmazonWorkDocs.delete_folder_contents

AmazonWorkDocs.delete_labels

AmazonWorkDocs.delete_notification_subscription

AmazonWorkDocs.delete_user

AmazonWorkDocs.describe_activities

AmazonWorkDocs.describe_comments

AmazonWorkDocs.describe_document_versions

AmazonWorkDocs.describe_folder_contents

AmazonWorkDocs.describe_notification_subscriptions

AmazonWorkDocs.describe_resource_permissions

AmazonWorkDocs.describe_root_folders

AmazonWorkDocs.describe_users

AmazonWorkDocs.get_current_user

AmazonWorkDocs.get_document

AmazonWorkDocs.get_document_path

AmazonWorkDocs.get_document_version

AmazonWorkDocs.get_folder

AmazonWorkDocs.get_folder_path

AmazonWorkDocs.initiate_document_version_upload

AmazonWorkDocs.remove_all_resource_permissions

AmazonWorkDocs.remove_resource_permission

AmazonWorkDocs.update_document

AmazonWorkDocs.update_document_version

AmazonWorkDocs.update_folder

AmazonWorkDocs.update_user

AmazonWorkDocs.abort_document_version_upload — Function.

abort_document_version_upload(::AWSConfig, arguments::Dict)
abort_document_version_upload([::AWSConfig]; DocumentId=, VersionId=, <keyword arguments>)

AbortDocumentVersionUpload Operation

Aborts the upload of the specified document version that was previously initiated by InitiateDocumentVersionUpload. The client should make this call only when it no longer intends to upload the document version, or fails to do so.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

VersionId = ::String – Required

The ID of the version.

Exceptions

EntityNotExistsException, ProhibitedStateException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.activate_user — Function.

activate_user(::AWSConfig, arguments::Dict)
activate_user([::AWSConfig]; UserId=, <keyword arguments>)

ActivateUser Operation

Activates the specified user. Only active users can access Amazon WorkDocs.

Arguments

UserId = ::String – Required

The ID of the user.

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

Returns

ActivateUserResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.add_resource_permissions — Function.

add_resource_permissions(::AWSConfig, arguments::Dict)
add_resource_permissions([::AWSConfig]; ResourceId=, Principals=, <keyword arguments>)

AddResourcePermissions Operation

Creates a set of permissions for the specified folder or document. The resource permissions are overwritten if the principals already have different permissions.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ResourceId = ::String – Required

The ID of the resource.

Principals = [[ ... ], ...] – Required

The users, groups, or organization being granted permission.

 Principals = [[
        "Id" => <required> ::String,
        "Type" => <required> "USER", "GROUP", "INVITE", "ANONYMOUS" or "ORGANIZATION",
        "Role" => <required> "VIEWER", "CONTRIBUTOR", "OWNER" or "COOWNER"
    ], ...]

Returns

AddResourcePermissionsResponse

Exceptions

UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.create_comment — Function.

create_comment(::AWSConfig, arguments::Dict)
create_comment([::AWSConfig]; DocumentId=, VersionId=, Text=, <keyword arguments>)

CreateComment Operation

Adds a new comment to the specified document version.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

VersionId = ::String – Required

The ID of the document version.

ParentId = ::String

The ID of the parent comment.

ThreadId = ::String

The ID of the root comment in the thread.

Text = ::String – Required

The text of the comment.

Visibility = "PUBLIC" or "PRIVATE"

The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

NotifyCollaborators = ::Bool

Set this parameter to TRUE to send an email out to the document collaborators after the comment is created.

Returns

CreateCommentResponse

Exceptions

EntityNotExistsException, ProhibitedStateException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException, ServiceUnavailableException or DocumentLockedForCommentsException.

See also: AWS API Documentation

AmazonWorkDocs.create_custom_metadata — Function.

create_custom_metadata(::AWSConfig, arguments::Dict)
create_custom_metadata([::AWSConfig]; ResourceId=, CustomMetadata=, <keyword arguments>)

CreateCustomMetadata Operation

Adds one or more custom properties to the specified resource (a folder, document, or version).

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ResourceId = ::String – Required

The ID of the resource.

versionid = ::String

The ID of the version, if the custom metadata is being added to a document version.

CustomMetadata = ::Dict{String,String} – Required

Custom metadata in the form of name-value pairs.

Returns

CreateCustomMetadataResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, ProhibitedStateException, CustomMetadataLimitExceededException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.create_folder — Function.

create_folder(::AWSConfig, arguments::Dict)
create_folder([::AWSConfig]; ParentFolderId=, <keyword arguments>)

CreateFolder Operation

Creates a folder with the specified name and parent folder.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

Name = ::String

The name of the new folder.

ParentFolderId = ::String – Required

The ID of the parent folder.

Returns

CreateFolderResponse

Exceptions

EntityNotExistsException, EntityAlreadyExistsException, ProhibitedStateException, LimitExceededException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.create_labels — Function.

create_labels(::AWSConfig, arguments::Dict)
create_labels([::AWSConfig]; ResourceId=, Labels=, <keyword arguments>)

CreateLabels Operation

Adds the specified list of labels to the given resource (a document or folder)

Arguments

ResourceId = ::String – Required

The ID of the resource.

Labels = [::String, ...] – Required

List of labels to add to the resource.

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

Returns

CreateLabelsResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException, ServiceUnavailableException or TooManyLabelsException.

See also: AWS API Documentation

AmazonWorkDocs.create_notification_subscription — Function.

create_notification_subscription(::AWSConfig, arguments::Dict)
create_notification_subscription([::AWSConfig]; OrganizationId=, Endpoint=, Protocol=, SubscriptionType=)

CreateNotificationSubscription Operation

Configure WorkDocs to use Amazon SNS notifications.

The endpoint receives a confirmation message, and must confirm the subscription. For more information, see Confirm the Subscription in the Amazon Simple Notification Service Developer Guide.

Arguments

OrganizationId = ::String – Required

The ID of the organization.

Endpoint = ::String – Required

The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with "https://".

Protocol = "HTTPS" – Required

The protocol to use. The supported value is https, which delivers JSON-encoded messasges using HTTPS POST.

SubscriptionType = "ALL" – Required

The notification type.

Returns

CreateNotificationSubscriptionResponse

Exceptions

UnauthorizedResourceAccessException, TooManySubscriptionsException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.create_user — Function.

create_user(::AWSConfig, arguments::Dict)
create_user([::AWSConfig]; Username=, GivenName=, Surname=, Password=, <keyword arguments>)

CreateUser Operation

Creates a user in a Simple AD or Microsoft AD directory. The status of a newly created user is "ACTIVE". New users can access Amazon WorkDocs.

Arguments

OrganizationId = ::String

The ID of the organization.

Username = ::String – Required

The login name of the user.

EmailAddress = ::String

The email address of the user.

GivenName = ::String – Required

The given name of the user.

Surname = ::String – Required

The surname of the user.

Password = ::String – Required

The password of the user.

TimeZoneId = ::String

The time zone ID of the user.

StorageRule = [ ... ]

The amount of storage for the user.

 StorageRule = [
        "StorageAllocatedInBytes" =>  ::Int,
        "StorageType" =>  "UNLIMITED" or "QUOTA"
    ]

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

Returns

CreateUserResponse

Exceptions

EntityAlreadyExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.deactivate_user — Function.

deactivate_user(::AWSConfig, arguments::Dict)
deactivate_user([::AWSConfig]; UserId=, <keyword arguments>)

DeactivateUser Operation

Deactivates the specified user, which revokes the user's access to Amazon WorkDocs.

Arguments

UserId = ::String – Required

The ID of the user.

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.delete_comment — Function.

delete_comment(::AWSConfig, arguments::Dict)
delete_comment([::AWSConfig]; DocumentId=, VersionId=, CommentId=, <keyword arguments>)

DeleteComment Operation

Deletes the specified comment from the document version.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

VersionId = ::String – Required

The ID of the document version.

CommentId = ::String – Required

The ID of the comment.

Exceptions

EntityNotExistsException, ProhibitedStateException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException, ServiceUnavailableException or DocumentLockedForCommentsException.

See also: AWS API Documentation

AmazonWorkDocs.delete_custom_metadata — Function.

delete_custom_metadata(::AWSConfig, arguments::Dict)
delete_custom_metadata([::AWSConfig]; ResourceId=, <keyword arguments>)

DeleteCustomMetadata Operation

Deletes custom metadata from the specified resource.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ResourceId = ::String – Required

The ID of the resource, either a document or folder.

versionId = ::String

The ID of the version, if the custom metadata is being deleted from a document version.

keys = [::String, ...]

List of properties to remove.

deleteAll = ::Bool

Flag to indicate removal of all custom metadata properties from the specified resource.

Returns

DeleteCustomMetadataResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, ProhibitedStateException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.delete_document — Function.

delete_document(::AWSConfig, arguments::Dict)
delete_document([::AWSConfig]; DocumentId=, <keyword arguments>)

DeleteDocument Operation

Permanently deletes the specified document and its associated metadata.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

Exceptions

EntityNotExistsException, ProhibitedStateException, ConcurrentModificationException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.delete_folder — Function.

delete_folder(::AWSConfig, arguments::Dict)
delete_folder([::AWSConfig]; FolderId=, <keyword arguments>)

DeleteFolder Operation

Permanently deletes the specified folder and its contents.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

FolderId = ::String – Required

The ID of the folder.

Exceptions

EntityNotExistsException, ProhibitedStateException, ConcurrentModificationException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.delete_folder_contents — Function.

delete_folder_contents(::AWSConfig, arguments::Dict)
delete_folder_contents([::AWSConfig]; FolderId=, <keyword arguments>)

DeleteFolderContents Operation

Deletes the contents of the specified folder.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

FolderId = ::String – Required

The ID of the folder.

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.delete_labels — Function.

delete_labels(::AWSConfig, arguments::Dict)
delete_labels([::AWSConfig]; ResourceId=, <keyword arguments>)

DeleteLabels Operation

Deletes the specified list of labels from a resource.

Arguments

ResourceId = ::String – Required

The ID of the resource.

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

labels = [::String, ...]

List of labels to delete from the resource.

deleteAll = ::Bool

Flag to request removal of all labels from the specified resource.

Returns

DeleteLabelsResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.delete_notification_subscription — Function.

delete_notification_subscription(::AWSConfig, arguments::Dict)
delete_notification_subscription([::AWSConfig]; SubscriptionId=, OrganizationId=)

DeleteNotificationSubscription Operation

Deletes the specified subscription from the specified organization.

Arguments

SubscriptionId = ::String – Required

The ID of the subscription.

OrganizationId = ::String – Required

The ID of the organization.

Exceptions

UnauthorizedResourceAccessException, EntityNotExistsException, ServiceUnavailableException or ProhibitedStateException.

See also: AWS API Documentation

AmazonWorkDocs.delete_user — Function.

delete_user(::AWSConfig, arguments::Dict)
delete_user([::AWSConfig]; UserId=, <keyword arguments>)

DeleteUser Operation

Deletes the specified user from a Simple AD or Microsoft AD directory.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

UserId = ::String – Required

The ID of the user.

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.describe_activities — Function.

describe_activities(::AWSConfig, arguments::Dict)
describe_activities([::AWSConfig]; <keyword arguments>)

DescribeActivities Operation

Describes the user activities in a specified time period.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

startTime = timestamp

The timestamp that determines the starting time of the activities; the response includes the activities performed after the specified timestamp.

endTime = timestamp

The timestamp that determines the end time of the activities; the response includes the activities performed before the specified timestamp.

organizationId = ::String

The ID of the organization. This is a mandatory parameter when using administrative API (SigV4) requests.

userId = ::String

The ID of the user who performed the action. The response includes activities pertaining to this user. This is an optional parameter and is only applicable for administrative API (SigV4) requests.

limit = ::Int

The maximum number of items to return.

marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

Returns

DescribeActivitiesResponse

Exceptions

UnauthorizedOperationException, UnauthorizedResourceAccessException, InvalidArgumentException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.describe_comments — Function.

describe_comments(::AWSConfig, arguments::Dict)
describe_comments([::AWSConfig]; DocumentId=, VersionId=, <keyword arguments>)

DescribeComments Operation

List all the comments for the specified document version.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

VersionId = ::String – Required

The ID of the document version.

limit = ::Int

The maximum number of items to return.

marker = ::String

The marker for the next set of results. This marker was received from a previous call.

Returns

DescribeCommentsResponse

Exceptions

EntityNotExistsException, ProhibitedStateException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.describe_document_versions — Function.

describe_document_versions(::AWSConfig, arguments::Dict)
describe_document_versions([::AWSConfig]; DocumentId=, <keyword arguments>)

DescribeDocumentVersions Operation

Retrieves the document versions for the specified document.

By default, only active versions are returned.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

limit = ::Int

The maximum number of versions to return with this call.

include = ::String

A comma-separated list of values. Specify "INITIALIZED" to include incomplete versions.

fields = ::String

Specify "SOURCE" to include initialized versions and a URL for the source document.

Returns

DescribeDocumentVersionsResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, InvalidArgumentException, FailedDependencyException, ServiceUnavailableException or ProhibitedStateException.

See also: AWS API Documentation

AmazonWorkDocs.describe_folder_contents — Function.

describe_folder_contents(::AWSConfig, arguments::Dict)
describe_folder_contents([::AWSConfig]; FolderId=, <keyword arguments>)

DescribeFolderContents Operation

Describes the contents of the specified folder, including its documents and subfolders.

By default, Amazon WorkDocs returns the first 100 active document and folder metadata items. If there are more results, the response includes a marker that you can use to request the next set of results. You can also request initialized documents.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

FolderId = ::String – Required

The ID of the folder.

sort = "DATE" or "NAME"

The sorting criteria.

order = "ASCENDING" or "DESCENDING"

The order for the contents of the folder.

limit = ::Int

The maximum number of items to return with this call.

marker = ::String

The marker for the next set of results. This marker was received from a previous call.

type = "ALL", "DOCUMENT" or "FOLDER"

The type of items.

include = ::String

The contents to include. Specify "INITIALIZED" to include initialized documents.

Returns

DescribeFolderContentsResponse

Exceptions

EntityNotExistsException, UnauthorizedResourceAccessException, InvalidArgumentException, FailedDependencyException, ServiceUnavailableException or ProhibitedStateException.

See also: AWS API Documentation

AmazonWorkDocs.describe_notification_subscriptions — Function.

describe_notification_subscriptions(::AWSConfig, arguments::Dict)
describe_notification_subscriptions([::AWSConfig]; OrganizationId=, <keyword arguments>)

DescribeNotificationSubscriptions Operation

Lists the specified notification subscriptions.

Arguments

OrganizationId = ::String – Required

The ID of the organization.

marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

limit = ::Int

The maximum number of items to return with this call.

Returns

DescribeNotificationSubscriptionsResponse

Exceptions

UnauthorizedResourceAccessException, EntityNotExistsException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.describe_resource_permissions — Function.

describe_resource_permissions(::AWSConfig, arguments::Dict)
describe_resource_permissions([::AWSConfig]; ResourceId=, <keyword arguments>)

DescribeResourcePermissions Operation

Describes the permissions of a specified resource.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ResourceId = ::String – Required

The ID of the resource.

limit = ::Int

The maximum number of items to return with this call.

marker = ::String

The marker for the next set of results. (You received this marker from a previous call)

Returns

DescribeResourcePermissionsResponse

Exceptions

UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.describe_root_folders — Function.

describe_root_folders(::AWSConfig, arguments::Dict)
describe_root_folders([::AWSConfig]; Authentication=, <keyword arguments>)

DescribeRootFolders Operation

Describes the current user's special folders; the RootFolder and the RecyleBin. RootFolder is the root of user's files and folders and RecyleBin is the root of recycled items. This is not a valid action for SigV4 (administrative API) clients.

Arguments

Authentication = ::String – Required

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

limit = ::Int

The maximum number of items to return.

marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

Returns

DescribeRootFoldersResponse

Exceptions

UnauthorizedOperationException, UnauthorizedResourceAccessException, InvalidArgumentException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.describe_users — Function.

describe_users(::AWSConfig, arguments::Dict)
describe_users([::AWSConfig]; <keyword arguments>)

DescribeUsers Operation

Describes the specified users. You can describe all users or filter the results (for example, by status or organization).

By default, Amazon WorkDocs returns the first 24 active or pending users. If there are more results, the response includes a marker that you can use to request the next set of results.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

organizationId = ::String

The ID of the organization.

userIds = ::String

The IDs of the users.

query = ::String

A query to filter users by user name.

include = "ALL" or "ACTIVE_PENDING"

The state of the users. Specify "ALL" to include inactive users.

order = "ASCENDING" or "DESCENDING"

The order for the results.

sort = "USER_NAME", "FULL_NAME", "STORAGE_LIMIT", "USER_STATUS" or "STORAGE_USED"

The sorting criteria.

marker = ::String

The marker for the next set of results. (You received this marker from a previous call.)

limit = ::Int

The maximum number of items to return.

fields = ::String

A comma-separated list of values. Specify "STORAGE_METADATA" to include the user storage quota and utilization information.

Returns

DescribeUsersResponse

Exceptions

UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException, ServiceUnavailableException or InvalidArgumentException.

See also: AWS API Documentation

AmazonWorkDocs.get_current_user — Function.

get_current_user(::AWSConfig, arguments::Dict)
get_current_user([::AWSConfig]; Authentication=)

GetCurrentUser Operation

Retrieves details of the current user for whom the authentication token was generated. This is not a valid action for SigV4 (administrative API) clients.

Arguments

Authentication = ::String – Required

Amazon WorkDocs authentication token.

Returns

GetCurrentUserResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.get_document — Function.

get_document(::AWSConfig, arguments::Dict)
get_document([::AWSConfig]; DocumentId=, <keyword arguments>)

GetDocument Operation

Retrieves details of a document.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

includeCustomMetadata = ::Bool

Set this to TRUE to include custom metadata in the response.

Returns

GetDocumentResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, InvalidArgumentException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.get_document_path — Function.

get_document_path(::AWSConfig, arguments::Dict)
get_document_path([::AWSConfig]; DocumentId=, <keyword arguments>)

GetDocumentPath Operation

Retrieves the path information (the hierarchy from the root folder) for the requested document.

By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested document and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the names of the parent folders.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

limit = ::Int

The maximum number of levels in the hierarchy to return.

fields = ::String

A comma-separated list of values. Specify NAME to include the names of the parent folders.

marker = ::String

This value is not supported.

Returns

GetDocumentPathResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.get_document_version — Function.

get_document_version(::AWSConfig, arguments::Dict)
get_document_version([::AWSConfig]; DocumentId=, VersionId=, <keyword arguments>)

GetDocumentVersion Operation

Retrieves version metadata for the specified document.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

VersionId = ::String – Required

The version ID of the document.

fields = ::String

A comma-separated list of values. Specify "SOURCE" to include a URL for the source document.

includeCustomMetadata = ::Bool

Set this to TRUE to include custom metadata in the response.

Returns

GetDocumentVersionResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException, ServiceUnavailableException or ProhibitedStateException.

See also: AWS API Documentation

AmazonWorkDocs.get_folder — Function.

get_folder(::AWSConfig, arguments::Dict)
get_folder([::AWSConfig]; FolderId=, <keyword arguments>)

GetFolder Operation

Retrieves the metadata of the specified folder.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

FolderId = ::String – Required

The ID of the folder.

includeCustomMetadata = ::Bool

Set to TRUE to include custom metadata in the response.

Returns

GetFolderResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, InvalidArgumentException, FailedDependencyException, ServiceUnavailableException or ProhibitedStateException.

See also: AWS API Documentation

AmazonWorkDocs.get_folder_path — Function.

get_folder_path(::AWSConfig, arguments::Dict)
get_folder_path([::AWSConfig]; FolderId=, <keyword arguments>)

GetFolderPath Operation

Retrieves the path information (the hierarchy from the root folder) for the specified folder.

By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested folder and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the parent folder names.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

FolderId = ::String – Required

The ID of the folder.

limit = ::Int

The maximum number of levels in the hierarchy to return.

fields = ::String

A comma-separated list of values. Specify "NAME" to include the names of the parent folders.

marker = ::String

This value is not supported.

Returns

GetFolderPathResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.initiate_document_version_upload — Function.

initiate_document_version_upload(::AWSConfig, arguments::Dict)
initiate_document_version_upload([::AWSConfig]; ParentFolderId=, <keyword arguments>)

InitiateDocumentVersionUpload Operation

Creates a new document object and version object.

The client specifies the parent folder ID and name of the document to upload. The ID is optionally specified when creating a new version of an existing document. This is the first step to upload a document. Next, upload the document to the URL returned from the call, and then call UpdateDocumentVersion.

To cancel the document upload, call AbortDocumentVersionUpload.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

Id = ::String

The ID of the document.

Name = ::String

The name of the document.

ContentCreatedTimestamp = timestamp

The time stamp when the content of the document was originally created.

ContentModifiedTimestamp = timestamp

The time stamp when the content of the document was modified.

ContentType = ::String

The content type of the document.

DocumentSizeInBytes = ::Int

The size of the document, in bytes.

ParentFolderId = ::String – Required

The ID of the parent folder.

Returns

InitiateDocumentVersionUploadResponse

Exceptions

EntityNotExistsException, EntityAlreadyExistsException, StorageLimitExceededException, StorageLimitWillExceedException, ProhibitedStateException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException, ServiceUnavailableException, DraftUploadOutOfSyncException or ResourceAlreadyCheckedOutException.

See also: AWS API Documentation

AmazonWorkDocs.remove_all_resource_permissions — Function.

remove_all_resource_permissions(::AWSConfig, arguments::Dict)
remove_all_resource_permissions([::AWSConfig]; ResourceId=, <keyword arguments>)

RemoveAllResourcePermissions Operation

Removes all the permissions from the specified resource.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ResourceId = ::String – Required

The ID of the resource.

Exceptions

UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.remove_resource_permission — Function.

remove_resource_permission(::AWSConfig, arguments::Dict)
remove_resource_permission([::AWSConfig]; ResourceId=, PrincipalId=, <keyword arguments>)

RemoveResourcePermission Operation

Removes the permission for the specified principal from the specified resource.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

ResourceId = ::String – Required

The ID of the resource.

PrincipalId = ::String – Required

The principal ID of the resource.

type = "USER", "GROUP", "INVITE", "ANONYMOUS" or "ORGANIZATION"

The principal type of the resource.

Exceptions

UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.update_document — Function.

update_document(::AWSConfig, arguments::Dict)
update_document([::AWSConfig]; DocumentId=, <keyword arguments>)

UpdateDocument Operation

Updates the specified attributes of a document. The user must have access to both the document and its parent folder, if applicable.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

Name = ::String

The name of the document.

ParentFolderId = ::String

The ID of the parent folder.

ResourceState = "ACTIVE", "RESTORING", "RECYCLING" or "RECYCLED"

The resource state of the document. Note that only ACTIVE and RECYCLED are supported.

Exceptions

EntityNotExistsException, EntityAlreadyExistsException, LimitExceededException, ProhibitedStateException, ConcurrentModificationException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.update_document_version — Function.

update_document_version(::AWSConfig, arguments::Dict)
update_document_version([::AWSConfig]; DocumentId=, VersionId=, <keyword arguments>)

UpdateDocumentVersion Operation

Changes the status of the document version to ACTIVE.

Amazon WorkDocs also sets its document container to ACTIVE. This is the last step in a document upload, after the client uploads the document to an S3-presigned URL returned by InitiateDocumentVersionUpload.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

DocumentId = ::String – Required

The ID of the document.

VersionId = ::String – Required

The version ID of the document.

VersionStatus = "ACTIVE"

The status of the version.

Exceptions

EntityNotExistsException, ProhibitedStateException, ConcurrentModificationException, InvalidOperationException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.update_folder — Function.

update_folder(::AWSConfig, arguments::Dict)
update_folder([::AWSConfig]; FolderId=, <keyword arguments>)

UpdateFolder Operation

Updates the specified attributes of the specified folder. The user must have access to both the folder and its parent folder, if applicable.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

FolderId = ::String – Required

The ID of the folder.

Name = ::String

The name of the folder.

ParentFolderId = ::String

The ID of the parent folder.

ResourceState = "ACTIVE", "RESTORING", "RECYCLING" or "RECYCLED"

The resource state of the folder. Note that only ACTIVE and RECYCLED are accepted values from the API.

Exceptions

EntityNotExistsException, EntityAlreadyExistsException, ProhibitedStateException, ConcurrentModificationException, LimitExceededException, UnauthorizedOperationException, UnauthorizedResourceAccessException, FailedDependencyException or ServiceUnavailableException.

See also: AWS API Documentation

AmazonWorkDocs.update_user — Function.

update_user(::AWSConfig, arguments::Dict)
update_user([::AWSConfig]; UserId=, <keyword arguments>)

UpdateUser Operation

Updates the specified attributes of the specified user, and grants or revokes administrative privileges to the Amazon WorkDocs site.

Arguments

Authentication = ::String

Amazon WorkDocs authentication token. This field should not be set when using administrative API actions, as in accessing the API using AWS credentials.

UserId = ::String – Required

The ID of the user.

GivenName = ::String

The given name of the user.

Surname = ::String

The surname of the user.

Type = "USER" or "ADMIN"

The type of the user.

StorageRule = [ ... ]

The amount of storage for the user.

 StorageRule = [
        "StorageAllocatedInBytes" =>  ::Int,
        "StorageType" =>  "UNLIMITED" or "QUOTA"
    ]

TimeZoneId = ::String

The time zone ID of the user.

Locale = "en", "fr", "ko", "de", "es", "ja", "ru", "zh_CN", "zh_TW", "pt_BR" or "default"

The locale of the user.

Returns

UpdateUserResponse

Exceptions

EntityNotExistsException, UnauthorizedOperationException, UnauthorizedResourceAccessException, IllegalUserStateException, FailedDependencyException, ServiceUnavailableException or DeactivatingLastSystemUserException.

See also: AWS API Documentation