AmazonXRay.jl

AmazonXRay

AWS X-Ray provides APIs for managing debug traces and retrieving service maps and other data created by processing those traces.

This document is generated from apis/xray-2016-04-12.normal.json. See JuliaCloud/AWSCore.jl.

AmazonXRay.batch_get_traces

AmazonXRay.get_service_graph

AmazonXRay.get_trace_graph

AmazonXRay.get_trace_summaries

AmazonXRay.put_telemetry_records

AmazonXRay.put_trace_segments

AmazonXRay.batch_get_traces — Function.

using AmazonXRay
batch_get_traces([::AWSConfig], arguments::Dict)
batch_get_traces([::AWSConfig]; TraceIds=, <keyword arguments>)

using AWSCore.Services.xray
request([::AWSConfig], "POST", "/Traces", args)

BatchGetTraces Operation

Retrieves a list of traces specified by ID. Each trace is a collection of segment documents that originates from a single request. Use GetTraceSummaries to get a list of trace IDs.

Arguments

TraceIds = [::String, ...] – Required

Specify the trace IDs of requests for which to retrieve segments.

NextToken = ::String

Pagination token. Not used.

Returns

BatchGetTracesResult

Exceptions

InvalidRequestException or ThrottledException.

See also: AWS API Documentation

AmazonXRay.get_service_graph — Function.

using AmazonXRay
get_service_graph([::AWSConfig], arguments::Dict)
get_service_graph([::AWSConfig]; StartTime=, EndTime=, <keyword arguments>)

using AWSCore.Services.xray
request([::AWSConfig], "POST", "/ServiceGraph", args)

GetServiceGraph Operation

Retrieves a document that describes services that process incoming requests, and downstream services that they call as a result. Root services process incoming requests and make calls to downstream services. Root services are applications that use the AWS X-Ray SDK. Downstream services can be other applications, AWS resources, HTTP web APIs, or SQL databases.

Arguments

StartTime = timestamp – Required

The start of the time frame for which to generate a graph.

EndTime = timestamp – Required

The end of the time frame for which to generate a graph.

NextToken = ::String

Pagination token. Not used.

Returns

GetServiceGraphResult

Exceptions

InvalidRequestException or ThrottledException.

See also: AWS API Documentation

AmazonXRay.get_trace_graph — Function.

using AmazonXRay
get_trace_graph([::AWSConfig], arguments::Dict)
get_trace_graph([::AWSConfig]; TraceIds=, <keyword arguments>)

using AWSCore.Services.xray
request([::AWSConfig], "POST", "/TraceGraph", args)

GetTraceGraph Operation

Retrieves a service graph for one or more specific trace IDs.

Arguments

TraceIds = [::String, ...] – Required

Trace IDs of requests for which to generate a service graph.

NextToken = ::String

Pagination token. Not used.

Returns

GetTraceGraphResult

Exceptions

InvalidRequestException or ThrottledException.

See also: AWS API Documentation

AmazonXRay.get_trace_summaries — Function.

using AmazonXRay
get_trace_summaries([::AWSConfig], arguments::Dict)
get_trace_summaries([::AWSConfig]; StartTime=, EndTime=, <keyword arguments>)

using AWSCore.Services.xray
request([::AWSConfig], "POST", "/TraceSummaries", args)

GetTraceSummaries Operation

Retrieves IDs and metadata for traces available for a specified time frame using an optional filter. To get the full traces, pass the trace IDs to BatchGetTraces.

A filter expression can target traced requests that hit specific service nodes or edges, have errors, or come from a known user. For example, the following filter expression targets traces that pass through api.example.com:

service("api.example.com")

This filter expression finds traces that have an annotation named account with the value 12345:

annotation.account = "12345"

For a full list of indexed fields and keywords that you can use in filter expressions, see Using Filter Expressions in the AWS X-Ray Developer Guide.

Arguments

StartTime = timestamp – Required

The start of the time frame for which to retrieve traces.

EndTime = timestamp – Required

The end of the time frame for which to retrieve traces.

Sampling = ::Bool

Set to true to get summaries for only a subset of available traces.

FilterExpression = ::String

Specify a filter expression to retrieve trace summaries for services or requests that meet certain requirements.

NextToken = ::String

Specify the pagination token returned by a previous request to retrieve the next page of results.

Returns

GetTraceSummariesResult

Exceptions

InvalidRequestException or ThrottledException.

See also: AWS API Documentation

AmazonXRay.put_telemetry_records — Function.

using AmazonXRay
put_telemetry_records([::AWSConfig], arguments::Dict)
put_telemetry_records([::AWSConfig]; TelemetryRecords=, <keyword arguments>)

using AWSCore.Services.xray
request([::AWSConfig], "POST", "/TelemetryRecords", args)

PutTelemetryRecords Operation

Used by the AWS X-Ray daemon to upload telemetry.

Arguments

TelemetryRecords = [[ ... ], ...] – Required

 TelemetryRecords = [[
        "Timestamp" =>  timestamp,
        "SegmentsReceivedCount" =>  ::Int,
        "SegmentsSentCount" =>  ::Int,
        "SegmentsSpilloverCount" =>  ::Int,
        "SegmentsRejectedCount" =>  ::Int,
        "BackendConnectionErrors" =>  [
            "TimeoutCount" =>  ::Int,
            "ConnectionRefusedCount" =>  ::Int,
            "HTTPCode4XXCount" =>  ::Int,
            "HTTPCode5XXCount" =>  ::Int,
            "UnknownHostCount" =>  ::Int,
            "OtherCount" =>  ::Int
        ]
    ], ...]

EC2InstanceId = ::String

Hostname = ::String

ResourceARN = ::String

Returns

PutTelemetryRecordsResult

Exceptions

InvalidRequestException or ThrottledException.

See also: AWS API Documentation

AmazonXRay.put_trace_segments — Function.

using AmazonXRay
put_trace_segments([::AWSConfig], arguments::Dict)
put_trace_segments([::AWSConfig]; TraceSegmentDocuments=)

using AWSCore.Services.xray
request([::AWSConfig], "POST", "/TraceSegments", args)

PutTraceSegments Operation

Uploads segment documents to AWS X-Ray. The X-Ray SDK generates segment documents and sends them to the X-Ray daemon, which uploads them in batches. A segment document can be a completed segment, an in-progress segment, or an array of subsegments.

Segments must include the following fields. For the full segment document schema, see AWS X-Ray Segment Documents in the AWS X-Ray Developer Guide.

Required Segment Document Fields

name - The name of the service that handled the request.

id - A 64-bit identifier for the segment, unique among segments in the same trace, in 16 hexadecimal digits.

trace_id - A unique identifier that connects all segments and subsegments originating from a single client request.

start_time - Time the segment or subsegment was created, in floating point seconds in epoch time, accurate to milliseconds. For example, 1480615200.010 or 1.480615200010E9.

end_time - Time the segment or subsegment was closed. For example, 1480615200.090 or 1.480615200090E9. Specify either an end_time or in_progress.

in_progress - Set to true instead of specifying an end_time to record that a segment has been started, but is not complete. Send an in progress segment when your application receives a request that will take a long time to serve, to trace the fact that the request was received. When the response is sent, send the complete segment to overwrite the in-progress segment.

A trace_id consists of three numbers separated by hyphens. For example, 1-58406520-a006649127e371903a2de979. This includes:

Trace ID Format

The version number, i.e. 1.

The time of the original request, in Unix epoch time, in 8 hexadecimal digits. For example, 10:00AM December 2nd, 2016 PST in epoch time is 1480615200 seconds, or 58406520 in hexadecimal.

A 96-bit identifier for the trace, globally unique, in 24 hexadecimal digits.

Arguments

TraceSegmentDocuments = [::String, ...] – Required

A string containing a JSON document defining one or more segments or subsegments.

Returns

PutTraceSegmentsResult

Exceptions

InvalidRequestException or ThrottledException.

See also: AWS API Documentation