AmazonApplicationAutoScaling.jl

AmazonApplicationAutoScaling

With Application Auto Scaling, you can automatically scale your AWS resources. The experience similar to that of Auto Scaling. You can use Application Auto Scaling to accomplish the following tasks:

Define scaling policies to automatically scale your AWS resources

Scale your resources in response to CloudWatch alarms

View the history of your scaling events

Application Auto Scaling can scale the following AWS resources:

Amazon ECS services. For more information, see Service Auto Scaling in the Amazon EC2 Container Service Developer Guide.

Amazon EC2 Spot fleets. For more information, see Automatic Scaling for Spot Fleet in the Amazon EC2 User Guide.

Amazon EMR clusters. For more information, see Using Automatic Scaling in Amazon EMR in the Amazon EMR Management Guide.

AppStream 2.0 fleets. For more information, see Fleet Auto Scaling for Amazon AppStream 2.0 in the Amazon AppStream 2.0 Developer Guide.

Provisioned read and write capacity for Amazon DynamoDB tables and global secondary indexes. For more information, see Auto Scaling for DynamoDB in the Amazon DynamoDB Developer Guide.

For a list of supported regions, see AWS Regions and Endpoints: Application Auto Scaling in the AWS General Reference.

This document is generated from apis/application-autoscaling-2016-02-06.normal.json. See JuliaCloud/AWSCore.jl.

AmazonApplicationAutoScaling.delete_scaling_policy

AmazonApplicationAutoScaling.deregister_scalable_target

AmazonApplicationAutoScaling.describe_scalable_targets

AmazonApplicationAutoScaling.describe_scaling_activities

AmazonApplicationAutoScaling.describe_scaling_policies

AmazonApplicationAutoScaling.put_scaling_policy

AmazonApplicationAutoScaling.register_scalable_target

AmazonApplicationAutoScaling.delete_scaling_policy — Method.

delete_scaling_policy(::AWSConfig, arguments::Dict)
delete_scaling_policy(::AWSConfig; PolicyName=, ServiceNamespace=, ResourceId=, ScalableDimension=)

DeleteScalingPolicy Operation

Deletes the specified Application Auto Scaling scaling policy.

Deleting a policy deletes the underlying alarm action, but does not delete the CloudWatch alarm associated with the scaling policy, even if it no longer has an associated action.

To create a scaling policy or update an existing one, see PutScalingPolicy.

Arguments

PolicyName::String – Required

The name of the scaling policy.

ServiceNamespace: "ecs", "elasticmapreduce", "ec2", "appstream" or "dynamodb" – Required

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

ResourceId::String – Required

The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

ScalableDimension: "ecs:service:DesiredCount", "ec2:spot-fleet-request:TargetCapacity", "elasticmapreduce:instancegroup:InstanceCount", "appstream:fleet:DesiredCapacity", "dynamodb:table:ReadCapacityUnits", "dynamodb:table:WriteCapacityUnits", "dynamodb:index:ReadCapacityUnits" or "dynamodb:index:WriteCapacityUnits" – Required

The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

Returns

DeleteScalingPolicyResponse

Exceptions

ValidationException, ObjectNotFoundException, ConcurrentUpdateException or InternalServiceException.

See also: AWS API Documentation

AmazonApplicationAutoScaling.deregister_scalable_target — Method.

deregister_scalable_target(::AWSConfig, arguments::Dict)
deregister_scalable_target(::AWSConfig; ServiceNamespace=, ResourceId=, ScalableDimension=)

DeregisterScalableTarget Operation

Deregisters a scalable target.

Deregistering a scalable target deletes the scaling policies that are associated with it.

To create a scalable target or update an existing one, see RegisterScalableTarget.

Arguments

ServiceNamespace: "ecs", "elasticmapreduce", "ec2", "appstream" or "dynamodb" – Required

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

ResourceId::String – Required

The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

ScalableDimension: "ecs:service:DesiredCount", "ec2:spot-fleet-request:TargetCapacity", "elasticmapreduce:instancegroup:InstanceCount", "appstream:fleet:DesiredCapacity", "dynamodb:table:ReadCapacityUnits", "dynamodb:table:WriteCapacityUnits", "dynamodb:index:ReadCapacityUnits" or "dynamodb:index:WriteCapacityUnits" – Required

The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

Returns

DeregisterScalableTargetResponse

Exceptions

ValidationException, ObjectNotFoundException, ConcurrentUpdateException or InternalServiceException.

See also: AWS API Documentation

AmazonApplicationAutoScaling.describe_scalable_targets — Method.

describe_scalable_targets(::AWSConfig, arguments::Dict)
describe_scalable_targets(::AWSConfig; ServiceNamespace=, <keyword arguments>)

DescribeScalableTargets Operation

Provides descriptive information about the scalable targets in the specified namespace.

You can filter the results using the ResourceIds and ScalableDimension parameters.

To create a scalable target or update an existing one, see RegisterScalableTarget. If you are no longer using a scalable target, you can deregister it using DeregisterScalableTarget.

Arguments

ServiceNamespace: "ecs", "elasticmapreduce", "ec2", "appstream" or "dynamodb" – Required

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

ResourceIds: [::String...]

The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier. If you specify a scalable dimension, you must also specify a resource ID.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

ScalableDimension: "ecs:service:DesiredCount", "ec2:spot-fleet-request:TargetCapacity", "elasticmapreduce:instancegroup:InstanceCount", "appstream:fleet:DesiredCapacity", "dynamodb:table:ReadCapacityUnits", "dynamodb:table:WriteCapacityUnits", "dynamodb:index:ReadCapacityUnits" or "dynamodb:index:WriteCapacityUnits"

The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property. If you specify a scalable dimension, you must also specify a resource ID.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

MaxResults::Int

The maximum number of scalable target results. This value can be between 1 and 50. The default value is 50.

If this parameter is used, the operation returns up to MaxResults results at a time, along with a NextToken value. To get the next set of results, include the NextToken value in a subsequent call. If this parameter is not used, the operation returns up to 50 results and a NextToken value, if applicable.

NextToken::String

The token for the next set of results.

Returns

DescribeScalableTargetsResponse

Exceptions

ValidationException, InvalidNextTokenException, ConcurrentUpdateException or InternalServiceException.

See also: AWS API Documentation

AmazonApplicationAutoScaling.describe_scaling_activities — Method.

describe_scaling_activities(::AWSConfig, arguments::Dict)
describe_scaling_activities(::AWSConfig; ServiceNamespace=, <keyword arguments>)

DescribeScalingActivities Operation

Provides descriptive information about the scaling activities in the specified namespace from the previous six weeks.

You can filter the results using the ResourceId and ScalableDimension parameters.

Scaling activities are triggered by CloudWatch alarms that are associated with scaling policies. To view the scaling policies for a service namespace, see DescribeScalingPolicies. To create a scaling policy or update an existing one, see PutScalingPolicy.

Arguments

ServiceNamespace: "ecs", "elasticmapreduce", "ec2", "appstream" or "dynamodb" – Required

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

ResourceId::String

The identifier of the resource associated with the scaling activity. This string consists of the resource type and unique identifier. If you specify a scalable dimension, you must also specify a resource ID.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

ScalableDimension: "ecs:service:DesiredCount", "ec2:spot-fleet-request:TargetCapacity", "elasticmapreduce:instancegroup:InstanceCount", "appstream:fleet:DesiredCapacity", "dynamodb:table:ReadCapacityUnits", "dynamodb:table:WriteCapacityUnits", "dynamodb:index:ReadCapacityUnits" or "dynamodb:index:WriteCapacityUnits"

The scalable dimension. This string consists of the service namespace, resource type, and scaling property. If you specify a scalable dimension, you must also specify a resource ID.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

MaxResults::Int

The maximum number of scalable target results. This value can be between 1 and 50. The default value is 50.

If this parameter is used, the operation returns up to MaxResults results at a time, along with a NextToken value. To get the next set of results, include the NextToken value in a subsequent call. If this parameter is not used, the operation returns up to 50 results and a NextToken value, if applicable.

NextToken::String

The token for the next set of results.

Returns

DescribeScalingActivitiesResponse

Exceptions

ValidationException, InvalidNextTokenException, ConcurrentUpdateException or InternalServiceException.

See also: AWS API Documentation

AmazonApplicationAutoScaling.describe_scaling_policies — Method.

describe_scaling_policies(::AWSConfig, arguments::Dict)
describe_scaling_policies(::AWSConfig; ServiceNamespace=, <keyword arguments>)

DescribeScalingPolicies Operation

Provides descriptive information about the scaling policies in the specified namespace.

You can filter the results using the ResourceId, ScalableDimension, and PolicyNames parameters.

To create a scaling policy or update an existing one, see PutScalingPolicy. If you are no longer using a scaling policy, you can delete it using DeleteScalingPolicy.

Arguments

PolicyNames: [::String...]

The names of the scaling policies to describe.

ServiceNamespace: "ecs", "elasticmapreduce", "ec2", "appstream" or "dynamodb" – Required

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

ResourceId::String

The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier. If you specify a scalable dimension, you must also specify a resource ID.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

ScalableDimension: "ecs:service:DesiredCount", "ec2:spot-fleet-request:TargetCapacity", "elasticmapreduce:instancegroup:InstanceCount", "appstream:fleet:DesiredCapacity", "dynamodb:table:ReadCapacityUnits", "dynamodb:table:WriteCapacityUnits", "dynamodb:index:ReadCapacityUnits" or "dynamodb:index:WriteCapacityUnits"

The scalable dimension. This string consists of the service namespace, resource type, and scaling property. If you specify a scalable dimension, you must also specify a resource ID.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

MaxResults::Int

The maximum number of scalable target results. This value can be between 1 and 50. The default value is 50.

If this parameter is used, the operation returns up to MaxResults results at a time, along with a NextToken value. To get the next set of results, include the NextToken value in a subsequent call. If this parameter is not used, the operation returns up to 50 results and a NextToken value, if applicable.

NextToken::String

The token for the next set of results.

Returns

DescribeScalingPoliciesResponse

Exceptions

ValidationException, FailedResourceAccessException, InvalidNextTokenException, ConcurrentUpdateException or InternalServiceException.

See also: AWS API Documentation

AmazonApplicationAutoScaling.put_scaling_policy — Method.

put_scaling_policy(::AWSConfig, arguments::Dict)
put_scaling_policy(::AWSConfig; PolicyName=, ServiceNamespace=, ResourceId=, ScalableDimension=, <keyword arguments>)

PutScalingPolicy Operation

Creates or updates a policy for an Application Auto Scaling scalable target.

Each scalable target is identified by a service namespace, resource ID, and scalable dimension. A scaling policy applies to the scalable target identified by those three attributes. You cannot create a scaling policy without first registering a scalable target using RegisterScalableTarget.

To update a policy, specify its policy name and the parameters that you want to change. Any parameters that you don't specify are not changed by this update request.

You can view the scaling policies for a service namespace using DescribeScalingPolicies. If you are no longer using a scaling policy, you can delete it using DeleteScalingPolicy.

Arguments

PolicyName::String – Required

The name of the scaling policy.

ServiceNamespace: "ecs", "elasticmapreduce", "ec2", "appstream" or "dynamodb" – Required

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

ResourceId::String – Required

The identifier of the resource associated with the scaling policy. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

ScalableDimension: "ecs:service:DesiredCount", "ec2:spot-fleet-request:TargetCapacity", "elasticmapreduce:instancegroup:InstanceCount", "appstream:fleet:DesiredCapacity", "dynamodb:table:ReadCapacityUnits", "dynamodb:table:WriteCapacityUnits", "dynamodb:index:ReadCapacityUnits" or "dynamodb:index:WriteCapacityUnits" – Required

The scalable dimension. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

PolicyType: "StepScaling" or "TargetTrackingScaling"

The policy type. If you are creating a new policy, this parameter is required. If you are updating a policy, this parameter is not required.

For DynamoDB, only TargetTrackingScaling is supported. For any other service, only StepScaling is supported.

StepScalingPolicyConfiguration:

A step scaling policy.

This parameter is required if you are creating a policy and the policy type is StepScaling.

    Dict(
        AdjustmentType  => : "ChangeInCapacity", "PercentChangeInCapacity" or "ExactCapacity",
        StepAdjustments  => [Dict(
            MetricIntervalLowerBound  => double,
            MetricIntervalUpperBound  => double,
            ScalingAdjustment * => ::Int
        )...],
        MinAdjustmentMagnitude  => ::Int,
        Cooldown  => ::Int,
        MetricAggregationType  => : "Average", "Minimum" or "Maximum"
    )

TargetTrackingScalingPolicyConfiguration:

A target tracking policy.

This parameter is required if you are creating a new policy and the policy type is TargetTrackingScaling.

    Dict(
        TargetValue * => double,
        PredefinedMetricSpecification  => Dict(
            PredefinedMetricType * => : "DynamoDBReadCapacityUtilization" or "DynamoDBWriteCapacityUtilization",
            ResourceLabel  => ::String
        ),
        CustomizedMetricSpecification  => Dict(
            MetricName * => ::String,
            Namespace * => ::String,
            Dimensions  => [Dict(
                Name * => ::String,
                Value * => ::String
            )...],
            Statistic * => : "Average", "Minimum", "Maximum", "SampleCount" or "Sum",
            Unit  => ::String
        ),
        ScaleOutCooldown  => ::Int,
        ScaleInCooldown  => ::Int
    )

Returns

PutScalingPolicyResponse

Exceptions

ValidationException, LimitExceededException, ObjectNotFoundException, ConcurrentUpdateException, FailedResourceAccessException or InternalServiceException.

See also: AWS API Documentation

AmazonApplicationAutoScaling.register_scalable_target — Method.

register_scalable_target(::AWSConfig, arguments::Dict)
register_scalable_target(::AWSConfig; ServiceNamespace=, ResourceId=, ScalableDimension=, <keyword arguments>)

RegisterScalableTarget Operation

Registers or updates a scalable target. A scalable target is a resource that Application Auto Scaling can scale out or scale in. After you have registered a scalable target, you can use this operation to update the minimum and maximum values for your scalable dimension.

After you register a scalable target, you can create and apply scaling policies using PutScalingPolicy. You can view the scaling policies for a service namespace using DescribeScalableTargets. If you are no longer using a scalable target, you can deregister it using DeregisterScalableTarget.

Arguments

ServiceNamespace: "ecs", "elasticmapreduce", "ec2", "appstream" or "dynamodb" – Required

The namespace of the AWS service. For more information, see AWS Service Namespaces in the Amazon Web Services General Reference.

ResourceId::String – Required

The identifier of the resource associated with the scalable target. This string consists of the resource type and unique identifier.

ECS service - The resource type is service and the unique identifier is the cluster name and service name. Example: service/default/sample-webapp.

Spot fleet request - The resource type is spot-fleet-request and the unique identifier is the Spot fleet request ID. Example: spot-fleet-request/sfr-73fbd2ce-aa30-494c-8788-1cee4EXAMPLE.

EMR cluster - The resource type is instancegroup and the unique identifier is the cluster ID and instance group ID. Example: instancegroup/j-2EEZNYKUA1NTV/ig-1791Y4E1L8YI0.

AppStream 2.0 fleet - The resource type is fleet and the unique identifier is the fleet name. Example: fleet/sample-fleet.

DynamoDB table - The resource type is table and the unique identifier is the resource ID. Example: table/my-table.

DynamoDB global secondary index - The resource type is index and the unique identifier is the resource ID. Example: table/my-table/index/my-table-index.

ScalableDimension: "ecs:service:DesiredCount", "ec2:spot-fleet-request:TargetCapacity", "elasticmapreduce:instancegroup:InstanceCount", "appstream:fleet:DesiredCapacity", "dynamodb:table:ReadCapacityUnits", "dynamodb:table:WriteCapacityUnits", "dynamodb:index:ReadCapacityUnits" or "dynamodb:index:WriteCapacityUnits" – Required

The scalable dimension associated with the scalable target. This string consists of the service namespace, resource type, and scaling property.

ecs:service:DesiredCount - The desired task count of an ECS service.

ec2:spot-fleet-request:TargetCapacity - The target capacity of a Spot fleet request.

elasticmapreduce:instancegroup:InstanceCount - The instance count of an EMR Instance Group.

appstream:fleet:DesiredCapacity - The desired capacity of an AppStream 2.0 fleet.

dynamodb:table:ReadCapacityUnits - The provisioned read capacity for a DynamoDB table.

dynamodb:table:WriteCapacityUnits - The provisioned write capacity for a DynamoDB table.

dynamodb:index:ReadCapacityUnits - The provisioned read capacity for a DynamoDB global secondary index.

dynamodb:index:WriteCapacityUnits - The provisioned write capacity for a DynamoDB global secondary index.

MinCapacity::Int

The minimum value to scale to in response to a scale in event. This parameter is required if you are registering a scalable target and optional if you are updating one.

MaxCapacity::Int

The maximum value to scale to in response to a scale out event. This parameter is required if you are registering a scalable target and optional if you are updating one.

RoleARN::String

The ARN of an IAM role that allows Application Auto Scaling to modify the scalable target on your behalf. This parameter is required when you register a scalable target and optional when you update one.

Returns

RegisterScalableTargetResponse

Exceptions

ValidationException, LimitExceededException, ConcurrentUpdateException or InternalServiceException.

See also: AWS API Documentation