AmazonBatch.jl

AmazonBatch

AWS Batch enables you to run batch computing workloads on the AWS Cloud. Batch computing is a common way for developers, scientists, and engineers to access large amounts of compute resources, and AWS Batch removes the undifferentiated heavy lifting of configuring and managing the required infrastructure. AWS Batch will be familiar to users of traditional batch computing software. This service can efficiently provision resources in response to jobs submitted in order to eliminate capacity constraints, reduce compute costs, and deliver results quickly.

As a fully managed service, AWS Batch enables developers, scientists, and engineers to run batch computing workloads of any scale. AWS Batch automatically provisions compute resources and optimizes the workload distribution based on the quantity and scale of the workloads. With AWS Batch, there is no need to install or manage batch computing software, which allows you to focus on analyzing results and solving problems. AWS Batch reduces operational complexities, saves time, and reduces costs, which makes it easy for developers, scientists, and engineers to run their batch jobs in the AWS Cloud.

This document is generated from apis/batch-2016-08-10.normal.json. See JuliaCloud/AWSCore.jl.

AmazonBatch.cancel_job

AmazonBatch.create_compute_environment

AmazonBatch.create_job_queue

AmazonBatch.delete_compute_environment

AmazonBatch.delete_job_queue

AmazonBatch.deregister_job_definition

AmazonBatch.describe_compute_environments

AmazonBatch.describe_job_definitions

AmazonBatch.describe_job_queues

AmazonBatch.describe_jobs

AmazonBatch.list_jobs

AmazonBatch.register_job_definition

AmazonBatch.submit_job

AmazonBatch.terminate_job

AmazonBatch.update_compute_environment

AmazonBatch.update_job_queue

AmazonBatch.cancel_job — Function.

cancel_job(::AWSConfig, arguments::Dict)
cancel_job([::AWSConfig]; jobId=, reason=)

CancelJob Operation

Cancels jobs in an AWS Batch job queue. Jobs that are in the SUBMITTED, PENDING, or RUNNABLE state are cancelled. Jobs that have progressed to STARTING or RUNNING are not cancelled (but the API operation still succeeds, even if no jobs are cancelled); these jobs must be terminated with the TerminateJob operation.

Arguments

jobId = ::String – Required

A list of up to 100 job IDs to cancel.

reason = ::String – Required

A message to attach to the job that explains the reason for cancelling it. This message is returned by future DescribeJobs operations on the job. This message is also recorded in the AWS Batch activity logs.

Returns

CancelJobResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.create_compute_environment — Function.

create_compute_environment(::AWSConfig, arguments::Dict)
create_compute_environment([::AWSConfig]; computeEnvironmentName=, type=, serviceRole=, <keyword arguments>)

CreateComputeEnvironment Operation

Creates an AWS Batch compute environment. You can create MANAGED or UNMANAGED compute environments.

In a managed compute environment, AWS Batch manages the compute resources within the environment, based on the compute resources that you specify. Instances launched into a managed compute environment use the latest Amazon ECS-optimized AMI. You can choose to use Amazon EC2 On-Demand instances in your managed compute environment, or you can use Amazon EC2 Spot instances that only launch when the Spot bid price is below a specified percentage of the On-Demand price.

In an unmanaged compute environment, you can manage your own compute resources. This provides more compute resource configuration options, such as using a custom AMI, but you must ensure that your AMI meets the Amazon ECS container instance AMI specification. For more information, see Container Instance AMIs in the Amazon EC2 Container Service Developer Guide. After you have created your unmanaged compute environment, you can use the DescribeComputeEnvironments operation to find the Amazon ECS cluster that is associated with it and then manually launch your container instances into that Amazon ECS cluster. For more information, see Launching an Amazon ECS Container Instance in the Amazon EC2 Container Service Developer Guide.

Arguments

computeEnvironmentName = ::String – Required

The name for your compute environment. Up to 128 letters (uppercase and lowercase), numbers, and underscores are allowed.

type = "MANAGED" or "UNMANAGED" – Required

The type of the compute environment.

state = "ENABLED" or "DISABLED"

The state of the compute environment. If the state is ENABLED, then the compute environment accepts jobs from a queue and can scale out automatically based on queues.

computeResources = Dict( ... )

Details of the compute resources managed by the compute environment. This parameter is required for managed compute environments.

 computeResources = Dict(
        "type" => <required> "EC2" or "SPOT",
        "minvCpus" => <required> ::Int,
        "maxvCpus" => <required> ::Int,
        "desiredvCpus" =>  ::Int,
        "instanceTypes" => <required> [::String...],
        "imageId" =>  ::String,
        "subnets" => <required> [::String...],
        "securityGroupIds" => <required> [::String...],
        "ec2KeyPair" =>  ::String,
        "instanceRole" => <required> ::String,
        "tags" =>  ::Dict{String,String},
        "bidPercentage" =>  ::Int,
        "spotIamFleetRole" =>  ::String
    )

serviceRole = ::String – Required

The full Amazon Resource Name (ARN) of the IAM role that allows AWS Batch to make calls to other AWS services on your behalf.

Returns

CreateComputeEnvironmentResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.create_job_queue — Function.

create_job_queue(::AWSConfig, arguments::Dict)
create_job_queue([::AWSConfig]; jobQueueName=, priority=, computeEnvironmentOrder=, <keyword arguments>)

CreateJobQueue Operation

Creates an AWS Batch job queue. When you create a job queue, you associate one or more compute environments to the queue and assign an order of preference for the compute environments.

You also set a priority to the job queue that determines the order in which the AWS Batch scheduler places jobs onto its associated compute environments. For example, if a compute environment is associated with more than one job queue, the job queue with a higher priority is given preference for scheduling jobs to that compute environment.

Arguments

jobQueueName = ::String – Required

The name of the job queue.

state = "ENABLED" or "DISABLED"

The state of the job queue. If the job queue state is ENABLED, it is able to accept jobs.

priority = ::Int – Required

The priority of the job queue. Job queues with a higher priority (or a lower integer value for the priority parameter) are evaluated first when associated with same compute environment. Priority is determined in ascending order, for example, a job queue with a priority value of 1 is given scheduling preference over a job queue with a priority value of 10.

computeEnvironmentOrder = [Dict( ... )...] – Required

The set of compute environments mapped to a job queue and their order relative to each other. The job scheduler uses this parameter to determine which compute environment should execute a given job. Compute environments must be in the VALID state before you can associate them with a job queue. You can associate up to 3 compute environments with a job queue.

 computeEnvironmentOrder = [Dict(
        "order" => <required> ::Int,
        "computeEnvironment" => <required> ::String
    )...]

Returns

CreateJobQueueResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.delete_compute_environment — Function.

delete_compute_environment(::AWSConfig, arguments::Dict)
delete_compute_environment([::AWSConfig]; computeEnvironment=)

DeleteComputeEnvironment Operation

Deletes an AWS Batch compute environment.

Before you can delete a compute environment, you must set its state to DISABLED with the UpdateComputeEnvironment API operation and disassociate it from any job queues with the UpdateJobQueue API operation.

Arguments

computeEnvironment = ::String – Required

The name or Amazon Resource Name (ARN) of the compute environment to delete.

Returns

DeleteComputeEnvironmentResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.delete_job_queue — Function.

delete_job_queue(::AWSConfig, arguments::Dict)
delete_job_queue([::AWSConfig]; jobQueue=)

DeleteJobQueue Operation

Deletes the specified job queue. You must first disable submissions for a queue with the UpdateJobQueue operation and terminate any jobs that have not completed with the TerminateJob.

It is not necessary to disassociate compute environments from a queue before submitting a DeleteJobQueue request.

Arguments

jobQueue = ::String – Required

The short name or full Amazon Resource Name (ARN) of the queue to delete.

Returns

DeleteJobQueueResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.deregister_job_definition — Function.

deregister_job_definition(::AWSConfig, arguments::Dict)
deregister_job_definition([::AWSConfig]; jobDefinition=)

DeregisterJobDefinition Operation

Deregisters an AWS Batch job definition.

Arguments

jobDefinition = ::String – Required

The name and revision (name:revision) or full Amazon Resource Name (ARN) of the job definition to deregister.

Returns

DeregisterJobDefinitionResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.describe_compute_environments — Function.

describe_compute_environments(::AWSConfig, arguments::Dict)
describe_compute_environments([::AWSConfig]; <keyword arguments>)

DescribeComputeEnvironments Operation

Describes one or more of your compute environments.

If you are using an unmanaged compute environment, you can use the DescribeComputeEnvironment operation to determine the ecsClusterArn that you should launch your Amazon ECS container instances into.

Arguments

computeEnvironments = [::String...]

A list of up to 100 compute environment names or full Amazon Resource Name (ARN) entries.

maxResults = ::Int

The maximum number of cluster results returned by DescribeComputeEnvironments in paginated output. When this parameter is used, DescribeComputeEnvironments only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeComputeEnvironments request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeComputeEnvironments returns up to 100 results and a nextToken value if applicable.

nextToken = ::String

The nextToken value returned from a previous paginated DescribeComputeEnvironments request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Note

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

Returns

DescribeComputeEnvironmentsResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.describe_job_definitions — Function.

describe_job_definitions(::AWSConfig, arguments::Dict)
describe_job_definitions([::AWSConfig]; <keyword arguments>)

DescribeJobDefinitions Operation

Describes a list of job definitions. You can specify a status (such as ACTIVE) to only return job definitions that match that status.

Arguments

jobDefinitions = [::String...]

A space-separated list of up to 100 job definition names or full Amazon Resource Name (ARN) entries.

maxResults = ::Int

The maximum number of results returned by DescribeJobDefinitions in paginated output. When this parameter is used, DescribeJobDefinitions only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeJobDefinitions request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeJobDefinitions returns up to 100 results and a nextToken value if applicable.

jobDefinitionName = ::String

The name of the job definition to describe.

status = ::String

The status with which to filter job definitions.

nextToken = ::String

The nextToken value returned from a previous paginated DescribeJobDefinitions request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Note

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

Returns

DescribeJobDefinitionsResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.describe_job_queues — Function.

describe_job_queues(::AWSConfig, arguments::Dict)
describe_job_queues([::AWSConfig]; <keyword arguments>)

DescribeJobQueues Operation

Describes one or more of your job queues.

Arguments

jobQueues = [::String...]

A list of up to 100 queue names or full queue Amazon Resource Name (ARN) entries.

maxResults = ::Int

The maximum number of results returned by DescribeJobQueues in paginated output. When this parameter is used, DescribeJobQueues only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeJobQueues request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then DescribeJobQueues returns up to 100 results and a nextToken value if applicable.

nextToken = ::String

The nextToken value returned from a previous paginated DescribeJobQueues request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Note

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

Returns

DescribeJobQueuesResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.describe_jobs — Function.

describe_jobs(::AWSConfig, arguments::Dict)
describe_jobs([::AWSConfig]; jobs=)

DescribeJobs Operation

Describes a list of AWS Batch jobs.

Arguments

jobs = [::String...] – Required

A space-separated list of up to 100 job IDs.

Returns

DescribeJobsResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.list_jobs — Function.

list_jobs(::AWSConfig, arguments::Dict)
list_jobs([::AWSConfig]; jobQueue=, <keyword arguments>)

ListJobs Operation

Returns a list of task jobs for a specified job queue. You can filter the results by job status with the jobStatus parameter.

Arguments

jobQueue = ::String – Required

The name or full Amazon Resource Name (ARN) of the job queue with which to list jobs.

jobStatus = "SUBMITTED", "PENDING", "RUNNABLE", "STARTING", "RUNNING", "SUCCEEDED" or "FAILED"

The job status with which to filter jobs in the specified queue.

maxResults = ::Int

The maximum number of results returned by ListJobs in paginated output. When this parameter is used, ListJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListJobs request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListJobs returns up to 100 results and a nextToken value if applicable.

nextToken = ::String

The nextToken value returned from a previous paginated ListJobs request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

Note

This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

Returns

ListJobsResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.register_job_definition — Function.

register_job_definition(::AWSConfig, arguments::Dict)
register_job_definition([::AWSConfig]; jobDefinitionName=, type=, <keyword arguments>)

RegisterJobDefinition Operation

Registers an AWS Batch job definition.

Arguments

jobDefinitionName = ::String – Required

The name of the job definition to register.

type = "container" – Required

The type of job definition.

parameters = ::Dict{String,String}

Default parameter substitution placeholders to set in the job definition. Parameters are specified as a key-value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

containerProperties = Dict( ... )

An object with various properties specific for container-based jobs. This parameter is required if the type parameter is container.

 containerProperties = Dict(
        "image" => <required> ::String,
        "vcpus" => <required> ::Int,
        "memory" => <required> ::Int,
        "command" =>  [::String...],
        "jobRoleArn" =>  ::String,
        "volumes" =>  [Dict(
            "host" =>  Dict("sourcePath" =>  ::String),
            "name" =>  ::String
        )...],
        "environment" =>  [Dict(
            "name" =>  ::String,
            "value" =>  ::String
        )...],
        "mountPoints" =>  [Dict(
            "containerPath" =>  ::String,
            "readOnly" =>  ::Bool,
            "sourceVolume" =>  ::String
        )...],
        "readonlyRootFilesystem" =>  ::Bool,
        "privileged" =>  ::Bool,
        "ulimits" =>  [Dict(
            "hardLimit" => <required> ::Int,
            "name" => <required> ::String,
            "softLimit" => <required> ::Int
        )...],
        "user" =>  ::String
    )

retryStrategy = Dict("attempts" => ::Int)

The retry strategy to use for failed jobs that are submitted with this job definition. Any retry strategy that is specified during a SubmitJob operation overrides the retry strategy defined here.

Returns

RegisterJobDefinitionResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.submit_job — Function.

submit_job(::AWSConfig, arguments::Dict)
submit_job([::AWSConfig]; jobName=, jobQueue=, jobDefinition=, <keyword arguments>)

SubmitJob Operation

Submits an AWS Batch job from a job definition. Parameters specified during SubmitJob override parameters defined in the job definition.

Arguments

jobName = ::String – Required

The name of the job. A name must be 1 to 128 characters in length.

Pattern: ^[a-zA-Z0-9_]+$

jobQueue = ::String – Required

The job queue into which the job will be submitted. You can specify either the name or the Amazon Resource Name (ARN) of the queue.

dependsOn = [Dict("jobId" => ::String)...]

A list of job IDs on which this job depends. A job can depend upon a maximum of 100 jobs.

jobDefinition = ::String – Required

The job definition used by this job. This value can be either a name:revision or the Amazon Resource Name (ARN) for the job definition.

parameters = ::Dict{String,String}

Additional parameters passed to the job that replace parameter substitution placeholders that are set in the job definition. Parameters are specified as a key and value pair mapping. Parameters in a SubmitJob request override any corresponding parameter defaults from the job definition.

containerOverrides = Dict( ... )

A list of container overrides in JSON format that specify the name of a container in the specified job definition and the overrides it should receive. You can override the default command for a container (that is specified in the job definition or the Docker image) with a command override. You can also override existing environment variables (that are specified in the job definition or Docker image) on a container or add new environment variables to it with an environment override.

 containerOverrides = Dict(
        "vcpus" =>  ::Int,
        "memory" =>  ::Int,
        "command" =>  [::String...],
        "environment" =>  [Dict(
            "name" =>  ::String,
            "value" =>  ::String
        )...]
    )

retryStrategy = Dict("attempts" => ::Int)

The retry strategy to use for failed jobs from this SubmitJob operation. When a retry strategy is specified here, it overrides the retry strategy defined in the job definition.

Returns

SubmitJobResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.terminate_job — Function.

terminate_job(::AWSConfig, arguments::Dict)
terminate_job([::AWSConfig]; jobId=, reason=)

TerminateJob Operation

Terminates jobs in a job queue. Jobs that are in the STARTING or RUNNING state are terminated, which causes them to transition to FAILED. Jobs that have not progressed to the STARTING state are cancelled.

Arguments

jobId = ::String – Required

Job IDs to be terminated. Up to 100 jobs can be specified.

reason = ::String – Required

A message to attach to the job that explains the reason for cancelling it. This message is returned by future DescribeJobs operations on the job. This message is also recorded in the AWS Batch activity logs.

Returns

TerminateJobResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.update_compute_environment — Function.

update_compute_environment(::AWSConfig, arguments::Dict)
update_compute_environment([::AWSConfig]; computeEnvironment=, <keyword arguments>)

UpdateComputeEnvironment Operation

Updates an AWS Batch compute environment.

Arguments

computeEnvironment = ::String – Required

The name or full Amazon Resource Name (ARN) of the compute environment to update.

state = "ENABLED" or "DISABLED"

The state of the compute environment. Compute environments in the ENABLED state can accept jobs from a queue and scale in or out automatically based on the workload demand of its associated queues.

computeResources = Dict( ... )

Details of the compute resources managed by the compute environment. Required for a managed compute environment.

 computeResources = Dict(
        "minvCpus" =>  ::Int,
        "maxvCpus" =>  ::Int,
        "desiredvCpus" =>  ::Int
    )

serviceRole = ::String

The name or full Amazon Resource Name (ARN) of the IAM role that allows AWS Batch to make calls to ECS, Auto Scaling, and EC2 on your behalf.

Returns

UpdateComputeEnvironmentResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation

AmazonBatch.update_job_queue — Function.

update_job_queue(::AWSConfig, arguments::Dict)
update_job_queue([::AWSConfig]; jobQueue=, <keyword arguments>)

UpdateJobQueue Operation

Updates a job queue.

Arguments

jobQueue = ::String – Required

The name or the Amazon Resource Name (ARN) of the job queue.

state = "ENABLED" or "DISABLED"

Describes the queue's ability to accept new jobs.

priority = ::Int

The priority of the job queue. Job queues with a higher priority (or a lower integer value for the priority parameter) are evaluated first when associated with same compute environment. Priority is determined in ascending order, for example, a job queue with a priority value of 1 is given scheduling preference over a job queue with a priority value of 10.

computeEnvironmentOrder = [Dict( ... )...]

Details the set of compute environments mapped to a job queue and their order relative to each other. This is one of the parameters used by the job scheduler to determine which compute environment should execute a given job.

 computeEnvironmentOrder = [Dict(
        "order" => <required> ::Int,
        "computeEnvironment" => <required> ::String
    )...]

Returns

UpdateJobQueueResponse

Exceptions

ClientException or ServerException.

See also: AWS API Documentation