AmazonCloudHSM.jl

AmazonCloudHSM

AWS CloudHSM Service

This document is generated from apis/cloudhsm-2014-05-30.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCloudHSM.add_tags_to_resource

AmazonCloudHSM.create_hapg

AmazonCloudHSM.create_hsm

AmazonCloudHSM.create_luna_client

AmazonCloudHSM.delete_hapg

AmazonCloudHSM.delete_hsm

AmazonCloudHSM.delete_luna_client

AmazonCloudHSM.describe_hapg

AmazonCloudHSM.describe_hsm

AmazonCloudHSM.describe_luna_client

AmazonCloudHSM.get_config

AmazonCloudHSM.list_available_zones

AmazonCloudHSM.list_hapgs

AmazonCloudHSM.list_hsms

AmazonCloudHSM.list_luna_clients

AmazonCloudHSM.list_tags_for_resource

AmazonCloudHSM.modify_hapg

AmazonCloudHSM.modify_hsm

AmazonCloudHSM.modify_luna_client

AmazonCloudHSM.remove_tags_from_resource

AmazonCloudHSM.add_tags_to_resource — Function.

add_tags_to_resource(::AWSConfig, arguments::Dict)
add_tags_to_resource([::AWSConfig]; ResourceArn=, TagList=)

AddTagsToResource Operation

Adds or overwrites one or more tags for the specified AWS CloudHSM resource.

Each tag consists of a key and a value. Tag keys must be unique to each resource.

Arguments

ResourceArn = ::String – Required

The Amazon Resource Name (ARN) of the AWS CloudHSM resource to tag.

TagList = [Dict( ... )...] – Required

One or more tags.

 TagList = [Dict(
        "Key" => <required> ::String,
        "Value" => <required> ::String
    )...]

Returns

AddTagsToResourceResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.create_hapg — Function.

create_hapg(::AWSConfig, arguments::Dict)
create_hapg([::AWSConfig]; Label=)

CreateHapg Operation

Creates a high-availability partition group. A high-availability partition group is a group of partitions that spans multiple physical HSMs.

Arguments

Label = ::String – Required

The label of the new high-availability partition group.

Returns

CreateHapgResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.create_hsm — Function.

create_hsm(::AWSConfig, arguments::Dict)
create_hsm([::AWSConfig]; SubnetId=, SshKey=, IamRoleArn=, SubscriptionType=, <keyword arguments>)

CreateHsm Operation

Creates an uninitialized HSM instance.

There is an upfront fee charged for each HSM instance that you create with the CreateHsm operation. If you accidentally provision an HSM and want to request a refund, delete the instance using the DeleteHsm operation, go to the AWS Support Center, create a new case, and select Account and Billing Support.

Important

It can take up to 20 minutes to create and provision an HSM. You can monitor the status of the HSM with the DescribeHsm operation. The HSM is ready to be initialized when the status changes to RUNNING.

Arguments

SubnetId = ::String – Required

The identifier of the subnet in your VPC in which to place the HSM.

SshKey = ::String – Required

The SSH public key to install on the HSM.

EniIp = ::String

The IP address to assign to the HSM's ENI.

If an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the subnet.

IamRoleArn = ::String – Required

The ARN of an IAM role to enable the AWS CloudHSM service to allocate an ENI on your behalf.

ExternalId = ::String

The external ID from IamRoleArn, if present.

SubscriptionType = "PRODUCTION" – Required

ClientToken = ::String

A user-defined token to ensure idempotence. Subsequent calls to this operation with the same token will be ignored.

SyslogIp = ::String

The IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

Returns

CreateHsmResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.create_luna_client — Function.

create_luna_client(::AWSConfig, arguments::Dict)
create_luna_client([::AWSConfig]; Certificate=, <keyword arguments>)

CreateLunaClient Operation

Creates an HSM client.

Arguments

Label = ::String

The label for the client.

Certificate = ::String – Required

The contents of a Base64-Encoded X.509 v3 certificate to be installed on the HSMs used by this client.

Returns

CreateLunaClientResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.delete_hapg — Function.

delete_hapg(::AWSConfig, arguments::Dict)
delete_hapg([::AWSConfig]; HapgArn=)

DeleteHapg Operation

Deletes a high-availability partition group.

Arguments

HapgArn = ::String – Required

The ARN of the high-availability partition group to delete.

Returns

DeleteHapgResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.delete_hsm — Function.

delete_hsm(::AWSConfig, arguments::Dict)
delete_hsm([::AWSConfig]; HsmArn=)

DeleteHsm Operation

Deletes an HSM. After completion, this operation cannot be undone and your key material cannot be recovered.

Arguments

HsmArn = ::String – Required

The ARN of the HSM to delete.

Returns

DeleteHsmResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.delete_luna_client — Function.

delete_luna_client(::AWSConfig, arguments::Dict)
delete_luna_client([::AWSConfig]; ClientArn=)

DeleteLunaClient Operation

Deletes a client.

Arguments

ClientArn = ::String – Required

The ARN of the client to delete.

Returns

DeleteLunaClientResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.describe_hapg — Function.

describe_hapg(::AWSConfig, arguments::Dict)
describe_hapg([::AWSConfig]; HapgArn=)

DescribeHapg Operation

Retrieves information about a high-availability partition group.

Arguments

HapgArn = ::String – Required

The ARN of the high-availability partition group to describe.

Returns

DescribeHapgResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.describe_hsm — Function.

describe_hsm(::AWSConfig, arguments::Dict)
describe_hsm([::AWSConfig]; <keyword arguments>)

DescribeHsm Operation

Retrieves information about an HSM. You can identify the HSM by its ARN or its serial number.

Arguments

HsmArn = ::String

The ARN of the HSM. Either the HsmArn or the SerialNumber parameter must be specified.

HsmSerialNumber = ::String

The serial number of the HSM. Either the HsmArn or the HsmSerialNumber parameter must be specified.

Returns

DescribeHsmResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.describe_luna_client — Function.

describe_luna_client(::AWSConfig, arguments::Dict)
describe_luna_client([::AWSConfig]; <keyword arguments>)

DescribeLunaClient Operation

Retrieves information about an HSM client.

Arguments

ClientArn = ::String

The ARN of the client.

CertificateFingerprint = ::String

The certificate fingerprint.

Returns

DescribeLunaClientResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.get_config — Function.

get_config(::AWSConfig, arguments::Dict)
get_config([::AWSConfig]; ClientArn=, ClientVersion=, HapgList=)

GetConfig Operation

Gets the configuration files necessary to connect to all high availability partition groups the client is associated with.

Arguments

ClientArn = ::String – Required

The ARN of the client.

ClientVersion = "5.1" or "5.3" – Required

The client version.

HapgList = [::String...] – Required

A list of ARNs that identify the high-availability partition groups that are associated with the client.

Returns

GetConfigResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.list_available_zones — Function.

list_available_zones(::AWSConfig, arguments::Dict)
list_available_zones([::AWSConfig]; )

ListAvailableZones Operation

Lists the Availability Zones that have available AWS CloudHSM capacity.

Arguments

Returns

ListAvailableZonesResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.list_hapgs — Function.

list_hapgs(::AWSConfig, arguments::Dict)
list_hapgs([::AWSConfig]; <keyword arguments>)

ListHapgs Operation

Lists the high-availability partition groups for the account.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListHapgs to retrieve the next set of items.

Arguments

NextToken = ::String

The NextToken value from a previous call to ListHapgs. Pass null if this is the first call.

Returns

ListHapgsResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.list_hsms — Function.

list_hsms(::AWSConfig, arguments::Dict)
list_hsms([::AWSConfig]; <keyword arguments>)

ListHsms Operation

Retrieves the identifiers of all of the HSMs provisioned for the current customer.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListHsms to retrieve the next set of items.

Arguments

NextToken = ::String

The NextToken value from a previous call to ListHsms. Pass null if this is the first call.

Returns

ListHsmsResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.list_luna_clients — Function.

list_luna_clients(::AWSConfig, arguments::Dict)
list_luna_clients([::AWSConfig]; <keyword arguments>)

ListLunaClients Operation

Lists all of the clients.

This operation supports pagination with the use of the NextToken member. If more results are available, the NextToken member of the response contains a token that you pass in the next call to ListLunaClients to retrieve the next set of items.

Arguments

NextToken = ::String

The NextToken value from a previous call to ListLunaClients. Pass null if this is the first call.

Returns

ListLunaClientsResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.list_tags_for_resource — Function.

list_tags_for_resource(::AWSConfig, arguments::Dict)
list_tags_for_resource([::AWSConfig]; ResourceArn=)

ListTagsForResource Operation

Returns a list of all tags for the specified AWS CloudHSM resource.

Arguments

ResourceArn = ::String – Required

The Amazon Resource Name (ARN) of the AWS CloudHSM resource.

Returns

ListTagsForResourceResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.modify_hapg — Function.

modify_hapg(::AWSConfig, arguments::Dict)
modify_hapg([::AWSConfig]; HapgArn=, <keyword arguments>)

ModifyHapg Operation

Modifies an existing high-availability partition group.

Arguments

HapgArn = ::String – Required

The ARN of the high-availability partition group to modify.

Label = ::String

The new label for the high-availability partition group.

PartitionSerialList = [::String...]

The list of partition serial numbers to make members of the high-availability partition group.

Returns

ModifyHapgResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.modify_hsm — Function.

modify_hsm(::AWSConfig, arguments::Dict)
modify_hsm([::AWSConfig]; HsmArn=, <keyword arguments>)

ModifyHsm Operation

Modifies an HSM.

Important

This operation can result in the HSM being offline for up to 15 minutes while the AWS CloudHSM service is reconfigured. If you are modifying a production HSM, you should ensure that your AWS CloudHSM service is configured for high availability, and consider executing this operation during a maintenance window.

Arguments

HsmArn = ::String – Required

The ARN of the HSM to modify.

SubnetId = ::String

The new identifier of the subnet that the HSM is in. The new subnet must be in the same Availability Zone as the current subnet.

EniIp = ::String

The new IP address for the elastic network interface (ENI) attached to the HSM.

If the HSM is moved to a different subnet, and an IP address is not specified, an IP address will be randomly chosen from the CIDR range of the new subnet.

IamRoleArn = ::String

The new IAM role ARN.

ExternalId = ::String

The new external ID.

SyslogIp = ::String

The new IP address for the syslog monitoring server. The AWS CloudHSM service only supports one syslog monitoring server.

Returns

ModifyHsmResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation

AmazonCloudHSM.modify_luna_client — Function.

modify_luna_client(::AWSConfig, arguments::Dict)
modify_luna_client([::AWSConfig]; ClientArn=, Certificate=)

ModifyLunaClient Operation

Modifies the certificate used by the client.

This action can potentially start a workflow to install the new certificate on the client's HSMs.

Arguments

ClientArn = ::String – Required

The ARN of the client.

Certificate = ::String – Required

The new certificate for the client.

Returns

ModifyLunaClientResponse

Exceptions

CloudHsmServiceException.

See also: AWS API Documentation

AmazonCloudHSM.remove_tags_from_resource — Function.

remove_tags_from_resource(::AWSConfig, arguments::Dict)
remove_tags_from_resource([::AWSConfig]; ResourceArn=, TagKeyList=)

RemoveTagsFromResource Operation

Removes one or more tags from the specified AWS CloudHSM resource.

To remove a tag, specify only the tag key to remove (not the value). To overwrite the value for an existing tag, use AddTagsToResource.

Arguments

ResourceArn = ::String – Required

The Amazon Resource Name (ARN) of the AWS CloudHSM resource.

TagKeyList = [::String...] – Required

The tag key or keys to remove.

Specify only the tag key to remove (not the value). To overwrite the value for an existing tag, use AddTagsToResource.

Returns

RemoveTagsFromResourceResponse

Exceptions

CloudHsmServiceException, CloudHsmInternalException or InvalidRequestException.

See also: AWS API Documentation