AmazonCloudWatch.jl

AmazonCloudWatch

Amazon CloudWatch monitors your Amazon Web Services (AWS) resources and the applications you run on AWS in real time. You can use CloudWatch to collect and track metrics, which are the variables you want to measure for your resources and applications.

CloudWatch alarms send notifications or automatically change the resources you are monitoring based on rules that you define. For example, you can monitor the CPU usage and disk reads and writes of your Amazon EC2 instances. Then, use this data to determine whether you should launch additional instances to handle increased load. You can also use this data to stop under-used instances to save money.

In addition to monitoring the built-in metrics that come with AWS, you can monitor your own custom metrics. With CloudWatch, you gain system-wide visibility into resource utilization, application performance, and operational health.

This document is generated from apis/monitoring-2010-08-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonCloudWatch.delete_alarms

AmazonCloudWatch.delete_dashboards

AmazonCloudWatch.describe_alarm_history

AmazonCloudWatch.describe_alarms

AmazonCloudWatch.describe_alarms_for_metric

AmazonCloudWatch.disable_alarm_actions

AmazonCloudWatch.enable_alarm_actions

AmazonCloudWatch.get_dashboard

AmazonCloudWatch.get_metric_statistics

AmazonCloudWatch.list_dashboards

AmazonCloudWatch.list_metrics

AmazonCloudWatch.put_dashboard

AmazonCloudWatch.put_metric_alarm

AmazonCloudWatch.put_metric_data

AmazonCloudWatch.set_alarm_state

AmazonCloudWatch.delete_alarms — Function.

delete_alarms(::AWSConfig, arguments::Dict)
delete_alarms([::AWSConfig]; AlarmNames=)

DeleteAlarms Operation

Deletes the specified alarms. In the event of an error, no alarms are deleted.

Arguments

AlarmNames = [::String...] – Required

The alarms to be deleted.

Exceptions

ResourceNotFound.

See also: AWS API Documentation

AmazonCloudWatch.delete_dashboards — Function.

delete_dashboards(::AWSConfig, arguments::Dict)
delete_dashboards([::AWSConfig]; <keyword arguments>)

DeleteDashboards Operation

Deletes all dashboards that you specify. You may specify up to 100 dashboards to delete. If there is an error during this call, no dashboards are deleted.

Arguments

DashboardNames = [::String...]

The dashboards to be deleted.

Returns

DeleteDashboardsOutput

Exceptions

InvalidParameterValueException, DashboardNotFoundError or InternalServiceFault.

See also: AWS API Documentation

AmazonCloudWatch.describe_alarm_history — Function.

describe_alarm_history(::AWSConfig, arguments::Dict)
describe_alarm_history([::AWSConfig]; <keyword arguments>)

DescribeAlarmHistory Operation

Retrieves the history for the specified alarm. You can filter the results by date range or item type. If an alarm name is not specified, the histories for all alarms are returned.

CloudWatch retains the history of an alarm even if you delete the alarm.

Arguments

AlarmName = ::String

The name of the alarm.

HistoryItemType = "ConfigurationUpdate", "StateUpdate" or "Action"

The type of alarm histories to retrieve.

StartDate = timestamp

The starting date to retrieve alarm history.

EndDate = timestamp

The ending date to retrieve alarm history.

MaxRecords = ::Int

The maximum number of alarm history records to retrieve.

NextToken = ::String

The token returned by a previous call to indicate that there is more data available.

Returns

DescribeAlarmHistoryOutput

Exceptions

InvalidNextToken.

See also: AWS API Documentation

AmazonCloudWatch.describe_alarms — Function.

describe_alarms(::AWSConfig, arguments::Dict)
describe_alarms([::AWSConfig]; <keyword arguments>)

DescribeAlarms Operation

Retrieves the specified alarms. If no alarms are specified, all alarms are returned. Alarms can be retrieved by using only a prefix for the alarm name, the alarm state, or a prefix for any action.

Arguments

AlarmNames = [::String...]

The names of the alarms.

AlarmNamePrefix = ::String

The alarm name prefix. If this parameter is specified, you cannot specify AlarmNames.

StateValue = "OK", "ALARM" or "INSUFFICIENT_DATA"

The state value to be used in matching alarms.

ActionPrefix = ::String

The action name prefix.

MaxRecords = ::Int

The maximum number of alarm descriptions to retrieve.

NextToken = ::String

The token returned by a previous call to indicate that there is more data available.

Returns

DescribeAlarmsOutput

Exceptions

InvalidNextToken.

See also: AWS API Documentation

AmazonCloudWatch.describe_alarms_for_metric — Function.

describe_alarms_for_metric(::AWSConfig, arguments::Dict)
describe_alarms_for_metric([::AWSConfig]; MetricName=, Namespace=, <keyword arguments>)

DescribeAlarmsForMetric Operation

Retrieves the alarms for the specified metric. To filter the results, specify a statistic, period, or unit.

Arguments

MetricName = ::String – Required

The name of the metric.

Namespace = ::String – Required

The namespace of the metric.

Statistic = "SampleCount", "Average", "Sum", "Minimum" or "Maximum"

The statistic for the metric, other than percentiles. For percentile statistics, use ExtendedStatistics.

ExtendedStatistic = ::String

The percentile statistic for the metric. Specify a value between p0.0 and p100.

Dimensions = [Dict( ... )...]

The dimensions associated with the metric. If the metric has any associated dimensions, you must specify them in order for the call to succeed.

 Dimensions = [Dict(
        "Name" => <required> ::String,
        "Value" => <required> ::String
    )...]

Period = ::Int

The period, in seconds, over which the statistic is applied.

Unit = "Seconds", "Microseconds", "Milliseconds", "Bytes", "Kilobytes", "Megabytes", "Gigabytes", "Terabytes", "Bits", "Kilobits", "Megabits", "Gigabits", "Terabits", "Percent", "Count", "Bytes/Second", "Kilobytes/Second", "Megabytes/Second", "Gigabytes/Second", "Terabytes/Second", "Bits/Second", "Kilobits/Second", "Megabits/Second", "Gigabits/Second", "Terabits/Second", "Count/Second" or "None"

The unit for the metric.

Returns

DescribeAlarmsForMetricOutput

See also: AWS API Documentation

AmazonCloudWatch.disable_alarm_actions — Function.

disable_alarm_actions(::AWSConfig, arguments::Dict)
disable_alarm_actions([::AWSConfig]; AlarmNames=)

DisableAlarmActions Operation

Disables the actions for the specified alarms. When an alarm's actions are disabled, the alarm actions do not execute when the alarm state changes.

Arguments

AlarmNames = [::String...] – Required

The names of the alarms.

See also: AWS API Documentation

AmazonCloudWatch.enable_alarm_actions — Function.

enable_alarm_actions(::AWSConfig, arguments::Dict)
enable_alarm_actions([::AWSConfig]; AlarmNames=)

EnableAlarmActions Operation

Enables the actions for the specified alarms.

Arguments

AlarmNames = [::String...] – Required

The names of the alarms.

See also: AWS API Documentation

AmazonCloudWatch.get_dashboard — Function.

get_dashboard(::AWSConfig, arguments::Dict)
get_dashboard([::AWSConfig]; <keyword arguments>)

GetDashboard Operation

Displays the details of the dashboard that you specify.

To copy an existing dashboard, use GetDashboard, and then use the data returned within DashboardBody as the template for the new dashboard when you call PutDashboard to create the copy.

Arguments

DashboardName = ::String

The name of the dashboard to be described.

Returns

GetDashboardOutput

Exceptions

InvalidParameterValueException, DashboardNotFoundError or InternalServiceFault.

See also: AWS API Documentation

AmazonCloudWatch.get_metric_statistics — Function.

get_metric_statistics(::AWSConfig, arguments::Dict)
get_metric_statistics([::AWSConfig]; Namespace=, MetricName=, StartTime=, EndTime=, Period=, <keyword arguments>)

GetMetricStatistics Operation

Gets statistics for the specified metric.

The maximum number of data points returned from a single call is 1,440. If you request more than 1,440 data points, CloudWatch returns an error. To reduce the number of data points, you can narrow the specified time range and make multiple requests across adjacent time ranges, or you can increase the specified period. Data points are not returned in chronological order.

CloudWatch aggregates data points based on the length of the period that you specify. For example, if you request statistics with a one-hour period, CloudWatch aggregates all data points with time stamps that fall within each one-hour period. Therefore, the number of values aggregated by CloudWatch is larger than the number of data points returned.

CloudWatch needs raw data points to calculate percentile statistics. If you publish data using a statistic set instead, you can only retrieve percentile statistics for this data if one of the following conditions is true:

The SampleCount value of the statistic set is 1.

The Min and the Max values of the statistic set are equal.

Amazon CloudWatch retains metric data as follows:

Data points with a period of less than 60 seconds are available for 3 hours. These data points are high-resolution metrics and are available only for custom metrics that have been defined with a StorageResolution of 1.

Data points with a period of 60 seconds (1-minute) are available for 15 days.

Data points with a period of 300 seconds (5-minute) are available for 63 days.

Data points with a period of 3600 seconds (1 hour) are available for 455 days (15 months).

Data points that are initially published with a shorter period are aggregated together for long-term storage. For example, if you collect data using a period of 1 minute, the data remains available for 15 days with 1-minute resolution. After 15 days, this data is still available, but is aggregated and retrievable only with a resolution of 5 minutes. After 63 days, the data is further aggregated and is available with a resolution of 1 hour.

CloudWatch started retaining 5-minute and 1-hour metric data as of July 9, 2016.

For information about metrics and dimensions supported by AWS services, see the Amazon CloudWatch Metrics and Dimensions Reference in the Amazon CloudWatch User Guide.

Arguments

Namespace = ::String – Required

The namespace of the metric, with or without spaces.

MetricName = ::String – Required

The name of the metric, with or without spaces.

Dimensions = [Dict( ... )...]

The dimensions. If the metric contains multiple dimensions, you must include a value for each dimension. CloudWatch treats each unique combination of dimensions as a separate metric. If a specific combination of dimensions was not published, you can't retrieve statistics for it. You must specify the same dimensions that were used when the metrics were created. For an example, see Dimension Combinations in the Amazon CloudWatch User Guide. For more information about specifying dimensions, see Publishing Metrics in the Amazon CloudWatch User Guide.

 Dimensions = [Dict(
        "Name" => <required> ::String,
        "Value" => <required> ::String
    )...]

StartTime = timestamp – Required

The time stamp that determines the first data point to return. Start times are evaluated relative to the time that CloudWatch receives the request.

The value specified is inclusive; results include data points with the specified time stamp. The time stamp must be in ISO 8601 UTC format (for example, 2016-10-03T23:00:00Z).

CloudWatch rounds the specified time stamp as follows:

Start time less than 15 days ago - Round down to the nearest whole minute. For example, 12:32:34 is rounded down to 12:32:00.

Start time between 15 and 63 days ago - Round down to the nearest 5-minute clock interval. For example, 12:32:34 is rounded down to 12:30:00.

Start time greater than 63 days ago - Round down to the nearest 1-hour clock interval. For example, 12:32:34 is rounded down to 12:00:00.

If you set Period to 5, 10, or 30, the start time of your request is rounded down to the nearest time that corresponds to even 5-, 10-, or 30-second divisions of a minute. For example, if you make a query at (HH:mm:ss) 01:05:23 for the previous 10-second period, the start time of your request is rounded down and you receive data from 01:05:10 to 01:05:20. If you make a query at 15:07:17 for the previous 5 minutes of data, using a period of 5 seconds, you receive data timestamped between 15:02:15 and 15:07:15.

EndTime = timestamp – Required

The time stamp that determines the last data point to return.

The value specified is exclusive; results include data points up to the specified time stamp. The time stamp must be in ISO 8601 UTC format (for example, 2016-10-10T23:00:00Z).

Period = ::Int – Required

The granularity, in seconds, of the returned data points. For metrics with regular resolution, a period can be as short as one minute (60 seconds) and must be a multiple of 60. For high-resolution metrics that are collected at intervals of less than one minute, the period can be 1, 5, 10, 30, 60, or any multiple of 60. High-resolution metrics are those metrics stored by a PutMetricData call that includes a StorageResolution of 1 second.

If the StartTime parameter specifies a time stamp that is greater than 3 hours ago, you must specify the period as follows or no data points in that time range is returned:

Start time between 3 hours and 15 days ago - Use a multiple of 60 seconds (1 minute).

Start time between 15 and 63 days ago - Use a multiple of 300 seconds (5 minutes).

Start time greater than 63 days ago - Use a multiple of 3600 seconds (1 hour).

Statistics = ["SampleCount", "Average", "Sum", "Minimum" or "Maximum"...]

The metric statistics, other than percentile. For percentile statistics, use ExtendedStatistics. When calling GetMetricStatistics, you must specify either Statistics or ExtendedStatistics, but not both.

ExtendedStatistics = [::String...]

The percentile statistics. Specify values between p0.0 and p100. When calling GetMetricStatistics, you must specify either Statistics or ExtendedStatistics, but not both.

Unit = "Seconds", "Microseconds", "Milliseconds", "Bytes", "Kilobytes", "Megabytes", "Gigabytes", "Terabytes", "Bits", "Kilobits", "Megabits", "Gigabits", "Terabits", "Percent", "Count", "Bytes/Second", "Kilobytes/Second", "Megabytes/Second", "Gigabytes/Second", "Terabytes/Second", "Bits/Second", "Kilobits/Second", "Megabits/Second", "Gigabits/Second", "Terabits/Second", "Count/Second" or "None"

The unit for a given metric. Metrics may be reported in multiple units. Not supplying a unit results in all units being returned. If the metric only ever reports one unit, specifying a unit has no effect.

Returns

GetMetricStatisticsOutput

Exceptions

InvalidParameterValueException, MissingRequiredParameterException, InvalidParameterCombinationException or InternalServiceFault.

See also: AWS API Documentation

AmazonCloudWatch.list_dashboards — Function.

list_dashboards(::AWSConfig, arguments::Dict)
list_dashboards([::AWSConfig]; <keyword arguments>)

ListDashboards Operation

Returns a list of the dashboards for your account. If you include DashboardNamePrefix, only those dashboards with names starting with the prefix are listed. Otherwise, all dashboards in your account are listed.

Arguments

DashboardNamePrefix = ::String

If you specify this parameter, only the dashboards with names starting with the specified string are listed. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, ".", "-", and "_".

NextToken = ::String

The token returned by a previous call to indicate that there is more data available.

Returns

ListDashboardsOutput

Exceptions

InvalidParameterValueException or InternalServiceFault.

See also: AWS API Documentation

AmazonCloudWatch.list_metrics — Function.

list_metrics(::AWSConfig, arguments::Dict)
list_metrics([::AWSConfig]; <keyword arguments>)

ListMetrics Operation

List the specified metrics. You can use the returned metrics with GetMetricStatistics to obtain statistical data.

Up to 500 results are returned for any one call. To retrieve additional results, use the returned token with subsequent calls.

After you create a metric, allow up to fifteen minutes before the metric appears. Statistics about the metric, however, are available sooner using GetMetricStatistics.

Arguments

Namespace = ::String

The namespace to filter against.

MetricName = ::String

The name of the metric to filter against.

Dimensions = [Dict( ... )...]

The dimensions to filter against.

 Dimensions = [Dict(
        "Name" => <required> ::String,
        "Value" =>  ::String
    )...]

NextToken = ::String

The token returned by a previous call to indicate that there is more data available.

Returns

ListMetricsOutput

Exceptions

InternalServiceFault or InvalidParameterValueException.

See also: AWS API Documentation

AmazonCloudWatch.put_dashboard — Function.

put_dashboard(::AWSConfig, arguments::Dict)
put_dashboard([::AWSConfig]; <keyword arguments>)

PutDashboard Operation

Creates a dashboard if it does not already exist, or updates an existing dashboard. If you update a dashboard, the entire contents are replaced with what you specify here.

You can have up to 500 dashboards per account. All dashboards in your account are global, not region-specific.

A simple way to create a dashboard using PutDashboard is to copy an existing dashboard. To copy an existing dashboard using the console, you can load the dashboard and then use the View/edit source command in the Actions menu to display the JSON block for that dashboard. Another way to copy a dashboard is to use GetDashboard, and then use the data returned within DashboardBody as the template for the new dashboard when you call PutDashboard.

When you create a dashboard with PutDashboard, a good practice is to add a text widget at the top of the dashboard with a message that the dashboard was created by script and should not be changed in the console. This message could also point console users to the location of the DashboardBody script or the CloudFormation template used to create the dashboard.

Arguments

DashboardName = ::String

The name of the dashboard. If a dashboard with this name already exists, this call modifies that dashboard, replacing its current contents. Otherwise, a new dashboard is created. The maximum length is 255, and valid characters are A-Z, a-z, 0-9, "-", and "_".

DashboardBody = ::String

The detailed information about the dashboard in JSON format, including the widgets to include and their location on the dashboard.

For more information about the syntax, see CloudWatch-Dashboard-Body-Structure.

Returns

PutDashboardOutput

Exceptions

DashboardInvalidInputError or InternalServiceFault.

See also: AWS API Documentation

AmazonCloudWatch.put_metric_alarm — Function.

put_metric_alarm(::AWSConfig, arguments::Dict)
put_metric_alarm([::AWSConfig]; AlarmName=, MetricName=, Namespace=, Period=, EvaluationPeriods=, Threshold=, ComparisonOperator=, <keyword arguments>)

PutMetricAlarm Operation

Creates or updates an alarm and associates it with the specified metric. Optionally, this operation can associate one or more Amazon SNS resources with the alarm.

When this operation creates an alarm, the alarm state is immediately set to INSUFFICIENT_DATA. The alarm is evaluated and its state is set appropriately. Any actions associated with the state are then executed.

When you update an existing alarm, its state is left unchanged, but the update completely overwrites the previous configuration of the alarm.

If you are an IAM user, you must have Amazon EC2 permissions for some operations:

ec2:DescribeInstanceStatus and ec2:DescribeInstances for all alarms on EC2 instance status metrics

ec2:StopInstances for alarms with stop actions

ec2:TerminateInstances for alarms with terminate actions

ec2:DescribeInstanceRecoveryAttribute and ec2:RecoverInstances for alarms with recover actions

If you have read/write permissions for Amazon CloudWatch but not for Amazon EC2, you can still create an alarm, but the stop or terminate actions are not performed. However, if you are later granted the required permissions, the alarm actions that you created earlier are performed.

If you are using an IAM role (for example, an EC2 instance profile), you cannot stop or terminate the instance using alarm actions. However, you can still see the alarm state and perform any other actions such as Amazon SNS notifications or Auto Scaling policies.

If you are using temporary security credentials granted using AWS STS, you cannot stop or terminate an EC2 instance using alarm actions.

You must create at least one stop, terminate, or reboot alarm using either the Amazon EC2 or CloudWatch consoles to create the EC2ActionsAccess IAM role. After this IAM role is created, you can create stop, terminate, or reboot alarms using a command-line interface or API.

Arguments

AlarmName = ::String – Required

The name for the alarm. This name must be unique within the AWS account.

AlarmDescription = ::String

The description for the alarm.

ActionsEnabled = ::Bool

Indicates whether actions should be executed during any changes to the alarm state.

OKActions = [::String...]

The actions to execute when this alarm transitions to an OK state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

AlarmActions = [::String...]

The actions to execute when this alarm transitions to the ALARM state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

InsufficientDataActions = [::String...]

The actions to execute when this alarm transitions to the INSUFFICIENT_DATA state from any other state. Each action is specified as an Amazon Resource Name (ARN).

Valid Values: arn:aws:automate:region:ec2:stop | arn:aws:automate:region:ec2:terminate | arn:aws:automate:region:ec2:recover

Valid Values (for use with IAM roles): arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Stop/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Terminate/1.0 | arn:aws:swf:us-east-1:{customer-account}:action/actions/AWS_EC2.InstanceId.Reboot/1.0

MetricName = ::String – Required

The name for the metric associated with the alarm.

Namespace = ::String – Required

The namespace for the metric associated with the alarm.

Statistic = "SampleCount", "Average", "Sum", "Minimum" or "Maximum"

The statistic for the metric associated with the alarm, other than percentile. For percentile statistics, use ExtendedStatistic.

ExtendedStatistic = ::String

The percentile statistic for the metric associated with the alarm. Specify a value between p0.0 and p100.

Dimensions = [Dict( ... )...]

The dimensions for the metric associated with the alarm.

 Dimensions = [Dict(
        "Name" => <required> ::String,
        "Value" => <required> ::String
    )...]

Period = ::Int – Required

The period, in seconds, over which the specified statistic is applied. Valid values are 10, 30, and any multiple of 60.

Be sure to specify 10 or 30 only for metrics that are stored by a PutMetricData call with a StorageResolution of 1. If you specify a Period of 10 or 30 for a metric that does not have sub-minute resolution, the alarm still attempts to gather data at the period rate that you specify. In this case, it does not receive data for the attempts that do not correspond to a one-minute data resolution, and the alarm may often lapse into INSUFFICENT_DATA status. Specifying 10 or 30 also sets this alarm as a high-resolution alarm, which has a higher charge than other alarms. For more information about pricing, see Amazon CloudWatch Pricing.

An alarm's total current evaluation period can be no longer than one day, so Period multiplied by EvaluationPeriods cannot be more than 86,400 seconds.

Unit = "Seconds", "Microseconds", "Milliseconds", "Bytes", "Kilobytes", "Megabytes", "Gigabytes", "Terabytes", "Bits", "Kilobits", "Megabits", "Gigabits", "Terabits", "Percent", "Count", "Bytes/Second", "Kilobytes/Second", "Megabytes/Second", "Gigabytes/Second", "Terabytes/Second", "Bits/Second", "Kilobits/Second", "Megabits/Second", "Gigabits/Second", "Terabits/Second", "Count/Second" or "None"

The unit of measure for the statistic. For example, the units for the Amazon EC2 NetworkIn metric are Bytes because NetworkIn tracks the number of bytes that an instance receives on all network interfaces. You can also specify a unit when you create a custom metric. Units help provide conceptual meaning to your data. Metric data points that specify a unit of measure, such as Percent, are aggregated separately.

If you specify a unit, you must use a unit that is appropriate for the metric. Otherwise, the CloudWatch alarm can get stuck in the INSUFFICIENT DATA state.

EvaluationPeriods = ::Int – Required

The number of periods over which data is compared to the specified threshold. An alarm's total current evaluation period can be no longer than one day, so this number multiplied by Period cannot be more than 86,400 seconds.

Threshold = double – Required

The value against which the specified statistic is compared.

ComparisonOperator = "GreaterThanOrEqualToThreshold", "GreaterThanThreshold", "LessThanThreshold" or "LessThanOrEqualToThreshold" – Required

The arithmetic operation to use when comparing the specified statistic and threshold. The specified statistic value is used as the first operand.

TreatMissingData = ::String

Sets how this alarm is to handle missing data points. If TreatMissingData is omitted, the default behavior of missing is used. For more information, see Configuring How CloudWatch Alarms Treats Missing Data.

Valid Values: breaching | notBreaching | ignore | missing

EvaluateLowSampleCountPercentile = ::String

Used only for alarms based on percentiles. If you specify ignore, the alarm state does not change during periods with too few data points to be statistically significant. If you specify evaluate or omit this parameter, the alarm is always evaluated and possibly changes state no matter how many data points are available. For more information, see Percentile-Based CloudWatch Alarms and Low Data Samples.

Valid Values: evaluate | ignore

Exceptions

LimitExceededFault.

See also: AWS API Documentation

AmazonCloudWatch.put_metric_data — Function.

put_metric_data(::AWSConfig, arguments::Dict)
put_metric_data([::AWSConfig]; Namespace=, MetricData=)

PutMetricData Operation

Publishes metric data points to Amazon CloudWatch. CloudWatch associates the data points with the specified metric. If the specified metric does not exist, CloudWatch creates the metric. When CloudWatch creates a metric, it can take up to fifteen minutes for the metric to appear in calls to ListMetrics.

Each PutMetricData request is limited to 40 KB in size for HTTP POST requests.

Although the Value parameter accepts numbers of type Double, CloudWatch rejects values that are either too small or too large. Values must be in the range of 8.515920e-109 to 1.174271e+108 (Base 10) or 2e-360 to 2e360 (Base 2). In addition, special values (for example, NaN, +Infinity, -Infinity) are not supported.

You can use up to 10 dimensions per metric to further clarify what data the metric collects. For more information about specifying dimensions, see Publishing Metrics in the Amazon CloudWatch User Guide.

Data points with time stamps from 24 hours ago or longer can take at least 48 hours to become available for GetMetricStatistics from the time they are submitted.

CloudWatch needs raw data points to calculate percentile statistics. If you publish data using a statistic set instead, you can only retrieve percentile statistics for this data if one of the following conditions is true:

The SampleCount value of the statistic set is 1

The Min and the Max values of the statistic set are equal

Arguments

Namespace = ::String – Required

The namespace for the metric data.

You cannot specify a namespace that begins with "AWS/". Namespaces that begin with "AWS/" are reserved for use by Amazon Web Services products.

MetricData = [Dict( ... )...] – Required

The data for the metric.

 MetricData = [Dict(
        "MetricName" => <required> ::String,
        "Dimensions" =>  [Dict(
            "Name" => <required> ::String,
            "Value" => <required> ::String
        )...],
        "Timestamp" =>  timestamp,
        "Value" =>  double,
        "StatisticValues" =>  Dict(
            "SampleCount" => <required> double,
            "Sum" => <required> double,
            "Minimum" => <required> double,
            "Maximum" => <required> double
        ),
        "Unit" =>  "Seconds", "Microseconds", "Milliseconds", "Bytes", "Kilobytes", "Megabytes", "Gigabytes", "Terabytes", "Bits", "Kilobits", "Megabits", "Gigabits", "Terabits", "Percent", "Count", "Bytes/Second", "Kilobytes/Second", "Megabytes/Second", "Gigabytes/Second", "Terabytes/Second", "Bits/Second", "Kilobits/Second", "Megabits/Second", "Gigabits/Second", "Terabits/Second", "Count/Second" or "None",
        "StorageResolution" =>  ::Int
    )...]

Exceptions

InvalidParameterValueException, MissingRequiredParameterException, InvalidParameterCombinationException or InternalServiceFault.

See also: AWS API Documentation

AmazonCloudWatch.set_alarm_state — Function.

set_alarm_state(::AWSConfig, arguments::Dict)
set_alarm_state([::AWSConfig]; AlarmName=, StateValue=, StateReason=, <keyword arguments>)

SetAlarmState Operation

Temporarily sets the state of an alarm for testing purposes. When the updated state differs from the previous value, the action configured for the appropriate state is invoked. For example, if your alarm is configured to send an Amazon SNS message when an alarm is triggered, temporarily changing the alarm state to ALARM sends an SNS message. The alarm returns to its actual state (often within seconds). Because the alarm state change happens quickly, it is typically only visible in the alarm's History tab in the Amazon CloudWatch console or through DescribeAlarmHistory.

Arguments

AlarmName = ::String – Required

The name for the alarm. This name must be unique within the AWS account. The maximum length is 255 characters.

StateValue = "OK", "ALARM" or "INSUFFICIENT_DATA" – Required

The value of the state.

StateReason = ::String – Required

The reason that this alarm is set to this specific state, in text format.

StateReasonData = ::String

The reason that this alarm is set to this specific state, in JSON format.

Exceptions

ResourceNotFound or InvalidFormatFault.

See also: AWS API Documentation