AmazonDeviceFarm.jl

AmazonDeviceFarm

AWS Device Farm is a service that enables mobile app developers to test Android, iOS, and Fire OS apps on physical phones, tablets, and other devices in the cloud.

This document is generated from apis/devicefarm-2015-06-23.normal.json. See JuliaCloud/AWSCore.jl.

AmazonDeviceFarm.create_device_pool

AmazonDeviceFarm.create_network_profile

AmazonDeviceFarm.create_project

AmazonDeviceFarm.create_remote_access_session

AmazonDeviceFarm.create_upload

AmazonDeviceFarm.delete_device_pool

AmazonDeviceFarm.delete_network_profile

AmazonDeviceFarm.delete_project

AmazonDeviceFarm.delete_remote_access_session

AmazonDeviceFarm.delete_run

AmazonDeviceFarm.delete_upload

AmazonDeviceFarm.get_account_settings

AmazonDeviceFarm.get_device

AmazonDeviceFarm.get_device_pool

AmazonDeviceFarm.get_device_pool_compatibility

AmazonDeviceFarm.get_job

AmazonDeviceFarm.get_network_profile

AmazonDeviceFarm.get_offering_status

AmazonDeviceFarm.get_project

AmazonDeviceFarm.get_remote_access_session

AmazonDeviceFarm.get_run

AmazonDeviceFarm.get_suite

AmazonDeviceFarm.get_test

AmazonDeviceFarm.get_upload

AmazonDeviceFarm.install_to_remote_access_session

AmazonDeviceFarm.list_artifacts

AmazonDeviceFarm.list_device_pools

AmazonDeviceFarm.list_devices

AmazonDeviceFarm.list_jobs

AmazonDeviceFarm.list_network_profiles

AmazonDeviceFarm.list_offering_promotions

AmazonDeviceFarm.list_offering_transactions

AmazonDeviceFarm.list_offerings

AmazonDeviceFarm.list_projects

AmazonDeviceFarm.list_remote_access_sessions

AmazonDeviceFarm.list_runs

AmazonDeviceFarm.list_samples

AmazonDeviceFarm.list_suites

AmazonDeviceFarm.list_tests

AmazonDeviceFarm.list_unique_problems

AmazonDeviceFarm.list_uploads

AmazonDeviceFarm.purchase_offering

AmazonDeviceFarm.renew_offering

AmazonDeviceFarm.schedule_run

AmazonDeviceFarm.stop_remote_access_session

AmazonDeviceFarm.stop_run

AmazonDeviceFarm.update_device_pool

AmazonDeviceFarm.update_network_profile

AmazonDeviceFarm.update_project

AmazonDeviceFarm.create_device_pool — Function.

create_device_pool(::AWSConfig, arguments::Dict)
create_device_pool([::AWSConfig]; projectArn=, name=, rules=, <keyword arguments>)

CreateDevicePool Operation

Creates a device pool.

Arguments

projectArn = ::String – Required

The ARN of the project for the device pool.

name = ::String – Required

The device pool's name.

description = ::String

The device pool's description.

rules = [Dict( ... )...] – Required

The device pool's rules.

 rules = [Dict(
        "attribute" =>  "ARN", "PLATFORM", "FORM_FACTOR", "MANUFACTURER", "REMOTE_ACCESS_ENABLED" or "APPIUM_VERSION",
        "operator" =>  "EQUALS", "LESS_THAN", "GREATER_THAN", "IN", "NOT_IN" or "CONTAINS",
        "value" =>  ::String
    )...]

Returns

CreateDevicePoolResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.create_network_profile — Function.

create_network_profile(::AWSConfig, arguments::Dict)
create_network_profile([::AWSConfig]; projectArn=, name=, <keyword arguments>)

CreateNetworkProfile Operation

Creates a network profile.

Arguments

projectArn = ::String – Required

The Amazon Resource Name (ARN) of the project for which you want to create a network profile.

name = ::String – Required

The name you wish to specify for the new network profile.

description = ::String

The description of the network profile.

type = "CURATED" or "PRIVATE"

The type of network profile you wish to create. Valid values are listed below.

uplinkBandwidthBits = ::Int

The data throughput rate in bits per second, as an integer from 0 to 104857600.

downlinkBandwidthBits = ::Int

The data throughput rate in bits per second, as an integer from 0 to 104857600.

uplinkDelayMs = ::Int

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

downlinkDelayMs = ::Int

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

uplinkJitterMs = ::Int

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

downlinkJitterMs = ::Int

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

uplinkLossPercent = ::Int

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

downlinkLossPercent = ::Int

Proportion of received packets that fail to arrive from 0 to 100 percent.

Returns

CreateNetworkProfileResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.create_project — Function.

create_project(::AWSConfig, arguments::Dict)
create_project([::AWSConfig]; name=, <keyword arguments>)

CreateProject Operation

Creates a new project.

Arguments

name = ::String – Required

The project's name.

defaultJobTimeoutMinutes = ::Int

Sets the execution timeout value (in minutes) for a project. All test runs in this project will use the specified execution timeout value unless overridden when scheduling a run.

Returns

CreateProjectResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.create_remote_access_session — Function.

create_remote_access_session(::AWSConfig, arguments::Dict)
create_remote_access_session([::AWSConfig]; projectArn=, deviceArn=, <keyword arguments>)

CreateRemoteAccessSession Operation

Specifies and starts a remote access session.

Arguments

projectArn = ::String – Required

The Amazon Resource Name (ARN) of the project for which you want to create a remote access session.

deviceArn = ::String – Required

The Amazon Resource Name (ARN) of the device for which you want to create a remote access session.

name = ::String

The name of the remote access session that you wish to create.

configuration = Dict("billingMethod" => "METERED" or "UNMETERED")

The configuration information for the remote access session request.

Returns

CreateRemoteAccessSessionResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.create_upload — Function.

create_upload(::AWSConfig, arguments::Dict)
create_upload([::AWSConfig]; projectArn=, name=, type=, <keyword arguments>)

CreateUpload Operation

Uploads an app or test scripts.

Arguments

projectArn = ::String – Required

The ARN of the project for the upload.

name = ::String – Required

The upload's file name. The name should not contain the '/' character. If uploading an iOS app, the file name needs to end with the .ipa extension. If uploading an Android app, the file name needs to end with the .apk extension. For all others, the file name must end with the .zip file extension.

type = "ANDROID_APP", "IOS_APP", "WEB_APP", "EXTERNAL_DATA", "APPIUM_JAVA_JUNIT_TEST_PACKAGE", "APPIUM_JAVA_TESTNG_TEST_PACKAGE", "APPIUM_PYTHON_TEST_PACKAGE", "APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE", "APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE", "APPIUM_WEB_PYTHON_TEST_PACKAGE", "CALABASH_TEST_PACKAGE", "INSTRUMENTATION_TEST_PACKAGE", "UIAUTOMATION_TEST_PACKAGE", "UIAUTOMATOR_TEST_PACKAGE", "XCTEST_TEST_PACKAGE" or "XCTEST_UI_TEST_PACKAGE" – Required

The upload's upload type.

Must be one of the following values:

ANDROID_APP: An Android upload.

IOS_APP: An iOS upload.

WEB_APP: A web appliction upload.

EXTERNAL_DATA: An external data upload.

APPIUM_JAVA_JUNIT_TEST_PACKAGE: An Appium Java JUnit test package upload.

APPIUM_JAVA_TESTNG_TEST_PACKAGE: An Appium Java TestNG test package upload.

APPIUM_PYTHON_TEST_PACKAGE: An Appium Python test package upload.

APPIUM_WEB_JAVA_JUNIT_TEST_PACKAGE: An Appium Java JUnit test package upload.

APPIUM_WEB_JAVA_TESTNG_TEST_PACKAGE: An Appium Java TestNG test package upload.

APPIUM_WEB_PYTHON_TEST_PACKAGE: An Appium Python test package upload.

CALABASH_TEST_PACKAGE: A Calabash test package upload.

INSTRUMENTATION_TEST_PACKAGE: An instrumentation upload.

UIAUTOMATION_TEST_PACKAGE: A uiautomation test package upload.

UIAUTOMATOR_TEST_PACKAGE: A uiautomator test package upload.

XCTEST_TEST_PACKAGE: An XCode test package upload.

XCTEST_UI_TEST_PACKAGE: An XCode UI test package upload.

Note If you call CreateUpload with WEB_APP specified, AWS Device Farm throws an ArgumentException error.

contentType = ::String

The upload's content type (for example, "application/octet-stream").

Returns

CreateUploadResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.delete_device_pool — Function.

delete_device_pool(::AWSConfig, arguments::Dict)
delete_device_pool([::AWSConfig]; arn=)

DeleteDevicePool Operation

Deletes a device pool given the pool ARN. Does not allow deletion of curated pools owned by the system.

Arguments

arn = ::String – Required

Represents the Amazon Resource Name (ARN) of the Device Farm device pool you wish to delete.

Returns

DeleteDevicePoolResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.delete_network_profile — Function.

delete_network_profile(::AWSConfig, arguments::Dict)
delete_network_profile([::AWSConfig]; arn=)

DeleteNetworkProfile Operation

Deletes a network profile.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the network profile you want to delete.

Returns

DeleteNetworkProfileResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.delete_project — Function.

delete_project(::AWSConfig, arguments::Dict)
delete_project([::AWSConfig]; arn=)

DeleteProject Operation

Deletes an AWS Device Farm project, given the project ARN.

Note Deleting this resource does not stop an in-progress run.

Arguments

arn = ::String – Required

Represents the Amazon Resource Name (ARN) of the Device Farm project you wish to delete.

Returns

DeleteProjectResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.delete_remote_access_session — Function.

delete_remote_access_session(::AWSConfig, arguments::Dict)
delete_remote_access_session([::AWSConfig]; arn=)

DeleteRemoteAccessSession Operation

Deletes a completed remote access session and its results.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the sesssion for which you want to delete remote access.

Returns

DeleteRemoteAccessSessionResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.delete_run — Function.

delete_run(::AWSConfig, arguments::Dict)
delete_run([::AWSConfig]; arn=)

DeleteRun Operation

Deletes the run, given the run ARN.

Note Deleting this resource does not stop an in-progress run.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) for the run you wish to delete.

Returns

DeleteRunResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.delete_upload — Function.

delete_upload(::AWSConfig, arguments::Dict)
delete_upload([::AWSConfig]; arn=)

DeleteUpload Operation

Deletes an upload given the upload ARN.

Arguments

arn = ::String – Required

Represents the Amazon Resource Name (ARN) of the Device Farm upload you wish to delete.

Returns

DeleteUploadResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_account_settings — Function.

get_account_settings(::AWSConfig, arguments::Dict)
get_account_settings([::AWSConfig]; )

GetAccountSettings Operation

Returns the number of unmetered iOS and/or unmetered Android devices that have been purchased by the account.

Arguments

Returns

GetAccountSettingsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_device — Function.

get_device(::AWSConfig, arguments::Dict)
get_device([::AWSConfig]; arn=)

GetDevice Operation

Gets information about a unique device type.

Arguments

arn = ::String – Required

The device type's ARN.

Returns

GetDeviceResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_device_pool — Function.

get_device_pool(::AWSConfig, arguments::Dict)
get_device_pool([::AWSConfig]; arn=)

GetDevicePool Operation

Gets information about a device pool.

Arguments

arn = ::String – Required

The device pool's ARN.

Returns

GetDevicePoolResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_device_pool_compatibility — Function.

get_device_pool_compatibility(::AWSConfig, arguments::Dict)
get_device_pool_compatibility([::AWSConfig]; devicePoolArn=, <keyword arguments>)

GetDevicePoolCompatibility Operation

Gets information about compatibility with a device pool.

Arguments

devicePoolArn = ::String – Required

The device pool's ARN.

appArn = ::String

The ARN of the app that is associated with the specified device pool.

testType = "BUILTIN_FUZZ", "BUILTIN_EXPLORER", "APPIUM_JAVA_JUNIT", "APPIUM_JAVA_TESTNG", "APPIUM_PYTHON", "APPIUM_WEB_JAVA_JUNIT", "APPIUM_WEB_JAVA_TESTNG", "APPIUM_WEB_PYTHON", "CALABASH", "INSTRUMENTATION", "UIAUTOMATION", "UIAUTOMATOR", "XCTEST" or "XCTEST_UI"

The test type for the specified device pool.

Allowed values include the following:

BUILTIN_FUZZ: The built-in fuzz type.

BUILTIN_EXPLORER: For Android, an app explorer that will traverse an Android app, interacting with it and capturing screenshots at the same time.

APPIUM_JAVA_JUNIT: The Appium Java JUnit type.

APPIUM_JAVA_TESTNG: The Appium Java TestNG type.

APPIUM_PYTHON: The Appium Python type.

APPIUM_WEB_JAVA_JUNIT: The Appium Java JUnit type for Web apps.

APPIUM_WEB_JAVA_TESTNG: The Appium Java TestNG type for Web apps.

APPIUM_WEB_PYTHON: The Appium Python type for Web apps.

CALABASH: The Calabash type.

INSTRUMENTATION: The Instrumentation type.

UIAUTOMATION: The uiautomation type.

UIAUTOMATOR: The uiautomator type.

XCTEST: The XCode test type.

XCTEST_UI: The XCode UI test type.

test = Dict( ... )

Information about the uploaded test to be run against the device pool.

 test = Dict(
        "type" => <required> "BUILTIN_FUZZ", "BUILTIN_EXPLORER", "APPIUM_JAVA_JUNIT", "APPIUM_JAVA_TESTNG", "APPIUM_PYTHON", "APPIUM_WEB_JAVA_JUNIT", "APPIUM_WEB_JAVA_TESTNG", "APPIUM_WEB_PYTHON", "CALABASH", "INSTRUMENTATION", "UIAUTOMATION", "UIAUTOMATOR", "XCTEST" or "XCTEST_UI",
        "testPackageArn" =>  ::String,
        "filter" =>  ::String,
        "parameters" =>  ::Dict{String,String}
    )

Returns

GetDevicePoolCompatibilityResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_job — Function.

get_job(::AWSConfig, arguments::Dict)
get_job([::AWSConfig]; arn=)

GetJob Operation

Gets information about a job.

Arguments

arn = ::String – Required

The job's ARN.

Returns

GetJobResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_network_profile — Function.

get_network_profile(::AWSConfig, arguments::Dict)
get_network_profile([::AWSConfig]; arn=)

GetNetworkProfile Operation

Returns information about a network profile.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the network profile you want to return information about.

Returns

GetNetworkProfileResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_offering_status — Function.

get_offering_status(::AWSConfig, arguments::Dict)
get_offering_status([::AWSConfig]; <keyword arguments>)

GetOfferingStatus Operation

Gets the current status and future status of all offerings purchased by an AWS account. The response indicates how many offerings are currently available and the offerings that will be available in the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

Arguments

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

GetOfferingStatusResult

Exceptions

ArgumentException, NotFoundException, NotEligibleException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_project — Function.

get_project(::AWSConfig, arguments::Dict)
get_project([::AWSConfig]; arn=)

GetProject Operation

Gets information about a project.

Arguments

arn = ::String – Required

The project's ARN.

Returns

GetProjectResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_remote_access_session — Function.

get_remote_access_session(::AWSConfig, arguments::Dict)
get_remote_access_session([::AWSConfig]; arn=)

GetRemoteAccessSession Operation

Returns a link to a currently running remote access session.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the remote access session about which you want to get session information.

Returns

GetRemoteAccessSessionResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_run — Function.

get_run(::AWSConfig, arguments::Dict)
get_run([::AWSConfig]; arn=)

GetRun Operation

Gets information about a run.

Arguments

arn = ::String – Required

The run's ARN.

Returns

GetRunResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_suite — Function.

get_suite(::AWSConfig, arguments::Dict)
get_suite([::AWSConfig]; arn=)

GetSuite Operation

Gets information about a suite.

Arguments

arn = ::String – Required

The suite's ARN.

Returns

GetSuiteResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_test — Function.

get_test(::AWSConfig, arguments::Dict)
get_test([::AWSConfig]; arn=)

GetTest Operation

Gets information about a test.

Arguments

arn = ::String – Required

The test's ARN.

Returns

GetTestResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.get_upload — Function.

get_upload(::AWSConfig, arguments::Dict)
get_upload([::AWSConfig]; arn=)

GetUpload Operation

Gets information about an upload.

Arguments

arn = ::String – Required

The upload's ARN.

Returns

GetUploadResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.install_to_remote_access_session — Function.

install_to_remote_access_session(::AWSConfig, arguments::Dict)
install_to_remote_access_session([::AWSConfig]; remoteAccessSessionArn=, appArn=)

InstallToRemoteAccessSession Operation

Installs an application to the device in a remote access session. For Android applications, the file must be in .apk format. For iOS applications, the file must be in .ipa format.

Arguments

remoteAccessSessionArn = ::String – Required

The Amazon Resource Name (ARN) of the remote access session about which you are requesting information.

appArn = ::String – Required

The Amazon Resource Name (ARN) of the app about which you are requesting information.

Returns

InstallToRemoteAccessSessionResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_artifacts — Function.

list_artifacts(::AWSConfig, arguments::Dict)
list_artifacts([::AWSConfig]; arn=, type=, <keyword arguments>)

ListArtifacts Operation

Gets information about artifacts.

Arguments

arn = ::String – Required

The Run, Job, Suite, or Test ARN.

type = "SCREENSHOT", "FILE" or "LOG" – Required

The artifacts' type.

Allowed values include:

FILE: The artifacts are files.

LOG: The artifacts are logs.

SCREENSHOT: The artifacts are screenshots.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListArtifactsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_device_pools — Function.

list_device_pools(::AWSConfig, arguments::Dict)
list_device_pools([::AWSConfig]; arn=, <keyword arguments>)

ListDevicePools Operation

Gets information about device pools.

Arguments

arn = ::String – Required

The project ARN.

type = "CURATED" or "PRIVATE"

The device pools' type.

Allowed values include:

CURATED: A device pool that is created and managed by AWS Device Farm.

PRIVATE: A device pool that is created and managed by the device pool developer.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListDevicePoolsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_devices — Function.

list_devices(::AWSConfig, arguments::Dict)
list_devices([::AWSConfig]; <keyword arguments>)

ListDevices Operation

Gets information about unique device types.

Arguments

arn = ::String

The Amazon Resource Name (ARN) of the project.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListDevicesResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_jobs — Function.

list_jobs(::AWSConfig, arguments::Dict)
list_jobs([::AWSConfig]; arn=, <keyword arguments>)

ListJobs Operation

Gets information about jobs.

Arguments

arn = ::String – Required

The jobs' ARNs.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListJobsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_network_profiles — Function.

list_network_profiles(::AWSConfig, arguments::Dict)
list_network_profiles([::AWSConfig]; arn=, <keyword arguments>)

ListNetworkProfiles Operation

Returns the list of available network profiles.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the project for which you want to list network profiles.

type = "CURATED" or "PRIVATE"

The type of network profile you wish to return information about. Valid values are listed below.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListNetworkProfilesResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_offering_promotions — Function.

list_offering_promotions(::AWSConfig, arguments::Dict)
list_offering_promotions([::AWSConfig]; <keyword arguments>)

ListOfferingPromotions Operation

Returns a list of offering promotions. Each offering promotion record contains the ID and description of the promotion. The API returns a NotEligible error if the caller is not permitted to invoke the operation. Contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

Arguments

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListOfferingPromotionsResult

Exceptions

ArgumentException, NotFoundException, NotEligibleException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_offering_transactions — Function.

list_offering_transactions(::AWSConfig, arguments::Dict)
list_offering_transactions([::AWSConfig]; <keyword arguments>)

ListOfferingTransactions Operation

Returns a list of all historical purchases, renewals, and system renewal transactions for an AWS account. The list is paginated and ordered by a descending timestamp (most recent transactions are first). The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

Arguments

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListOfferingTransactionsResult

Exceptions

ArgumentException, NotFoundException, NotEligibleException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_offerings — Function.

list_offerings(::AWSConfig, arguments::Dict)
list_offerings([::AWSConfig]; <keyword arguments>)

ListOfferings Operation

Returns a list of products or offerings that the user can manage through the API. Each offering record indicates the recurring price per unit and the frequency for that offering. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

Arguments

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListOfferingsResult

Exceptions

ArgumentException, NotFoundException, NotEligibleException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_projects — Function.

list_projects(::AWSConfig, arguments::Dict)
list_projects([::AWSConfig]; <keyword arguments>)

ListProjects Operation

Gets information about projects.

Arguments

arn = ::String

Optional. If no Amazon Resource Name (ARN) is specified, then AWS Device Farm returns a list of all projects for the AWS account. You can also specify a project ARN.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListProjectsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_remote_access_sessions — Function.

list_remote_access_sessions(::AWSConfig, arguments::Dict)
list_remote_access_sessions([::AWSConfig]; arn=, <keyword arguments>)

ListRemoteAccessSessions Operation

Returns a list of all currently running remote access sessions.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the remote access session about which you are requesting information.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListRemoteAccessSessionsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_runs — Function.

list_runs(::AWSConfig, arguments::Dict)
list_runs([::AWSConfig]; arn=, <keyword arguments>)

ListRuns Operation

Gets information about runs, given an AWS Device Farm project ARN.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the project for which you want to list runs.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListRunsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_samples — Function.

list_samples(::AWSConfig, arguments::Dict)
list_samples([::AWSConfig]; arn=, <keyword arguments>)

ListSamples Operation

Gets information about samples, given an AWS Device Farm project ARN

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the project for which you want to list samples.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListSamplesResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_suites — Function.

list_suites(::AWSConfig, arguments::Dict)
list_suites([::AWSConfig]; arn=, <keyword arguments>)

ListSuites Operation

Gets information about suites.

Arguments

arn = ::String – Required

The suites' ARNs.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListSuitesResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_tests — Function.

list_tests(::AWSConfig, arguments::Dict)
list_tests([::AWSConfig]; arn=, <keyword arguments>)

ListTests Operation

Gets information about tests.

Arguments

arn = ::String – Required

The tests' ARNs.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListTestsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_unique_problems — Function.

list_unique_problems(::AWSConfig, arguments::Dict)
list_unique_problems([::AWSConfig]; arn=, <keyword arguments>)

ListUniqueProblems Operation

Gets information about unique problems.

Arguments

arn = ::String – Required

The unique problems' ARNs.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListUniqueProblemsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.list_uploads — Function.

list_uploads(::AWSConfig, arguments::Dict)
list_uploads([::AWSConfig]; arn=, <keyword arguments>)

ListUploads Operation

Gets information about uploads, given an AWS Device Farm project ARN.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the project for which you want to list uploads.

nextToken = ::String

An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Returns

ListUploadsResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.purchase_offering — Function.

purchase_offering(::AWSConfig, arguments::Dict)
purchase_offering([::AWSConfig]; <keyword arguments>)

PurchaseOffering Operation

Immediately purchases offerings for an AWS account. Offerings renew with the latest total purchased quantity for an offering, unless the renewal was overridden. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

Arguments

offeringId = ::String

The ID of the offering.

quantity = ::Int

The number of device slots you wish to purchase in an offering request.

offeringPromotionId = ::String

The ID of the offering promotion to be applied to the purchase.

Returns

PurchaseOfferingResult

Exceptions

ArgumentException, NotFoundException, NotEligibleException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.renew_offering — Function.

renew_offering(::AWSConfig, arguments::Dict)
renew_offering([::AWSConfig]; <keyword arguments>)

RenewOffering Operation

Explicitly sets the quantity of devices to renew for an offering, starting from the effectiveDate of the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. Please contact aws-devicefarm-support@amazon.com if you believe that you should be able to invoke this operation.

Arguments

offeringId = ::String

The ID of a request to renew an offering.

quantity = ::Int

The quantity requested in an offering renewal.

Returns

RenewOfferingResult

Exceptions

ArgumentException, NotFoundException, NotEligibleException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.schedule_run — Function.

schedule_run(::AWSConfig, arguments::Dict)
schedule_run([::AWSConfig]; projectArn=, devicePoolArn=, test=, <keyword arguments>)

ScheduleRun Operation

Schedules a run.

Arguments

projectArn = ::String – Required

The ARN of the project for the run to be scheduled.

appArn = ::String

The ARN of the app to schedule a run.

devicePoolArn = ::String – Required

The ARN of the device pool for the run to be scheduled.

name = ::String

The name for the run to be scheduled.

test = Dict( ... ) – Required

Information about the test for the run to be scheduled.

 test = Dict(
        "type" => <required> "BUILTIN_FUZZ", "BUILTIN_EXPLORER", "APPIUM_JAVA_JUNIT", "APPIUM_JAVA_TESTNG", "APPIUM_PYTHON", "APPIUM_WEB_JAVA_JUNIT", "APPIUM_WEB_JAVA_TESTNG", "APPIUM_WEB_PYTHON", "CALABASH", "INSTRUMENTATION", "UIAUTOMATION", "UIAUTOMATOR", "XCTEST" or "XCTEST_UI",
        "testPackageArn" =>  ::String,
        "filter" =>  ::String,
        "parameters" =>  ::Dict{String,String}
    )

configuration = Dict( ... )

Information about the settings for the run to be scheduled.

 configuration = Dict(
        "extraDataPackageArn" =>  ::String,
        "networkProfileArn" =>  ::String,
        "locale" =>  ::String,
        "location" =>  Dict(
            "latitude" => <required> double,
            "longitude" => <required> double
        ),
        "radios" =>  Dict(
            "wifi" =>  ::Bool,
            "bluetooth" =>  ::Bool,
            "nfc" =>  ::Bool,
            "gps" =>  ::Bool
        ),
        "auxiliaryApps" =>  [::String...],
        "billingMethod" =>  "METERED" or "UNMETERED"
    )

executionConfiguration = Dict( ... )

Specifies configuration information about a test run, such as the execution timeout (in minutes).

 executionConfiguration = Dict(
        "jobTimeoutMinutes" =>  ::Int,
        "accountsCleanup" =>  ::Bool,
        "appPackagesCleanup" =>  ::Bool
    )

Returns

ScheduleRunResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException, IdempotencyException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.stop_remote_access_session — Function.

stop_remote_access_session(::AWSConfig, arguments::Dict)
stop_remote_access_session([::AWSConfig]; arn=)

StopRemoteAccessSession Operation

Ends a specified remote access session.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the remote access session you wish to stop.

Returns

StopRemoteAccessSessionResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.stop_run — Function.

stop_run(::AWSConfig, arguments::Dict)
stop_run([::AWSConfig]; arn=)

StopRun Operation

Initiates a stop request for the current test run. AWS Device Farm will immediately stop the run on devices where tests have not started executing, and you will not be billed for these devices. On devices where tests have started executing, Setup Suite and Teardown Suite tests will run to completion before stopping execution on those devices. You will be billed for Setup, Teardown, and any tests that were in progress or already completed.

Arguments

arn = ::String – Required

Represents the Amazon Resource Name (ARN) of the Device Farm run you wish to stop.

Returns

StopRunResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.update_device_pool — Function.

update_device_pool(::AWSConfig, arguments::Dict)
update_device_pool([::AWSConfig]; arn=, <keyword arguments>)

UpdateDevicePool Operation

Modifies the name, description, and rules in a device pool given the attributes and the pool ARN. Rule updates are all-or-nothing, meaning they can only be updated as a whole (or not at all).

Arguments

arn = ::String – Required

The Amazon Resourc Name (ARN) of the Device Farm device pool you wish to update.

name = ::String

A string representing the name of the device pool you wish to update.

description = ::String

A description of the device pool you wish to update.

rules = [Dict( ... )...]

Represents the rules you wish to modify for the device pool. Updating rules is optional; however, if you choose to update rules for your request, the update will replace the existing rules.

 rules = [Dict(
        "attribute" =>  "ARN", "PLATFORM", "FORM_FACTOR", "MANUFACTURER", "REMOTE_ACCESS_ENABLED" or "APPIUM_VERSION",
        "operator" =>  "EQUALS", "LESS_THAN", "GREATER_THAN", "IN", "NOT_IN" or "CONTAINS",
        "value" =>  ::String
    )...]

Returns

UpdateDevicePoolResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.update_network_profile — Function.

update_network_profile(::AWSConfig, arguments::Dict)
update_network_profile([::AWSConfig]; arn=, <keyword arguments>)

UpdateNetworkProfile Operation

Updates the network profile with specific settings.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the project that you wish to update network profile settings.

name = ::String

The name of the network profile about which you are returning information.

description = ::String

The descriptoin of the network profile about which you are returning information.

type = "CURATED" or "PRIVATE"

The type of network profile you wish to return information about. Valid values are listed below.

uplinkBandwidthBits = ::Int

The data throughput rate in bits per second, as an integer from 0 to 104857600.

downlinkBandwidthBits = ::Int

The data throughput rate in bits per second, as an integer from 0 to 104857600.

uplinkDelayMs = ::Int

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

downlinkDelayMs = ::Int

Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

uplinkJitterMs = ::Int

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

downlinkJitterMs = ::Int

Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

uplinkLossPercent = ::Int

Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

downlinkLossPercent = ::Int

Proportion of received packets that fail to arrive from 0 to 100 percent.

Returns

UpdateNetworkProfileResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation

AmazonDeviceFarm.update_project — Function.

update_project(::AWSConfig, arguments::Dict)
update_project([::AWSConfig]; arn=, <keyword arguments>)

UpdateProject Operation

Modifies the specified project name, given the project ARN and a new name.

Arguments

arn = ::String – Required

The Amazon Resource Name (ARN) of the project whose name you wish to update.

name = ::String

A string representing the new name of the project that you are updating.

defaultJobTimeoutMinutes = ::Int

The number of minutes a test run in the project will execute before it times out.

Returns

UpdateProjectResult

Exceptions

ArgumentException, NotFoundException, LimitExceededException or ServiceAccountException.

See also: AWS API Documentation