AmazonDirectoryService.jl

AmazonDirectoryService

AWS Directory Service

AWS Directory Service is a web service that makes it easy for you to setup and run directories in the AWS cloud, or connect your AWS resources with an existing on-premises Microsoft Active Directory. This guide provides detailed information about AWS Directory Service operations, data types, parameters, and errors. For information about AWS Directory Services features, see AWS Directory Service and the AWS Directory Service Administration Guide.

Note

AWS provides SDKs that consist of libraries and sample code for various programming languages and platforms (Java, Ruby, .Net, iOS, Android, etc.). The SDKs provide a convenient way to create programmatic access to AWS Directory Service and other AWS services. For more information about the AWS SDKs, including how to download and install them, see Tools for Amazon Web Services.

This document is generated from apis/ds-2015-04-16.normal.json. See JuliaCloud/AWSCore.jl.

AmazonDirectoryService.add_ip_routes

AmazonDirectoryService.add_tags_to_resource

AmazonDirectoryService.cancel_schema_extension

AmazonDirectoryService.connect_directory

AmazonDirectoryService.create_alias

AmazonDirectoryService.create_computer

AmazonDirectoryService.create_conditional_forwarder

AmazonDirectoryService.create_directory

AmazonDirectoryService.create_microsoft_ad

AmazonDirectoryService.create_snapshot

AmazonDirectoryService.create_trust

AmazonDirectoryService.delete_conditional_forwarder

AmazonDirectoryService.delete_directory

AmazonDirectoryService.delete_snapshot

AmazonDirectoryService.delete_trust

AmazonDirectoryService.deregister_event_topic

AmazonDirectoryService.describe_conditional_forwarders

AmazonDirectoryService.describe_directories

AmazonDirectoryService.describe_domain_controllers

AmazonDirectoryService.describe_event_topics

AmazonDirectoryService.describe_snapshots

AmazonDirectoryService.describe_trusts

AmazonDirectoryService.disable_radius

AmazonDirectoryService.disable_sso

AmazonDirectoryService.enable_radius

AmazonDirectoryService.enable_sso

AmazonDirectoryService.get_directory_limits

AmazonDirectoryService.get_snapshot_limits

AmazonDirectoryService.list_ip_routes

AmazonDirectoryService.list_schema_extensions

AmazonDirectoryService.list_tags_for_resource

AmazonDirectoryService.register_event_topic

AmazonDirectoryService.remove_ip_routes

AmazonDirectoryService.remove_tags_from_resource

AmazonDirectoryService.restore_from_snapshot

AmazonDirectoryService.start_schema_extension

AmazonDirectoryService.update_conditional_forwarder

AmazonDirectoryService.update_number_of_domain_controllers

AmazonDirectoryService.update_radius

AmazonDirectoryService.verify_trust

AmazonDirectoryService.add_ip_routes — Function.

add_ip_routes(::AWSConfig, arguments::Dict)
add_ip_routes([::AWSConfig]; DirectoryId=, IpRoutes=, <keyword arguments>)

AddIpRoutes Operation

If the DNS server for your on-premises domain uses a publicly addressable IP address, you must add a CIDR address block to correctly route traffic to and from your Microsoft AD on Amazon Web Services. AddIpRoutes adds this address block. You can also use AddIpRoutes to facilitate routing traffic that uses public IP ranges from your Microsoft AD on AWS to a peer VPC.

Before you call AddIpRoutes, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the AddIpRoutes operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Arguments

DirectoryId = ::String – Required

Identifier (ID) of the directory to which to add the address block.

IpRoutes = [Dict( ... )...] – Required

IP address blocks, using CIDR format, of the traffic to route. This is often the IP address block of the DNS server used for your on-premises domain.

 IpRoutes = [Dict(
        "CidrIp" =>  ::String,
        "Description" =>  ::String
    )...]

UpdateSecurityGroupForDirectoryControllers = ::Bool

If set to true, updates the inbound and outbound rules of the security group that has the description: "AWS created security group for directory ID directory controllers." Following are the new rules:

Inbound:

Type: Custom UDP Rule, Protocol: UDP, Range: 88, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 123, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 138, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 389, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 464, Source: 0.0.0.0/0

Type: Custom UDP Rule, Protocol: UDP, Range: 445, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 88, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 135, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 445, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 464, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 636, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 1024-65535, Source: 0.0.0.0/0

Type: Custom TCP Rule, Protocol: TCP, Range: 3268-33269, Source: 0.0.0.0/0

Type: DNS (UDP), Protocol: UDP, Range: 53, Source: 0.0.0.0/0

Type: DNS (TCP), Protocol: TCP, Range: 53, Source: 0.0.0.0/0

Type: LDAP, Protocol: TCP, Range: 389, Source: 0.0.0.0/0

Type: All ICMP, Protocol: All, Range: N/A, Source: 0.0.0.0/0

Outbound:

Type: All traffic, Protocol: All, Range: All, Destination: 0.0.0.0/0

These security rules impact an internal network interface that is not exposed publicly.

Returns

AddIpRoutesResult

Exceptions

EntityDoesNotExistException, EntityAlreadyExistsException, InvalidParameterException, DirectoryUnavailableException, IpRouteLimitExceededException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.add_tags_to_resource — Function.

add_tags_to_resource(::AWSConfig, arguments::Dict)
add_tags_to_resource([::AWSConfig]; ResourceId=, Tags=)

AddTagsToResource Operation

Adds or overwrites one or more tags for the specified directory. Each directory can have a maximum of 50 tags. Each tag consists of a key and optional value. Tag keys must be unique to each resource.

Arguments

ResourceId = ::String – Required

Identifier (ID) for the directory to which to add the tag.

Tags = [Dict( ... )...] – Required

The tags to be assigned to the directory.

 Tags = [Dict(
        "Key" => <required> ::String,
        "Value" => <required> ::String
    )...]

Returns

AddTagsToResourceResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, TagLimitExceededException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.cancel_schema_extension — Function.

cancel_schema_extension(::AWSConfig, arguments::Dict)
cancel_schema_extension([::AWSConfig]; DirectoryId=, SchemaExtensionId=)

CancelSchemaExtension Operation

Cancels an in-progress schema extension to a Microsoft AD directory. Once a schema extension has started replicating to all domain controllers, the task can no longer be canceled. A schema extension can be canceled during any of the following states; Initializing, CreatingSnapshot, and UpdatingSchema.

Arguments

DirectoryId = ::String – Required

The identifier of the directory whose schema extension will be canceled.

SchemaExtensionId = ::String – Required

The identifier of the schema extension that will be canceled.

Returns

CancelSchemaExtensionResult

Exceptions

EntityDoesNotExistException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.connect_directory — Function.

connect_directory(::AWSConfig, arguments::Dict)
connect_directory([::AWSConfig]; Name=, Password=, Size=, ConnectSettings=, <keyword arguments>)

ConnectDirectory Operation

Creates an AD Connector to connect to an on-premises directory.

Before you call ConnectDirectory, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the ConnectDirectory operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Arguments

Name = ::String – Required

The fully-qualified name of the on-premises directory, such as corp.example.com.

ShortName = ::String

The NetBIOS name of the on-premises directory, such as CORP.

Password = ::String – Required

The password for the on-premises user account.

Description = ::String

A textual description for the directory.

Size = "Small" or "Large" – Required

The size of the directory.

ConnectSettings = Dict( ... ) – Required

A DirectoryConnectSettings object that contains additional information for the operation.

 ConnectSettings = Dict(
        "VpcId" => <required> ::String,
        "SubnetIds" => <required> [::String...],
        "CustomerDnsIps" => <required> [::String...],
        "CustomerUserName" => <required> ::String
    )

Returns

ConnectDirectoryResult

Exceptions

DirectoryLimitExceededException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.create_alias — Function.

create_alias(::AWSConfig, arguments::Dict)
create_alias([::AWSConfig]; DirectoryId=, Alias=)

CreateAlias Operation

Creates an alias for a directory and assigns the alias to the directory. The alias is used to construct the access URL for the directory, such as http://<alias>.awsapps.com.

Important

After an alias has been created, it cannot be deleted or reused, so this operation should only be used when absolutely necessary.

Arguments

DirectoryId = ::String – Required

The identifier of the directory for which to create the alias.

Alias = ::String – Required

The requested alias.

The alias must be unique amongst all aliases in AWS. This operation throws an EntityAlreadyExistsException error if the alias already exists.

Returns

CreateAliasResult

Exceptions

EntityAlreadyExistsException, EntityDoesNotExistException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.create_computer — Function.

create_computer(::AWSConfig, arguments::Dict)
create_computer([::AWSConfig]; DirectoryId=, ComputerName=, Password=, <keyword arguments>)

CreateComputer Operation

Creates a computer account in the specified directory, and joins the computer to the directory.

Arguments

DirectoryId = ::String – Required

The identifier of the directory in which to create the computer account.

ComputerName = ::String – Required

The name of the computer account.

Password = ::String – Required

A one-time password that is used to join the computer to the directory. You should generate a random, strong password to use for this parameter.

OrganizationalUnitDistinguishedName = ::String

The fully-qualified distinguished name of the organizational unit to place the computer account in.

ComputerAttributes = [Dict( ... )...]

An array of Attribute objects that contain any LDAP attributes to apply to the computer account.

 ComputerAttributes = [Dict(
        "Name" =>  ::String,
        "Value" =>  ::String
    )...]

Returns

CreateComputerResult

Exceptions

AuthenticationFailedException, DirectoryUnavailableException, EntityAlreadyExistsException, EntityDoesNotExistException, InvalidParameterException, UnsupportedOperationException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.create_conditional_forwarder — Function.

create_conditional_forwarder(::AWSConfig, arguments::Dict)
create_conditional_forwarder([::AWSConfig]; DirectoryId=, RemoteDomainName=, DnsIpAddrs=)

CreateConditionalForwarder Operation

Creates a conditional forwarder associated with your AWS directory. Conditional forwarders are required in order to set up a trust relationship with another domain. The conditional forwarder points to the trusted domain.

Arguments

DirectoryId = ::String – Required

The directory ID of the AWS directory for which you are creating the conditional forwarder.

RemoteDomainName = ::String – Required

The fully qualified domain name (FQDN) of the remote domain with which you will set up a trust relationship.

DnsIpAddrs = [::String...] – Required

The IP addresses of the remote DNS server associated with RemoteDomainName.

Returns

CreateConditionalForwarderResult

Exceptions

EntityAlreadyExistsException, EntityDoesNotExistException, DirectoryUnavailableException, InvalidParameterException, UnsupportedOperationException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.create_directory — Function.

create_directory(::AWSConfig, arguments::Dict)
create_directory([::AWSConfig]; Name=, Password=, Size=, <keyword arguments>)

CreateDirectory Operation

Creates a Simple AD directory.

Before you call CreateDirectory, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the CreateDirectory operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Arguments

Name = ::String – Required

The fully qualified name for the directory, such as corp.example.com.

ShortName = ::String

The short name of the directory, such as CORP.

Password = ::String – Required

The password for the directory administrator. The directory creation process creates a directory administrator account with the username Administrator and this password.

Description = ::String

A textual description for the directory.

Size = "Small" or "Large" – Required

The size of the directory.

VpcSettings = Dict( ... )

A DirectoryVpcSettings object that contains additional information for the operation.

 VpcSettings = Dict(
        "VpcId" => <required> ::String,
        "SubnetIds" => <required> [::String...]
    )

Returns

CreateDirectoryResult

Exceptions

DirectoryLimitExceededException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.create_microsoft_ad — Function.

create_microsoft_ad(::AWSConfig, arguments::Dict)
create_microsoft_ad([::AWSConfig]; Name=, Password=, VpcSettings=, <keyword arguments>)

CreateMicrosoftAD Operation

Creates a Microsoft AD in the AWS cloud.

Before you call CreateMicrosoftAD, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the CreateMicrosoftAD operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Arguments

Name = ::String – Required

The fully qualified domain name for the directory, such as corp.example.com. This name will resolve inside your VPC only. It does not need to be publicly resolvable.

ShortName = ::String

The NetBIOS name for your domain. A short identifier for your domain, such as CORP. If you don't specify a NetBIOS name, it will default to the first part of your directory DNS. For example, CORP for the directory DNS corp.example.com.

Password = ::String – Required

The password for the default administrative user named Admin.

Description = ::String

A textual description for the directory. This label will appear on the AWS console Directory Details page after the directory is created.

VpcSettings = Dict( ... ) – Required

Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.

 VpcSettings = Dict(
        "VpcId" => <required> ::String,
        "SubnetIds" => <required> [::String...]
    )

Returns

CreateMicrosoftADResult

Exceptions

DirectoryLimitExceededException, InvalidParameterException, ClientException, ServiceException or UnsupportedOperationException.

See also: AWS API Documentation

AmazonDirectoryService.create_snapshot — Function.

create_snapshot(::AWSConfig, arguments::Dict)
create_snapshot([::AWSConfig]; DirectoryId=, <keyword arguments>)

CreateSnapshot Operation

Creates a snapshot of a Simple AD or Microsoft AD directory in the AWS cloud.

Note

You cannot take snapshots of AD Connector directories.

Arguments

DirectoryId = ::String – Required

The identifier of the directory of which to take a snapshot.

Name = ::String

The descriptive name to apply to the snapshot.

Returns

CreateSnapshotResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, SnapshotLimitExceededException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.create_trust — Function.

create_trust(::AWSConfig, arguments::Dict)
create_trust([::AWSConfig]; DirectoryId=, RemoteDomainName=, TrustPassword=, TrustDirection=, <keyword arguments>)

CreateTrust Operation

AWS Directory Service for Microsoft Active Directory allows you to configure trust relationships. For example, you can establish a trust between your Microsoft AD in the AWS cloud, and your existing on-premises Microsoft Active Directory. This would allow you to provide users and groups access to resources in either domain, with a single set of credentials.

This action initiates the creation of the AWS side of a trust relationship between a Microsoft AD in the AWS cloud and an external domain.

Arguments

DirectoryId = ::String – Required

The Directory ID of the Microsoft AD in the AWS cloud for which to establish the trust relationship.

RemoteDomainName = ::String – Required

The Fully Qualified Domain Name (FQDN) of the external domain for which to create the trust relationship.

TrustPassword = ::String – Required

The trust password. The must be the same password that was used when creating the trust relationship on the external domain.

TrustDirection = "One-Way: Outgoing", "One-Way: Incoming" or "Two-Way" – Required

The direction of the trust relationship.

TrustType = "Forest"

The trust relationship type.

ConditionalForwarderIpAddrs = [::String...]

The IP addresses of the remote DNS server associated with RemoteDomainName.

Returns

CreateTrustResult

Exceptions

EntityAlreadyExistsException, EntityDoesNotExistException, InvalidParameterException, ClientException, ServiceException or UnsupportedOperationException.

See also: AWS API Documentation

AmazonDirectoryService.delete_conditional_forwarder — Function.

delete_conditional_forwarder(::AWSConfig, arguments::Dict)
delete_conditional_forwarder([::AWSConfig]; DirectoryId=, RemoteDomainName=)

DeleteConditionalForwarder Operation

Deletes a conditional forwarder that has been set up for your AWS directory.

Arguments

DirectoryId = ::String – Required

The directory ID for which you are deleting the conditional forwarder.

RemoteDomainName = ::String – Required

The fully qualified domain name (FQDN) of the remote domain with which you are deleting the conditional forwarder.

Returns

DeleteConditionalForwarderResult

Exceptions

EntityDoesNotExistException, DirectoryUnavailableException, InvalidParameterException, UnsupportedOperationException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.delete_directory — Function.

delete_directory(::AWSConfig, arguments::Dict)
delete_directory([::AWSConfig]; DirectoryId=)

DeleteDirectory Operation

Deletes an AWS Directory Service directory.

Before you call DeleteDirectory, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the DeleteDirectory operation, see AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference.

Arguments

DirectoryId = ::String – Required

The identifier of the directory to delete.

Returns

DeleteDirectoryResult

Exceptions

EntityDoesNotExistException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.delete_snapshot — Function.

delete_snapshot(::AWSConfig, arguments::Dict)
delete_snapshot([::AWSConfig]; SnapshotId=)

DeleteSnapshot Operation

Deletes a directory snapshot.

Arguments

SnapshotId = ::String – Required

The identifier of the directory snapshot to be deleted.

Returns

DeleteSnapshotResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.delete_trust — Function.

delete_trust(::AWSConfig, arguments::Dict)
delete_trust([::AWSConfig]; TrustId=, <keyword arguments>)

DeleteTrust Operation

Deletes an existing trust relationship between your Microsoft AD in the AWS cloud and an external domain.

Arguments

TrustId = ::String – Required

The Trust ID of the trust relationship to be deleted.

DeleteAssociatedConditionalForwarder = ::Bool

Delete a conditional forwarder as part of a DeleteTrustRequest.

Returns

DeleteTrustResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, ClientException, ServiceException or UnsupportedOperationException.

See also: AWS API Documentation

AmazonDirectoryService.deregister_event_topic — Function.

deregister_event_topic(::AWSConfig, arguments::Dict)
deregister_event_topic([::AWSConfig]; DirectoryId=, TopicName=)

DeregisterEventTopic Operation

Removes the specified directory as a publisher to the specified SNS topic.

Arguments

DirectoryId = ::String – Required

The Directory ID to remove as a publisher. This directory will no longer send messages to the specified SNS topic.

TopicName = ::String – Required

The name of the SNS topic from which to remove the directory as a publisher.

Returns

DeregisterEventTopicResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.describe_conditional_forwarders — Function.

describe_conditional_forwarders(::AWSConfig, arguments::Dict)
describe_conditional_forwarders([::AWSConfig]; DirectoryId=, <keyword arguments>)

DescribeConditionalForwarders Operation

Obtains information about the conditional forwarders for this account.

If no input parameters are provided for RemoteDomainNames, this request describes all conditional forwarders for the specified directory ID.

Arguments

DirectoryId = ::String – Required

The directory ID for which to get the list of associated conditional forwarders.

RemoteDomainNames = [::String...]

The fully qualified domain names (FQDN) of the remote domains for which to get the list of associated conditional forwarders. If this member is null, all conditional forwarders are returned.

Returns

DescribeConditionalForwardersResult

Exceptions

EntityDoesNotExistException, DirectoryUnavailableException, InvalidParameterException, UnsupportedOperationException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.describe_directories — Function.

describe_directories(::AWSConfig, arguments::Dict)
describe_directories([::AWSConfig]; <keyword arguments>)

DescribeDirectories Operation

Obtains information about the directories that belong to this account.

You can retrieve information about specific directories by passing the directory identifiers in the DirectoryIds parameter. Otherwise, all directories that belong to the current account are returned.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeDirectoriesResult.NextToken member contains a token that you pass in the next call to DescribeDirectories to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

Arguments

DirectoryIds = [::String...]

A list of identifiers of the directories for which to obtain the information. If this member is null, all directories that belong to the current account are returned.

An empty list results in an InvalidParameterException being thrown.

NextToken = ::String

The DescribeDirectoriesResult.NextToken value from a previous call to DescribeDirectories. Pass null if this is the first call.

Limit = ::Int

The maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

Returns

DescribeDirectoriesResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, InvalidNextTokenException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.describe_domain_controllers — Function.

describe_domain_controllers(::AWSConfig, arguments::Dict)
describe_domain_controllers([::AWSConfig]; DirectoryId=, <keyword arguments>)

DescribeDomainControllers Operation

Provides information about any domain controllers in your directory.

Arguments

DirectoryId = ::String – Required

Identifier of the directory for which to retrieve the domain controller information.

DomainControllerIds = [::String...]

A list of identifiers for the domain controllers whose information will be provided.

NextToken = ::String

The DescribeDomainControllers.NextToken value from a previous call to DescribeDomainControllers. Pass null if this is the first call.

Limit = ::Int

The maximum number of items to return.

Returns

DescribeDomainControllersResult

Exceptions

EntityDoesNotExistException, InvalidNextTokenException, InvalidParameterException, ClientException, ServiceException or UnsupportedOperationException.

See also: AWS API Documentation

AmazonDirectoryService.describe_event_topics — Function.

describe_event_topics(::AWSConfig, arguments::Dict)
describe_event_topics([::AWSConfig]; <keyword arguments>)

DescribeEventTopics Operation

Obtains information about which SNS topics receive status messages from the specified directory.

If no input parameters are provided, such as DirectoryId or TopicName, this request describes all of the associations in the account.

Arguments

DirectoryId = ::String

The Directory ID for which to get the list of associated SNS topics. If this member is null, associations for all Directory IDs are returned.

TopicNames = [::String...]

A list of SNS topic names for which to obtain the information. If this member is null, all associations for the specified Directory ID are returned.

An empty list results in an InvalidParameterException being thrown.

Returns

DescribeEventTopicsResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.describe_snapshots — Function.

describe_snapshots(::AWSConfig, arguments::Dict)
describe_snapshots([::AWSConfig]; <keyword arguments>)

DescribeSnapshots Operation

Obtains information about the directory snapshots that belong to this account.

This operation supports pagination with the use of the NextToken request and response parameters. If more results are available, the DescribeSnapshots.NextToken member contains a token that you pass in the next call to DescribeSnapshots to retrieve the next set of items.

You can also specify a maximum number of return results with the Limit parameter.

Arguments

DirectoryId = ::String

The identifier of the directory for which to retrieve snapshot information.

SnapshotIds = [::String...]

A list of identifiers of the snapshots to obtain the information for. If this member is null or empty, all snapshots are returned using the Limit and NextToken members.

NextToken = ::String

The DescribeSnapshotsResult.NextToken value from a previous call to DescribeSnapshots. Pass null if this is the first call.

Limit = ::Int

The maximum number of objects to return.

Returns

DescribeSnapshotsResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, InvalidNextTokenException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.describe_trusts — Function.

describe_trusts(::AWSConfig, arguments::Dict)
describe_trusts([::AWSConfig]; <keyword arguments>)

DescribeTrusts Operation

Obtains information about the trust relationships for this account.

If no input parameters are provided, such as DirectoryId or TrustIds, this request describes all the trust relationships belonging to the account.

Arguments

DirectoryId = ::String

The Directory ID of the AWS directory that is a part of the requested trust relationship.

TrustIds = [::String...]

A list of identifiers of the trust relationships for which to obtain the information. If this member is null, all trust relationships that belong to the current account are returned.

An empty list results in an InvalidParameterException being thrown.

NextToken = ::String

The DescribeTrustsResult.NextToken value from a previous call to DescribeTrusts. Pass null if this is the first call.

Limit = ::Int

The maximum number of objects to return.

Returns

DescribeTrustsResult

Exceptions

EntityDoesNotExistException, InvalidNextTokenException, InvalidParameterException, ClientException, ServiceException or UnsupportedOperationException.

See also: AWS API Documentation

AmazonDirectoryService.disable_radius — Function.

disable_radius(::AWSConfig, arguments::Dict)
disable_radius([::AWSConfig]; DirectoryId=)

DisableRadius Operation

Disables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector directory.

Arguments

DirectoryId = ::String – Required

The identifier of the directory for which to disable MFA.

Returns

DisableRadiusResult

Exceptions

EntityDoesNotExistException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.disable_sso — Function.

disable_sso(::AWSConfig, arguments::Dict)
disable_sso([::AWSConfig]; DirectoryId=, <keyword arguments>)

DisableSso Operation

Disables single-sign on for a directory.

Arguments

DirectoryId = ::String – Required

The identifier of the directory for which to disable single-sign on.

UserName = ::String

The username of an alternate account to use to disable single-sign on. This is only used for AD Connector directories. This account must have privileges to remove a service principal name.

If the AD Connector service account does not have privileges to remove a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to disable single sign-on and are not stored by the service. The AD Connector service account is not changed.

Password = ::String

The password of an alternate account to use to disable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

Returns

DisableSsoResult

Exceptions

EntityDoesNotExistException, InsufficientPermissionsException, AuthenticationFailedException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.enable_radius — Function.

enable_radius(::AWSConfig, arguments::Dict)
enable_radius([::AWSConfig]; DirectoryId=, RadiusSettings=)

EnableRadius Operation

Enables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector directory.

Arguments

DirectoryId = ::String – Required

The identifier of the directory for which to enable MFA.

RadiusSettings = Dict( ... ) – Required

A RadiusSettings object that contains information about the RADIUS server.

 RadiusSettings = Dict(
        "RadiusServers" =>  [::String...],
        "RadiusPort" =>  ::Int,
        "RadiusTimeout" =>  ::Int,
        "RadiusRetries" =>  ::Int,
        "SharedSecret" =>  ::String,
        "AuthenticationProtocol" =>  "PAP", "CHAP", "MS-CHAPv1" or "MS-CHAPv2",
        "DisplayLabel" =>  ::String,
        "UseSameUsername" =>  ::Bool
    )

Returns

EnableRadiusResult

Exceptions

InvalidParameterException, EntityAlreadyExistsException, EntityDoesNotExistException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.enable_sso — Function.

enable_sso(::AWSConfig, arguments::Dict)
enable_sso([::AWSConfig]; DirectoryId=, <keyword arguments>)

EnableSso Operation

Enables single sign-on for a directory.

Arguments

DirectoryId = ::String – Required

The identifier of the directory for which to enable single-sign on.

UserName = ::String

The username of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. This account must have privileges to add a service principal name.

If the AD Connector service account does not have privileges to add a service principal name, you can specify an alternate account with the UserName and Password parameters. These credentials are only used to enable single sign-on and are not stored by the service. The AD Connector service account is not changed.

Password = ::String

The password of an alternate account to use to enable single-sign on. This is only used for AD Connector directories. For more information, see the UserName parameter.

Returns

EnableSsoResult

Exceptions

EntityDoesNotExistException, InsufficientPermissionsException, AuthenticationFailedException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.get_directory_limits — Function.

get_directory_limits(::AWSConfig, arguments::Dict)
get_directory_limits([::AWSConfig]; )

GetDirectoryLimits Operation

Obtains directory limit information for the current region.

Arguments

Returns

GetDirectoryLimitsResult

Exceptions

EntityDoesNotExistException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.get_snapshot_limits — Function.

get_snapshot_limits(::AWSConfig, arguments::Dict)
get_snapshot_limits([::AWSConfig]; DirectoryId=)

GetSnapshotLimits Operation

Obtains the manual snapshot limits for a directory.

Arguments

DirectoryId = ::String – Required

Contains the identifier of the directory to obtain the limits for.

Returns

GetSnapshotLimitsResult

Exceptions

EntityDoesNotExistException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.list_ip_routes — Function.

list_ip_routes(::AWSConfig, arguments::Dict)
list_ip_routes([::AWSConfig]; DirectoryId=, <keyword arguments>)

ListIpRoutes Operation

Lists the address blocks that you have added to a directory.

Arguments

DirectoryId = ::String – Required

Identifier (ID) of the directory for which you want to retrieve the IP addresses.

NextToken = ::String

The ListIpRoutes.NextToken value from a previous call to ListIpRoutes. Pass null if this is the first call.

Limit = ::Int

Maximum number of items to return. If this value is zero, the maximum number of items is specified by the limitations of the operation.

Returns

ListIpRoutesResult

Exceptions

EntityDoesNotExistException, InvalidNextTokenException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.list_schema_extensions — Function.

list_schema_extensions(::AWSConfig, arguments::Dict)
list_schema_extensions([::AWSConfig]; DirectoryId=, <keyword arguments>)

ListSchemaExtensions Operation

Lists all schema extensions applied to a Microsoft AD Directory.

Arguments

DirectoryId = ::String – Required

The identifier of the directory from which to retrieve the schema extension information.

NextToken = ::String

The ListSchemaExtensions.NextToken value from a previous call to ListSchemaExtensions. Pass null if this is the first call.

Limit = ::Int

The maximum number of items to return.

Returns

ListSchemaExtensionsResult

Exceptions

InvalidNextTokenException, EntityDoesNotExistException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.list_tags_for_resource — Function.

list_tags_for_resource(::AWSConfig, arguments::Dict)
list_tags_for_resource([::AWSConfig]; ResourceId=, <keyword arguments>)

ListTagsForResource Operation

Lists all tags on a directory.

Arguments

ResourceId = ::String – Required

Identifier (ID) of the directory for which you want to retrieve tags.

NextToken = ::String

Reserved for future use.

Limit = ::Int

Reserved for future use.

Returns

ListTagsForResourceResult

Exceptions

EntityDoesNotExistException, InvalidNextTokenException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.register_event_topic — Function.

register_event_topic(::AWSConfig, arguments::Dict)
register_event_topic([::AWSConfig]; DirectoryId=, TopicName=)

RegisterEventTopic Operation

Associates a directory with an SNS topic. This establishes the directory as a publisher to the specified SNS topic. You can then receive email or text (SMS) messages when the status of your directory changes. You get notified if your directory goes from an Active status to an Impaired or Inoperable status. You also receive a notification when the directory returns to an Active status.

Arguments

DirectoryId = ::String – Required

The Directory ID that will publish status messages to the SNS topic.

TopicName = ::String – Required

The SNS topic name to which the directory will publish status messages. This SNS topic must be in the same region as the specified Directory ID.

Returns

RegisterEventTopicResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.remove_ip_routes — Function.

remove_ip_routes(::AWSConfig, arguments::Dict)
remove_ip_routes([::AWSConfig]; DirectoryId=, CidrIps=)

RemoveIpRoutes Operation

Removes IP address blocks from a directory.

Arguments

DirectoryId = ::String – Required

Identifier (ID) of the directory from which you want to remove the IP addresses.

CidrIps = [::String...] – Required

IP address blocks that you want to remove.

Returns

RemoveIpRoutesResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, DirectoryUnavailableException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.remove_tags_from_resource — Function.

remove_tags_from_resource(::AWSConfig, arguments::Dict)
remove_tags_from_resource([::AWSConfig]; ResourceId=, TagKeys=)

RemoveTagsFromResource Operation

Removes tags from a directory.

Arguments

ResourceId = ::String – Required

Identifier (ID) of the directory from which to remove the tag.

TagKeys = [::String...] – Required

The tag key (name) of the tag to be removed.

Returns

RemoveTagsFromResourceResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.restore_from_snapshot — Function.

restore_from_snapshot(::AWSConfig, arguments::Dict)
restore_from_snapshot([::AWSConfig]; SnapshotId=)

RestoreFromSnapshot Operation

Restores a directory using an existing directory snapshot.

When you restore a directory from a snapshot, any changes made to the directory after the snapshot date are overwritten.

This action returns as soon as the restore operation is initiated. You can monitor the progress of the restore operation by calling the DescribeDirectories operation with the directory identifier. When the DirectoryDescription.Stage value changes to Active, the restore operation is complete.

Arguments

SnapshotId = ::String – Required

The identifier of the snapshot to restore from.

Returns

RestoreFromSnapshotResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.start_schema_extension — Function.

start_schema_extension(::AWSConfig, arguments::Dict)
start_schema_extension([::AWSConfig]; DirectoryId=, CreateSnapshotBeforeSchemaExtension=, LdifContent=, Description=)

StartSchemaExtension Operation

Applies a schema extension to a Microsoft AD directory.

Arguments

DirectoryId = ::String – Required

The identifier of the directory for which the schema extension will be applied to.

CreateSnapshotBeforeSchemaExtension = ::Bool – Required

If true, creates a snapshot of the directory before applying the schema extension.

LdifContent = ::String – Required

The LDIF file represented as a string. To construct the LdifContent string, precede each line as it would be formatted in an ldif file with \n. See the example request below for more details. The file size can be no larger than 1MB.

Description = ::String – Required

A description of the schema extension.

Returns

StartSchemaExtensionResult

Exceptions

DirectoryUnavailableException, EntityDoesNotExistException, InvalidParameterException, SnapshotLimitExceededException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.update_conditional_forwarder — Function.

update_conditional_forwarder(::AWSConfig, arguments::Dict)
update_conditional_forwarder([::AWSConfig]; DirectoryId=, RemoteDomainName=, DnsIpAddrs=)

UpdateConditionalForwarder Operation

Updates a conditional forwarder that has been set up for your AWS directory.

Arguments

DirectoryId = ::String – Required

The directory ID of the AWS directory for which to update the conditional forwarder.

RemoteDomainName = ::String – Required

The fully qualified domain name (FQDN) of the remote domain with which you will set up a trust relationship.

DnsIpAddrs = [::String...] – Required

The updated IP addresses of the remote DNS server associated with the conditional forwarder.

Returns

UpdateConditionalForwarderResult

Exceptions

EntityDoesNotExistException, DirectoryUnavailableException, InvalidParameterException, UnsupportedOperationException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.update_number_of_domain_controllers — Function.

update_number_of_domain_controllers(::AWSConfig, arguments::Dict)
update_number_of_domain_controllers([::AWSConfig]; DirectoryId=, DesiredNumber=)

UpdateNumberOfDomainControllers Operation

Adds or removes domain controllers to or from the directory. Based on the difference between current value and new value (provided through this API call), domain controllers will be added or removed. It may take up to 45 minutes for any new domain controllers to become fully active once the requested number of domain controllers is updated. During this time, you cannot make another update request.

Arguments

DirectoryId = ::String – Required

Identifier of the directory to which the domain controllers will be added or removed.

DesiredNumber = ::Int – Required

The number of domain controllers desired in the directory.

Returns

UpdateNumberOfDomainControllersResult

Exceptions

EntityDoesNotExistException, DirectoryUnavailableException, DomainControllerLimitExceededException, InvalidParameterException, UnsupportedOperationException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.update_radius — Function.

update_radius(::AWSConfig, arguments::Dict)
update_radius([::AWSConfig]; DirectoryId=, RadiusSettings=)

UpdateRadius Operation

Updates the Remote Authentication Dial In User Service (RADIUS) server information for an AD Connector directory.

Arguments

DirectoryId = ::String – Required

The identifier of the directory for which to update the RADIUS server information.

RadiusSettings = Dict( ... ) – Required

A RadiusSettings object that contains information about the RADIUS server.

 RadiusSettings = Dict(
        "RadiusServers" =>  [::String...],
        "RadiusPort" =>  ::Int,
        "RadiusTimeout" =>  ::Int,
        "RadiusRetries" =>  ::Int,
        "SharedSecret" =>  ::String,
        "AuthenticationProtocol" =>  "PAP", "CHAP", "MS-CHAPv1" or "MS-CHAPv2",
        "DisplayLabel" =>  ::String,
        "UseSameUsername" =>  ::Bool
    )

Returns

UpdateRadiusResult

Exceptions

InvalidParameterException, EntityDoesNotExistException, ClientException or ServiceException.

See also: AWS API Documentation

AmazonDirectoryService.verify_trust — Function.

verify_trust(::AWSConfig, arguments::Dict)
verify_trust([::AWSConfig]; TrustId=)

VerifyTrust Operation

AWS Directory Service for Microsoft Active Directory allows you to configure and verify trust relationships.

This action verifies a trust relationship between your Microsoft AD in the AWS cloud and an external domain.

Arguments

TrustId = ::String – Required

The unique Trust ID of the trust relationship to verify.

Returns

VerifyTrustResult

Exceptions

EntityDoesNotExistException, InvalidParameterException, ClientException, ServiceException or UnsupportedOperationException.

See also: AWS API Documentation