AmazonDynamoDBStreams.jl

AmazonDynamoDBStreams

Amazon DynamoDB

Amazon DynamoDB Streams provides API actions for accessing streams and processing stream records. To learn more about application development with Streams, see Capturing Table Activity with DynamoDB Streams in the Amazon DynamoDB Developer Guide.

This document is generated from apis/streams.dynamodb-2012-08-10.normal.json. See JuliaCloud/AWSCore.jl.

AmazonDynamoDBStreams.describe_stream

AmazonDynamoDBStreams.get_records

AmazonDynamoDBStreams.get_shard_iterator

AmazonDynamoDBStreams.list_streams

AmazonDynamoDBStreams.describe_stream — Function.

describe_stream(::AWSConfig, arguments::Dict)
describe_stream([::AWSConfig]; StreamArn=, <keyword arguments>)

DescribeStream Operation

Returns information about a stream, including the current status of the stream, its Amazon Resource Name (ARN), the composition of its shards, and its corresponding DynamoDB table.

Note

You can call DescribeStream at a maximum rate of 10 times per second.

Each shard in the stream has a SequenceNumberRange associated with it. If the SequenceNumberRange has a StartingSequenceNumber but no EndingSequenceNumber, then the shard is still open (able to receive more stream records). If both StartingSequenceNumber and EndingSequenceNumber are present, then that shard is closed and can no longer receive more data.

Arguments

StreamArn = ::String – Required

The Amazon Resource Name (ARN) for the stream.

Limit = ::Int

The maximum number of shard objects to return. The upper limit is 100.

ExclusiveStartShardId = ::String

The shard ID of the first item that this operation will evaluate. Use the value that was returned for LastEvaluatedShardId in the previous operation.

Returns

DescribeStreamOutput

Exceptions

ResourceNotFoundException or InternalServerError.

See also: AWS API Documentation

AmazonDynamoDBStreams.get_records — Function.

get_records(::AWSConfig, arguments::Dict)
get_records([::AWSConfig]; ShardIterator=, <keyword arguments>)

GetRecords Operation

Retrieves the stream records from a given shard.

Specify a shard iterator using the ShardIterator parameter. The shard iterator specifies the position in the shard from which you want to start reading stream records sequentially. If there are no stream records available in the portion of the shard that the iterator points to, GetRecords returns an empty list. Note that it might take multiple calls to get to a portion of the shard that contains stream records.

Note

GetRecords can retrieve a maximum of 1 MB of data or 1000 stream records, whichever comes first.

Arguments

ShardIterator = ::String – Required

A shard iterator that was retrieved from a previous GetShardIterator operation. This iterator can be used to access the stream records in this shard.

Limit = ::Int

The maximum number of records to return from the shard. The upper limit is 1000.

Returns

GetRecordsOutput

Exceptions

ResourceNotFoundException, LimitExceededException, InternalServerError, ExpiredIteratorException or TrimmedDataAccessException.

See also: AWS API Documentation

AmazonDynamoDBStreams.get_shard_iterator — Function.

get_shard_iterator(::AWSConfig, arguments::Dict)
get_shard_iterator([::AWSConfig]; StreamArn=, ShardId=, ShardIteratorType=, <keyword arguments>)

GetShardIterator Operation

Returns a shard iterator. A shard iterator provides information about how to retrieve the stream records from within a shard. Use the shard iterator in a subsequent GetRecords request to read the stream records from the shard.

Note

A shard iterator expires 15 minutes after it is returned to the requester.

Arguments

StreamArn = ::String – Required

The Amazon Resource Name (ARN) for the stream.

ShardId = ::String – Required

The identifier of the shard. The iterator will be returned for this shard ID.

ShardIteratorType = "TRIM_HORIZON", "LATEST", "AT_SEQUENCE_NUMBER" or "AFTER_SEQUENCE_NUMBER" – Required

Determines how the shard iterator is used to start reading stream records from the shard:

AT_SEQUENCE_NUMBER - Start reading exactly from the position denoted by a specific sequence number.

AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number.

TRIM_HORIZON - Start reading at the last (untrimmed) stream record, which is the oldest record in the shard. In DynamoDB Streams, there is a 24 hour limit on data retention. Stream records whose age exceeds this limit are subject to removal (trimming) from the stream.

LATEST - Start reading just after the most recent stream record in the shard, so that you always read the most recent data in the shard.

SequenceNumber = ::String

The sequence number of a stream record in the shard from which to start reading.

Returns

GetShardIteratorOutput

Exceptions

ResourceNotFoundException, InternalServerError or TrimmedDataAccessException.

See also: AWS API Documentation

AmazonDynamoDBStreams.list_streams — Function.

list_streams(::AWSConfig, arguments::Dict)
list_streams([::AWSConfig]; <keyword arguments>)

ListStreams Operation

Returns an array of stream ARNs associated with the current account and endpoint. If the TableName parameter is present, then ListStreams will return only the streams ARNs for that table.

Note

You can call ListStreams at a maximum rate of 5 times per second.

Arguments

TableName = ::String

If this parameter is provided, then only the streams associated with this table name are returned.

Limit = ::Int

The maximum number of streams to return. The upper limit is 100.

ExclusiveStartStreamArn = ::String

The ARN (Amazon Resource Name) of the first item that this operation will evaluate. Use the value that was returned for LastEvaluatedStreamArn in the previous operation.

Returns

ListStreamsOutput

Exceptions

ResourceNotFoundException or InternalServerError.

See also: AWS API Documentation