AmazonEMR.jl

AmazonEMR

Amazon EMR is a web service that makes it easy to process large amounts of data efficiently. Amazon EMR uses Hadoop processing combined with several AWS products to do tasks such as web indexing, data mining, log file analysis, machine learning, scientific simulation, and data warehousing.

This document is generated from apis/elasticmapreduce-2009-03-31.normal.json. See JuliaCloud/AWSCore.jl.

AmazonEMR.add_instance_fleet

AmazonEMR.add_instance_groups

AmazonEMR.add_job_flow_steps

AmazonEMR.add_tags

AmazonEMR.cancel_steps

AmazonEMR.create_security_configuration

AmazonEMR.delete_security_configuration

AmazonEMR.describe_cluster

AmazonEMR.describe_job_flows

AmazonEMR.describe_security_configuration

AmazonEMR.describe_step

AmazonEMR.list_bootstrap_actions

AmazonEMR.list_clusters

AmazonEMR.list_instance_fleets

AmazonEMR.list_instance_groups

AmazonEMR.list_instances

AmazonEMR.list_security_configurations

AmazonEMR.list_steps

AmazonEMR.modify_instance_fleet

AmazonEMR.modify_instance_groups

AmazonEMR.put_auto_scaling_policy

AmazonEMR.remove_auto_scaling_policy

AmazonEMR.remove_tags

AmazonEMR.run_job_flow

AmazonEMR.set_termination_protection

AmazonEMR.set_visible_to_all_users

AmazonEMR.terminate_job_flows

AmazonEMR.add_instance_fleet — Function.

add_instance_fleet(::AWSConfig, arguments::Dict)
add_instance_fleet([::AWSConfig]; ClusterId=, InstanceFleet=)

AddInstanceFleet Operation

Adds an instance fleet to a running cluster.

Note

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x.

Arguments

ClusterId = ::String – Required

The unique identifier of the cluster.

InstanceFleet = Dict( ... ) – Required

Specifies the configuration of the instance fleet.

 InstanceFleet = Dict(
        "Name" =>  ::String,
        "InstanceFleetType" => <required> "MASTER", "CORE" or "TASK",
        "TargetOnDemandCapacity" =>  ::Int,
        "TargetSpotCapacity" =>  ::Int,
        "InstanceTypeConfigs" =>  [Dict(
            "InstanceType" => <required> ::String,
            "WeightedCapacity" =>  ::Int,
            "BidPrice" =>  ::String,
            "BidPriceAsPercentageOfOnDemandPrice" =>  double,
            "EbsConfiguration" =>  Dict(
                "EbsBlockDeviceConfigs" =>  [Dict(
                    "VolumeSpecification" => <required> Dict(
                        "VolumeType" => <required> ::String,
                        "Iops" =>  ::Int,
                        "SizeInGB" => <required> ::Int
                    ),
                    "VolumesPerInstance" =>  ::Int
                )...],
                "EbsOptimized" =>  ::Bool
            ),
            "Configurations" =>  [Dict(
                "Classification" =>  ::String,
                "Configurations" =>  list,
                "Properties" =>  ::Dict{String,String}
            )...]
        )...],
        "LaunchSpecifications" =>  Dict("SpotSpecification" => <required> Dict(
                "TimeoutDurationMinutes" => <required> ::Int,
                "TimeoutAction" => <required> "SWITCH_TO_ON_DEMAND" or "TERMINATE_CLUSTER",
                "BlockDurationMinutes" =>  ::Int
            ))
    )

Returns

AddInstanceFleetOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.add_instance_groups — Function.

add_instance_groups(::AWSConfig, arguments::Dict)
add_instance_groups([::AWSConfig]; InstanceGroups=, JobFlowId=)

AddInstanceGroups Operation

Adds one or more instance groups to a running cluster.

Arguments

InstanceGroups = [Dict( ... )...] – Required

Instance groups to add.

 InstanceGroups = [Dict(
        "Name" =>  ::String,
        "Market" =>  "ON_DEMAND" or "SPOT",
        "InstanceRole" => <required> "MASTER", "CORE" or "TASK",
        "BidPrice" =>  ::String,
        "InstanceType" => <required> ::String,
        "InstanceCount" => <required> ::Int,
        "Configurations" =>  [Dict(
            "Classification" =>  ::String,
            "Configurations" =>  list,
            "Properties" =>  ::Dict{String,String}
        )...],
        "EbsConfiguration" =>  Dict(
            "EbsBlockDeviceConfigs" =>  [Dict(
                "VolumeSpecification" => <required> Dict(
                    "VolumeType" => <required> ::String,
                    "Iops" =>  ::Int,
                    "SizeInGB" => <required> ::Int
                ),
                "VolumesPerInstance" =>  ::Int
            )...],
            "EbsOptimized" =>  ::Bool
        ),
        "AutoScalingPolicy" =>  Dict(
            "Constraints" => <required> Dict(
                "MinCapacity" => <required> ::Int,
                "MaxCapacity" => <required> ::Int
            ),
            "Rules" => <required> [Dict(
                "Name" => <required> ::String,
                "Description" =>  ::String,
                "Action" => <required> Dict(
                    "Market" =>  "ON_DEMAND" or "SPOT",
                    "SimpleScalingPolicyConfiguration" => <required> Dict(
                        "AdjustmentType" =>  "CHANGE_IN_CAPACITY", "PERCENT_CHANGE_IN_CAPACITY" or "EXACT_CAPACITY",
                        "ScalingAdjustment" => <required> ::Int,
                        "CoolDown" =>  ::Int
                    )
                ),
                "Trigger" => <required> Dict("CloudWatchAlarmDefinition" => <required> Dict(
                        "ComparisonOperator" => <required> "GREATER_THAN_OR_EQUAL", "GREATER_THAN", "LESS_THAN" or "LESS_THAN_OR_EQUAL",
                        "EvaluationPeriods" =>  ::Int,
                        "MetricName" => <required> ::String,
                        "Namespace" =>  ::String,
                        "Period" => <required> ::Int,
                        "Statistic" =>  "SAMPLE_COUNT", "AVERAGE", "SUM", "MINIMUM" or "MAXIMUM",
                        "Threshold" => <required> double,
                        "Unit" =>  "NONE", "SECONDS", "MICRO_SECONDS", "MILLI_SECONDS", "BYTES", "KILO_BYTES", "MEGA_BYTES", "GIGA_BYTES", "TERA_BYTES", "BITS", "KILO_BITS", "MEGA_BITS", "GIGA_BITS", "TERA_BITS", "PERCENT", "COUNT", "BYTES_PER_SECOND", "KILO_BYTES_PER_SECOND", "MEGA_BYTES_PER_SECOND", "GIGA_BYTES_PER_SECOND", "TERA_BYTES_PER_SECOND", "BITS_PER_SECOND", "KILO_BITS_PER_SECOND", "MEGA_BITS_PER_SECOND", "GIGA_BITS_PER_SECOND", "TERA_BITS_PER_SECOND" or "COUNT_PER_SECOND",
                        "Dimensions" =>  [Dict(
                            "Key" =>  ::String,
                            "Value" =>  ::String
                        )...]
                    ))
            )...]
        )
    )...]

JobFlowId = ::String – Required

Job flow in which to add the instance groups.

Returns

AddInstanceGroupsOutput

Exceptions

InternalServerError.

See also: AWS API Documentation

AmazonEMR.add_job_flow_steps — Function.

add_job_flow_steps(::AWSConfig, arguments::Dict)
add_job_flow_steps([::AWSConfig]; JobFlowId=, Steps=)

AddJobFlowSteps Operation

AddJobFlowSteps adds new steps to a running cluster. A maximum of 256 steps are allowed in each job flow.

If your cluster is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using SSH to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, see Add More than 256 Steps to a Cluster in the Amazon EMR Management Guide.

A step specifies the location of a JAR file stored either on the master node of the cluster or in Amazon S3. Each step is performed by the main function of the main class of the JAR file. The main class can be specified either in the manifest of the JAR or by using the MainFunction parameter of the step.

Amazon EMR executes each step in the order listed. For a step to be considered complete, the main function must exit with a zero exit code and all Hadoop jobs started while the step was running must have completed and run successfully.

You can only add steps to a cluster that is in one of the following states: STARTING, BOOTSTRAPPING, RUNNING, or WAITING.

Arguments

JobFlowId = ::String – Required

A string that uniquely identifies the job flow. This identifier is returned by RunJobFlow and can also be obtained from ListClusters.

Steps = [Dict( ... )...] – Required

A list of StepConfig to be executed by the job flow.

 Steps = [Dict(
        "Name" => <required> ::String,
        "ActionOnFailure" =>  "TERMINATE_JOB_FLOW", "TERMINATE_CLUSTER", "CANCEL_AND_WAIT" or "CONTINUE",
        "HadoopJarStep" => <required> Dict(
            "Properties" =>  [Dict(
                "Key" =>  ::String,
                "Value" =>  ::String
            )...],
            "Jar" => <required> ::String,
            "MainClass" =>  ::String,
            "Args" =>  [::String...]
        )
    )...]

Returns

AddJobFlowStepsOutput

Exceptions

InternalServerError.

See also: AWS API Documentation

AmazonEMR.add_tags — Function.

add_tags(::AWSConfig, arguments::Dict)
add_tags([::AWSConfig]; ResourceId=, Tags=)

AddTags Operation

Adds tags to an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMR Resources.

Arguments

ResourceId = ::String – Required

The Amazon EMR resource identifier to which tags will be added. This value must be a cluster identifier.

Tags = [Dict( ... )...] – Required

A list of tags to associate with a cluster and propagate to EC2 instances. Tags are user-defined key/value pairs that consist of a required key string with a maximum of 128 characters, and an optional value string with a maximum of 256 characters.

 Tags = [Dict(
        "Key" =>  ::String,
        "Value" =>  ::String
    )...]

Returns

AddTagsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.cancel_steps — Function.

cancel_steps(::AWSConfig, arguments::Dict)
cancel_steps([::AWSConfig]; <keyword arguments>)

CancelSteps Operation

Cancels a pending step or steps in a running cluster. Available only in Amazon EMR versions 4.8.0 and later, excluding version 5.0.0. A maximum of 256 steps are allowed in each CancelSteps request. CancelSteps is idempotent but asynchronous; it does not guarantee a step will be canceled, even if the request is successfully submitted. You can only cancel steps that are in a PENDING state.

Arguments

ClusterId = ::String

The ClusterID for which specified steps will be canceled. Use RunJobFlow and ListClusters to get ClusterIDs.

StepIds = [::String...]

The list of StepIDs to cancel. Use ListSteps to get steps and their states for the specified cluster.

Returns

CancelStepsOutput

Exceptions

InternalServerError or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.create_security_configuration — Function.

create_security_configuration(::AWSConfig, arguments::Dict)
create_security_configuration([::AWSConfig]; Name=, SecurityConfiguration=)

CreateSecurityConfiguration Operation

Creates a security configuration, which is stored in the service and can be specified when a cluster is created.

Arguments

Name = ::String – Required

The name of the security configuration.

SecurityConfiguration = ::String – Required

The security configuration details in JSON format.

Returns

CreateSecurityConfigurationOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.delete_security_configuration — Function.

delete_security_configuration(::AWSConfig, arguments::Dict)
delete_security_configuration([::AWSConfig]; Name=)

DeleteSecurityConfiguration Operation

Deletes a security configuration.

Arguments

Name = ::String – Required

The name of the security configuration.

Returns

DeleteSecurityConfigurationOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.describe_cluster — Function.

describe_cluster(::AWSConfig, arguments::Dict)
describe_cluster([::AWSConfig]; ClusterId=)

DescribeCluster Operation

Provides cluster-level details including status, hardware and software configuration, VPC settings, and so on. For information about the cluster steps, see ListSteps.

Arguments

ClusterId = ::String – Required

The identifier of the cluster to describe.

Returns

DescribeClusterOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.describe_job_flows — Function.

describe_job_flows(::AWSConfig, arguments::Dict)
describe_job_flows([::AWSConfig]; <keyword arguments>)

DescribeJobFlows Operation

This API is deprecated and will eventually be removed. We recommend you use ListClusters, DescribeCluster, ListSteps, ListInstanceGroups and ListBootstrapActions instead.

DescribeJobFlows returns a list of job flows that match all of the supplied parameters. The parameters can include a list of job flow IDs, job flow states, and restrictions on job flow creation date and time.

Regardless of supplied parameters, only job flows created within the last two months are returned.

If no parameters are supplied, then job flows matching either of the following criteria are returned:

Job flows created and completed in the last two weeks

Job flows created within the last two months that are in one of the following states: RUNNING, WAITING, SHUTTING_DOWN, STARTING

Amazon EMR can return a maximum of 512 job flow descriptions.

Arguments

CreatedAfter = timestamp

Return only job flows created after this date and time.

CreatedBefore = timestamp

Return only job flows created before this date and time.

JobFlowIds = [::String...]

Return only job flows whose job flow ID is contained in this list.

JobFlowStates = ["STARTING", "BOOTSTRAPPING", "RUNNING", "WAITING", "SHUTTING_DOWN", "TERMINATED", "COMPLETED" or "FAILED"...]

Return only job flows whose state is contained in this list.

Returns

DescribeJobFlowsOutput

Exceptions

InternalServerError.

See also: AWS API Documentation

AmazonEMR.describe_security_configuration — Function.

describe_security_configuration(::AWSConfig, arguments::Dict)
describe_security_configuration([::AWSConfig]; Name=)

DescribeSecurityConfiguration Operation

Provides the details of a security configuration by returning the configuration JSON.

Arguments

Name = ::String – Required

The name of the security configuration.

Returns

DescribeSecurityConfigurationOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.describe_step — Function.

describe_step(::AWSConfig, arguments::Dict)
describe_step([::AWSConfig]; ClusterId=, StepId=)

DescribeStep Operation

Provides more detail about the cluster step.

Arguments

ClusterId = ::String – Required

The identifier of the cluster with steps to describe.

StepId = ::String – Required

The identifier of the step to describe.

Returns

DescribeStepOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.list_bootstrap_actions — Function.

list_bootstrap_actions(::AWSConfig, arguments::Dict)
list_bootstrap_actions([::AWSConfig]; ClusterId=, <keyword arguments>)

ListBootstrapActions Operation

Provides information about the bootstrap actions associated with a cluster.

Arguments

ClusterId = ::String – Required

The cluster identifier for the bootstrap actions to list.

Marker = ::String

The pagination token that indicates the next set of results to retrieve.

Returns

ListBootstrapActionsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.list_clusters — Function.

list_clusters(::AWSConfig, arguments::Dict)
list_clusters([::AWSConfig]; <keyword arguments>)

ListClusters Operation

Provides the status of all clusters visible to this AWS account. Allows you to filter the list of clusters based on certain criteria; for example, filtering by cluster creation date and time or by status. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListClusters calls.

Arguments

CreatedAfter = timestamp

The creation date and time beginning value filter for listing clusters.

CreatedBefore = timestamp

The creation date and time end value filter for listing clusters.

ClusterStates = ["STARTING", "BOOTSTRAPPING", "RUNNING", "WAITING", "TERMINATING", "TERMINATED" or "TERMINATED_WITH_ERRORS"...]

The cluster state filters to apply when listing clusters.

Marker = ::String

The pagination token that indicates the next set of results to retrieve.

Returns

ListClustersOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.list_instance_fleets — Function.

list_instance_fleets(::AWSConfig, arguments::Dict)
list_instance_fleets([::AWSConfig]; ClusterId=, <keyword arguments>)

ListInstanceFleets Operation

Lists all available details about the instance fleets in a cluster.

Note

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

Arguments

ClusterId = ::String – Required

The unique identifier of the cluster.

Marker = ::String

The pagination token that indicates the next set of results to retrieve.

Returns

ListInstanceFleetsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.list_instance_groups — Function.

list_instance_groups(::AWSConfig, arguments::Dict)
list_instance_groups([::AWSConfig]; ClusterId=, <keyword arguments>)

ListInstanceGroups Operation

Provides all available details about the instance groups in a cluster.

Arguments

ClusterId = ::String – Required

The identifier of the cluster for which to list the instance groups.

Marker = ::String

The pagination token that indicates the next set of results to retrieve.

Returns

ListInstanceGroupsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.list_instances — Function.

list_instances(::AWSConfig, arguments::Dict)
list_instances([::AWSConfig]; ClusterId=, <keyword arguments>)

ListInstances Operation

Provides information for all active EC2 instances and EC2 instances terminated in the last 30 days, up to a maximum of 2,000. EC2 instances in any of the following states are considered active: AWAITING_FULFILLMENT, PROVISIONING, BOOTSTRAPPING, RUNNING.

Arguments

ClusterId = ::String – Required

The identifier of the cluster for which to list the instances.

InstanceGroupId = ::String

The identifier of the instance group for which to list the instances.

InstanceGroupTypes = ["MASTER", "CORE" or "TASK"...]

The type of instance group for which to list the instances.

InstanceFleetId = ::String

The unique identifier of the instance fleet.

InstanceFleetType = "MASTER", "CORE" or "TASK"

The node type of the instance fleet. For example MASTER, CORE, or TASK.

InstanceStates = ["AWAITING_FULFILLMENT", "PROVISIONING", "BOOTSTRAPPING", "RUNNING" or "TERMINATED"...]

A list of instance states that will filter the instances returned with this request.

Marker = ::String

The pagination token that indicates the next set of results to retrieve.

Returns

ListInstancesOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.list_security_configurations — Function.

list_security_configurations(::AWSConfig, arguments::Dict)
list_security_configurations([::AWSConfig]; <keyword arguments>)

ListSecurityConfigurations Operation

Lists all the security configurations visible to this account, providing their creation dates and times, and their names. This call returns a maximum of 50 clusters per call, but returns a marker to track the paging of the cluster list across multiple ListSecurityConfigurations calls.

Arguments

Marker = ::String

The pagination token that indicates the set of results to retrieve.

Returns

ListSecurityConfigurationsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.list_steps — Function.

list_steps(::AWSConfig, arguments::Dict)
list_steps([::AWSConfig]; ClusterId=, <keyword arguments>)

ListSteps Operation

Provides a list of steps for the cluster in reverse order unless you specify stepIds with the request.

Arguments

ClusterId = ::String – Required

The identifier of the cluster for which to list the steps.

StepStates = ["PENDING", "CANCEL_PENDING", "RUNNING", "COMPLETED", "CANCELLED", "FAILED" or "INTERRUPTED"...]

The filter to limit the step list based on certain states.

StepIds = [::String...]

The filter to limit the step list based on the identifier of the steps.

Marker = ::String

The pagination token that indicates the next set of results to retrieve.

Returns

ListStepsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.modify_instance_fleet — Function.

modify_instance_fleet(::AWSConfig, arguments::Dict)
modify_instance_fleet([::AWSConfig]; ClusterId=, InstanceFleet=)

ModifyInstanceFleet Operation

Modifies the target On-Demand and target Spot capacities for the instance fleet with the specified InstanceFleetID within the cluster specified using ClusterID. The call either succeeds or fails atomically.

Note

The instance fleet configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions.

Arguments

ClusterId = ::String – Required

The unique identifier of the cluster.

InstanceFleet = Dict( ... ) – Required

The unique identifier of the instance fleet.

 InstanceFleet = Dict(
        "InstanceFleetId" => <required> ::String,
        "TargetOnDemandCapacity" =>  ::Int,
        "TargetSpotCapacity" =>  ::Int
    )

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.modify_instance_groups — Function.

modify_instance_groups(::AWSConfig, arguments::Dict)
modify_instance_groups([::AWSConfig]; <keyword arguments>)

ModifyInstanceGroups Operation

ModifyInstanceGroups modifies the number of nodes and configuration settings of an instance group. The input parameters include the new target instance count for the group and the instance group ID. The call will either succeed or fail atomically.

Arguments

ClusterId = ::String

The ID of the cluster to which the instance group belongs.

InstanceGroups = [Dict( ... )...]

Instance groups to change.

 InstanceGroups = [Dict(
        "InstanceGroupId" => <required> ::String,
        "InstanceCount" =>  ::Int,
        "EC2InstanceIdsToTerminate" =>  [::String...],
        "ShrinkPolicy" =>  Dict(
            "DecommissionTimeout" =>  ::Int,
            "InstanceResizePolicy" =>  Dict(
                "InstancesToTerminate" =>  [::String...],
                "InstancesToProtect" =>  [::String...],
                "InstanceTerminationTimeout" =>  ::Int
            )
        )
    )...]

Exceptions

InternalServerError.

See also: AWS API Documentation

AmazonEMR.put_auto_scaling_policy — Function.

put_auto_scaling_policy(::AWSConfig, arguments::Dict)
put_auto_scaling_policy([::AWSConfig]; ClusterId=, InstanceGroupId=, AutoScalingPolicy=)

PutAutoScalingPolicy Operation

Creates or updates an automatic scaling policy for a core instance group or task instance group in an Amazon EMR cluster. The automatic scaling policy defines how an instance group dynamically adds and terminates EC2 instances in response to the value of a CloudWatch metric.

Arguments

ClusterId = ::String – Required

Specifies the ID of a cluster. The instance group to which the automatic scaling policy is applied is within this cluster.

InstanceGroupId = ::String – Required

Specifies the ID of the instance group to which the automatic scaling policy is applied.

AutoScalingPolicy = Dict( ... ) – Required

Specifies the definition of the automatic scaling policy.

 AutoScalingPolicy = Dict(
        "Constraints" => <required> Dict(
            "MinCapacity" => <required> ::Int,
            "MaxCapacity" => <required> ::Int
        ),
        "Rules" => <required> [Dict(
            "Name" => <required> ::String,
            "Description" =>  ::String,
            "Action" => <required> Dict(
                "Market" =>  "ON_DEMAND" or "SPOT",
                "SimpleScalingPolicyConfiguration" => <required> Dict(
                    "AdjustmentType" =>  "CHANGE_IN_CAPACITY", "PERCENT_CHANGE_IN_CAPACITY" or "EXACT_CAPACITY",
                    "ScalingAdjustment" => <required> ::Int,
                    "CoolDown" =>  ::Int
                )
            ),
            "Trigger" => <required> Dict("CloudWatchAlarmDefinition" => <required> Dict(
                    "ComparisonOperator" => <required> "GREATER_THAN_OR_EQUAL", "GREATER_THAN", "LESS_THAN" or "LESS_THAN_OR_EQUAL",
                    "EvaluationPeriods" =>  ::Int,
                    "MetricName" => <required> ::String,
                    "Namespace" =>  ::String,
                    "Period" => <required> ::Int,
                    "Statistic" =>  "SAMPLE_COUNT", "AVERAGE", "SUM", "MINIMUM" or "MAXIMUM",
                    "Threshold" => <required> double,
                    "Unit" =>  "NONE", "SECONDS", "MICRO_SECONDS", "MILLI_SECONDS", "BYTES", "KILO_BYTES", "MEGA_BYTES", "GIGA_BYTES", "TERA_BYTES", "BITS", "KILO_BITS", "MEGA_BITS", "GIGA_BITS", "TERA_BITS", "PERCENT", "COUNT", "BYTES_PER_SECOND", "KILO_BYTES_PER_SECOND", "MEGA_BYTES_PER_SECOND", "GIGA_BYTES_PER_SECOND", "TERA_BYTES_PER_SECOND", "BITS_PER_SECOND", "KILO_BITS_PER_SECOND", "MEGA_BITS_PER_SECOND", "GIGA_BITS_PER_SECOND", "TERA_BITS_PER_SECOND" or "COUNT_PER_SECOND",
                    "Dimensions" =>  [Dict(
                        "Key" =>  ::String,
                        "Value" =>  ::String
                    )...]
                ))
        )...]
    )

Returns

PutAutoScalingPolicyOutput

See also: AWS API Documentation

AmazonEMR.remove_auto_scaling_policy — Function.

remove_auto_scaling_policy(::AWSConfig, arguments::Dict)
remove_auto_scaling_policy([::AWSConfig]; ClusterId=, InstanceGroupId=)

RemoveAutoScalingPolicy Operation

Removes an automatic scaling policy from a specified instance group within an EMR cluster.

Arguments

ClusterId = ::String – Required

Specifies the ID of a cluster. The instance group to which the automatic scaling policy is applied is within this cluster.

InstanceGroupId = ::String – Required

Specifies the ID of the instance group to which the scaling policy is applied.

Returns

RemoveAutoScalingPolicyOutput

See also: AWS API Documentation

AmazonEMR.remove_tags — Function.

remove_tags(::AWSConfig, arguments::Dict)
remove_tags([::AWSConfig]; ResourceId=, TagKeys=)

RemoveTags Operation

Removes tags from an Amazon EMR resource. Tags make it easier to associate clusters in various ways, such as grouping clusters to track your Amazon EMR resource allocation costs. For more information, see Tagging Amazon EMR Resources.

The following example removes the stack tag with value Prod from a cluster:

Arguments

ResourceId = ::String – Required

The Amazon EMR resource identifier from which tags will be removed. This value must be a cluster identifier.

TagKeys = [::String...] – Required

A list of tag keys to remove from a resource.

Returns

RemoveTagsOutput

Exceptions

InternalServerException or InvalidRequestException.

See also: AWS API Documentation

AmazonEMR.run_job_flow — Function.

run_job_flow(::AWSConfig, arguments::Dict)
run_job_flow([::AWSConfig]; Name=, Instances=, <keyword arguments>)

RunJobFlow Operation

RunJobFlow creates and starts running a new cluster (job flow). The cluster runs the steps specified. After the steps complete, the cluster stops and the HDFS partition is lost. To prevent loss of data, configure the last step of the job flow to store results in Amazon S3. If the JobFlowInstancesConfig KeepJobFlowAliveWhenNoSteps parameter is set to TRUE, the cluster transitions to the WAITING state rather than shutting down after the steps have completed.

For additional protection, you can set the JobFlowInstancesConfig TerminationProtected parameter to TRUE to lock the cluster and prevent it from being terminated by API call, user intervention, or in the event of a job flow error.

A maximum of 256 steps are allowed in each job flow.

If your cluster is long-running (such as a Hive data warehouse) or complex, you may require more than 256 steps to process your data. You can bypass the 256-step limitation in various ways, including using the SSH shell to connect to the master node and submitting queries directly to the software running on the master node, such as Hive and Hadoop. For more information on how to do this, see Add More than 256 Steps to a Cluster in the Amazon EMR Management Guide.

For long running clusters, we recommend that you periodically store your results.

Note

The instance fleets configuration is available only in Amazon EMR versions 4.8.0 and later, excluding 5.0.x versions. The RunJobFlow request can contain InstanceFleets parameters or InstanceGroups parameters, but not both.

Arguments

Name = ::String – Required

The name of the job flow.

LogUri = ::String

The location in Amazon S3 to write the log files of the job flow. If a value is not provided, logs are not created.

AdditionalInfo = ::String

A JSON string for selecting additional features.

AmiVersion = ::String

For Amazon EMR AMI versions 3.x and 2.x. For Amazon EMR releases 4.0 and later, the Linux AMI is determined by the ReleaseLabel specified or by CustomAmiID. The version of the Amazon Machine Image (AMI) to use when launching Amazon EC2 instances in the job flow. For details about the AMI versions currently supported in EMR version 3.x and 2.x, see AMI Versions Supported in EMR in the Amazon EMR Developer Guide.

If the AMI supports multiple versions of Hadoop (for example, AMI 1.0 supports both Hadoop 0.18 and 0.20), you can use the JobFlowInstancesConfig HadoopVersion parameter to modify the version of Hadoop from the defaults shown above.

Note

Previously, the EMR AMI version API parameter options allowed you to use latest for the latest AMI version rather than specify a numerical value. Some regions no longer support this deprecated option as they only have a newer release label version of EMR, which requires you to specify an EMR release label release (EMR 4.x or later).

ReleaseLabel = ::String

The release label for the Amazon EMR release. For Amazon EMR 3.x and 2.x AMIs, use AmiVersion instead.

Instances = Dict( ... ) – Required

A specification of the number and type of Amazon EC2 instances.

 Instances = Dict(
        "MasterInstanceType" =>  ::String,
        "SlaveInstanceType" =>  ::String,
        "InstanceCount" =>  ::Int,
        "InstanceGroups" =>  [Dict(
            "Name" =>  ::String,
            "Market" =>  "ON_DEMAND" or "SPOT",
            "InstanceRole" => <required> "MASTER", "CORE" or "TASK",
            "BidPrice" =>  ::String,
            "InstanceType" => <required> ::String,
            "InstanceCount" => <required> ::Int,
            "Configurations" =>  [Dict(
                "Classification" =>  ::String,
                "Configurations" =>  list,
                "Properties" =>  ::Dict{String,String}
            )...],
            "EbsConfiguration" =>  Dict(
                "EbsBlockDeviceConfigs" =>  [Dict(
                    "VolumeSpecification" => <required> Dict(
                        "VolumeType" => <required> ::String,
                        "Iops" =>  ::Int,
                        "SizeInGB" => <required> ::Int
                    ),
                    "VolumesPerInstance" =>  ::Int
                )...],
                "EbsOptimized" =>  ::Bool
            ),
            "AutoScalingPolicy" =>  Dict(
                "Constraints" => <required> Dict(
                    "MinCapacity" => <required> ::Int,
                    "MaxCapacity" => <required> ::Int
                ),
                "Rules" => <required> [Dict(
                    "Name" => <required> ::String,
                    "Description" =>  ::String,
                    "Action" => <required> Dict(
                        "Market" =>  "ON_DEMAND" or "SPOT",
                        "SimpleScalingPolicyConfiguration" => <required> Dict(
                            "AdjustmentType" =>  "CHANGE_IN_CAPACITY", "PERCENT_CHANGE_IN_CAPACITY" or "EXACT_CAPACITY",
                            "ScalingAdjustment" => <required> ::Int,
                            "CoolDown" =>  ::Int
                        )
                    ),
                    "Trigger" => <required> Dict("CloudWatchAlarmDefinition" => <required> Dict(
                            "ComparisonOperator" => <required> "GREATER_THAN_OR_EQUAL", "GREATER_THAN", "LESS_THAN" or "LESS_THAN_OR_EQUAL",
                            "EvaluationPeriods" =>  ::Int,
                            "MetricName" => <required> ::String,
                            "Namespace" =>  ::String,
                            "Period" => <required> ::Int,
                            "Statistic" =>  "SAMPLE_COUNT", "AVERAGE", "SUM", "MINIMUM" or "MAXIMUM",
                            "Threshold" => <required> double,
                            "Unit" =>  "NONE", "SECONDS", "MICRO_SECONDS", "MILLI_SECONDS", "BYTES", "KILO_BYTES", "MEGA_BYTES", "GIGA_BYTES", "TERA_BYTES", "BITS", "KILO_BITS", "MEGA_BITS", "GIGA_BITS", "TERA_BITS", "PERCENT", "COUNT", "BYTES_PER_SECOND", "KILO_BYTES_PER_SECOND", "MEGA_BYTES_PER_SECOND", "GIGA_BYTES_PER_SECOND", "TERA_BYTES_PER_SECOND", "BITS_PER_SECOND", "KILO_BITS_PER_SECOND", "MEGA_BITS_PER_SECOND", "GIGA_BITS_PER_SECOND", "TERA_BITS_PER_SECOND" or "COUNT_PER_SECOND",
                            "Dimensions" =>  [Dict(
                                "Key" =>  ::String,
                                "Value" =>  ::String
                            )...]
                        ))
                )...]
            )
        )...],
        "InstanceFleets" =>  [Dict(
            "Name" =>  ::String,
            "InstanceFleetType" => <required> "MASTER", "CORE" or "TASK",
            "TargetOnDemandCapacity" =>  ::Int,
            "TargetSpotCapacity" =>  ::Int,
            "InstanceTypeConfigs" =>  [Dict(
                "InstanceType" => <required> ::String,
                "WeightedCapacity" =>  ::Int,
                "BidPrice" =>  ::String,
                "BidPriceAsPercentageOfOnDemandPrice" =>  double,
                "EbsConfiguration" =>  Dict(
                    "EbsBlockDeviceConfigs" =>  [Dict(
                        "VolumeSpecification" => <required> Dict(
                            "VolumeType" => <required> ::String,
                            "Iops" =>  ::Int,
                            "SizeInGB" => <required> ::Int
                        ),
                        "VolumesPerInstance" =>  ::Int
                    )...],
                    "EbsOptimized" =>  ::Bool
                ),
                "Configurations" =>  [Dict(
                    "Classification" =>  ::String,
                    "Configurations" =>  list,
                    "Properties" =>  ::Dict{String,String}
                )...]
            )...],
            "LaunchSpecifications" =>  Dict("SpotSpecification" => <required> Dict(
                    "TimeoutDurationMinutes" => <required> ::Int,
                    "TimeoutAction" => <required> "SWITCH_TO_ON_DEMAND" or "TERMINATE_CLUSTER",
                    "BlockDurationMinutes" =>  ::Int
                ))
        )...],
        "Ec2KeyName" =>  ::String,
        "Placement" =>  Dict(
            "AvailabilityZone" =>  ::String,
            "AvailabilityZones" =>  [::String...]
        ),
        "KeepJobFlowAliveWhenNoSteps" =>  ::Bool,
        "TerminationProtected" =>  ::Bool,
        "HadoopVersion" =>  ::String,
        "Ec2SubnetId" =>  ::String,
        "Ec2SubnetIds" =>  [::String...],
        "EmrManagedMasterSecurityGroup" =>  ::String,
        "EmrManagedSlaveSecurityGroup" =>  ::String,
        "ServiceAccessSecurityGroup" =>  ::String,
        "AdditionalMasterSecurityGroups" =>  [::String...],
        "AdditionalSlaveSecurityGroups" =>  [::String...]
    )

Steps = [Dict( ... )...]

A list of steps to run.

 Steps = [Dict(
        "Name" => <required> ::String,
        "ActionOnFailure" =>  "TERMINATE_JOB_FLOW", "TERMINATE_CLUSTER", "CANCEL_AND_WAIT" or "CONTINUE",
        "HadoopJarStep" => <required> Dict(
            "Properties" =>  [Dict(
                "Key" =>  ::String,
                "Value" =>  ::String
            )...],
            "Jar" => <required> ::String,
            "MainClass" =>  ::String,
            "Args" =>  [::String...]
        )
    )...]

BootstrapActions = [Dict( ... )...]

A list of bootstrap actions to run before Hadoop starts on the cluster nodes.

 BootstrapActions = [Dict(
        "Name" => <required> ::String,
        "ScriptBootstrapAction" => <required> Dict(
            "Path" => <required> ::String,
            "Args" =>  [::String...]
        )
    )...]

SupportedProducts = [::String...]

Note

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use. For more information, see Use Third Party Applications with Amazon EMR. Currently supported values are:

"mapr-m3" - launch the job flow using MapR M3 Edition.

"mapr-m5" - launch the job flow using MapR M5 Edition.

NewSupportedProducts = [Dict( ... )...]

Note

For Amazon EMR releases 3.x and 2.x. For Amazon EMR releases 4.x and later, use Applications.

A list of strings that indicates third-party software to use with the job flow that accepts a user argument list. EMR accepts and forwards the argument list to the corresponding installation script as bootstrap action arguments. For more information, see "Launch a Job Flow on the MapR Distribution for Hadoop" in the Amazon EMR Developer Guide. Supported values are:

"mapr-m3" - launch the cluster using MapR M3 Edition.

"mapr-m5" - launch the cluster using MapR M5 Edition.

"mapr" with the user arguments specifying "–edition,m3" or "–edition,m5" - launch the job flow using MapR M3 or M5 Edition respectively.

"mapr-m7" - launch the cluster using MapR M7 Edition.

"hunk" - launch the cluster with the Hunk Big Data Analtics Platform.

"hue"- launch the cluster with Hue installed.

"spark" - launch the cluster with Apache Spark installed.

"ganglia" - launch the cluster with the Ganglia Monitoring System installed.

 NewSupportedProducts = [Dict(
        "Name" =>  ::String,
        "Args" =>  [::String...]
    )...]

Applications = [Dict( ... )...]

For Amazon EMR releases 4.0 and later. A list of applications for the cluster. Valid values are: "Hadoop", "Hive", "Mahout", "Pig", and "Spark." They are case insensitive.

 Applications = [Dict(
        "Name" =>  ::String,
        "Version" =>  ::String,
        "Args" =>  [::String...],
        "AdditionalInfo" =>  ::Dict{String,String}
    )...]

Configurations = [Dict( ... )...]

For Amazon EMR releases 4.0 and later. The list of configurations supplied for the EMR cluster you are creating.

 Configurations = [Dict(
        "Classification" =>  ::String,
        "Configurations" =>  list,
        "Properties" =>  ::Dict{String,String}
    )...]

VisibleToAllUsers = ::Bool

Whether the cluster is visible to all IAM users of the AWS account associated with the cluster. If this value is set to true, all IAM users of that AWS account can view and (if they have the proper policy permissions set) manage the cluster. If it is set to false, only the IAM user that created the cluster can view and manage it.

JobFlowRole = ::String

Also called instance profile and EC2 role. An IAM role for an EMR cluster. The EC2 instances of the cluster assume this role. The default role is EMR_EC2_DefaultRole. In order to use the default role, you must have already created it using the CLI or console.

ServiceRole = ::String

The IAM role that will be assumed by the Amazon EMR service to access AWS resources on your behalf.

Tags = [Dict( ... )...]

A list of tags to associate with a cluster and propagate to Amazon EC2 instances.

 Tags = [Dict(
        "Key" =>  ::String,
        "Value" =>  ::String
    )...]

SecurityConfiguration = ::String

The name of a security configuration to apply to the cluster.

AutoScalingRole = ::String

An IAM role for automatic scaling policies. The default role is EMR_AutoScaling_DefaultRole. The IAM role provides permissions that the automatic scaling feature requires to launch and terminate EC2 instances in an instance group.

ScaleDownBehavior = "TERMINATE_AT_INSTANCE_HOUR" or "TERMINATE_AT_TASK_COMPLETION"

Specifies the way that individual Amazon EC2 instances terminate when an automatic scale-in activity occurs or an instance group is resized. TERMINATE_AT_INSTANCE_HOUR indicates that Amazon EMR terminates nodes at the instance-hour boundary, regardless of when the request to terminate the instance was submitted. This option is only available with Amazon EMR 5.1.0 and later and is the default for clusters created using that version. TERMINATE_AT_TASK_COMPLETION indicates that Amazon EMR blacklists and drains tasks from nodes before terminating the Amazon EC2 instances, regardless of the instance-hour boundary. With either behavior, Amazon EMR removes the least active nodes first and blocks instance termination if it could lead to HDFS corruption. TERMINATE_AT_TASK_COMPLETION available only in Amazon EMR version 4.1.0 and later, and is the default for versions of Amazon EMR earlier than 5.1.0.

CustomAmiId = ::String

Available only in Amazon EMR version 5.7.0 and later. The ID of a custom Amazon EBS-backed Linux AMI. If specified, Amazon EMR uses this AMI when it launches cluster EC2 instances. For more information about custom AMIs in Amazon EMR, see Using a Custom AMI in the Amazon EMR Management Guide. If omitted, the cluster uses the base Linux AMI for the ReleaseLabel specified. For Amazon EMR versions 2.x and 3.x, use AmiVersion instead.

For information about creating a custom AMI, see Creating an Amazon EBS-Backed Linux AMI in the Amazon Elastic Compute Cloud User Guide for Linux Instances. For information about finding an AMI ID, see Finding a Linux AMI.

EbsRootVolumeSize = ::Int

The size, in GiB, of the EBS root device volume of the Linux AMI that is used for each EC2 instance. Available in Amazon EMR version 4.x and later.

RepoUpgradeOnBoot = "SECURITY" or "NONE"

Applies only when CustomAmiID is used. Specifies which updates from the Amazon Linux AMI package repositories to apply automatically when the instance boots using the AMI. If omitted, the default is SECURITY, which indicates that only security updates are applied. If NONE is specified, no updates are applied, and all updates must be applied manually.

Returns

RunJobFlowOutput

Exceptions

InternalServerError.

See also: AWS API Documentation

AmazonEMR.set_termination_protection — Function.

set_termination_protection(::AWSConfig, arguments::Dict)
set_termination_protection([::AWSConfig]; JobFlowIds=, TerminationProtected=)

SetTerminationProtection Operation

SetTerminationProtection locks a cluster (job flow) so the EC2 instances in the cluster cannot be terminated by user intervention, an API call, or in the event of a job-flow error. The cluster still terminates upon successful completion of the job flow. Calling SetTerminationProtection on a cluster is similar to calling the Amazon EC2 DisableAPITermination API on all EC2 instances in a cluster.

SetTerminationProtection is used to prevent accidental termination of a cluster and to ensure that in the event of an error, the instances persist so that you can recover any data stored in their ephemeral instance storage.

To terminate a cluster that has been locked by setting SetTerminationProtection to true, you must first unlock the job flow by a subsequent call to SetTerminationProtection in which you set the value to false.

For more information, seeManaging Cluster Termination in the Amazon EMR Management Guide.

Arguments

JobFlowIds = [::String...] – Required

A list of strings that uniquely identify the clusters to protect. This identifier is returned by RunJobFlow and can also be obtained from DescribeJobFlows .

TerminationProtected = ::Bool – Required

A Boolean that indicates whether to protect the cluster and prevent the Amazon EC2 instances in the cluster from shutting down due to API calls, user intervention, or job-flow error.

Exceptions

InternalServerError.

See also: AWS API Documentation

AmazonEMR.set_visible_to_all_users — Function.

set_visible_to_all_users(::AWSConfig, arguments::Dict)
set_visible_to_all_users([::AWSConfig]; JobFlowIds=, VisibleToAllUsers=)

SetVisibleToAllUsers Operation

Sets whether all AWS Identity and Access Management (IAM) users under your account can access the specified clusters (job flows). This action works on running clusters. You can also set the visibility of a cluster when you launch it using the VisibleToAllUsers parameter of RunJobFlow. The SetVisibleToAllUsers action can be called only by an IAM user who created the cluster or the AWS account that owns the cluster.

Arguments

JobFlowIds = [::String...] – Required

Identifiers of the job flows to receive the new visibility setting.

VisibleToAllUsers = ::Bool – Required

Whether the specified clusters are visible to all IAM users of the AWS account associated with the cluster. If this value is set to True, all IAM users of that AWS account can view and, if they have the proper IAM policy permissions set, manage the clusters. If it is set to False, only the IAM user that created a cluster can view and manage it.

Exceptions

InternalServerError.

See also: AWS API Documentation

AmazonEMR.terminate_job_flows — Function.

terminate_job_flows(::AWSConfig, arguments::Dict)
terminate_job_flows([::AWSConfig]; JobFlowIds=)

TerminateJobFlows Operation

TerminateJobFlows shuts a list of clusters (job flows) down. When a job flow is shut down, any step not yet completed is canceled and the EC2 instances on which the cluster is running are stopped. Any log files not already saved are uploaded to Amazon S3 if a LogUri was specified when the cluster was created.

The maximum number of clusters allowed is 10. The call to TerminateJobFlows is asynchronous. Depending on the configuration of the cluster, it may take up to 1-5 minutes for the cluster to completely terminate and release allocated resources, such as Amazon EC2 instances.

Arguments

JobFlowIds = [::String...] – Required

A list of job flows to be shutdown.

Exceptions

InternalServerError.

See also: AWS API Documentation