AmazonES.jl

AmazonES

Amazon Elasticsearch Configuration Service

Use the Amazon Elasticsearch configuration API to create, configure, and manage Elasticsearch domains.

The endpoint for configuration service requests is region-specific: es.region.amazonaws.com. For example, es.us-east-1.amazonaws.com. For a current list of supported regions and endpoints, see Regions and Endpoints.

This document is generated from apis/es-2015-01-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonES.add_tags

AmazonES.create_elasticsearch_domain

AmazonES.delete_elasticsearch_domain

AmazonES.describe_elasticsearch_domain

AmazonES.describe_elasticsearch_domain_config

AmazonES.describe_elasticsearch_domains

AmazonES.describe_elasticsearch_instance_type_limits

AmazonES.list_domain_names

AmazonES.list_elasticsearch_instance_types

AmazonES.list_elasticsearch_versions

AmazonES.list_tags

AmazonES.remove_tags

AmazonES.update_elasticsearch_domain_config

AmazonES.add_tags — Method.

add_tags(::AWSConfig, arguments::Dict)
add_tags(::AWSConfig; ARN=, TagList=)

AddTags Operation

Attaches tags to an existing Elasticsearch domain. Tags are a set of case-sensitive key value pairs. An Elasticsearch domain may have up to 10 tags. See Tagging Amazon Elasticsearch Service Domains for more information.

Arguments

ARN::String – Required

Specify the ARN for which you want to add the tags.

TagList: – Required

List of Tag that need to be added for the Elasticsearch domain.

    [Dict(
        Key * => ::String,
        Value * => ::String
    )...]

Exceptions

BaseException, LimitExceededException, ValidationException or InternalException.

See also: AWS API Documentation

AmazonES.create_elasticsearch_domain — Method.

create_elasticsearch_domain(::AWSConfig, arguments::Dict)
create_elasticsearch_domain(::AWSConfig; DomainName=, <keyword arguments>)

CreateElasticsearchDomain Operation

Creates a new Elasticsearch domain. For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

Arguments

DomainName::String – Required

The name of the Elasticsearch domain that you are creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a letter or number and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

ElasticsearchVersion::String

String of format X.Y to specify version for the Elasticsearch domain eg. "1.5" or "2.3". For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

ElasticsearchClusterConfig:

Configuration options for an Elasticsearch domain. Specifies the instance type and number of instances in the domain cluster.

    Dict(
        InstanceType  => : "m3.medium.elasticsearch", "m3.large.elasticsearch", "m3.xlarge.elasticsearch", "m3.2xlarge.elasticsearch", "m4.large.elasticsearch", "m4.xlarge.elasticsearch", "m4.2xlarge.elasticsearch", "m4.4xlarge.elasticsearch", "m4.10xlarge.elasticsearch", "t2.micro.elasticsearch", "t2.small.elasticsearch", "t2.medium.elasticsearch", "r3.large.elasticsearch", "r3.xlarge.elasticsearch", "r3.2xlarge.elasticsearch", "r3.4xlarge.elasticsearch", "r3.8xlarge.elasticsearch", "i2.xlarge.elasticsearch", "i2.2xlarge.elasticsearch", "d2.xlarge.elasticsearch", "d2.2xlarge.elasticsearch", "d2.4xlarge.elasticsearch", "d2.8xlarge.elasticsearch", "c4.large.elasticsearch", "c4.xlarge.elasticsearch", "c4.2xlarge.elasticsearch", "c4.4xlarge.elasticsearch", "c4.8xlarge.elasticsearch", "r4.large.elasticsearch", "r4.xlarge.elasticsearch", "r4.2xlarge.elasticsearch", "r4.4xlarge.elasticsearch", "r4.8xlarge.elasticsearch" or "r4.16xlarge.elasticsearch",
        InstanceCount  => ::Int,
        DedicatedMasterEnabled  => ::Bool,
        ZoneAwarenessEnabled  => ::Bool,
        DedicatedMasterType  => : "m3.medium.elasticsearch", "m3.large.elasticsearch", "m3.xlarge.elasticsearch", "m3.2xlarge.elasticsearch", "m4.large.elasticsearch", "m4.xlarge.elasticsearch", "m4.2xlarge.elasticsearch", "m4.4xlarge.elasticsearch", "m4.10xlarge.elasticsearch", "t2.micro.elasticsearch", "t2.small.elasticsearch", "t2.medium.elasticsearch", "r3.large.elasticsearch", "r3.xlarge.elasticsearch", "r3.2xlarge.elasticsearch", "r3.4xlarge.elasticsearch", "r3.8xlarge.elasticsearch", "i2.xlarge.elasticsearch", "i2.2xlarge.elasticsearch", "d2.xlarge.elasticsearch", "d2.2xlarge.elasticsearch", "d2.4xlarge.elasticsearch", "d2.8xlarge.elasticsearch", "c4.large.elasticsearch", "c4.xlarge.elasticsearch", "c4.2xlarge.elasticsearch", "c4.4xlarge.elasticsearch", "c4.8xlarge.elasticsearch", "r4.large.elasticsearch", "r4.xlarge.elasticsearch", "r4.2xlarge.elasticsearch", "r4.4xlarge.elasticsearch", "r4.8xlarge.elasticsearch" or "r4.16xlarge.elasticsearch",
        DedicatedMasterCount  => ::Int
    )

EBSOptions:

Options to enable, disable and specify the type and size of EBS storage volumes.

    Dict(
        EBSEnabled  => ::Bool,
        VolumeType  => : "standard", "gp2" or "io1",
        VolumeSize  => ::Int,
        Iops  => ::Int
    )

AccessPolicies::String

IAM access policy as a JSON-formatted string.

SnapshotOptions::Int

Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

AdvancedOptions::Dict{String,String}

Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

Returns

CreateElasticsearchDomainResponse

Exceptions

BaseException, DisabledOperationException, InternalException, InvalidTypeException, LimitExceededException, ResourceAlreadyExistsException or ValidationException.

See also: AWS API Documentation

AmazonES.delete_elasticsearch_domain — Method.

delete_elasticsearch_domain(::AWSConfig, arguments::Dict)
delete_elasticsearch_domain(::AWSConfig; DomainName=)

DeleteElasticsearchDomain Operation

Permanently deletes the specified Elasticsearch domain and all of its data. Once a domain is deleted, it cannot be recovered.

Arguments

DomainName::String – Required

The name of the Elasticsearch domain that you want to permanently delete.

Returns

DeleteElasticsearchDomainResponse

Exceptions

BaseException, InternalException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonES.describe_elasticsearch_domain — Method.

describe_elasticsearch_domain(::AWSConfig, arguments::Dict)
describe_elasticsearch_domain(::AWSConfig; DomainName=)

DescribeElasticsearchDomain Operation

Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.

Arguments

DomainName::String – Required

The name of the Elasticsearch domain for which you want information.

Returns

DescribeElasticsearchDomainResponse

Exceptions

BaseException, InternalException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonES.describe_elasticsearch_domain_config — Method.

describe_elasticsearch_domain_config(::AWSConfig, arguments::Dict)
describe_elasticsearch_domain_config(::AWSConfig; DomainName=)

DescribeElasticsearchDomainConfig Operation

Provides cluster configuration information about the specified Elasticsearch domain, such as the state, creation date, update version, and update date for cluster options.

Arguments

DomainName::String – Required

The Elasticsearch domain that you want to get information about.

Returns

DescribeElasticsearchDomainConfigResponse

Exceptions

BaseException, InternalException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonES.describe_elasticsearch_domains — Method.

describe_elasticsearch_domains(::AWSConfig, arguments::Dict)
describe_elasticsearch_domains(::AWSConfig; DomainNames=)

DescribeElasticsearchDomains Operation

Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.

Arguments

DomainNames: [::String...] – Required

The Elasticsearch domains for which you want information.

Returns

DescribeElasticsearchDomainsResponse

Exceptions

BaseException, InternalException or ValidationException.

See also: AWS API Documentation

AmazonES.describe_elasticsearch_instance_type_limits — Method.

describe_elasticsearch_instance_type_limits(::AWSConfig, arguments::Dict)
describe_elasticsearch_instance_type_limits(::AWSConfig; InstanceType=, ElasticsearchVersion=, <keyword arguments>)

DescribeElasticsearchInstanceTypeLimits Operation

Describe Elasticsearch Limits for a given InstanceType and ElasticsearchVersion. When modifying existing Domain, specify the [DomainName](@ref) to know what Limits are supported for modifying.

Arguments

DomainName::String

DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for Elasticsearch [Limits](@ref) for existing domain.

InstanceType: "m3.medium.elasticsearch", "m3.large.elasticsearch", "m3.xlarge.elasticsearch", "m3.2xlarge.elasticsearch", "m4.large.elasticsearch", "m4.xlarge.elasticsearch", "m4.2xlarge.elasticsearch", "m4.4xlarge.elasticsearch", "m4.10xlarge.elasticsearch", "t2.micro.elasticsearch", "t2.small.elasticsearch", "t2.medium.elasticsearch", "r3.large.elasticsearch", "r3.xlarge.elasticsearch", "r3.2xlarge.elasticsearch", "r3.4xlarge.elasticsearch", "r3.8xlarge.elasticsearch", "i2.xlarge.elasticsearch", "i2.2xlarge.elasticsearch", "d2.xlarge.elasticsearch", "d2.2xlarge.elasticsearch", "d2.4xlarge.elasticsearch", "d2.8xlarge.elasticsearch", "c4.large.elasticsearch", "c4.xlarge.elasticsearch", "c4.2xlarge.elasticsearch", "c4.4xlarge.elasticsearch", "c4.8xlarge.elasticsearch", "r4.large.elasticsearch", "r4.xlarge.elasticsearch", "r4.2xlarge.elasticsearch", "r4.4xlarge.elasticsearch", "r4.8xlarge.elasticsearch" or "r4.16xlarge.elasticsearch" – Required

The instance type for an Elasticsearch cluster for which Elasticsearch [Limits](@ref) are needed.

ElasticsearchVersion::String – Required

Version of Elasticsearch for which [Limits](@ref) are needed.

Returns

DescribeElasticsearchInstanceTypeLimitsResponse

Exceptions

BaseException, InternalException, InvalidTypeException, LimitExceededException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonES.list_domain_names — Method.

list_domain_names(::AWSConfig)

ListDomainNames Operation

Returns the name of all Elasticsearch domains owned by the current user's account.

Returns

ListDomainNamesResponse

Exceptions

BaseException or ValidationException.

See also: AWS API Documentation

AmazonES.list_elasticsearch_instance_types — Method.

list_elasticsearch_instance_types(::AWSConfig, arguments::Dict)
list_elasticsearch_instance_types(::AWSConfig; ElasticsearchVersion=, <keyword arguments>)

ListElasticsearchInstanceTypes Operation

List all Elasticsearch instance types that are supported for given ElasticsearchVersion

Arguments

ElasticsearchVersion::String – Required

Version of Elasticsearch for which list of supported elasticsearch instance types are needed.

DomainName::String

DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for list of available Elasticsearch instance types when modifying existing domain.

MaxResults::Int

Set this value to limit the number of results returned. Value provided must be greater than 30 else it wont be honored.

NextToken::String

NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

Returns

ListElasticsearchInstanceTypesResponse

Exceptions

BaseException, InternalException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonES.list_elasticsearch_versions — Method.

list_elasticsearch_versions(::AWSConfig, arguments::Dict)
list_elasticsearch_versions(::AWSConfig; <keyword arguments>)

ListElasticsearchVersions Operation

List all supported Elasticsearch versions

Arguments

MaxResults::Int

Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.

NextToken::String

Returns

ListElasticsearchVersionsResponse

Exceptions

BaseException, InternalException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonES.list_tags — Method.

list_tags(::AWSConfig, arguments::Dict)
list_tags(::AWSConfig; ARN=)

ListTags Operation

Returns all tags for the given Elasticsearch domain.

Arguments

ARN::String – Required

Specify the ARN for the Elasticsearch domain to which the tags are attached that you want to view.

Returns

ListTagsResponse

Exceptions

BaseException, ResourceNotFoundException, ValidationException or InternalException.

See also: AWS API Documentation

AmazonES.remove_tags — Method.

remove_tags(::AWSConfig, arguments::Dict)
remove_tags(::AWSConfig; ARN=, TagKeys=)

RemoveTags Operation

Removes the specified set of tags from the specified Elasticsearch domain.

Arguments

ARN::String – Required

Specifies the ARN for the Elasticsearch domain from which you want to delete the specified tags.

TagKeys: [::String...] – Required

Specifies the TagKey list which you want to remove from the Elasticsearch domain.

Exceptions

BaseException, ValidationException or InternalException.

See also: AWS API Documentation

AmazonES.update_elasticsearch_domain_config — Method.

update_elasticsearch_domain_config(::AWSConfig, arguments::Dict)
update_elasticsearch_domain_config(::AWSConfig; DomainName=, <keyword arguments>)

UpdateElasticsearchDomainConfig Operation

Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances.

Arguments

DomainName::String – Required

The name of the Elasticsearch domain that you are updating.

ElasticsearchClusterConfig:

The type and number of instances to instantiate for the domain cluster.

    Dict(
        InstanceType  => : "m3.medium.elasticsearch", "m3.large.elasticsearch", "m3.xlarge.elasticsearch", "m3.2xlarge.elasticsearch", "m4.large.elasticsearch", "m4.xlarge.elasticsearch", "m4.2xlarge.elasticsearch", "m4.4xlarge.elasticsearch", "m4.10xlarge.elasticsearch", "t2.micro.elasticsearch", "t2.small.elasticsearch", "t2.medium.elasticsearch", "r3.large.elasticsearch", "r3.xlarge.elasticsearch", "r3.2xlarge.elasticsearch", "r3.4xlarge.elasticsearch", "r3.8xlarge.elasticsearch", "i2.xlarge.elasticsearch", "i2.2xlarge.elasticsearch", "d2.xlarge.elasticsearch", "d2.2xlarge.elasticsearch", "d2.4xlarge.elasticsearch", "d2.8xlarge.elasticsearch", "c4.large.elasticsearch", "c4.xlarge.elasticsearch", "c4.2xlarge.elasticsearch", "c4.4xlarge.elasticsearch", "c4.8xlarge.elasticsearch", "r4.large.elasticsearch", "r4.xlarge.elasticsearch", "r4.2xlarge.elasticsearch", "r4.4xlarge.elasticsearch", "r4.8xlarge.elasticsearch" or "r4.16xlarge.elasticsearch",
        InstanceCount  => ::Int,
        DedicatedMasterEnabled  => ::Bool,
        ZoneAwarenessEnabled  => ::Bool,
        DedicatedMasterType  => : "m3.medium.elasticsearch", "m3.large.elasticsearch", "m3.xlarge.elasticsearch", "m3.2xlarge.elasticsearch", "m4.large.elasticsearch", "m4.xlarge.elasticsearch", "m4.2xlarge.elasticsearch", "m4.4xlarge.elasticsearch", "m4.10xlarge.elasticsearch", "t2.micro.elasticsearch", "t2.small.elasticsearch", "t2.medium.elasticsearch", "r3.large.elasticsearch", "r3.xlarge.elasticsearch", "r3.2xlarge.elasticsearch", "r3.4xlarge.elasticsearch", "r3.8xlarge.elasticsearch", "i2.xlarge.elasticsearch", "i2.2xlarge.elasticsearch", "d2.xlarge.elasticsearch", "d2.2xlarge.elasticsearch", "d2.4xlarge.elasticsearch", "d2.8xlarge.elasticsearch", "c4.large.elasticsearch", "c4.xlarge.elasticsearch", "c4.2xlarge.elasticsearch", "c4.4xlarge.elasticsearch", "c4.8xlarge.elasticsearch", "r4.large.elasticsearch", "r4.xlarge.elasticsearch", "r4.2xlarge.elasticsearch", "r4.4xlarge.elasticsearch", "r4.8xlarge.elasticsearch" or "r4.16xlarge.elasticsearch",
        DedicatedMasterCount  => ::Int
    )

EBSOptions:

Specify the type and size of the EBS volume that you want to use.

    Dict(
        EBSEnabled  => ::Bool,
        VolumeType  => : "standard", "gp2" or "io1",
        VolumeSize  => ::Int,
        Iops  => ::Int
    )

SnapshotOptions::Int

Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

AdvancedOptions::Dict{String,String}

Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

AccessPolicies::String

IAM access policy as a JSON-formatted string.

Returns

UpdateElasticsearchDomainConfigResponse

Exceptions

BaseException, InternalException, InvalidTypeException, LimitExceededException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation