AmazonOpsWorksCM.jl

AmazonOpsWorksCM

AWS OpsWorks for Chef Automate

AWS OpsWorks for Chef Automate is a service that runs and manages configuration management servers.

Glossary of terms

Server: A configuration management server that can be highly-available. The configuration manager runs on your instances by using various AWS services, such as Amazon Elastic Compute Cloud (EC2), and potentially Amazon Relational Database Service (RDS). A server is a generic abstraction over the configuration manager that you want to use, much like Amazon RDS. In AWS OpsWorks for Chef Automate, you do not start or stop servers. After you create servers, they continue to run until they are deleted.

Engine: The specific configuration manager that you want to use (such as Chef) is the engine.

Backup: This is an application-level backup of the data that the configuration manager stores. A backup creates a .tar.gz file that is stored in an Amazon Simple Storage Service (S3) bucket in your account. AWS OpsWorks for Chef Automate creates the S3 bucket when you launch the first instance. A backup maintains a snapshot of all of a server's important attributes at the time of the backup.

Events: Events are always related to a server. Events are written during server creation, when health checks run, when backups are created, etc. When you delete a server, the server's events are also deleted.

AccountAttributes: Every account has attributes that are assigned in the AWS OpsWorks for Chef Automate database. These attributes store information about configuration limits (servers, backups, etc.) and your customer account.

Endpoints

AWS OpsWorks for Chef Automate supports the following endpoints, all HTTPS. You must connect to one of the following endpoints. Chef servers can only be accessed or managed within the endpoint in which they are created.

opsworks-cm.us-east-1.amazonaws.com

opsworks-cm.us-west-2.amazonaws.com

opsworks-cm.eu-west-1.amazonaws.com

Throttling limits

All API operations allow for five requests per second with a burst of 10 requests per second.

This document is generated from apis/opsworkscm-2016-11-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonOpsWorksCM.associate_node

AmazonOpsWorksCM.create_backup

AmazonOpsWorksCM.create_server

AmazonOpsWorksCM.delete_backup

AmazonOpsWorksCM.delete_server

AmazonOpsWorksCM.describe_account_attributes

AmazonOpsWorksCM.describe_backups

AmazonOpsWorksCM.describe_events

AmazonOpsWorksCM.describe_node_association_status

AmazonOpsWorksCM.describe_servers

AmazonOpsWorksCM.disassociate_node

AmazonOpsWorksCM.restore_server

AmazonOpsWorksCM.start_maintenance

AmazonOpsWorksCM.update_server

AmazonOpsWorksCM.update_server_engine_attributes

AmazonOpsWorksCM.associate_node — Method.

associate_node(::AWSConfig, arguments::Dict)
associate_node(::AWSConfig; ServerName=, NodeName=, EngineAttributes=)

AssociateNode Operation

Associates a new node with the Chef server. This command is an alternative to knife bootstrap. For more information about how to disassociate a node, see DisassociateNode.

A node can can only be associated with servers that are in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid. The AssociateNode API call can be integrated into Auto Scaling configurations, AWS Cloudformation templates, or the user data of a server's instance.

Example: aws opsworks-cm associate-node --server-name *MyServer* --node-name *MyManagedNode* --engine-attributes "Name=*MyOrganization*,Value=default" "Name=*Chef_node_public_key*,Value=*Public_key_contents*"

Arguments

ServerName::String – Required

The name of the server with which to associate the node.

NodeName::String – Required

The name of the Chef client node.

EngineAttributes: – Required

Engine attributes used for associating the node.

Attributes accepted in a AssociateNode request:

CHEF_ORGANIZATION: The Chef organization with which the node is associated. By default only one organization named default can exist.

CHEF_NODE_PUBLIC_KEY: A PEM-formatted public key. This key is required for the chef-client agent to access the Chef API.

    [Dict(
        Name  => ::String,
        Value  => ::String
    )...]

Returns

AssociateNodeResponse

Exceptions

InvalidStateException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.create_backup — Method.

create_backup(::AWSConfig, arguments::Dict)
create_backup(::AWSConfig; ServerName=, <keyword arguments>)

CreateBackup Operation

Creates an application-level backup of a server. While the server is in the BACKING_UP state, the server cannot be changed, and no additional backup can be created.

Backups can be created for servers in RUNNING, HEALTHY, and UNHEALTHY states. By default, you can create a maximum of 50 manual backups.

This operation is asynchronous.

A LimitExceededException is thrown when the maximum number of manual backups is reached. An InvalidStateException is thrown when the server is not in any of the following states: RUNNING, HEALTHY, or UNHEALTHY. A ResourceNotFoundException is thrown when the server is not found. A ValidationException is thrown when parameters of the request are not valid.

Arguments

ServerName::String – Required

The name of the server that you want to back up.

Description::String

A user-defined description of the backup.

Returns

CreateBackupResponse

Exceptions

InvalidStateException, LimitExceededException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.create_server — Method.

create_server(::AWSConfig, arguments::Dict)
create_server(::AWSConfig; ServerName=, InstanceProfileArn=, InstanceType=, ServiceRoleArn=, <keyword arguments>)

CreateServer Operation

Creates and immedately starts a new server. The server is ready to use when it is in the HEALTHY state. By default, you can create a maximum of 10 servers.

This operation is asynchronous.

A LimitExceededException is thrown when you have created the maximum number of servers (10). A ResourceAlreadyExistsException is thrown when a server with the same name already exists in the account. A ResourceNotFoundException is thrown when you specify a backup ID that is not valid or is for a backup that does not exist. A ValidationException is thrown when parameters of the request are not valid.

If you do not specify a security group by adding the SecurityGroupIds parameter, AWS OpsWorks creates a new security group. The default security group opens the Chef server to the world on TCP port 443. If a KeyName is present, AWS OpsWorks enables SSH access. SSH is also open to the world on TCP port 22.

By default, the Chef Server is accessible from any IP address. We recommend that you update your security group rules to allow access from known IP addresses and address ranges only. To edit security group rules, open Security Groups in the navigation pane of the EC2 management console.

Arguments

AssociatePublicIpAddress::Bool

Associate a public IP address with a server that you are launching. Valid values are true or false. The default value is true.

DisableAutomatedBackup::Bool

Enable or disable scheduled backups. Valid values are true or false. The default value is true.

Engine::String

The configuration management engine to use. Valid values include Chef.

EngineModel::String

The engine model, or option. Valid values include Single.

EngineVersion::String

The major release version of the engine that you want to use. Values depend on the engine that you choose.

EngineAttributes:

Optional engine attributes on a specified server.

Attributes accepted in a createServer request:

CHEF_PIVOTAL_KEY: A base64-encoded RSA private key that is not stored by AWS OpsWorks for Chef. This private key is required to access the Chef API. When no CHEF_PIVOTAL_KEY is set, one is generated and returned in the response.

CHEF_DELIVERY_ADMIN_PASSWORD: The password for the administrative user in the Chef Automate GUI. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#%^&+=_). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEF_DELIVERY_ADMIN_PASSWORD is set, one is generated and returned in the response.

    [Dict(
        Name  => ::String,
        Value  => ::String
    )...]

BackupRetentionCount::Int

The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks for Chef Automate deletes the oldest backups if this number is exceeded. The default value is 1.

ServerName::String – Required

The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

InstanceProfileArn::String – Required

The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the instance profile you need.

InstanceType::String – Required

The Amazon EC2 instance type to use. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m4.large. Valid values are t2.medium, m4.large, or m4.2xlarge.

KeyPair::String

The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

PreferredMaintenanceWindow::String

The start time for a one-hour period each week during which AWS OpsWorks for Chef Automate performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM. The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

PreferredBackupWindow::String

The start time for a one-hour period during which AWS OpsWorks for Chef Automate backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats:

HH:MM for daily backups

DDD:HH:MM for weekly backups

The specified time is in coordinated universal time (UTC). The default value is a random, daily start time.

Example: 08:00, which represents a daily start time of 08:00 UTC.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

SecurityGroupIds: [::String...]

A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds.

If you do not specify this parameter, AWS OpsWorks for Chef Automate creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

ServiceRoleArn::String – Required

The service role that the AWS OpsWorks for Chef Automate service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-stuff/latest/service-role-creation.yaml. This template creates a CloudFormation stack that includes the service role that you need.

SubnetIds: [::String...]

The IDs of subnets in which to launch the server EC2 instance.

Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled.

EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled.

For more information about supported Amazon EC2 platforms, see Supported Platforms.

BackupId::String

If you specify this field, AWS OpsWorks for Chef Automate creates the server by using the backup represented by BackupId.

Returns

CreateServerResponse

Exceptions

LimitExceededException, ResourceAlreadyExistsException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.delete_backup — Method.

delete_backup(::AWSConfig, arguments::Dict)
delete_backup(::AWSConfig; BackupId=)

DeleteBackup Operation

Deletes a backup. You can delete both manual and automated backups. This operation is asynchronous.

An InvalidStateException is thrown when a backup deletion is already in progress. A ResourceNotFoundException is thrown when the backup does not exist. A ValidationException is thrown when parameters of the request are not valid.

Arguments

BackupId::String – Required

The ID of the backup to delete. Run the DescribeBackups command to get a list of backup IDs. Backup IDs are in the format ServerName-yyyyMMddHHmmssSSS.

Returns

DeleteBackupResponse

Exceptions

InvalidStateException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.delete_server — Method.

delete_server(::AWSConfig, arguments::Dict)
delete_server(::AWSConfig; ServerName=)

DeleteServer Operation

Deletes the server and the underlying AWS CloudFormation stack (including the server's EC2 instance). When you run this command, the server state is updated to DELETING. After the server is deleted, it is no longer returned by DescribeServer requests. If the AWS CloudFormation stack cannot be deleted, the server cannot be deleted.

This operation is asynchronous.

An InvalidStateException is thrown when a server deletion is already in progress. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

ServerName::String – Required

The ID of the server to delete.

Returns

DeleteServerResponse

Exceptions

InvalidStateException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.describe_account_attributes — Method.

describe_account_attributes(::AWSConfig, arguments::Dict)
describe_account_attributes(::AWSConfig; )

DescribeAccountAttributes Operation

Describes your account attributes, and creates requests to increase limits before they are reached or exceeded.

This operation is synchronous.

Arguments

Returns

DescribeAccountAttributesResponse

See also: AWS API Documentation

AmazonOpsWorksCM.describe_backups — Method.

describe_backups(::AWSConfig, arguments::Dict)
describe_backups(::AWSConfig; <keyword arguments>)

DescribeBackups Operation

Describes backups. The results are ordered by time, with newest backups first. If you do not specify a BackupId or ServerName, the command returns all backups.

This operation is synchronous.

A ResourceNotFoundException is thrown when the backup does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

BackupId::String

Describes a single backup.

ServerName::String

Returns backups for the server with the specified ServerName.

NextToken::String

NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeBackups again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null. Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

MaxResults::Int

To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Returns

DescribeBackupsResponse

Exceptions

ValidationException, ResourceNotFoundException or InvalidNextTokenException.

See also: AWS API Documentation

AmazonOpsWorksCM.describe_events — Method.

describe_events(::AWSConfig, arguments::Dict)
describe_events(::AWSConfig; ServerName=, <keyword arguments>)

DescribeEvents Operation

Describes events for a specified server. Results are ordered by time, with newest events first.

This operation is synchronous.

A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

ServerName::String – Required

The name of the server for which you want to view events.

NextToken::String

NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeEvents again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null. Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

MaxResults::Int

To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Returns

DescribeEventsResponse

Exceptions

ValidationException, InvalidNextTokenException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonOpsWorksCM.describe_node_association_status — Method.

describe_node_association_status(::AWSConfig, arguments::Dict)
describe_node_association_status(::AWSConfig; NodeAssociationStatusToken=, ServerName=)

DescribeNodeAssociationStatus Operation

Returns the current status of an existing association or disassociation request.

A ResourceNotFoundException is thrown when no recent association or disassociation request with the specified token is found, or when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

NodeAssociationStatusToken::String – Required

ServerName::String – Required

The name of the server from which to disassociate the node.

Returns

DescribeNodeAssociationStatusResponse

Exceptions

ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.describe_servers — Method.

describe_servers(::AWSConfig, arguments::Dict)
describe_servers(::AWSConfig; <keyword arguments>)

DescribeServers Operation

Lists all configuration management servers that are identified with your account. Only the stored results from Amazon DynamoDB are returned. AWS OpsWorks for Chef Automate does not query other services.

This operation is synchronous.

A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

ServerName::String

Describes the server with the specified ServerName.

NextToken::String

NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeServers again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null. Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

MaxResults::Int

To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

Returns

DescribeServersResponse

Exceptions

ValidationException, ResourceNotFoundException or InvalidNextTokenException.

See also: AWS API Documentation

AmazonOpsWorksCM.disassociate_node — Method.

disassociate_node(::AWSConfig, arguments::Dict)
disassociate_node(::AWSConfig; ServerName=, NodeName=, <keyword arguments>)

DisassociateNode Operation

Disassociates a node from a Chef server, and removes the node from the Chef server's managed nodes. After a node is disassociated, the node key pair is no longer valid for accessing the Chef API. For more information about how to associate a node, see AssociateNode.

A node can can only be disassociated from a server that is in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

ServerName::String – Required

The name of the server from which to disassociate the node.

NodeName::String – Required

The name of the Chef client node.

EngineAttributes:

Engine attributes used for disassociating the node.

Attributes accepted in a DisassociateNode request:

CHEF_ORGANIZATION: The Chef organization with which the node was associated. By default only one organization named default can exist.

    [Dict(
        Name  => ::String,
        Value  => ::String
    )...]

Returns

DisassociateNodeResponse

Exceptions

InvalidStateException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.restore_server — Method.

restore_server(::AWSConfig, arguments::Dict)
restore_server(::AWSConfig; BackupId=, ServerName=, <keyword arguments>)

RestoreServer Operation

Restores a backup to a server that is in a CONNECTION_LOST, HEALTHY, RUNNING, UNHEALTHY, or TERMINATED state. When you run RestoreServer, the server's EC2 instance is deleted, and a new EC2 instance is configured. RestoreServer maintains the existing server endpoint, so configuration management of the server's client devices (nodes) should continue to work.

This operation is asynchronous.

An InvalidStateException is thrown when the server is not in a valid state. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

BackupId::String – Required

The ID of the backup that you want to use to restore a server.

ServerName::String – Required

The name of the server that you want to restore.

InstanceType::String

The type of the instance to create. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m4.large. Valid values are t2.medium, m4.large, and m4.2xlarge. If you do not specify this parameter, RestoreServer uses the instance type from the specified backup.

KeyPair::String

The name of the key pair to set on the new EC2 instance. This can be helpful if the administrator no longer has the SSH key.

Returns

RestoreServerResponse

Exceptions

InvalidStateException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.start_maintenance — Method.

start_maintenance(::AWSConfig, arguments::Dict)
start_maintenance(::AWSConfig; ServerName=)

StartMaintenance Operation

Manually starts server maintenance. This command can be useful if an earlier maintenance attempt failed, and the underlying cause of maintenance failure has been resolved. The server is in an UNDER_MAINTENANCE state while maintenance is in progress.

Maintenance can only be started on servers in HEALTHY and UNHEALTHY states. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

ServerName::String – Required

The name of the server on which to run maintenance.

Returns

StartMaintenanceResponse

Exceptions

InvalidStateException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.update_server — Method.

update_server(::AWSConfig, arguments::Dict)
update_server(::AWSConfig; ServerName=, <keyword arguments>)

UpdateServer Operation

Updates settings for a server.

This operation is synchronous.

Arguments

DisableAutomatedBackup::Bool

Setting DisableAutomatedBackup to true disables automated or scheduled backups. Automated backups are enabled by default.

BackupRetentionCount::Int

Sets the number of automated backups that you want to keep.

ServerName::String – Required

The name of the server to update.

PreferredMaintenanceWindow::String

PreferredBackupWindow::String

Returns

UpdateServerResponse

Exceptions

InvalidStateException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation

AmazonOpsWorksCM.update_server_engine_attributes — Method.

update_server_engine_attributes(::AWSConfig, arguments::Dict)
update_server_engine_attributes(::AWSConfig; ServerName=, AttributeName=, <keyword arguments>)

UpdateServerEngineAttributes Operation

Updates engine-specific attributes on a specified server. The server enters the MODIFYING state when this operation is in progress. Only one update can occur at a time. You can use this command to reset the Chef server's private key (CHEF_PIVOTAL_KEY).

This operation is asynchronous.

This operation can only be called for servers in HEALTHY or UNHEALTHY states. Otherwise, an InvalidStateException is raised. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

ServerName::String – Required

The name of the server to update.

AttributeName::String – Required

The name of the engine attribute to update.

AttributeValue::String

The value to set for the attribute.

Returns

UpdateServerEngineAttributesResponse

Exceptions

InvalidStateException, ResourceNotFoundException or ValidationException.

See also: AWS API Documentation