AmazonRekognition.jl

AmazonRekognition

This is the Amazon Rekognition API reference.

This document is generated from apis/rekognition-2016-06-27.normal.json. See JuliaCloud/AWSCore.jl.

AmazonRekognition.compare_faces

AmazonRekognition.create_collection

AmazonRekognition.delete_collection

AmazonRekognition.delete_faces

AmazonRekognition.detect_faces

AmazonRekognition.detect_labels

AmazonRekognition.detect_moderation_labels

AmazonRekognition.get_celebrity_info

AmazonRekognition.index_faces

AmazonRekognition.list_collections

AmazonRekognition.list_faces

AmazonRekognition.recognize_celebrities

AmazonRekognition.search_faces

AmazonRekognition.search_faces_by_image

AmazonRekognition.compare_faces — Function.

compare_faces(::AWSConfig, arguments::Dict)
compare_faces([::AWSConfig]; SourceImage=, TargetImage=, <keyword arguments>)

CompareFaces Operation

Compares a face in the source input image with each face detected in the target input image.

Note

If the source image contains multiple faces, the service detects the largest face and compares it with each face detected in the target image.

In response, the operation returns an array of face matches ordered by similarity score in descending order. For each face match, the response provides a bounding box of the face, facial landmarks, pose details (pitch, role, and yaw), quality (brightness and sharpness), and confidence value (indicating the level of confidence that the bounding box contains a face). The response also provides a similarity score, which indicates how closely the faces match.

Note

By default, only faces with a similarity score of greater than or equal to 80% are returned in the response. You can change this value by specifying the SimilarityThreshold parameter.

CompareFaces also returns an array of faces that don't match the source image. For each face, it returns a bounding box, confidence value, landmarks, pose details, and quality. The response also returns information about the face in the source image, including the bounding box of the face and confidence value.

If the image doesn't contain Exif metadata, CompareFaces returns orientation information for the source and target images. Use these values to display the images with the correct image orientation.

Note

This is a stateless API operation. That is, data returned by this operation doesn't persist.

For an example, see get-started-exercise-compare-faces.

This operation requires permissions to perform the rekognition:CompareFaces action.

Arguments

SourceImage = Dict( ... ) – Required

The source image, either as bytes or as an S3 object.

 SourceImage = Dict(
        "Bytes" =>  blob,
        "S3Object" =>  Dict(
            "Bucket" =>  ::String,
            "Name" =>  ::String,
            "Version" =>  ::String
        )
    )

TargetImage = Dict( ... ) – Required

The target image, either as bytes or as an S3 object.

 TargetImage = Dict(
        "Bytes" =>  blob,
        "S3Object" =>  Dict(
            "Bucket" =>  ::String,
            "Name" =>  ::String,
            "Version" =>  ::String
        )
    )

SimilarityThreshold = float

The minimum level of confidence in the face matches that a match must meet to be included in the FaceMatches array.

Returns

CompareFacesResponse

Exceptions

InvalidParameterException, InvalidS3ObjectException, ImageTooLargeException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or InvalidImageFormatException.

See also: AWS API Documentation

AmazonRekognition.create_collection — Function.

create_collection(::AWSConfig, arguments::Dict)
create_collection([::AWSConfig]; CollectionId=)

CreateCollection Operation

Creates a collection in an AWS Region. You can add faces to the collection using the operation.

For example, you might create collections, one for each of your application users. A user can then index faces using the IndexFaces operation and persist results in a specific collection. Then, a user can search the collection for faces in the user-specific container.

Note

Collection names are case-sensitive.

For an example, see example1.

This operation requires permissions to perform the rekognition:CreateCollection action.

Arguments

CollectionId = ::String – Required

ID for the collection that you are creating.

Returns

CreateCollectionResponse

Exceptions

InvalidParameterException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or ResourceAlreadyExistsException.

See also: AWS API Documentation

AmazonRekognition.delete_collection — Function.

delete_collection(::AWSConfig, arguments::Dict)
delete_collection([::AWSConfig]; CollectionId=)

DeleteCollection Operation

Deletes the specified collection. Note that this operation removes all faces in the collection. For an example, see example1.

This operation requires permissions to perform the rekognition:DeleteCollection action.

Arguments

CollectionId = ::String – Required

ID of the collection to delete.

Returns

DeleteCollectionResponse

Exceptions

InvalidParameterException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonRekognition.delete_faces — Function.

delete_faces(::AWSConfig, arguments::Dict)
delete_faces([::AWSConfig]; CollectionId=, FaceIds=)

DeleteFaces Operation

Deletes faces from a collection. You specify a collection ID and an array of face IDs to remove from the collection.

This operation requires permissions to perform the rekognition:DeleteFaces action.

Arguments

CollectionId = ::String – Required

Collection from which to remove the specific faces.

FaceIds = [::String...] – Required

An array of face IDs to delete.

Returns

DeleteFacesResponse

Exceptions

InvalidParameterException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonRekognition.detect_faces — Function.

detect_faces(::AWSConfig, arguments::Dict)
detect_faces([::AWSConfig]; Image=, <keyword arguments>)

DetectFaces Operation

Detects faces within an image (JPEG or PNG) that is provided as input.

For each face detected, the operation returns face details including a bounding box of the face, a confidence value (that the bounding box contains a face), and a fixed set of attributes such as facial landmarks (for example, coordinates of eye and mouth), gender, presence of beard, sunglasses, etc.

The face-detection algorithm is most effective on frontal faces. For non-frontal or obscured faces, the algorithm may not detect the faces or might detect faces with lower confidence.

Note

This is a stateless API operation. That is, the operation does not persist any data.

For an example, see get-started-exercise-detect-faces.

This operation requires permissions to perform the rekognition:DetectFaces action.

Arguments

Image = Dict( ... ) – Required

The image in which you want to detect faces. You can specify a blob or an S3 object.

 Image = Dict(
        "Bytes" =>  blob,
        "S3Object" =>  Dict(
            "Bucket" =>  ::String,
            "Name" =>  ::String,
            "Version" =>  ::String
        )
    )

Attributes = ["DEFAULT" or "ALL"...]

An array of facial attributes you want to be returned. This can be the default list of attributes or all attributes. If you don't specify a value for Attributes or if you specify ["DEFAULT"], the API returns the following subset of facial attributes: BoundingBox, Confidence, Pose, Quality and Landmarks. If you provide ["ALL"], all facial attributes are returned but the operation will take longer to complete.

If you provide both, ["ALL", "DEFAULT"], the service uses a logical AND operator to determine which attributes to return (in this case, all attributes).

Returns

DetectFacesResponse

Exceptions

InvalidS3ObjectException, InvalidParameterException, ImageTooLargeException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or InvalidImageFormatException.

See also: AWS API Documentation

AmazonRekognition.detect_labels — Function.

detect_labels(::AWSConfig, arguments::Dict)
detect_labels([::AWSConfig]; Image=, <keyword arguments>)

DetectLabels Operation

Detects instances of real-world labels within an image (JPEG or PNG) provided as input. This includes objects like flower, tree, and table; events like wedding, graduation, and birthday party; and concepts like landscape, evening, and nature. For an example, see get-started-exercise-detect-labels.

For each object, scene, and concept the API returns one or more labels. Each label provides the object name, and the level of confidence that the image contains the object. For example, suppose the input image has a lighthouse, the sea, and a rock. The response will include all three labels, one for each object.

{Name: lighthouse, Confidence: 98.4629}

{Name: rock,Confidence: 79.2097}

{Name: sea,Confidence: 75.061}

In the preceding example, the operation returns one label for each of the three objects. The operation can also return multiple labels for the same object in the image. For example, if the input image shows a flower (for example, a tulip), the operation might return the following three labels.

{Name: flower,Confidence: 99.0562}

{Name: plant,Confidence: 99.0562}

{Name: tulip,Confidence: 99.0562}

In this example, the detection algorithm more precisely identifies the flower as a tulip.

You can provide the input image as an S3 object or as base64-encoded bytes. In response, the API returns an array of labels. In addition, the response also includes the orientation correction. Optionally, you can specify MinConfidence to control the confidence threshold for the labels returned. The default is 50%. You can also add the MaxLabels parameter to limit the number of labels returned.

Note

If the object detected is a person, the operation doesn't provide the same facial details that the DetectFaces operation provides.

This is a stateless API operation. That is, the operation does not persist any data.

This operation requires permissions to perform the rekognition:DetectLabels action.

Arguments

Image = Dict( ... ) – Required

The input image. You can provide a blob of image bytes or an S3 object.

 Image = Dict(
        "Bytes" =>  blob,
        "S3Object" =>  Dict(
            "Bucket" =>  ::String,
            "Name" =>  ::String,
            "Version" =>  ::String
        )
    )

MaxLabels = ::Int

Maximum number of labels you want the service to return in the response. The service returns the specified number of highest confidence labels.

MinConfidence = float

Specifies the minimum confidence level for the labels to return. Amazon Rekognition doesn't return any labels with confidence lower than this specified value.

If MinConfidence is not specified, the operation returns labels with a confidence values greater than or equal to 50 percent.

Returns

DetectLabelsResponse

Exceptions

InvalidS3ObjectException, InvalidParameterException, ImageTooLargeException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or InvalidImageFormatException.

See also: AWS API Documentation

AmazonRekognition.detect_moderation_labels — Function.

detect_moderation_labels(::AWSConfig, arguments::Dict)
detect_moderation_labels([::AWSConfig]; Image=, <keyword arguments>)

DetectModerationLabels Operation

Detects explicit or suggestive adult content in a specified JPEG or PNG format image. Use DetectModerationLabels to moderate images depending on your requirements. For example, you might want to filter images that contain nudity, but not images containing suggestive content.

To filter images, use the labels returned by DetectModerationLabels to determine which types of content are appropriate. For information about moderation labels, see image-moderation.

Arguments

Image = Dict( ... ) – Required

The input image as bytes or an S3 object.

 Image = Dict(
        "Bytes" =>  blob,
        "S3Object" =>  Dict(
            "Bucket" =>  ::String,
            "Name" =>  ::String,
            "Version" =>  ::String
        )
    )

MinConfidence = float

Specifies the minimum confidence level for the labels to return. Amazon Rekognition doesn't return any labels with a confidence level lower than this specified value.

If you don't specify MinConfidence, the operation returns labels with confidence values greater than or equal to 50 percent.

Returns

DetectModerationLabelsResponse

Exceptions

InvalidS3ObjectException, InvalidParameterException, ImageTooLargeException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or InvalidImageFormatException.

See also: AWS API Documentation

AmazonRekognition.get_celebrity_info — Function.

get_celebrity_info(::AWSConfig, arguments::Dict)
get_celebrity_info([::AWSConfig]; Id=)

GetCelebrityInfo Operation

Gets the name and additional information about a celebrity based on his or her Rekognition ID. The additional information is returned as an array of URLs. If there is no additional information about the celebrity, this list is empty. For more information, see celebrity-recognition.

This operation requires permissions to perform the rekognition:GetCelebrityInfo action.

Arguments

Id = ::String – Required

The ID for the celebrity. You get the celebrity ID from a call to the operation, which recognizes celebrities in an image.

Returns

GetCelebrityInfoResponse

Exceptions

InvalidParameterException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonRekognition.index_faces — Function.

index_faces(::AWSConfig, arguments::Dict)
index_faces([::AWSConfig]; CollectionId=, Image=, <keyword arguments>)

IndexFaces Operation

Detects faces in the input image and adds them to the specified collection.

Amazon Rekognition does not save the actual faces detected. Instead, the underlying detection algorithm first detects the faces in the input image, and for each face extracts facial features into a feature vector, and stores it in the back-end database. Amazon Rekognition uses feature vectors when performing face match and search operations using the and operations.

If you provide the optional externalImageID for the input image you provided, Amazon Rekognition associates this ID with all faces that it detects. When you call the operation, the response returns the external ID. You can use this external image ID to create a client-side index to associate the faces with each image. You can then use the index to find all faces in an image.

In response, the operation returns an array of metadata for all detected faces. This includes, the bounding box of the detected face, confidence value (indicating the bounding box contains a face), a face ID assigned by the service for each face that is detected and stored, and an image ID assigned by the service for the input image. If you request all facial attributes (using the detectionAttributes parameter, Amazon Rekognition returns detailed facial attributes such as facial landmarks (for example, location of eye and mount) and other facial attributes such gender. If you provide the same image, specify the same collection, and use the same external ID in the IndexFaces operation, Amazon Rekognition doesn't save duplicate face metadata.

For an example, see example2.

This operation requires permissions to perform the rekognition:IndexFaces action.

Arguments

CollectionId = ::String – Required

The ID of an existing collection to which you want to add the faces that are detected in the input images.

Image = Dict( ... ) – Required

The input image as bytes or an S3 object.

 Image = Dict(
        "Bytes" =>  blob,
        "S3Object" =>  Dict(
            "Bucket" =>  ::String,
            "Name" =>  ::String,
            "Version" =>  ::String
        )
    )

ExternalImageId = ::String

ID you want to assign to all the faces detected in the image.

DetectionAttributes = ["DEFAULT" or "ALL"...]

An array of facial attributes that you want to be returned. This can be the default list of attributes or all attributes. If you don't specify a value for Attributes or if you specify ["DEFAULT"], the API returns the following subset of facial attributes: BoundingBox, Confidence, Pose, Quality and Landmarks. If you provide ["ALL"], all facial attributes are returned but the operation will take longer to complete.

If you provide both, ["ALL", "DEFAULT"], the service uses a logical AND operator to determine which attributes to return (in this case, all attributes).

Returns

IndexFacesResponse

Exceptions

InvalidS3ObjectException, InvalidParameterException, ImageTooLargeException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException, ResourceNotFoundException or InvalidImageFormatException.

See also: AWS API Documentation

AmazonRekognition.list_collections — Function.

list_collections(::AWSConfig, arguments::Dict)
list_collections([::AWSConfig]; <keyword arguments>)

ListCollections Operation

Returns list of collection IDs in your account. If the result is truncated, the response also provides a NextToken that you can use in the subsequent request to fetch the next set of collection IDs.

For an example, see example1.

This operation requires permissions to perform the rekognition:ListCollections action.

Arguments

NextToken = ::String

Pagination token from the previous response.

MaxResults = ::Int

Maximum number of collection IDs to return.

Returns

ListCollectionsResponse

Exceptions

InvalidParameterException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException, InvalidPaginationTokenException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonRekognition.list_faces — Function.

list_faces(::AWSConfig, arguments::Dict)
list_faces([::AWSConfig]; CollectionId=, <keyword arguments>)

ListFaces Operation

Returns metadata for faces in the specified collection. This metadata includes information such as the bounding box coordinates, the confidence (that the bounding box contains a face), and face ID. For an example, see example3.

This operation requires permissions to perform the rekognition:ListFaces action.

Arguments

CollectionId = ::String – Required

ID of the collection from which to list the faces.

NextToken = ::String

If the previous response was incomplete (because there is more data to retrieve), Amazon Rekognition returns a pagination token in the response. You can use this pagination token to retrieve the next set of faces.

MaxResults = ::Int

Maximum number of faces to return.

Returns

ListFacesResponse

Exceptions

InvalidParameterException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException, InvalidPaginationTokenException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonRekognition.recognize_celebrities — Function.

recognize_celebrities(::AWSConfig, arguments::Dict)
recognize_celebrities([::AWSConfig]; Image=)

RecognizeCelebrities Operation

Returns an array of celebrities recognized in the input image. The image is passed either as base64-encoded image bytes or as a reference to an image in an Amazon S3 bucket. The image must be either a PNG or JPEG formatted file. For more information, see celebrity-recognition.

RecognizeCelebrities returns the 15 largest faces in the image. It lists recognized celebrities in the CelebrityFaces list and unrecognized faces in the UnrecognizedFaces list. The operation doesn't return celebrities whose face sizes are smaller than the largest 15 faces in the image.

For each celebrity recognized, the API returns a Celebrity object. The Celebrity object contains the celebrity name, ID, URL links to additional information, match confidence, and a ComparedFace object that you can use to locate the celebrity's face on the image.

Rekognition does not retain information about which images a celebrity has been recognized in. Your application must store this information and use the Celebrity ID property as a unique identifier for the celebrity. If you don't store the celebrity name or additional information URLs returned by RecognizeCelebrities, you will need the ID to identify the celebrity in a call to the operation.

For an example, see recognize-celebrities-tutorial.

This operation requires permissions to perform the rekognition:RecognizeCelebrities operation.

Arguments

Image = Dict( ... ) – Required

The input image to use for celebrity recognition.

 Image = Dict(
        "Bytes" =>  blob,
        "S3Object" =>  Dict(
            "Bucket" =>  ::String,
            "Name" =>  ::String,
            "Version" =>  ::String
        )
    )

Returns

RecognizeCelebritiesResponse

Exceptions

InvalidS3ObjectException, InvalidParameterException, InvalidImageFormatException, ImageTooLargeException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or InvalidImageFormatException.

See also: AWS API Documentation

AmazonRekognition.search_faces — Function.

search_faces(::AWSConfig, arguments::Dict)
search_faces([::AWSConfig]; CollectionId=, FaceId=, <keyword arguments>)

SearchFaces Operation

For a given input face ID, searches for matching faces in the collection the face belongs to. You get a face ID when you add a face to the collection using the IndexFaces operation. The operation compares the features of the input face with faces in the specified collection.

Note

You can also search faces without indexing faces by using the SearchFacesByImage operation.

The operation response returns an array of faces that match, ordered by similarity score with the highest similarity first. More specifically, it is an array of metadata for each face match that is found. Along with the metadata, the response also includes a confidence value for each face match, indicating the confidence that the specific face matches the input face.

For an example, see example3.

This operation requires permissions to perform the rekognition:SearchFaces action.

Arguments

CollectionId = ::String – Required

ID of the collection the face belongs to.

FaceId = ::String – Required

ID of a face to find matches for in the collection.

MaxFaces = ::Int

Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

FaceMatchThreshold = float

Optional value specifying the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%.

Returns

SearchFacesResponse

Exceptions

InvalidParameterException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonRekognition.search_faces_by_image — Function.

search_faces_by_image(::AWSConfig, arguments::Dict)
search_faces_by_image([::AWSConfig]; CollectionId=, Image=, <keyword arguments>)

SearchFacesByImage Operation

For a given input image, first detects the largest face in the image, and then searches the specified collection for matching faces. The operation compares the features of the input face with faces in the specified collection.

Note

To search for all faces in an input image, you might first call the operation, and then use the face IDs returned in subsequent calls to the operation.

You can also call the DetectFaces operation and use the bounding boxes in the response to make face crops, which then you can pass in to the SearchFacesByImage operation.

The response returns an array of faces that match, ordered by similarity score with the highest similarity first. More specifically, it is an array of metadata for each face match found. Along with the metadata, the response also includes a similarity indicating how similar the face is to the input face. In the response, the operation also returns the bounding box (and a confidence level that the bounding box contains a face) of the face that Amazon Rekognition used for the input image.

For an example, see example3.

This operation requires permissions to perform the rekognition:SearchFacesByImage action.

Arguments

CollectionId = ::String – Required

ID of the collection to search.

Image = Dict( ... ) – Required

The input image as bytes or an S3 object.

 Image = Dict(
        "Bytes" =>  blob,
        "S3Object" =>  Dict(
            "Bucket" =>  ::String,
            "Name" =>  ::String,
            "Version" =>  ::String
        )
    )

MaxFaces = ::Int

Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

FaceMatchThreshold = float

(Optional) Specifies the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%.

Returns

SearchFacesByImageResponse

Exceptions

InvalidS3ObjectException, InvalidParameterException, ImageTooLargeException, AccessDeniedException, InternalServerError, ThrottlingException, ProvisionedThroughputExceededException, ResourceNotFoundException or InvalidImageFormatException.

See also: AWS API Documentation