AmazonServiceCatalog.jl

AmazonServiceCatalog

AWS Service Catalog

Overview

AWS Service Catalog allows organizations to create and manage catalogs of IT services that are approved for use on AWS. This documentation provides reference material for the AWS Service Catalog end user API. To get the most out of this documentation, be familiar with the terminology discussed in AWS Service Catalog Concepts.

Additional Resources

AWS Service Catalog Administrator Guide

AWS Service Catalog User Guide

This document is generated from apis/servicecatalog-2015-12-10.normal.json. See JuliaCloud/AWSCore.jl.

AmazonServiceCatalog.accept_portfolio_share

AmazonServiceCatalog.associate_principal_with_portfolio

AmazonServiceCatalog.associate_product_with_portfolio

AmazonServiceCatalog.associate_tag_option_with_resource

AmazonServiceCatalog.create_constraint

AmazonServiceCatalog.create_portfolio

AmazonServiceCatalog.create_portfolio_share

AmazonServiceCatalog.create_product

AmazonServiceCatalog.create_provisioning_artifact

AmazonServiceCatalog.create_tag_option

AmazonServiceCatalog.delete_constraint

AmazonServiceCatalog.delete_portfolio

AmazonServiceCatalog.delete_portfolio_share

AmazonServiceCatalog.delete_product

AmazonServiceCatalog.delete_provisioning_artifact

AmazonServiceCatalog.describe_constraint

AmazonServiceCatalog.describe_portfolio

AmazonServiceCatalog.describe_product

AmazonServiceCatalog.describe_product_as_admin

AmazonServiceCatalog.describe_product_view

AmazonServiceCatalog.describe_provisioned_product

AmazonServiceCatalog.describe_provisioning_artifact

AmazonServiceCatalog.describe_provisioning_parameters

AmazonServiceCatalog.describe_record

AmazonServiceCatalog.describe_tag_option

AmazonServiceCatalog.disassociate_principal_from_portfolio

AmazonServiceCatalog.disassociate_product_from_portfolio

AmazonServiceCatalog.disassociate_tag_option_from_resource

AmazonServiceCatalog.list_accepted_portfolio_shares

AmazonServiceCatalog.list_constraints_for_portfolio

AmazonServiceCatalog.list_launch_paths

AmazonServiceCatalog.list_portfolio_access

AmazonServiceCatalog.list_portfolios

AmazonServiceCatalog.list_portfolios_for_product

AmazonServiceCatalog.list_principals_for_portfolio

AmazonServiceCatalog.list_provisioning_artifacts

AmazonServiceCatalog.list_record_history

AmazonServiceCatalog.list_resources_for_tag_option

AmazonServiceCatalog.list_tag_options

AmazonServiceCatalog.provision_product

AmazonServiceCatalog.reject_portfolio_share

AmazonServiceCatalog.scan_provisioned_products

AmazonServiceCatalog.search_products

AmazonServiceCatalog.search_products_as_admin

AmazonServiceCatalog.terminate_provisioned_product

AmazonServiceCatalog.update_constraint

AmazonServiceCatalog.update_portfolio

AmazonServiceCatalog.update_product

AmazonServiceCatalog.update_provisioned_product

AmazonServiceCatalog.update_provisioning_artifact

AmazonServiceCatalog.update_tag_option

AmazonServiceCatalog.accept_portfolio_share — Method.

accept_portfolio_share(::AWSConfig, arguments::Dict)
accept_portfolio_share(::AWSConfig; PortfolioId=, <keyword arguments>)

AcceptPortfolioShare Operation

Accepts an offer to share a portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

Returns

AcceptPortfolioShareOutput

Exceptions

InvalidParametersException, ResourceNotFoundException or LimitExceededException.

See also: AWS API Documentation

AmazonServiceCatalog.associate_principal_with_portfolio — Method.

associate_principal_with_portfolio(::AWSConfig, arguments::Dict)
associate_principal_with_portfolio(::AWSConfig; PortfolioId=, PrincipalARN=, PrincipalType=, <keyword arguments>)

AssociatePrincipalWithPortfolio Operation

Associates the specified principal ARN with the specified portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

PrincipalARN::String – Required

The ARN representing the principal (IAM user, role, or group).

PrincipalType: "IAM" – Required

The principal type. Must be IAM

Returns

AssociatePrincipalWithPortfolioOutput

Exceptions

InvalidParametersException, ResourceNotFoundException or LimitExceededException.

See also: AWS API Documentation

AmazonServiceCatalog.associate_product_with_portfolio — Method.

associate_product_with_portfolio(::AWSConfig, arguments::Dict)
associate_product_with_portfolio(::AWSConfig; ProductId=, PortfolioId=, <keyword arguments>)

AssociateProductWithPortfolio Operation

Associates a product with a portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

PortfolioId::String – Required

The portfolio identifier.

SourcePortfolioId::String

The identifier of the source portfolio to use with this association.

Returns

AssociateProductWithPortfolioOutput

Exceptions

InvalidParametersException, ResourceNotFoundException or LimitExceededException.

See also: AWS API Documentation

AmazonServiceCatalog.associate_tag_option_with_resource — Method.

associate_tag_option_with_resource(::AWSConfig, arguments::Dict)
associate_tag_option_with_resource(::AWSConfig; ResourceId=, TagOptionId=)

AssociateTagOptionWithResource Operation

Associate a TagOption identifier with a resource identifier.

Arguments

ResourceId::String – Required

The resource identifier.

TagOptionId::String – Required

The TagOption identifier.

Returns

AssociateTagOptionWithResourceOutput

Exceptions

TagOptionNotMigratedException, ResourceNotFoundException, InvalidParametersException, LimitExceededException, DuplicateResourceException or InvalidStateException.

See also: AWS API Documentation

AmazonServiceCatalog.create_constraint — Method.

create_constraint(::AWSConfig, arguments::Dict)
create_constraint(::AWSConfig; PortfolioId=, ProductId=, Parameters=, Type=, IdempotencyToken=, <keyword arguments>)

CreateConstraint Operation

Creates a new constraint. For more information, see Using Constraints.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

ProductId::String – Required

The product identifier.

Parameters::String – Required

The constraint parameters. Expected values vary depending on which Type is specified. For examples, see the bottom of this topic.

For Type LAUNCH, the RoleArn property is required.

For Type NOTIFICATION, the NotificationArns property is required.

For Type TEMPLATE, the Rules property is required.

Type::String – Required

The type of the constraint. Case-sensitive valid values are: LAUNCH, NOTIFICATION, or TEMPLATE.

Description::String

The text description of the constraint.

IdempotencyToken::String – Required

A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

Returns

CreateConstraintOutput

Exceptions

ResourceNotFoundException, InvalidParametersException, LimitExceededException or DuplicateResourceException.

See also: AWS API Documentation

AmazonServiceCatalog.create_portfolio — Method.

create_portfolio(::AWSConfig, arguments::Dict)
create_portfolio(::AWSConfig; DisplayName=, ProviderName=, IdempotencyToken=, <keyword arguments>)

CreatePortfolio Operation

Creates a new portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

DisplayName::String – Required

The name to use for display purposes.

Description::String

The text description of the portfolio.

ProviderName::String – Required

The name of the portfolio provider.

Tags:

Tags to associate with the new portfolio.

    [Dict(
        Key * => ::String,
        Value * => ::String
    )...]

IdempotencyToken::String – Required

A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

Returns

CreatePortfolioOutput

Exceptions

InvalidParametersException, LimitExceededException or TagOptionNotMigratedException.

See also: AWS API Documentation

AmazonServiceCatalog.create_portfolio_share — Method.

create_portfolio_share(::AWSConfig, arguments::Dict)
create_portfolio_share(::AWSConfig; PortfolioId=, AccountId=, <keyword arguments>)

CreatePortfolioShare Operation

Creates a new portfolio share.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

AccountId::String – Required

The account ID with which to share the portfolio.

Returns

CreatePortfolioShareOutput

Exceptions

ResourceNotFoundException, LimitExceededException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.create_product — Method.

create_product(::AWSConfig, arguments::Dict)
create_product(::AWSConfig; Name=, Owner=, ProductType=, ProvisioningArtifactParameters=, IdempotencyToken=, <keyword arguments>)

CreateProduct Operation

Creates a new product.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Name::String – Required

The name of the product.

Owner::String – Required

The owner of the product.

Description::String

The text description of the product.

Distributor::String

The distributor of the product.

SupportDescription::String

Support information about the product.

SupportEmail::String

Contact email for product support.

SupportUrl::String

Contact URL for product support.

ProductType: "CLOUD_FORMATION_TEMPLATE" or "MARKETPLACE" – Required

The type of the product to create.

Tags:

Tags to associate with the new product.

    [Dict(
        Key * => ::String,
        Value * => ::String
    )...]

ProvisioningArtifactParameters: – Required

Parameters for the provisioning artifact.

    Dict(
        Name  => ::String,
        Description  => ::String,
        Info * => ::Dict{String,String},
        Type  => : "CLOUD_FORMATION_TEMPLATE", "MARKETPLACE_AMI" or "MARKETPLACE_CAR"
    )

IdempotencyToken::String – Required

A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

Returns

CreateProductOutput

Exceptions

InvalidParametersException, LimitExceededException or TagOptionNotMigratedException.

See also: AWS API Documentation

AmazonServiceCatalog.create_provisioning_artifact — Method.

create_provisioning_artifact(::AWSConfig, arguments::Dict)
create_provisioning_artifact(::AWSConfig; ProductId=, Parameters=, IdempotencyToken=, <keyword arguments>)

CreateProvisioningArtifact Operation

Create a new provisioning artifact for the specified product. This operation does not work with a product that has been shared with you.

See the bottom of this topic for an example JSON request.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

Parameters: – Required

The parameters to use when creating the new provisioning artifact.

    Dict(
        Name  => ::String,
        Description  => ::String,
        Info * => ::Dict{String,String},
        Type  => : "CLOUD_FORMATION_TEMPLATE", "MARKETPLACE_AMI" or "MARKETPLACE_CAR"
    )

IdempotencyToken::String – Required

A token to disambiguate duplicate requests. You can create multiple resources using the same input in multiple requests, provided that you also specify a different idempotency token for each request.

Returns

CreateProvisioningArtifactOutput

Exceptions

ResourceNotFoundException, InvalidParametersException or LimitExceededException.

See also: AWS API Documentation

AmazonServiceCatalog.create_tag_option — Method.

create_tag_option(::AWSConfig, arguments::Dict)
create_tag_option(::AWSConfig; Key=, Value=)

CreateTagOption Operation

Create a new TagOption.

Arguments

Key::String – Required

The TagOption key.

Value::String – Required

The TagOption value.

Returns

CreateTagOptionOutput

Exceptions

TagOptionNotMigratedException, DuplicateResourceException or LimitExceededException.

See also: AWS API Documentation

AmazonServiceCatalog.delete_constraint — Method.

delete_constraint(::AWSConfig, arguments::Dict)
delete_constraint(::AWSConfig; Id=, <keyword arguments>)

DeleteConstraint Operation

Deletes the specified constraint.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the constraint to delete.

Returns

DeleteConstraintOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.delete_portfolio — Method.

delete_portfolio(::AWSConfig, arguments::Dict)
delete_portfolio(::AWSConfig; Id=, <keyword arguments>)

DeletePortfolio Operation

Deletes the specified portfolio. This operation does not work with a portfolio that has been shared with you or if it has products, users, constraints, or shared accounts associated with it.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the portfolio for the delete request.

Returns

DeletePortfolioOutput

Exceptions

ResourceNotFoundException, InvalidParametersException, ResourceInUseException or TagOptionNotMigratedException.

See also: AWS API Documentation

AmazonServiceCatalog.delete_portfolio_share — Method.

delete_portfolio_share(::AWSConfig, arguments::Dict)
delete_portfolio_share(::AWSConfig; PortfolioId=, AccountId=, <keyword arguments>)

DeletePortfolioShare Operation

Deletes the specified portfolio share.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

AccountId::String – Required

The account ID associated with the share to delete.

Returns

DeletePortfolioShareOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.delete_product — Method.

delete_product(::AWSConfig, arguments::Dict)
delete_product(::AWSConfig; Id=, <keyword arguments>)

DeleteProduct Operation

Deletes the specified product. This operation does not work with a product that has been shared with you or is associated with a portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the product for the delete request.

Returns

DeleteProductOutput

Exceptions

ResourceNotFoundException, ResourceInUseException, InvalidParametersException or TagOptionNotMigratedException.

See also: AWS API Documentation

AmazonServiceCatalog.delete_provisioning_artifact — Method.

delete_provisioning_artifact(::AWSConfig, arguments::Dict)
delete_provisioning_artifact(::AWSConfig; ProductId=, ProvisioningArtifactId=, <keyword arguments>)

DeleteProvisioningArtifact Operation

Deletes the specified provisioning artifact. This operation does not work on a provisioning artifact associated with a product that has been shared with you, or on the last provisioning artifact associated with a product (a product must have at least one provisioning artifact).

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

ProvisioningArtifactId::String – Required

The identifier of the provisioning artifact for the delete request. This is sometimes referred to as the product version.

Returns

DeleteProvisioningArtifactOutput

Exceptions

ResourceNotFoundException, ResourceInUseException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_constraint — Method.

describe_constraint(::AWSConfig, arguments::Dict)
describe_constraint(::AWSConfig; Id=, <keyword arguments>)

DescribeConstraint Operation

Retrieves detailed information for a specified constraint.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the constraint.

Returns

DescribeConstraintOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_portfolio — Method.

describe_portfolio(::AWSConfig, arguments::Dict)
describe_portfolio(::AWSConfig; Id=, <keyword arguments>)

DescribePortfolio Operation

Retrieves detailed information and any tags associated with the specified portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the portfolio for which to retrieve information.

Returns

DescribePortfolioOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_product — Method.

describe_product(::AWSConfig, arguments::Dict)
describe_product(::AWSConfig; Id=, <keyword arguments>)

DescribeProduct Operation

Retrieves information about a specified product.

This operation is functionally identical to DescribeProductView except that it takes as input ProductId instead of ProductViewId.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The ProductId of the product to describe.

Returns

DescribeProductOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_product_as_admin — Method.

describe_product_as_admin(::AWSConfig, arguments::Dict)
describe_product_as_admin(::AWSConfig; Id=, <keyword arguments>)

DescribeProductAsAdmin Operation

Retrieves information about a specified product, run with administrator access.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the product for which to retrieve information.

Returns

DescribeProductAsAdminOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_product_view — Method.

describe_product_view(::AWSConfig, arguments::Dict)
describe_product_view(::AWSConfig; Id=, <keyword arguments>)

DescribeProductView Operation

Retrieves information about a specified product.

This operation is functionally identical to DescribeProduct except that it takes as input ProductViewId instead of ProductId.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The ProductViewId of the product to describe.

Returns

DescribeProductViewOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_provisioned_product — Method.

describe_provisioned_product(::AWSConfig, arguments::Dict)
describe_provisioned_product(::AWSConfig; Id=, <keyword arguments>)

DescribeProvisionedProduct Operation

Retrieve detailed information about the provisioned product.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The provisioned product identifier.

Returns

DescribeProvisionedProductOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_provisioning_artifact — Method.

describe_provisioning_artifact(::AWSConfig, arguments::Dict)
describe_provisioning_artifact(::AWSConfig; ProvisioningArtifactId=, ProductId=, <keyword arguments>)

DescribeProvisioningArtifact Operation

Retrieves detailed information about the specified provisioning artifact.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProvisioningArtifactId::String – Required

The identifier of the provisioning artifact. This is sometimes referred to as the product version.

ProductId::String – Required

The product identifier.

Verbose::Bool

Enable a verbose level of details for the provisioning artifact.

Returns

DescribeProvisioningArtifactOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_provisioning_parameters — Method.

describe_provisioning_parameters(::AWSConfig, arguments::Dict)
describe_provisioning_parameters(::AWSConfig; ProductId=, ProvisioningArtifactId=, <keyword arguments>)

DescribeProvisioningParameters Operation

Provides information about parameters required to provision a specified product in a specified manner. Use this operation to obtain the list of ProvisioningArtifactParameters parameters available to call the ProvisionProduct operation for the specified product.

If the output contains a TagOption key with an empty list of values, there is a TagOption conflict for that key. The end user cannot take action to fix the conflict, and launch is not blocked. In subsequent calls to the ProvisionProduct operation, do not include conflicted TagOption keys as tags. Calls to ProvisionProduct with empty TagOption values cause the error "Parameter validation failed: Missing required parameter in Tags[N]:Value ". Calls to ProvisionProduct with conflicted TagOption keys automatically tag the provisioned product with the conflicted keys with the value "sc-tagoption-conflict-portfolioId-productId".

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

ProvisioningArtifactId::String – Required

The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

PathId::String

The identifier of the path for this product's provisioning. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

Returns

DescribeProvisioningParametersOutput

Exceptions

InvalidParametersException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_record — Method.

describe_record(::AWSConfig, arguments::Dict)
describe_record(::AWSConfig; Id=, <keyword arguments>)

DescribeRecord Operation

Retrieves a paginated list of the full details of a specific request. Use this operation after calling a request operation (ProvisionProduct, TerminateProvisionedProduct, or UpdateProvisionedProduct).

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The record identifier of the ProvisionedProduct object for which to retrieve output information. This is the RecordDetail.RecordId obtained from the request operation's response.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

Returns

DescribeRecordOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.describe_tag_option — Method.

describe_tag_option(::AWSConfig, arguments::Dict)
describe_tag_option(::AWSConfig; Id=)

DescribeTagOption Operation

Describes a TagOption.

Arguments

Id::String – Required

The identifier of the TagOption.

Returns

DescribeTagOptionOutput

Exceptions

TagOptionNotMigratedException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.disassociate_principal_from_portfolio — Method.

disassociate_principal_from_portfolio(::AWSConfig, arguments::Dict)
disassociate_principal_from_portfolio(::AWSConfig; PortfolioId=, PrincipalARN=, <keyword arguments>)

DisassociatePrincipalFromPortfolio Operation

Disassociates a previously associated principal ARN from a specified portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

PrincipalARN::String – Required

The ARN representing the principal (IAM user, role, or group).

Returns

DisassociatePrincipalFromPortfolioOutput

Exceptions

InvalidParametersException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.disassociate_product_from_portfolio — Method.

disassociate_product_from_portfolio(::AWSConfig, arguments::Dict)
disassociate_product_from_portfolio(::AWSConfig; ProductId=, PortfolioId=, <keyword arguments>)

DisassociateProductFromPortfolio Operation

Disassociates the specified product from the specified portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

PortfolioId::String – Required

The portfolio identifier.

Returns

DisassociateProductFromPortfolioOutput

Exceptions

ResourceNotFoundException, ResourceInUseException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.disassociate_tag_option_from_resource — Method.

disassociate_tag_option_from_resource(::AWSConfig, arguments::Dict)
disassociate_tag_option_from_resource(::AWSConfig; ResourceId=, TagOptionId=)

DisassociateTagOptionFromResource Operation

Disassociates a TagOption from a resource.

Arguments

ResourceId::String – Required

Identifier of the resource from which to disassociate the TagOption.

TagOptionId::String – Required

Identifier of the TagOption to disassociate from the resource.

Returns

DisassociateTagOptionFromResourceOutput

Exceptions

TagOptionNotMigratedException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.list_accepted_portfolio_shares — Method.

list_accepted_portfolio_shares(::AWSConfig, arguments::Dict)
list_accepted_portfolio_shares(::AWSConfig; <keyword arguments>)

ListAcceptedPortfolioShares Operation

Lists details of all portfolios for which sharing was accepted by this account.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

Returns

ListAcceptedPortfolioSharesOutput

Exceptions

InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.list_constraints_for_portfolio — Method.

list_constraints_for_portfolio(::AWSConfig, arguments::Dict)
list_constraints_for_portfolio(::AWSConfig; PortfolioId=, <keyword arguments>)

ListConstraintsForPortfolio Operation

Retrieves detailed constraint information for the specified portfolio and product.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

ProductId::String

The product identifier.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

Returns

ListConstraintsForPortfolioOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.list_launch_paths — Method.

list_launch_paths(::AWSConfig, arguments::Dict)
list_launch_paths(::AWSConfig; ProductId=, <keyword arguments>)

ListLaunchPaths Operation

Returns a paginated list of all paths to a specified product. A path is how the user has access to a specified product, and is necessary when provisioning a product. A path also determines the constraints put on the product.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier. Identifies the product for which to retrieve LaunchPathSummaries information.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

Returns

ListLaunchPathsOutput

Exceptions

InvalidParametersException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.list_portfolio_access — Method.

list_portfolio_access(::AWSConfig, arguments::Dict)
list_portfolio_access(::AWSConfig; PortfolioId=, <keyword arguments>)

ListPortfolioAccess Operation

Lists the account IDs that have been authorized sharing of the specified portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

Returns

ListPortfolioAccessOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.list_portfolios — Method.

list_portfolios(::AWSConfig, arguments::Dict)
list_portfolios(::AWSConfig; <keyword arguments>)

ListPortfolios Operation

Lists all portfolios in the catalog.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

Returns

ListPortfoliosOutput

Exceptions

InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.list_portfolios_for_product — Method.

list_portfolios_for_product(::AWSConfig, arguments::Dict)
list_portfolios_for_product(::AWSConfig; ProductId=, <keyword arguments>)

ListPortfoliosForProduct Operation

Lists all portfolios that the specified product is associated with.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

Returns

ListPortfoliosForProductOutput

Exceptions

InvalidParametersException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.list_principals_for_portfolio — Method.

list_principals_for_portfolio(::AWSConfig, arguments::Dict)
list_principals_for_portfolio(::AWSConfig; PortfolioId=, <keyword arguments>)

ListPrincipalsForPortfolio Operation

Lists all principal ARNs associated with the specified portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

Returns

ListPrincipalsForPortfolioOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.list_provisioning_artifacts — Method.

list_provisioning_artifacts(::AWSConfig, arguments::Dict)
list_provisioning_artifacts(::AWSConfig; ProductId=, <keyword arguments>)

ListProvisioningArtifacts Operation

Lists all provisioning artifacts associated with the specified product.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

Returns

ListProvisioningArtifactsOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.list_record_history — Method.

list_record_history(::AWSConfig, arguments::Dict)
list_record_history(::AWSConfig; <keyword arguments>)

ListRecordHistory Operation

Returns a paginated list of all performed requests, in the form of RecordDetails objects that are filtered as specified.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

AccessLevelFilter:

The access level for obtaining results. If left unspecified, User level access is used.

    Dict(
        Key  => : "Account", "Role" or "User",
        Value  => ::String
    )

SearchFilter:

The filter to limit search results.

    Dict(
        Key  => ::String,
        Value  => ::String
    )

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

Returns

ListRecordHistoryOutput

Exceptions

InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.list_resources_for_tag_option — Method.

list_resources_for_tag_option(::AWSConfig, arguments::Dict)
list_resources_for_tag_option(::AWSConfig; TagOptionId=, <keyword arguments>)

ListResourcesForTagOption Operation

Lists resources associated with a TagOption.

Arguments

TagOptionId::String – Required

Identifier of the TagOption.

ResourceType::String

Resource type.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

Returns

ListResourcesForTagOptionOutput

Exceptions

TagOptionNotMigratedException, ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.list_tag_options — Method.

list_tag_options(::AWSConfig, arguments::Dict)
list_tag_options(::AWSConfig; <keyword arguments>)

ListTagOptions Operation

Lists detailed TagOptions information.

Arguments

Filters:

The list of filters with which to limit search results. If no search filters are specified, the output is all TagOptions.

    Dict(
        Key  => ::String,
        Value  => ::String,
        Active  => ::Bool
    )

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

Returns

ListTagOptionsOutput

Exceptions

TagOptionNotMigratedException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.provision_product — Method.

provision_product(::AWSConfig, arguments::Dict)
provision_product(::AWSConfig; ProductId=, ProvisioningArtifactId=, ProvisionedProductName=, ProvisionToken=, <keyword arguments>)

ProvisionProduct Operation

Requests a provision of a specified product. A provisioned product is a resourced instance for a product. For example, provisioning a CloudFormation-template-backed product results in launching a CloudFormation stack and all the underlying resources that come with it.

You can check the status of this request using the DescribeRecord operation. The error "Parameter validation failed: Missing required parameter in Tags[N]:Value" indicates that your request contains a tag which has a tag key but no corresponding tag value (value is empty or null). Your call may have included values returned from a DescribeProvisioningParameters call that resulted in a TagOption key with an empty list. This happens when TagOption keys are in conflict. For more information, see DescribeProvisioningParameters.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

ProvisioningArtifactId::String – Required

The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

PathId::String

The identifier of the path for this product's provisioning. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

ProvisionedProductName::String – Required

A user-friendly name to identify the ProvisionedProduct object. This value must be unique for the AWS account and cannot be updated after the product is provisioned.

ProvisioningParameters:

Parameters specified by the administrator that are required for provisioning the product.

    [Dict(
        Key  => ::String,
        Value  => ::String
    )...]

Tags:

A list of tags to use as provisioning options.

    [Dict(
        Key * => ::String,
        Value * => ::String
    )...]

NotificationArns: [::String...]

Passed to CloudFormation. The SNS topic ARNs to which to publish stack-related events.

ProvisionToken::String – Required

An idempotency token that uniquely identifies the provisioning request.

Returns

ProvisionProductOutput

Exceptions

InvalidParametersException, ResourceNotFoundException or DuplicateResourceException.

See also: AWS API Documentation

AmazonServiceCatalog.reject_portfolio_share — Method.

reject_portfolio_share(::AWSConfig, arguments::Dict)
reject_portfolio_share(::AWSConfig; PortfolioId=, <keyword arguments>)

RejectPortfolioShare Operation

Rejects an offer to share a portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String – Required

The portfolio identifier.

Returns

RejectPortfolioShareOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.scan_provisioned_products — Method.

scan_provisioned_products(::AWSConfig, arguments::Dict)
scan_provisioned_products(::AWSConfig; <keyword arguments>)

ScanProvisionedProducts Operation

Returns a paginated list of all the ProvisionedProduct objects that are currently available (not terminated).

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

AccessLevelFilter:

The access level for obtaining results. If left unspecified, User level access is used.

    Dict(
        Key  => : "Account", "Role" or "User",
        Value  => ::String
    )

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

Returns

ScanProvisionedProductsOutput

Exceptions

InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.search_products — Method.

search_products(::AWSConfig, arguments::Dict)
search_products(::AWSConfig; <keyword arguments>)

SearchProducts Operation

Returns a paginated list all of the Products objects to which the caller has access.

The output of this operation can be used as input for other operations, such as DescribeProductView.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Filters::Dict{String,String}

The list of filters with which to limit search results. If no search filters are specified, the output is all the products to which the calling user has access.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

SortBy: "Title", "VersionCount" or "CreationDate"

The sort field specifier. If no value is specified, results are not sorted.

SortOrder: "ASCENDING" or "DESCENDING"

The sort order specifier. If no value is specified, results are not sorted.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

Returns

SearchProductsOutput

Exceptions

InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.search_products_as_admin — Method.

search_products_as_admin(::AWSConfig, arguments::Dict)
search_products_as_admin(::AWSConfig; <keyword arguments>)

SearchProductsAsAdmin Operation

Retrieves summary and status information about all products created within the caller's account. If a portfolio ID is provided, this operation retrieves information for only those products that are associated with the specified portfolio.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

PortfolioId::String

The portfolio identifier.

Filters::Dict{String,String}

The list of filters with which to limit search results. If no search filters are specified, the output is all the products to which the administrator has access.

SortBy: "Title", "VersionCount" or "CreationDate"

The sort field specifier. If no value is specified, results are not sorted.

SortOrder: "ASCENDING" or "DESCENDING"

The sort order specifier. If no value is specified, results are not sorted.

PageToken::String

The page token of the first page retrieved. If null, this retrieves the first page of size PageSize.

PageSize::Int

The maximum number of items to return in the results. If more results exist than fit in the specified PageSize, the value of NextPageToken in the response is non-null.

ProductSource: "ACCOUNT"

Access level of the source of the product.

Returns

SearchProductsAsAdminOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.terminate_provisioned_product — Method.

terminate_provisioned_product(::AWSConfig, arguments::Dict)
terminate_provisioned_product(::AWSConfig; TerminateToken=, <keyword arguments>)

TerminateProvisionedProduct Operation

Requests termination of an existing ProvisionedProduct object. If there are Tags associated with the object, they are terminated when the ProvisionedProduct object is terminated.

This operation does not delete any records associated with the ProvisionedProduct object.

You can check the status of this request using the DescribeRecord operation.

Arguments

ProvisionedProductName::String

The name of the ProvisionedProduct object to terminate. Specify either ProvisionedProductName or ProvisionedProductId, but not both.

ProvisionedProductId::String

The identifier of the ProvisionedProduct object to terminate. Specify either ProvisionedProductName or ProvisionedProductId, but not both.

TerminateToken::String – Required

An idempotency token that uniquely identifies the termination request. This token is only valid during the termination process. After the ProvisionedProduct object is terminated, further requests to terminate the same ProvisionedProduct object always return ResourceNotFound regardless of the value of TerminateToken.

IgnoreErrors::Bool

If set to true, AWS Service Catalog stops managing the specified ProvisionedProduct object even if it cannot delete the underlying resources.

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Returns

TerminateProvisionedProductOutput

Exceptions

ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.update_constraint — Method.

update_constraint(::AWSConfig, arguments::Dict)
update_constraint(::AWSConfig; Id=, <keyword arguments>)

UpdateConstraint Operation

Updates an existing constraint.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the constraint to update.

Description::String

The updated text description of the constraint.

Returns

UpdateConstraintOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.update_portfolio — Method.

update_portfolio(::AWSConfig, arguments::Dict)
update_portfolio(::AWSConfig; Id=, <keyword arguments>)

UpdatePortfolio Operation

Updates the specified portfolio's details. This operation does not work with a product that has been shared with you.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the portfolio for the update request.

DisplayName::String

The name to use for display purposes.

Description::String

The updated text description of the portfolio.

ProviderName::String

The updated name of the portfolio provider.

AddTags:

Tags to add to the existing list of tags associated with the portfolio.

    [Dict(
        Key * => ::String,
        Value * => ::String
    )...]

RemoveTags: [::String...]

Tags to remove from the existing list of tags associated with the portfolio.

Returns

UpdatePortfolioOutput

Exceptions

InvalidParametersException, ResourceNotFoundException, LimitExceededException or TagOptionNotMigratedException.

See also: AWS API Documentation

AmazonServiceCatalog.update_product — Method.

update_product(::AWSConfig, arguments::Dict)
update_product(::AWSConfig; Id=, <keyword arguments>)

UpdateProduct Operation

Updates an existing product.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

Id::String – Required

The identifier of the product for the update request.

Name::String

The updated product name.

Owner::String

The updated owner of the product.

Description::String

The updated text description of the product.

Distributor::String

The updated distributor of the product.

SupportDescription::String

The updated support description for the product.

SupportEmail::String

The updated support email for the product.

SupportUrl::String

The updated support URL for the product.

AddTags:

Tags to add to the existing list of tags associated with the product.

    [Dict(
        Key * => ::String,
        Value * => ::String
    )...]

RemoveTags: [::String...]

Tags to remove from the existing list of tags associated with the product.

Returns

UpdateProductOutput

Exceptions

ResourceNotFoundException, InvalidParametersException or TagOptionNotMigratedException.

See also: AWS API Documentation

AmazonServiceCatalog.update_provisioned_product — Method.

update_provisioned_product(::AWSConfig, arguments::Dict)
update_provisioned_product(::AWSConfig; UpdateToken=, <keyword arguments>)

UpdateProvisionedProduct Operation

Requests updates to the configuration of an existing ProvisionedProduct object. If there are tags associated with the object, they cannot be updated or added with this operation. Depending on the specific updates requested, this operation may update with no interruption, with some interruption, or replace the ProvisionedProduct object entirely.

You can check the status of this request using the DescribeRecord operation.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProvisionedProductName::String

The updated name of the ProvisionedProduct object. Specify either ProvisionedProductName or ProvisionedProductId, but not both.

ProvisionedProductId::String

The identifier of the ProvisionedProduct object to update. Specify either ProvisionedProductName or ProvisionedProductId, but not both.

ProductId::String

The identifier of the ProvisionedProduct object.

ProvisioningArtifactId::String

The provisioning artifact identifier for this product. This is sometimes referred to as the product version.

PathId::String

The identifier of the path to use in the updated ProvisionedProduct object. This value is optional if the product has a default path, and is required if there is more than one path for the specified product.

ProvisioningParameters:

A list of ProvisioningParameter objects used to update the ProvisionedProduct object.

    [Dict(
        Key  => ::String,
        Value  => ::String,
        UsePreviousValue  => ::Bool
    )...]

UpdateToken::String – Required

The idempotency token that uniquely identifies the provisioning update request.

Returns

UpdateProvisionedProductOutput

Exceptions

InvalidParametersException or ResourceNotFoundException.

See also: AWS API Documentation

AmazonServiceCatalog.update_provisioning_artifact — Method.

update_provisioning_artifact(::AWSConfig, arguments::Dict)
update_provisioning_artifact(::AWSConfig; ProductId=, ProvisioningArtifactId=, <keyword arguments>)

UpdateProvisioningArtifact Operation

Updates an existing provisioning artifact's information. This operation does not work on a provisioning artifact associated with a product that has been shared with you.

Arguments

AcceptLanguage::String

The language code to use for this operation. Supported language codes are as follows:

"en" (English)

"jp" (Japanese)

"zh" (Chinese)

If no code is specified, "en" is used as the default.

ProductId::String – Required

The product identifier.

ProvisioningArtifactId::String – Required

The identifier of the provisioning artifact for the update request. This is sometimes referred to as the product version.

Name::String

The updated name of the provisioning artifact.

Description::String

The updated text description of the provisioning artifact.

Returns

UpdateProvisioningArtifactOutput

Exceptions

ResourceNotFoundException or InvalidParametersException.

See also: AWS API Documentation

AmazonServiceCatalog.update_tag_option — Method.

update_tag_option(::AWSConfig, arguments::Dict)
update_tag_option(::AWSConfig; Id=, <keyword arguments>)

UpdateTagOption Operation

Updates an existing TagOption.

Arguments

Id::String – Required

The identifier of the constraint to update.

Value::String

The updated value.

Active::Bool

The updated active state.

Returns

UpdateTagOptionOutput

Exceptions

TagOptionNotMigratedException, ResourceNotFoundException, DuplicateResourceException or InvalidParametersException.

See also: AWS API Documentation