AmazonPinpoint.jl

AmazonPinpoint

This document is generated from apis/pinpoint-2016-12-01.normal.json. See JuliaCloud/AWSCore.jl.

AmazonPinpoint.create_campaign

AmazonPinpoint.create_import_job

AmazonPinpoint.create_segment

AmazonPinpoint.delete_apns_channel

AmazonPinpoint.delete_apns_sandbox_channel

AmazonPinpoint.delete_campaign

AmazonPinpoint.delete_email_channel

AmazonPinpoint.delete_event_stream

AmazonPinpoint.delete_gcm_channel

AmazonPinpoint.delete_segment

AmazonPinpoint.delete_sms_channel

AmazonPinpoint.get_apns_channel

AmazonPinpoint.get_apns_sandbox_channel

AmazonPinpoint.get_application_settings

AmazonPinpoint.get_campaign

AmazonPinpoint.get_campaign_activities

AmazonPinpoint.get_campaign_version

AmazonPinpoint.get_campaign_versions

AmazonPinpoint.get_campaigns

AmazonPinpoint.get_email_channel

AmazonPinpoint.get_endpoint

AmazonPinpoint.get_event_stream

AmazonPinpoint.get_gcm_channel

AmazonPinpoint.get_import_job

AmazonPinpoint.get_import_jobs

AmazonPinpoint.get_segment

AmazonPinpoint.get_segment_import_jobs

AmazonPinpoint.get_segment_version

AmazonPinpoint.get_segment_versions

AmazonPinpoint.get_segments

AmazonPinpoint.get_sms_channel

AmazonPinpoint.put_event_stream

AmazonPinpoint.send_messages

AmazonPinpoint.update_apns_channel

AmazonPinpoint.update_apns_sandbox_channel

AmazonPinpoint.update_application_settings

AmazonPinpoint.update_campaign

AmazonPinpoint.update_email_channel

AmazonPinpoint.update_endpoint

AmazonPinpoint.update_endpoints_batch

AmazonPinpoint.update_gcm_channel

AmazonPinpoint.update_segment

AmazonPinpoint.update_sms_channel

AmazonPinpoint.create_campaign — Function.

create_campaign(::AWSConfig, arguments::Dict)
create_campaign([::AWSConfig]; application-id=, WriteCampaignRequest=)

CreateCampaign Operation

Creates or updates a campaign.

Arguments

application-id = ::String – Required

WriteCampaignRequest = [ ... ] – Required

 WriteCampaignRequest = [
        "AdditionalTreatments" =>  [[
            "MessageConfiguration" =>  [
                "APNSMessage" =>  [
                    "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                    "Body" =>  ::String,
                    "ImageIconUrl" =>  ::String,
                    "ImageSmallIconUrl" =>  ::String,
                    "ImageUrl" =>  ::String,
                    "JsonBody" =>  ::String,
                    "MediaUrl" =>  ::String,
                    "SilentPush" =>  ::Bool,
                    "Title" =>  ::String,
                    "Url" =>  ::String
                ],
                "DefaultMessage" =>  [
                    "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                    "Body" =>  ::String,
                    "ImageIconUrl" =>  ::String,
                    "ImageSmallIconUrl" =>  ::String,
                    "ImageUrl" =>  ::String,
                    "JsonBody" =>  ::String,
                    "MediaUrl" =>  ::String,
                    "SilentPush" =>  ::Bool,
                    "Title" =>  ::String,
                    "Url" =>  ::String
                ],
                "EmailMessage" =>  [
                    "Body" =>  ::String,
                    "HtmlBody" =>  ::String,
                    "Title" =>  ::String
                ],
                "GCMMessage" =>  [
                    "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                    "Body" =>  ::String,
                    "ImageIconUrl" =>  ::String,
                    "ImageSmallIconUrl" =>  ::String,
                    "ImageUrl" =>  ::String,
                    "JsonBody" =>  ::String,
                    "MediaUrl" =>  ::String,
                    "SilentPush" =>  ::Bool,
                    "Title" =>  ::String,
                    "Url" =>  ::String
                ],
                "SMSMessage" =>  [
                    "Body" =>  ::String,
                    "MessageType" =>  "TRANSACTIONAL" or "PROMOTIONAL",
                    "SenderId" =>  ::String
                ]
            ],
            "Schedule" =>  [
                "EndTime" =>  ::String,
                "Frequency" =>  "ONCE", "HOURLY", "DAILY", "WEEKLY" or "MONTHLY",
                "IsLocalTime" =>  ::Bool,
                "QuietTime" =>  [
                    "End" =>  ::String,
                    "Start" =>  ::String
                ],
                "StartTime" =>  ::String,
                "Timezone" =>  ::String
            ],
            "SizePercent" =>  ::Int,
            "TreatmentDescription" =>  ::String,
            "TreatmentName" =>  ::String
        ], ...],
        "Description" =>  ::String,
        "HoldoutPercent" =>  ::Int,
        "IsPaused" =>  ::Bool,
        "Limits" =>  [
            "Daily" =>  ::Int,
            "Total" =>  ::Int
        ],
        "MessageConfiguration" =>  [
            "APNSMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Body" =>  ::String,
                "ImageIconUrl" =>  ::String,
                "ImageSmallIconUrl" =>  ::String,
                "ImageUrl" =>  ::String,
                "JsonBody" =>  ::String,
                "MediaUrl" =>  ::String,
                "SilentPush" =>  ::Bool,
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "DefaultMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Body" =>  ::String,
                "ImageIconUrl" =>  ::String,
                "ImageSmallIconUrl" =>  ::String,
                "ImageUrl" =>  ::String,
                "JsonBody" =>  ::String,
                "MediaUrl" =>  ::String,
                "SilentPush" =>  ::Bool,
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "EmailMessage" =>  [
                "Body" =>  ::String,
                "HtmlBody" =>  ::String,
                "Title" =>  ::String
            ],
            "GCMMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Body" =>  ::String,
                "ImageIconUrl" =>  ::String,
                "ImageSmallIconUrl" =>  ::String,
                "ImageUrl" =>  ::String,
                "JsonBody" =>  ::String,
                "MediaUrl" =>  ::String,
                "SilentPush" =>  ::Bool,
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "SMSMessage" =>  [
                "Body" =>  ::String,
                "MessageType" =>  "TRANSACTIONAL" or "PROMOTIONAL",
                "SenderId" =>  ::String
            ]
        ],
        "Name" =>  ::String,
        "Schedule" =>  [
            "EndTime" =>  ::String,
            "Frequency" =>  "ONCE", "HOURLY", "DAILY", "WEEKLY" or "MONTHLY",
            "IsLocalTime" =>  ::Bool,
            "QuietTime" =>  [
                "End" =>  ::String,
                "Start" =>  ::String
            ],
            "StartTime" =>  ::String,
            "Timezone" =>  ::String
        ],
        "SegmentId" =>  ::String,
        "SegmentVersion" =>  ::Int,
        "TreatmentDescription" =>  ::String,
        "TreatmentName" =>  ::String
    ]

Returns

CreateCampaignResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.create_import_job — Function.

create_import_job(::AWSConfig, arguments::Dict)
create_import_job([::AWSConfig]; application-id=, ImportJobRequest=)

CreateImportJob Operation

Creates or updates an import job.

Arguments

application-id = ::String – Required

ImportJobRequest = [ ... ] – Required

 ImportJobRequest = [
        "DefineSegment" =>  ::Bool,
        "ExternalId" =>  ::String,
        "Format" =>  "CSV" or "JSON",
        "RegisterEndpoints" =>  ::Bool,
        "RoleArn" =>  ::String,
        "S3Url" =>  ::String,
        "SegmentId" =>  ::String,
        "SegmentName" =>  ::String
    ]

Returns

CreateImportJobResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.create_segment — Function.

create_segment(::AWSConfig, arguments::Dict)
create_segment([::AWSConfig]; application-id=, WriteSegmentRequest=)

CreateSegment Operation

Used to create or update a segment.

Arguments

application-id = ::String – Required

WriteSegmentRequest = [ ... ] – Required

 WriteSegmentRequest = [
        "Dimensions" =>  [
            "Attributes" =>  ::Dict{String,String},
            "Behavior" =>  ["Recency" =>  [
                    "Duration" =>  "HR_24", "DAY_7", "DAY_14" or "DAY_30",
                    "RecencyType" =>  "ACTIVE" or "INACTIVE"
                ]],
            "Demographic" =>  [
                "AppVersion" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "Channel" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "DeviceType" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "Make" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "Model" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "Platform" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ]
            ],
            "Location" =>  ["Country" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ]],
            "UserAttributes" =>  ::Dict{String,String}
        ],
        "Name" =>  ::String
    ]

Returns

CreateSegmentResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_apns_channel — Function.

delete_apns_channel(::AWSConfig, arguments::Dict)
delete_apns_channel([::AWSConfig]; application-id=)

DeleteApnsChannel Operation

Deletes the APNs channel for an app.

Arguments

application-id = ::String – Required

Returns

DeleteApnsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_apns_sandbox_channel — Function.

delete_apns_sandbox_channel(::AWSConfig, arguments::Dict)
delete_apns_sandbox_channel([::AWSConfig]; application-id=)

DeleteApnsSandboxChannel Operation

Delete an APNS sandbox channel

Arguments

application-id = ::String – Required

Returns

DeleteApnsSandboxChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_campaign — Function.

delete_campaign(::AWSConfig, arguments::Dict)
delete_campaign([::AWSConfig]; application-id=, campaign-id=)

DeleteCampaign Operation

Deletes a campaign.

Arguments

application-id = ::String – Required

campaign-id = ::String – Required

Returns

DeleteCampaignResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_email_channel — Function.

delete_email_channel(::AWSConfig, arguments::Dict)
delete_email_channel([::AWSConfig]; application-id=)

DeleteEmailChannel Operation

Delete an email channel

Arguments

application-id = ::String – Required

Returns

DeleteEmailChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_event_stream — Function.

delete_event_stream(::AWSConfig, arguments::Dict)
delete_event_stream([::AWSConfig]; application-id=)

DeleteEventStream Operation

Deletes the event stream for an app.

Arguments

application-id = ::String – Required

ApplicationId

Returns

DeleteEventStreamResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_gcm_channel — Function.

delete_gcm_channel(::AWSConfig, arguments::Dict)
delete_gcm_channel([::AWSConfig]; application-id=)

DeleteGcmChannel Operation

Deletes the GCM channel for an app.

Arguments

application-id = ::String – Required

Returns

DeleteGcmChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_segment — Function.

delete_segment(::AWSConfig, arguments::Dict)
delete_segment([::AWSConfig]; application-id=, segment-id=)

DeleteSegment Operation

Deletes a segment.

Arguments

application-id = ::String – Required

segment-id = ::String – Required

Returns

DeleteSegmentResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.delete_sms_channel — Function.

delete_sms_channel(::AWSConfig, arguments::Dict)
delete_sms_channel([::AWSConfig]; application-id=)

DeleteSmsChannel Operation

Delete an SMS channel

Arguments

application-id = ::String – Required

Returns

DeleteSmsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_apns_channel — Function.

get_apns_channel(::AWSConfig, arguments::Dict)
get_apns_channel([::AWSConfig]; application-id=)

GetApnsChannel Operation

Returns information about the APNs channel for an app.

Arguments

application-id = ::String – Required

Returns

GetApnsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_apns_sandbox_channel — Function.

get_apns_sandbox_channel(::AWSConfig, arguments::Dict)
get_apns_sandbox_channel([::AWSConfig]; application-id=)

GetApnsSandboxChannel Operation

Get an APNS sandbox channel

Arguments

application-id = ::String – Required

Returns

GetApnsSandboxChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_application_settings — Function.

get_application_settings(::AWSConfig, arguments::Dict)
get_application_settings([::AWSConfig]; application-id=)

GetApplicationSettings Operation

Used to request the settings for an app.

Arguments

application-id = ::String – Required

Returns

GetApplicationSettingsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaign — Function.

get_campaign(::AWSConfig, arguments::Dict)
get_campaign([::AWSConfig]; application-id=, campaign-id=)

GetCampaign Operation

Returns information about a campaign.

Arguments

application-id = ::String – Required

campaign-id = ::String – Required

Returns

GetCampaignResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaign_activities — Function.

get_campaign_activities(::AWSConfig, arguments::Dict)
get_campaign_activities([::AWSConfig]; application-id=, campaign-id=, <keyword arguments>)

GetCampaignActivities Operation

Returns information about the activity performed by a campaign.

Arguments

application-id = ::String – Required

campaign-id = ::String – Required

page-size = ::String

token = ::String

Returns

GetCampaignActivitiesResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaign_version — Function.

get_campaign_version(::AWSConfig, arguments::Dict)
get_campaign_version([::AWSConfig]; application-id=, campaign-id=, version=)

GetCampaignVersion Operation

Returns information about a specific version of a campaign.

Arguments

application-id = ::String – Required

campaign-id = ::String – Required

version = ::String – Required

Returns

GetCampaignVersionResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaign_versions — Function.

get_campaign_versions(::AWSConfig, arguments::Dict)
get_campaign_versions([::AWSConfig]; application-id=, campaign-id=, <keyword arguments>)

GetCampaignVersions Operation

Returns information about your campaign versions.

Arguments

application-id = ::String – Required

campaign-id = ::String – Required

page-size = ::String

token = ::String

Returns

GetCampaignVersionsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_campaigns — Function.

get_campaigns(::AWSConfig, arguments::Dict)
get_campaigns([::AWSConfig]; application-id=, <keyword arguments>)

GetCampaigns Operation

Returns information about your campaigns.

Arguments

application-id = ::String – Required

page-size = ::String

token = ::String

Returns

GetCampaignsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_email_channel — Function.

get_email_channel(::AWSConfig, arguments::Dict)
get_email_channel([::AWSConfig]; application-id=)

GetEmailChannel Operation

Get an email channel

Arguments

application-id = ::String – Required

Returns

GetEmailChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_endpoint — Function.

get_endpoint(::AWSConfig, arguments::Dict)
get_endpoint([::AWSConfig]; application-id=, endpoint-id=)

GetEndpoint Operation

Returns information about an endpoint.

Arguments

application-id = ::String – Required

endpoint-id = ::String – Required

Returns

GetEndpointResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_event_stream — Function.

get_event_stream(::AWSConfig, arguments::Dict)
get_event_stream([::AWSConfig]; application-id=)

GetEventStream Operation

Returns the event stream for an app.

Arguments

application-id = ::String – Required

ApplicationId

Returns

GetEventStreamResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_gcm_channel — Function.

get_gcm_channel(::AWSConfig, arguments::Dict)
get_gcm_channel([::AWSConfig]; application-id=)

GetGcmChannel Operation

Returns information about the GCM channel for an app.

Arguments

application-id = ::String – Required

Returns

GetGcmChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_import_job — Function.

get_import_job(::AWSConfig, arguments::Dict)
get_import_job([::AWSConfig]; application-id=, job-id=)

GetImportJob Operation

Returns information about an import job.

Arguments

application-id = ::String – Required

job-id = ::String – Required

Returns

GetImportJobResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_import_jobs — Function.

get_import_jobs(::AWSConfig, arguments::Dict)
get_import_jobs([::AWSConfig]; application-id=, <keyword arguments>)

GetImportJobs Operation

Returns information about your import jobs.

Arguments

application-id = ::String – Required

page-size = ::String

token = ::String

Returns

GetImportJobsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segment — Function.

get_segment(::AWSConfig, arguments::Dict)
get_segment([::AWSConfig]; application-id=, segment-id=)

GetSegment Operation

Returns information about a segment.

Arguments

application-id = ::String – Required

segment-id = ::String – Required

Returns

GetSegmentResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segment_import_jobs — Function.

get_segment_import_jobs(::AWSConfig, arguments::Dict)
get_segment_import_jobs([::AWSConfig]; application-id=, segment-id=, <keyword arguments>)

GetSegmentImportJobs Operation

Returns a list of import jobs for a specific segment.

Arguments

application-id = ::String – Required

page-size = ::String

segment-id = ::String – Required

token = ::String

Returns

GetSegmentImportJobsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segment_version — Function.

get_segment_version(::AWSConfig, arguments::Dict)
get_segment_version([::AWSConfig]; application-id=, segment-id=, version=)

GetSegmentVersion Operation

Returns information about a segment version.

Arguments

application-id = ::String – Required

segment-id = ::String – Required

version = ::String – Required

Returns

GetSegmentVersionResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segment_versions — Function.

get_segment_versions(::AWSConfig, arguments::Dict)
get_segment_versions([::AWSConfig]; application-id=, segment-id=, <keyword arguments>)

GetSegmentVersions Operation

Returns information about your segment versions.

Arguments

application-id = ::String – Required

page-size = ::String

segment-id = ::String – Required

token = ::String

Returns

GetSegmentVersionsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_segments — Function.

get_segments(::AWSConfig, arguments::Dict)
get_segments([::AWSConfig]; application-id=, <keyword arguments>)

GetSegments Operation

Used to get information about your segments.

Arguments

application-id = ::String – Required

page-size = ::String

token = ::String

Returns

GetSegmentsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.get_sms_channel — Function.

get_sms_channel(::AWSConfig, arguments::Dict)
get_sms_channel([::AWSConfig]; application-id=)

GetSmsChannel Operation

Get an SMS channel

Arguments

application-id = ::String – Required

Returns

GetSmsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.put_event_stream — Function.

put_event_stream(::AWSConfig, arguments::Dict)
put_event_stream([::AWSConfig]; application-id=, WriteEventStream=)

PutEventStream Operation

Use to create or update the event stream for an app.

Arguments

application-id = ::String – Required

ApplicationId

WriteEventStream = [ ... ] – Required

EventStream to write.

 WriteEventStream = [
        "DestinationStreamArn" =>  ::String,
        "ExternalId" =>  ::String,
        "RoleArn" =>  ::String
    ]

Returns

PutEventStreamResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.send_messages — Function.

send_messages(::AWSConfig, arguments::Dict)
send_messages([::AWSConfig]; application-id=, MessageRequest=)

SendMessages Operation

Send a batch of messages

Arguments

application-id = ::String – Required

MessageRequest = [ ... ] – Required

 MessageRequest = [
        "Addresses" =>  ::Dict{String,String},
        "Context" =>  ::Dict{String,String},
        "MessageConfiguration" =>  [
            "APNSMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Badge" =>  ::Int,
                "Body" =>  ::String,
                "Category" =>  ::String,
                "Data" =>  ::Dict{String,String},
                "MediaUrl" =>  ::String,
                "RawContent" =>  ::String,
                "SilentPush" =>  ::Bool,
                "Sound" =>  ::String,
                "Substitutions" =>  ::Dict{String,String},
                "ThreadId" =>  ::String,
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "DefaultMessage" =>  [
                "Body" =>  ::String,
                "Substitutions" =>  ::Dict{String,String}
            ],
            "DefaultPushNotificationMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Body" =>  ::String,
                "Data" =>  ::Dict{String,String},
                "SilentPush" =>  ::Bool,
                "Substitutions" =>  ::Dict{String,String},
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "GCMMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Body" =>  ::String,
                "CollapseKey" =>  ::String,
                "Data" =>  ::Dict{String,String},
                "IconReference" =>  ::String,
                "ImageIconUrl" =>  ::String,
                "ImageUrl" =>  ::String,
                "RawContent" =>  ::String,
                "RestrictedPackageName" =>  ::String,
                "SilentPush" =>  ::Bool,
                "SmallImageIconUrl" =>  ::String,
                "Sound" =>  ::String,
                "Substitutions" =>  ::Dict{String,String},
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "SMSMessage" =>  [
                "Body" =>  ::String,
                "MessageType" =>  "TRANSACTIONAL" or "PROMOTIONAL",
                "SenderId" =>  ::String,
                "Substitutions" =>  ::Dict{String,String}
            ]
        ]
    ]

Returns

SendMessagesResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_apns_channel — Function.

update_apns_channel(::AWSConfig, arguments::Dict)
update_apns_channel([::AWSConfig]; APNSChannelRequest=, application-id=)

UpdateApnsChannel Operation

Use to update the APNs channel for an app.

Arguments

APNSChannelRequest = [ ... ] – Required

 APNSChannelRequest = [
        "Certificate" =>  ::String,
        "Enabled" =>  ::Bool,
        "PrivateKey" =>  ::String
    ]

application-id = ::String – Required

Returns

UpdateApnsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_apns_sandbox_channel — Function.

update_apns_sandbox_channel(::AWSConfig, arguments::Dict)
update_apns_sandbox_channel([::AWSConfig]; APNSSandboxChannelRequest=, application-id=)

UpdateApnsSandboxChannel Operation

Update an APNS sandbox channel

Arguments

APNSSandboxChannelRequest = [ ... ] – Required

 APNSSandboxChannelRequest = [
        "Certificate" =>  ::String,
        "Enabled" =>  ::Bool,
        "PrivateKey" =>  ::String
    ]

application-id = ::String – Required

Returns

UpdateApnsSandboxChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_application_settings — Function.

update_application_settings(::AWSConfig, arguments::Dict)
update_application_settings([::AWSConfig]; application-id=, WriteApplicationSettingsRequest=)

UpdateApplicationSettings Operation

Used to update the settings for an app.

Arguments

application-id = ::String – Required

WriteApplicationSettingsRequest = [ ... ] – Required

 WriteApplicationSettingsRequest = [
        "Limits" =>  [
            "Daily" =>  ::Int,
            "Total" =>  ::Int
        ],
        "QuietTime" =>  [
            "End" =>  ::String,
            "Start" =>  ::String
        ]
    ]

Returns

UpdateApplicationSettingsResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_campaign — Function.

update_campaign(::AWSConfig, arguments::Dict)
update_campaign([::AWSConfig]; application-id=, campaign-id=, WriteCampaignRequest=)

UpdateCampaign Operation

Use to update a campaign.

Arguments

application-id = ::String – Required

campaign-id = ::String – Required

WriteCampaignRequest = [ ... ] – Required

 WriteCampaignRequest = [
        "AdditionalTreatments" =>  [[
            "MessageConfiguration" =>  [
                "APNSMessage" =>  [
                    "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                    "Body" =>  ::String,
                    "ImageIconUrl" =>  ::String,
                    "ImageSmallIconUrl" =>  ::String,
                    "ImageUrl" =>  ::String,
                    "JsonBody" =>  ::String,
                    "MediaUrl" =>  ::String,
                    "SilentPush" =>  ::Bool,
                    "Title" =>  ::String,
                    "Url" =>  ::String
                ],
                "DefaultMessage" =>  [
                    "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                    "Body" =>  ::String,
                    "ImageIconUrl" =>  ::String,
                    "ImageSmallIconUrl" =>  ::String,
                    "ImageUrl" =>  ::String,
                    "JsonBody" =>  ::String,
                    "MediaUrl" =>  ::String,
                    "SilentPush" =>  ::Bool,
                    "Title" =>  ::String,
                    "Url" =>  ::String
                ],
                "EmailMessage" =>  [
                    "Body" =>  ::String,
                    "HtmlBody" =>  ::String,
                    "Title" =>  ::String
                ],
                "GCMMessage" =>  [
                    "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                    "Body" =>  ::String,
                    "ImageIconUrl" =>  ::String,
                    "ImageSmallIconUrl" =>  ::String,
                    "ImageUrl" =>  ::String,
                    "JsonBody" =>  ::String,
                    "MediaUrl" =>  ::String,
                    "SilentPush" =>  ::Bool,
                    "Title" =>  ::String,
                    "Url" =>  ::String
                ],
                "SMSMessage" =>  [
                    "Body" =>  ::String,
                    "MessageType" =>  "TRANSACTIONAL" or "PROMOTIONAL",
                    "SenderId" =>  ::String
                ]
            ],
            "Schedule" =>  [
                "EndTime" =>  ::String,
                "Frequency" =>  "ONCE", "HOURLY", "DAILY", "WEEKLY" or "MONTHLY",
                "IsLocalTime" =>  ::Bool,
                "QuietTime" =>  [
                    "End" =>  ::String,
                    "Start" =>  ::String
                ],
                "StartTime" =>  ::String,
                "Timezone" =>  ::String
            ],
            "SizePercent" =>  ::Int,
            "TreatmentDescription" =>  ::String,
            "TreatmentName" =>  ::String
        ], ...],
        "Description" =>  ::String,
        "HoldoutPercent" =>  ::Int,
        "IsPaused" =>  ::Bool,
        "Limits" =>  [
            "Daily" =>  ::Int,
            "Total" =>  ::Int
        ],
        "MessageConfiguration" =>  [
            "APNSMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Body" =>  ::String,
                "ImageIconUrl" =>  ::String,
                "ImageSmallIconUrl" =>  ::String,
                "ImageUrl" =>  ::String,
                "JsonBody" =>  ::String,
                "MediaUrl" =>  ::String,
                "SilentPush" =>  ::Bool,
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "DefaultMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Body" =>  ::String,
                "ImageIconUrl" =>  ::String,
                "ImageSmallIconUrl" =>  ::String,
                "ImageUrl" =>  ::String,
                "JsonBody" =>  ::String,
                "MediaUrl" =>  ::String,
                "SilentPush" =>  ::Bool,
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "EmailMessage" =>  [
                "Body" =>  ::String,
                "HtmlBody" =>  ::String,
                "Title" =>  ::String
            ],
            "GCMMessage" =>  [
                "Action" =>  "OPEN_APP", "DEEP_LINK" or "URL",
                "Body" =>  ::String,
                "ImageIconUrl" =>  ::String,
                "ImageSmallIconUrl" =>  ::String,
                "ImageUrl" =>  ::String,
                "JsonBody" =>  ::String,
                "MediaUrl" =>  ::String,
                "SilentPush" =>  ::Bool,
                "Title" =>  ::String,
                "Url" =>  ::String
            ],
            "SMSMessage" =>  [
                "Body" =>  ::String,
                "MessageType" =>  "TRANSACTIONAL" or "PROMOTIONAL",
                "SenderId" =>  ::String
            ]
        ],
        "Name" =>  ::String,
        "Schedule" =>  [
            "EndTime" =>  ::String,
            "Frequency" =>  "ONCE", "HOURLY", "DAILY", "WEEKLY" or "MONTHLY",
            "IsLocalTime" =>  ::Bool,
            "QuietTime" =>  [
                "End" =>  ::String,
                "Start" =>  ::String
            ],
            "StartTime" =>  ::String,
            "Timezone" =>  ::String
        ],
        "SegmentId" =>  ::String,
        "SegmentVersion" =>  ::Int,
        "TreatmentDescription" =>  ::String,
        "TreatmentName" =>  ::String
    ]

Returns

UpdateCampaignResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_email_channel — Function.

update_email_channel(::AWSConfig, arguments::Dict)
update_email_channel([::AWSConfig]; application-id=, EmailChannelRequest=)

UpdateEmailChannel Operation

Update an email channel

Arguments

application-id = ::String – Required

EmailChannelRequest = [ ... ] – Required

 EmailChannelRequest = [
        "Enabled" =>  ::Bool,
        "FromAddress" =>  ::String,
        "Identity" =>  ::String,
        "RoleArn" =>  ::String
    ]

Returns

UpdateEmailChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_endpoint — Function.

update_endpoint(::AWSConfig, arguments::Dict)
update_endpoint([::AWSConfig]; application-id=, endpoint-id=, EndpointRequest=)

UpdateEndpoint Operation

Use to update an endpoint.

Arguments

application-id = ::String – Required

endpoint-id = ::String – Required

EndpointRequest = [ ... ] – Required

 EndpointRequest = [
        "Address" =>  ::String,
        "Attributes" =>  ::Dict{String,String},
        "ChannelType" =>  "GCM", "APNS", "APNS_SANDBOX", "ADM", "SMS" or "EMAIL",
        "Demographic" =>  [
            "AppVersion" =>  ::String,
            "Locale" =>  ::String,
            "Make" =>  ::String,
            "Model" =>  ::String,
            "ModelVersion" =>  ::String,
            "Platform" =>  ::String,
            "PlatformVersion" =>  ::String,
            "Timezone" =>  ::String
        ],
        "EffectiveDate" =>  ::String,
        "EndpointStatus" =>  ::String,
        "Location" =>  [
            "City" =>  ::String,
            "Country" =>  ::String,
            "Latitude" =>  double,
            "Longitude" =>  double,
            "PostalCode" =>  ::String,
            "Region" =>  ::String
        ],
        "Metrics" =>  ::Dict{String,String},
        "OptOut" =>  ::String,
        "RequestId" =>  ::String,
        "User" =>  [
            "UserAttributes" =>  ::Dict{String,String},
            "UserId" =>  ::String
        ]
    ]

Returns

UpdateEndpointResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_endpoints_batch — Function.

update_endpoints_batch(::AWSConfig, arguments::Dict)
update_endpoints_batch([::AWSConfig]; application-id=, EndpointBatchRequest=)

UpdateEndpointsBatch Operation

Use to update a batch of endpoints.

Arguments

application-id = ::String – Required

EndpointBatchRequest = ["Item" => [[ ... ], ...]] – Required

 EndpointBatchRequest = ["Item" =>  [[
            "Address" =>  ::String,
            "Attributes" =>  ::Dict{String,String},
            "ChannelType" =>  "GCM", "APNS", "APNS_SANDBOX", "ADM", "SMS" or "EMAIL",
            "Demographic" =>  [
                "AppVersion" =>  ::String,
                "Locale" =>  ::String,
                "Make" =>  ::String,
                "Model" =>  ::String,
                "ModelVersion" =>  ::String,
                "Platform" =>  ::String,
                "PlatformVersion" =>  ::String,
                "Timezone" =>  ::String
            ],
            "EffectiveDate" =>  ::String,
            "EndpointStatus" =>  ::String,
            "Id" =>  ::String,
            "Location" =>  [
                "City" =>  ::String,
                "Country" =>  ::String,
                "Latitude" =>  double,
                "Longitude" =>  double,
                "PostalCode" =>  ::String,
                "Region" =>  ::String
            ],
            "Metrics" =>  ::Dict{String,String},
            "OptOut" =>  ::String,
            "RequestId" =>  ::String,
            "User" =>  [
                "UserAttributes" =>  ::Dict{String,String},
                "UserId" =>  ::String
            ]
        ], ...]]

Returns

UpdateEndpointsBatchResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_gcm_channel — Function.

update_gcm_channel(::AWSConfig, arguments::Dict)
update_gcm_channel([::AWSConfig]; application-id=, GCMChannelRequest=)

UpdateGcmChannel Operation

Use to update the GCM channel for an app.

Arguments

application-id = ::String – Required

GCMChannelRequest = [ ... ] – Required

 GCMChannelRequest = [
        "ApiKey" =>  ::String,
        "Enabled" =>  ::Bool
    ]

Returns

UpdateGcmChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_segment — Function.

update_segment(::AWSConfig, arguments::Dict)
update_segment([::AWSConfig]; application-id=, segment-id=, WriteSegmentRequest=)

UpdateSegment Operation

Use to update a segment.

Arguments

application-id = ::String – Required

segment-id = ::String – Required

WriteSegmentRequest = [ ... ] – Required

 WriteSegmentRequest = [
        "Dimensions" =>  [
            "Attributes" =>  ::Dict{String,String},
            "Behavior" =>  ["Recency" =>  [
                    "Duration" =>  "HR_24", "DAY_7", "DAY_14" or "DAY_30",
                    "RecencyType" =>  "ACTIVE" or "INACTIVE"
                ]],
            "Demographic" =>  [
                "AppVersion" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "Channel" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "DeviceType" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "Make" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "Model" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ],
                "Platform" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ]
            ],
            "Location" =>  ["Country" =>  [
                    "DimensionType" =>  "INCLUSIVE" or "EXCLUSIVE",
                    "Values" =>  [::String, ...]
                ]],
            "UserAttributes" =>  ::Dict{String,String}
        ],
        "Name" =>  ::String
    ]

Returns

UpdateSegmentResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation

AmazonPinpoint.update_sms_channel — Function.

update_sms_channel(::AWSConfig, arguments::Dict)
update_sms_channel([::AWSConfig]; application-id=, SMSChannelRequest=)

UpdateSmsChannel Operation

Update an SMS channel

Arguments

application-id = ::String – Required

SMSChannelRequest = [ ... ] – Required

 SMSChannelRequest = [
        "Enabled" =>  ::Bool,
        "SenderId" =>  ::String
    ]

Returns

UpdateSmsChannelResponse

Exceptions

BadRequestException, InternalServerErrorException, ForbiddenException, NotFoundException, MethodNotAllowedException or TooManyRequestsException.

See also: AWS API Documentation